R6 Class defining statistical tables that can be perturbed

This class allows to define statistical tables and perturb both count and numerical variables.

ck_setup(x, rkey, dims, w = NULL, countvars = NULL, numvars = NULL)

Arguments

x
an object coercible to a data.frame

rkey
either a column name within x referring to a variable containing record keys or a single integer(ish) number > 5 that referns to the number of digits for record keys that will be generated internally.

dims
a list containing slots for each variable that should be tabulated. Each slot consists should be created/modified using sdcHierarchies::hier_create(), sdcHierarchies::hier_add() and other functionality from package sdcHierarchies.

w
(character) a scalar character referring to a variable in x holding sampling weights. If w is NULL (the default), all weights are assumed to be 1

countvars
(character) an optional vector containing names of binary (0/1 coded) variables withing x that should be included in the problem instance. These variables can later be perturbed.

numvars
(character) an optional vector of numerical variables that can later be tabulated.

Details

Such objects are typically generated using ck_setup().

Methods

Public methods

ck_class$new()

ck_class$perturb()

ck_class$freqtab()

ck_class$numtab()

ck_class$measures_cnts()

ck_class$measures_nums()

ck_class$allvars()

ck_class$cntvars()

ck_class$numvars()

ck_class$hierarchy_info()

ck_class$mod_cnts()

ck_class$mod_nums()

ck_class$supp_freq()

ck_class$supp_val()

ck_class$supp_p()

ck_class$supp_pq()

ck_class$supp_nk()

ck_class$params_cnts_get()

ck_class$params_cnts_set()

ck_class$reset_cntvars()

ck_class$reset_numvars()

ck_class$reset_allvars()

ck_class$params_nums_get()

ck_class$params_nums_set()

ck_class$summary()

ck_class$print()

Method new()

Create a new table instance

Usage

ck_class$new(x, rkey, dims, w = NULL, countvars = NULL, numvars = NULL)

Arguments

x

an object coercible to a data.frame

rkey

either a column name within x referring to a variable containing record keys or a single integer(ish) number > 5 that referns to the number of digits for record keys that will be generated internally.

dims

a list containing slots for each variable that should be tabulated. Each slot consists should be created/modified using sdcHierarchies::hier_create(), sdcHierarchies::hier_add() and other functionality from package sdcHierarchies.

w

(character) a scalar character referring to a variable in x holding sampling weights. If w is NULL (the default), all weights are assumed to be 1

countvars

(character) an optional vector containing names of binary (0/1 coded) variables withing x that should be included in the problem instance. These variables can later be perturbed.

numvars

(character) an optional vector of numerical variables that can later be tabulated.

Returns

A new cellkey_obj object.

Method perturb()

Perturb a count- or magnitude variable

Usage

ck_class$perturb(v)

Arguments

v

name(s) of count- or magnitude variables that should be perturbed.

Method freqtab()

Extract results from already perturbed count variables as a data.table

Usage

ck_class$freqtab(v = NULL, path = NULL)

Arguments

v

a vector of variable names for count variables. If NULL (the default), the results are returned for all available count variables. For variables that have not yet perturbed, columns puwc and pwc are filled with NA.

path

if not NULL, a scalar character defining a (relative or absolute) path to which the result table should be written. A csv file will be generated and, if specified, path must have ".csv" as file-ending

Returns

This method returns a data.table containing all combinations of the dimensional variables in the first n columns. Additionally, the following columns are shown:

vname: name of the perturbed variable

uwc: unweighted counts

wc: weighted counts

puwc: perturbed unweighted counts or NA if vname was not yet perturbed

pwc: perturbed weighted counts or NA if vname was not yet perturbed

Method numtab()

Extract results from already perturbed continuous variables as a data.table.

Usage

ck_class$numtab(v = NULL, mean_before_sum = FALSE, path = NULL)

Arguments

v

a vector of variable names of continuous variables. If NULL (the default), the results are returned for all available numeric variables.

mean_before_sum

(logical); if TRUE, the perturbed values are adjusted by a factor ((n+p))⁄n with

n: the original weighted cell value

p: the perturbed cell value

This makes sense if the the accuracy of the variable mean is considered to be more important than accuracy of sums of the variable. The default value is FALSE (no adjustment is done)

path

if not NULL, a scalar character defining a (relative or absolute) path to which the result table should be written. A csv file will be generated and, if specified, path must have ".csv" as file-ending

Returns

This method returns a data.table containing all combinations of the dimensional variables in the first n columns. Additionally, the following columns are shown:

vname: name of the perturbed variable

uws: unweighted sum of the given variable

ws: weighted cellsum

pws: perturbed weighted sum of the given cell or NA if vname has not not perturbed

Method measures_cnts()

Utility measures for perturbed count variables

Usage

ck_class$measures_cnts(v, exclude_zeros = TRUE)

Arguments

v

name of a count variable for which utility measures should be computed.

exclude_zeros

should empty (zero) cells in the original values be excluded when computing distance measures

Returns

This method returns a list containing a set of utility measures based on some distance functions. For a detailed description of the computed measures, see ck_cnt_measures()

Method measures_nums()

Utility measures for continuous variables (not yet implemented)

Usage

ck_class$measures_nums(v)

Arguments

v

name of a continuosu variable for which utility measures should be computed.

Method allvars()

returns a list with the following two elements:

cntvars: character vector with names of available count variables for perturbation

numvars: character vector with names of available numerical variables for perturbation

Usage

ck_class$allvars()

Method cntvars()

character vector with names of count variables that could be perturbed

Usage

ck_class$cntvars()

Method numvars()

character vector with names of continuous variables that could be perturbed

Usage

ck_class$numvars()

Method hierarchy_info()

Information about hierarchies

Usage

ck_class$hierarchy_info()

Returns

a list (for each dimensional variable) with information on the hierarchies. This may be used to restrict output tables to specific levels or codes. Each list element is a data.table containing the following variables:

code: the name of a code within the hierarchy

level: number defining the level of the code; the higher the number, the lower the hierarchy with 1 being the overall total

is_leaf: if TRUE, this code is a leaf node which means no other codes contribute to it

parent: name of the parent code

Method mod_cnts()

Modifications applied to count variables

Usage

ck_class$mod_cnts()

Returns

a data.table containing modifications applied to count variables

Method mod_nums()

Modifications applied to numerical variables

Usage

ck_class$mod_nums()

Returns

a data.table containing modifications applied to numerical variables

Method supp_freq()

Identify sensitive cells based on minimum frequency rule

Usage

ck_class$supp_freq(v, n, weighted = TRUE)

Arguments

v

a single variable name of a continuous variable (see method numvars())

n

a number defining the threshold. All cells <= n are considered as unsafe.

weighted

if TRUE, the weighted number of contributors to a cell are compared to the threshold specified in n (default); else the unweighted number of contributors is used.

Method supp_val()

Identify sensitive cells based on weighted or unweighted cell values.minimum frequency rule

Usage

ck_class$supp_val(v, n, weighted = TRUE)

Arguments

v

a single variable name of a continuous variable (see method numvars())

n

a number defining the threshold. All cells <= n are considered as unsafe.

weighted

if TRUE, the weighted cell value of variable v is compared to the threshold specified in n (default); else the unweighted number is used.

Method supp_p()

Identify sensitive cells based on the p%-rule rule. Please note that this rule can only be applied to positive-only variables.

Usage

ck_class$supp_p(v, p)

Arguments

v

a single variable name of a continuous variable (see method numvars())

p

a number defining a percentage between 1 and 99.

Method supp_pq()

Identify sensitive cells based on the pq-rule. Please note that this rule can only be applied to positive-only variables.

Usage

ck_class$supp_pq(v, p, q)

Arguments

v

a single variable name of a continuous variable (see method numvars())

p

a number defining a percentage between 1 and 99.

q

a number defining a percentage between 1 and 99. This value must be larger than p.

Method supp_nk()

Identify sensitive cells based on the nk-dominance rule. Please note that this rule can only be applied to positive-only variables.

Usage

ck_class$supp_nk(v, n, k)

Arguments

v

a single variable name of a continuous variable (see method numvars())

n

an integerish number >= 2

k

a number defining a percentage between 1 and 99. All cells to which the top n contributers contribute more than k% is considered unsafe

Method params_cnts_get()

Return perturbation parameters of count variables

Usage

ck_class$params_cnts_get()

Returns

a named list in which each list-element contains the active perturbation parameters for the specific count variable defined by the list-name.

Method params_cnts_set()

Set perturbation parameters for count variables

Usage

ck_class$params_cnts_set(val, v = NULL)

Arguments

val

a perturbation object created with ck_params_cnts()

v

a character vector (or NULL). If NULL (the default), the perturbation parameters provided in val are set for all count variables; otherwise one may specify the names of the count variables for which the parameters should be set.

Method reset_cntvars()

reset results and parameters for already perturbed count variables

Usage

ck_class$reset_cntvars(v = NULL)

Arguments

v

if v equals NULL (the default), the results are reset for all perturbed count variables; otherwise it is possible to specify the names of already perturbed count variables.

Method reset_numvars()

reset results and parameters for already perturbed numerical variables

Usage

ck_class$reset_numvars(v = NULL)

Arguments

v

if v equals NULL (the default), the results are reset for all perturbed numerical variables; otherwise it is possible to specify the names of already perturbed continuous variables.

Method reset_allvars()

reset results and parameters for all already perturbed variables.

Usage

ck_class$reset_allvars()

Method params_nums_get()

Return perturbation parameters of continuous variables

Usage

ck_class$params_nums_get()

Returns

a named list in which each list-element contains the active perturbation parameters for the specific continuous variable defined by the list-name.

Method params_nums_set()

set perturbation parameters for continuous variables.

Usage

ck_class$params_nums_set(val, v = NULL)

Arguments

val

a perturbation object created with ck_params_nums()

v

a character vector (or NULL); if NULL (the default), the perturbation parameters provided in val are set for all continuous variables; otherwise one may specify the names of the numeric variables for which the parameters should be set.

Method summary()

some aggregated summary statistics about perturbed variables

Usage

ck_class$summary()

Method print()

prints information about the current table

Usage

ck_class$print()

Examples

x <- ck_create_testdata()

# create some 0/1 variables that should be perturbed later
x[, cnt_females := ifelse(sex == "male", 0, 1)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              64          25.00000           0
#>    2:   2530      28       1              42          25.00000           1
#>    3:   6920     550       1              95          25.00000           0
#>    4:   7960     870       1              82          25.00000           0
#>    5:   9030      20       2              66          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              67          16.66667           1
#> 4577:   1420     987    1000              51          16.66667           0
#> 4578:   8900     684    1000              92          16.66667           0
#> 4579:   3880     294    1000              40          16.66667           1
#> 4580:   4830     911    1000              39          16.66667           0

x[, cnt_males := ifelse(sex == "male", 1, 0)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              64          25.00000           0
#>    2:   2530      28       1              42          25.00000           1
#>    3:   6920     550       1              95          25.00000           0
#>    4:   7960     870       1              82          25.00000           0
#>    5:   9030      20       2              66          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              67          16.66667           1
#> 4577:   1420     987    1000              51          16.66667           0
#> 4578:   8900     684    1000              92          16.66667           0
#> 4579:   3880     294    1000              40          16.66667           1
#> 4580:   4830     911    1000              39          16.66667           0
#>       cnt_males
#>    1:         1
#>    2:         0
#>    3:         1
#>    4:         1
#>    5:         1
#>   ---          
#> 4576:         0
#> 4577:         1
#> 4578:         1
#> 4579:         0
#> 4580:         1

x[, cnt_highincome := ifelse(income >= 9000, 1, 0)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              64          25.00000           0
#>    2:   2530      28       1              42          25.00000           1
#>    3:   6920     550       1              95          25.00000           0
#>    4:   7960     870       1              82          25.00000           0
#>    5:   9030      20       2              66          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              67          16.66667           1
#> 4577:   1420     987    1000              51          16.66667           0
#> 4578:   8900     684    1000              92          16.66667           0
#> 4579:   3880     294    1000              40          16.66667           1
#> 4580:   4830     911    1000              39          16.66667           0
#>       cnt_males cnt_highincome
#>    1:         1              0
#>    2:         0              0
#>    3:         1              0
#>    4:         1              0
#>    5:         1              1
#>   ---                         
#> 4576:         0              0
#> 4577:         1              0
#> 4578:         1              0
#> 4579:         0              0
#> 4580:         1              0

# a variable with positive and negative contributions
x[, mixed := sample(-10:10, nrow(x), replace = TRUE)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              64          25.00000           0
#>    2:   2530      28       1              42          25.00000           1
#>    3:   6920     550       1              95          25.00000           0
#>    4:   7960     870       1              82          25.00000           0
#>    5:   9030      20       2              66          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              67          16.66667           1
#> 4577:   1420     987    1000              51          16.66667           0
#> 4578:   8900     684    1000              92          16.66667           0
#> 4579:   3880     294    1000              40          16.66667           1
#> 4580:   4830     911    1000              39          16.66667           0
#>       cnt_males cnt_highincome mixed
#>    1:         1              0    -6
#>    2:         0              0     2
#>    3:         1              0    -2
#>    4:         1              0     0
#>    5:         1              1     8
#>   ---                               
#> 4576:         0              0    -2
#> 4577:         1              0     5
#> 4578:         1              0    -4
#> 4579:         0              0     7
#> 4580:         1              0    -9

# create record keys
x$rkey <- ck_generate_rkeys(dat = x)

# define required inputs

# hierarchy with some bogus codes
d_sex <- hier_create(root = "Total", nodes = c("male", "female"))
d_sex <- hier_add(d_sex, root = "female", "f")
d_sex <- hier_add(d_sex, root = "male", "m")

d_age <- hier_create(root = "Total", nodes = paste0("age_group", 1:6))
d_age <- hier_add(d_age, root = "age_group1", "ag1a")
d_age <- hier_add(d_age, root = "age_group2", "ag2a")

# define the cell key object
countvars <- c("cnt_females", "cnt_males", "cnt_highincome")
numvars <- c("expend", "income", "savings", "mixed")
tab <- ck_setup(
  x = x,
  rkey = "rkey",
  dims = list(sex = d_sex, age = d_age),
  w = "sampling_weight",
  countvars = countvars,
  numvars = numvars)

# show some information about this table instance
tab$print() # identical with print(tab)

#> ── Table Information ───────────────────────────────────────────────────────────
#> ✔ 45 cells in 2 dimensions ('sex', 'age')
#> ✔ weights: yes
#> ── Tabulated / Perturbed countvars ─────────────────────────────────────────────
#> ☐ 'total'
#> ☐ 'cnt_females'
#> ☐ 'cnt_males'
#> ☐ 'cnt_highincome'
#> ── Tabulated / Perturbed numvars ───────────────────────────────────────────────
#> ☐ 'expend'
#> ☐ 'income'
#> ☐ 'savings'
#> ☐ 'mixed'

# information about the hierarchies
tab$hierarchy_info()

#> $sex
#>      code level is_leaf parent
#> 1:  Total     1   FALSE   <NA>
#> 2:   male     2   FALSE  Total
#> 3: female     2   FALSE  Total
#> 4:      f     3    TRUE female
#> 5:      m     3    TRUE   male
#> 
#> $age
#>          code level is_leaf     parent
#> 1:      Total     1   FALSE       <NA>
#> 2: age_group1     2   FALSE      Total
#> 3: age_group2     2   FALSE      Total
#> 4: age_group3     2    TRUE      Total
#> 5: age_group4     2    TRUE      Total
#> 6: age_group5     2    TRUE      Total
#> 7: age_group6     2    TRUE      Total
#> 8:       ag1a     3    TRUE age_group1
#> 9:       ag2a     3    TRUE age_group2
#> 

# which variables have been defined?
tab$allvars()

#> $cntvars
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"
#> 
#> $numvars
#> [1] "expend"  "income"  "savings" "mixed"  
#> 

# count variables
tab$cntvars()

#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"

# continuous variables
tab$numvars()

#> [1] "expend"  "income"  "savings" "mixed"  

# create perturbation parameters for "total" variable and
# write to yaml-file

# create a ptable using functionality from the ptable-pkg
f_yaml <- tempfile(fileext = ".yaml")
p_cnts1 <- ck_params_cnts(
  ptab = ptable::pt_ex_cnts(),
  path = f_yaml)

#> yaml configuration '/tmp/RtmptOwavS/fileeda2f15e8d6.yaml' successfully written.

# read parameters from yaml-file and set them for variable `"total"`
p_cnts1 <- ck_read_yaml(path = f_yaml)

tab$params_cnts_set(val = p_cnts1, v = "total")

#> --> setting perturbation parameters for variable 'total'

# create alternative perturbation parameters by specifying parameters
para2 <- ptable::pt_create_pParams(
  D = 8, V = 3, js = 2, table = "cnts")

#> Ignoring argument `step` for freqency tables

p_cnts2 <- ck_params_cnts(ptab = para2)

# use these ptable it for the remaining variables
tab$params_cnts_set(val = p_cnts2, v = countvars)

#> --> setting perturbation parameters for variable 'cnt_females'

#> --> setting perturbation parameters for variable 'cnt_males'

#> --> setting perturbation parameters for variable 'cnt_highincome'

# perturb a variable
tab$perturb(v = "total")

#> Count variable 'total' was perturbed.

# multiple variables can be perturbed as well
tab$perturb(v = c("cnt_males", "cnt_highincome"))

#> Count variable 'cnt_males' was perturbed.

#> Count variable 'cnt_highincome' was perturbed.

# return weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))

#>        sex        age     vname  uwc     wc puwc        pwc
#>  1:  Total      Total     total 4580 274707 4582 274826.959
#>  2:  Total age_group1     total 1969 118802 1969 118802.000
#>  3:  Total       ag1a     total 1969 118802 1969 118802.000
#>  4:  Total age_group2     total 1143  68583 1141  68462.995
#>  5:  Total       ag2a     total 1143  68583 1141  68462.995
#>  6:  Total age_group3     total  864  51473  864  51473.000
#>  7:  Total age_group4     total  423  25121  423  25121.000
#>  8:  Total age_group5     total  168   9970  167   9910.655
#>  9:  Total age_group6     total   13    758   13    758.000
#> 10: female      Total     total 2284 137362 2284 137362.000
#> 11:      f      Total     total 2284 137362 2284 137362.000
#> 12: female age_group1     total  954  57468  953  57407.761
#> 13:      f age_group1     total  954  57468  953  57407.761
#> 14: female       ag1a     total  954  57468  953  57407.761
#> 15:      f       ag1a     total  954  57468  953  57407.761
#> 16: female age_group2     total  572  34475  571  34414.729
#> 17:      f age_group2     total  572  34475  571  34414.729
#> 18: female       ag2a     total  572  34475  571  34414.729
#> 19:      f       ag2a     total  572  34475  571  34414.729
#> 20: female age_group3     total  440  26490  439  26429.795
#> 21:      f age_group3     total  440  26490  439  26429.795
#> 22: female age_group4     total  228  13373  228  13373.000
#> 23:      f age_group4     total  228  13373  228  13373.000
#> 24: female age_group5     total   84   5274   84   5274.000
#> 25:      f age_group5     total   84   5274   84   5274.000
#> 26: female age_group6     total    6    282    7    329.000
#> 27:      f age_group6     total    6    282    7    329.000
#> 28:   male      Total     total 2296 137345 2296 137345.000
#> 29:      m      Total     total 2296 137345 2296 137345.000
#> 30:   male age_group1     total 1015  61334 1014  61273.572
#> 31:      m age_group1     total 1015  61334 1014  61273.572
#> 32:   male       ag1a     total 1015  61334 1014  61273.572
#> 33:      m       ag1a     total 1015  61334 1014  61273.572
#> 34:   male age_group2     total  571  34108  572  34167.734
#> 35:      m age_group2     total  571  34108  572  34167.734
#> 36:   male       ag2a     total  571  34108  572  34167.734
#> 37:      m       ag2a     total  571  34108  572  34167.734
#> 38:   male age_group3     total  424  24983  424  24983.000
#> 39:      m age_group3     total  424  24983  424  24983.000
#> 40:   male age_group4     total  195  11748  196  11808.246
#> 41:      m age_group4     total  195  11748  196  11808.246
#> 42:   male age_group5     total   84   4696   84   4696.000
#> 43:      m age_group5     total   84   4696   84   4696.000
#> 44:   male age_group6     total    7    476    7    476.000
#> 45:      m age_group6     total    7    476    7    476.000
#> 46:  Total      Total cnt_males 2296 137345 2296 137345.000
#> 47:  Total age_group1 cnt_males 1015  61334 1014  61273.572
#> 48:  Total       ag1a cnt_males 1015  61334 1014  61273.572
#> 49:  Total age_group2 cnt_males  571  34108  572  34167.734
#> 50:  Total       ag2a cnt_males  571  34108  572  34167.734
#> 51:  Total age_group3 cnt_males  424  24983  424  24983.000
#> 52:  Total age_group4 cnt_males  195  11748  196  11808.246
#> 53:  Total age_group5 cnt_males   84   4696   85   4751.905
#> 54:  Total age_group6 cnt_males    7    476    7    476.000
#> 55: female      Total cnt_males    0      0    0      0.000
#> 56:      f      Total cnt_males    0      0    0      0.000
#> 57: female age_group1 cnt_males    0      0    0      0.000
#> 58:      f age_group1 cnt_males    0      0    0      0.000
#> 59: female       ag1a cnt_males    0      0    0      0.000
#> 60:      f       ag1a cnt_males    0      0    0      0.000
#> 61: female age_group2 cnt_males    0      0    0      0.000
#> 62:      f age_group2 cnt_males    0      0    0      0.000
#> 63: female       ag2a cnt_males    0      0    0      0.000
#> 64:      f       ag2a cnt_males    0      0    0      0.000
#> 65: female age_group3 cnt_males    0      0    0      0.000
#> 66:      f age_group3 cnt_males    0      0    0      0.000
#> 67: female age_group4 cnt_males    0      0    0      0.000
#> 68:      f age_group4 cnt_males    0      0    0      0.000
#> 69: female age_group5 cnt_males    0      0    0      0.000
#> 70:      f age_group5 cnt_males    0      0    0      0.000
#> 71: female age_group6 cnt_males    0      0    0      0.000
#> 72:      f age_group6 cnt_males    0      0    0      0.000
#> 73:   male      Total cnt_males 2296 137345 2296 137345.000
#> 74:      m      Total cnt_males 2296 137345 2296 137345.000
#> 75:   male age_group1 cnt_males 1015  61334 1014  61273.572
#> 76:      m age_group1 cnt_males 1015  61334 1014  61273.572
#> 77:   male       ag1a cnt_males 1015  61334 1014  61273.572
#> 78:      m       ag1a cnt_males 1015  61334 1014  61273.572
#> 79:   male age_group2 cnt_males  571  34108  572  34167.734
#> 80:      m age_group2 cnt_males  571  34108  572  34167.734
#> 81:   male       ag2a cnt_males  571  34108  572  34167.734
#> 82:      m       ag2a cnt_males  571  34108  572  34167.734
#> 83:   male age_group3 cnt_males  424  24983  424  24983.000
#> 84:      m age_group3 cnt_males  424  24983  424  24983.000
#> 85:   male age_group4 cnt_males  195  11748  196  11808.246
#> 86:      m age_group4 cnt_males  195  11748  196  11808.246
#> 87:   male age_group5 cnt_males   84   4696   85   4751.905
#> 88:      m age_group5 cnt_males   84   4696   85   4751.905
#> 89:   male age_group6 cnt_males    7    476    7    476.000
#> 90:      m age_group6 cnt_males    7    476    7    476.000
#>        sex        age     vname  uwc     wc puwc        pwc

# numerical variables (positive variables using flex-function)
# we also write the config to a yaml file
f_yaml <- tempfile(fileext = ".yaml")

# create a ptable using functionality from the ptable-pkg
# a single ptable for all cells
ptab1 <- ptable::pt_ex_nums(parity = TRUE, separation = FALSE)

# a single ptab for all cells except for very small ones
ptab2 <- ptable::pt_ex_nums(parity = TRUE, separation = TRUE)

# different ptables for cells with even/odd number of contributors
# and very small cells
ptab3 <- ptable::pt_ex_nums(parity = FALSE, separation = TRUE)

p_nums1 <- ck_params_nums(
  ptab = ptab1,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.30, 0.03),
    epsilon = c(1, 0.5, 0.2),
    q = 3),
  mu_c = 2,
  same_key = FALSE,
  use_zero_rkeys = FALSE,
  path = f_yaml)

#> yaml configuration '/tmp/RtmptOwavS/fileeda5049da9.yaml' successfully written.

# we read the parameters from the yaml-file
p_nums1 <- ck_read_yaml(path = f_yaml)

# for variables with positive and negative values
p_nums2 <- ck_params_nums(
  ptab = ptab2,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.15, 0.02),
    epsilon = c(1, 0.4, 0.15),
    q = 3),
  mu_c = 2,
  same_key = FALSE)

# simple perturbation parameters (not using the flex-function approach)
p_nums3 <- ck_params_nums(
  ptab = ptab3,
  type = "mean",
  mult_params = ck_simpleparams(p = 0.25),
  mu_c = 2,
  same_key = FALSE)

# use `p_nums1` for all variables
tab$params_nums_set(p_nums1, c("savings", "income", "expend"))

#> --> setting perturbation parameters for variable 'savings'

#> --> setting perturbation parameters for variable 'income'

#> --> setting perturbation parameters for variable 'expend'

# use different parameters for variable `mixed`
tab$params_nums_set(p_nums2, v = "mixed")

#> --> setting perturbation parameters for variable 'mixed'

# identify sensitive cells to which extra protection (`mu_c`) is added.
tab$supp_p(v = "income", p = 85)

#> p%-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)

tab$supp_pq(v = "income", p = 85, q = 90)

#> pq-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)

tab$supp_nk(v = "income", n = 2, k = 90)

#> nk-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)

tab$supp_freq(v = "income", n = 14, weighted = FALSE)

#> freq-rule: 5 new sensitive cells (incl. duplicates) found (total: 5)

tab$supp_val(v = "income", n = 10000, weighted = TRUE)

#> val-rule: 0 new sensitive cells (incl. duplicates) found (total: 5)

# perturb variables
tab$perturb(v = c("income", "savings"))

#> Numeric variable 'income' was perturbed.

#> Numeric variable 'savings' was perturbed.

# extract results
tab$numtab("income", mean_before_sum = TRUE)

#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1377005797 1376945138
#>  2:  Total age_group1 income  9810547  590371940  590376154
#>  3:  Total       ag1a income  9810547  590371940  590376154
#>  4:  Total age_group2 income  5692119  340327551  340187928
#>  5:  Total       ag2a income  5692119  340327551  340187928
#>  6:  Total age_group3 income  4406946  263870748  263807267
#>  7:  Total age_group4 income  2133543  128635475  128623136
#>  8:  Total age_group5 income   848151   50513336   50403682
#>  9:  Total age_group6 income    61672    3286747    3051762
#> 10: female      Total income 11690929  702755232  703051791
#> 11:      f      Total income 11690929  702755232  703051791
#> 12: female age_group1 income  4933383  296150572  296142272
#> 13:      f age_group1 income  4933383  296150572  296142272
#> 14: female       ag1a income  4933383  296150572  296142272
#> 15:      f       ag1a income  4933383  296150572  296142272
#> 16: female age_group2 income  2880740  173368984  173383331
#> 17:      f age_group2 income  2880740  173368984  173383331
#> 18: female       ag2a income  2880740  173368984  173383331
#> 19:      f       ag2a income  2880740  173368984  173383331
#> 20: female age_group3 income  2238777  134883636  134914416
#> 21:      f age_group3 income  2238777  134883636  134914416
#> 22: female age_group4 income  1155033   68276741   68470582
#> 23:      f age_group4 income  1155033   68276741   68470582
#> 24: female age_group5 income   455017   28913759   28953004
#> 25:      f age_group5 income   455017   28913759   28953004
#> 26: female age_group6 income    27979    1161540    1111365
#> 27:      f age_group6 income    27979    1161540    1111365
#> 28:   male      Total income 11262049  674250565  674175625
#> 29:      m      Total income 11262049  674250565  674175625
#> 30:   male age_group1 income  4877164  294221368  294321199
#> 31:      m age_group1 income  4877164  294221368  294321199
#> 32:   male       ag1a income  4877164  294221368  294321199
#> 33:      m       ag1a income  4877164  294221368  294321199
#> 34:   male age_group2 income  2811379  166958567  166998608
#> 35:      m age_group2 income  2811379  166958567  166998608
#> 36:   male       ag2a income  2811379  166958567  166998608
#> 37:      m       ag2a income  2811379  166958567  166998608
#> 38:   male age_group3 income  2168169  128987112  129015795
#> 39:      m age_group3 income  2168169  128987112  129015795
#> 40:   male age_group4 income   978510   60358734   60375054
#> 41:      m age_group4 income   978510   60358734   60375054
#> 42:   male age_group5 income   393134   21599577   21687227
#> 43:      m age_group5 income   393134   21599577   21687227
#> 44:   male age_group6 income    33693    2125207    2198004
#> 45:      m age_group6 income    33693    2125207    2198004
#>        sex        age  vname      uws         ws        pws

tab$numtab("income", mean_before_sum = FALSE)

#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1377005797 1376975467
#>  2:  Total age_group1 income  9810547  590371940  590374047
#>  3:  Total       ag1a income  9810547  590371940  590374047
#>  4:  Total age_group2 income  5692119  340327551  340257732
#>  5:  Total       ag2a income  5692119  340327551  340257732
#>  6:  Total age_group3 income  4406946  263870748  263839006
#>  7:  Total age_group4 income  2133543  128635475  128629305
#>  8:  Total age_group5 income   848151   50513336   50458479
#>  9:  Total age_group6 income    61672    3286747    3167076
#> 10: female      Total income 11690929  702755232  702903496
#> 11:      f      Total income 11690929  702755232  702903496
#> 12: female age_group1 income  4933383  296150572  296146422
#> 13:      f age_group1 income  4933383  296150572  296146422
#> 14: female       ag1a income  4933383  296150572  296146422
#> 15:      f       ag1a income  4933383  296150572  296146422
#> 16: female age_group2 income  2880740  173368984  173376158
#> 17:      f age_group2 income  2880740  173368984  173376158
#> 18: female       ag2a income  2880740  173368984  173376158
#> 19:      f       ag2a income  2880740  173368984  173376158
#> 20: female age_group3 income  2238777  134883636  134899025
#> 21:      f age_group3 income  2238777  134883636  134899025
#> 22: female age_group4 income  1155033   68276741   68373593
#> 23:      f age_group4 income  1155033   68276741   68373593
#> 24: female age_group5 income   455017   28913759   28933375
#> 25:      f age_group5 income   455017   28913759   28933375
#> 26: female age_group6 income    27979    1161540    1136176
#> 27:      f age_group6 income    27979    1161540    1136176
#> 28:   male      Total income 11262049  674250565  674213094
#> 29:      m      Total income 11262049  674250565  674213094
#> 30:   male age_group1 income  4877164  294221368  294271279
#> 31:      m age_group1 income  4877164  294221368  294271279
#> 32:   male       ag1a income  4877164  294221368  294271279
#> 33:      m       ag1a income  4877164  294221368  294271279
#> 34:   male age_group2 income  2811379  166958567  166978586
#> 35:      m age_group2 income  2811379  166958567  166978586
#> 36:   male       ag2a income  2811379  166958567  166978586
#> 37:      m       ag2a income  2811379  166958567  166978586
#> 38:   male age_group3 income  2168169  128987112  129001453
#> 39:      m age_group3 income  2168169  128987112  129001453
#> 40:   male age_group4 income   978510   60358734   60366894
#> 41:      m age_group4 income   978510   60358734   60366894
#> 42:   male age_group5 income   393134   21599577   21643358
#> 43:      m age_group5 income   393134   21599577   21643358
#> 44:   male age_group6 income    33693    2125207    2161299
#> 45:      m age_group6 income    33693    2125207    2161299
#>        sex        age  vname      uws         ws        pws

tab$numtab("savings")

#>        sex        age   vname     uws        ws         pws
#>  1:  Total      Total savings 2273532 136873818 136870309.5
#>  2:  Total age_group1 savings  982386  59422265  59420448.2
#>  3:  Total       ag1a savings  982386  59422265  59420448.2
#>  4:  Total age_group2 savings  552336  33246815  33243636.4
#>  5:  Total       ag2a savings  552336  33246815  33243636.4
#>  6:  Total age_group3 savings  437101  26180709  26181760.6
#>  7:  Total age_group4 savings  214661  12776589  12774656.3
#>  8:  Total age_group5 savings   80451   4886105   4878388.5
#>  9:  Total age_group6 savings    6597    361335    353629.9
#> 10: female      Total savings 1113716  67211781  67214410.0
#> 11:      f      Total savings 1113716  67211781  67214410.0
#> 12: female age_group1 savings  464726  27991793  27991815.6
#> 13:      f age_group1 savings  464726  27991793  27991815.6
#> 14: female       ag1a savings  464726  27991793  27991815.6
#> 15:      f       ag1a savings  464726  27991793  27991815.6
#> 16: female age_group2 savings  271413  16410574  16411672.7
#> 17:      f age_group2 savings  271413  16410574  16411672.7
#> 18: female       ag2a savings  271413  16410574  16411672.7
#> 19:      f       ag2a savings  271413  16410574  16411672.7
#> 20: female age_group3 savings  222131  13497780  13501050.9
#> 21:      f age_group3 savings  222131  13497780  13501050.9
#> 22: female age_group4 savings  115241   6810459   6819421.9
#> 23:      f age_group4 savings  115241   6810459   6819421.9
#> 24: female age_group5 savings   37218   2377583   2380704.9
#> 25:      f age_group5 savings   37218   2377583   2380704.9
#> 26: female age_group6 savings    2987    123592    122549.0
#> 27:      f age_group6 savings    2987    123592    122549.0
#> 28:   male      Total savings 1159816  69662037  69662037.0
#> 29:      m      Total savings 1159816  69662037  69662037.0
#> 30:   male age_group1 savings  517660  31430472  31435293.2
#> 31:      m age_group1 savings  517660  31430472  31435293.2
#> 32:   male       ag1a savings  517660  31430472  31435293.2
#> 33:      m       ag1a savings  517660  31430472  31435293.2
#> 34:   male age_group2 savings  280923  16836241  16836943.0
#> 35:      m age_group2 savings  280923  16836241  16836943.0
#> 36:   male       ag2a savings  280923  16836241  16836943.0
#> 37:      m       ag2a savings  280923  16836241  16836943.0
#> 38:   male age_group3 savings  214970  12682929  12680329.3
#> 39:      m age_group3 savings  214970  12682929  12680329.3
#> 40:   male age_group4 savings   99420   5966130   5964685.1
#> 41:      m age_group4 savings   99420   5966130   5964685.1
#> 42:   male age_group5 savings   43233   2508522   2514248.6
#> 43:      m age_group5 savings   43233   2508522   2514248.6
#> 44:   male age_group6 savings    3610    237743    236468.4
#> 45:      m age_group6 savings    3610    237743    236468.4
#>        sex        age   vname     uws        ws         pws

# results can be resetted, too
tab$reset_cntvars(v = "cnt_males")

# we can then set other parameters and perturb again
tab$params_cnts_set(val = p_cnts1, v = "cnt_males")

#> --> setting perturbation parameters for variable 'cnt_males'

tab$perturb(v = "cnt_males")

#> Count variable 'cnt_males' was perturbed.

# write to a file "outtab.csv" (.csv is automatically added to the path)
if (FALSE) {
tab$freqtab(v = c("total", "cnt_males"), path = "outtab.csv")
}

# show results containing weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))

#>        sex        age     vname  uwc     wc puwc        pwc
#>  1:  Total      Total     total 4580 274707 4582 274826.959
#>  2:  Total age_group1     total 1969 118802 1969 118802.000
#>  3:  Total       ag1a     total 1969 118802 1969 118802.000
#>  4:  Total age_group2     total 1143  68583 1141  68462.995
#>  5:  Total       ag2a     total 1143  68583 1141  68462.995
#>  6:  Total age_group3     total  864  51473  864  51473.000
#>  7:  Total age_group4     total  423  25121  423  25121.000
#>  8:  Total age_group5     total  168   9970  167   9910.655
#>  9:  Total age_group6     total   13    758   13    758.000
#> 10: female      Total     total 2284 137362 2284 137362.000
#> 11:      f      Total     total 2284 137362 2284 137362.000
#> 12: female age_group1     total  954  57468  953  57407.761
#> 13:      f age_group1     total  954  57468  953  57407.761
#> 14: female       ag1a     total  954  57468  953  57407.761
#> 15:      f       ag1a     total  954  57468  953  57407.761
#> 16: female age_group2     total  572  34475  571  34414.729
#> 17:      f age_group2     total  572  34475  571  34414.729
#> 18: female       ag2a     total  572  34475  571  34414.729
#> 19:      f       ag2a     total  572  34475  571  34414.729
#> 20: female age_group3     total  440  26490  439  26429.795
#> 21:      f age_group3     total  440  26490  439  26429.795
#> 22: female age_group4     total  228  13373  228  13373.000
#> 23:      f age_group4     total  228  13373  228  13373.000
#> 24: female age_group5     total   84   5274   84   5274.000
#> 25:      f age_group5     total   84   5274   84   5274.000
#> 26: female age_group6     total    6    282    7    329.000
#> 27:      f age_group6     total    6    282    7    329.000
#> 28:   male      Total     total 2296 137345 2296 137345.000
#> 29:      m      Total     total 2296 137345 2296 137345.000
#> 30:   male age_group1     total 1015  61334 1014  61273.572
#> 31:      m age_group1     total 1015  61334 1014  61273.572
#> 32:   male       ag1a     total 1015  61334 1014  61273.572
#> 33:      m       ag1a     total 1015  61334 1014  61273.572
#> 34:   male age_group2     total  571  34108  572  34167.734
#> 35:      m age_group2     total  571  34108  572  34167.734
#> 36:   male       ag2a     total  571  34108  572  34167.734
#> 37:      m       ag2a     total  571  34108  572  34167.734
#> 38:   male age_group3     total  424  24983  424  24983.000
#> 39:      m age_group3     total  424  24983  424  24983.000
#> 40:   male age_group4     total  195  11748  196  11808.246
#> 41:      m age_group4     total  195  11748  196  11808.246
#> 42:   male age_group5     total   84   4696   84   4696.000
#> 43:      m age_group5     total   84   4696   84   4696.000
#> 44:   male age_group6     total    7    476    7    476.000
#> 45:      m age_group6     total    7    476    7    476.000
#> 46:  Total      Total cnt_males 2296 137345 2296 137345.000
#> 47:  Total age_group1 cnt_males 1015  61334 1014  61273.572
#> 48:  Total       ag1a cnt_males 1015  61334 1014  61273.572
#> 49:  Total age_group2 cnt_males  571  34108  572  34167.734
#> 50:  Total       ag2a cnt_males  571  34108  572  34167.734
#> 51:  Total age_group3 cnt_males  424  24983  424  24983.000
#> 52:  Total age_group4 cnt_males  195  11748  196  11808.246
#> 53:  Total age_group5 cnt_males   84   4696   84   4696.000
#> 54:  Total age_group6 cnt_males    7    476    7    476.000
#> 55: female      Total cnt_males    0      0    0      0.000
#> 56:      f      Total cnt_males    0      0    0      0.000
#> 57: female age_group1 cnt_males    0      0    0      0.000
#> 58:      f age_group1 cnt_males    0      0    0      0.000
#> 59: female       ag1a cnt_males    0      0    0      0.000
#> 60:      f       ag1a cnt_males    0      0    0      0.000
#> 61: female age_group2 cnt_males    0      0    0      0.000
#> 62:      f age_group2 cnt_males    0      0    0      0.000
#> 63: female       ag2a cnt_males    0      0    0      0.000
#> 64:      f       ag2a cnt_males    0      0    0      0.000
#> 65: female age_group3 cnt_males    0      0    0      0.000
#> 66:      f age_group3 cnt_males    0      0    0      0.000
#> 67: female age_group4 cnt_males    0      0    0      0.000
#> 68:      f age_group4 cnt_males    0      0    0      0.000
#> 69: female age_group5 cnt_males    0      0    0      0.000
#> 70:      f age_group5 cnt_males    0      0    0      0.000
#> 71: female age_group6 cnt_males    0      0    0      0.000
#> 72:      f age_group6 cnt_males    0      0    0      0.000
#> 73:   male      Total cnt_males 2296 137345 2296 137345.000
#> 74:      m      Total cnt_males 2296 137345 2296 137345.000
#> 75:   male age_group1 cnt_males 1015  61334 1014  61273.572
#> 76:      m age_group1 cnt_males 1015  61334 1014  61273.572
#> 77:   male       ag1a cnt_males 1015  61334 1014  61273.572
#> 78:      m       ag1a cnt_males 1015  61334 1014  61273.572
#> 79:   male age_group2 cnt_males  571  34108  572  34167.734
#> 80:      m age_group2 cnt_males  571  34108  572  34167.734
#> 81:   male       ag2a cnt_males  571  34108  572  34167.734
#> 82:      m       ag2a cnt_males  571  34108  572  34167.734
#> 83:   male age_group3 cnt_males  424  24983  424  24983.000
#> 84:      m age_group3 cnt_males  424  24983  424  24983.000
#> 85:   male age_group4 cnt_males  195  11748  196  11808.246
#> 86:      m age_group4 cnt_males  195  11748  196  11808.246
#> 87:   male age_group5 cnt_males   84   4696   84   4696.000
#> 88:      m age_group5 cnt_males   84   4696   84   4696.000
#> 89:   male age_group6 cnt_males    7    476    7    476.000
#> 90:      m age_group6 cnt_males    7    476    7    476.000
#>        sex        age     vname  uwc     wc puwc        pwc

# utility measures for a count variable
tab$measures_cnts(v = "total", exclude_zeros = TRUE)

#> $overview
#>    noise cnt        pct
#> 1:    -2   1 0.02222222
#> 2:    -1   8 0.17777778
#> 3:     0  19 0.42222222
#> 4:     1  15 0.33333333
#> 5:     2   2 0.04444444
#> 
#> $measures
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 0.000 0.000 0.000
#>  6:   Mean 0.644 0.008 0.020
#>  7: Median 1.000 0.001 0.016
#>  8:    Q60 1.000 0.001 0.016
#>  9:    Q70 1.000 0.002 0.021
#> 10:    Q80 1.000 0.002 0.022
#> 11:    Q90 1.000 0.004 0.033
#> 12:    Q95 1.800 0.006 0.038
#> 13:    Q99 2.000 0.167 0.196
#> 14:    Max 2.000 0.167 0.196
#> 
#> $cumdistr_d1
#>    cat cnt       pct
#> 1:   0  19 0.4222222
#> 2:   1  42 0.9333333
#> 3:   2  45 1.0000000
#> 
#> $cumdistr_d2
#>            cat cnt       pct
#> 1:    [0,0.02]  43 0.9555556
#> 2: (0.02,0.05]  43 0.9555556
#> 3:  (0.05,0.1]  43 0.9555556
#> 4:   (0.1,0.2]  45 1.0000000
#> 5:   (0.2,0.3]  45 1.0000000
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $cumdistr_d3
#>            cat cnt       pct
#> 1:    [0,0.02]  28 0.6222222
#> 2: (0.02,0.05]  43 0.9555556
#> 3:  (0.05,0.1]  43 0.9555556
#> 4:   (0.1,0.2]  45 1.0000000
#> 5:   (0.2,0.3]  45 1.0000000
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $false_zero
#> [1] 0
#> 
#> $false_nonzero
#> [1] 0
#> 
#> $exclude_zeros
#> [1] TRUE
#> 

# modifications for perturbed count variables
tab$mod_cnts()

#>        sex        age row_nr pert       ckey  countvar
#>   1: Total      Total     17    2 0.95532247     total
#>   2: Total age_group1     15    0 0.44707334     total
#>   3: Total       ag1a     15    0 0.44707334     total
#>   4: Total age_group2     13   -2 0.01327944     total
#>   5: Total       ag2a     13   -2 0.01327944     total
#>  ---                                                  
#> 131:     m age_group4     16    1 0.72903399 cnt_males
#> 132:  male age_group5     15    0 0.68482099 cnt_males
#> 133:     m age_group5     15    0 0.68482099 cnt_males
#> 134:  male age_group6     15    0 0.48143486 cnt_males
#> 135:     m age_group6     15    0 0.48143486 cnt_males

# display a summary about utility measures
tab$summary()

#> ┌──────────────────────────────────────────────┐
#> │Utility measures for perturbed count variables│
#> └──────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#>          countvar Min Q10 Q20 Q30  Q40   Mean Median Q60 Q70 Q80 Q90 Q95  Q99
#> 1:          total  -2  -1  -1  -1  0.0 -0.200      0   0   0 0.2   1   1 1.56
#> 2: cnt_highincome  -4  -3  -2  -2 -1.4 -0.778     -1   0   0 1.0   2   2 2.00
#> 3:      cnt_males  -1  -1   0   0  0.0  0.067      0   0   0 0.2   1   1 1.00
#>    Max
#> 1:   2
#> 2:   2
#> 3:   1
#> 
#> ── Distance-based measures ─────────────────────────────────────────────────────
#> ✔ Variable: 'total'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 0.000 0.000 0.000
#>  6:   Mean 0.644 0.008 0.020
#>  7: Median 1.000 0.001 0.016
#>  8:    Q60 1.000 0.001 0.016
#>  9:    Q70 1.000 0.002 0.021
#> 10:    Q80 1.000 0.002 0.022
#> 11:    Q90 1.000 0.004 0.033
#> 12:    Q95 1.800 0.006 0.038
#> 13:    Q99 2.000 0.167 0.196
#> 14:    Max 2.000 0.167 0.196
#> 
#> ✔ Variable: 'cnt_males'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 0.000 0.000 0.000
#>  6:   Mean 0.556 0.001 0.012
#>  7: Median 1.000 0.001 0.016
#>  8:    Q60 1.000 0.001 0.016
#>  9:    Q70 1.000 0.002 0.021
#> 10:    Q80 1.000 0.002 0.021
#> 11:    Q90 1.000 0.003 0.027
#> 12:    Q95 1.000 0.005 0.036
#> 13:    Q99 1.000 0.005 0.036
#> 14:    Max 1.000 0.005 0.036
#> 
#> ✔ Variable: 'cnt_highincome'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 1.000 0.005 0.034
#>  3:    Q20 1.000 0.012 0.072
#>  4:    Q30 1.000 0.020 0.084
#>  5:    Q40 1.600 0.024 0.100
#>  6:   Mean 1.775 0.043 0.119
#>  7: Median 2.000 0.024 0.106
#>  8:    Q60 2.000 0.030 0.118
#>  9:    Q70 2.000 0.035 0.126
#> 10:    Q80 2.000 0.039 0.144
#> 11:    Q90 3.100 0.062 0.200
#> 12:    Q95 4.000 0.150 0.210
#> 13:    Q99 4.000 0.286 0.410
#> 14:    Max 4.000 0.286 0.410
#> 
#> ┌──────────────────────────────────────────────────┐
#> │Utility measures for perturbed numerical variables│
#> └──────────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────

#> Warning: no non-missing arguments to min; returning Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to min; returning Inf

#> Warning: no non-missing arguments to max; returning -Inf

#>      vname         Min        Q10        Q20       Q30      Q40     Mean
#> 1:  expend         Inf         NA         NA        NA       NA      NaN
#> 2:  income -119670.898 -37471.122 -25364.211 -4149.947 7173.574 12287.54
#> 3: savings   -7716.547  -2947.004  -1816.778 -1228.303   13.560   699.44
#> 4:   mixed         Inf         NA         NA        NA       NA      NaN
#>      Median      Q60       Q70       Q80       Q90       Q95        Q99
#> 1:       NA       NA        NA        NA        NA        NA         NA
#> 2: 8159.516 17079.81 20019.388 43780.682 49911.087 96851.629 148263.749
#> 3:  702.049  1070.46  2322.902  3270.909  4821.238  5726.561   8962.924
#> 4:       NA       NA        NA        NA        NA        NA         NA
#>           Max
#> 1:       -Inf
#> 2: 148263.749
#> 3:   8962.924
#> 4:       -Inf