ck_cnt_measures_basic

compute utility/information loss measures two numeric vectors (original and perturbed)

ck_cnt_measures_basic(orig, pert)

Arguments

orig
a numeric vector holding original values

pert
a numeric vector holding perturbed values

Value

a list containing the following elements:

measures: a data.table containing measures of the distribution of three different distances between original and perturbed values of the unweighted counts. The three distances are

absolute distance (column vals_abs), D1

relative distance (column vals_rel), D2

absolute distance between square-roots of original and perturbed values (column vals_r), D3

cumDistrA: a data.table with 3 columns showing the cummulative sum of absolute distances

kat: possible values for distance D1

val_abs: number of records smaller or equal the value in column kat

prop_abs: proportion or records smaller or equal the value in column kat

cumDistrB: a data.table with 5 columns which show the cummulative distributions for distances D2 (columns val_rel and prop_rel) and D3 (columns val_r and prop_r) for a set of intervals shown in column kat.

kat: a specific interval

val_rel: number of records smaller or equal the value in column kat for distance D2

prop_rel proportion of records smaller or equal the value in column kat for distance D2

val_r: number of records smaller or equal the value in column kat for distance D3

prop_r: proportion of records smaller or equal the value in column kat for distance D3

false_zero: number of cells that were perturbed to zero

false_positives: number of cells that were initially zero but have been perturbed to a number different from zero

Examples

orig <- 1:10
pert <- orig; pert[c(1, 5, 7)] <- c(0, 6, 9)
ck_cnt_measures_basic(
  orig = orig,
  pert = pert
)

#> $measures
#>       what vals_abs vals_rel vals_r
#>  1:    Min      0.0    0.000  0.000
#>  2:    Q10      0.0    0.000  0.000
#>  3:    Q20      0.0    0.000  0.000
#>  4:    Q30      0.0    0.000  0.000
#>  5:    Q40      0.0    0.000  0.000
#>  6:   Mean      0.4    0.149  0.157
#>  7: Median      0.0    0.000  0.000
#>  8:    Q60      0.0    0.000  0.000
#>  9:    Q70      0.3    0.060  0.064
#> 10:    Q80      1.0    0.217  0.242
#> 11:    Q90      1.1    0.357  0.419
#> 12:    Max      2.0    1.000  1.000
#> 
#> $cumdistrA
#>    kat val_abs prop_abs
#> 1:   0       7      0.7
#> 2:   1       9      0.9
#> 3:   2      10      1.0
#> 
#> $cumdistrB
#>             kat val_rel prop_rel val_r prop_r
#>  1:    [0,0.01]       7      0.7     7    0.7
#>  2: (0.01,0.02]       7      0.7     7    0.7
#>  3: (0.02,0.03]       7      0.7     7    0.7
#>  4: (0.03,0.04]       7      0.7     7    0.7
#>  5: (0.04,0.05]       7      0.7     7    0.7
#>  6: (0.05,0.06]       7      0.7     7    0.7
#>  7: (0.06,0.07]       7      0.7     7    0.7
#>  8: (0.07,0.08]       7      0.7     7    0.7
#>  9: (0.08,0.09]       7      0.7     7    0.7
#> 10:  (0.09,0.1]       7      0.7     7    0.7
#> 11:  (0.1,0.15]       7      0.7     7    0.7
#> 12:  (0.15,0.2]       8      0.8     7    0.8
#> 13:  (0.2,0.25]       8      0.8     8    0.8
#> 14:  (0.25,0.3]       9      0.9     8    0.9
#> 15:  (0.3,0.35]       9      0.9     8    0.9
#> 16:  (0.35,0.4]       9      0.9     9    0.9
#> 17:  (0.4,0.45]       9      0.9     9    0.9
#> 18:  (0.45,0.5]       9      0.9     9    0.9
#> 19:  (0.5,0.55]       9      0.9     9    0.9
#> 20:  (0.55,0.6]       9      0.9     9    0.9
#> 21:  (0.6,0.65]       9      0.9     9    0.9
#> 22:  (0.65,0.7]       9      0.9     9    0.9
#> 23:  (0.7,0.75]       9      0.9     9    0.9
#> 24:  (0.75,0.8]       9      0.9     9    0.9
#> 25:  (0.8,0.85]       9      0.9     9    0.9
#> 26:  (0.85,0.9]       9      0.9     9    0.9
#> 27:  (0.9,0.95]       9      0.9     9    0.9
#> 28:    (0.95,1]      10      1.0    10    1.0
#> 29:     (1,Inf]      10      1.0    10    1.0
#>             kat val_rel prop_rel val_r prop_r
#> 
#> $false_zero
#> [1] 1
#> 
#> $false_positives
#> [1] 0
#>