ck_cont_table

this function allows to return perturbed magnitude-tables from oututs of pert_table objects typically generated using perturbTable.

ck_cont_table(inp, vname = NULL, meanBeforeSum = TRUE)

Arguments

inp
an object of class pert_table generated with perturbTable

vname
either NULL or the name of a variable for which a perturbed frequeny table has been generated. If NULL, the function prints the variables for which a perturbation has been applied.

meanBeforeSum
(logical) defines if the weighted means should be computed before weighted sums or the other way round.

Value

a data.table containing all combinations of the dimensional variables in the first n columns. Afterwards, the following columns are shown:

UW_{vname}: unweighted sum for variable specified in vname

pUW_{vname}: perturbed unweighted sum for given variable

WS_{vname}: weighted sum for given variable

pWS_{vname}: perturbed weighted sum for variable specified in vname

pWM_{vname}: perturbed weighted mean for variable specified in vname

The return argument has a attribute called "modifications" which contains a data.table holding all the specific perturbation on unit-level for reproducibility. This can be extracted using attributes(ck_cont_table(...), "modifications"). .

Details

The computation of the perturbed weighted mean and the perturbed weighted sum depends on argument meanBeforeSum. If meanBeforeSum is TRUE, the perturbed mean is computed first and the perturbed sum is derived from the perturbed mean an the perturbed weighted counts. If meanBeforeSum is FALSE, the computation is the other way round. Details are available in the ABS paper Methodology for the Automatic Confidentialisation of Statistical Outputs from Remote Servers at the Australian Bureau of Statistics (Thompson, Broadfoot, Elazar).

See also

Examples

# see examples in perturbTable()