ck_create_pTable

generates perturbation tables using functionality from package ptable

ck_create_pTable(D, V, type, js = 0, pstay = NULL, optim = 1,
  mono = TRUE, epsilon = 0.0000001, pTableSize = 70)

Arguments

D
perturbation parameter for maximum perturbation (scalar or vector)

V
perturbation parameter for variance (scalar)

type
character specifying the format of the input table. Valid choices are "abs" and "destatis".

js
treshold value for blocking of small frequencies (i.e. there won't occur positive target frequencies below the treshold value)

pstay
optional parameter to set

optim
optimization parameter: 1 standard approach (default)

mono
(logical) vector specifying optimization parameter for monotony condition

epsilon
(double)

pTableSize
(number) defining the number of required columns in the "abs"-approach

Value

an object of class ptable defined in the ptable-package.

See also

Examples

# abs-format of perturbation table
ck_create_pTable(
  D = 5,
  V = 3,
  js = 2,
  pTableSize = 70,
  type = "abs"
)

#> An object of class "ptable"
#> Slot "pMatrix":
#>            0 1 2          3          4          5           6            7
#> 0 1.00000000 0 0 0.00000000 0.00000000 0.00000000 0.000000000 0.0000000000
#> 1 0.73446954 0 0 0.13672019 0.07279631 0.03742956 0.018584399 0.0000000000
#> 2 0.40880704 0 0 0.40664356 0.14762399 0.03231882 0.004266872 0.0003397187
#> 3 0.20398618 0 0 0.39366253 0.24749239 0.11056847 0.035102043 0.0079189056
#> 4 0.08485487 0 0 0.27792123 0.27792123 0.18939370 0.105125439 0.0454553635
#> 5 0.02014748 0 0 0.17334747 0.22063304 0.22131829 0.174967655 0.1090164422
#> 6 0.00000000 0 0 0.07202982 0.13671708 0.19735138 0.216652933 0.1808822565
#> 7 0.00000000 0 0 0.01843920 0.05548389 0.12162104 0.194208159 0.2259140240
#> 8 0.00000000 0 0 0.00379186 0.01659093 0.05229263 0.118730369 0.1941934604
#>             8           9          10          11          12         13
#> 0 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000 0.00000000
#> 1 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000 0.00000000
#> 2 0.000000000 0.000000000 0.000000000 0.000000000 0.000000000 0.00000000
#> 3 0.001269491 0.000000000 0.000000000 0.000000000 0.000000000 0.00000000
#> 4 0.015310774 0.004017389 0.000000000 0.000000000 0.000000000 0.00000000
#> 5 0.053532802 0.020717704 0.006319125 0.000000000 0.000000000 0.00000000
#> 6 0.114850887 0.055459983 0.020367237 0.005688418 0.000000000 0.00000000
#> 7 0.191441397 0.118180404 0.053146176 0.017410675 0.004155048 0.00000000
#> 8 0.228801497 0.194193460 0.118730369 0.052292633 0.016590928 0.00379186
#> 
#> Slot "pClasses":
#> [1] 0 1 2 3 4 5 6 7 8
#> 
#> Slot "pTable":
#>      V1 V2 V3 V4 V5 V6 V7 V8 V9 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V20 V21
#>   1: -1 -2  0 -1  1 -2 -1  0 -2  -2   0   2   0  -1  -1  -2  -1   2   0  -1  -3
#>   2: -1 -2  1  0 -1 -3  1 -2 -1   2  -2   1  -1   0  -1   3   4  -1   1  -2  -1
#>   3: -1  1  0  0 -2  2  0  1 -2   3   2  -3  -4   1  -1   4  -1  -2   0  -1   0
#>   4: -1  1 -3 -1  1 -1 -1 -2 -1   2   0  -1  -2  -1   2   0   0  -1   3   0  -3
#>   5:  3 -2  0  1 -1  1 -3 -1 -3   0   0   0   3   1   0   1  -2   1   1   2  -1
#>  ---                                                                           
#> 252: -1  1  3  1  1  2 -1 -1  1   0   1   1  -2   1  -1  -2  -1  -3   0   2   0
#> 253:  4 -2  0  0 -1 -2  0  0  3   3   0   1  -1   1  -3   2   1  -1   0  -1   0
#> 254: -1  1  0  4 -1  2 -1 -3  0   0   0  -3   3   2   0  -3  -1   1  -2   1  -1
#> 255: -1  1  0 -1  0 -2  1  1 -4  -2   1   0  -2   3   0   0  -1  -3   2   0  -2
#> 256: -1  1  0  2 -1  1  2  0  2  -2   0  -3  -1   1   0  -1   1  -2  -4   1   0
#>      V22 V23 V24 V25 V26 V27 V28 V29 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39
#>   1:   0   0   0   0  -1   3   0  -3  -1  -3   0   0  -2  -3   2  -2   2  -2
#>   2:   1   1   1   2   0   1  -2  -1   0   0   0  -1   1  -1  -1   2   1  -1
#>   3:  -5   0  -4  -1  -3   1  -2  -1   0   1  -3  -1   1   3   0  -1   0   1
#>   4:   2  -2   1   1   0  -2  -2   0   2   3   0  -3  -3  -2   0  -2   0  -1
#>   5:   1   2  -1   0  -1  -1   1   0   2   3   0  -3   3   0  -3   0  -2   0
#>  ---                                                                        
#> 252:  -1   0   3  -3   1  -2   0   1   2   1  -3   2  -3  -1   2   0  -1   0
#> 253:   4  -1   0  -1   3  -1   3  -2  -1  -1   2   2   0   2   1   1  -1  -1
#> 254:   0  -1   0   1   0   1  -1  -1  -1   2  -1   2   2  -1   1   5   2   0
#> 255:  -4   1   1   1   1   2   4   3  -4   0   0  -2   2  -3  -3   1  -2   0
#> 256:  -2   1  -3  -2   1   0   0   1   0  -5   1   4   1  -4  -4  -3  -4  -3
#>      V40 V41 V42 V43 V44 V45 V46 V47 V48 V49 V50 V51 V52 V53 V54 V55 V56 V57
#>   1:  -3   2   1  -1   0   1   1   0  -3   1   0   2   0   2   0  -1   2   0
#>   2:   2   2  -2  -1   0  -1   1   0  -1   1   1  -1   1  -2   0   3  -4   0
#>   3:  -1   0   1   1   4   0  -2   0  -2  -1  -1  -2   4  -1   0   0  -1  -2
#>   4:   0   1   1   1  -3   1  -1   2   3   1  -2   5   1   0   1  -1   1   1
#>   5:   0  -2   1   1   2   1  -4   1   0   3  -1   4   2   3   0  -2   1   0
#>  ---                                                                        
#> 252:   1   1   1  -1   1  -1   0   0   0   1  -1   0   0  -1   2  -1   1   1
#> 253:  -2   1   0   1  -4   0   1  -2   0   1   1  -1   2  -2  -2   0   0  -3
#> 254:  -2   2  -4   0  -1   0  -1   0   3   2   0   0   1  -2   2   0  -2   2
#> 255:  -3   0   1  -2   2  -1   0   1  -1   0   1  -3   0   0   0   3   2  -1
#> 256:   2   3   0  -2  -1   1  -1   2   0   0   0   0   0   1   3  -2   2  -1
#>      V58 V59 V60 V61 V62 V63 V64 V65 V66 V67 V68 V69 V70
#>   1:  -3   2   1  -1   0   2   0   0  -1   1   3   0   0
#>   2:   0  -1  -1  -3   0   1   2  -1   2  -1   2   3  -2
#>   3:  -1   3  -1   2   1   0   2  -1  -2  -3   2  -2  -2
#>   4:  -1   3  -2  -1   1  -1  -2   1   1  -1   2  -4  -1
#>   5:  -3   1   3  -2  -2   2   1  -2   1   2  -1  -4  -1
#>  ---                                                    
#> 252:  -1  -1  -1   0   1   0   3   4  -2   0   1   0   0
#> 253:  -2   1  -1  -1   3  -2  -1   4   0   1   2   0  -1
#> 254:  -2  -2  -3  -1   0   1  -2   1   1   2  -4   0  -1
#> 255:  -3   0   0  -1  -1   2   0   2   2   0   1   0   0
#> 256:   1   1   2   0  -1   1  -1  -2   2  -1  -3   3   2
#> 
#> Slot "dFrame":
#>      i  j          p v check  p_int_lb  p_int_ub          i_info
#>   1: 0  0 1.00000000 0  TRUE 0.0000000 1.0000000             i=0
#>   2: 0  1 0.00000000 1 FALSE 1.0000000 1.0000000             i=0
#>   3: 0  2 0.00000000 2 FALSE 1.0000000 1.0000000             i=0
#>   4: 0  3 0.00000000 3  TRUE 1.0000000 1.0000000             i=0
#>   5: 0  4 0.00000000 4  TRUE 1.0000000 1.0000000             i=0
#>  ---                                                            
#> 122: 8  9 0.19419346 1  TRUE 0.6144007 0.8085942 i>=8 (symmtery)
#> 123: 8 10 0.11873037 2  TRUE 0.8085942 0.9273246 i>=8 (symmtery)
#> 124: 8 11 0.05229263 3  TRUE 0.9273246 0.9796172 i>=8 (symmtery)
#> 125: 8 12 0.01659093 4  TRUE 0.9796172 0.9962081 i>=8 (symmtery)
#> 126: 8 13 0.00379186 5  TRUE 0.9962081 1.0000000 i>=8 (symmtery)
#> 
#> Slot "empResults":
#>    i p_mean p_var p_sum    p_stay iter
#> 1: 0      0     0     1 1.0000000    0
#> 2: 1      0     3     1 0.0000000    1
#> 3: 2      0     3     1 0.0000000    1
#> 4: 3      0     3     1 0.3936625    1
#> 5: 4      0     3     1 0.2779212    1
#> 6: 5      0     3     1 0.2213183    1
#> 7: 6      0     3     1 0.2166529    1
#> 8: 7      0     3     1 0.2259140    1
#> 9: 8      0     3     1 0.2288015    1
#> 
#> Slot "pParams":
#> An object of class "ptable_params"
#> Slot "D":
#> [1] 5
#> 
#> Slot "V":
#> [1] 3
#> 
#> Slot "js":
#> [1] 2
#> 
#> Slot "ncat":
#> [1] 8
#> 
#> Slot "epsilon":
#> [1] 0.0000001 0.0000001 0.0000001 0.0000001 0.0000001 0.0000001 0.0000001
#> [8] 0.0000001
#> 
#> Slot "pstay":
#> [1] NA NA NA NA NA NA NA NA
#> 
#> Slot "optim":
#> [1] 1 1 1 1 1 1 1 1
#> 
#> Slot "mono":
#> [1] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> 
#> Slot "pTableSize":
#> [1] 70
#> 
#> Slot "label":
#> [1] "D5V300"
#> 
#> 
#> Slot "tStamp":
#> [1] "20190128132433"
#> 
#> Slot "type":
#> [1] "abs"
#> 

# destatis-format of perturbation table
ck_create_pTable(
  D=5,
  V=3,
  js=2,
  pstay=0.5,
  optim=1,
  mono=TRUE,
  typ="destatis"
)

#> An object of class "ptable"
#> Slot "pMatrix":
#>            0 1 2          3          4          5           6            7
#> 0 1.00000000 0 0 0.00000000 0.00000000 0.00000000 0.000000000 0.0000000000
#> 1 0.73446954 0 0 0.13672019 0.07279631 0.03742956 0.018584399 0.0000000000
#> 2 0.40880704 0 0 0.40664356 0.14762399 0.03231882 0.004266872 0.0003397187
#> 3 0.18737598 0 0 0.50000000 0.15629816 0.09081388 0.043242014 0.0168738807
#> 4 0.09146068 0 0 0.15340420 0.50000000 0.10745598 0.072882739 0.0429685201
#> 5 0.03934302 0 0 0.10580374 0.11427070 0.50000000 0.091229785 0.0674379852
#> 6 0.00000000 0 0 0.09063278 0.09217047 0.09217047 0.500000000 0.0712627178
#> 7 0.00000000 0 0 0.03393943 0.05458754 0.07607160 0.091852693 0.5000000000
#> 8 0.00000000 0 0 0.01158491 0.02539548 0.04675981 0.072317176 0.0939426281
#>             8           9         10         11         12         13
#> 0 0.000000000 0.000000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 1 0.000000000 0.000000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 2 0.000000000 0.000000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 3 0.005396092 0.000000000 0.00000000 0.00000000 0.00000000 0.00000000
#> 4 0.022019514 0.009808366 0.00000000 0.00000000 0.00000000 0.00000000
#> 5 0.043932295 0.025221711 0.01276076 0.00000000 0.00000000 0.00000000
#> 6 0.057594091 0.043786679 0.03131516 0.02106764 0.00000000 0.00000000
#> 7 0.087106597 0.068413356 0.04655550 0.02744993 0.01402335 0.00000000
#> 8 0.500000000 0.093942634 0.07231718 0.04675979 0.02539548 0.01158492
#> 
#> Slot "pClasses":
#> [1] 0 1 2 3 4 5 6 7 8
#> 
#> Slot "pTable":
#>     i  j            p  v   p_int_lb   p_int_ub
#>  1: 0  0 1.0000000000  0 0.00000000 1.00000000
#>  2: 1  0 0.7344695411 -1 0.00000000 0.73446954
#>  3: 1  3 0.1367201903  2 0.73446954 0.87118973
#>  4: 1  4 0.0727963132  3 0.87118973 0.94398604
#>  5: 1  5 0.0374295563  4 0.94398604 0.98141560
#>  6: 1  6 0.0185843991  5 0.98141560 1.00000000
#>  7: 2  0 0.4088070415 -2 0.00000000 0.40880704
#>  8: 2  3 0.4066435573  1 0.40880704 0.81545060
#>  9: 2  4 0.1476239875  2 0.81545060 0.96307459
#> 10: 2  5 0.0323188230  3 0.96307459 0.99539341
#> 11: 2  6 0.0042668720  4 0.99539341 0.99966028
#> 12: 2  7 0.0003397187  5 0.99966028 1.00000000
#> 13: 3  0 0.1873759789 -3 0.00000000 0.18737598
#> 14: 3  3 0.5000000000  0 0.18737598 0.68737598
#> 15: 3  4 0.1562981579  1 0.68737598 0.84367414
#> 16: 3  5 0.0908138759  2 0.84367414 0.93448801
#> 17: 3  6 0.0432420142  3 0.93448801 0.97773003
#> 18: 3  7 0.0168738807  4 0.97773003 0.99460391
#> 19: 3  8 0.0053960924  5 0.99460391 1.00000000
#> 20: 4  0 0.0914606752 -4 0.00000000 0.09146068
#> 21: 4  3 0.1534042047 -1 0.09146068 0.24486488
#> 22: 4  4 0.5000000000  0 0.24486488 0.74486488
#> 23: 4  5 0.1074559811  1 0.74486488 0.85232086
#> 24: 4  6 0.0728827391  2 0.85232086 0.92520360
#> 25: 4  7 0.0429685201  3 0.92520360 0.96817212
#> 26: 4  8 0.0220195138  4 0.96817212 0.99019163
#> 27: 4  9 0.0098083662  5 0.99019163 1.00000000
#> 28: 5  0 0.0393430213 -5 0.00000000 0.03934302
#> 29: 5  3 0.1058037441 -2 0.03934302 0.14514677
#> 30: 5  4 0.1142706969 -1 0.14514677 0.25941746
#> 31: 5  5 0.5000000000  0 0.25941746 0.75941746
#> 32: 5  6 0.0912297850  1 0.75941746 0.85064725
#> 33: 5  7 0.0674379852  2 0.85064725 0.91808523
#> 34: 5  8 0.0439322946  3 0.91808523 0.96201753
#> 35: 5  9 0.0252217115  4 0.96201753 0.98723924
#> 36: 5 10 0.0127607613  5 0.98723924 1.00000000
#> 37: 6  3 0.0906327817 -3 0.00000000 0.09063278
#> 38: 6  4 0.0921704682 -2 0.09063278 0.18280325
#> 39: 6  5 0.0921704682 -1 0.18280325 0.27497372
#> 40: 6  6 0.5000000000  0 0.27497372 0.77497372
#> 41: 6  7 0.0712627178  1 0.77497372 0.84623644
#> 42: 6  8 0.0575940906  2 0.84623644 0.90383053
#> 43: 6  9 0.0437866786  3 0.90383053 0.94761721
#> 44: 6 10 0.0313151596  4 0.94761721 0.97893236
#> 45: 6 11 0.0210676353  5 0.97893236 1.00000000
#> 46: 7  3 0.0339394348 -4 0.00000000 0.03393943
#> 47: 7  4 0.0545875396 -3 0.03393943 0.08852697
#> 48: 7  5 0.0760716044 -2 0.08852697 0.16459858
#> 49: 7  6 0.0918526932 -1 0.16459858 0.25645127
#> 50: 7  7 0.5000000000  0 0.25645127 0.75645127
#> 51: 7  8 0.0871065970  1 0.75645127 0.84355787
#> 52: 7  9 0.0684133561  2 0.84355787 0.91197123
#> 53: 7 10 0.0465554964  3 0.91197123 0.95852672
#> 54: 7 11 0.0274499306  4 0.95852672 0.98597665
#> 55: 7 12 0.0140233479  5 0.98597665 1.00000000
#> 56: 8  3 0.0115849067 -5 0.00000000 0.01158491
#> 57: 8  4 0.0253954829 -4 0.01158491 0.03698039
#> 58: 8  5 0.0467598059 -3 0.03698039 0.08374020
#> 59: 8  6 0.0723171756 -2 0.08374020 0.15605737
#> 60: 8  7 0.0939426281 -1 0.15605737 0.25000000
#> 61: 8  8 0.5000000000  0 0.25000000 0.75000000
#> 62: 8  9 0.0939426341  1 0.75000000 0.84394263
#> 63: 8 10 0.0723171799  2 0.84394263 0.91625981
#> 64: 8 11 0.0467597947  3 0.91625981 0.96301961
#> 65: 8 12 0.0253954762  4 0.96301961 0.98841508
#> 66: 8 13 0.0115849159  5 0.98841508 1.00000000
#>     i  j            p  v   p_int_lb   p_int_ub
#> 
#> Slot "dFrame":
#>      i  j          p v check  p_int_lb  p_int_ub          i_info
#>   1: 0  0 1.00000000 0  TRUE 0.0000000 1.0000000             i=0
#>   2: 0  1 0.00000000 1 FALSE 1.0000000 1.0000000             i=0
#>   3: 0  2 0.00000000 2 FALSE 1.0000000 1.0000000             i=0
#>   4: 0  3 0.00000000 3  TRUE 1.0000000 1.0000000             i=0
#>   5: 0  4 0.00000000 4  TRUE 1.0000000 1.0000000             i=0
#>  ---                                                            
#> 122: 8  9 0.09394263 1  TRUE 0.7500000 0.8439426 i>=8 (symmtery)
#> 123: 8 10 0.07231718 2  TRUE 0.8439426 0.9162598 i>=8 (symmtery)
#> 124: 8 11 0.04675979 3  TRUE 0.9162598 0.9630196 i>=8 (symmtery)
#> 125: 8 12 0.02539548 4  TRUE 0.9630196 0.9884151 i>=8 (symmtery)
#> 126: 8 13 0.01158492 5  TRUE 0.9884151 1.0000000 i>=8 (symmtery)
#> 
#> Slot "empResults":
#>    i p_mean p_var p_sum p_stay iter
#> 1: 0      0     0     1    1.0    0
#> 2: 1      0     3     1    0.0    1
#> 3: 2      0     3     1    0.0    1
#> 4: 3      0     3     1    0.5    1
#> 5: 4      0     3     1    0.5    1
#> 6: 5      0     3     1    0.5    1
#> 7: 6      0     3     1    0.5    1
#> 8: 7      0     3     1    0.5    1
#> 9: 8      0     3     1    0.5    1
#> 
#> Slot "pParams":
#> An object of class "ptable_params"
#> Slot "D":
#> [1] 5
#> 
#> Slot "V":
#> [1] 3
#> 
#> Slot "js":
#> [1] 2
#> 
#> Slot "ncat":
#> [1] 8
#> 
#> Slot "epsilon":
#> [1] 0.0000001 0.0000001 0.0000001 0.0000001 0.0000001 0.0000001 0.0000001
#> [8] 0.0000001
#> 
#> Slot "pstay":
#> [1] 0.5 0.5 0.5 0.5 0.5 0.5 0.5 0.5
#> 
#> Slot "optim":
#> [1] 1 1 1 1 1 1 1 1
#> 
#> Slot "mono":
#> [1] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE
#> 
#> Slot "pTableSize":
#> [1] 70
#> 
#> Slot "label":
#> [1] "D5V300"
#> 
#> 
#> Slot "tStamp":
#> [1] "20190128132433"
#> 
#> Slot "type":
#> [1] "destatis"
#>