Set parameters required to perturb numeric variables using a flex function

ck_flexparams() allows to define a flex function that is used to lookup perturbation magnitudes (percentages) used when perturbing continuous variables.

ck_flexparams(fp, p = c(0.25, 0.05), epsilon = 1, q = 3)

Arguments

fp

(numeric scalar); at which point should the noise coefficient function reaches its desired maximum (defined by the first element of p)

p

a numeric vector of length 2 where both elements specify a percentage. The first value refers to the desired maximum perturbation percentage for small cells (depending on fp) while the second element refers to the desired maximum perturbation percentage for large cells. Both values must be between 0 and 1 and need to be in descending order.

epsilon

a numeric vector in descending order with all values >= 0 and <= 1 with the first element forced to equal 1. The length of this vector must correspond with the number top_k specified in ck_params_nums() when creating parameters for type == "top_contr" which is checked at runtime. This setting allows to use different flex-functions for the largest top_k contributors.

q

(numeric scalar); Parameter of the function; q needs to be >= 1

Value

an object suitable as input for ck_params_nums().

Details

details about the flex function can be found in Deliverable D4.2, Part I in SGA "Open Source tools for perturbative confidentiality methods"

See also

Examples

x <- ck_create_testdata()

# create some 0/1 variables that should be perturbed later
x[, cnt_females := ifelse(sex == "male", 0, 1)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              52          25.00000           0
#>    2:   2530      28       1              23          25.00000           1
#>    3:   6920     550       1              70          25.00000           0
#>    4:   7960     870       1              57          25.00000           0
#>    5:   9030      20       2              58          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              52          16.66667           1
#> 4577:   1420     987    1000              46          16.66667           0
#> 4578:   8900     684    1000              66          16.66667           0
#> 4579:   3880     294    1000              70          16.66667           1
#> 4580:   4830     911    1000              80          16.66667           0
x[, cnt_males := ifelse(sex == "male", 1, 0)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              52          25.00000           0
#>    2:   2530      28       1              23          25.00000           1
#>    3:   6920     550       1              70          25.00000           0
#>    4:   7960     870       1              57          25.00000           0
#>    5:   9030      20       2              58          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              52          16.66667           1
#> 4577:   1420     987    1000              46          16.66667           0
#> 4578:   8900     684    1000              66          16.66667           0
#> 4579:   3880     294    1000              70          16.66667           1
#> 4580:   4830     911    1000              80          16.66667           0
#>       cnt_males
#>    1:         1
#>    2:         0
#>    3:         1
#>    4:         1
#>    5:         1
#>   ---          
#> 4576:         0
#> 4577:         1
#> 4578:         1
#> 4579:         0
#> 4580:         1
x[, cnt_highincome := ifelse(income >= 9000, 1, 0)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              52          25.00000           0
#>    2:   2530      28       1              23          25.00000           1
#>    3:   6920     550       1              70          25.00000           0
#>    4:   7960     870       1              57          25.00000           0
#>    5:   9030      20       2              58          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              52          16.66667           1
#> 4577:   1420     987    1000              46          16.66667           0
#> 4578:   8900     684    1000              66          16.66667           0
#> 4579:   3880     294    1000              70          16.66667           1
#> 4580:   4830     911    1000              80          16.66667           0
#>       cnt_males cnt_highincome
#>    1:         1              0
#>    2:         0              0
#>    3:         1              0
#>    4:         1              0
#>    5:         1              1
#>   ---                         
#> 4576:         0              0
#> 4577:         1              0
#> 4578:         1              0
#> 4579:         0              0
#> 4580:         1              0
# a variable with positive and negative contributions
x[, mixed := sample(-10:10, nrow(x), replace = TRUE)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              52          25.00000           0
#>    2:   2530      28       1              23          25.00000           1
#>    3:   6920     550       1              70          25.00000           0
#>    4:   7960     870       1              57          25.00000           0
#>    5:   9030      20       2              58          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              52          16.66667           1
#> 4577:   1420     987    1000              46          16.66667           0
#> 4578:   8900     684    1000              66          16.66667           0
#> 4579:   3880     294    1000              70          16.66667           1
#> 4580:   4830     911    1000              80          16.66667           0
#>       cnt_males cnt_highincome mixed
#>    1:         1              0     4
#>    2:         0              0    -1
#>    3:         1              0     6
#>    4:         1              0     4
#>    5:         1              1     7
#>   ---                               
#> 4576:         0              0     4
#> 4577:         1              0    -3
#> 4578:         1              0    -9
#> 4579:         0              0     7
#> 4580:         1              0    -7

# create record keys
x$rkey <- ck_generate_rkeys(dat = x)

# define required inputs

# hierarchy with some bogus codes
d_sex <- hier_create(root = "Total", nodes = c("male", "female"))
d_sex <- hier_add(d_sex, root = "female", "f")
d_sex <- hier_add(d_sex, root = "male", "m")

d_age <- hier_create(root = "Total", nodes = paste0("age_group", 1:6))
d_age <- hier_add(d_age, root = "age_group1", "ag1a")
d_age <- hier_add(d_age, root = "age_group2", "ag2a")

# define the cell key object
countvars <- c("cnt_females", "cnt_males", "cnt_highincome")
numvars <- c("expend", "income", "savings", "mixed")
tab <- ck_setup(
  x = x,
  rkey = "rkey",
  dims = list(sex = d_sex, age = d_age),
  w = "sampling_weight",
  countvars = countvars,
  numvars = numvars)

# show some information about this table instance
tab$print() # identical with print(tab)
#> ── Table Information ───────────────────────────────────────────────────────────
#> ✔ 45 cells in 2 dimensions ('sex', 'age')
#> ✔ weights: yes
#> ── Tabulated / Perturbed countvars ─────────────────────────────────────────────
#> ☐ 'total'
#> ☐ 'cnt_females'
#> ☐ 'cnt_males'
#> ☐ 'cnt_highincome'
#> ── Tabulated / Perturbed numvars ───────────────────────────────────────────────
#> ☐ 'expend'
#> ☐ 'income'
#> ☐ 'savings'
#> ☐ 'mixed'

# information about the hierarchies
tab$hierarchy_info()
#> $sex
#>      code level is_leaf parent
#> 1:  Total     1   FALSE  Total
#> 2:   male     2   FALSE  Total
#> 3:      m     3    TRUE   male
#> 4: female     2   FALSE  Total
#> 5:      f     3    TRUE female
#> 
#> $age
#>          code level is_leaf     parent
#> 1:      Total     1   FALSE      Total
#> 2: age_group1     2   FALSE      Total
#> 3:       ag1a     3    TRUE age_group1
#> 4: age_group2     2   FALSE      Total
#> 5:       ag2a     3    TRUE age_group2
#> 6: age_group3     2    TRUE      Total
#> 7: age_group4     2    TRUE      Total
#> 8: age_group5     2    TRUE      Total
#> 9: age_group6     2    TRUE      Total
#> 

# which variables have been defined?
tab$allvars()
#> $cntvars
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"
#> 
#> $numvars
#> [1] "expend"  "income"  "savings" "mixed"  
#> 

# count variables
tab$cntvars()
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"

# continuous variables
tab$numvars()
#> [1] "expend"  "income"  "savings" "mixed"  

# create perturbation parameters for "total" variable and
# write to yaml-file

# create a ptable using functionality from the ptable-pkg
f_yaml <- tempfile(fileext = ".yaml")
p_cnts1 <- ck_params_cnts(
  ptab = ptable::pt_ex_cnts(),
  path = f_yaml)
#> yaml configuration '/tmp/RtmpRLPm23/file1e1b53f1467c.yaml' successfully written.

# read parameters from yaml-file and set them for variable `"total"`
p_cnts1 <- ck_read_yaml(path = f_yaml)

tab$params_cnts_set(val = p_cnts1, v = "total")
#> --> setting perturbation parameters for variable 'total'

# create alternative perturbation parameters by specifying parameters
para2 <- ptable::create_num_ptable(
  D = 8, V = 3, js = 2, create = FALSE)
#> Error in ptable::create_num_ptable(D = 8, V = 3, js = 2, create = FALSE): unused argument (js = 2)

p_cnts2 <- ck_params_cnts(ptab = para2)
#> Error in ck_params_cnts(ptab = para2): object 'para2' not found

# use these ptable it for the remaining variables
tab$params_cnts_set(val = p_cnts2, v = countvars)
#> Error in tab$params_cnts_set(val = p_cnts2, v = countvars): object 'p_cnts2' not found

# perturb a variable
tab$perturb(v = "total")
#> Count variable 'total' was perturbed.

# multiple variables can be perturbed as well
tab$perturb(v = c("cnt_males", "cnt_highincome"))
#> Error: No perturbation parameters found for variable 'cnt_males' - Please use method `$params_cnts_set().

# return weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))
#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 277386 4582 277507.1293
#>  2:  Total age_group1     total 1969 119210 1967 119088.9132
#>  3:  Total       ag1a     total 1969 119210 1967 119088.9132
#>  4:  Total age_group2     total 1143  69592 1143  69592.0000
#>  5:  Total       ag2a     total 1143  69592 1143  69592.0000
#>  6:  Total age_group3     total  864  52213  863  52152.5683
#>  7:  Total age_group4     total  423  25750  424  25810.8747
#>  8:  Total age_group5     total  168   9822  168   9822.0000
#>  9:  Total age_group6     total   13    799   14    860.4615
#> 10:   male      Total     total 2296 139424 2296 139424.0000
#> 11:      m      Total     total 2296 139424 2296 139424.0000
#> 12:   male age_group1     total 1015  61800 1016  61860.8867
#> 13:      m age_group1     total 1015  61800 1016  61860.8867
#> 14:   male       ag1a     total 1015  61800 1016  61860.8867
#> 15:      m       ag1a     total 1015  61800 1016  61860.8867
#> 16:   male age_group2     total  571  35145  570  35083.4501
#> 17:      m age_group2     total  571  35145  570  35083.4501
#> 18:   male       ag2a     total  571  35145  570  35083.4501
#> 19:      m       ag2a     total  571  35145  570  35083.4501
#> 20:   male age_group3     total  424  25460  426  25580.0943
#> 21:      m age_group3     total  424  25460  426  25580.0943
#> 22:   male age_group4     total  195  11756  195  11756.0000
#> 23:      m age_group4     total  195  11756  195  11756.0000
#> 24:   male age_group5     total   84   4742   85   4798.4524
#> 25:      m age_group5     total   84   4742   85   4798.4524
#> 26:   male age_group6     total    7    521    8    595.4286
#> 27:      m age_group6     total    7    521    8    595.4286
#> 28: female      Total     total 2284 137962 2284 137962.0000
#> 29:      f      Total     total 2284 137962 2284 137962.0000
#> 30: female age_group1     total  954  57410  953  57349.8218
#> 31:      f age_group1     total  954  57410  953  57349.8218
#> 32: female       ag1a     total  954  57410  953  57349.8218
#> 33:      f       ag1a     total  954  57410  953  57349.8218
#> 34: female age_group2     total  572  34447  571  34386.7780
#> 35:      f age_group2     total  572  34447  571  34386.7780
#> 36: female       ag2a     total  572  34447  571  34386.7780
#> 37:      f       ag2a     total  572  34447  571  34386.7780
#> 38: female age_group3     total  440  26753  439  26692.1977
#> 39:      f age_group3     total  440  26753  439  26692.1977
#> 40: female age_group4     total  228  13994  227  13932.6228
#> 41:      f age_group4     total  228  13994  227  13932.6228
#> 42: female age_group5     total   84   5080   85   5140.4762
#> 43:      f age_group5     total   84   5080   85   5140.4762
#> 44: female age_group6     total    6    278    8    370.6667
#> 45:      f age_group6     total    6    278    8    370.6667
#> 46:  Total      Total cnt_males 2296 139424   NA          NA
#> 47:  Total age_group1 cnt_males 1015  61800   NA          NA
#> 48:  Total       ag1a cnt_males 1015  61800   NA          NA
#> 49:  Total age_group2 cnt_males  571  35145   NA          NA
#> 50:  Total       ag2a cnt_males  571  35145   NA          NA
#> 51:  Total age_group3 cnt_males  424  25460   NA          NA
#> 52:  Total age_group4 cnt_males  195  11756   NA          NA
#> 53:  Total age_group5 cnt_males   84   4742   NA          NA
#> 54:  Total age_group6 cnt_males    7    521   NA          NA
#> 55:   male      Total cnt_males 2296 139424   NA          NA
#> 56:      m      Total cnt_males 2296 139424   NA          NA
#> 57:   male age_group1 cnt_males 1015  61800   NA          NA
#> 58:      m age_group1 cnt_males 1015  61800   NA          NA
#> 59:   male       ag1a cnt_males 1015  61800   NA          NA
#> 60:      m       ag1a cnt_males 1015  61800   NA          NA
#> 61:   male age_group2 cnt_males  571  35145   NA          NA
#> 62:      m age_group2 cnt_males  571  35145   NA          NA
#> 63:   male       ag2a cnt_males  571  35145   NA          NA
#> 64:      m       ag2a cnt_males  571  35145   NA          NA
#> 65:   male age_group3 cnt_males  424  25460   NA          NA
#> 66:      m age_group3 cnt_males  424  25460   NA          NA
#> 67:   male age_group4 cnt_males  195  11756   NA          NA
#> 68:      m age_group4 cnt_males  195  11756   NA          NA
#> 69:   male age_group5 cnt_males   84   4742   NA          NA
#> 70:      m age_group5 cnt_males   84   4742   NA          NA
#> 71:   male age_group6 cnt_males    7    521   NA          NA
#> 72:      m age_group6 cnt_males    7    521   NA          NA
#> 73: female      Total cnt_males    0      0   NA          NA
#> 74:      f      Total cnt_males    0      0   NA          NA
#> 75: female age_group1 cnt_males    0      0   NA          NA
#> 76:      f age_group1 cnt_males    0      0   NA          NA
#> 77: female       ag1a cnt_males    0      0   NA          NA
#> 78:      f       ag1a cnt_males    0      0   NA          NA
#> 79: female age_group2 cnt_males    0      0   NA          NA
#> 80:      f age_group2 cnt_males    0      0   NA          NA
#> 81: female       ag2a cnt_males    0      0   NA          NA
#> 82:      f       ag2a cnt_males    0      0   NA          NA
#> 83: female age_group3 cnt_males    0      0   NA          NA
#> 84:      f age_group3 cnt_males    0      0   NA          NA
#> 85: female age_group4 cnt_males    0      0   NA          NA
#> 86:      f age_group4 cnt_males    0      0   NA          NA
#> 87: female age_group5 cnt_males    0      0   NA          NA
#> 88:      f age_group5 cnt_males    0      0   NA          NA
#> 89: female age_group6 cnt_males    0      0   NA          NA
#> 90:      f age_group6 cnt_males    0      0   NA          NA
#>        sex        age     vname  uwc     wc puwc         pwc

# numerical variables (positive variables using flex-function)
# we also write the config to a yaml file
f_yaml <- tempfile(fileext = ".yaml")

# create a ptable using functionality from the ptable-pkg
# a single ptable for all cells
ptab1 <- ptable::pt_ex_nums(parity = TRUE, separation = FALSE)

# a single ptab for all cells except for very small ones
ptab2 <- ptable::pt_ex_nums(parity = TRUE, separation = TRUE)

# different ptables for cells with even/odd number of contributors
# and very small cells
ptab3 <- ptable::pt_ex_nums(parity = FALSE, separation = TRUE)

p_nums1 <- ck_params_nums(
  ptab = ptab1,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.30, 0.03),
    epsilon = c(1, 0.5, 0.2),
    q = 3),
  mu_c = 2,
  same_key = FALSE,
  use_zero_rkeys = FALSE,
  path = f_yaml)
#> yaml configuration '/tmp/RtmpRLPm23/file1e1b1196e3a9.yaml' successfully written.

# we read the parameters from the yaml-file
p_nums1 <- ck_read_yaml(path = f_yaml)

# for variables with positive and negative values
p_nums2 <- ck_params_nums(
  ptab = ptab2,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.15, 0.02),
    epsilon = c(1, 0.4, 0.15),
    q = 3),
  mu_c = 2,
  same_key = FALSE)

# simple perturbation parameters (not using the flex-function approach)
p_nums3 <- ck_params_nums(
  ptab = ptab3,
  type = "mean",
  mult_params = ck_simpleparams(p = 0.25),
  mu_c = 2,
  same_key = FALSE)

# use `p_nums1` for all variables
tab$params_nums_set(p_nums1, c("savings", "income", "expend"))
#> --> setting perturbation parameters for variable 'savings'
#> --> setting perturbation parameters for variable 'income'
#> --> setting perturbation parameters for variable 'expend'

# use different parameters for variable `mixed`
tab$params_nums_set(p_nums2, v = "mixed")
#> --> setting perturbation parameters for variable 'mixed'

# identify sensitive cells to which extra protection (`mu_c`) is added.
tab$supp_p(v = "income", p = 85)
#> computing contributing indices | rawdata <--> table; this might take a while
#> p%-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_pq(v = "income", p = 85, q = 90)
#> computing contributing indices | rawdata <--> table; this might take a while
#> pq-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_nk(v = "income", n = 2, k = 90)
#> computing contributing indices | rawdata <--> table; this might take a while
#> nk-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_freq(v = "income", n = 14, weighted = FALSE)
#> freq-rule: 5 new sensitive cells (incl. duplicates) found (total: 5)
tab$supp_val(v = "income", n = 10000, weighted = TRUE)
#> val-rule: 0 new sensitive cells (incl. duplicates) found (total: 5)
tab$supp_cells(
  v = "income",
  inp = data.frame(
    sex = c("female", "female"),
    "age" = c("age_group1", "age_group3")
  )
)
#> cell-rule: 2 new sensitive cells (incl. duplicates) found (total: 7)

# perturb variables
tab$perturb(v = c("income", "savings"))
#> Numeric variable 'income' was perturbed.
#> Numeric variable 'savings' was perturbed.

# extract results
tab$numtab("income", mean_before_sum = TRUE)
#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1393165240 1393157536
#>  2:  Total age_group1 income  9810547  596086116  596034836
#>  3:  Total       ag1a income  9810547  596086116  596034836
#>  4:  Total age_group2 income  5692119  345056675  345072409
#>  5:  Total       ag2a income  5692119  345056675  345072409
#>  6:  Total age_group3 income  4406946  268115875  268070960
#>  7:  Total age_group4 income  2133543  129537902  129628206
#>  8:  Total age_group5 income   848151   50241001   50241001
#>  9:  Total age_group6 income    61672    4127671    4224800
#> 10:   male      Total income 11262049  684391371  684361846
#> 11:      m      Total income 11262049  684391371  684361846
#> 12:   male age_group1 income  4877164  296210946  296195018
#> 13:      m age_group1 income  4877164  296210946  296195018
#> 14:   male       ag1a income  4877164  296210946  296195018
#> 15:      m       ag1a income  4877164  296210946  296195018
#> 16:   male age_group2 income  2811379  171994169  172188759
#> 17:      m age_group2 income  2811379  171994169  172188759
#> 18:   male       ag2a income  2811379  171994169  172188759
#> 19:      m       ag2a income  2811379  171994169  172188759
#> 20:   male age_group3 income  2168169  130231491  130210698
#> 21:      m age_group3 income  2168169  130231491  130210698
#> 22:   male age_group4 income   978510   59566415   59504058
#> 23:      m age_group4 income   978510   59566415   59504058
#> 24:   male age_group5 income   393134   23735837   23686855
#> 25:      m age_group5 income   393134   23735837   23686855
#> 26:   male age_group6 income    33693    2652513    2637237
#> 27:      m age_group6 income    33693    2652513    2637237
#> 28: female      Total income 11690929  708773869  708559658
#> 29:      f      Total income 11690929  708773869  708559658
#> 30: female age_group1 income  4933383  299875170  299693375
#> 31:      f age_group1 income  4933383  299875170  299693375
#> 32: female       ag1a income  4933383  299875170  299693375
#> 33:      f       ag1a income  4933383  299875170  299693375
#> 34: female age_group2 income  2880740  173062506  173023614
#> 35:      f age_group2 income  2880740  173062506  173023614
#> 36: female       ag2a income  2880740  173062506  173023614
#> 37:      f       ag2a income  2880740  173062506  173023614
#> 38: female age_group3 income  2238777  137884384  137828496
#> 39:      f age_group3 income  2238777  137884384  137828496
#> 40: female age_group4 income  1155033   69971487   69907514
#> 41:      f age_group4 income  1155033   69971487   69907514
#> 42: female age_group5 income   455017   26505164   26527299
#> 43:      f age_group5 income   455017   26505164   26527299
#> 44: female age_group6 income    27979    1475158    1455416
#> 45:      f age_group6 income    27979    1475158    1455416
#>        sex        age  vname      uws         ws        pws
tab$numtab("income", mean_before_sum = FALSE)
#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1393165240 1393161388
#>  2:  Total age_group1 income  9810547  596086116  596060475
#>  3:  Total       ag1a income  9810547  596086116  596060475
#>  4:  Total age_group2 income  5692119  345056675  345064542
#>  5:  Total       ag2a income  5692119  345056675  345064542
#>  6:  Total age_group3 income  4406946  268115875  268093417
#>  7:  Total age_group4 income  2133543  129537902  129583046
#>  8:  Total age_group5 income   848151   50241001   50241001
#>  9:  Total age_group6 income    61672    4127671    4175953
#> 10:   male      Total income 11262049  684391371  684376609
#> 11:      m      Total income 11262049  684391371  684376609
#> 12:   male age_group1 income  4877164  296210946  296202982
#> 13:      m age_group1 income  4877164  296210946  296202982
#> 14:   male       ag1a income  4877164  296210946  296202982
#> 15:      m       ag1a income  4877164  296210946  296202982
#> 16:   male age_group2 income  2811379  171994169  172091436
#> 17:      m age_group2 income  2811379  171994169  172091436
#> 18:   male       ag2a income  2811379  171994169  172091436
#> 19:      m       ag2a income  2811379  171994169  172091436
#> 20:   male age_group3 income  2168169  130231491  130221094
#> 21:      m age_group3 income  2168169  130231491  130221094
#> 22:   male age_group4 income   978510   59566415   59535229
#> 23:      m age_group4 income   978510   59566415   59535229
#> 24:   male age_group5 income   393134   23735837   23711333
#> 25:      m age_group5 income   393134   23735837   23711333
#> 26:   male age_group6 income    33693    2652513    2644864
#> 27:      m age_group6 income    33693    2652513    2644864
#> 28: female      Total income 11690929  708773869  708666755
#> 29:      f      Total income 11690929  708773869  708666755
#> 30: female age_group1 income  4933383  299875170  299784259
#> 31:      f age_group1 income  4933383  299875170  299784259
#> 32: female       ag1a income  4933383  299875170  299784259
#> 33:      f       ag1a income  4933383  299875170  299784259
#> 34: female age_group2 income  2880740  173062506  173043059
#> 35:      f age_group2 income  2880740  173062506  173043059
#> 36: female       ag2a income  2880740  173062506  173043059
#> 37:      f       ag2a income  2880740  173062506  173043059
#> 38: female age_group3 income  2238777  137884384  137856437
#> 39:      f age_group3 income  2238777  137884384  137856437
#> 40: female age_group4 income  1155033   69971487   69939493
#> 41:      f age_group4 income  1155033   69971487   69939493
#> 42: female age_group5 income   455017   26505164   26516229
#> 43:      f age_group5 income   455017   26505164   26516229
#> 44: female age_group6 income    27979    1475158    1465254
#> 45:      f age_group6 income    27979    1475158    1465254
#>        sex        age  vname      uws         ws        pws
tab$numtab("savings")
#>        sex        age   vname     uws        ws         pws
#>  1:  Total      Total savings 2273532 137752665 137746416.6
#>  2:  Total age_group1 savings  982386  59310054  59312358.6
#>  3:  Total       ag1a savings  982386  59310054  59312358.6
#>  4:  Total age_group2 savings  552336  33863857  33864496.8
#>  5:  Total       ag2a savings  552336  33863857  33864496.8
#>  6:  Total age_group3 savings  437101  26443930  26446187.9
#>  7:  Total age_group4 savings  214661  12923313  12925982.8
#>  8:  Total age_group5 savings   80451   4788185   4788185.0
#>  9:  Total age_group6 savings    6597    423326    424163.8
#> 10:   male      Total savings 1159816  70575537  70576780.8
#> 11:      m      Total savings 1159816  70575537  70576780.8
#> 12:   male age_group1 savings  517660  31333625  31332409.6
#> 13:      m age_group1 savings  517660  31333625  31332409.6
#> 14:   male       ag1a savings  517660  31333625  31332409.6
#> 15:      m       ag1a savings  517660  31333625  31332409.6
#> 16:   male age_group2 savings  280923  17471194  17480364.7
#> 17:      m age_group2 savings  280923  17471194  17480364.7
#> 18:   male       ag2a savings  280923  17471194  17480364.7
#> 19:      m       ag2a savings  280923  17471194  17480364.7
#> 20:   male age_group3 savings  214970  12952414  12956412.6
#> 21:      m age_group3 savings  214970  12952414  12956412.6
#> 22:   male age_group4 savings   99420   6028643   6026777.3
#> 23:      m age_group4 savings   99420   6028643   6026777.3
#> 24:   male age_group5 savings   43233   2522407   2517301.8
#> 25:      m age_group5 savings   43233   2522407   2517301.8
#> 26:   male age_group6 savings    3610    267254    266693.2
#> 27:      m age_group6 savings    3610    267254    266693.2
#> 28: female      Total savings 1113716  67177128  67178750.8
#> 29:      f      Total savings 1113716  67177128  67178750.8
#> 30: female age_group1 savings  464726  27976429  27978635.6
#> 31:      f age_group1 savings  464726  27976429  27978635.6
#> 32: female       ag1a savings  464726  27976429  27978635.6
#> 33:      f       ag1a savings  464726  27976429  27978635.6
#> 34: female age_group2 savings  271413  16392663  16392253.6
#> 35:      f age_group2 savings  271413  16392663  16392253.6
#> 36: female       ag2a savings  271413  16392663  16392253.6
#> 37:      f       ag2a savings  271413  16392663  16392253.6
#> 38: female age_group3 savings  222131  13491516  13490482.7
#> 39:      f age_group3 savings  222131  13491516  13490482.7
#> 40: female age_group4 savings  115241   6894670   6892106.7
#> 41:      f age_group4 savings  115241   6894670   6892106.7
#> 42: female age_group5 savings   37218   2265778   2268679.9
#> 43:      f age_group5 savings   37218   2265778   2268679.9
#> 44: female age_group6 savings    2987    156072    156438.1
#> 45:      f age_group6 savings    2987    156072    156438.1
#>        sex        age   vname     uws        ws         pws

# results can be resetted, too
tab$reset_cntvars(v = "cnt_males")
#> Error: Invalid variables specified in `v`: Possible choices are:
#>  'total'

# we can then set other parameters and perturb again
tab$params_cnts_set(val = p_cnts1, v = "cnt_males")
#> --> setting perturbation parameters for variable 'cnt_males'
tab$perturb(v = "cnt_males")
#> Count variable 'cnt_males' was perturbed.

# write to a file "outtab.csv" (.csv is automatically added to the path)
if (FALSE) {
tab$freqtab(v = c("total", "cnt_males"), path = "outtab.csv")
}

# show results containing weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))
#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 277386 4582 277507.1293
#>  2:  Total age_group1     total 1969 119210 1967 119088.9132
#>  3:  Total       ag1a     total 1969 119210 1967 119088.9132
#>  4:  Total age_group2     total 1143  69592 1143  69592.0000
#>  5:  Total       ag2a     total 1143  69592 1143  69592.0000
#>  6:  Total age_group3     total  864  52213  863  52152.5683
#>  7:  Total age_group4     total  423  25750  424  25810.8747
#>  8:  Total age_group5     total  168   9822  168   9822.0000
#>  9:  Total age_group6     total   13    799   14    860.4615
#> 10:   male      Total     total 2296 139424 2296 139424.0000
#> 11:      m      Total     total 2296 139424 2296 139424.0000
#> 12:   male age_group1     total 1015  61800 1016  61860.8867
#> 13:      m age_group1     total 1015  61800 1016  61860.8867
#> 14:   male       ag1a     total 1015  61800 1016  61860.8867
#> 15:      m       ag1a     total 1015  61800 1016  61860.8867
#> 16:   male age_group2     total  571  35145  570  35083.4501
#> 17:      m age_group2     total  571  35145  570  35083.4501
#> 18:   male       ag2a     total  571  35145  570  35083.4501
#> 19:      m       ag2a     total  571  35145  570  35083.4501
#> 20:   male age_group3     total  424  25460  426  25580.0943
#> 21:      m age_group3     total  424  25460  426  25580.0943
#> 22:   male age_group4     total  195  11756  195  11756.0000
#> 23:      m age_group4     total  195  11756  195  11756.0000
#> 24:   male age_group5     total   84   4742   85   4798.4524
#> 25:      m age_group5     total   84   4742   85   4798.4524
#> 26:   male age_group6     total    7    521    8    595.4286
#> 27:      m age_group6     total    7    521    8    595.4286
#> 28: female      Total     total 2284 137962 2284 137962.0000
#> 29:      f      Total     total 2284 137962 2284 137962.0000
#> 30: female age_group1     total  954  57410  953  57349.8218
#> 31:      f age_group1     total  954  57410  953  57349.8218
#> 32: female       ag1a     total  954  57410  953  57349.8218
#> 33:      f       ag1a     total  954  57410  953  57349.8218
#> 34: female age_group2     total  572  34447  571  34386.7780
#> 35:      f age_group2     total  572  34447  571  34386.7780
#> 36: female       ag2a     total  572  34447  571  34386.7780
#> 37:      f       ag2a     total  572  34447  571  34386.7780
#> 38: female age_group3     total  440  26753  439  26692.1977
#> 39:      f age_group3     total  440  26753  439  26692.1977
#> 40: female age_group4     total  228  13994  227  13932.6228
#> 41:      f age_group4     total  228  13994  227  13932.6228
#> 42: female age_group5     total   84   5080   85   5140.4762
#> 43:      f age_group5     total   84   5080   85   5140.4762
#> 44: female age_group6     total    6    278    8    370.6667
#> 45:      f age_group6     total    6    278    8    370.6667
#> 46:  Total      Total cnt_males 2296 139424 2296 139424.0000
#> 47:  Total age_group1 cnt_males 1015  61800 1016  61860.8867
#> 48:  Total       ag1a cnt_males 1015  61800 1016  61860.8867
#> 49:  Total age_group2 cnt_males  571  35145  570  35083.4501
#> 50:  Total       ag2a cnt_males  571  35145  570  35083.4501
#> 51:  Total age_group3 cnt_males  424  25460  426  25580.0943
#> 52:  Total age_group4 cnt_males  195  11756  195  11756.0000
#> 53:  Total age_group5 cnt_males   84   4742   85   4798.4524
#> 54:  Total age_group6 cnt_males    7    521    8    595.4286
#> 55:   male      Total cnt_males 2296 139424 2296 139424.0000
#> 56:      m      Total cnt_males 2296 139424 2296 139424.0000
#> 57:   male age_group1 cnt_males 1015  61800 1016  61860.8867
#> 58:      m age_group1 cnt_males 1015  61800 1016  61860.8867
#> 59:   male       ag1a cnt_males 1015  61800 1016  61860.8867
#> 60:      m       ag1a cnt_males 1015  61800 1016  61860.8867
#> 61:   male age_group2 cnt_males  571  35145  570  35083.4501
#> 62:      m age_group2 cnt_males  571  35145  570  35083.4501
#> 63:   male       ag2a cnt_males  571  35145  570  35083.4501
#> 64:      m       ag2a cnt_males  571  35145  570  35083.4501
#> 65:   male age_group3 cnt_males  424  25460  426  25580.0943
#> 66:      m age_group3 cnt_males  424  25460  426  25580.0943
#> 67:   male age_group4 cnt_males  195  11756  195  11756.0000
#> 68:      m age_group4 cnt_males  195  11756  195  11756.0000
#> 69:   male age_group5 cnt_males   84   4742   85   4798.4524
#> 70:      m age_group5 cnt_males   84   4742   85   4798.4524
#> 71:   male age_group6 cnt_males    7    521    8    595.4286
#> 72:      m age_group6 cnt_males    7    521    8    595.4286
#> 73: female      Total cnt_males    0      0    0      0.0000
#> 74:      f      Total cnt_males    0      0    0      0.0000
#> 75: female age_group1 cnt_males    0      0    0      0.0000
#> 76:      f age_group1 cnt_males    0      0    0      0.0000
#> 77: female       ag1a cnt_males    0      0    0      0.0000
#> 78:      f       ag1a cnt_males    0      0    0      0.0000
#> 79: female age_group2 cnt_males    0      0    0      0.0000
#> 80:      f age_group2 cnt_males    0      0    0      0.0000
#> 81: female       ag2a cnt_males    0      0    0      0.0000
#> 82:      f       ag2a cnt_males    0      0    0      0.0000
#> 83: female age_group3 cnt_males    0      0    0      0.0000
#> 84:      f age_group3 cnt_males    0      0    0      0.0000
#> 85: female age_group4 cnt_males    0      0    0      0.0000
#> 86:      f age_group4 cnt_males    0      0    0      0.0000
#> 87: female age_group5 cnt_males    0      0    0      0.0000
#> 88:      f age_group5 cnt_males    0      0    0      0.0000
#> 89: female age_group6 cnt_males    0      0    0      0.0000
#> 90:      f age_group6 cnt_males    0      0    0      0.0000
#>        sex        age     vname  uwc     wc puwc         pwc

# utility measures for a count variable
tab$measures_cnts(v = "total", exclude_zeros = TRUE)
#> $overview
#>    noise cnt        pct
#> 1:    -2   5 0.11111111
#> 2:    -1  12 0.26666667
#> 3:     0   9 0.20000000
#> 4:     1  17 0.37777778
#> 5:     2   2 0.04444444
#> 
#> $measures
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.800 0.000 0.012
#>  4:    Q30 1.000 0.001 0.016
#>  5:    Q40 1.000 0.001 0.017
#>  6:   Mean 0.956 0.025 0.046
#>  7: Median 1.000 0.002 0.021
#>  8:    Q60 1.000 0.002 0.022
#>  9:    Q70 1.000 0.002 0.024
#> 10:    Q80 1.000 0.006 0.050
#> 11:    Q90 2.000 0.051 0.103
#> 12:    Q95 2.000 0.143 0.183
#> 13:    Q99 2.000 0.333 0.379
#> 14:    Max 2.000 0.333 0.379
#> 
#> $cumdistr_d1
#>    cat cnt       pct
#> 1:   0   9 0.2000000
#> 2:   1  38 0.8444444
#> 3:   2  45 1.0000000
#> 
#> $cumdistr_d2
#>            cat cnt       pct
#> 1:    [0,0.02]  40 0.8888889
#> 2: (0.02,0.05]  40 0.8888889
#> 3:  (0.05,0.1]  41 0.9111111
#> 4:   (0.1,0.2]  43 0.9555556
#> 5:   (0.2,0.3]  43 0.9555556
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $cumdistr_d3
#>            cat cnt       pct
#> 1:    [0,0.02]  19 0.4222222
#> 2: (0.02,0.05]  36 0.8000000
#> 3:  (0.05,0.1]  40 0.8888889
#> 4:   (0.1,0.2]  43 0.9555556
#> 5:   (0.2,0.3]  43 0.9555556
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $false_zero
#> [1] 0
#> 
#> $false_nonzero
#> [1] 0
#> 
#> $exclude_zeros
#> [1] TRUE
#> 

# modifications for perturbed count variables
tab$mod_cnts()
#>        sex        age row_nr pert       ckey  countvar
#>  1:  Total      Total     17    2 0.96712471     total
#>  2:  Total age_group1     13   -2 0.01953388     total
#>  3:  Total       ag1a     13   -2 0.01953388     total
#>  4:  Total age_group2     15    0 0.46621700     total
#>  5:  Total       ag2a     15    0 0.46621700     total
#>  6:  Total age_group3     14   -1 0.18277532     total
#>  7:  Total age_group4     16    1 0.77764823     total
#>  8:  Total age_group5     15    0 0.65454175     total
#>  9:  Total age_group6     16    1 0.86640853     total
#> 10:   male      Total     15    0 0.36564886     total
#> 11:      m      Total     15    0 0.36564886     total
#> 12:   male age_group1     16    1 0.81705599     total
#> 13:      m age_group1     16    1 0.81705599     total
#> 14:   male       ag1a     16    1 0.81705599     total
#> 15:      m       ag1a     16    1 0.81705599     total
#> 16:   male age_group2     14   -1 0.29477679     total
#> 17:      m age_group2     14   -1 0.29477679     total
#> 18:   male       ag2a     14   -1 0.29477679     total
#> 19:      m       ag2a     14   -1 0.29477679     total
#> 20:   male age_group3     17    2 0.94980535     total
#> 21:      m age_group3     17    2 0.94980535     total
#> 22:   male age_group4     15    0 0.57022205     total
#> 23:      m age_group4     15    0 0.57022205     total
#> 24:   male age_group5     16    1 0.80597197     total
#> 25:      m age_group5     16    1 0.80597197     total
#> 26:   male age_group6     16    1 0.92781671     total
#> 27:      m age_group6     16    1 0.92781671     total
#> 28: female      Total     15    0 0.60147585     total
#> 29:      f      Total     15    0 0.60147585     total
#> 30: female age_group1     14   -1 0.20247789     total
#> 31:      f age_group1     14   -1 0.20247789     total
#> 32: female       ag1a     14   -1 0.20247789     total
#> 33:      f       ag1a     14   -1 0.20247789     total
#> 34: female age_group2     14   -1 0.17144021     total
#> 35:      f age_group2     14   -1 0.17144021     total
#> 36: female       ag2a     14   -1 0.17144021     total
#> 37:      f       ag2a     14   -1 0.17144021     total
#> 38: female age_group3     14   -1 0.23296997     total
#> 39:      f age_group3     14   -1 0.23296997     total
#> 40: female age_group4     14   -1 0.20742618     total
#> 41:      f age_group4     14   -1 0.20742618     total
#> 42: female age_group5     16    1 0.84856978     total
#> 43:      f age_group5     16    1 0.84856978     total
#> 44: female age_group6     17    2 0.93859182     total
#> 45:      f age_group6     17    2 0.93859182     total
#> 46:  Total      Total     15    0 0.36564886 cnt_males
#> 47:  Total age_group1     16    1 0.81705599 cnt_males
#> 48:  Total       ag1a     16    1 0.81705599 cnt_males
#> 49:  Total age_group2     14   -1 0.29477679 cnt_males
#> 50:  Total       ag2a     14   -1 0.29477679 cnt_males
#> 51:  Total age_group3     17    2 0.94980535 cnt_males
#> 52:  Total age_group4     15    0 0.57022205 cnt_males
#> 53:  Total age_group5     16    1 0.80597197 cnt_males
#> 54:  Total age_group6     16    1 0.92781671 cnt_males
#> 55:   male      Total     15    0 0.36564886 cnt_males
#> 56:      m      Total     15    0 0.36564886 cnt_males
#> 57:   male age_group1     16    1 0.81705599 cnt_males
#> 58:      m age_group1     16    1 0.81705599 cnt_males
#> 59:   male       ag1a     16    1 0.81705599 cnt_males
#> 60:      m       ag1a     16    1 0.81705599 cnt_males
#> 61:   male age_group2     14   -1 0.29477679 cnt_males
#> 62:      m age_group2     14   -1 0.29477679 cnt_males
#> 63:   male       ag2a     14   -1 0.29477679 cnt_males
#> 64:      m       ag2a     14   -1 0.29477679 cnt_males
#> 65:   male age_group3     17    2 0.94980535 cnt_males
#> 66:      m age_group3     17    2 0.94980535 cnt_males
#> 67:   male age_group4     15    0 0.57022205 cnt_males
#> 68:      m age_group4     15    0 0.57022205 cnt_males
#> 69:   male age_group5     16    1 0.80597197 cnt_males
#> 70:      m age_group5     16    1 0.80597197 cnt_males
#> 71:   male age_group6     16    1 0.92781671 cnt_males
#> 72:      m age_group6     16    1 0.92781671 cnt_males
#> 73: female      Total     -1    0 0.00000000 cnt_males
#> 74:      f      Total     -1    0 0.00000000 cnt_males
#> 75: female age_group1     -1    0 0.00000000 cnt_males
#> 76:      f age_group1     -1    0 0.00000000 cnt_males
#> 77: female       ag1a     -1    0 0.00000000 cnt_males
#> 78:      f       ag1a     -1    0 0.00000000 cnt_males
#> 79: female age_group2     -1    0 0.00000000 cnt_males
#> 80:      f age_group2     -1    0 0.00000000 cnt_males
#> 81: female       ag2a     -1    0 0.00000000 cnt_males
#> 82:      f       ag2a     -1    0 0.00000000 cnt_males
#> 83: female age_group3     -1    0 0.00000000 cnt_males
#> 84:      f age_group3     -1    0 0.00000000 cnt_males
#> 85: female age_group4     -1    0 0.00000000 cnt_males
#> 86:      f age_group4     -1    0 0.00000000 cnt_males
#> 87: female age_group5     -1    0 0.00000000 cnt_males
#> 88:      f age_group5     -1    0 0.00000000 cnt_males
#> 89: female age_group6     -1    0 0.00000000 cnt_males
#> 90:      f age_group6     -1    0 0.00000000 cnt_males
#>        sex        age row_nr pert       ckey  countvar

# display a summary about utility measures
tab$summary()
#> ┌──────────────────────────────────────────────┐
#> │Utility measures for perturbed count variables│
#> └──────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#>     countvar Min Q10 Q20 Q30 Q40  Mean Median Q60 Q70 Q80 Q90 Q95 Q99 Max
#> 1:     total  -2  -1  -1  -1  -1 0.022      0   0   1   1 1.6 2.0   2   2
#> 2: cnt_males  -1  -1   0   0   0 0.267      0   0   1   1 1.0 1.8   2   2
#> 
#> ── Distance-based measures ─────────────────────────────────────────────────────
#> ✔ Variable: 'total'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.800 0.000 0.012
#>  4:    Q30 1.000 0.001 0.016
#>  5:    Q40 1.000 0.001 0.017
#>  6:   Mean 0.956 0.025 0.046
#>  7: Median 1.000 0.002 0.021
#>  8:    Q60 1.000 0.002 0.022
#>  9:    Q70 1.000 0.002 0.024
#> 10:    Q80 1.000 0.006 0.050
#> 11:    Q90 2.000 0.051 0.103
#> 12:    Q95 2.000 0.143 0.183
#> 13:    Q99 2.000 0.333 0.379
#> 14:    Max 2.000 0.333 0.379
#> 
#> ✔ Variable: 'cnt_males'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.200 0.000 0.003
#>  4:    Q30 1.000 0.001 0.016
#>  5:    Q40 1.000 0.001 0.016
#>  6:   Mean 0.889 0.018 0.040
#>  7: Median 1.000 0.002 0.021
#>  8:    Q60 1.000 0.002 0.021
#>  9:    Q70 1.000 0.005 0.049
#> 10:    Q80 1.000 0.010 0.053
#> 11:    Q90 1.400 0.064 0.106
#> 12:    Q95 2.000 0.143 0.183
#> 13:    Q99 2.000 0.143 0.183
#> 14:    Max 2.000 0.143 0.183
#> 
#> ┌──────────────────────────────────────────────────┐
#> │Utility measures for perturbed numerical variables│
#> └──────────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#>      vname         Min       Q10        Q20        Q30        Q40       Mean
#> 1:  expend         Inf        NA         NA         NA         NA        NaN
#> 2:  income -107113.514 -90911.08 -31186.420 -25413.231 -19447.037 -12476.308
#> 3: savings   -6248.398  -2284.27  -1215.388   -560.833   -409.409    942.801
#> 4:   mixed         Inf        NA         NA         NA         NA        NaN
#>        Median       Q60       Q70      Q80       Q90       Q95       Q99
#> 1:         NA        NA        NA       NA        NA        NA        NA
#> 2: -14762.443 -9128.155 -7712.173 7866.883 47026.824 97267.454 97267.454
#> 3:    639.845  1395.405  2206.641 2377.617  3998.609  9170.732  9170.732
#> 4:         NA        NA        NA       NA        NA        NA        NA
#>          Max
#> 1:      -Inf
#> 2: 97267.454
#> 3:  9170.732
#> 4:      -Inf