ck_generate_rkeys

creates random record keys by sampling between 1:max_val

ck_generate_rkeys(dat, max_val = nrow(dat), max_digits = 10,
  type = "abs", seed = NULL, verbose = TRUE)

Arguments

dat
microdata used to generated hash for random seed

max_val
maximum number for record-key

max_digits
maximum number of digits in the record keys. This parameter is only used if type equals abs. This parameter is only used if type equals destatis.

type
(character) choice how to compute record keys. Valid choices are abs and destatis.

seed
if not NULL a number specifying the initial seed value for the random number generator

verbose
(logical) if TRUE, print additional information

Value

a numeric vector with N record keys given a seed depending on dat

Examples

dat <- ck_create_testdata()
dat$rkeys <- ck_generate_rkeys(
  dat = dat,
  max_val = 2 * nrow(dat),
  type = "abs"
)

# rkeys for destatis method
# NOT RUN {
ck_generate_rkeys(
  dat = dat,
  max_digits = 6,
  type = "destatis"
)
# }