Create and modify the structure of hierarchies

Functions ck_create_nodes(), ck_add_nodes() and ck_delete_nodes() allow to define and modify hierarchical structures represented as trees. These objects can be used in perturbTable to define the (hierarchical) structure of tables.

ck_create_node(total_lab = "Total", node_labs = NULL)

ck_add_nodes(node, node_labs, reference_node)

ck_delete_nodes(node, node_labs, reference_node = NULL)

ck_rename_nodes(node, node_labs, node_labs_new)

Arguments

total_lab
the name of the overall total (summation over all contributing)

node_labs
character name(s) of new elements that should be inserted to or deleted from a hierarchical structure

node
an object as created in ck_create_node() or returned from ck_add_nodes() or ck_delete_nodes().

reference_node
character name of an existing node in the hierarchical structure. When using ck_add_nodes(), the new elements are created as children of the reference node. In ck_delete_nodes(), all children of the reference node that match the names with argument node_labs are deleted from the hierarchy.

node_labs_new
a character vector of new node names replacing the existing node names given in argument node_labs when using ck_rename_nodes.

Value

a Node object that can be used to specify a hierarchy used to define table inputs in perturbTable.

Examples

dim <- ck_create_node(total_lab="Total")
dim <- ck_add_nodes(dim, reference_node="Total", node_labs=LETTERS[1:4])
print(dim)

#>     root  leaf
#> 1: Total Total
#> 2: Total     A
#> 3: Total     B
#> 4: Total     C
#> 5: Total     D

## add some levels below "A" and "C"
dim <- ck_add_nodes(dim, reference_node="A", node_labs=paste0("a", 1:5))
dim <- ck_add_nodes(dim, reference_node="C", node_labs=paste0("c", 1:5))
print(dim)

#>      root  leaf
#>  1: Total Total
#>  2: Total     A
#>  3: Total     B
#>  4: Total     C
#>  5: Total     D
#>  6:     A    a1
#>  7:     A    a2
#>  8:     A    a3
#>  9:     A    a4
#> 10:     A    a5
#> 11:     C    c1
#> 12:     C    c2
#> 13:     C    c3
#> 14:     C    c4
#> 15:     C    c5

## delete some specific levels
dim <- ck_delete_nodes(dim, reference_node="A", node_labs=c("a1", "a4"))
print(dim)

#>      root  leaf
#>  1: Total Total
#>  2: Total     A
#>  3: Total     B
#>  4: Total     C
#>  5: Total     D
#>  6:     A    a2
#>  7:     A    a3
#>  8:     A    a5
#>  9:     C    c1
#> 10:     C    c2
#> 11:     C    c3
#> 12:     C    c4
#> 13:     C    c5

## delete entire subtree
dim <- ck_delete_nodes(dim, reference_node="Total", node_labs=c("C"))
print(dim)

#>     root  leaf
#> 1: Total Total
#> 2: Total     A
#> 3: Total     B
#> 4: Total     D
#> 5:     A    a2
#> 6:     A    a3
#> 7:     A    a5

# plot(dim)