Create perturbation parameters for count variables

This function allows to generate required perturbation parameters that are used to perturb count variables.

ck_params_cnts(ptab, path = NULL)

Arguments

ptab

an object created with ptable::create_ptable(), or ptable::create_cnt_ptable()

path

a scalar character specifying a path to which the parameters created with this functions should be written to (in yaml format)

Value

an object suitable as input to method $params_cnts_set() for the perturbation of counts and frequencies.

See also

Examples

# \donttest{
x <- ck_create_testdata()

# create some 0/1 variables that should be perturbed later
x[, cnt_females := ifelse(sex == "male", 0, 1)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              96          25.00000           0
#>    2:   2530      28       1              88          25.00000           1
#>    3:   6920     550       1              63          25.00000           0
#>    4:   7960     870       1              54          25.00000           0
#>    5:   9030      20       2              93          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000             100          16.66667           1
#> 4577:   1420     987    1000              50          16.66667           0
#> 4578:   8900     684    1000              27          16.66667           0
#> 4579:   3880     294    1000              45          16.66667           1
#> 4580:   4830     911    1000              25          16.66667           0
x[, cnt_males := ifelse(sex == "male", 1, 0)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              96          25.00000           0
#>    2:   2530      28       1              88          25.00000           1
#>    3:   6920     550       1              63          25.00000           0
#>    4:   7960     870       1              54          25.00000           0
#>    5:   9030      20       2              93          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000             100          16.66667           1
#> 4577:   1420     987    1000              50          16.66667           0
#> 4578:   8900     684    1000              27          16.66667           0
#> 4579:   3880     294    1000              45          16.66667           1
#> 4580:   4830     911    1000              25          16.66667           0
#>       cnt_males
#>    1:         1
#>    2:         0
#>    3:         1
#>    4:         1
#>    5:         1
#>   ---          
#> 4576:         0
#> 4577:         1
#> 4578:         1
#> 4579:         0
#> 4580:         1
x[, cnt_highincome := ifelse(income >= 9000, 1, 0)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              96          25.00000           0
#>    2:   2530      28       1              88          25.00000           1
#>    3:   6920     550       1              63          25.00000           0
#>    4:   7960     870       1              54          25.00000           0
#>    5:   9030      20       2              93          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000             100          16.66667           1
#> 4577:   1420     987    1000              50          16.66667           0
#> 4578:   8900     684    1000              27          16.66667           0
#> 4579:   3880     294    1000              45          16.66667           1
#> 4580:   4830     911    1000              25          16.66667           0
#>       cnt_males cnt_highincome
#>    1:         1              0
#>    2:         0              0
#>    3:         1              0
#>    4:         1              0
#>    5:         1              1
#>   ---                         
#> 4576:         0              0
#> 4577:         1              0
#> 4578:         1              0
#> 4579:         0              0
#> 4580:         1              0
# a variable with positive and negative contributions
x[, mixed := sample(-10:10, nrow(x), replace = TRUE)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              96          25.00000           0
#>    2:   2530      28       1              88          25.00000           1
#>    3:   6920     550       1              63          25.00000           0
#>    4:   7960     870       1              54          25.00000           0
#>    5:   9030      20       2              93          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000             100          16.66667           1
#> 4577:   1420     987    1000              50          16.66667           0
#> 4578:   8900     684    1000              27          16.66667           0
#> 4579:   3880     294    1000              45          16.66667           1
#> 4580:   4830     911    1000              25          16.66667           0
#>       cnt_males cnt_highincome mixed
#>    1:         1              0     7
#>    2:         0              0    -1
#>    3:         1              0     5
#>    4:         1              0   -10
#>    5:         1              1    -3
#>   ---                               
#> 4576:         0              0     8
#> 4577:         1              0     1
#> 4578:         1              0     8
#> 4579:         0              0    -3
#> 4580:         1              0    -9

# create record keys
x$rkey <- ck_generate_rkeys(dat = x)

# define required inputs

# hierarchy with some bogus codes
d_sex <- hier_create(root = "Total", nodes = c("male", "female"))
d_sex <- hier_add(d_sex, root = "female", "f")
d_sex <- hier_add(d_sex, root = "male", "m")

d_age <- hier_create(root = "Total", nodes = paste0("age_group", 1:6))
d_age <- hier_add(d_age, root = "age_group1", "ag1a")
d_age <- hier_add(d_age, root = "age_group2", "ag2a")

# define the cell key object
countvars <- c("cnt_females", "cnt_males", "cnt_highincome")
numvars <- c("expend", "income", "savings", "mixed")
tab <- ck_setup(
  x = x,
  rkey = "rkey",
  dims = list(sex = d_sex, age = d_age),
  w = "sampling_weight",
  countvars = countvars,
  numvars = numvars)

# show some information about this table instance
tab$print() # identical with print(tab)
#> ── Table Information ───────────────────────────────────────────────────────────
#> ✔ 45 cells in 2 dimensions ('sex', 'age')
#> ✔ weights: yes
#> ── Tabulated / Perturbed countvars ─────────────────────────────────────────────
#> ☐ 'total'
#> ☐ 'cnt_females'
#> ☐ 'cnt_males'
#> ☐ 'cnt_highincome'
#> ── Tabulated / Perturbed numvars ───────────────────────────────────────────────
#> ☐ 'expend'
#> ☐ 'income'
#> ☐ 'savings'
#> ☐ 'mixed'

# information about the hierarchies
tab$hierarchy_info()
#> $sex
#>      code level is_leaf parent
#> 1:  Total     1   FALSE  Total
#> 2:   male     2   FALSE  Total
#> 3:      m     3    TRUE   male
#> 4: female     2   FALSE  Total
#> 5:      f     3    TRUE female
#> 
#> $age
#>          code level is_leaf     parent
#> 1:      Total     1   FALSE      Total
#> 2: age_group1     2   FALSE      Total
#> 3:       ag1a     3    TRUE age_group1
#> 4: age_group2     2   FALSE      Total
#> 5:       ag2a     3    TRUE age_group2
#> 6: age_group3     2    TRUE      Total
#> 7: age_group4     2    TRUE      Total
#> 8: age_group5     2    TRUE      Total
#> 9: age_group6     2    TRUE      Total
#> 

# which variables have been defined?
tab$allvars()
#> $cntvars
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"
#> 
#> $numvars
#> [1] "expend"  "income"  "savings" "mixed"  
#> 

# count variables
tab$cntvars()
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"

# continuous variables
tab$numvars()
#> [1] "expend"  "income"  "savings" "mixed"  

# create perturbation parameters for "total" variable and
# write to yaml-file

# create a ptable using functionality from the ptable-pkg
f_yaml <- tempfile(fileext = ".yaml")
p_cnts1 <- ck_params_cnts(
  ptab = ptable::pt_ex_cnts(),
  path = f_yaml)
#> yaml configuration '/tmp/RtmpojDFXK/file19cd2c8af6f4.yaml' successfully written.

# read parameters from yaml-file and set them for variable `"total"`
p_cnts1 <- ck_read_yaml(path = f_yaml)

tab$params_cnts_set(val = p_cnts1, v = "total")
#> --> setting perturbation parameters for variable 'total'

# create alternative perturbation parameters by specifying parameters
para2 <- ptable::create_cnt_ptable(
  D = 8, V = 3, js = 2, create = FALSE)

p_cnts2 <- ck_params_cnts(ptab = para2)

# use these ptable it for the remaining variables
tab$params_cnts_set(val = p_cnts2, v = countvars)
#> --> setting perturbation parameters for variable 'cnt_females'
#> --> setting perturbation parameters for variable 'cnt_males'
#> --> setting perturbation parameters for variable 'cnt_highincome'

# perturb a variable
tab$perturb(v = "total")
#> Count variable 'total' was perturbed.

# multiple variables can be perturbed as well
tab$perturb(v = c("cnt_males", "cnt_highincome"))
#> Count variable 'cnt_males' was perturbed.
#> Count variable 'cnt_highincome' was perturbed.

# return weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))
#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 275043 4581 275103.0531
#>  2:  Total age_group1     total 1969 118786 1969 118786.0000
#>  3:  Total       ag1a     total 1969 118786 1969 118786.0000
#>  4:  Total age_group2     total 1143  68714 1143  68714.0000
#>  5:  Total       ag2a     total 1143  68714 1143  68714.0000
#>  6:  Total age_group3     total  864  51511  865  51570.6192
#>  7:  Total age_group4     total  423  25215  423  25215.0000
#>  8:  Total age_group5     total  168   9996  168   9996.0000
#>  9:  Total age_group6     total   13    821   13    821.0000
#> 10:   male      Total     total 2296 137342 2295 137282.1821
#> 11:      m      Total     total 2296 137342 2295 137282.1821
#> 12:   male age_group1     total 1015  61562 1015  61562.0000
#> 13:      m age_group1     total 1015  61562 1015  61562.0000
#> 14:   male       ag1a     total 1015  61562 1015  61562.0000
#> 15:      m       ag1a     total 1015  61562 1015  61562.0000
#> 16:   male age_group2     total  571  33966  571  33966.0000
#> 17:      m age_group2     total  571  33966  571  33966.0000
#> 18:   male       ag2a     total  571  33966  571  33966.0000
#> 19:      m       ag2a     total  571  33966  571  33966.0000
#> 20:   male age_group3     total  424  25064  424  25064.0000
#> 21:      m age_group3     total  424  25064  424  25064.0000
#> 22:   male age_group4     total  195  11258  196  11315.7333
#> 23:      m age_group4     total  195  11258  196  11315.7333
#> 24:   male age_group5     total   84   5090   86   5211.1905
#> 25:      m age_group5     total   84   5090   86   5211.1905
#> 26:   male age_group6     total    7    402    8    459.4286
#> 27:      m age_group6     total    7    402    8    459.4286
#> 28: female      Total     total 2284 137701 2284 137701.0000
#> 29:      f      Total     total 2284 137701 2284 137701.0000
#> 30: female age_group1     total  954  57224  952  57104.0335
#> 31:      f age_group1     total  954  57224  952  57104.0335
#> 32: female       ag1a     total  954  57224  952  57104.0335
#> 33:      f       ag1a     total  954  57224  952  57104.0335
#> 34: female age_group2     total  572  34748  570  34626.5035
#> 35:      f age_group2     total  572  34748  570  34626.5035
#> 36: female       ag2a     total  572  34748  570  34626.5035
#> 37:      f       ag2a     total  572  34748  570  34626.5035
#> 38: female age_group3     total  440  26447  440  26447.0000
#> 39:      f age_group3     total  440  26447  440  26447.0000
#> 40: female age_group4     total  228  13957  229  14018.2149
#> 41:      f age_group4     total  228  13957  229  14018.2149
#> 42: female age_group5     total   84   4906   84   4906.0000
#> 43:      f age_group5     total   84   4906   84   4906.0000
#> 44: female age_group6     total    6    419    7    488.8333
#> 45:      f age_group6     total    6    419    7    488.8333
#> 46:  Total      Total cnt_males 2296 137342 2295 137282.1821
#> 47:  Total age_group1 cnt_males 1015  61562 1015  61562.0000
#> 48:  Total       ag1a cnt_males 1015  61562 1015  61562.0000
#> 49:  Total age_group2 cnt_males  571  33966  571  33966.0000
#> 50:  Total       ag2a cnt_males  571  33966  571  33966.0000
#> 51:  Total age_group3 cnt_males  424  25064  424  25064.0000
#> 52:  Total age_group4 cnt_males  195  11258  197  11373.4667
#> 53:  Total age_group5 cnt_males   84   5090   87   5271.7857
#> 54:  Total age_group6 cnt_males    7    402    8    459.4286
#> 55:   male      Total cnt_males 2296 137342 2295 137282.1821
#> 56:      m      Total cnt_males 2296 137342 2295 137282.1821
#> 57:   male age_group1 cnt_males 1015  61562 1015  61562.0000
#> 58:      m age_group1 cnt_males 1015  61562 1015  61562.0000
#> 59:   male       ag1a cnt_males 1015  61562 1015  61562.0000
#> 60:      m       ag1a cnt_males 1015  61562 1015  61562.0000
#> 61:   male age_group2 cnt_males  571  33966  571  33966.0000
#> 62:      m age_group2 cnt_males  571  33966  571  33966.0000
#> 63:   male       ag2a cnt_males  571  33966  571  33966.0000
#> 64:      m       ag2a cnt_males  571  33966  571  33966.0000
#> 65:   male age_group3 cnt_males  424  25064  424  25064.0000
#> 66:      m age_group3 cnt_males  424  25064  424  25064.0000
#> 67:   male age_group4 cnt_males  195  11258  197  11373.4667
#> 68:      m age_group4 cnt_males  195  11258  197  11373.4667
#> 69:   male age_group5 cnt_males   84   5090   87   5271.7857
#> 70:      m age_group5 cnt_males   84   5090   87   5271.7857
#> 71:   male age_group6 cnt_males    7    402    8    459.4286
#> 72:      m age_group6 cnt_males    7    402    8    459.4286
#> 73: female      Total cnt_males    0      0    0      0.0000
#> 74:      f      Total cnt_males    0      0    0      0.0000
#> 75: female age_group1 cnt_males    0      0    0      0.0000
#> 76:      f age_group1 cnt_males    0      0    0      0.0000
#> 77: female       ag1a cnt_males    0      0    0      0.0000
#> 78:      f       ag1a cnt_males    0      0    0      0.0000
#> 79: female age_group2 cnt_males    0      0    0      0.0000
#> 80:      f age_group2 cnt_males    0      0    0      0.0000
#> 81: female       ag2a cnt_males    0      0    0      0.0000
#> 82:      f       ag2a cnt_males    0      0    0      0.0000
#> 83: female age_group3 cnt_males    0      0    0      0.0000
#> 84:      f age_group3 cnt_males    0      0    0      0.0000
#> 85: female age_group4 cnt_males    0      0    0      0.0000
#> 86:      f age_group4 cnt_males    0      0    0      0.0000
#> 87: female age_group5 cnt_males    0      0    0      0.0000
#> 88:      f age_group5 cnt_males    0      0    0      0.0000
#> 89: female age_group6 cnt_males    0      0    0      0.0000
#> 90:      f age_group6 cnt_males    0      0    0      0.0000
#>        sex        age     vname  uwc     wc puwc         pwc

# numerical variables (positive variables using flex-function)
# we also write the config to a yaml file
f_yaml <- tempfile(fileext = ".yaml")

# create a ptable using functionality from the ptable-pkg
# a single ptable for all cells
ptab1 <- ptable::pt_ex_nums(parity = TRUE, separation = FALSE)

# a single ptab for all cells except for very small ones
ptab2 <- ptable::pt_ex_nums(parity = TRUE, separation = TRUE)

# different ptables for cells with even/odd number of contributors
# and very small cells
ptab3 <- ptable::pt_ex_nums(parity = FALSE, separation = TRUE)

p_nums1 <- ck_params_nums(
  ptab = ptab1,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.30, 0.03),
    epsilon = c(1, 0.5, 0.2),
    q = 3),
  mu_c = 2,
  same_key = FALSE,
  use_zero_rkeys = FALSE,
  path = f_yaml)
#> yaml configuration '/tmp/RtmpojDFXK/file19cd11bc82cc.yaml' successfully written.

# we read the parameters from the yaml-file
p_nums1 <- ck_read_yaml(path = f_yaml)

# for variables with positive and negative values
p_nums2 <- ck_params_nums(
  ptab = ptab2,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.15, 0.02),
    epsilon = c(1, 0.4, 0.15),
    q = 3),
  mu_c = 2,
  same_key = FALSE)

# simple perturbation parameters (not using the flex-function approach)
p_nums3 <- ck_params_nums(
  ptab = ptab3,
  type = "mean",
  mult_params = ck_simpleparams(p = 0.25),
  mu_c = 2,
  same_key = FALSE)

# use `p_nums1` for all variables
tab$params_nums_set(p_nums1, c("savings", "income", "expend"))
#> --> setting perturbation parameters for variable 'savings'
#> --> setting perturbation parameters for variable 'income'
#> --> setting perturbation parameters for variable 'expend'

# use different parameters for variable `mixed`
tab$params_nums_set(p_nums2, v = "mixed")
#> --> setting perturbation parameters for variable 'mixed'

# identify sensitive cells to which extra protection (`mu_c`) is added.
tab$supp_p(v = "income", p = 85)
#> computing contributing indices | rawdata <--> table; this might take a while
#> p%-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_pq(v = "income", p = 85, q = 90)
#> computing contributing indices | rawdata <--> table; this might take a while
#> pq-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_nk(v = "income", n = 2, k = 90)
#> computing contributing indices | rawdata <--> table; this might take a while
#> nk-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_freq(v = "income", n = 14, weighted = FALSE)
#> freq-rule: 5 new sensitive cells (incl. duplicates) found (total: 5)
tab$supp_val(v = "income", n = 10000, weighted = TRUE)
#> val-rule: 0 new sensitive cells (incl. duplicates) found (total: 5)
tab$supp_cells(
  v = "income",
  inp = data.frame(
    sex = c("female", "female"),
    "age" = c("age_group1", "age_group3")
  )
)
#> cell-rule: 2 new sensitive cells (incl. duplicates) found (total: 7)

# perturb variables
tab$perturb(v = c("income", "savings"))
#> Numeric variable 'income' was perturbed.
#> Numeric variable 'savings' was perturbed.

# extract results
tab$numtab("income", mean_before_sum = TRUE)
#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1385231746 1385337488
#>  2:  Total age_group1 income  9810547  598656368  598556446
#>  3:  Total       ag1a income  9810547  598656368  598556446
#>  4:  Total age_group2 income  5692119  341706068  341696746
#>  5:  Total       ag2a income  5692119  341706068  341696746
#>  6:  Total age_group3 income  4406946  263609757  263664365
#>  7:  Total age_group4 income  2133543  126320267  126312850
#>  8:  Total age_group5 income   848151   50797143   50952670
#>  9:  Total age_group6 income    61672    4142143    4367387
#> 10:   male      Total income 11262049  677829443  677666801
#> 11:      m      Total income 11262049  677829443  677666801
#> 12:   male age_group1 income  4877164  301734253  301541248
#> 13:      m age_group1 income  4877164  301734253  301541248
#> 14:   male       ag1a income  4877164  301734253  301541248
#> 15:      m       ag1a income  4877164  301734253  301541248
#> 16:   male age_group2 income  2811379  164071664  164126053
#> 17:      m age_group2 income  2811379  164071664  164126053
#> 18:   male       ag2a income  2811379  164071664  164126053
#> 19:      m       ag2a income  2811379  164071664  164126053
#> 20:   male age_group3 income  2168169  129602078  129586826
#> 21:      m age_group3 income  2168169  129602078  129586826
#> 22:   male age_group4 income   978510   55873379   55972152
#> 23:      m age_group4 income   978510   55873379   55972152
#> 24:   male age_group5 income   393134   24445205   24446662
#> 25:      m age_group5 income   393134   24445205   24446662
#> 26:   male age_group6 income    33693    2102864    2284361
#> 27:      m age_group6 income    33693    2102864    2284361
#> 28: female      Total income 11690929  707402303  707430380
#> 29:      f      Total income 11690929  707402303  707430380
#> 30: female age_group1 income  4933383  296922115  296868750
#> 31:      f age_group1 income  4933383  296922115  296868750
#> 32: female       ag1a income  4933383  296922115  296868750
#> 33:      f       ag1a income  4933383  296922115  296868750
#> 34: female age_group2 income  2880740  177634404  177620912
#> 35:      f age_group2 income  2880740  177634404  177620912
#> 36: female       ag2a income  2880740  177634404  177620912
#> 37:      f       ag2a income  2880740  177634404  177620912
#> 38: female age_group3 income  2238777  134007679  133942381
#> 39:      f age_group3 income  2238777  134007679  133942381
#> 40: female age_group4 income  1155033   70446888   70576767
#> 41:      f age_group4 income  1155033   70446888   70576767
#> 42: female age_group5 income   455017   26351938   26256459
#> 43:      f age_group5 income   455017   26351938   26256459
#> 44: female age_group6 income    27979    2039279    2258736
#> 45:      f age_group6 income    27979    2039279    2258736
#>        sex        age  vname      uws         ws        pws
tab$numtab("income", mean_before_sum = FALSE)
#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1385231746 1385284616
#>  2:  Total age_group1 income  9810547  598656368  598606405
#>  3:  Total       ag1a income  9810547  598656368  598606405
#>  4:  Total age_group2 income  5692119  341706068  341701407
#>  5:  Total       ag2a income  5692119  341706068  341701407
#>  6:  Total age_group3 income  4406946  263609757  263637060
#>  7:  Total age_group4 income  2133543  126320267  126316558
#>  8:  Total age_group5 income   848151   50797143   50874847
#>  9:  Total age_group6 income    61672    4142143    4253274
#> 10:   male      Total income 11262049  677829443  677748117
#> 11:      m      Total income 11262049  677829443  677748117
#> 12:   male age_group1 income  4877164  301734253  301637735
#> 13:      m age_group1 income  4877164  301734253  301637735
#> 14:   male       ag1a income  4877164  301734253  301637735
#> 15:      m       ag1a income  4877164  301734253  301637735
#> 16:   male age_group2 income  2811379  164071664  164098856
#> 17:      m age_group2 income  2811379  164071664  164098856
#> 18:   male       ag2a income  2811379  164071664  164098856
#> 19:      m       ag2a income  2811379  164071664  164098856
#> 20:   male age_group3 income  2168169  129602078  129594452
#> 21:      m age_group3 income  2168169  129602078  129594452
#> 22:   male age_group4 income   978510   55873379   55922744
#> 23:      m age_group4 income   978510   55873379   55922744
#> 24:   male age_group5 income   393134   24445205   24445933
#> 25:      m age_group5 income   393134   24445205   24445933
#> 26:   male age_group6 income    33693    2102864    2191735
#> 27:      m age_group6 income    33693    2102864    2191735
#> 28: female      Total income 11690929  707402303  707416341
#> 29:      f      Total income 11690929  707402303  707416341
#> 30: female age_group1 income  4933383  296922115  296895431
#> 31:      f age_group1 income  4933383  296922115  296895431
#> 32: female       ag1a income  4933383  296922115  296895431
#> 33:      f       ag1a income  4933383  296922115  296895431
#> 34: female age_group2 income  2880740  177634404  177627658
#> 35:      f age_group2 income  2880740  177634404  177627658
#> 36: female       ag2a income  2880740  177634404  177627658
#> 37:      f       ag2a income  2880740  177634404  177627658
#> 38: female age_group3 income  2238777  134007679  133975026
#> 39:      f age_group3 income  2238777  134007679  133975026
#> 40: female age_group4 income  1155033   70446888   70511797
#> 41:      f age_group4 income  1155033   70446888   70511797
#> 42: female age_group5 income   455017   26351938   26304155
#> 43:      f age_group5 income   455017   26351938   26304155
#> 44: female age_group6 income    27979    2039279    2146204
#> 45:      f age_group6 income    27979    2039279    2146204
#>        sex        age  vname      uws         ws        pws
tab$numtab("savings")
#>        sex        age   vname     uws        ws         pws
#>  1:  Total      Total savings 2273532 135685280 135685887.0
#>  2:  Total age_group1 savings  982386  58941186  58943797.7
#>  3:  Total       ag1a savings  982386  58941186  58943797.7
#>  4:  Total age_group2 savings  552336  32980869  32980947.1
#>  5:  Total       ag2a savings  552336  32980869  32980947.1
#>  6:  Total age_group3 savings  437101  26011884  26006612.1
#>  7:  Total age_group4 savings  214661  12582264  12582288.7
#>  8:  Total age_group5 savings   80451   4810176   4812927.6
#>  9:  Total age_group6 savings    6597    358901    362950.1
#> 10:   male      Total savings 1159816  68684263  68686922.7
#> 11:      m      Total savings 1159816  68684263  68686922.7
#> 12:   male age_group1 savings  517660  31158219  31144098.7
#> 13:      m age_group1 savings  517660  31158219  31144098.7
#> 14:   male       ag1a savings  517660  31158219  31144098.7
#> 15:      m       ag1a savings  517660  31158219  31144098.7
#> 16:   male age_group2 savings  280923  16571971  16574166.9
#> 17:      m age_group2 savings  280923  16571971  16574166.9
#> 18:   male       ag2a savings  280923  16571971  16574166.9
#> 19:      m       ag2a savings  280923  16571971  16574166.9
#> 20:   male age_group3 savings  214970  12565243  12571102.5
#> 21:      m age_group3 savings  214970  12565243  12571102.5
#> 22:   male age_group4 savings   99420   5617952   5620455.0
#> 23:      m age_group4 savings   99420   5617952   5620455.0
#> 24:   male age_group5 savings   43233   2602302   2601029.0
#> 25:      m age_group5 savings   43233   2602302   2601029.0
#> 26:   male age_group6 savings    3610    168576    168909.8
#> 27:      m age_group6 savings    3610    168576    168909.8
#> 28: female      Total savings 1113716  67001017  66997341.7
#> 29:      f      Total savings 1113716  67001017  66997341.7
#> 30: female age_group1 savings  464726  27782967  27784076.9
#> 31:      f age_group1 savings  464726  27782967  27784076.9
#> 32: female       ag1a savings  464726  27782967  27784076.9
#> 33:      f       ag1a savings  464726  27782967  27784076.9
#> 34: female age_group2 savings  271413  16408898  16408546.4
#> 35:      f age_group2 savings  271413  16408898  16408546.4
#> 36: female       ag2a savings  271413  16408898  16408546.4
#> 37:      f       ag2a savings  271413  16408898  16408546.4
#> 38: female age_group3 savings  222131  13446641  13439034.9
#> 39:      f age_group3 savings  222131  13446641  13439034.9
#> 40: female age_group4 savings  115241   6964312   6966341.7
#> 41:      f age_group4 savings  115241   6964312   6966341.7
#> 42: female age_group5 savings   37218   2207874   2204617.9
#> 43:      f age_group5 savings   37218   2207874   2204617.9
#> 44: female age_group6 savings    2987    190325    195608.5
#> 45:      f age_group6 savings    2987    190325    195608.5
#>        sex        age   vname     uws        ws         pws

# results can be resetted, too
tab$reset_cntvars(v = "cnt_males")

# we can then set other parameters and perturb again
tab$params_cnts_set(val = p_cnts1, v = "cnt_males")
#> --> setting perturbation parameters for variable 'cnt_males'

tab$perturb(v = "cnt_males")
#> Count variable 'cnt_males' was perturbed.

# write results to a .csv file
tab$freqtab(
  v = c("total", "cnt_males"),
  path = file.path(tempdir(), "outtab.csv")
)
#> File '/tmp/RtmpojDFXK/outtab.csv' successfully written to disk.
#> NULL

# show results containing weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))
#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 275043 4581 275103.0531
#>  2:  Total age_group1     total 1969 118786 1969 118786.0000
#>  3:  Total       ag1a     total 1969 118786 1969 118786.0000
#>  4:  Total age_group2     total 1143  68714 1143  68714.0000
#>  5:  Total       ag2a     total 1143  68714 1143  68714.0000
#>  6:  Total age_group3     total  864  51511  865  51570.6192
#>  7:  Total age_group4     total  423  25215  423  25215.0000
#>  8:  Total age_group5     total  168   9996  168   9996.0000
#>  9:  Total age_group6     total   13    821   13    821.0000
#> 10:   male      Total     total 2296 137342 2295 137282.1821
#> 11:      m      Total     total 2296 137342 2295 137282.1821
#> 12:   male age_group1     total 1015  61562 1015  61562.0000
#> 13:      m age_group1     total 1015  61562 1015  61562.0000
#> 14:   male       ag1a     total 1015  61562 1015  61562.0000
#> 15:      m       ag1a     total 1015  61562 1015  61562.0000
#> 16:   male age_group2     total  571  33966  571  33966.0000
#> 17:      m age_group2     total  571  33966  571  33966.0000
#> 18:   male       ag2a     total  571  33966  571  33966.0000
#> 19:      m       ag2a     total  571  33966  571  33966.0000
#> 20:   male age_group3     total  424  25064  424  25064.0000
#> 21:      m age_group3     total  424  25064  424  25064.0000
#> 22:   male age_group4     total  195  11258  196  11315.7333
#> 23:      m age_group4     total  195  11258  196  11315.7333
#> 24:   male age_group5     total   84   5090   86   5211.1905
#> 25:      m age_group5     total   84   5090   86   5211.1905
#> 26:   male age_group6     total    7    402    8    459.4286
#> 27:      m age_group6     total    7    402    8    459.4286
#> 28: female      Total     total 2284 137701 2284 137701.0000
#> 29:      f      Total     total 2284 137701 2284 137701.0000
#> 30: female age_group1     total  954  57224  952  57104.0335
#> 31:      f age_group1     total  954  57224  952  57104.0335
#> 32: female       ag1a     total  954  57224  952  57104.0335
#> 33:      f       ag1a     total  954  57224  952  57104.0335
#> 34: female age_group2     total  572  34748  570  34626.5035
#> 35:      f age_group2     total  572  34748  570  34626.5035
#> 36: female       ag2a     total  572  34748  570  34626.5035
#> 37:      f       ag2a     total  572  34748  570  34626.5035
#> 38: female age_group3     total  440  26447  440  26447.0000
#> 39:      f age_group3     total  440  26447  440  26447.0000
#> 40: female age_group4     total  228  13957  229  14018.2149
#> 41:      f age_group4     total  228  13957  229  14018.2149
#> 42: female age_group5     total   84   4906   84   4906.0000
#> 43:      f age_group5     total   84   4906   84   4906.0000
#> 44: female age_group6     total    6    419    7    488.8333
#> 45:      f age_group6     total    6    419    7    488.8333
#> 46:  Total      Total cnt_males 2296 137342 2295 137282.1821
#> 47:  Total age_group1 cnt_males 1015  61562 1015  61562.0000
#> 48:  Total       ag1a cnt_males 1015  61562 1015  61562.0000
#> 49:  Total age_group2 cnt_males  571  33966  571  33966.0000
#> 50:  Total       ag2a cnt_males  571  33966  571  33966.0000
#> 51:  Total age_group3 cnt_males  424  25064  424  25064.0000
#> 52:  Total age_group4 cnt_males  195  11258  196  11315.7333
#> 53:  Total age_group5 cnt_males   84   5090   86   5211.1905
#> 54:  Total age_group6 cnt_males    7    402    8    459.4286
#> 55:   male      Total cnt_males 2296 137342 2295 137282.1821
#> 56:      m      Total cnt_males 2296 137342 2295 137282.1821
#> 57:   male age_group1 cnt_males 1015  61562 1015  61562.0000
#> 58:      m age_group1 cnt_males 1015  61562 1015  61562.0000
#> 59:   male       ag1a cnt_males 1015  61562 1015  61562.0000
#> 60:      m       ag1a cnt_males 1015  61562 1015  61562.0000
#> 61:   male age_group2 cnt_males  571  33966  571  33966.0000
#> 62:      m age_group2 cnt_males  571  33966  571  33966.0000
#> 63:   male       ag2a cnt_males  571  33966  571  33966.0000
#> 64:      m       ag2a cnt_males  571  33966  571  33966.0000
#> 65:   male age_group3 cnt_males  424  25064  424  25064.0000
#> 66:      m age_group3 cnt_males  424  25064  424  25064.0000
#> 67:   male age_group4 cnt_males  195  11258  196  11315.7333
#> 68:      m age_group4 cnt_males  195  11258  196  11315.7333
#> 69:   male age_group5 cnt_males   84   5090   86   5211.1905
#> 70:      m age_group5 cnt_males   84   5090   86   5211.1905
#> 71:   male age_group6 cnt_males    7    402    8    459.4286
#> 72:      m age_group6 cnt_males    7    402    8    459.4286
#> 73: female      Total cnt_males    0      0    0      0.0000
#> 74:      f      Total cnt_males    0      0    0      0.0000
#> 75: female age_group1 cnt_males    0      0    0      0.0000
#> 76:      f age_group1 cnt_males    0      0    0      0.0000
#> 77: female       ag1a cnt_males    0      0    0      0.0000
#> 78:      f       ag1a cnt_males    0      0    0      0.0000
#> 79: female age_group2 cnt_males    0      0    0      0.0000
#> 80:      f age_group2 cnt_males    0      0    0      0.0000
#> 81: female       ag2a cnt_males    0      0    0      0.0000
#> 82:      f       ag2a cnt_males    0      0    0      0.0000
#> 83: female age_group3 cnt_males    0      0    0      0.0000
#> 84:      f age_group3 cnt_males    0      0    0      0.0000
#> 85: female age_group4 cnt_males    0      0    0      0.0000
#> 86:      f age_group4 cnt_males    0      0    0      0.0000
#> 87: female age_group5 cnt_males    0      0    0      0.0000
#> 88:      f age_group5 cnt_males    0      0    0      0.0000
#> 89: female age_group6 cnt_males    0      0    0      0.0000
#> 90:      f age_group6 cnt_males    0      0    0      0.0000
#>        sex        age     vname  uwc     wc puwc         pwc

# utility measures for a count variable
tab$measures_cnts(v = "total", exclude_zeros = TRUE)
#> $overview
#>    noise cnt        pct
#> 1:    -2   2 0.04444444
#> 2:    -1  10 0.22222222
#> 3:     0  23 0.51111111
#> 4:     1   2 0.04444444
#> 5:     2   8 0.17777778
#> 
#> $measures
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 0.000 0.000 0.000
#>  6:   Mean 0.711 0.016 0.032
#>  7: Median 0.000 0.000 0.000
#>  8:    Q60 1.000 0.002 0.023
#>  9:    Q70 1.000 0.003 0.033
#> 10:    Q80 2.000 0.004 0.042
#> 11:    Q90 2.000 0.024 0.108
#> 12:    Q95 2.000 0.143 0.183
#> 13:    Q99 2.000 0.167 0.196
#> 14:    Max 2.000 0.167 0.196
#> 
#> $cumdistr_d1
#>    cat cnt       pct
#> 1:   0  23 0.5111111
#> 2:   1  35 0.7777778
#> 3:   2  45 1.0000000
#> 
#> $cumdistr_d2
#>            cat cnt       pct
#> 1:    [0,0.02]  39 0.8666667
#> 2: (0.02,0.05]  41 0.9111111
#> 3:  (0.05,0.1]  41 0.9111111
#> 4:   (0.1,0.2]  45 1.0000000
#> 5:   (0.2,0.3]  45 1.0000000
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $cumdistr_d3
#>            cat cnt       pct
#> 1:    [0,0.02]  27 0.6000000
#> 2: (0.02,0.05]  39 0.8666667
#> 3:  (0.05,0.1]  39 0.8666667
#> 4:   (0.1,0.2]  45 1.0000000
#> 5:   (0.2,0.3]  45 1.0000000
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $false_zero
#> [1] 0
#> 
#> $false_nonzero
#> [1] 0
#> 
#> $exclude_zeros
#> [1] TRUE
#> 

# modifications for perturbed count variables
tab$mod_cnts()
#>         sex        age row_nr pert      ckey  countvar
#>   1:  Total      Total     16    1 0.8970891     total
#>   2:  Total age_group1     15    0 0.6110620     total
#>   3:  Total       ag1a     15    0 0.6110620     total
#>   4:  Total age_group2     15    0 0.6256478     total
#>   5:  Total       ag2a     15    0 0.6256478     total
#>  ---                                                  
#> 131:      f age_group4     -1    0 0.0000000 cnt_males
#> 132: female age_group5     -1    0 0.0000000 cnt_males
#> 133:      f age_group5     -1    0 0.0000000 cnt_males
#> 134: female age_group6     -1    0 0.0000000 cnt_males
#> 135:      f age_group6     -1    0 0.0000000 cnt_males

# display a summary about utility measures
tab$summary()
#> ┌──────────────────────────────────────────────┐
#> │Utility measures for perturbed count variables│
#> └──────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#>          countvar Min Q10 Q20 Q30 Q40   Mean Median Q60 Q70 Q80 Q90 Q95  Q99
#> 1:          total  -2  -2  -1   0   0 -0.089      0   0   0 1.0   1 1.0 2.00
#> 2: cnt_highincome  -3  -2  -2  -1   0 -0.333      0   0   0 1.0   1 1.8 2.56
#> 3:      cnt_males  -1   0   0   0   0  0.200      0   0   0 0.2   1 1.8 2.00
#>    Max
#> 1:   2
#> 2:   3
#> 3:   2
#> 
#> ── Distance-based measures ─────────────────────────────────────────────────────
#> ✔ Variable: 'total'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 0.000 0.000 0.000
#>  6:   Mean 0.711 0.016 0.032
#>  7: Median 0.000 0.000 0.000
#>  8:    Q60 1.000 0.002 0.023
#>  9:    Q70 1.000 0.003 0.033
#> 10:    Q80 2.000 0.004 0.042
#> 11:    Q90 2.000 0.024 0.108
#> 12:    Q95 2.000 0.143 0.183
#> 13:    Q99 2.000 0.167 0.196
#> 14:    Max 2.000 0.167 0.196
#> 
#> ✔ Variable: 'cnt_males'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 0.000 0.000 0.000
#>  6:   Mean 0.556 0.019 0.037
#>  7: Median 0.000 0.000 0.000
#>  8:    Q60 1.000 0.000 0.010
#>  9:    Q70 1.000 0.005 0.036
#> 10:    Q80 1.000 0.020 0.094
#> 11:    Q90 1.400 0.071 0.138
#> 12:    Q95 2.000 0.143 0.183
#> 13:    Q99 2.000 0.143 0.183
#> 14:    Max 2.000 0.143 0.183
#> 
#> ✔ Variable: 'cnt_highincome'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.700 0.005 0.037
#>  5:    Q40 1.000 0.011 0.058
#>  6:   Mean 1.175 0.043 0.101
#>  7: Median 1.000 0.015 0.066
#>  8:    Q60 1.000 0.016 0.091
#>  9:    Q70 2.000 0.049 0.155
#> 10:    Q80 2.000 0.078 0.179
#> 11:    Q90 2.100 0.143 0.200
#> 12:    Q95 3.000 0.146 0.244
#> 13:    Q99 3.000 0.200 0.409
#> 14:    Max 3.000 0.200 0.409
#> 
#> ┌──────────────────────────────────────────────────┐
#> │Utility measures for perturbed numerical variables│
#> └──────────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#>      vname       Min        Q10        Q20        Q30       Q40     Mean
#> 1:  expend       Inf         NA         NA         NA        NA      NaN
#> 2:  income -96518.10 -81325.764 -47782.645 -26683.771 -7098.073 1242.801
#> 3: savings -14120.29  -7606.079  -3339.935   -351.571    56.779 -697.920
#> 4:   mixed       Inf         NA         NA         NA        NA      NaN
#>       Median      Q60       Q70       Q80       Q90        Q95        Q99
#> 1:        NA       NA        NA        NA        NA         NA         NA
#> 2: -4661.057 6052.398 27192.256 50065.610 84404.072 103314.297 109280.602
#> 3:   607.023 1477.846  2195.869  2611.747  3530.078   5283.536   5859.527
#> 4:        NA       NA        NA        NA        NA         NA         NA
#>           Max
#> 1:       -Inf
#> 2: 111131.262
#> 3:   5859.527
#> 4:       -Inf
# }