Read perturbation parameters from yaml-files

ck_read_yaml() allows to create perturbation parameter inputs from yaml-files that were previously created using ck_params_cnts() or ck_params_nums().

ck_read_yaml(path)

Arguments

path

a path to a yaml-input file

Value

an object object suitable as input to method $params_nums_set() for the perturbation of continous variables in case path was created using ck_params_nums() or an object suitable as input for $params_cnts_set() for the perturbation of counts and frequencies if the input file was generated using ck_params_cnts().

Examples

x <- ck_create_testdata()

# create some 0/1 variables that should be perturbed later
x[, cnt_females := ifelse(sex == "male", 0, 1)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              77          25.00000           0
#>    2:   2530      28       1              45          25.00000           1
#>    3:   6920     550       1              84          25.00000           0
#>    4:   7960     870       1              50          25.00000           0
#>    5:   9030      20       2              96          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              49          16.66667           1
#> 4577:   1420     987    1000              49          16.66667           0
#> 4578:   8900     684    1000              91          16.66667           0
#> 4579:   3880     294    1000              94          16.66667           1
#> 4580:   4830     911    1000              23          16.66667           0
x[, cnt_males := ifelse(sex == "male", 1, 0)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              77          25.00000           0
#>    2:   2530      28       1              45          25.00000           1
#>    3:   6920     550       1              84          25.00000           0
#>    4:   7960     870       1              50          25.00000           0
#>    5:   9030      20       2              96          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              49          16.66667           1
#> 4577:   1420     987    1000              49          16.66667           0
#> 4578:   8900     684    1000              91          16.66667           0
#> 4579:   3880     294    1000              94          16.66667           1
#> 4580:   4830     911    1000              23          16.66667           0
#>       cnt_males
#>    1:         1
#>    2:         0
#>    3:         1
#>    4:         1
#>    5:         1
#>   ---          
#> 4576:         0
#> 4577:         1
#> 4578:         1
#> 4579:         0
#> 4580:         1
x[, cnt_highincome := ifelse(income >= 9000, 1, 0)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              77          25.00000           0
#>    2:   2530      28       1              45          25.00000           1
#>    3:   6920     550       1              84          25.00000           0
#>    4:   7960     870       1              50          25.00000           0
#>    5:   9030      20       2              96          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              49          16.66667           1
#> 4577:   1420     987    1000              49          16.66667           0
#> 4578:   8900     684    1000              91          16.66667           0
#> 4579:   3880     294    1000              94          16.66667           1
#> 4580:   4830     911    1000              23          16.66667           0
#>       cnt_males cnt_highincome
#>    1:         1              0
#>    2:         0              0
#>    3:         1              0
#>    4:         1              0
#>    5:         1              1
#>   ---                         
#> 4576:         0              0
#> 4577:         1              0
#> 4578:         1              0
#> 4579:         0              0
#> 4580:         1              0
# a variable with positive and negative contributions
x[, mixed := sample(-10:10, nrow(x), replace = TRUE)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              77          25.00000           0
#>    2:   2530      28       1              45          25.00000           1
#>    3:   6920     550       1              84          25.00000           0
#>    4:   7960     870       1              50          25.00000           0
#>    5:   9030      20       2              96          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              49          16.66667           1
#> 4577:   1420     987    1000              49          16.66667           0
#> 4578:   8900     684    1000              91          16.66667           0
#> 4579:   3880     294    1000              94          16.66667           1
#> 4580:   4830     911    1000              23          16.66667           0
#>       cnt_males cnt_highincome mixed
#>    1:         1              0     7
#>    2:         0              0     2
#>    3:         1              0    -2
#>    4:         1              0    -8
#>    5:         1              1     6
#>   ---                               
#> 4576:         0              0     5
#> 4577:         1              0     2
#> 4578:         1              0     1
#> 4579:         0              0    10
#> 4580:         1              0     6

# create record keys
x$rkey <- ck_generate_rkeys(dat = x)

# define required inputs

# hierarchy with some bogus codes
d_sex <- hier_create(root = "Total", nodes = c("male", "female"))
d_sex <- hier_add(d_sex, root = "female", "f")
d_sex <- hier_add(d_sex, root = "male", "m")

d_age <- hier_create(root = "Total", nodes = paste0("age_group", 1:6))
d_age <- hier_add(d_age, root = "age_group1", "ag1a")
d_age <- hier_add(d_age, root = "age_group2", "ag2a")

# define the cell key object
countvars <- c("cnt_females", "cnt_males", "cnt_highincome")
numvars <- c("expend", "income", "savings", "mixed")
tab <- ck_setup(
  x = x,
  rkey = "rkey",
  dims = list(sex = d_sex, age = d_age),
  w = "sampling_weight",
  countvars = countvars,
  numvars = numvars)

# show some information about this table instance
tab$print() # identical with print(tab)
#> ── Table Information ───────────────────────────────────────────────────────────
#> ✔ 45 cells in 2 dimensions ('sex', 'age')
#> ✔ weights: yes
#> ── Tabulated / Perturbed countvars ─────────────────────────────────────────────
#> ☐ 'total'
#> ☐ 'cnt_females'
#> ☐ 'cnt_males'
#> ☐ 'cnt_highincome'
#> ── Tabulated / Perturbed numvars ───────────────────────────────────────────────
#> ☐ 'expend'
#> ☐ 'income'
#> ☐ 'savings'
#> ☐ 'mixed'

# information about the hierarchies
tab$hierarchy_info()
#> $sex
#>      code level is_leaf parent
#> 1:  Total     1   FALSE  Total
#> 2:   male     2   FALSE  Total
#> 3:      m     3    TRUE   male
#> 4: female     2   FALSE  Total
#> 5:      f     3    TRUE female
#> 
#> $age
#>          code level is_leaf     parent
#> 1:      Total     1   FALSE      Total
#> 2: age_group1     2   FALSE      Total
#> 3:       ag1a     3    TRUE age_group1
#> 4: age_group2     2   FALSE      Total
#> 5:       ag2a     3    TRUE age_group2
#> 6: age_group3     2    TRUE      Total
#> 7: age_group4     2    TRUE      Total
#> 8: age_group5     2    TRUE      Total
#> 9: age_group6     2    TRUE      Total
#> 

# which variables have been defined?
tab$allvars()
#> $cntvars
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"
#> 
#> $numvars
#> [1] "expend"  "income"  "savings" "mixed"  
#> 

# count variables
tab$cntvars()
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"

# continuous variables
tab$numvars()
#> [1] "expend"  "income"  "savings" "mixed"  

# create perturbation parameters for "total" variable and
# write to yaml-file

# create a ptable using functionality from the ptable-pkg
f_yaml <- tempfile(fileext = ".yaml")
p_cnts1 <- ck_params_cnts(
  ptab = ptable::pt_ex_cnts(),
  path = f_yaml)
#> yaml configuration '/tmp/RtmpRLPm23/file1e1b200d87f1.yaml' successfully written.

# read parameters from yaml-file and set them for variable `"total"`
p_cnts1 <- ck_read_yaml(path = f_yaml)

tab$params_cnts_set(val = p_cnts1, v = "total")
#> --> setting perturbation parameters for variable 'total'

# create alternative perturbation parameters by specifying parameters
para2 <- ptable::create_num_ptable(
  D = 8, V = 3, js = 2, create = FALSE)
#> Error in ptable::create_num_ptable(D = 8, V = 3, js = 2, create = FALSE): unused argument (js = 2)

p_cnts2 <- ck_params_cnts(ptab = para2)
#> Error in ck_params_cnts(ptab = para2): object 'para2' not found

# use these ptable it for the remaining variables
tab$params_cnts_set(val = p_cnts2, v = countvars)
#> Error in tab$params_cnts_set(val = p_cnts2, v = countvars): object 'p_cnts2' not found

# perturb a variable
tab$perturb(v = "total")
#> Count variable 'total' was perturbed.

# multiple variables can be perturbed as well
tab$perturb(v = c("cnt_males", "cnt_highincome"))
#> Error: No perturbation parameters found for variable 'cnt_males' - Please use method `$params_cnts_set().

# return weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))
#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 275654 4580 275654.0000
#>  2:  Total age_group1     total 1969 119029 1968 118968.5485
#>  3:  Total       ag1a     total 1969 119029 1968 118968.5485
#>  4:  Total age_group2     total 1143  67463 1143  67463.0000
#>  5:  Total       ag2a     total 1143  67463 1143  67463.0000
#>  6:  Total age_group3     total  864  53049  863  52987.6007
#>  7:  Total age_group4     total  423  25364  422  25304.0378
#>  8:  Total age_group5     total  168   9847  168   9847.0000
#>  9:  Total age_group6     total   13    902   14    971.3846
#> 10:   male      Total     total 2296 137712 2297 137771.9791
#> 11:      m      Total     total 2296 137712 2297 137771.9791
#> 12:   male age_group1     total 1015  61622 1013  61500.5773
#> 13:      m age_group1     total 1015  61622 1013  61500.5773
#> 14:   male       ag1a     total 1015  61622 1013  61500.5773
#> 15:      m       ag1a     total 1015  61622 1013  61500.5773
#> 16:   male age_group2     total  571  33767  571  33767.0000
#> 17:      m age_group2     total  571  33767  571  33767.0000
#> 18:   male       ag2a     total  571  33767  571  33767.0000
#> 19:      m       ag2a     total  571  33767  571  33767.0000
#> 20:   male age_group3     total  424  25515  425  25575.1769
#> 21:      m age_group3     total  424  25515  425  25575.1769
#> 22:   male age_group4     total  195  11354  196  11412.2256
#> 23:      m age_group4     total  195  11354  196  11412.2256
#> 24:   male age_group5     total   84   4960   83   4900.9524
#> 25:      m age_group5     total   84   4960   83   4900.9524
#> 26:   male age_group6     total    7    494    6    423.4286
#> 27:      m age_group6     total    7    494    6    423.4286
#> 28: female      Total     total 2284 137942 2285 138002.3949
#> 29:      f      Total     total 2284 137942 2285 138002.3949
#> 30: female age_group1     total  954  57407  953  57346.8249
#> 31:      f age_group1     total  954  57407  953  57346.8249
#> 32: female       ag1a     total  954  57407  953  57346.8249
#> 33:      f       ag1a     total  954  57407  953  57346.8249
#> 34: female age_group2     total  572  33696  573  33754.9091
#> 35:      f age_group2     total  572  33696  573  33754.9091
#> 36: female       ag2a     total  572  33696  573  33754.9091
#> 37:      f       ag2a     total  572  33696  573  33754.9091
#> 38: female age_group3     total  440  27534  440  27534.0000
#> 39:      f age_group3     total  440  27534  440  27534.0000
#> 40: female age_group4     total  228  14010  228  14010.0000
#> 41:      f age_group4     total  228  14010  228  14010.0000
#> 42: female age_group5     total   84   4887   84   4887.0000
#> 43:      f age_group5     total   84   4887   84   4887.0000
#> 44: female age_group6     total    6    408    6    408.0000
#> 45:      f age_group6     total    6    408    6    408.0000
#> 46:  Total      Total cnt_males 2296 137712   NA          NA
#> 47:  Total age_group1 cnt_males 1015  61622   NA          NA
#> 48:  Total       ag1a cnt_males 1015  61622   NA          NA
#> 49:  Total age_group2 cnt_males  571  33767   NA          NA
#> 50:  Total       ag2a cnt_males  571  33767   NA          NA
#> 51:  Total age_group3 cnt_males  424  25515   NA          NA
#> 52:  Total age_group4 cnt_males  195  11354   NA          NA
#> 53:  Total age_group5 cnt_males   84   4960   NA          NA
#> 54:  Total age_group6 cnt_males    7    494   NA          NA
#> 55:   male      Total cnt_males 2296 137712   NA          NA
#> 56:      m      Total cnt_males 2296 137712   NA          NA
#> 57:   male age_group1 cnt_males 1015  61622   NA          NA
#> 58:      m age_group1 cnt_males 1015  61622   NA          NA
#> 59:   male       ag1a cnt_males 1015  61622   NA          NA
#> 60:      m       ag1a cnt_males 1015  61622   NA          NA
#> 61:   male age_group2 cnt_males  571  33767   NA          NA
#> 62:      m age_group2 cnt_males  571  33767   NA          NA
#> 63:   male       ag2a cnt_males  571  33767   NA          NA
#> 64:      m       ag2a cnt_males  571  33767   NA          NA
#> 65:   male age_group3 cnt_males  424  25515   NA          NA
#> 66:      m age_group3 cnt_males  424  25515   NA          NA
#> 67:   male age_group4 cnt_males  195  11354   NA          NA
#> 68:      m age_group4 cnt_males  195  11354   NA          NA
#> 69:   male age_group5 cnt_males   84   4960   NA          NA
#> 70:      m age_group5 cnt_males   84   4960   NA          NA
#> 71:   male age_group6 cnt_males    7    494   NA          NA
#> 72:      m age_group6 cnt_males    7    494   NA          NA
#> 73: female      Total cnt_males    0      0   NA          NA
#> 74:      f      Total cnt_males    0      0   NA          NA
#> 75: female age_group1 cnt_males    0      0   NA          NA
#> 76:      f age_group1 cnt_males    0      0   NA          NA
#> 77: female       ag1a cnt_males    0      0   NA          NA
#> 78:      f       ag1a cnt_males    0      0   NA          NA
#> 79: female age_group2 cnt_males    0      0   NA          NA
#> 80:      f age_group2 cnt_males    0      0   NA          NA
#> 81: female       ag2a cnt_males    0      0   NA          NA
#> 82:      f       ag2a cnt_males    0      0   NA          NA
#> 83: female age_group3 cnt_males    0      0   NA          NA
#> 84:      f age_group3 cnt_males    0      0   NA          NA
#> 85: female age_group4 cnt_males    0      0   NA          NA
#> 86:      f age_group4 cnt_males    0      0   NA          NA
#> 87: female age_group5 cnt_males    0      0   NA          NA
#> 88:      f age_group5 cnt_males    0      0   NA          NA
#> 89: female age_group6 cnt_males    0      0   NA          NA
#> 90:      f age_group6 cnt_males    0      0   NA          NA
#>        sex        age     vname  uwc     wc puwc         pwc

# numerical variables (positive variables using flex-function)
# we also write the config to a yaml file
f_yaml <- tempfile(fileext = ".yaml")

# create a ptable using functionality from the ptable-pkg
# a single ptable for all cells
ptab1 <- ptable::pt_ex_nums(parity = TRUE, separation = FALSE)

# a single ptab for all cells except for very small ones
ptab2 <- ptable::pt_ex_nums(parity = TRUE, separation = TRUE)

# different ptables for cells with even/odd number of contributors
# and very small cells
ptab3 <- ptable::pt_ex_nums(parity = FALSE, separation = TRUE)

p_nums1 <- ck_params_nums(
  ptab = ptab1,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.30, 0.03),
    epsilon = c(1, 0.5, 0.2),
    q = 3),
  mu_c = 2,
  same_key = FALSE,
  use_zero_rkeys = FALSE,
  path = f_yaml)
#> yaml configuration '/tmp/RtmpRLPm23/file1e1b4753ed2d.yaml' successfully written.

# we read the parameters from the yaml-file
p_nums1 <- ck_read_yaml(path = f_yaml)

# for variables with positive and negative values
p_nums2 <- ck_params_nums(
  ptab = ptab2,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.15, 0.02),
    epsilon = c(1, 0.4, 0.15),
    q = 3),
  mu_c = 2,
  same_key = FALSE)

# simple perturbation parameters (not using the flex-function approach)
p_nums3 <- ck_params_nums(
  ptab = ptab3,
  type = "mean",
  mult_params = ck_simpleparams(p = 0.25),
  mu_c = 2,
  same_key = FALSE)

# use `p_nums1` for all variables
tab$params_nums_set(p_nums1, c("savings", "income", "expend"))
#> --> setting perturbation parameters for variable 'savings'
#> --> setting perturbation parameters for variable 'income'
#> --> setting perturbation parameters for variable 'expend'

# use different parameters for variable `mixed`
tab$params_nums_set(p_nums2, v = "mixed")
#> --> setting perturbation parameters for variable 'mixed'

# identify sensitive cells to which extra protection (`mu_c`) is added.
tab$supp_p(v = "income", p = 85)
#> computing contributing indices | rawdata <--> table; this might take a while
#> p%-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_pq(v = "income", p = 85, q = 90)
#> computing contributing indices | rawdata <--> table; this might take a while
#> pq-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_nk(v = "income", n = 2, k = 90)
#> computing contributing indices | rawdata <--> table; this might take a while
#> nk-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_freq(v = "income", n = 14, weighted = FALSE)
#> freq-rule: 5 new sensitive cells (incl. duplicates) found (total: 5)
tab$supp_val(v = "income", n = 10000, weighted = TRUE)
#> val-rule: 0 new sensitive cells (incl. duplicates) found (total: 5)
tab$supp_cells(
  v = "income",
  inp = data.frame(
    sex = c("female", "female"),
    "age" = c("age_group1", "age_group3")
  )
)
#> cell-rule: 2 new sensitive cells (incl. duplicates) found (total: 7)

# perturb variables
tab$perturb(v = c("income", "savings"))
#> Numeric variable 'income' was perturbed.
#> Numeric variable 'savings' was perturbed.

# extract results
tab$numtab("income", mean_before_sum = TRUE)
#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1378798016 1378816199
#>  2:  Total age_group1 income  9810547  591503483  591517471
#>  3:  Total       ag1a income  9810547  591503483  591517471
#>  4:  Total age_group2 income  5692119  332688820  332697096
#>  5:  Total       ag2a income  5692119  332688820  332697096
#>  6:  Total age_group3 income  4406946  271040544  270977583
#>  7:  Total age_group4 income  2133543  129232678  129134064
#>  8:  Total age_group5 income   848151   50014695   50000305
#>  9:  Total age_group6 income    61672    4317796    4475139
#> 10:   male      Total income 11262049  675328222  675229587
#> 11:      m      Total income 11262049  675328222  675229587
#> 12:   male age_group1 income  4877164  296188978  296179823
#> 13:      m age_group1 income  4877164  296188978  296179823
#> 14:   male       ag1a income  4877164  296188978  296179823
#> 15:      m       ag1a income  4877164  296188978  296179823
#> 16:   male age_group2 income  2811379  165821914  165937258
#> 17:      m age_group2 income  2811379  165821914  165937258
#> 18:   male       ag2a income  2811379  165821914  165937258
#> 19:      m       ag2a income  2811379  165821914  165937258
#> 20:   male age_group3 income  2168169  129661874  129751388
#> 21:      m age_group3 income  2168169  129661874  129751388
#> 22:   male age_group4 income   978510   57656313   57636258
#> 23:      m age_group4 income   978510   57656313   57636258
#> 24:   male age_group5 income   393134   23415745   23407045
#> 25:      m age_group5 income   393134   23415745   23407045
#> 26:   male age_group6 income    33693    2583398    2725650
#> 27:      m age_group6 income    33693    2583398    2725650
#> 28: female      Total income 11690929  703469794  703511197
#> 29:      f      Total income 11690929  703469794  703511197
#> 30: female age_group1 income  4933383  295314505  295321932
#> 31:      f age_group1 income  4933383  295314505  295321932
#> 32: female       ag1a income  4933383  295314505  295321932
#> 33:      f       ag1a income  4933383  295314505  295321932
#> 34: female age_group2 income  2880740  166866906  166877324
#> 35:      f age_group2 income  2880740  166866906  166877324
#> 36: female       ag2a income  2880740  166866906  166877324
#> 37:      f       ag2a income  2880740  166866906  166877324
#> 38: female age_group3 income  2238777  141378670  141327504
#> 39:      f age_group3 income  2238777  141378670  141327504
#> 40: female age_group4 income  1155033   71576365   71623606
#> 41:      f age_group4 income  1155033   71576365   71623606
#> 42: female age_group5 income   455017   26598950   26643841
#> 43:      f age_group5 income   455017   26598950   26643841
#> 44: female age_group6 income    27979    1734398    1722577
#> 45:      f age_group6 income    27979    1734398    1722577
#>        sex        age  vname      uws         ws        pws
tab$numtab("income", mean_before_sum = FALSE)
#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1378798016 1378807108
#>  2:  Total age_group1 income  9810547  591503483  591510477
#>  3:  Total       ag1a income  9810547  591503483  591510477
#>  4:  Total age_group2 income  5692119  332688820  332692958
#>  5:  Total       ag2a income  5692119  332688820  332692958
#>  6:  Total age_group3 income  4406946  271040544  271009061
#>  7:  Total age_group4 income  2133543  129232678  129183362
#>  8:  Total age_group5 income   848151   50014695   50007500
#>  9:  Total age_group6 income    61672    4317796    4395763
#> 10:   male      Total income 11262049  675328222  675278903
#> 11:      m      Total income 11262049  675328222  675278903
#> 12:   male age_group1 income  4877164  296188978  296184401
#> 13:      m age_group1 income  4877164  296188978  296184401
#> 14:   male       ag1a income  4877164  296188978  296184401
#> 15:      m       ag1a income  4877164  296188978  296184401
#> 16:   male age_group2 income  2811379  165821914  165879576
#> 17:      m age_group2 income  2811379  165821914  165879576
#> 18:   male       ag2a income  2811379  165821914  165879576
#> 19:      m       ag2a income  2811379  165821914  165879576
#> 20:   male age_group3 income  2168169  129661874  129706623
#> 21:      m age_group3 income  2168169  129661874  129706623
#> 22:   male age_group4 income   978510   57656313   57646284
#> 23:      m age_group4 income   978510   57656313   57646284
#> 24:   male age_group5 income   393134   23415745   23411395
#> 25:      m age_group5 income   393134   23415745   23411395
#> 26:   male age_group6 income    33693    2583398    2653571
#> 27:      m age_group6 income    33693    2583398    2653571
#> 28: female      Total income 11690929  703469794  703490495
#> 29:      f      Total income 11690929  703469794  703490495
#> 30: female age_group1 income  4933383  295314505  295318218
#> 31:      f age_group1 income  4933383  295314505  295318218
#> 32: female       ag1a income  4933383  295314505  295318218
#> 33:      f       ag1a income  4933383  295314505  295318218
#> 34: female age_group2 income  2880740  166866906  166872115
#> 35:      f age_group2 income  2880740  166866906  166872115
#> 36: female       ag2a income  2880740  166866906  166872115
#> 37:      f       ag2a income  2880740  166866906  166872115
#> 38: female age_group3 income  2238777  141378670  141353085
#> 39:      f age_group3 income  2238777  141378670  141353085
#> 40: female age_group4 income  1155033   71576365   71599981
#> 41:      f age_group4 income  1155033   71576365   71599981
#> 42: female age_group5 income   455017   26598950   26621386
#> 43:      f age_group5 income   455017   26598950   26621386
#> 44: female age_group6 income    27979    1734398    1728477
#> 45:      f age_group6 income    27979    1734398    1728477
#>        sex        age  vname      uws         ws        pws
tab$numtab("savings")
#>        sex        age   vname     uws        ws         pws
#>  1:  Total      Total savings 2273532 135853153 135853657.5
#>  2:  Total age_group1 savings  982386  59253733  59255044.8
#>  3:  Total       ag1a savings  982386  59253733  59255044.8
#>  4:  Total age_group2 savings  552336  32139945  32139528.5
#>  5:  Total       ag2a savings  552336  32139945  32139528.5
#>  6:  Total age_group3 savings  437101  26833456  26824256.1
#>  7:  Total age_group4 savings  214661  12716337  12712863.9
#>  8:  Total age_group5 savings   80451   4517559   4517657.8
#>  9:  Total age_group6 savings    6597    392123    393324.3
#> 10:   male      Total savings 1159816  69236247  69241762.9
#> 11:      m      Total savings 1159816  69236247  69241762.9
#> 12:   male age_group1 savings  517660  31351659  31352088.5
#> 13:      m age_group1 savings  517660  31351659  31352088.5
#> 14:   male       ag1a savings  517660  31351659  31352088.5
#> 15:      m       ag1a savings  517660  31351659  31352088.5
#> 16:   male age_group2 savings  280923  16408805  16411780.9
#> 17:      m age_group2 savings  280923  16408805  16411780.9
#> 18:   male       ag2a savings  280923  16408805  16411780.9
#> 19:      m       ag2a savings  280923  16408805  16411780.9
#> 20:   male age_group3 savings  214970  13053972  13058190.9
#> 21:      m age_group3 savings  214970  13053972  13058190.9
#> 22:   male age_group4 savings   99420   5725818   5724139.2
#> 23:      m age_group4 savings   99420   5725818   5724139.2
#> 24:   male age_group5 savings   43233   2476527   2475982.4
#> 25:      m age_group5 savings   43233   2476527   2475982.4
#> 26:   male age_group6 savings    3610    219466    224876.4
#> 27:      m age_group6 savings    3610    219466    224876.4
#> 28: female      Total savings 1113716  66616906  66615595.8
#> 29:      f      Total savings 1113716  66616906  66615595.8
#> 30: female age_group1 savings  464726  27902074  27905142.3
#> 31:      f age_group1 savings  464726  27902074  27905142.3
#> 32: female       ag1a savings  464726  27902074  27905142.3
#> 33:      f       ag1a savings  464726  27902074  27905142.3
#> 34: female age_group2 savings  271413  15731140  15731864.5
#> 35:      f age_group2 savings  271413  15731140  15731864.5
#> 36: female       ag2a savings  271413  15731140  15731864.5
#> 37:      f       ag2a savings  271413  15731140  15731864.5
#> 38: female age_group3 savings  222131  13779484  13777791.8
#> 39:      f age_group3 savings  222131  13779484  13777791.8
#> 40: female age_group4 savings  115241   6990519   6991912.6
#> 41:      f age_group4 savings  115241   6990519   6991912.6
#> 42: female age_group5 savings   37218   2041032   2042191.2
#> 43:      f age_group5 savings   37218   2041032   2042191.2
#> 44: female age_group6 savings    2987    172657    172162.6
#> 45:      f age_group6 savings    2987    172657    172162.6
#>        sex        age   vname     uws        ws         pws

# results can be resetted, too
tab$reset_cntvars(v = "cnt_males")
#> Error: Invalid variables specified in `v`: Possible choices are:
#>  'total'

# we can then set other parameters and perturb again
tab$params_cnts_set(val = p_cnts1, v = "cnt_males")
#> --> setting perturbation parameters for variable 'cnt_males'
tab$perturb(v = "cnt_males")
#> Count variable 'cnt_males' was perturbed.

# write to a file "outtab.csv" (.csv is automatically added to the path)
if (FALSE) {
tab$freqtab(v = c("total", "cnt_males"), path = "outtab.csv")
}

# show results containing weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))
#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 275654 4580 275654.0000
#>  2:  Total age_group1     total 1969 119029 1968 118968.5485
#>  3:  Total       ag1a     total 1969 119029 1968 118968.5485
#>  4:  Total age_group2     total 1143  67463 1143  67463.0000
#>  5:  Total       ag2a     total 1143  67463 1143  67463.0000
#>  6:  Total age_group3     total  864  53049  863  52987.6007
#>  7:  Total age_group4     total  423  25364  422  25304.0378
#>  8:  Total age_group5     total  168   9847  168   9847.0000
#>  9:  Total age_group6     total   13    902   14    971.3846
#> 10:   male      Total     total 2296 137712 2297 137771.9791
#> 11:      m      Total     total 2296 137712 2297 137771.9791
#> 12:   male age_group1     total 1015  61622 1013  61500.5773
#> 13:      m age_group1     total 1015  61622 1013  61500.5773
#> 14:   male       ag1a     total 1015  61622 1013  61500.5773
#> 15:      m       ag1a     total 1015  61622 1013  61500.5773
#> 16:   male age_group2     total  571  33767  571  33767.0000
#> 17:      m age_group2     total  571  33767  571  33767.0000
#> 18:   male       ag2a     total  571  33767  571  33767.0000
#> 19:      m       ag2a     total  571  33767  571  33767.0000
#> 20:   male age_group3     total  424  25515  425  25575.1769
#> 21:      m age_group3     total  424  25515  425  25575.1769
#> 22:   male age_group4     total  195  11354  196  11412.2256
#> 23:      m age_group4     total  195  11354  196  11412.2256
#> 24:   male age_group5     total   84   4960   83   4900.9524
#> 25:      m age_group5     total   84   4960   83   4900.9524
#> 26:   male age_group6     total    7    494    6    423.4286
#> 27:      m age_group6     total    7    494    6    423.4286
#> 28: female      Total     total 2284 137942 2285 138002.3949
#> 29:      f      Total     total 2284 137942 2285 138002.3949
#> 30: female age_group1     total  954  57407  953  57346.8249
#> 31:      f age_group1     total  954  57407  953  57346.8249
#> 32: female       ag1a     total  954  57407  953  57346.8249
#> 33:      f       ag1a     total  954  57407  953  57346.8249
#> 34: female age_group2     total  572  33696  573  33754.9091
#> 35:      f age_group2     total  572  33696  573  33754.9091
#> 36: female       ag2a     total  572  33696  573  33754.9091
#> 37:      f       ag2a     total  572  33696  573  33754.9091
#> 38: female age_group3     total  440  27534  440  27534.0000
#> 39:      f age_group3     total  440  27534  440  27534.0000
#> 40: female age_group4     total  228  14010  228  14010.0000
#> 41:      f age_group4     total  228  14010  228  14010.0000
#> 42: female age_group5     total   84   4887   84   4887.0000
#> 43:      f age_group5     total   84   4887   84   4887.0000
#> 44: female age_group6     total    6    408    6    408.0000
#> 45:      f age_group6     total    6    408    6    408.0000
#> 46:  Total      Total cnt_males 2296 137712 2297 137771.9791
#> 47:  Total age_group1 cnt_males 1015  61622 1013  61500.5773
#> 48:  Total       ag1a cnt_males 1015  61622 1013  61500.5773
#> 49:  Total age_group2 cnt_males  571  33767  571  33767.0000
#> 50:  Total       ag2a cnt_males  571  33767  571  33767.0000
#> 51:  Total age_group3 cnt_males  424  25515  425  25575.1769
#> 52:  Total age_group4 cnt_males  195  11354  196  11412.2256
#> 53:  Total age_group5 cnt_males   84   4960   83   4900.9524
#> 54:  Total age_group6 cnt_males    7    494    6    423.4286
#> 55:   male      Total cnt_males 2296 137712 2297 137771.9791
#> 56:      m      Total cnt_males 2296 137712 2297 137771.9791
#> 57:   male age_group1 cnt_males 1015  61622 1013  61500.5773
#> 58:      m age_group1 cnt_males 1015  61622 1013  61500.5773
#> 59:   male       ag1a cnt_males 1015  61622 1013  61500.5773
#> 60:      m       ag1a cnt_males 1015  61622 1013  61500.5773
#> 61:   male age_group2 cnt_males  571  33767  571  33767.0000
#> 62:      m age_group2 cnt_males  571  33767  571  33767.0000
#> 63:   male       ag2a cnt_males  571  33767  571  33767.0000
#> 64:      m       ag2a cnt_males  571  33767  571  33767.0000
#> 65:   male age_group3 cnt_males  424  25515  425  25575.1769
#> 66:      m age_group3 cnt_males  424  25515  425  25575.1769
#> 67:   male age_group4 cnt_males  195  11354  196  11412.2256
#> 68:      m age_group4 cnt_males  195  11354  196  11412.2256
#> 69:   male age_group5 cnt_males   84   4960   83   4900.9524
#> 70:      m age_group5 cnt_males   84   4960   83   4900.9524
#> 71:   male age_group6 cnt_males    7    494    6    423.4286
#> 72:      m age_group6 cnt_males    7    494    6    423.4286
#> 73: female      Total cnt_males    0      0    0      0.0000
#> 74:      f      Total cnt_males    0      0    0      0.0000
#> 75: female age_group1 cnt_males    0      0    0      0.0000
#> 76:      f age_group1 cnt_males    0      0    0      0.0000
#> 77: female       ag1a cnt_males    0      0    0      0.0000
#> 78:      f       ag1a cnt_males    0      0    0      0.0000
#> 79: female age_group2 cnt_males    0      0    0      0.0000
#> 80:      f age_group2 cnt_males    0      0    0      0.0000
#> 81: female       ag2a cnt_males    0      0    0      0.0000
#> 82:      f       ag2a cnt_males    0      0    0      0.0000
#> 83: female age_group3 cnt_males    0      0    0      0.0000
#> 84:      f age_group3 cnt_males    0      0    0      0.0000
#> 85: female age_group4 cnt_males    0      0    0      0.0000
#> 86:      f age_group4 cnt_males    0      0    0      0.0000
#> 87: female age_group5 cnt_males    0      0    0      0.0000
#> 88:      f age_group5 cnt_males    0      0    0      0.0000
#> 89: female age_group6 cnt_males    0      0    0      0.0000
#> 90:      f age_group6 cnt_males    0      0    0      0.0000
#>        sex        age     vname  uwc     wc puwc         pwc

# utility measures for a count variable
tab$measures_cnts(v = "total", exclude_zeros = TRUE)
#> $overview
#>    noise cnt        pct
#> 1:    -1  13 0.28888889
#> 2:     0  16 0.35555556
#> 3:     1  12 0.26666667
#> 4:     2   4 0.08888889
#> 
#> $measures
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 1.000 0.000 0.010
#>  6:   Mean 0.733 0.009 0.025
#>  7: Median 1.000 0.001 0.016
#>  8:    Q60 1.000 0.001 0.019
#>  9:    Q70 1.000 0.002 0.024
#> 10:    Q80 1.000 0.002 0.031
#> 11:    Q90 1.000 0.009 0.047
#> 12:    Q95 2.000 0.064 0.120
#> 13:    Q99 2.000 0.143 0.196
#> 14:    Max 2.000 0.143 0.196
#> 
#> $cumdistr_d1
#>    cat cnt       pct
#> 1:   0  16 0.3555556
#> 2:   1  41 0.9111111
#> 3:   2  45 1.0000000
#> 
#> $cumdistr_d2
#>            cat cnt       pct
#> 1:    [0,0.02]  42 0.9333333
#> 2: (0.02,0.05]  42 0.9333333
#> 3:  (0.05,0.1]  43 0.9555556
#> 4:   (0.1,0.2]  45 1.0000000
#> 5:   (0.2,0.3]  45 1.0000000
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $cumdistr_d3
#>            cat cnt       pct
#> 1:    [0,0.02]  27 0.6000000
#> 2: (0.02,0.05]  40 0.8888889
#> 3:  (0.05,0.1]  42 0.9333333
#> 4:   (0.1,0.2]  45 1.0000000
#> 5:   (0.2,0.3]  45 1.0000000
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $false_zero
#> [1] 0
#> 
#> $false_nonzero
#> [1] 0
#> 
#> $exclude_zeros
#> [1] TRUE
#> 

# modifications for perturbed count variables
tab$mod_cnts()
#>        sex        age row_nr pert       ckey  countvar
#>  1:  Total      Total     15    0 0.56112821     total
#>  2:  Total age_group1     14   -1 0.28132810     total
#>  3:  Total       ag1a     14   -1 0.28132810     total
#>  4:  Total age_group2     15    0 0.34396863     total
#>  5:  Total       ag2a     15    0 0.34396863     total
#>  6:  Total age_group3     14   -1 0.20284939     total
#>  7:  Total age_group4     14   -1 0.24014911     total
#>  8:  Total age_group5     15    0 0.62718981     total
#>  9:  Total age_group6     16    1 0.86564317     total
#> 10:   male      Total     16    1 0.78467977     total
#> 11:      m      Total     16    1 0.78467977     total
#> 12:   male age_group1     13   -2 0.03662473     total
#> 13:      m age_group1     13   -2 0.03662473     total
#> 14:   male       ag1a     13   -2 0.03662473     total
#> 15:      m       ag1a     13   -2 0.03662473     total
#> 16:   male age_group2     15    0 0.58834634     total
#> 17:      m age_group2     15    0 0.58834634     total
#> 18:   male       ag2a     15    0 0.58834634     total
#> 19:      m       ag2a     15    0 0.58834634     total
#> 20:   male age_group3     16    1 0.84974742     total
#> 21:      m age_group3     16    1 0.84974742     total
#> 22:   male age_group4     16    1 0.86214426     total
#> 23:      m age_group4     16    1 0.86214426     total
#> 24:   male age_group5     14   -1 0.15357983     total
#> 25:      m age_group5     14   -1 0.15357983     total
#> 26:   male age_group6     14   -1 0.29423719     total
#> 27:      m age_group6     14   -1 0.29423719     total
#> 28: female      Total     16    1 0.77644844     total
#> 29:      f      Total     16    1 0.77644844     total
#> 30: female age_group1     14   -1 0.24470337     total
#> 31:      f age_group1     14   -1 0.24470337     total
#> 32: female       ag1a     14   -1 0.24470337     total
#> 33:      f       ag1a     14   -1 0.24470337     total
#> 34: female age_group2     16    1 0.75562229     total
#> 35:      f age_group2     16    1 0.75562229     total
#> 36: female       ag2a     16    1 0.75562229     total
#> 37:      f       ag2a     16    1 0.75562229     total
#> 38: female age_group3     15    0 0.35310197     total
#> 39:      f age_group3     15    0 0.35310197     total
#> 40: female age_group4     15    0 0.37800485     total
#> 41:      f age_group4     15    0 0.37800485     total
#> 42: female age_group5     15    0 0.47360998     total
#> 43:      f age_group5     15    0 0.47360998     total
#> 44: female age_group6     15    0 0.57140598     total
#> 45:      f age_group6     15    0 0.57140598     total
#> 46:  Total      Total     16    1 0.78467977 cnt_males
#> 47:  Total age_group1     13   -2 0.03662473 cnt_males
#> 48:  Total       ag1a     13   -2 0.03662473 cnt_males
#> 49:  Total age_group2     15    0 0.58834634 cnt_males
#> 50:  Total       ag2a     15    0 0.58834634 cnt_males
#> 51:  Total age_group3     16    1 0.84974742 cnt_males
#> 52:  Total age_group4     16    1 0.86214426 cnt_males
#> 53:  Total age_group5     14   -1 0.15357983 cnt_males
#> 54:  Total age_group6     14   -1 0.29423719 cnt_males
#> 55:   male      Total     16    1 0.78467977 cnt_males
#> 56:      m      Total     16    1 0.78467977 cnt_males
#> 57:   male age_group1     13   -2 0.03662473 cnt_males
#> 58:      m age_group1     13   -2 0.03662473 cnt_males
#> 59:   male       ag1a     13   -2 0.03662473 cnt_males
#> 60:      m       ag1a     13   -2 0.03662473 cnt_males
#> 61:   male age_group2     15    0 0.58834634 cnt_males
#> 62:      m age_group2     15    0 0.58834634 cnt_males
#> 63:   male       ag2a     15    0 0.58834634 cnt_males
#> 64:      m       ag2a     15    0 0.58834634 cnt_males
#> 65:   male age_group3     16    1 0.84974742 cnt_males
#> 66:      m age_group3     16    1 0.84974742 cnt_males
#> 67:   male age_group4     16    1 0.86214426 cnt_males
#> 68:      m age_group4     16    1 0.86214426 cnt_males
#> 69:   male age_group5     14   -1 0.15357983 cnt_males
#> 70:      m age_group5     14   -1 0.15357983 cnt_males
#> 71:   male age_group6     14   -1 0.29423719 cnt_males
#> 72:      m age_group6     14   -1 0.29423719 cnt_males
#> 73: female      Total     -1    0 0.00000000 cnt_males
#> 74:      f      Total     -1    0 0.00000000 cnt_males
#> 75: female age_group1     -1    0 0.00000000 cnt_males
#> 76:      f age_group1     -1    0 0.00000000 cnt_males
#> 77: female       ag1a     -1    0 0.00000000 cnt_males
#> 78:      f       ag1a     -1    0 0.00000000 cnt_males
#> 79: female age_group2     -1    0 0.00000000 cnt_males
#> 80:      f age_group2     -1    0 0.00000000 cnt_males
#> 81: female       ag2a     -1    0 0.00000000 cnt_males
#> 82:      f       ag2a     -1    0 0.00000000 cnt_males
#> 83: female age_group3     -1    0 0.00000000 cnt_males
#> 84:      f age_group3     -1    0 0.00000000 cnt_males
#> 85: female age_group4     -1    0 0.00000000 cnt_males
#> 86:      f age_group4     -1    0 0.00000000 cnt_males
#> 87: female age_group5     -1    0 0.00000000 cnt_males
#> 88:      f age_group5     -1    0 0.00000000 cnt_males
#> 89: female age_group6     -1    0 0.00000000 cnt_males
#> 90:      f age_group6     -1    0 0.00000000 cnt_males
#>        sex        age row_nr pert       ckey  countvar

# display a summary about utility measures
tab$summary()
#> ┌──────────────────────────────────────────────┐
#> │Utility measures for perturbed count variables│
#> └──────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#>     countvar Min Q10 Q20 Q30 Q40   Mean Median Q60 Q70 Q80 Q90 Q95 Q99 Max
#> 1:     total  -2  -1  -1  -1   0 -0.156      0   0   0 1.0   1   1   1   1
#> 2: cnt_males  -2  -2  -1   0   0 -0.200      0   0   0 0.2   1   1   1   1
#> 
#> ── Distance-based measures ─────────────────────────────────────────────────────
#> ✔ Variable: 'total'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 1.000 0.000 0.010
#>  6:   Mean 0.733 0.009 0.025
#>  7: Median 1.000 0.001 0.016
#>  8:    Q60 1.000 0.001 0.019
#>  9:    Q70 1.000 0.002 0.024
#> 10:    Q80 1.000 0.002 0.031
#> 11:    Q90 1.000 0.009 0.047
#> 12:    Q95 2.000 0.064 0.120
#> 13:    Q99 2.000 0.143 0.196
#> 14:    Max 2.000 0.143 0.196
#> 
#> ✔ Variable: 'cnt_males'
#> 
#>       what  d1    d2    d3
#>  1:    Min 0.0 0.000 0.000
#>  2:    Q10 0.0 0.000 0.000
#>  3:    Q20 0.2 0.000 0.002
#>  4:    Q30 1.0 0.000 0.010
#>  5:    Q40 1.0 0.002 0.024
#>  6:   Mean 1.0 0.019 0.043
#>  7: Median 1.0 0.002 0.031
#>  8:    Q60 1.0 0.002 0.031
#>  9:    Q70 1.0 0.005 0.036
#> 10:    Q80 1.8 0.011 0.051
#> 11:    Q90 2.0 0.064 0.111
#> 12:    Q95 2.0 0.143 0.196
#> 13:    Q99 2.0 0.143 0.196
#> 14:    Max 2.0 0.143 0.196
#> 
#> ┌──────────────────────────────────────────────────┐
#> │Utility measures for perturbed numerical variables│
#> └──────────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#>      vname        Min        Q10       Q20       Q30      Q40     Mean   Median
#> 1:  expend        Inf         NA        NA        NA       NA      NaN       NA
#> 2:  income -49319.151 -25585.079 -6175.597 -4577.326 3713.332 9828.941 4138.005
#> 3: savings  -9199.918  -1678.822  -544.572  -313.448  429.536  970.472  724.501
#> 4:   mixed        Inf         NA        NA        NA       NA      NaN       NA
#>         Q60       Q70       Q80       Q90       Q95       Q99       Max
#> 1:       NA        NA        NA        NA        NA        NA      -Inf
#> 2: 5923.057 20701.206 27842.996 57662.178 67670.833 74537.917 77967.496
#> 3: 1245.531  2659.461  3068.287  4218.948  5410.428  5515.888  5515.888
#> 4:       NA        NA        NA        NA        NA        NA      -Inf