Set parameters required to perturb numeric variables using a simple approach

ck_simpleparams() allows to define parameters for a simple perturbation approach based on a single magnitude parameter (m). The values of epsilon are used to "weight" parameter m in case type == "top_contr" is set in ck_params_nums().

ck_simpleparams(p, epsilon = 1)

Arguments

p

a percentage value used as magnitude for perturbation

epsilon

a numeric vector in descending order with all values >= 0 and <= 1 with the first element forced to equal 1. The length of this vector must correspond with the number top_k specified in ck_params_nums() when creating parameters for type == "top_contr" which is checked at runtime. This setting allows to use different flex-functions for the largest top_k contributors.

Value

an object suitable as input for ck_params_nums().

Details

details about the flex function can be found in Deliverable D4.2, Part I in SGA "Open Source tools for perturbative confidentiality methods"

See also

Examples

# \donttest{
x <- ck_create_testdata()

# create some 0/1 variables that should be perturbed later
x[, cnt_females := ifelse(sex == "male", 0, 1)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              94          25.00000           0
#>    2:   2530      28       1              41          25.00000           1
#>    3:   6920     550       1              52          25.00000           0
#>    4:   7960     870       1              41          25.00000           0
#>    5:   9030      20       2              49          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              37          16.66667           1
#> 4577:   1420     987    1000              32          16.66667           0
#> 4578:   8900     684    1000              94          16.66667           0
#> 4579:   3880     294    1000              79          16.66667           1
#> 4580:   4830     911    1000              94          16.66667           0
x[, cnt_males := ifelse(sex == "male", 1, 0)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              94          25.00000           0
#>    2:   2530      28       1              41          25.00000           1
#>    3:   6920     550       1              52          25.00000           0
#>    4:   7960     870       1              41          25.00000           0
#>    5:   9030      20       2              49          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              37          16.66667           1
#> 4577:   1420     987    1000              32          16.66667           0
#> 4578:   8900     684    1000              94          16.66667           0
#> 4579:   3880     294    1000              79          16.66667           1
#> 4580:   4830     911    1000              94          16.66667           0
#>       cnt_males
#>    1:         1
#>    2:         0
#>    3:         1
#>    4:         1
#>    5:         1
#>   ---          
#> 4576:         0
#> 4577:         1
#> 4578:         1
#> 4579:         0
#> 4580:         1
x[, cnt_highincome := ifelse(income >= 9000, 1, 0)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              94          25.00000           0
#>    2:   2530      28       1              41          25.00000           1
#>    3:   6920     550       1              52          25.00000           0
#>    4:   7960     870       1              41          25.00000           0
#>    5:   9030      20       2              49          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              37          16.66667           1
#> 4577:   1420     987    1000              32          16.66667           0
#> 4578:   8900     684    1000              94          16.66667           0
#> 4579:   3880     294    1000              79          16.66667           1
#> 4580:   4830     911    1000              94          16.66667           0
#>       cnt_males cnt_highincome
#>    1:         1              0
#>    2:         0              0
#>    3:         1              0
#>    4:         1              0
#>    5:         1              1
#>   ---                         
#> 4576:         0              0
#> 4577:         1              0
#> 4578:         1              0
#> 4579:         0              0
#> 4580:         1              0
# a variable with positive and negative contributions
x[, mixed := sample(-10:10, nrow(x), replace = TRUE)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              94          25.00000           0
#>    2:   2530      28       1              41          25.00000           1
#>    3:   6920     550       1              52          25.00000           0
#>    4:   7960     870       1              41          25.00000           0
#>    5:   9030      20       2              49          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              37          16.66667           1
#> 4577:   1420     987    1000              32          16.66667           0
#> 4578:   8900     684    1000              94          16.66667           0
#> 4579:   3880     294    1000              79          16.66667           1
#> 4580:   4830     911    1000              94          16.66667           0
#>       cnt_males cnt_highincome mixed
#>    1:         1              0    -2
#>    2:         0              0     3
#>    3:         1              0    -2
#>    4:         1              0     5
#>    5:         1              1    -4
#>   ---                               
#> 4576:         0              0    -6
#> 4577:         1              0    -6
#> 4578:         1              0     9
#> 4579:         0              0     5
#> 4580:         1              0     3

# create record keys
x$rkey <- ck_generate_rkeys(dat = x)

# define required inputs

# hierarchy with some bogus codes
d_sex <- hier_create(root = "Total", nodes = c("male", "female"))
d_sex <- hier_add(d_sex, root = "female", "f")
d_sex <- hier_add(d_sex, root = "male", "m")

d_age <- hier_create(root = "Total", nodes = paste0("age_group", 1:6))
d_age <- hier_add(d_age, root = "age_group1", "ag1a")
d_age <- hier_add(d_age, root = "age_group2", "ag2a")

# define the cell key object
countvars <- c("cnt_females", "cnt_males", "cnt_highincome")
numvars <- c("expend", "income", "savings", "mixed")
tab <- ck_setup(
  x = x,
  rkey = "rkey",
  dims = list(sex = d_sex, age = d_age),
  w = "sampling_weight",
  countvars = countvars,
  numvars = numvars)

# show some information about this table instance
tab$print() # identical with print(tab)
#> ── Table Information ───────────────────────────────────────────────────────────
#> ✔ 45 cells in 2 dimensions ('sex', 'age')
#> ✔ weights: yes
#> ── Tabulated / Perturbed countvars ─────────────────────────────────────────────
#> ☐ 'total'
#> ☐ 'cnt_females'
#> ☐ 'cnt_males'
#> ☐ 'cnt_highincome'
#> ── Tabulated / Perturbed numvars ───────────────────────────────────────────────
#> ☐ 'expend'
#> ☐ 'income'
#> ☐ 'savings'
#> ☐ 'mixed'

# information about the hierarchies
tab$hierarchy_info()
#> $sex
#>      code level is_leaf parent
#> 1:  Total     1   FALSE  Total
#> 2:   male     2   FALSE  Total
#> 3:      m     3    TRUE   male
#> 4: female     2   FALSE  Total
#> 5:      f     3    TRUE female
#> 
#> $age
#>          code level is_leaf     parent
#> 1:      Total     1   FALSE      Total
#> 2: age_group1     2   FALSE      Total
#> 3:       ag1a     3    TRUE age_group1
#> 4: age_group2     2   FALSE      Total
#> 5:       ag2a     3    TRUE age_group2
#> 6: age_group3     2    TRUE      Total
#> 7: age_group4     2    TRUE      Total
#> 8: age_group5     2    TRUE      Total
#> 9: age_group6     2    TRUE      Total
#> 

# which variables have been defined?
tab$allvars()
#> $cntvars
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"
#> 
#> $numvars
#> [1] "expend"  "income"  "savings" "mixed"  
#> 

# count variables
tab$cntvars()
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"

# continuous variables
tab$numvars()
#> [1] "expend"  "income"  "savings" "mixed"  

# create perturbation parameters for "total" variable and
# write to yaml-file

# create a ptable using functionality from the ptable-pkg
f_yaml <- tempfile(fileext = ".yaml")
p_cnts1 <- ck_params_cnts(
  ptab = ptable::pt_ex_cnts(),
  path = f_yaml)
#> yaml configuration '/tmp/Rtmp2sERaX/file1a9042ab07f.yaml' successfully written.

# read parameters from yaml-file and set them for variable `"total"`
p_cnts1 <- ck_read_yaml(path = f_yaml)

tab$params_cnts_set(val = p_cnts1, v = "total")
#> --> setting perturbation parameters for variable 'total'

# create alternative perturbation parameters by specifying parameters
para2 <- ptable::create_cnt_ptable(
  D = 8, V = 3, js = 2, create = FALSE)

p_cnts2 <- ck_params_cnts(ptab = para2)

# use these ptable it for the remaining variables
tab$params_cnts_set(val = p_cnts2, v = countvars)
#> --> setting perturbation parameters for variable 'cnt_females'
#> --> setting perturbation parameters for variable 'cnt_males'
#> --> setting perturbation parameters for variable 'cnt_highincome'

# perturb a variable
tab$perturb(v = "total")
#> Count variable 'total' was perturbed.

# multiple variables can be perturbed as well
tab$perturb(v = c("cnt_males", "cnt_highincome"))
#> Count variable 'cnt_males' was perturbed.
#> Count variable 'cnt_highincome' was perturbed.

# return weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))
#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 274065 4579 274005.1605
#>  2:  Total age_group1     total 1969 117705 1968 117645.2209
#>  3:  Total       ag1a     total 1969 117705 1968 117645.2209
#>  4:  Total age_group2     total 1143  68914 1144  68974.2922
#>  5:  Total       ag2a     total 1143  68914 1144  68974.2922
#>  6:  Total age_group3     total  864  51238  865  51297.3032
#>  7:  Total age_group4     total  423  25315  425  25434.6927
#>  8:  Total age_group5     total  168  10164  170  10285.0000
#>  9:  Total age_group6     total   13    729   13    729.0000
#> 10:   male      Total     total 2296 136191 2295 136131.6834
#> 11:      m      Total     total 2296 136191 2295 136131.6834
#> 12:   male age_group1     total 1015  60214 1015  60214.0000
#> 13:      m age_group1     total 1015  60214 1015  60214.0000
#> 14:   male       ag1a     total 1015  60214 1015  60214.0000
#> 15:      m       ag1a     total 1015  60214 1015  60214.0000
#> 16:   male age_group2     total  571  34345  571  34345.0000
#> 17:      m age_group2     total  571  34345  571  34345.0000
#> 18:   male       ag2a     total  571  34345  571  34345.0000
#> 19:      m       ag2a     total  571  34345  571  34345.0000
#> 20:   male age_group3     total  424  24573  425  24630.9552
#> 21:      m age_group3     total  424  24573  425  24630.9552
#> 22:   male age_group4     total  195  11530  195  11530.0000
#> 23:      m age_group4     total  195  11530  195  11530.0000
#> 24:   male age_group5     total   84   5058   85   5118.2143
#> 25:      m age_group5     total   84   5058   85   5118.2143
#> 26:   male age_group6     total    7    471    6    403.7143
#> 27:      m age_group6     total    7    471    6    403.7143
#> 28: female      Total     total 2284 137874 2286 137994.7303
#> 29:      f      Total     total 2284 137874 2286 137994.7303
#> 30: female age_group1     total  954  57491  955  57551.2631
#> 31:      f age_group1     total  954  57491  955  57551.2631
#> 32: female       ag1a     total  954  57491  955  57551.2631
#> 33:      f       ag1a     total  954  57491  955  57551.2631
#> 34: female age_group2     total  572  34569  572  34569.0000
#> 35:      f age_group2     total  572  34569  572  34569.0000
#> 36: female       ag2a     total  572  34569  572  34569.0000
#> 37:      f       ag2a     total  572  34569  572  34569.0000
#> 38: female age_group3     total  440  26665  438  26543.7955
#> 39:      f age_group3     total  440  26665  438  26543.7955
#> 40: female age_group4     total  228  13785  228  13785.0000
#> 41:      f age_group4     total  228  13785  228  13785.0000
#> 42: female age_group5     total   84   5106   83   5045.2143
#> 43:      f age_group5     total   84   5106   83   5045.2143
#> 44: female age_group6     total    6    258    5    215.0000
#> 45:      f age_group6     total    6    258    5    215.0000
#> 46:  Total      Total cnt_males 2296 136191 2294 136072.3667
#> 47:  Total age_group1 cnt_males 1015  60214 1014  60154.6759
#> 48:  Total       ag1a cnt_males 1015  60214 1014  60154.6759
#> 49:  Total age_group2 cnt_males  571  34345  571  34345.0000
#> 50:  Total       ag2a cnt_males  571  34345  571  34345.0000
#> 51:  Total age_group3 cnt_males  424  24573  425  24630.9552
#> 52:  Total age_group4 cnt_males  195  11530  195  11530.0000
#> 53:  Total age_group5 cnt_males   84   5058   86   5178.4286
#> 54:  Total age_group6 cnt_males    7    471    5    336.4286
#> 55:   male      Total cnt_males 2296 136191 2294 136072.3667
#> 56:      m      Total cnt_males 2296 136191 2294 136072.3667
#> 57:   male age_group1 cnt_males 1015  60214 1014  60154.6759
#> 58:      m age_group1 cnt_males 1015  60214 1014  60154.6759
#> 59:   male       ag1a cnt_males 1015  60214 1014  60154.6759
#> 60:      m       ag1a cnt_males 1015  60214 1014  60154.6759
#> 61:   male age_group2 cnt_males  571  34345  571  34345.0000
#> 62:      m age_group2 cnt_males  571  34345  571  34345.0000
#> 63:   male       ag2a cnt_males  571  34345  571  34345.0000
#> 64:      m       ag2a cnt_males  571  34345  571  34345.0000
#> 65:   male age_group3 cnt_males  424  24573  425  24630.9552
#> 66:      m age_group3 cnt_males  424  24573  425  24630.9552
#> 67:   male age_group4 cnt_males  195  11530  195  11530.0000
#> 68:      m age_group4 cnt_males  195  11530  195  11530.0000
#> 69:   male age_group5 cnt_males   84   5058   86   5178.4286
#> 70:      m age_group5 cnt_males   84   5058   86   5178.4286
#> 71:   male age_group6 cnt_males    7    471    5    336.4286
#> 72:      m age_group6 cnt_males    7    471    5    336.4286
#> 73: female      Total cnt_males    0      0    0      0.0000
#> 74:      f      Total cnt_males    0      0    0      0.0000
#> 75: female age_group1 cnt_males    0      0    0      0.0000
#> 76:      f age_group1 cnt_males    0      0    0      0.0000
#> 77: female       ag1a cnt_males    0      0    0      0.0000
#> 78:      f       ag1a cnt_males    0      0    0      0.0000
#> 79: female age_group2 cnt_males    0      0    0      0.0000
#> 80:      f age_group2 cnt_males    0      0    0      0.0000
#> 81: female       ag2a cnt_males    0      0    0      0.0000
#> 82:      f       ag2a cnt_males    0      0    0      0.0000
#> 83: female age_group3 cnt_males    0      0    0      0.0000
#> 84:      f age_group3 cnt_males    0      0    0      0.0000
#> 85: female age_group4 cnt_males    0      0    0      0.0000
#> 86:      f age_group4 cnt_males    0      0    0      0.0000
#> 87: female age_group5 cnt_males    0      0    0      0.0000
#> 88:      f age_group5 cnt_males    0      0    0      0.0000
#> 89: female age_group6 cnt_males    0      0    0      0.0000
#> 90:      f age_group6 cnt_males    0      0    0      0.0000
#>        sex        age     vname  uwc     wc puwc         pwc

# numerical variables (positive variables using flex-function)
# we also write the config to a yaml file
f_yaml <- tempfile(fileext = ".yaml")

# create a ptable using functionality from the ptable-pkg
# a single ptable for all cells
ptab1 <- ptable::pt_ex_nums(parity = TRUE, separation = FALSE)

# a single ptab for all cells except for very small ones
ptab2 <- ptable::pt_ex_nums(parity = TRUE, separation = TRUE)

# different ptables for cells with even/odd number of contributors
# and very small cells
ptab3 <- ptable::pt_ex_nums(parity = FALSE, separation = TRUE)

p_nums1 <- ck_params_nums(
  ptab = ptab1,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.30, 0.03),
    epsilon = c(1, 0.5, 0.2),
    q = 3),
  mu_c = 2,
  same_key = FALSE,
  use_zero_rkeys = FALSE,
  path = f_yaml)
#> yaml configuration '/tmp/Rtmp2sERaX/file1a905e693c83.yaml' successfully written.

# we read the parameters from the yaml-file
p_nums1 <- ck_read_yaml(path = f_yaml)

# for variables with positive and negative values
p_nums2 <- ck_params_nums(
  ptab = ptab2,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.15, 0.02),
    epsilon = c(1, 0.4, 0.15),
    q = 3),
  mu_c = 2,
  same_key = FALSE)

# simple perturbation parameters (not using the flex-function approach)
p_nums3 <- ck_params_nums(
  ptab = ptab3,
  type = "mean",
  mult_params = ck_simpleparams(p = 0.25),
  mu_c = 2,
  same_key = FALSE)

# use `p_nums1` for all variables
tab$params_nums_set(p_nums1, c("savings", "income", "expend"))
#> --> setting perturbation parameters for variable 'savings'
#> --> setting perturbation parameters for variable 'income'
#> --> setting perturbation parameters for variable 'expend'

# use different parameters for variable `mixed`
tab$params_nums_set(p_nums2, v = "mixed")
#> --> setting perturbation parameters for variable 'mixed'

# identify sensitive cells to which extra protection (`mu_c`) is added.
tab$supp_p(v = "income", p = 85)
#> computing contributing indices | rawdata <--> table; this might take a while
#> p%-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_pq(v = "income", p = 85, q = 90)
#> computing contributing indices | rawdata <--> table; this might take a while
#> pq-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_nk(v = "income", n = 2, k = 90)
#> computing contributing indices | rawdata <--> table; this might take a while
#> nk-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_freq(v = "income", n = 14, weighted = FALSE)
#> freq-rule: 5 new sensitive cells (incl. duplicates) found (total: 5)
tab$supp_val(v = "income", n = 10000, weighted = TRUE)
#> val-rule: 0 new sensitive cells (incl. duplicates) found (total: 5)
tab$supp_cells(
  v = "income",
  inp = data.frame(
    sex = c("female", "female"),
    "age" = c("age_group1", "age_group3")
  )
)
#> cell-rule: 2 new sensitive cells (incl. duplicates) found (total: 7)

# perturb variables
tab$perturb(v = c("income", "savings"))
#> Numeric variable 'income' was perturbed.
#> Numeric variable 'savings' was perturbed.

# extract results
tab$numtab("income", mean_before_sum = TRUE)
#>        sex        age  vname      uws         ws          pws
#>  1:  Total      Total income 22952978 1372075860 1371930046.2
#>  2:  Total age_group1 income  9810547  588028354  587986472.2
#>  3:  Total       ag1a income  9810547  588028354  587986472.2
#>  4:  Total age_group2 income  5692119  341667352  341672947.7
#>  5:  Total       ag2a income  5692119  341667352  341672947.7
#>  6:  Total age_group3 income  4406946  262545002  262349882.8
#>  7:  Total age_group4 income  2133543  126089712  126027645.2
#>  8:  Total age_group5 income   848151   50500709   50469891.5
#>  9:  Total age_group6 income    61672    3244731    3301715.2
#> 10:   male      Total income 11262049  671284327  671323672.3
#> 11:      m      Total income 11262049  671284327  671323672.3
#> 12:   male age_group1 income  4877164  291187287  291184129.0
#> 13:      m age_group1 income  4877164  291187287  291184129.0
#> 14:   male       ag1a income  4877164  291187287  291184129.0
#> 15:      m       ag1a income  4877164  291187287  291184129.0
#> 16:   male age_group2 income  2811379  169998902  169798083.4
#> 17:      m age_group2 income  2811379  169998902  169798083.4
#> 18:   male       ag2a income  2811379  169998902  169798083.4
#> 19:      m       ag2a income  2811379  169998902  169798083.4
#> 20:   male age_group3 income  2168169  124952035  124954488.8
#> 21:      m age_group3 income  2168169  124952035  124954488.8
#> 22:   male age_group4 income   978510   58441082   58510931.4
#> 23:      m age_group4 income   978510   58441082   58510931.4
#> 24:   male age_group5 income   393134   24398281   24438201.6
#> 25:      m age_group5 income   393134   24398281   24438201.6
#> 26:   male age_group6 income    33693    2306740    2392343.0
#> 27:      m age_group6 income    33693    2306740    2392343.0
#> 28: female      Total income 11690929  700791533  700876622.8
#> 29:      f      Total income 11690929  700791533  700876622.8
#> 30: female age_group1 income  4933383  296841067  296725587.4
#> 31:      f age_group1 income  4933383  296841067  296725587.4
#> 32: female       ag1a income  4933383  296841067  296725587.4
#> 33:      f       ag1a income  4933383  296841067  296725587.4
#> 34: female age_group2 income  2880740  171668450  171618989.2
#> 35:      f age_group2 income  2880740  171668450  171618989.2
#> 36: female       ag2a income  2880740  171668450  171618989.2
#> 37:      f       ag2a income  2880740  171668450  171618989.2
#> 38: female age_group3 income  2238777  137592967  137620485.1
#> 39:      f age_group3 income  2238777  137592967  137620485.1
#> 40: female age_group4 income  1155033   67648630   67547391.3
#> 41:      f age_group4 income  1155033   67648630   67547391.3
#> 42: female age_group5 income   455017   26102428   26193951.7
#> 43:      f age_group5 income   455017   26102428   26193951.7
#> 44: female age_group6 income    27979     937991     934584.3
#> 45:      f age_group6 income    27979     937991     934584.3
#>        sex        age  vname      uws         ws          pws
tab$numtab("income", mean_before_sum = FALSE)
#>        sex        age  vname      uws         ws          pws
#>  1:  Total      Total income 22952978 1372075860 1372002951.1
#>  2:  Total age_group1 income  9810547  588028354  588007412.7
#>  3:  Total       ag1a income  9810547  588028354  588007412.7
#>  4:  Total age_group2 income  5692119  341667352  341670149.9
#>  5:  Total       ag2a income  5692119  341667352  341670149.9
#>  6:  Total age_group3 income  4406946  262545002  262447424.3
#>  7:  Total age_group4 income  2133543  126089712  126058674.8
#>  8:  Total age_group5 income   848151   50500709   50485297.9
#>  9:  Total age_group6 income    61672    3244731    3273099.1
#> 10:   male      Total income 11262049  671284327  671303999.4
#> 11:      m      Total income 11262049  671284327  671303999.4
#> 12:   male age_group1 income  4877164  291187287  291185708.0
#> 13:      m age_group1 income  4877164  291187287  291185708.0
#> 14:   male       ag1a income  4877164  291187287  291185708.0
#> 15:      m       ag1a income  4877164  291187287  291185708.0
#> 16:   male age_group2 income  2811379  169998902  169898463.0
#> 17:      m age_group2 income  2811379  169998902  169898463.0
#> 18:   male       ag2a income  2811379  169998902  169898463.0
#> 19:      m       ag2a income  2811379  169998902  169898463.0
#> 20:   male age_group3 income  2168169  124952035  124953261.9
#> 21:      m age_group3 income  2168169  124952035  124953261.9
#> 22:   male age_group4 income   978510   58441082   58475996.3
#> 23:      m age_group4 income   978510   58441082   58475996.3
#> 24:   male age_group5 income   393134   24398281   24418233.1
#> 25:      m age_group5 income   393134   24398281   24418233.1
#> 26:   male age_group6 income    33693    2306740    2349151.6
#> 27:      m age_group6 income    33693    2306740    2349151.6
#> 28: female      Total income 11690929  700791533  700834076.6
#> 29:      f      Total income 11690929  700791533  700834076.6
#> 30: female age_group1 income  4933383  296841067  296783321.6
#> 31:      f age_group1 income  4933383  296841067  296783321.6
#> 32: female       ag1a income  4933383  296841067  296783321.6
#> 33:      f       ag1a income  4933383  296841067  296783321.6
#> 34: female age_group2 income  2880740  171668450  171643717.8
#> 35:      f age_group2 income  2880740  171668450  171643717.8
#> 36: female       ag2a income  2880740  171668450  171643717.8
#> 37:      f       ag2a income  2880740  171668450  171643717.8
#> 38: female age_group3 income  2238777  137592967  137606725.4
#> 39:      f age_group3 income  2238777  137592967  137606725.4
#> 40: female age_group4 income  1155033   67648630   67597991.7
#> 41:      f age_group4 income  1155033   67648630   67597991.7
#> 42: female age_group5 income   455017   26102428   26148149.8
#> 43:      f age_group5 income   455017   26102428   26148149.8
#> 44: female age_group6 income    27979     937991     936286.1
#> 45:      f age_group6 income    27979     937991     936286.1
#>        sex        age  vname      uws         ws          pws
tab$numtab("savings")
#>        sex        age   vname     uws        ws          pws
#>  1:  Total      Total savings 2273532 135900260 135900243.93
#>  2:  Total age_group1 savings  982386  58649117  58648822.97
#>  3:  Total       ag1a savings  982386  58649117  58648822.97
#>  4:  Total age_group2 savings  552336  33401875  33406156.03
#>  5:  Total       ag2a savings  552336  33401875  33406156.03
#>  6:  Total age_group3 savings  437101  25729606  25723223.30
#>  7:  Total age_group4 savings  214661  12850055  12845357.59
#>  8:  Total age_group5 savings   80451   4939498   4939008.85
#>  9:  Total age_group6 savings    6597    330109    330052.89
#> 10:   male      Total savings 1159816  68952609  68953219.90
#> 11:      m      Total savings 1159816  68952609  68953219.90
#> 12:   male age_group1 savings  517660  30787167  30787263.16
#> 13:      m age_group1 savings  517660  30787167  30787263.16
#> 14:   male       ag1a savings  517660  30787167  30787263.16
#> 15:      m       ag1a savings  517660  30787167  30787263.16
#> 16:   male age_group2 savings  280923  16979795  16971970.58
#> 17:      m age_group2 savings  280923  16979795  16971970.58
#> 18:   male       ag2a savings  280923  16979795  16971970.58
#> 19:      m       ag2a savings  280923  16979795  16971970.58
#> 20:   male age_group3 savings  214970  12377280  12378128.71
#> 21:      m age_group3 savings  214970  12377280  12378128.71
#> 22:   male age_group4 savings   99420   5938588   5940472.81
#> 23:      m age_group4 savings   99420   5938588   5940472.81
#> 24:   male age_group5 savings   43233   2638540   2638753.34
#> 25:      m age_group5 savings   43233   2638540   2638753.34
#> 26:   male age_group6 savings    3610    231239    232067.82
#> 27:      m age_group6 savings    3610    231239    232067.82
#> 28: female      Total savings 1113716  66947651  66953929.21
#> 29:      f      Total savings 1113716  66947651  66953929.21
#> 30: female age_group1 savings  464726  27861950  27857185.26
#> 31:      f age_group1 savings  464726  27861950  27857185.26
#> 32: female       ag1a savings  464726  27861950  27857185.26
#> 33:      f       ag1a savings  464726  27861950  27857185.26
#> 34: female age_group2 savings  271413  16422080  16422866.91
#> 35:      f age_group2 savings  271413  16422080  16422866.91
#> 36: female       ag2a savings  271413  16422080  16422866.91
#> 37:      f       ag2a savings  271413  16422080  16422866.91
#> 38: female age_group3 savings  222131  13352326  13352570.34
#> 39:      f age_group3 savings  222131  13352326  13352570.34
#> 40: female age_group4 savings  115241   6911467   6904765.75
#> 41:      f age_group4 savings  115241   6911467   6904765.75
#> 42: female age_group5 savings   37218   2300958   2308756.31
#> 43:      f age_group5 savings   37218   2300958   2308756.31
#> 44: female age_group6 savings    2987     98870     98388.76
#> 45:      f age_group6 savings    2987     98870     98388.76
#>        sex        age   vname     uws        ws          pws

# results can be resetted, too
tab$reset_cntvars(v = "cnt_males")

# we can then set other parameters and perturb again
tab$params_cnts_set(val = p_cnts1, v = "cnt_males")
#> --> setting perturbation parameters for variable 'cnt_males'

tab$perturb(v = "cnt_males")
#> Count variable 'cnt_males' was perturbed.

# write results to a .csv file
tab$freqtab(
  v = c("total", "cnt_males"),
  path = file.path(tempdir(), "outtab.csv")
)
#> File '/tmp/Rtmp2sERaX/outtab.csv' successfully written to disk.
#> NULL

# show results containing weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))
#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 274065 4579 274005.1605
#>  2:  Total age_group1     total 1969 117705 1968 117645.2209
#>  3:  Total       ag1a     total 1969 117705 1968 117645.2209
#>  4:  Total age_group2     total 1143  68914 1144  68974.2922
#>  5:  Total       ag2a     total 1143  68914 1144  68974.2922
#>  6:  Total age_group3     total  864  51238  865  51297.3032
#>  7:  Total age_group4     total  423  25315  425  25434.6927
#>  8:  Total age_group5     total  168  10164  170  10285.0000
#>  9:  Total age_group6     total   13    729   13    729.0000
#> 10:   male      Total     total 2296 136191 2295 136131.6834
#> 11:      m      Total     total 2296 136191 2295 136131.6834
#> 12:   male age_group1     total 1015  60214 1015  60214.0000
#> 13:      m age_group1     total 1015  60214 1015  60214.0000
#> 14:   male       ag1a     total 1015  60214 1015  60214.0000
#> 15:      m       ag1a     total 1015  60214 1015  60214.0000
#> 16:   male age_group2     total  571  34345  571  34345.0000
#> 17:      m age_group2     total  571  34345  571  34345.0000
#> 18:   male       ag2a     total  571  34345  571  34345.0000
#> 19:      m       ag2a     total  571  34345  571  34345.0000
#> 20:   male age_group3     total  424  24573  425  24630.9552
#> 21:      m age_group3     total  424  24573  425  24630.9552
#> 22:   male age_group4     total  195  11530  195  11530.0000
#> 23:      m age_group4     total  195  11530  195  11530.0000
#> 24:   male age_group5     total   84   5058   85   5118.2143
#> 25:      m age_group5     total   84   5058   85   5118.2143
#> 26:   male age_group6     total    7    471    6    403.7143
#> 27:      m age_group6     total    7    471    6    403.7143
#> 28: female      Total     total 2284 137874 2286 137994.7303
#> 29:      f      Total     total 2284 137874 2286 137994.7303
#> 30: female age_group1     total  954  57491  955  57551.2631
#> 31:      f age_group1     total  954  57491  955  57551.2631
#> 32: female       ag1a     total  954  57491  955  57551.2631
#> 33:      f       ag1a     total  954  57491  955  57551.2631
#> 34: female age_group2     total  572  34569  572  34569.0000
#> 35:      f age_group2     total  572  34569  572  34569.0000
#> 36: female       ag2a     total  572  34569  572  34569.0000
#> 37:      f       ag2a     total  572  34569  572  34569.0000
#> 38: female age_group3     total  440  26665  438  26543.7955
#> 39:      f age_group3     total  440  26665  438  26543.7955
#> 40: female age_group4     total  228  13785  228  13785.0000
#> 41:      f age_group4     total  228  13785  228  13785.0000
#> 42: female age_group5     total   84   5106   83   5045.2143
#> 43:      f age_group5     total   84   5106   83   5045.2143
#> 44: female age_group6     total    6    258    5    215.0000
#> 45:      f age_group6     total    6    258    5    215.0000
#> 46:  Total      Total cnt_males 2296 136191 2295 136131.6834
#> 47:  Total age_group1 cnt_males 1015  60214 1015  60214.0000
#> 48:  Total       ag1a cnt_males 1015  60214 1015  60214.0000
#> 49:  Total age_group2 cnt_males  571  34345  571  34345.0000
#> 50:  Total       ag2a cnt_males  571  34345  571  34345.0000
#> 51:  Total age_group3 cnt_males  424  24573  425  24630.9552
#> 52:  Total age_group4 cnt_males  195  11530  195  11530.0000
#> 53:  Total age_group5 cnt_males   84   5058   85   5118.2143
#> 54:  Total age_group6 cnt_males    7    471    6    403.7143
#> 55:   male      Total cnt_males 2296 136191 2295 136131.6834
#> 56:      m      Total cnt_males 2296 136191 2295 136131.6834
#> 57:   male age_group1 cnt_males 1015  60214 1015  60214.0000
#> 58:      m age_group1 cnt_males 1015  60214 1015  60214.0000
#> 59:   male       ag1a cnt_males 1015  60214 1015  60214.0000
#> 60:      m       ag1a cnt_males 1015  60214 1015  60214.0000
#> 61:   male age_group2 cnt_males  571  34345  571  34345.0000
#> 62:      m age_group2 cnt_males  571  34345  571  34345.0000
#> 63:   male       ag2a cnt_males  571  34345  571  34345.0000
#> 64:      m       ag2a cnt_males  571  34345  571  34345.0000
#> 65:   male age_group3 cnt_males  424  24573  425  24630.9552
#> 66:      m age_group3 cnt_males  424  24573  425  24630.9552
#> 67:   male age_group4 cnt_males  195  11530  195  11530.0000
#> 68:      m age_group4 cnt_males  195  11530  195  11530.0000
#> 69:   male age_group5 cnt_males   84   5058   85   5118.2143
#> 70:      m age_group5 cnt_males   84   5058   85   5118.2143
#> 71:   male age_group6 cnt_males    7    471    6    403.7143
#> 72:      m age_group6 cnt_males    7    471    6    403.7143
#> 73: female      Total cnt_males    0      0    0      0.0000
#> 74:      f      Total cnt_males    0      0    0      0.0000
#> 75: female age_group1 cnt_males    0      0    0      0.0000
#> 76:      f age_group1 cnt_males    0      0    0      0.0000
#> 77: female       ag1a cnt_males    0      0    0      0.0000
#> 78:      f       ag1a cnt_males    0      0    0      0.0000
#> 79: female age_group2 cnt_males    0      0    0      0.0000
#> 80:      f age_group2 cnt_males    0      0    0      0.0000
#> 81: female       ag2a cnt_males    0      0    0      0.0000
#> 82:      f       ag2a cnt_males    0      0    0      0.0000
#> 83: female age_group3 cnt_males    0      0    0      0.0000
#> 84:      f age_group3 cnt_males    0      0    0      0.0000
#> 85: female age_group4 cnt_males    0      0    0      0.0000
#> 86:      f age_group4 cnt_males    0      0    0      0.0000
#> 87: female age_group5 cnt_males    0      0    0      0.0000
#> 88:      f age_group5 cnt_males    0      0    0      0.0000
#> 89: female age_group6 cnt_males    0      0    0      0.0000
#> 90:      f age_group6 cnt_males    0      0    0      0.0000
#>        sex        age     vname  uwc     wc puwc         pwc

# utility measures for a count variable
tab$measures_cnts(v = "total", exclude_zeros = TRUE)
#> $overview
#>    noise cnt        pct
#> 1:    -2   4 0.08888889
#> 2:    -1  11 0.24444444
#> 3:     0  17 0.37777778
#> 4:     1  11 0.24444444
#> 5:     2   2 0.04444444
#> 
#> $measures
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 1.000 0.000 0.009
#>  6:   Mean 0.756 0.016 0.034
#>  7: Median 1.000 0.001 0.015
#>  8:    Q60 1.000 0.001 0.016
#>  9:    Q70 1.000 0.002 0.024
#> 10:    Q80 1.000 0.006 0.050
#> 11:    Q90 2.000 0.012 0.068
#> 12:    Q95 2.000 0.143 0.196
#> 13:    Q99 2.000 0.167 0.213
#> 14:    Max 2.000 0.167 0.213
#> 
#> $cumdistr_d1
#>    cat cnt       pct
#> 1:   0  17 0.3777778
#> 2:   1  39 0.8666667
#> 3:   2  45 1.0000000
#> 
#> $cumdistr_d2
#>            cat cnt       pct
#> 1:    [0,0.02]  41 0.9111111
#> 2: (0.02,0.05]  41 0.9111111
#> 3:  (0.05,0.1]  41 0.9111111
#> 4:   (0.1,0.2]  45 1.0000000
#> 5:   (0.2,0.3]  45 1.0000000
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $cumdistr_d3
#>            cat cnt       pct
#> 1:    [0,0.02]  29 0.6444444
#> 2: (0.02,0.05]  36 0.8000000
#> 3:  (0.05,0.1]  41 0.9111111
#> 4:   (0.1,0.2]  43 0.9555556
#> 5:   (0.2,0.3]  45 1.0000000
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $false_zero
#> [1] 0
#> 
#> $false_nonzero
#> [1] 0
#> 
#> $exclude_zeros
#> [1] TRUE
#> 

# modifications for perturbed count variables
tab$mod_cnts()
#>         sex        age row_nr pert      ckey  countvar
#>   1:  Total      Total     14   -1 0.1147697     total
#>   2:  Total age_group1     14   -1 0.1513646     total
#>   3:  Total       ag1a     14   -1 0.1513646     total
#>   4:  Total age_group2     16    1 0.8906504     total
#>   5:  Total       ag2a     16    1 0.8906504     total
#>  ---                                                  
#> 131:      f age_group4     -1    0 0.0000000 cnt_males
#> 132: female age_group5     -1    0 0.0000000 cnt_males
#> 133:      f age_group5     -1    0 0.0000000 cnt_males
#> 134: female age_group6     -1    0 0.0000000 cnt_males
#> 135:      f age_group6     -1    0 0.0000000 cnt_males

# display a summary about utility measures
tab$summary()
#> ┌──────────────────────────────────────────────┐
#> │Utility measures for perturbed count variables│
#> └──────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#>          countvar Min Q10 Q20 Q30 Q40   Mean Median Q60 Q70 Q80 Q90 Q95 Q99 Max
#> 1:          total  -2  -1  -1   0   0  0.089      0 0.0   1   1   1   2   2   2
#> 2: cnt_highincome  -3  -2  -2  -1   0 -0.111      0 0.4   1   1   2   2   2   2
#> 3:      cnt_males  -1  -1   0   0   0  0.000      0 0.0   0   0   1   1   1   1
#> 
#> ── Distance-based measures ─────────────────────────────────────────────────────
#> ✔ Variable: 'total'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 1.000 0.000 0.009
#>  6:   Mean 0.756 0.016 0.034
#>  7: Median 1.000 0.001 0.015
#>  8:    Q60 1.000 0.001 0.016
#>  9:    Q70 1.000 0.002 0.024
#> 10:    Q80 1.000 0.006 0.050
#> 11:    Q90 2.000 0.012 0.068
#> 12:    Q95 2.000 0.143 0.196
#> 13:    Q99 2.000 0.167 0.213
#> 14:    Max 2.000 0.167 0.213
#> 
#> ✔ Variable: 'cnt_males'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 0.000 0.000 0.000
#>  6:   Mean 0.444 0.018 0.032
#>  7: Median 0.000 0.000 0.000
#>  8:    Q60 1.000 0.000 0.010
#>  9:    Q70 1.000 0.002 0.024
#> 10:    Q80 1.000 0.010 0.048
#> 11:    Q90 1.000 0.064 0.111
#> 12:    Q95 1.000 0.143 0.196
#> 13:    Q99 1.000 0.143 0.196
#> 14:    Max 1.000 0.143 0.196
#> 
#> ✔ Variable: 'cnt_highincome'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 1.000 0.004 0.034
#>  4:    Q30 1.000 0.010 0.053
#>  5:    Q40 1.000 0.011 0.061
#>  6:   Mean 1.325 0.032 0.090
#>  7: Median 1.000 0.015 0.072
#>  8:    Q60 2.000 0.020 0.100
#>  9:    Q70 2.000 0.035 0.116
#> 10:    Q80 2.000 0.040 0.134
#> 11:    Q90 2.000 0.133 0.183
#> 12:    Q95 2.000 0.143 0.250
#> 13:    Q99 2.610 0.143 0.267
#> 14:    Max 3.000 0.143 0.278
#> 
#> ┌──────────────────────────────────────────────────┐
#> │Utility measures for perturbed numerical variables│
#> └──────────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#>      vname        Min        Q10       Q20        Q30        Q40      Mean
#> 1:  expend        Inf         NA        NA         NA         NA       NaN
#> 2:  income -100438.98 -87710.188 -57745.40 -24732.192 -20941.284 -13936.01
#> 3: savings   -7824.42  -6701.245  -4764.74   -481.243    -32.086   -609.82
#> 4:   mixed        Inf         NA        NA         NA         NA       NaN
#>       Median      Q60       Q70       Q80       Q90       Q95       Q99
#> 1:        NA       NA        NA        NA        NA        NA        NA
#> 2: -1579.006 1226.893 13758.360 21635.318 42411.593 42543.593 45721.798
#> 3:    96.160  390.962   786.911   848.714  4281.035  6278.211  7798.314
#> 4:        NA       NA        NA        NA        NA        NA        NA
#>          Max
#> 1:      -Inf
#> 2: 45721.798
#> 3:  7798.314
#> 4:      -Inf
# }