An S4 class to represent a perturbed table

An S4 class to represent a perturbed table

Slots

tab

a data.table containing original and pertubed values. The following variables are always present:

UWC: unweighted counts

pUWC: perturbed unweighted counts

WC: weighted counts

pWC: perturbed weighted counts

WCavg: average weight for each cell

Additionally, for each numerical variable the perturbed variable named vName.pert is included in this table.

count_modifications

a data.table with 3 columns (row_indices, col_indices and applied_perturbation) that contains information for each cell, where the applied perturbation was extracted from the perturbation table (useful for debugging)

numvars_modifications

(list) containing for each numerical variable that was tabulated a list element with a data.table showing which values have been modified prior to tabulation.

dimVars

(character) vector of variable names used to define the table structure

numVars

(character) variable names of numeric variables that have been tabulated

countVars

(character) vector containing names of variables that have been tabulated using the frequency approach

by

(character) variable name (that must also be listed in countVars and that is used to create perturbed tables for numVars given the groups defined in by

is_weighted

(logical) TRUE if sampling weights have been used

type

(character) either "abs" or "destatis" depending on the format of the perturbation table that was used.