cellKey - an post-tabular perturbation method

This class allows to define statistical tables and perturb both count and numerical variables.

ck_setup(x, rkey, dims, w, countvars = NULL, params_cnts)

Arguments

x
an object coercible to a data.frame

rkey
either a column name within x referring to a variable containing record keys or a single integer(ish) number > 5 that referns to the number of digits for record keys that will be generated internally.

dims
a list containing slots for each variable that should be tabulated. Each slot consists should be created/modified using sdcHierarchies::hier_create(), sdcHierarchies::hier_add() and other functionality from package sdcHierarchies.

w
(character) a scalar character referring to a variable in x holding sampling weights. If w is NULL (the default), all weights are assumed to be 1

countvars
(character) an optional vector containing names of binary (0/1 coded) variables withing x that should be included in the problem instance. These variables can later be perturbed.

params_cnts
an object containing perturbation parameters for count variables that needs to be created with ck_params_cnts()

Format

R6::R6Class object.

Usage

For usage details see the Methods and Examples sections.

Methods

The following methods are available and can be used:

print(): prints information about the current table

summary(): some aggregated summary statistics about perturbed variables

perturb(v): Perturb a count- or magnitude variable. The method has the following arguments:

v: name(s) of count or magnitude variables that should be perturbed.

freqtab(v): get results from already perturbed count variables as a data.table. The required arguments are:

v: a vector of variable names for already perturbed count variables

type: a scalar character depending what variables should be available in the output. Allowed values are "both" (the default), "weighted" and "unweighted"

path: if not NULL, a scalar character defining a (relative or absolute) path to which the result table should be written. A csv file will be generated and .csv will be appended to the value provided.

This method returns a data.table containing all combinations of the dimensional variables in the first n columns. Additionally, the following columns are shown:

vname: name of the perturbed variable

uwc: unweighted counts (if type is "both" or "unweighted")

wc: weighted counts (if type is "both" or "weighted")

puwc: perturbed unweighted counts (if type is "both" or "unweighted")

pwc: perturbed weighted counts (if type is "both" or "weighted")

measures_cnts(v, exclude_zeros = TRUE): utility measures for perturbed count variables. The required arguments are:

v: name of a count variable for which utility measures should be computed.

exclude_zeros: should empty (zero) cells in the original values be excluded when computing distance measures This method returns a list containing a set of utility measures based on some distance functions. For a detailed description of the computed measures, see ck_cnt_measures()

mod_cnts(): returns a data.table containing modifications applied to count variables

params_cnts(val): returns or modifies perturbation parameters used for count variables. If val is not provided, the current perturbation object is returned, otherwise the current perturbation parameters are replaced with the ones provided (which need to be created using (ck_params_cnts())

Examples

x <- ck_create_testdata()

# create some 0/1 variables that should be perturbed later
x[, cnt_females := ifelse(sex == "male", 0, 1)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              64          25.00000           0
#>    2:   2530      28       1              42          25.00000           1
#>    3:   6920     550       1              95          25.00000           0
#>    4:   7960     870       1              82          25.00000           0
#>    5:   9030      20       2              66          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              67          16.66667           1
#> 4577:   1420     987    1000              51          16.66667           0
#> 4578:   8900     684    1000              92          16.66667           0
#> 4579:   3880     294    1000              40          16.66667           1
#> 4580:   4830     911    1000              39          16.66667           0

x[, cnt_males := ifelse(sex == "male", 1, 0)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              64          25.00000           0
#>    2:   2530      28       1              42          25.00000           1
#>    3:   6920     550       1              95          25.00000           0
#>    4:   7960     870       1              82          25.00000           0
#>    5:   9030      20       2              66          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              67          16.66667           1
#> 4577:   1420     987    1000              51          16.66667           0
#> 4578:   8900     684    1000              92          16.66667           0
#> 4579:   3880     294    1000              40          16.66667           1
#> 4580:   4830     911    1000              39          16.66667           0
#>       cnt_males
#>    1:         1
#>    2:         0
#>    3:         1
#>    4:         1
#>    5:         1
#>   ---          
#> 4576:         0
#> 4577:         1
#> 4578:         1
#> 4579:         0
#> 4580:         1

x[, cnt_highincome := ifelse(income >= 9000, 1, 0)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              64          25.00000           0
#>    2:   2530      28       1              42          25.00000           1
#>    3:   6920     550       1              95          25.00000           0
#>    4:   7960     870       1              82          25.00000           0
#>    5:   9030      20       2              66          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              67          16.66667           1
#> 4577:   1420     987    1000              51          16.66667           0
#> 4578:   8900     684    1000              92          16.66667           0
#> 4579:   3880     294    1000              40          16.66667           1
#> 4580:   4830     911    1000              39          16.66667           0
#>       cnt_males cnt_highincome
#>    1:         1              0
#>    2:         0              0
#>    3:         1              0
#>    4:         1              0
#>    5:         1              1
#>   ---                         
#> 4576:         0              0
#> 4577:         1              0
#> 4578:         1              0
#> 4579:         0              0
#> 4580:         1              0

# create record keys
x$rkey <- ck_generate_rkeys(dat = x)

# define required inputs

# hierarchy with some bogus codes
d_sex <- hier_create(root = "Total", nodes = c("male", "female"))
d_sex <- hier_add(d_sex, root = "female", "f")
d_sex <- hier_add(d_sex, root = "male", "m")

d_age <- hier_create(root = "Total", nodes = paste0("age_group", 1:6))
d_age <- hier_add(d_age, root = "age_group1", "ag1a")
d_age <- hier_add(d_age, root = "age_group2", "ag2a")

# create a named list
dims <- list(sex = d_sex, age = d_age)
w <- "sampling_weight"
countvars <- c("cnt_females", "cnt_males", "cnt_highincome")
rkey <- "rkey"

# perturbation parameters for count variables
params_cnts <- ck_params_cnts(
  D = 5,
  V = 3,
  js = 2,
  pstay = 0.5,
  optim = 1,
  mono = TRUE
)

# define the cell key object
tab <- ck_setup(
  x = x,
  rkey = 5,
  dims = dims,
  w = w,
  countvars = countvars,
  params_cnts = params_cnts
)

# show some information about this table instance
print(tab) # identical with tab$print()

#> ── Table Information ───────────────────────────────────────────────────────────
#> ✔ 45 cells in 2 dimensions ('sex', 'age')
#> ✔ weights: yes
#> ── Tabulated / Perturbed countvars ─────────────────────────────────────────────
#> ☐ 'total'
#> ☐ 'cnt_females'
#> ☐ 'cnt_males'
#> ☐ 'cnt_highincome'

# perturb a single variable
tab$perturb(v = "total")

#> Count variable 'total' was perturbed.

# modify the perturbation parameters
# create alternative perturbation parameters
params_cnts_alt <- ck_params_cnts(
  D = 8,
  V = 3,
  js = 2,
  pstay = 0.5,
  optim = 1,
  mono = TRUE
)
# we creplace the current parameters with the alternatives
# they will be used from now on
tab$params_cnts(params_cnts_alt)

#> Perturbation parameters for count variables were modified.

# multiple variables can be perturbed as well
tab$perturb(v = c("cnt_males", "cnt_highincome"))

#> Count variable 'cnt_males' was perturbed.

#> Count variable 'cnt_highincome' was perturbed.

# return results (weighted and unweighted)
tab$freqtab(v = c("total", "cnt_males"))

#>        sex        age     vname  uwc     wc puwc        pwc
#>  1:  Total      Total     total 4580 274707 4580 4639.97969
#>  2:  Total age_group1     total 1969 118802 1969 2029.33621
#>  3:  Total       ag1a     total 1969 118802 1969 2029.33621
#>  4:  Total age_group2     total 1143  68583 1143 1203.00262
#>  5:  Total       ag2a     total 1143  68583 1143 1203.00262
#>  6:  Total age_group3     total  864  51473  865  924.57523
#>  7:  Total age_group4     total  423  25121  425  484.38771
#>  8:  Total age_group5     total  168   9970  170  229.34524
#>  9:  Total age_group6     total   13    758   13   71.30769
#> 10: female      Total     total 2284 137362 2284 2344.14098
#> 11:      f      Total     total 2284 137362 2284 2344.14098
#> 12: female age_group1     total  954  57468  955 1015.23899
#> 13:      f age_group1     total  954  57468  955 1015.23899
#> 14: female       ag1a     total  954  57468  955 1015.23899
#> 15:      f       ag1a     total  954  57468  955 1015.23899
#> 16: female age_group2     total  572  34475  572  632.27098
#> 17:      f age_group2     total  572  34475  572  632.27098
#> 18: female       ag2a     total  572  34475  572  632.27098
#> 19:      f       ag2a     total  572  34475  572  632.27098
#> 20: female age_group3     total  440  26490  445  505.20455
#> 21:      f age_group3     total  440  26490  445  505.20455
#> 22: female age_group4     total  228  13373  228  286.65351
#> 23:      f age_group4     total  228  13373  228  286.65351
#> 24: female age_group5     total   84   5274   83  145.78571
#> 25:      f age_group5     total   84   5274   83  145.78571
#> 26: female age_group6     total    6    282    8   55.00000
#> 27:      f age_group6     total    6    282    8   55.00000
#> 28:   male      Total     total 2296 137345 2296 2355.81925
#> 29:      m      Total     total 2296 137345 2296 2355.81925
#> 30:   male age_group1     total 1015  61334 1016 1076.42759
#> 31:      m age_group1     total 1015  61334 1016 1076.42759
#> 32:   male       ag1a     total 1015  61334 1016 1076.42759
#> 33:      m       ag1a     total 1015  61334 1016 1076.42759
#> 34:   male age_group2     total  571  34108  570  629.73380
#> 35:      m age_group2     total  571  34108  570  629.73380
#> 36:   male       ag2a     total  571  34108  570  629.73380
#> 37:      m       ag2a     total  571  34108  570  629.73380
#> 38:   male age_group3     total  424  24983  425  483.92217
#> 39:      m age_group3     total  424  24983  425  483.92217
#> 40:   male age_group4     total  195  11748  195  255.24615
#> 41:      m age_group4     total  195  11748  195  255.24615
#> 42:   male age_group5     total   84   4696   84  139.90476
#> 43:      m age_group5     total   84   4696   84  139.90476
#> 44:   male age_group6     total    7    476    7   75.00000
#> 45:      m age_group6     total    7    476    7   75.00000
#> 46:  Total      Total cnt_males 2296 137345 2296 2355.81925
#> 47:  Total age_group1 cnt_males 1015  61334 1016 1076.42759
#> 48:  Total       ag1a cnt_males 1015  61334 1016 1076.42759
#> 49:  Total age_group2 cnt_males  571  34108  570  629.73380
#> 50:  Total       ag2a cnt_males  571  34108  570  629.73380
#> 51:  Total age_group3 cnt_males  424  24983  425  483.92217
#> 52:  Total age_group4 cnt_males  195  11748  195  255.24615
#> 53:  Total age_group5 cnt_males   84   4696   84  139.90476
#> 54:  Total age_group6 cnt_males    7    476    7   75.00000
#> 55: female      Total cnt_males    0      0    0    0.00000
#> 56:      f      Total cnt_males    0      0    0    0.00000
#> 57: female age_group1 cnt_males    0      0    0    0.00000
#> 58:      f age_group1 cnt_males    0      0    0    0.00000
#> 59: female       ag1a cnt_males    0      0    0    0.00000
#> 60:      f       ag1a cnt_males    0      0    0    0.00000
#> 61: female age_group2 cnt_males    0      0    0    0.00000
#> 62:      f age_group2 cnt_males    0      0    0    0.00000
#> 63: female       ag2a cnt_males    0      0    0    0.00000
#> 64:      f       ag2a cnt_males    0      0    0    0.00000
#> 65: female age_group3 cnt_males    0      0    0    0.00000
#> 66:      f age_group3 cnt_males    0      0    0    0.00000
#> 67: female age_group4 cnt_males    0      0    0    0.00000
#> 68:      f age_group4 cnt_males    0      0    0    0.00000
#> 69: female age_group5 cnt_males    0      0    0    0.00000
#> 70:      f age_group5 cnt_males    0      0    0    0.00000
#> 71: female age_group6 cnt_males    0      0    0    0.00000
#> 72:      f age_group6 cnt_males    0      0    0    0.00000
#> 73:   male      Total cnt_males 2296 137345 2296 2355.81925
#> 74:      m      Total cnt_males 2296 137345 2296 2355.81925
#> 75:   male age_group1 cnt_males 1015  61334 1016 1076.42759
#> 76:      m age_group1 cnt_males 1015  61334 1016 1076.42759
#> 77:   male       ag1a cnt_males 1015  61334 1016 1076.42759
#> 78:      m       ag1a cnt_males 1015  61334 1016 1076.42759
#> 79:   male age_group2 cnt_males  571  34108  570  629.73380
#> 80:      m age_group2 cnt_males  571  34108  570  629.73380
#> 81:   male       ag2a cnt_males  571  34108  570  629.73380
#> 82:      m       ag2a cnt_males  571  34108  570  629.73380
#> 83:   male age_group3 cnt_males  424  24983  425  483.92217
#> 84:      m age_group3 cnt_males  424  24983  425  483.92217
#> 85:   male age_group4 cnt_males  195  11748  195  255.24615
#> 86:      m age_group4 cnt_males  195  11748  195  255.24615
#> 87:   male age_group5 cnt_males   84   4696   84  139.90476
#> 88:      m age_group5 cnt_males   84   4696   84  139.90476
#> 89:   male age_group6 cnt_males    7    476    7   75.00000
#> 90:      m age_group6 cnt_males    7    476    7   75.00000
#>        sex        age     vname  uwc     wc puwc        pwc

# only unweighted and perturbed unweighted counts
tab$freqtab(v = c("total", "cnt_males"), type = "unweighted")

#>        sex        age     vname  uwc puwc
#>  1:  Total      Total     total 4580 4580
#>  2:  Total age_group1     total 1969 1969
#>  3:  Total       ag1a     total 1969 1969
#>  4:  Total age_group2     total 1143 1143
#>  5:  Total       ag2a     total 1143 1143
#>  6:  Total age_group3     total  864  865
#>  7:  Total age_group4     total  423  425
#>  8:  Total age_group5     total  168  170
#>  9:  Total age_group6     total   13   13
#> 10: female      Total     total 2284 2284
#> 11:      f      Total     total 2284 2284
#> 12: female age_group1     total  954  955
#> 13:      f age_group1     total  954  955
#> 14: female       ag1a     total  954  955
#> 15:      f       ag1a     total  954  955
#> 16: female age_group2     total  572  572
#> 17:      f age_group2     total  572  572
#> 18: female       ag2a     total  572  572
#> 19:      f       ag2a     total  572  572
#> 20: female age_group3     total  440  445
#> 21:      f age_group3     total  440  445
#> 22: female age_group4     total  228  228
#> 23:      f age_group4     total  228  228
#> 24: female age_group5     total   84   83
#> 25:      f age_group5     total   84   83
#> 26: female age_group6     total    6    8
#> 27:      f age_group6     total    6    8
#> 28:   male      Total     total 2296 2296
#> 29:      m      Total     total 2296 2296
#> 30:   male age_group1     total 1015 1016
#> 31:      m age_group1     total 1015 1016
#> 32:   male       ag1a     total 1015 1016
#> 33:      m       ag1a     total 1015 1016
#> 34:   male age_group2     total  571  570
#> 35:      m age_group2     total  571  570
#> 36:   male       ag2a     total  571  570
#> 37:      m       ag2a     total  571  570
#> 38:   male age_group3     total  424  425
#> 39:      m age_group3     total  424  425
#> 40:   male age_group4     total  195  195
#> 41:      m age_group4     total  195  195
#> 42:   male age_group5     total   84   84
#> 43:      m age_group5     total   84   84
#> 44:   male age_group6     total    7    7
#> 45:      m age_group6     total    7    7
#> 46:  Total      Total cnt_males 2296 2296
#> 47:  Total age_group1 cnt_males 1015 1016
#> 48:  Total       ag1a cnt_males 1015 1016
#> 49:  Total age_group2 cnt_males  571  570
#> 50:  Total       ag2a cnt_males  571  570
#> 51:  Total age_group3 cnt_males  424  425
#> 52:  Total age_group4 cnt_males  195  195
#> 53:  Total age_group5 cnt_males   84   84
#> 54:  Total age_group6 cnt_males    7    7
#> 55: female      Total cnt_males    0    0
#> 56:      f      Total cnt_males    0    0
#> 57: female age_group1 cnt_males    0    0
#> 58:      f age_group1 cnt_males    0    0
#> 59: female       ag1a cnt_males    0    0
#> 60:      f       ag1a cnt_males    0    0
#> 61: female age_group2 cnt_males    0    0
#> 62:      f age_group2 cnt_males    0    0
#> 63: female       ag2a cnt_males    0    0
#> 64:      f       ag2a cnt_males    0    0
#> 65: female age_group3 cnt_males    0    0
#> 66:      f age_group3 cnt_males    0    0
#> 67: female age_group4 cnt_males    0    0
#> 68:      f age_group4 cnt_males    0    0
#> 69: female age_group5 cnt_males    0    0
#> 70:      f age_group5 cnt_males    0    0
#> 71: female age_group6 cnt_males    0    0
#> 72:      f age_group6 cnt_males    0    0
#> 73:   male      Total cnt_males 2296 2296
#> 74:      m      Total cnt_males 2296 2296
#> 75:   male age_group1 cnt_males 1015 1016
#> 76:      m age_group1 cnt_males 1015 1016
#> 77:   male       ag1a cnt_males 1015 1016
#> 78:      m       ag1a cnt_males 1015 1016
#> 79:   male age_group2 cnt_males  571  570
#> 80:      m age_group2 cnt_males  571  570
#> 81:   male       ag2a cnt_males  571  570
#> 82:      m       ag2a cnt_males  571  570
#> 83:   male age_group3 cnt_males  424  425
#> 84:      m age_group3 cnt_males  424  425
#> 85:   male age_group4 cnt_males  195  195
#> 86:      m age_group4 cnt_males  195  195
#> 87:   male age_group5 cnt_males   84   84
#> 88:      m age_group5 cnt_males   84   84
#> 89:   male age_group6 cnt_males    7    7
#> 90:      m age_group6 cnt_males    7    7
#>        sex        age     vname  uwc puwc

# write to a file "outtab.csv" (.csv is automatically added to the path)

# NOT RUN {
tab$freqtab(v = c("total", "cnt_males"), path = "outtab")
# }

# only weighted and perturbed weighted counts
tab$freqtab(v = c("total", "cnt_males"), type = "weighted")

#>        sex        age     vname     wc        pwc
#>  1:  Total      Total     total 274707 4639.97969
#>  2:  Total age_group1     total 118802 2029.33621
#>  3:  Total       ag1a     total 118802 2029.33621
#>  4:  Total age_group2     total  68583 1203.00262
#>  5:  Total       ag2a     total  68583 1203.00262
#>  6:  Total age_group3     total  51473  924.57523
#>  7:  Total age_group4     total  25121  484.38771
#>  8:  Total age_group5     total   9970  229.34524
#>  9:  Total age_group6     total    758   71.30769
#> 10: female      Total     total 137362 2344.14098
#> 11:      f      Total     total 137362 2344.14098
#> 12: female age_group1     total  57468 1015.23899
#> 13:      f age_group1     total  57468 1015.23899
#> 14: female       ag1a     total  57468 1015.23899
#> 15:      f       ag1a     total  57468 1015.23899
#> 16: female age_group2     total  34475  632.27098
#> 17:      f age_group2     total  34475  632.27098
#> 18: female       ag2a     total  34475  632.27098
#> 19:      f       ag2a     total  34475  632.27098
#> 20: female age_group3     total  26490  505.20455
#> 21:      f age_group3     total  26490  505.20455
#> 22: female age_group4     total  13373  286.65351
#> 23:      f age_group4     total  13373  286.65351
#> 24: female age_group5     total   5274  145.78571
#> 25:      f age_group5     total   5274  145.78571
#> 26: female age_group6     total    282   55.00000
#> 27:      f age_group6     total    282   55.00000
#> 28:   male      Total     total 137345 2355.81925
#> 29:      m      Total     total 137345 2355.81925
#> 30:   male age_group1     total  61334 1076.42759
#> 31:      m age_group1     total  61334 1076.42759
#> 32:   male       ag1a     total  61334 1076.42759
#> 33:      m       ag1a     total  61334 1076.42759
#> 34:   male age_group2     total  34108  629.73380
#> 35:      m age_group2     total  34108  629.73380
#> 36:   male       ag2a     total  34108  629.73380
#> 37:      m       ag2a     total  34108  629.73380
#> 38:   male age_group3     total  24983  483.92217
#> 39:      m age_group3     total  24983  483.92217
#> 40:   male age_group4     total  11748  255.24615
#> 41:      m age_group4     total  11748  255.24615
#> 42:   male age_group5     total   4696  139.90476
#> 43:      m age_group5     total   4696  139.90476
#> 44:   male age_group6     total    476   75.00000
#> 45:      m age_group6     total    476   75.00000
#> 46:  Total      Total cnt_males 137345 2355.81925
#> 47:  Total age_group1 cnt_males  61334 1076.42759
#> 48:  Total       ag1a cnt_males  61334 1076.42759
#> 49:  Total age_group2 cnt_males  34108  629.73380
#> 50:  Total       ag2a cnt_males  34108  629.73380
#> 51:  Total age_group3 cnt_males  24983  483.92217
#> 52:  Total age_group4 cnt_males  11748  255.24615
#> 53:  Total age_group5 cnt_males   4696  139.90476
#> 54:  Total age_group6 cnt_males    476   75.00000
#> 55: female      Total cnt_males      0    0.00000
#> 56:      f      Total cnt_males      0    0.00000
#> 57: female age_group1 cnt_males      0    0.00000
#> 58:      f age_group1 cnt_males      0    0.00000
#> 59: female       ag1a cnt_males      0    0.00000
#> 60:      f       ag1a cnt_males      0    0.00000
#> 61: female age_group2 cnt_males      0    0.00000
#> 62:      f age_group2 cnt_males      0    0.00000
#> 63: female       ag2a cnt_males      0    0.00000
#> 64:      f       ag2a cnt_males      0    0.00000
#> 65: female age_group3 cnt_males      0    0.00000
#> 66:      f age_group3 cnt_males      0    0.00000
#> 67: female age_group4 cnt_males      0    0.00000
#> 68:      f age_group4 cnt_males      0    0.00000
#> 69: female age_group5 cnt_males      0    0.00000
#> 70:      f age_group5 cnt_males      0    0.00000
#> 71: female age_group6 cnt_males      0    0.00000
#> 72:      f age_group6 cnt_males      0    0.00000
#> 73:   male      Total cnt_males 137345 2355.81925
#> 74:      m      Total cnt_males 137345 2355.81925
#> 75:   male age_group1 cnt_males  61334 1076.42759
#> 76:      m age_group1 cnt_males  61334 1076.42759
#> 77:   male       ag1a cnt_males  61334 1076.42759
#> 78:      m       ag1a cnt_males  61334 1076.42759
#> 79:   male age_group2 cnt_males  34108  629.73380
#> 80:      m age_group2 cnt_males  34108  629.73380
#> 81:   male       ag2a cnt_males  34108  629.73380
#> 82:      m       ag2a cnt_males  34108  629.73380
#> 83:   male age_group3 cnt_males  24983  483.92217
#> 84:      m age_group3 cnt_males  24983  483.92217
#> 85:   male age_group4 cnt_males  11748  255.24615
#> 86:      m age_group4 cnt_males  11748  255.24615
#> 87:   male age_group5 cnt_males   4696  139.90476
#> 88:      m age_group5 cnt_males   4696  139.90476
#> 89:   male age_group6 cnt_males    476   75.00000
#> 90:      m age_group6 cnt_males    476   75.00000
#>        sex        age     vname     wc        pwc

# utility measures for a count variable
tab$measures_cnts(v = "total", exclude_zeros = TRUE)

#> $overview
#>    noise cnt        pct
#> 1:    -5   2 0.04444444
#> 2:    -2   4 0.08888889
#> 3:    -1  11 0.24444444
#> 4:     0  22 0.48888889
#> 5:     1   6 0.13333333
#> 
#> $measures
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 0.000 0.000 0.000
#>  6:   Mean 0.778 0.017 0.033
#>  7: Median 1.000 0.001 0.016
#>  8:    Q60 1.000 0.001 0.016
#>  9:    Q70 1.000 0.002 0.020
#> 10:    Q80 1.000 0.002 0.024
#> 11:    Q90 2.000 0.012 0.068
#> 12:    Q95 2.000 0.012 0.119
#> 13:    Q99 5.000 0.333 0.379
#> 14:    Max 5.000 0.333 0.379
#> 
#> $cumdistr_d1
#>    cat cnt       pct
#> 1:   0  22 0.4888889
#> 2:   1  39 0.8666667
#> 3:   2  43 0.9555556
#> 4:   5  45 1.0000000
#> 
#> $cumdistr_d2
#>            cat cnt       pct
#> 1:    [0,0.02]  43 0.9555556
#> 2: (0.02,0.05]  43 0.9555556
#> 3:  (0.05,0.1]  43 0.9555556
#> 4:   (0.1,0.2]  43 0.9555556
#> 5:   (0.2,0.3]  43 0.9555556
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $cumdistr_d3
#>            cat cnt       pct
#> 1:    [0,0.02]  31 0.6888889
#> 2: (0.02,0.05]  38 0.8444444
#> 3:  (0.05,0.1]  41 0.9111111
#> 4:   (0.1,0.2]  43 0.9555556
#> 5:   (0.2,0.3]  43 0.9555556
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $false_zero
#> [1] 0
#> 
#> $false_nonzero
#> [1] 0
#> 
#> $exclude_zeros
#> [1] TRUE
#> 

# modifications for perturbed count variables
tab$mod_cnts()

#>        sex        age row_nr pert    ckey       countvar
#>   1: Total      Total     61    0 0.31681          total
#>   2: Total age_group1     61    0 0.57772          total
#>   3: Total       ag1a     61    0 0.57772          total
#>   4: Total age_group2     61    0 0.50002          total
#>   5: Total       ag2a     61    0 0.50002          total
#>  ---                                                    
#> 131:     m age_group4    133    0 0.47569 cnt_highincome
#> 132:  male age_group5     74    4 0.97326 cnt_highincome
#> 133:     m age_group5     74    4 0.97326 cnt_highincome
#> 134:  male age_group6     -1    0 0.00000 cnt_highincome
#> 135:     m age_group6     -1    0 0.00000 cnt_highincome

# display a summary about utility measures
tab$summary()

#> ┌──────────────────────────────────────────────┐
#> │Utility measures for perturbed count variables│
#> └──────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#>          countvar Min Q10 Q20 Q30 Q40   Mean Median Q60 Q70 Q80 Q90 Q95 Q99 Max
#> 1:          total  -1  -1   0   0   0  0.511      0   0   1 1.0 2.0   2   5   5
#> 2:      cnt_males  -1  -1   0   0   0  0.067      0   0   0 0.2 1.0   1   1   1
#> 3: cnt_highincome  -3  -2  -1   0   0 -0.022      0   0   0 0.0 1.6   3   4   4
#> 
#> ── Distance-based measures ─────────────────────────────────────────────────────
#> ✔ Variable: 'total'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 0.000 0.000 0.000
#>  6:   Mean 0.778 0.017 0.033
#>  7: Median 1.000 0.001 0.016
#>  8:    Q60 1.000 0.001 0.016
#>  9:    Q70 1.000 0.002 0.020
#> 10:    Q80 1.000 0.002 0.024
#> 11:    Q90 2.000 0.012 0.068
#> 12:    Q95 2.000 0.012 0.119
#> 13:    Q99 5.000 0.333 0.379
#> 14:    Max 5.000 0.333 0.379
#> 
#> ✔ Variable: 'cnt_males'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 0.000 0.000 0.000
#>  6:   Mean 0.556 0.001 0.011
#>  7: Median 1.000 0.001 0.016
#>  8:    Q60 1.000 0.001 0.016
#>  9:    Q70 1.000 0.002 0.021
#> 10:    Q80 1.000 0.002 0.021
#> 11:    Q90 1.000 0.002 0.022
#> 12:    Q95 1.000 0.002 0.024
#> 13:    Q99 1.000 0.002 0.024
#> 14:    Max 1.000 0.002 0.024
#> 
#> ✔ Variable: 'cnt_highincome'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 0.000 0.000 0.000
#>  6:   Mean 0.925 0.040 0.077
#>  7: Median 0.000 0.000 0.000
#>  8:    Q60 1.000 0.005 0.034
#>  9:    Q70 1.000 0.015 0.062
#> 10:    Q80 2.000 0.035 0.134
#> 11:    Q90 3.000 0.060 0.176
#> 12:    Q95 3.050 0.098 0.260
#> 13:    Q99 4.000 0.571 0.671
#> 14:    Max 4.000 0.571 0.671
#>