cellKey - an post-tabular perturbation method

This class allows to define statistical tables and perturb both count and numerical variables.

ck_setup(x, rkey, dims, w, countvars = NULL, numvars = NULL)

Arguments

x
an object coercible to a data.frame

rkey
either a column name within x referring to a variable containing record keys or a single integer(ish) number > 5 that referns to the number of digits for record keys that will be generated internally.

dims
a list containing slots for each variable that should be tabulated. Each slot consists should be created/modified using sdcHierarchies::hier_create(), sdcHierarchies::hier_add() and other functionality from package sdcHierarchies.

w
(character) a scalar character referring to a variable in x holding sampling weights. If w is NULL (the default), all weights are assumed to be 1

countvars
(character) an optional vector containing names of binary (0/1 coded) variables withing x that should be included in the problem instance. These variables can later be perturbed.

numvars
(character) an optional vector of numerical variables that can later be tabulated.

Format

R6::R6Class object.

Usage

For usage details of the implemented methods, see the Methods and Examples sections.

Methods

The following methods are available and can be used:

print(): prints information about the current table

summary(): some aggregated summary statistics about perturbed variables

allvars(): a list with elements cntvars and numvars with each being a character vector containing the pre-defined count- and continuous variables in the current instance.

cntvars(): returns a character vector of available count variables.

numvars(): returns a character vector of available numeric variables.

supp_freq(v, n, weighted = TRUE): mark sensitive cells based on a minimal-frequency rule that refers to the (weighted or unweighted) number of contributors to a given cell. The required inputs are:

v: a single variable name of a continuous variable (see method numvars())

n: a number defining the threshold. All cells <= n are considered as unsafe.

weighted: if TRUE, the weighted number of contributors to a cell are compared to the threshold specified in n (default); else the unweighted number of contributors is used.

supp_val(v, n, weighted = TRUE): mark sensitive cells based on (weighted or unweighted) cell values that are below the given threshold n. The required inputs are:

v: a single variable name of a continuous variable (see method numvars())

n: a number defining the threshold. All cells <= n are considered as unsafe.

weighted: if TRUE, the weighted cell value of variable v is compared to the threshold specified in n (default); else the unweighted number is used.

supp_p(v, p): mark sensitive cells based on the p applied to positive-only variables and the required inputs are:

v: a single variable name of a continuous variable (see method numvars())

p: a number defining a percentage between 1 and 99.

supp_pq(v, p, q): mark sensitive cells based on the pq-rule rule. This rule can only be applied to positive-only variables and the required inputs are:

v: a single variable name of a continuous variable (see method numvars())

p: a number defining a percentage between 1 and 99.

q: a number defining a percentage between 1 and 99. This value must be larger than p.

supp_nk(v, n, k): mark sensitive cells based on the nk-dominance rule. This rule can only be applied to positive-only variables and the required inputs are:

v: a single variable name of a continuous variable (see method numvars())

n: an integerish number >= 2.

k: a number defining a percentage between 1 and 99. All cells to which the top n contributers contribute more than k% is considered unsafe

perturb(v): Perturb a count- or magnitude variable. The method has the following arguments:

v: name(s) of count or magnitude variables that should be perturbed.

freqtab(v, path): get results from already perturbed count variables as a data.table. The required arguments are:

v: a vector of variable names for count variables. If NULL (the default), the results are returned for all available count variables. For variables that have not yet perturbed, columns puwc and pwc are filled with NA.

path: if not NULL, a scalar character defining a (relative or absolute) path to which the result table should be written. A csv file will be generated and, if specified, path must have ".csv" as file-ending

This method returns a data.table containing all combinations of the dimensional variables in the first n columns. Additionally, the following columns are shown:

vname: name of the perturbed variable

uwc: unweighted counts

wc: weighted counts

puwc: perturbed unweighted counts or NA if vname was not yet perturbed

pwc: perturbed weighted counts or NA if vname was not yet perturbed

numtab(v, mean_before_sum = FALSE, path = NULL): get results from already perturbed continuous variables as a data.table. The required arguments are:

v: a vector of variable names of continuous variables. If NULL (the default), the results are returned for all available numeric variables.

mean_before_sum: (logical); if TRUE, the perturbed values are adjusted by a factor ((n+p))⁄n with n being the original weighted cellvalue and p the perturbed cell value. This makes sense if the the accuracy of the variable mean is considered more important than accuracy of sums of the variable. The default value is FALSE (no adjustment is done)

path: if not NULL, a scalar character defining a (relative or absolute) path to which the result table should be written. A csv file will be generated and, if specified, path must have ".csv" as file-ending

This method returns a data.table containing all combinations of the dimensional variables in the first n columns. Additionally, the following columns are shown:

vname: name of the perturbed variable

uws: unweighted sum of the given variable

ws: weighted cellsum

pws: perturbed weighted sum of the given cell or NA if vname has not not perturbed

measures_cnts(v, exclude_zeros = TRUE): utility measures for perturbed count variables. The required arguments are:

v: name of a count variable for which utility measures should be computed.

exclude_zeros: should empty (zero) cells in the original values be excluded when computing distance measures This method returns a list containing a set of utility measures based on some distance functions. For a detailed description of the computed measures, see ck_cnt_measures()

mod_cnts(): returns a data.table containing modifications applied to count variables

mod_nums(): returns a data.table containing modifications applied to numerical variables.

id: cell id

ckey: cell key used in the lookup step

cv: original weighted cell value

cv_pert: perturbed weighted cell value

x_delta: multiplier used when computing the actual perturbation amount

lookup: a character vector specifying in which block of the provided table the lookup happened.

params_cnts_get(): returns a named list in which each list-element contains the active perturbation parameters for the specific count variable defined by the list-name.

params_cnts_set(val, v = NULL): allows to set perturbation parameters for count variables. The following arguments are expected:

val: a perturbation object created with ck_params_cnts()

v: a character vector (or NULL). If NULL (the default), the perturbation parameters provided in val are set for all count variables; otherwise one may specify the names of the count variables for which the parameters should be set.

params_nums_get(): returns a named list in which each list-element contains the active perturbation parameters for the specific continuous variable defined by the list-name.

params_nums_set(val, v = NULL): allows to set perturbation parameters for continuous variables. The following arguments are expected:

val: a perturbation object created with ck_params_nums()

v: a character vector (or NULL). If NULL (the default), the perturbation parameters provided in val are set for all continuous variables; otherwise one may specify the names of the numeric variables for which the parameters should be set.

reset_cntvars(v = NULL): allows to reset results and parameters for already perturbed count variables specified in v. If v equals NULL (the default), the results are reset for all perturbed count variables.

reset_numvars(v = NULL): allows to reset results and parameters for already perturbed numerical variables specified in v. If v equals NULL (the default), the results are reset for all perturbed numerical variables.

reset_allvars(): allows to reset results and parameters for all already perturbed variables.

Examples

x <- ck_create_testdata()

# create some 0/1 variables that should be perturbed later
x[, cnt_females := ifelse(sex == "male", 0, 1)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              64          25.00000           0
#>    2:   2530      28       1              42          25.00000           1
#>    3:   6920     550       1              95          25.00000           0
#>    4:   7960     870       1              82          25.00000           0
#>    5:   9030      20       2              66          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              67          16.66667           1
#> 4577:   1420     987    1000              51          16.66667           0
#> 4578:   8900     684    1000              92          16.66667           0
#> 4579:   3880     294    1000              40          16.66667           1
#> 4580:   4830     911    1000              39          16.66667           0

x[, cnt_males := ifelse(sex == "male", 1, 0)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              64          25.00000           0
#>    2:   2530      28       1              42          25.00000           1
#>    3:   6920     550       1              95          25.00000           0
#>    4:   7960     870       1              82          25.00000           0
#>    5:   9030      20       2              66          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              67          16.66667           1
#> 4577:   1420     987    1000              51          16.66667           0
#> 4578:   8900     684    1000              92          16.66667           0
#> 4579:   3880     294    1000              40          16.66667           1
#> 4580:   4830     911    1000              39          16.66667           0
#>       cnt_males
#>    1:         1
#>    2:         0
#>    3:         1
#>    4:         1
#>    5:         1
#>   ---          
#> 4576:         0
#> 4577:         1
#> 4578:         1
#> 4579:         0
#> 4580:         1

x[, cnt_highincome := ifelse(income >= 9000, 1, 0)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              64          25.00000           0
#>    2:   2530      28       1              42          25.00000           1
#>    3:   6920     550       1              95          25.00000           0
#>    4:   7960     870       1              82          25.00000           0
#>    5:   9030      20       2              66          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              67          16.66667           1
#> 4577:   1420     987    1000              51          16.66667           0
#> 4578:   8900     684    1000              92          16.66667           0
#> 4579:   3880     294    1000              40          16.66667           1
#> 4580:   4830     911    1000              39          16.66667           0
#>       cnt_males cnt_highincome
#>    1:         1              0
#>    2:         0              0
#>    3:         1              0
#>    4:         1              0
#>    5:         1              1
#>   ---                         
#> 4576:         0              0
#> 4577:         1              0
#> 4578:         1              0
#> 4579:         0              0
#> 4580:         1              0

# a variable with positive and negative contributions
x[, mixed := sample(-10:10, nrow(x), replace = TRUE)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              64          25.00000           0
#>    2:   2530      28       1              42          25.00000           1
#>    3:   6920     550       1              95          25.00000           0
#>    4:   7960     870       1              82          25.00000           0
#>    5:   9030      20       2              66          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              67          16.66667           1
#> 4577:   1420     987    1000              51          16.66667           0
#> 4578:   8900     684    1000              92          16.66667           0
#> 4579:   3880     294    1000              40          16.66667           1
#> 4580:   4830     911    1000              39          16.66667           0
#>       cnt_males cnt_highincome mixed
#>    1:         1              0    -6
#>    2:         0              0     2
#>    3:         1              0    -2
#>    4:         1              0     0
#>    5:         1              1     8
#>   ---                               
#> 4576:         0              0    -2
#> 4577:         1              0     5
#> 4578:         1              0    -4
#> 4579:         0              0     7
#> 4580:         1              0    -9

# create record keys
x$rkey <- ck_generate_rkeys(dat = x)

# define required inputs

# hierarchy with some bogus codes
d_sex <- hier_create(root = "Total", nodes = c("male", "female"))
d_sex <- hier_add(d_sex, root = "female", "f")
d_sex <- hier_add(d_sex, root = "male", "m")

d_age <- hier_create(root = "Total", nodes = paste0("age_group", 1:6))
d_age <- hier_add(d_age, root = "age_group1", "ag1a")
d_age <- hier_add(d_age, root = "age_group2", "ag2a")

# define the cell key object
countvars <- c("cnt_females", "cnt_males", "cnt_highincome")
numvars <- c("expend", "income", "savings", "mixed")
tab <- ck_setup(
  x = x,
  rkey = "rkey",
  dims = list(sex = d_sex, age = d_age),
  w = "sampling_weight",
  countvars = countvars,
  numvars = numvars)

# show some information about this table instance
tab$print() # identical with print(tab)

#> ── Table Information ───────────────────────────────────────────────────────────
#> ✔ 45 cells in 2 dimensions ('sex', 'age')
#> ✔ weights: yes
#> ── Tabulated / Perturbed countvars ─────────────────────────────────────────────

#> Warning: numerical expression has 4 elements: only the first used

#> ☐ 'total'
#> ── Tabulated / Perturbed numvars ───────────────────────────────────────────────
#> ☐ 'expend'
#> ☐ 'income'
#> ☐ 'savings'
#> ☐ 'mixed'

# which variables have been defined?
tab$allvars()

#> $cntvars
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"
#> 
#> $numvars
#> [1] "expend"  "income"  "savings" "mixed"  
#> 

# count variables
tab$cntvars()

#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"

# continuous variables
tab$numvars()

#> [1] "expend"  "income"  "savings" "mixed"  

# create perturbation parameters for "total" variable and
# write to yaml-file

# create a ptable using functionality from the ptable-pkg
f_yaml <- tempfile(fileext = ".yaml")
p_cnts1 <- ck_params_cnts(
  ptab = ptable::pt_ex_cnts(),
  path = f_yaml)

#> yaml configuration '/tmp/RtmpMxxMGp/file3ad030d9b5d1.yaml' successfully written.

# read parameters from yaml-file and set them for variable `"total"`
p_cnts1 <- ck_read_yaml(path = f_yaml)

tab$params_cnts_set(val = p_cnts1, v = "total")

#> --> setting perturbation parameters for variable 'total'

# create alternative perturbation parameters by specifying parameters
para2 <- ptable::pt_create_pParams(
  D = 8, V = 3, js = 2, table = "cnts")

#> Since type of table is frequency table (argument table is set to 'cnts'), the input parameters 'step' and 'icat' will be ignored.

p_cnts2 <- ck_params_cnts(ptab = para2)

# use these ptable it for the remaining variables
tab$params_cnts_set(val = p_cnts2, v = countvars)

#> --> setting perturbation parameters for variable 'cnt_females'

#> --> setting perturbation parameters for variable 'cnt_males'

#> --> setting perturbation parameters for variable 'cnt_highincome'

# perturb a variable
tab$perturb(v = "total")

#> Count variable 'total' was perturbed.

# multiple variables can be perturbed as well
tab$perturb(v = c("cnt_males", "cnt_highincome"))

#> Count variable 'cnt_males' was perturbed.

#> Count variable 'cnt_highincome' was perturbed.

# return weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))

#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 274707 4579 274647.0203
#>  2:  Total age_group1     total 1969 118802 1968 118741.6638
#>  3:  Total       ag1a     total 1969 118802 1968 118741.6638
#>  4:  Total age_group2     total 1143  68583 1143  68583.0000
#>  5:  Total       ag2a     total 1143  68583 1143  68583.0000
#>  6:  Total age_group3     total  864  51473  864  51473.0000
#>  7:  Total age_group4     total  423  25121  423  25121.0000
#>  8:  Total age_group5     total  168   9970  167   9910.6548
#>  9:  Total age_group6     total   13    758   12    699.6923
#> 10: female      Total     total 2284 137362 2286 137482.2820
#> 11:      f      Total     total 2284 137362 2286 137482.2820
#> 12: female age_group1     total  954  57468  955  57528.2390
#> 13:      f age_group1     total  954  57468  955  57528.2390
#> 14: female       ag1a     total  954  57468  955  57528.2390
#> 15:      f       ag1a     total  954  57468  955  57528.2390
#> 16: female age_group2     total  572  34475  573  34535.2710
#> 17:      f age_group2     total  572  34475  573  34535.2710
#> 18: female       ag2a     total  572  34475  573  34535.2710
#> 19:      f       ag2a     total  572  34475  573  34535.2710
#> 20: female age_group3     total  440  26490  438  26369.5909
#> 21:      f age_group3     total  440  26490  438  26369.5909
#> 22: female age_group4     total  228  13373  228  13373.0000
#> 23:      f age_group4     total  228  13373  228  13373.0000
#> 24: female age_group5     total   84   5274   84   5274.0000
#> 25:      f age_group5     total   84   5274   84   5274.0000
#> 26: female age_group6     total    6    282    5    235.0000
#> 27:      f age_group6     total    6    282    5    235.0000
#> 28:   male      Total     total 2296 137345 2295 137285.1807
#> 29:      m      Total     total 2296 137345 2295 137285.1807
#> 30:   male age_group1     total 1015  61334 1014  61273.5724
#> 31:      m age_group1     total 1015  61334 1014  61273.5724
#> 32:   male       ag1a     total 1015  61334 1014  61273.5724
#> 33:      m       ag1a     total 1015  61334 1014  61273.5724
#> 34:   male age_group2     total  571  34108  572  34167.7338
#> 35:      m age_group2     total  571  34108  572  34167.7338
#> 36:   male       ag2a     total  571  34108  572  34167.7338
#> 37:      m       ag2a     total  571  34108  572  34167.7338
#> 38:   male age_group3     total  424  24983  424  24983.0000
#> 39:      m age_group3     total  424  24983  424  24983.0000
#> 40:   male age_group4     total  195  11748  197  11868.4923
#> 41:      m age_group4     total  195  11748  197  11868.4923
#> 42:   male age_group5     total   84   4696   85   4751.9048
#> 43:      m age_group5     total   84   4696   85   4751.9048
#> 44:   male age_group6     total    7    476    5    340.0000
#> 45:      m age_group6     total    7    476    5    340.0000
#> 46:  Total      Total cnt_males 2296 137345 2295 137285.1807
#> 47:  Total age_group1 cnt_males 1015  61334 1014  61273.5724
#> 48:  Total       ag1a cnt_males 1015  61334 1014  61273.5724
#> 49:  Total age_group2 cnt_males  571  34108  572  34167.7338
#> 50:  Total       ag2a cnt_males  571  34108  572  34167.7338
#> 51:  Total age_group3 cnt_males  424  24983  424  24983.0000
#> 52:  Total age_group4 cnt_males  195  11748  198  11928.7385
#> 53:  Total age_group5 cnt_males   84   4696   85   4751.9048
#> 54:  Total age_group6 cnt_males    7    476    4    272.0000
#> 55: female      Total cnt_males    0      0    0      0.0000
#> 56:      f      Total cnt_males    0      0    0      0.0000
#> 57: female age_group1 cnt_males    0      0    0      0.0000
#> 58:      f age_group1 cnt_males    0      0    0      0.0000
#> 59: female       ag1a cnt_males    0      0    0      0.0000
#> 60:      f       ag1a cnt_males    0      0    0      0.0000
#> 61: female age_group2 cnt_males    0      0    0      0.0000
#> 62:      f age_group2 cnt_males    0      0    0      0.0000
#> 63: female       ag2a cnt_males    0      0    0      0.0000
#> 64:      f       ag2a cnt_males    0      0    0      0.0000
#> 65: female age_group3 cnt_males    0      0    0      0.0000
#> 66:      f age_group3 cnt_males    0      0    0      0.0000
#> 67: female age_group4 cnt_males    0      0    0      0.0000
#> 68:      f age_group4 cnt_males    0      0    0      0.0000
#> 69: female age_group5 cnt_males    0      0    0      0.0000
#> 70:      f age_group5 cnt_males    0      0    0      0.0000
#> 71: female age_group6 cnt_males    0      0    0      0.0000
#> 72:      f age_group6 cnt_males    0      0    0      0.0000
#> 73:   male      Total cnt_males 2296 137345 2295 137285.1807
#> 74:      m      Total cnt_males 2296 137345 2295 137285.1807
#> 75:   male age_group1 cnt_males 1015  61334 1014  61273.5724
#> 76:      m age_group1 cnt_males 1015  61334 1014  61273.5724
#> 77:   male       ag1a cnt_males 1015  61334 1014  61273.5724
#> 78:      m       ag1a cnt_males 1015  61334 1014  61273.5724
#> 79:   male age_group2 cnt_males  571  34108  572  34167.7338
#> 80:      m age_group2 cnt_males  571  34108  572  34167.7338
#> 81:   male       ag2a cnt_males  571  34108  572  34167.7338
#> 82:      m       ag2a cnt_males  571  34108  572  34167.7338
#> 83:   male age_group3 cnt_males  424  24983  424  24983.0000
#> 84:      m age_group3 cnt_males  424  24983  424  24983.0000
#> 85:   male age_group4 cnt_males  195  11748  198  11928.7385
#> 86:      m age_group4 cnt_males  195  11748  198  11928.7385
#> 87:   male age_group5 cnt_males   84   4696   85   4751.9048
#> 88:      m age_group5 cnt_males   84   4696   85   4751.9048
#> 89:   male age_group6 cnt_males    7    476    4    272.0000
#> 90:      m age_group6 cnt_males    7    476    4    272.0000
#>        sex        age     vname  uwc     wc puwc         pwc

# numerical variables (positive variables using flex-function)
# we also write the config to a yaml file
f_yaml <- tempfile(fileext = ".yaml")

# create a ptable using functionality from the ptable-pkg
# a single ptable for all cells
ptab1 <- ptable::pt_ex_nums(parity = TRUE, separation = FALSE)

# a single ptab for all cells except for very small ones
ptab2 <- ptable::pt_ex_nums(parity = TRUE, separation = TRUE)

# different ptables for cells with even/odd number of contributors
# and very small cells
ptab3 <- ptable::pt_ex_nums(parity = FALSE, separation = TRUE)

p_nums1 <- ck_params_nums(
  ptab = ptab1,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.30, 0.03),
    epsilon = c(1, 0.5, 0.2),
    q = 3),
  mu_c = 2,
  same_key = FALSE,
  use_zero_rkeys = FALSE,
  path = f_yaml)

#> yaml configuration '/tmp/RtmpMxxMGp/file3ad0487a9812.yaml' successfully written.

# we read the parameters from the yaml-file
p_nums1 <- ck_read_yaml(path = f_yaml)

# for variables with positive and negative values
p_nums2 <- ck_params_nums(
  ptab = ptab2,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.15, 0.02),
    epsilon = c(1, 0.4, 0.15),
    q = 3),
  mu_c = 2,
  same_key = FALSE)

# simple perturbation parameters (not using the flex-function approach)
p_nums3 <- ck_params_nums(
  ptab = ptab3,
  type = "mean",
  mult_params = ck_simpleparams(p = 0.25),
  mu_c = 2,
  same_key = FALSE)

# use `p_nums1` for all variables
tab$params_nums_set(p_nums1, c("savings", "income", "expend"))

#> --> setting perturbation parameters for variable 'savings'

#> --> setting perturbation parameters for variable 'income'

#> --> setting perturbation parameters for variable 'expend'

# use different parameters for variable `mixed`
tab$params_nums_set(p_nums2, v = "mixed")

#> --> setting perturbation parameters for variable 'mixed'

# identify sensitive cells to which extra protection (`mu_c`) is added.
tab$supp_p(v = "income", p = 85)

#> p%-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)

tab$supp_pq(v = "income", p = 85, q = 90)

#> pq-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)

tab$supp_nk(v = "income", n = 2, k = 90)

#> nk-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)

tab$supp_freq(v = "income", n = 14, weighted = FALSE)

#> freq-rule: 5 new sensitive cells (incl. duplicates) found (total: 5)

tab$supp_val(v = "income", n = 10000, weighted = TRUE)

#> val-rule: 0 new sensitive cells (incl. duplicates) found (total: 5)

# perturb variables
tab$perturb(v = c("income", "savings"))

#> Numeric variable 'income' was perturbed.

#> Numeric variable 'savings' was perturbed.

# extract results
tab$numtab("income", mean_before_sum = TRUE)

#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1377005797 1377073424
#>  2:  Total age_group1 income  9810547  590371940  590365458
#>  3:  Total       ag1a income  9810547  590371940  590365458
#>  4:  Total age_group2 income  5692119  340327551  340309861
#>  5:  Total       ag2a income  5692119  340327551  340309861
#>  6:  Total age_group3 income  4406946  263870748  263887794
#>  7:  Total age_group4 income  2133543  128635475  128657949
#>  8:  Total age_group5 income   848151   50513336   50541476
#>  9:  Total age_group6 income    61672    3286747    3319040
#> 10: female      Total income 11690929  702755232  702797498
#> 11:      f      Total income 11690929  702755232  702797498
#> 12: female age_group1 income  4933383  296150572  296123078
#> 13:      f age_group1 income  4933383  296150572  296123078
#> 14: female       ag1a income  4933383  296150572  296123078
#> 15:      f       ag1a income  4933383  296150572  296123078
#> 16: female age_group2 income  2880740  173368984  173388146
#> 17:      f age_group2 income  2880740  173368984  173388146
#> 18: female       ag2a income  2880740  173368984  173388146
#> 19:      f       ag2a income  2880740  173368984  173388146
#> 20: female age_group3 income  2238777  134883636  134851158
#> 21:      f age_group3 income  2238777  134883636  134851158
#> 22: female age_group4 income  1155033   68276741   68178323
#> 23:      f age_group4 income  1155033   68276741   68178323
#> 24: female age_group5 income   455017   28913759   28940085
#> 25:      f age_group5 income   455017   28913759   28940085
#> 26: female age_group6 income    27979    1161540    1267084
#> 27:      f age_group6 income    27979    1161540    1267084
#> 28:   male      Total income 11262049  674250565  674338251
#> 29:      m      Total income 11262049  674250565  674338251
#> 30:   male age_group1 income  4877164  294221368  294071892
#> 31:      m age_group1 income  4877164  294221368  294071892
#> 32:   male       ag1a income  4877164  294221368  294071892
#> 33:      m       ag1a income  4877164  294221368  294071892
#> 34:   male age_group2 income  2811379  166958567  167049714
#> 35:      m age_group2 income  2811379  166958567  167049714
#> 36:   male       ag2a income  2811379  166958567  167049714
#> 37:      m       ag2a income  2811379  166958567  167049714
#> 38:   male age_group3 income  2168169  128987112  128939361
#> 39:      m age_group3 income  2168169  128987112  128939361
#> 40:   male age_group4 income   978510   60358734   60252043
#> 41:      m age_group4 income   978510   60358734   60252043
#> 42:   male age_group5 income   393134   21599577   21610203
#> 43:      m age_group5 income   393134   21599577   21610203
#> 44:   male age_group6 income    33693    2125207    2195229
#> 45:      m age_group6 income    33693    2125207    2195229
#>        sex        age  vname      uws         ws        pws

tab$numtab("income", mean_before_sum = FALSE)

#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1377005797 1377039610
#>  2:  Total age_group1 income  9810547  590371940  590368699
#>  3:  Total       ag1a income  9810547  590371940  590368699
#>  4:  Total age_group2 income  5692119  340327551  340318706
#>  5:  Total       ag2a income  5692119  340327551  340318706
#>  6:  Total age_group3 income  4406946  263870748  263879271
#>  7:  Total age_group4 income  2133543  128635475  128646712
#>  8:  Total age_group5 income   848151   50513336   50527404
#>  9:  Total age_group6 income    61672    3286747    3302854
#> 10: female      Total income 11690929  702755232  702776364
#> 11:      f      Total income 11690929  702755232  702776364
#> 12: female age_group1 income  4933383  296150572  296136825
#> 13:      f age_group1 income  4933383  296150572  296136825
#> 14: female       ag1a income  4933383  296150572  296136825
#> 15:      f       ag1a income  4933383  296150572  296136825
#> 16: female age_group2 income  2880740  173368984  173378565
#> 17:      f age_group2 income  2880740  173368984  173378565
#> 18: female       ag2a income  2880740  173368984  173378565
#> 19:      f       ag2a income  2880740  173368984  173378565
#> 20: female age_group3 income  2238777  134883636  134867396
#> 21:      f age_group3 income  2238777  134883636  134867396
#> 22: female age_group4 income  1155033   68276741   68227514
#> 23:      f age_group4 income  1155033   68276741   68227514
#> 24: female age_group5 income   455017   28913759   28926919
#> 25:      f age_group5 income   455017   28913759   28926919
#> 26: female age_group6 income    27979    1161540    1213165
#> 27:      f age_group6 income    27979    1161540    1213165
#> 28:   male      Total income 11262049  674250565  674294406
#> 29:      m      Total income 11262049  674250565  674294406
#> 30:   male age_group1 income  4877164  294221368  294146620
#> 31:      m age_group1 income  4877164  294221368  294146620
#> 32:   male       ag1a income  4877164  294221368  294146620
#> 33:      m       ag1a income  4877164  294221368  294146620
#> 34:   male age_group2 income  2811379  166958567  167004134
#> 35:      m age_group2 income  2811379  166958567  167004134
#> 36:   male       ag2a income  2811379  166958567  167004134
#> 37:      m       ag2a income  2811379  166958567  167004134
#> 38:   male age_group3 income  2168169  128987112  128963234
#> 39:      m age_group3 income  2168169  128987112  128963234
#> 40:   male age_group4 income   978510   60358734   60305365
#> 41:      m age_group4 income   978510   60358734   60305365
#> 42:   male age_group5 income   393134   21599577   21604889
#> 43:      m age_group5 income   393134   21599577   21604889
#> 44:   male age_group6 income    33693    2125207    2159934
#> 45:      m age_group6 income    33693    2125207    2159934
#>        sex        age  vname      uws         ws        pws

tab$numtab("savings")

#>        sex        age   vname     uws        ws         pws
#>  1:  Total      Total savings 2273532 136873818 136875564.6
#>  2:  Total age_group1 savings  982386  59422265  59425129.7
#>  3:  Total       ag1a savings  982386  59422265  59425129.7
#>  4:  Total age_group2 savings  552336  33246815  33246439.1
#>  5:  Total       ag2a savings  552336  33246815  33246439.1
#>  6:  Total age_group3 savings  437101  26180709  26172177.1
#>  7:  Total age_group4 savings  214661  12776589  12778694.5
#>  8:  Total age_group5 savings   80451   4886105   4888072.2
#>  9:  Total age_group6 savings    6597    361335    361434.3
#> 10: female      Total savings 1113716  67211781  67215028.9
#> 11:      f      Total savings 1113716  67211781  67215028.9
#> 12: female age_group1 savings  464726  27991793  27991593.7
#> 13:      f age_group1 savings  464726  27991793  27991593.7
#> 14: female       ag1a savings  464726  27991793  27991593.7
#> 15:      f       ag1a savings  464726  27991793  27991593.7
#> 16: female age_group2 savings  271413  16410574  16411757.8
#> 17:      f age_group2 savings  271413  16410574  16411757.8
#> 18: female       ag2a savings  271413  16410574  16411757.8
#> 19:      f       ag2a savings  271413  16410574  16411757.8
#> 20: female age_group3 savings  222131  13497780  13497555.7
#> 21:      f age_group3 savings  222131  13497780  13497555.7
#> 22: female age_group4 savings  115241   6810459   6806070.6
#> 23:      f age_group4 savings  115241   6810459   6806070.6
#> 24: female age_group5 savings   37218   2377583   2378435.9
#> 25:      f age_group5 savings   37218   2377583   2378435.9
#> 26: female age_group6 savings    2987    123592    126756.7
#> 27:      f age_group6 savings    2987    123592    126756.7
#> 28:   male      Total savings 1159816  69662037  69660447.5
#> 29:      m      Total savings 1159816  69662037  69660447.5
#> 30:   male age_group1 savings  517660  31430472  31422620.1
#> 31:      m age_group1 savings  517660  31430472  31422620.1
#> 32:   male       ag1a savings  517660  31430472  31422620.1
#> 33:      m       ag1a savings  517660  31430472  31422620.1
#> 34:   male age_group2 savings  280923  16836241  16839487.2
#> 35:      m age_group2 savings  280923  16836241  16839487.2
#> 36:   male       ag2a savings  280923  16836241  16839487.2
#> 37:      m       ag2a savings  280923  16836241  16839487.2
#> 38:   male age_group3 savings  214970  12682929  12683174.2
#> 39:      m age_group3 savings  214970  12682929  12683174.2
#> 40:   male age_group4 savings   99420   5966130   5961942.5
#> 41:      m age_group4 savings   99420   5966130   5961942.5
#> 42:   male age_group5 savings   43233   2508522   2508834.0
#> 43:      m age_group5 savings   43233   2508522   2508834.0
#> 44:   male age_group6 savings    3610    237743    236666.3
#> 45:      m age_group6 savings    3610    237743    236666.3
#>        sex        age   vname     uws        ws         pws

# results can be resetted, too
tab$reset_cntvars(v = "cnt_males")

# we can then set other parameters and perturb again
tab$params_cnts_set(val = p_cnts1, v = "cnt_males")

#> --> setting perturbation parameters for variable 'cnt_males'

tab$perturb(v = "cnt_males")

#> Count variable 'cnt_males' was perturbed.

# write to a file "outtab.csv" (.csv is automatically added to the path)
if (FALSE) {
tab$freqtab(v = c("total", "cnt_males"), path = "outtab.csv")
}

# show results containing weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))

#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 274707 4579 274647.0203
#>  2:  Total age_group1     total 1969 118802 1968 118741.6638
#>  3:  Total       ag1a     total 1969 118802 1968 118741.6638
#>  4:  Total age_group2     total 1143  68583 1143  68583.0000
#>  5:  Total       ag2a     total 1143  68583 1143  68583.0000
#>  6:  Total age_group3     total  864  51473  864  51473.0000
#>  7:  Total age_group4     total  423  25121  423  25121.0000
#>  8:  Total age_group5     total  168   9970  167   9910.6548
#>  9:  Total age_group6     total   13    758   12    699.6923
#> 10: female      Total     total 2284 137362 2286 137482.2820
#> 11:      f      Total     total 2284 137362 2286 137482.2820
#> 12: female age_group1     total  954  57468  955  57528.2390
#> 13:      f age_group1     total  954  57468  955  57528.2390
#> 14: female       ag1a     total  954  57468  955  57528.2390
#> 15:      f       ag1a     total  954  57468  955  57528.2390
#> 16: female age_group2     total  572  34475  573  34535.2710
#> 17:      f age_group2     total  572  34475  573  34535.2710
#> 18: female       ag2a     total  572  34475  573  34535.2710
#> 19:      f       ag2a     total  572  34475  573  34535.2710
#> 20: female age_group3     total  440  26490  438  26369.5909
#> 21:      f age_group3     total  440  26490  438  26369.5909
#> 22: female age_group4     total  228  13373  228  13373.0000
#> 23:      f age_group4     total  228  13373  228  13373.0000
#> 24: female age_group5     total   84   5274   84   5274.0000
#> 25:      f age_group5     total   84   5274   84   5274.0000
#> 26: female age_group6     total    6    282    5    235.0000
#> 27:      f age_group6     total    6    282    5    235.0000
#> 28:   male      Total     total 2296 137345 2295 137285.1807
#> 29:      m      Total     total 2296 137345 2295 137285.1807
#> 30:   male age_group1     total 1015  61334 1014  61273.5724
#> 31:      m age_group1     total 1015  61334 1014  61273.5724
#> 32:   male       ag1a     total 1015  61334 1014  61273.5724
#> 33:      m       ag1a     total 1015  61334 1014  61273.5724
#> 34:   male age_group2     total  571  34108  572  34167.7338
#> 35:      m age_group2     total  571  34108  572  34167.7338
#> 36:   male       ag2a     total  571  34108  572  34167.7338
#> 37:      m       ag2a     total  571  34108  572  34167.7338
#> 38:   male age_group3     total  424  24983  424  24983.0000
#> 39:      m age_group3     total  424  24983  424  24983.0000
#> 40:   male age_group4     total  195  11748  197  11868.4923
#> 41:      m age_group4     total  195  11748  197  11868.4923
#> 42:   male age_group5     total   84   4696   85   4751.9048
#> 43:      m age_group5     total   84   4696   85   4751.9048
#> 44:   male age_group6     total    7    476    5    340.0000
#> 45:      m age_group6     total    7    476    5    340.0000
#> 46:  Total      Total cnt_males 2296 137345 2295 137285.1807
#> 47:  Total age_group1 cnt_males 1015  61334 1014  61273.5724
#> 48:  Total       ag1a cnt_males 1015  61334 1014  61273.5724
#> 49:  Total age_group2 cnt_males  571  34108  572  34167.7338
#> 50:  Total       ag2a cnt_males  571  34108  572  34167.7338
#> 51:  Total age_group3 cnt_males  424  24983  424  24983.0000
#> 52:  Total age_group4 cnt_males  195  11748  197  11868.4923
#> 53:  Total age_group5 cnt_males   84   4696   85   4751.9048
#> 54:  Total age_group6 cnt_males    7    476    5    340.0000
#> 55: female      Total cnt_males    0      0    0      0.0000
#> 56:      f      Total cnt_males    0      0    0      0.0000
#> 57: female age_group1 cnt_males    0      0    0      0.0000
#> 58:      f age_group1 cnt_males    0      0    0      0.0000
#> 59: female       ag1a cnt_males    0      0    0      0.0000
#> 60:      f       ag1a cnt_males    0      0    0      0.0000
#> 61: female age_group2 cnt_males    0      0    0      0.0000
#> 62:      f age_group2 cnt_males    0      0    0      0.0000
#> 63: female       ag2a cnt_males    0      0    0      0.0000
#> 64:      f       ag2a cnt_males    0      0    0      0.0000
#> 65: female age_group3 cnt_males    0      0    0      0.0000
#> 66:      f age_group3 cnt_males    0      0    0      0.0000
#> 67: female age_group4 cnt_males    0      0    0      0.0000
#> 68:      f age_group4 cnt_males    0      0    0      0.0000
#> 69: female age_group5 cnt_males    0      0    0      0.0000
#> 70:      f age_group5 cnt_males    0      0    0      0.0000
#> 71: female age_group6 cnt_males    0      0    0      0.0000
#> 72:      f age_group6 cnt_males    0      0    0      0.0000
#> 73:   male      Total cnt_males 2296 137345 2295 137285.1807
#> 74:      m      Total cnt_males 2296 137345 2295 137285.1807
#> 75:   male age_group1 cnt_males 1015  61334 1014  61273.5724
#> 76:      m age_group1 cnt_males 1015  61334 1014  61273.5724
#> 77:   male       ag1a cnt_males 1015  61334 1014  61273.5724
#> 78:      m       ag1a cnt_males 1015  61334 1014  61273.5724
#> 79:   male age_group2 cnt_males  571  34108  572  34167.7338
#> 80:      m age_group2 cnt_males  571  34108  572  34167.7338
#> 81:   male       ag2a cnt_males  571  34108  572  34167.7338
#> 82:      m       ag2a cnt_males  571  34108  572  34167.7338
#> 83:   male age_group3 cnt_males  424  24983  424  24983.0000
#> 84:      m age_group3 cnt_males  424  24983  424  24983.0000
#> 85:   male age_group4 cnt_males  195  11748  197  11868.4923
#> 86:      m age_group4 cnt_males  195  11748  197  11868.4923
#> 87:   male age_group5 cnt_males   84   4696   85   4751.9048
#> 88:      m age_group5 cnt_males   84   4696   85   4751.9048
#> 89:   male age_group6 cnt_males    7    476    5    340.0000
#> 90:      m age_group6 cnt_males    7    476    5    340.0000
#>        sex        age     vname  uwc     wc puwc         pwc

# utility measures for a count variable
tab$measures_cnts(v = "total", exclude_zeros = TRUE)

#> $overview
#>    noise cnt        pct
#> 1:    -2   4 0.08888889
#> 2:    -1  14 0.31111111
#> 3:     0  10 0.22222222
#> 4:     1  13 0.28888889
#> 5:     2   4 0.08888889
#> 
#> $measures
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 1.000 0.001 0.011
#>  5:    Q40 1.000 0.001 0.016
#>  6:   Mean 0.956 0.024 0.048
#>  7: Median 1.000 0.001 0.016
#>  8:    Q60 1.000 0.002 0.021
#>  9:    Q70 1.000 0.002 0.021
#> 10:    Q80 1.000 0.007 0.049
#> 11:    Q90 2.000 0.051 0.113
#> 12:    Q95 2.000 0.167 0.213
#> 13:    Q99 2.000 0.286 0.410
#> 14:    Max 2.000 0.286 0.410
#> 
#> $cumdistr_d1
#>    cat cnt       pct
#> 1:   0  10 0.2222222
#> 2:   1  37 0.8222222
#> 3:   2  45 1.0000000
#> 
#> $cumdistr_d2
#>            cat cnt       pct
#> 1:    [0,0.02]  40 0.8888889
#> 2: (0.02,0.05]  40 0.8888889
#> 3:  (0.05,0.1]  41 0.9111111
#> 4:   (0.1,0.2]  43 0.9555556
#> 5:   (0.2,0.3]  45 1.0000000
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $cumdistr_d3
#>            cat cnt       pct
#> 1:    [0,0.02]  23 0.5111111
#> 2: (0.02,0.05]  36 0.8000000
#> 3:  (0.05,0.1]  40 0.8888889
#> 4:   (0.1,0.2]  41 0.9111111
#> 5:   (0.2,0.3]  43 0.9555556
#> 6:   (0.3,0.4]  43 0.9555556
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $false_zero
#> [1] 0
#> 
#> $false_nonzero
#> [1] 0
#> 
#> $exclude_zeros
#> [1] TRUE
#> 

# modifications for perturbed count variables
tab$mod_cnts()

#>        sex        age row_nr pert       ckey  countvar
#>   1: Total      Total     14   -1 0.24410263     total
#>   2: Total age_group1     14   -1 0.22869726     total
#>   3: Total       ag1a     14   -1 0.22869726     total
#>   4: Total age_group2     15    0 0.63623435     total
#>   5: Total       ag2a     15    0 0.63623435     total
#>  ---                                                  
#> 131:     m age_group4     17    2 0.95029257 cnt_males
#> 132:  male age_group5     16    1 0.76251335 cnt_males
#> 133:     m age_group5     16    1 0.76251335 cnt_males
#> 134:  male age_group6     13   -2 0.06163708 cnt_males
#> 135:     m age_group6     13   -2 0.06163708 cnt_males

# display a summary about utility measures
tab$summary()

#> ┌──────────────────────────────────────────────┐
#> │Utility measures for perturbed count variables│
#> └──────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#>          countvar Min Q10 Q20 Q30 Q40  Mean Median Q60 Q70 Q80 Q90 Q95 Q99 Max
#> 1:          total  -2  -1  -1  -1   0 0.022      0 0.4 1.0   1 1.0 2.0   2   2
#> 2: cnt_highincome  -2  -2   0   0   0 0.311      0 0.0 0.8   1 2.6 3.0   4   4
#> 3:      cnt_males  -2  -1  -1   0   0 0.000      0 0.0 0.0   1 1.0 1.8   2   2
#> 
#> ── Distance-based measures ─────────────────────────────────────────────────────
#> ✔ Variable: 'total'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 1.000 0.001 0.011
#>  5:    Q40 1.000 0.001 0.016
#>  6:   Mean 0.956 0.024 0.048
#>  7: Median 1.000 0.001 0.016
#>  8:    Q60 1.000 0.002 0.021
#>  9:    Q70 1.000 0.002 0.021
#> 10:    Q80 1.000 0.007 0.049
#> 11:    Q90 2.000 0.051 0.113
#> 12:    Q95 2.000 0.167 0.213
#> 13:    Q99 2.000 0.286 0.410
#> 14:    Max 2.000 0.286 0.410
#> 
#> ✔ Variable: 'cnt_males'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.600 0.000 0.006
#>  3:    Q20 1.000 0.001 0.011
#>  4:    Q30 1.000 0.001 0.016
#>  5:    Q40 1.000 0.001 0.016
#>  6:   Mean 1.111 0.035 0.069
#>  7: Median 1.000 0.002 0.021
#>  8:    Q60 1.000 0.002 0.021
#>  9:    Q70 1.000 0.010 0.054
#> 10:    Q80 1.800 0.012 0.068
#> 11:    Q90 2.000 0.121 0.207
#> 12:    Q95 2.000 0.286 0.410
#> 13:    Q99 2.000 0.286 0.410
#> 14:    Max 2.000 0.286 0.410
#> 
#> ✔ Variable: 'cnt_highincome'
#> 
#>       what   d1    d2    d3
#>  1:    Min 0.00 0.000 0.000
#>  2:    Q10 0.00 0.000 0.000
#>  3:    Q20 0.00 0.000 0.000
#>  4:    Q30 0.00 0.000 0.000
#>  5:    Q40 0.00 0.000 0.000
#>  6:   Mean 1.05 0.033 0.080
#>  7: Median 1.00 0.009 0.060
#>  8:    Q60 1.00 0.016 0.068
#>  9:    Q70 2.00 0.018 0.108
#> 10:    Q80 2.00 0.030 0.124
#> 11:    Q90 3.00 0.143 0.183
#> 12:    Q95 3.05 0.149 0.288
#> 13:    Q99 4.00 0.267 0.486
#> 14:    Max 4.00 0.267 0.486
#> 
#> ┌──────────────────────────────────────────────────┐
#> │Utility measures for perturbed numerical variables│
#> └──────────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────

#> Warning: no non-missing arguments to min; returning Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to min; returning Inf

#> Warning: no non-missing arguments to max; returning -Inf

#>      vname        Min        Q10        Q20        Q30      Q40     Mean
#> 1:  expend        Inf         NA         NA         NA       NA      NaN
#> 2:  income -74747.579 -53369.292 -23877.935 -13747.282 -5482.80 -436.548
#> 3: savings  -8531.892  -6466.508  -2109.131   -375.942  -199.27 -431.285
#> 4:   mixed        Inf         NA         NA         NA       NA      NaN
#>      Median       Q60       Q70       Q80       Q90       Q95      Q99      Max
#> 1:       NA        NA        NA        NA        NA        NA       NA     -Inf
#> 2: 8522.635 10243.028 15699.244 34727.399 45567.471 45567.471 51624.70 51624.70
#> 3:  245.153   852.934  1183.768  2864.709  3246.228  3246.228  3247.92  3247.92
#> 4:       NA        NA        NA        NA        NA        NA       NA     -Inf