R6 Class defining statistical tables that can be perturbed

This class allows to define statistical tables and perturb both count and numerical variables.

ck_setup(x, rkey, dims, w = NULL, countvars = NULL, numvars = NULL)

Arguments

x

an object coercible to a data.frame

rkey

either a column name within x referring to a variable containing record keys or a single integer(ish) number > 5 that referns to the number of digits for record keys that will be generated internally.

dims

a list containing slots for each variable that should be tabulated. Each slot consists should be created/modified using sdcHierarchies::hier_create(), sdcHierarchies::hier_add() and other functionality from package sdcHierarchies.

w

(character) a scalar character referring to a variable in x holding sampling weights. If w is NULL (the default), all weights are assumed to be 1

countvars

(character) an optional vector containing names of binary (0/1 coded) variables withing x that should be included in the problem instance. These variables can later be perturbed.

numvars

(character) an optional vector of numerical variables that can later be tabulated.

Value

A new cellkey_obj object. Such objects (internally) contain the fully computed statistical tables given input microdata (x), the hierarchical definitionals (dims) as well as the remaining inputs. Intermediate results are stored internally and can only be modified / accessed via the exported public methods described below.

Details

Such objects are typically generated using ck_setup().

Methods

Public methods

ck_class$new()

ck_class$perturb()

ck_class$freqtab()

ck_class$numtab()

ck_class$measures_cnts()

ck_class$measures_nums()

ck_class$allvars()

ck_class$cntvars()

ck_class$numvars()

ck_class$hierarchy_info()

ck_class$mod_cnts()

ck_class$mod_nums()

ck_class$supp_freq()

ck_class$supp_val()

ck_class$supp_cells()

ck_class$supp_p()

ck_class$supp_pq()

ck_class$supp_nk()

ck_class$params_cnts_get()

ck_class$params_cnts_set()

ck_class$reset_cntvars()

ck_class$reset_numvars()

ck_class$reset_allvars()

ck_class$params_nums_get()

ck_class$params_nums_set()

ck_class$summary()

ck_class$print()

Method new()

Create a new table instance

Usage

ck_class$new(x, rkey, dims, w = NULL, countvars = NULL, numvars = NULL)

Arguments

x

an object coercible to a data.frame

rkey

either a column name within x referring to a variable containing record keys or a single integer(ish) number > 5 that referns to the number of digits for record keys that will be generated internally.

dims

a list containing slots for each variable that should be tabulated. Each slot consists should be created/modified using sdcHierarchies::hier_create(), sdcHierarchies::hier_add() and other functionality from package sdcHierarchies.

w

(character) a scalar character referring to a variable in x holding sampling weights. If w is NULL (the default), all weights are assumed to be 1

countvars

(character) an optional vector containing names of binary (0/1 coded) variables withing x that should be included in the problem instance. These variables can later be perturbed.

numvars

(character) an optional vector of numerical variables that can later be tabulated.

Returns

A new cellkey_obj object. Such objects (internally) contain the fully computed statistical tables given input microdata (x), the hierarchical definitionals (dims) as well as the remaining inputs. Intermediate results are stored internally and can only be modified / accessed via the exported public methods described below.

Method perturb()

Perturb a count- or magnitude variable

Usage

ck_class$perturb(v)

Arguments

v

name(s) of count- or magnitude variables that should be perturbed.

Returns

A modified cellkey_obj object in which private slots were updated for side-effects. Updated data can be accessed using other exported methods like $freqtab() or $numtab().

Method freqtab()

Extract results from already perturbed count variables as a data.table

Usage

ck_class$freqtab(v = NULL, path = NULL)

Arguments

v

a vector of variable names for count variables. If NULL (the default), the results are returned for all available count variables. For variables that have not yet perturbed, columns puwc and pwc are filled with NA.

path

if not NULL, a scalar character defining a (relative or absolute) path to which the result table should be written. A csv file will be generated and, if specified, path must have ".csv" as file-ending

Returns

This method returns a data.table containing all combinations of the dimensional variables in the first n columns. Additionally, the following columns are shown:

vname: name of the perturbed variable

uwc: unweighted counts

wc: weighted counts

puwc: perturbed unweighted counts or NA if vname was not yet perturbed

pwc: perturbed weighted counts or NA if vname was not yet perturbed

Method numtab()

Extract results from already perturbed continuous variables as a data.table.

Usage

ck_class$numtab(v = NULL, mean_before_sum = FALSE, path = NULL)

Arguments

v

a vector of variable names of continuous variables. If NULL (the default), the results are returned for all available numeric variables.

mean_before_sum

(logical); if TRUE, the perturbed values are adjusted by a factor ((n+p))⁄n with

n: the original weighted cell value

p: the perturbed cell value

This makes sense if the the accuracy of the variable mean is considered to be more important than accuracy of sums of the variable. The default value is FALSE (no adjustment is done)

path

if not NULL, a scalar character defining a (relative or absolute) path to which the result table should be written. A csv file will be generated and, if specified, path must have ".csv" as file-ending

Returns

This method returns a data.table containing all combinations of the dimensional variables in the first n columns. Additionally, the following columns are shown:

vname: name of the perturbed variable

uws: unweighted sum of the given variable

ws: weighted cellsum

pws: perturbed weighted sum of the given cell or NA if vname has not not perturbed

Method measures_cnts()

Utility measures for perturbed count variables

Usage

ck_class$measures_cnts(v, exclude_zeros = TRUE)

Arguments

v

name of a count variable for which utility measures should be computed.

exclude_zeros

should empty (zero) cells in the original values be excluded when computing distance measures

Returns

This method returns a list containing a set of utility measures based on some distance functions. For a detailed description of the computed measures, see ck_cnt_measures()

Method measures_nums()

Utility measures for continuous variables (not yet implemented)

Usage

ck_class$measures_nums(v)

Arguments

v

name of a continuous variable for which utility measures should be computed.

Returns

for (now) an empty list; In future versions of the package, the Method will return utility measures for perturbed magnitude tables.

Method allvars()

Names of variables that can be perturbed / tabulated

Usage

ck_class$allvars()

Returns

returns a list with the following two elements:

cntvars: character vector with names of available count variables for perturbation

numvars: character vector with names of available numerical variables for perturbation

Method cntvars()

Names of count variables that can be perturbed

Usage

ck_class$cntvars()

Returns

a character vector containing variable names

Method numvars()

Names of continuous variables that can be perturbed

Usage

ck_class$numvars()

Returns

a character vector containing variable names

Method hierarchy_info()

Information about hierarchies

Usage

ck_class$hierarchy_info()

Returns

a list (for each dimensional variable) with information on the hierarchies. This may be used to restrict output tables to specific levels or codes. Each list element is a data.table containing the following variables:

code: the name of a code within the hierarchy

level: number defining the level of the code; the higher the number, the lower the hierarchy with 1 being the overall total

is_leaf: if TRUE, this code is a leaf node which means no other codes contribute to it

parent: name of the parent code

Method mod_cnts()

Modifications applied to count variables

Usage

ck_class$mod_cnts()

Returns

a data.table containing modifications applied to count variables

Method mod_nums()

Modifications applied to numerical variables

Usage

ck_class$mod_nums()

Returns

a data.table containing modifications applied to numerical variables

Method supp_freq()

Identify sensitive cells based on minimum frequency rule

Usage

ck_class$supp_freq(v, n, weighted = TRUE)

Arguments

v

a single variable name of a continuous variable (see method numvars())

n

a number defining the threshold. All cells <= n are considered as unsafe.

weighted

if TRUE, the weighted number of contributors to a cell are compared to the threshold specified in n (default); else the unweighted number of contributors is used.

Returns

A modified cellkey_obj object in which private slots were updated for side-effects. These updated values are used by other methods (e.g $perturb()).

Method supp_val()

Identify sensitive cells based on weighted or unweighted cell value

Usage

ck_class$supp_val(v, n, weighted = TRUE)

Arguments

v

a single variable name of a continuous variable (see method numvars())

n

a number defining the threshold. All cells <= n are considered as unsafe.

weighted

if TRUE, the weighted cell value of variable v is compared to the threshold specified in n (default); else the unweighted number is used.

Returns

A modified cellkey_obj object in which private slots were updated for side-effects. These updated values are used by other methods (e.g $perturb()).

Method supp_cells()

Identify sensitive cells based on their names

Usage

ck_class$supp_cells(v, inp)

Arguments

v

a single variable name of a continuous variable (see method numvars())

inp

a data.frame where each colum represents a dimensional variable. Each row of this input is then used to compute the relevant cells to be identified as sensitive where NA-values are possible and used to match any characteristics of the dimensional variable.

Returns

A modified cellkey_obj object in which private slots were updated for side-effects. These updated values are used by other methods (e.g $perturb()).

Method supp_p()

Identify sensitive cells based on the p%-rule rule. Please note that this rule can only be applied to positive-only variables.

Usage

ck_class$supp_p(v, p)

Arguments

v

a single variable name of a continuous variable (see method numvars())

p

a number defining a percentage between 1 and 99.

Returns

A modified cellkey_obj object in which private slots were updated for side-effects. These updated values are used by other methods (e.g $perturb()).

Method supp_pq()

Identify sensitive cells based on the pq-rule. Please note that this rule can only be applied to positive-only variables.

Usage

ck_class$supp_pq(v, p, q)

Arguments

v

a single variable name of a continuous variable (see method numvars())

p

a number defining a percentage between 1 and 99.

q

a number defining a percentage between 1 and 99. This value must be larger than p.

Returns

A modified cellkey_obj object in which private slots were updated for side-effects. These updated values are used by other methods (e.g $perturb()).

Method supp_nk()

Identify sensitive cells based on the nk-dominance rule. Please note that this rule can only be applied to positive-only variables.

Usage

ck_class$supp_nk(v, n, k)

Arguments

v

a single variable name of a continuous variable (see method numvars())

n

an integerish number >= 2

k

a number defining a percentage between 1 and 99. All cells to which the top n contributers contribute more than k% is considered unsafe

Returns

A modified cellkey_obj object in which private slots were updated for side-effects. These updated values are used by other methods (e.g $perturb()).

Method params_cnts_get()

Return perturbation parameters of count variables

Usage

ck_class$params_cnts_get()

Returns

a named list in which each list-element contains the active perturbation parameters for the specific count variable defined by the list-name.

Method params_cnts_set()

Set perturbation parameters for count variables

Usage

ck_class$params_cnts_set(val, v = NULL)

Arguments

val

a perturbation object created with ck_params_cnts()

v

a character vector (or NULL). If NULL (the default), the perturbation parameters provided in val are set for all count variables; otherwise one may specify the names of the count variables for which the parameters should be set.

Returns

A modified cellkey_obj object in which private slots were updated for side-effects. These updated values are used by other methods (e.g $perturb()).

Method reset_cntvars()

reset results and parameters for already perturbed count variables

Usage

ck_class$reset_cntvars(v = NULL)

Arguments

v

if v equals NULL (the default), the results are reset for all perturbed count variables; otherwise it is possible to specify the names of already perturbed count variables.

Returns

A modified cellkey_obj object in which private slots were updated for side-effects. These updated values are used by other methods (e.g $perturb() or $freqtab()).

Method reset_numvars()

reset results and parameters for already perturbed numerical variables

Usage

ck_class$reset_numvars(v = NULL)

Arguments

v

if v equals NULL (the default), the results are reset for all perturbed numerical variables; otherwise it is possible to specify the names of already perturbed continuous variables.

Returns

A modified cellkey_obj object in which private slots were updated for side-effects. These updated values are used by other methods (e.g $perturb() or $numtab()).

Method reset_allvars()

reset results and parameters for all already perturbed variables.

Usage

ck_class$reset_allvars()

Returns

A modified cellkey_obj object in which private slots were updated for side-effects. These updated values are used by other methods (e.g $perturb(), $freqtab() or $numtab()).

Method params_nums_get()

Return perturbation parameters of continuous variables

Usage

ck_class$params_nums_get()

Returns

a named list in which each list-element contains the active perturbation parameters for the specific continuous variable defined by the list-name.

Method params_nums_set()

set perturbation parameters for continuous variables.

Usage

ck_class$params_nums_set(val, v = NULL)

Arguments

val

a perturbation object created with ck_params_nums()

v

a character vector (or NULL); if NULL (the default), the perturbation parameters provided in val are set for all continuous variables; otherwise one may specify the names of the numeric variables for which the parameters should be set.

Returns

A modified cellkey_obj object in which private slots were updated for side-effects. These updated values are used by other methods (e.g $perturb()).

Method summary()

some aggregated summary statistics about perturbed variables

Usage

ck_class$summary()

Returns

invisible NULL

Method print()

prints information about the current table

Usage

ck_class$print()

Returns

invisible NULL

Examples

# \donttest{
x <- ck_create_testdata()

# create some 0/1 variables that should be perturbed later
x[, cnt_females := ifelse(sex == "male", 0, 1)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              37          25.00000           0
#>    2:   2530      28       1              64          25.00000           1
#>    3:   6920     550       1              60          25.00000           0
#>    4:   7960     870       1              94          25.00000           0
#>    5:   9030      20       2              99          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              67          16.66667           1
#> 4577:   1420     987    1000              97          16.66667           0
#> 4578:   8900     684    1000              81          16.66667           0
#> 4579:   3880     294    1000              42          16.66667           1
#> 4580:   4830     911    1000              51          16.66667           0
x[, cnt_males := ifelse(sex == "male", 1, 0)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              37          25.00000           0
#>    2:   2530      28       1              64          25.00000           1
#>    3:   6920     550       1              60          25.00000           0
#>    4:   7960     870       1              94          25.00000           0
#>    5:   9030      20       2              99          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              67          16.66667           1
#> 4577:   1420     987    1000              97          16.66667           0
#> 4578:   8900     684    1000              81          16.66667           0
#> 4579:   3880     294    1000              42          16.66667           1
#> 4580:   4830     911    1000              51          16.66667           0
#>       cnt_males
#>    1:         1
#>    2:         0
#>    3:         1
#>    4:         1
#>    5:         1
#>   ---          
#> 4576:         0
#> 4577:         1
#> 4578:         1
#> 4579:         0
#> 4580:         1
x[, cnt_highincome := ifelse(income >= 9000, 1, 0)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              37          25.00000           0
#>    2:   2530      28       1              64          25.00000           1
#>    3:   6920     550       1              60          25.00000           0
#>    4:   7960     870       1              94          25.00000           0
#>    5:   9030      20       2              99          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              67          16.66667           1
#> 4577:   1420     987    1000              97          16.66667           0
#> 4578:   8900     684    1000              81          16.66667           0
#> 4579:   3880     294    1000              42          16.66667           1
#> 4580:   4830     911    1000              51          16.66667           0
#>       cnt_males cnt_highincome
#>    1:         1              0
#>    2:         0              0
#>    3:         1              0
#>    4:         1              0
#>    5:         1              1
#>   ---                         
#> 4576:         0              0
#> 4577:         1              0
#> 4578:         1              0
#> 4579:         0              0
#> 4580:         1              0
# a variable with positive and negative contributions
x[, mixed := sample(-10:10, nrow(x), replace = TRUE)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              37          25.00000           0
#>    2:   2530      28       1              64          25.00000           1
#>    3:   6920     550       1              60          25.00000           0
#>    4:   7960     870       1              94          25.00000           0
#>    5:   9030      20       2              99          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              67          16.66667           1
#> 4577:   1420     987    1000              97          16.66667           0
#> 4578:   8900     684    1000              81          16.66667           0
#> 4579:   3880     294    1000              42          16.66667           1
#> 4580:   4830     911    1000              51          16.66667           0
#>       cnt_males cnt_highincome mixed
#>    1:         1              0    -5
#>    2:         0              0     2
#>    3:         1              0     9
#>    4:         1              0    10
#>    5:         1              1    -5
#>   ---                               
#> 4576:         0              0    10
#> 4577:         1              0    -5
#> 4578:         1              0    -9
#> 4579:         0              0    -1
#> 4580:         1              0     3

# create record keys
x$rkey <- ck_generate_rkeys(dat = x)

# define required inputs

# hierarchy with some bogus codes
d_sex <- hier_create(root = "Total", nodes = c("male", "female"))
d_sex <- hier_add(d_sex, root = "female", "f")
d_sex <- hier_add(d_sex, root = "male", "m")

d_age <- hier_create(root = "Total", nodes = paste0("age_group", 1:6))
d_age <- hier_add(d_age, root = "age_group1", "ag1a")
d_age <- hier_add(d_age, root = "age_group2", "ag2a")

# define the cell key object
countvars <- c("cnt_females", "cnt_males", "cnt_highincome")
numvars <- c("expend", "income", "savings", "mixed")
tab <- ck_setup(
  x = x,
  rkey = "rkey",
  dims = list(sex = d_sex, age = d_age),
  w = "sampling_weight",
  countvars = countvars,
  numvars = numvars)

# show some information about this table instance
tab$print() # identical with print(tab)
#> ── Table Information ───────────────────────────────────────────────────────────
#> ✔ 45 cells in 2 dimensions ('sex', 'age')
#> ✔ weights: yes
#> ── Tabulated / Perturbed countvars ─────────────────────────────────────────────
#> ☐ 'total'
#> ☐ 'cnt_females'
#> ☐ 'cnt_males'
#> ☐ 'cnt_highincome'
#> ── Tabulated / Perturbed numvars ───────────────────────────────────────────────
#> ☐ 'expend'
#> ☐ 'income'
#> ☐ 'savings'
#> ☐ 'mixed'

# information about the hierarchies
tab$hierarchy_info()
#> $sex
#>      code level is_leaf parent
#> 1:  Total     1   FALSE  Total
#> 2:   male     2   FALSE  Total
#> 3:      m     3    TRUE   male
#> 4: female     2   FALSE  Total
#> 5:      f     3    TRUE female
#> 
#> $age
#>          code level is_leaf     parent
#> 1:      Total     1   FALSE      Total
#> 2: age_group1     2   FALSE      Total
#> 3:       ag1a     3    TRUE age_group1
#> 4: age_group2     2   FALSE      Total
#> 5:       ag2a     3    TRUE age_group2
#> 6: age_group3     2    TRUE      Total
#> 7: age_group4     2    TRUE      Total
#> 8: age_group5     2    TRUE      Total
#> 9: age_group6     2    TRUE      Total
#> 

# which variables have been defined?
tab$allvars()
#> $cntvars
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"
#> 
#> $numvars
#> [1] "expend"  "income"  "savings" "mixed"  
#> 

# count variables
tab$cntvars()
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"

# continuous variables
tab$numvars()
#> [1] "expend"  "income"  "savings" "mixed"  

# create perturbation parameters for "total" variable and
# write to yaml-file

# create a ptable using functionality from the ptable-pkg
f_yaml <- tempfile(fileext = ".yaml")
p_cnts1 <- ck_params_cnts(
  ptab = ptable::pt_ex_cnts(),
  path = f_yaml)
#> yaml configuration '/tmp/RtmpAtvqZz/file1a0a2552689c.yaml' successfully written.

# read parameters from yaml-file and set them for variable `"total"`
p_cnts1 <- ck_read_yaml(path = f_yaml)

tab$params_cnts_set(val = p_cnts1, v = "total")
#> --> setting perturbation parameters for variable 'total'

# create alternative perturbation parameters by specifying parameters
para2 <- ptable::create_cnt_ptable(
  D = 8, V = 3, js = 2, create = FALSE)

p_cnts2 <- ck_params_cnts(ptab = para2)

# use these ptable it for the remaining variables
tab$params_cnts_set(val = p_cnts2, v = countvars)
#> --> setting perturbation parameters for variable 'cnt_females'
#> --> setting perturbation parameters for variable 'cnt_males'
#> --> setting perturbation parameters for variable 'cnt_highincome'

# perturb a variable
tab$perturb(v = "total")
#> Count variable 'total' was perturbed.

# multiple variables can be perturbed as well
tab$perturb(v = c("cnt_males", "cnt_highincome"))
#> Count variable 'cnt_males' was perturbed.
#> Count variable 'cnt_highincome' was perturbed.

# return weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))
#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 276703 4579 276642.5845
#>  2:  Total age_group1     total 1969 118641 1968 118580.7456
#>  3:  Total       ag1a     total 1969 118641 1968 118580.7456
#>  4:  Total age_group2     total 1143  68573 1142  68513.0061
#>  5:  Total       ag2a     total 1143  68573 1142  68513.0061
#>  6:  Total age_group3     total  864  52725  863  52663.9757
#>  7:  Total age_group4     total  423  25729  423  25729.0000
#>  8:  Total age_group5     total  168  10350  170  10473.2143
#>  9:  Total age_group6     total   13    685   14    737.6923
#> 10:   male      Total     total 2296 138307 2298 138427.4765
#> 11:      m      Total     total 2296 138307 2298 138427.4765
#> 12:   male age_group1     total 1015  60859 1013  60739.0808
#> 13:      m age_group1     total 1015  60859 1013  60739.0808
#> 14:   male       ag1a     total 1015  60859 1013  60739.0808
#> 15:      m       ag1a     total 1015  60859 1013  60739.0808
#> 16:   male age_group2     total  571  34232  571  34232.0000
#> 17:      m age_group2     total  571  34232  571  34232.0000
#> 18:   male       ag2a     total  571  34232  571  34232.0000
#> 19:      m       ag2a     total  571  34232  571  34232.0000
#> 20:   male age_group3     total  424  25681  425  25741.5684
#> 21:      m age_group3     total  424  25681  425  25741.5684
#> 22:   male age_group4     total  195  12114  193  11989.7538
#> 23:      m age_group4     total  195  12114  193  11989.7538
#> 24:   male age_group5     total   84   5051   86   5171.2619
#> 25:      m age_group5     total   84   5051   86   5171.2619
#> 26:   male age_group6     total    7    370    8    422.8571
#> 27:      m age_group6     total    7    370    8    422.8571
#> 28: female      Total     total 2284 138396 2283 138335.4063
#> 29:      f      Total     total 2284 138396 2283 138335.4063
#> 30: female age_group1     total  954  57782  953  57721.4319
#> 31:      f age_group1     total  954  57782  953  57721.4319
#> 32: female       ag1a     total  954  57782  953  57721.4319
#> 33:      f       ag1a     total  954  57782  953  57721.4319
#> 34: female age_group2     total  572  34341  572  34341.0000
#> 35:      f age_group2     total  572  34341  572  34341.0000
#> 36: female       ag2a     total  572  34341  572  34341.0000
#> 37:      f       ag2a     total  572  34341  572  34341.0000
#> 38: female age_group3     total  440  27044  440  27044.0000
#> 39:      f age_group3     total  440  27044  440  27044.0000
#> 40: female age_group4     total  228  13615  228  13615.0000
#> 41:      f age_group4     total  228  13615  228  13615.0000
#> 42: female age_group5     total   84   5299   82   5172.8333
#> 43:      f age_group5     total   84   5299   82   5172.8333
#> 44: female age_group6     total    6    315    5    262.5000
#> 45:      f age_group6     total    6    315    5    262.5000
#> 46:  Total      Total cnt_males 2296 138307 2299 138487.7147
#> 47:  Total age_group1 cnt_males 1015  60859 1010  60559.2020
#> 48:  Total       ag1a cnt_males 1015  60859 1010  60559.2020
#> 49:  Total age_group2 cnt_males  571  34232  571  34232.0000
#> 50:  Total       ag2a cnt_males  571  34232  571  34232.0000
#> 51:  Total age_group3 cnt_males  424  25681  425  25741.5684
#> 52:  Total age_group4 cnt_males  195  12114  190  11803.3846
#> 53:  Total age_group5 cnt_males   84   5051   87   5231.3929
#> 54:  Total age_group6 cnt_males    7    370    8    422.8571
#> 55:   male      Total cnt_males 2296 138307 2299 138487.7147
#> 56:      m      Total cnt_males 2296 138307 2299 138487.7147
#> 57:   male age_group1 cnt_males 1015  60859 1010  60559.2020
#> 58:      m age_group1 cnt_males 1015  60859 1010  60559.2020
#> 59:   male       ag1a cnt_males 1015  60859 1010  60559.2020
#> 60:      m       ag1a cnt_males 1015  60859 1010  60559.2020
#> 61:   male age_group2 cnt_males  571  34232  571  34232.0000
#> 62:      m age_group2 cnt_males  571  34232  571  34232.0000
#> 63:   male       ag2a cnt_males  571  34232  571  34232.0000
#> 64:      m       ag2a cnt_males  571  34232  571  34232.0000
#> 65:   male age_group3 cnt_males  424  25681  425  25741.5684
#> 66:      m age_group3 cnt_males  424  25681  425  25741.5684
#> 67:   male age_group4 cnt_males  195  12114  190  11803.3846
#> 68:      m age_group4 cnt_males  195  12114  190  11803.3846
#> 69:   male age_group5 cnt_males   84   5051   87   5231.3929
#> 70:      m age_group5 cnt_males   84   5051   87   5231.3929
#> 71:   male age_group6 cnt_males    7    370    8    422.8571
#> 72:      m age_group6 cnt_males    7    370    8    422.8571
#> 73: female      Total cnt_males    0      0    0      0.0000
#> 74:      f      Total cnt_males    0      0    0      0.0000
#> 75: female age_group1 cnt_males    0      0    0      0.0000
#> 76:      f age_group1 cnt_males    0      0    0      0.0000
#> 77: female       ag1a cnt_males    0      0    0      0.0000
#> 78:      f       ag1a cnt_males    0      0    0      0.0000
#> 79: female age_group2 cnt_males    0      0    0      0.0000
#> 80:      f age_group2 cnt_males    0      0    0      0.0000
#> 81: female       ag2a cnt_males    0      0    0      0.0000
#> 82:      f       ag2a cnt_males    0      0    0      0.0000
#> 83: female age_group3 cnt_males    0      0    0      0.0000
#> 84:      f age_group3 cnt_males    0      0    0      0.0000
#> 85: female age_group4 cnt_males    0      0    0      0.0000
#> 86:      f age_group4 cnt_males    0      0    0      0.0000
#> 87: female age_group5 cnt_males    0      0    0      0.0000
#> 88:      f age_group5 cnt_males    0      0    0      0.0000
#> 89: female age_group6 cnt_males    0      0    0      0.0000
#> 90:      f age_group6 cnt_males    0      0    0      0.0000
#>        sex        age     vname  uwc     wc puwc         pwc

# numerical variables (positive variables using flex-function)
# we also write the config to a yaml file
f_yaml <- tempfile(fileext = ".yaml")

# create a ptable using functionality from the ptable-pkg
# a single ptable for all cells
ptab1 <- ptable::pt_ex_nums(parity = TRUE, separation = FALSE)

# a single ptab for all cells except for very small ones
ptab2 <- ptable::pt_ex_nums(parity = TRUE, separation = TRUE)

# different ptables for cells with even/odd number of contributors
# and very small cells
ptab3 <- ptable::pt_ex_nums(parity = FALSE, separation = TRUE)

p_nums1 <- ck_params_nums(
  ptab = ptab1,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.30, 0.03),
    epsilon = c(1, 0.5, 0.2),
    q = 3),
  mu_c = 2,
  same_key = FALSE,
  use_zero_rkeys = FALSE,
  path = f_yaml)
#> yaml configuration '/tmp/RtmpAtvqZz/file1a0a602ebcbc.yaml' successfully written.

# we read the parameters from the yaml-file
p_nums1 <- ck_read_yaml(path = f_yaml)

# for variables with positive and negative values
p_nums2 <- ck_params_nums(
  ptab = ptab2,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.15, 0.02),
    epsilon = c(1, 0.4, 0.15),
    q = 3),
  mu_c = 2,
  same_key = FALSE)

# simple perturbation parameters (not using the flex-function approach)
p_nums3 <- ck_params_nums(
  ptab = ptab3,
  type = "mean",
  mult_params = ck_simpleparams(p = 0.25),
  mu_c = 2,
  same_key = FALSE)

# use `p_nums1` for all variables
tab$params_nums_set(p_nums1, c("savings", "income", "expend"))
#> --> setting perturbation parameters for variable 'savings'
#> --> setting perturbation parameters for variable 'income'
#> --> setting perturbation parameters for variable 'expend'

# use different parameters for variable `mixed`
tab$params_nums_set(p_nums2, v = "mixed")
#> --> setting perturbation parameters for variable 'mixed'

# identify sensitive cells to which extra protection (`mu_c`) is added.
tab$supp_p(v = "income", p = 85)
#> computing contributing indices | rawdata <--> table; this might take a while
#> p%-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_pq(v = "income", p = 85, q = 90)
#> computing contributing indices | rawdata <--> table; this might take a while
#> pq-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_nk(v = "income", n = 2, k = 90)
#> computing contributing indices | rawdata <--> table; this might take a while
#> nk-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_freq(v = "income", n = 14, weighted = FALSE)
#> freq-rule: 5 new sensitive cells (incl. duplicates) found (total: 5)
tab$supp_val(v = "income", n = 10000, weighted = TRUE)
#> val-rule: 0 new sensitive cells (incl. duplicates) found (total: 5)
tab$supp_cells(
  v = "income",
  inp = data.frame(
    sex = c("female", "female"),
    "age" = c("age_group1", "age_group3")
  )
)
#> cell-rule: 2 new sensitive cells (incl. duplicates) found (total: 7)

# perturb variables
tab$perturb(v = c("income", "savings"))
#> Numeric variable 'income' was perturbed.
#> Numeric variable 'savings' was perturbed.

# extract results
tab$numtab("income", mean_before_sum = TRUE)
#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1390547170 1390662924
#>  2:  Total age_group1 income  9810547  590017386  590173834
#>  3:  Total       ag1a income  9810547  590017386  590173834
#>  4:  Total age_group2 income  5692119  344824821  344816086
#>  5:  Total       ag2a income  5692119  344824821  344816086
#>  6:  Total age_group3 income  4406946  270477959  270379566
#>  7:  Total age_group4 income  2133543  130042137  130070327
#>  8:  Total age_group5 income   848151   51556315   51564416
#>  9:  Total age_group6 income    61672    3628552    3620618
#> 10:   male      Total income 11262049  676250889  675968446
#> 11:      m      Total income 11262049  676250889  675968446
#> 12:   male age_group1 income  4877164  289944637  289736236
#> 13:      m age_group1 income  4877164  289944637  289736236
#> 14:   male       ag1a income  4877164  289944637  289736236
#> 15:      m       ag1a income  4877164  289944637  289736236
#> 16:   male age_group2 income  2811379  169318656  169353418
#> 17:      m age_group2 income  2811379  169318656  169353418
#> 18:   male       ag2a income  2811379  169318656  169353418
#> 19:      m       ag2a income  2811379  169318656  169353418
#> 20:   male age_group3 income  2168169  131456498  131502685
#> 21:      m age_group3 income  2168169  131456498  131502685
#> 22:   male age_group4 income   978510   60673098   60547312
#> 23:      m age_group4 income   978510   60673098   60547312
#> 24:   male age_group5 income   393134   22891345   22867673
#> 25:      m age_group5 income   393134   22891345   22867673
#> 26:   male age_group6 income    33693    1966655    2095461
#> 27:      m age_group6 income    33693    1966655    2095461
#> 28: female      Total income 11690929  714296281  714387398
#> 29:      f      Total income 11690929  714296281  714387398
#> 30: female age_group1 income  4933383  300072749  300182667
#> 31:      f age_group1 income  4933383  300072749  300182667
#> 32: female       ag1a income  4933383  300072749  300182667
#> 33:      f       ag1a income  4933383  300072749  300182667
#> 34: female age_group2 income  2880740  175506165  175500965
#> 35:      f age_group2 income  2880740  175506165  175500965
#> 36: female       ag2a income  2880740  175506165  175500965
#> 37:      f       ag2a income  2880740  175506165  175500965
#> 38: female age_group3 income  2238777  139021461  139012753
#> 39:      f age_group3 income  2238777  139021461  139012753
#> 40: female age_group4 income  1155033   69369039   69250348
#> 41:      f age_group4 income  1155033   69369039   69250348
#> 42: female age_group5 income   455017   28664970   28640277
#> 43:      f age_group5 income   455017   28664970   28640277
#> 44: female age_group6 income    27979    1661897    1478777
#> 45:      f age_group6 income    27979    1661897    1478777
#>        sex        age  vname      uws         ws        pws
tab$numtab("income", mean_before_sum = FALSE)
#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1390547170 1390605046
#>  2:  Total age_group1 income  9810547  590017386  590095605
#>  3:  Total       ag1a income  9810547  590017386  590095605
#>  4:  Total age_group2 income  5692119  344824821  344820453
#>  5:  Total       ag2a income  5692119  344824821  344820453
#>  6:  Total age_group3 income  4406946  270477959  270428758
#>  7:  Total age_group4 income  2133543  130042137  130056231
#>  8:  Total age_group5 income   848151   51556315   51560365
#>  9:  Total age_group6 income    61672    3628552    3624583
#> 10:   male      Total income 11262049  676250889  676109653
#> 11:      m      Total income 11262049  676250889  676109653
#> 12:   male age_group1 income  4877164  289944637  289840418
#> 13:      m age_group1 income  4877164  289944637  289840418
#> 14:   male       ag1a income  4877164  289944637  289840418
#> 15:      m       ag1a income  4877164  289944637  289840418
#> 16:   male age_group2 income  2811379  169318656  169336036
#> 17:      m age_group2 income  2811379  169318656  169336036
#> 18:   male       ag2a income  2811379  169318656  169336036
#> 19:      m       ag2a income  2811379  169318656  169336036
#> 20:   male age_group3 income  2168169  131456498  131479590
#> 21:      m age_group3 income  2168169  131456498  131479590
#> 22:   male age_group4 income   978510   60673098   60610172
#> 23:      m age_group4 income   978510   60673098   60610172
#> 24:   male age_group5 income   393134   22891345   22879506
#> 25:      m age_group5 income   393134   22891345   22879506
#> 26:   male age_group6 income    33693    1966655    2030037
#> 27:      m age_group6 income    33693    1966655    2030037
#> 28: female      Total income 11690929  714296281  714341838
#> 29:      f      Total income 11690929  714296281  714341838
#> 30: female age_group1 income  4933383  300072749  300127703
#> 31:      f age_group1 income  4933383  300072749  300127703
#> 32: female       ag1a income  4933383  300072749  300127703
#> 33:      f       ag1a income  4933383  300072749  300127703
#> 34: female age_group2 income  2880740  175506165  175503565
#> 35:      f age_group2 income  2880740  175506165  175503565
#> 36: female       ag2a income  2880740  175506165  175503565
#> 37:      f       ag2a income  2880740  175506165  175503565
#> 38: female age_group3 income  2238777  139021461  139017107
#> 39:      f age_group3 income  2238777  139021461  139017107
#> 40: female age_group4 income  1155033   69369039   69309668
#> 41:      f age_group4 income  1155033   69369039   69309668
#> 42: female age_group5 income   455017   28664970   28652621
#> 43:      f age_group5 income   455017   28664970   28652621
#> 44: female age_group6 income    27979    1661897    1567665
#> 45:      f age_group6 income    27979    1661897    1567665
#>        sex        age  vname      uws         ws        pws
tab$numtab("savings")
#>        sex        age   vname     uws        ws         pws
#>  1:  Total      Total savings 2273532 137468654 137471548.0
#>  2:  Total age_group1 savings  982386  58731465  58729490.2
#>  3:  Total       ag1a savings  982386  58731465  58729490.2
#>  4:  Total age_group2 savings  552336  33205691  33205223.9
#>  5:  Total       ag2a savings  552336  33205691  33205223.9
#>  6:  Total age_group3 savings  437101  27056143  27060769.7
#>  7:  Total age_group4 savings  214661  13188779  13189462.5
#>  8:  Total age_group5 savings   80451   4911708   4912610.6
#>  9:  Total age_group6 savings    6597    374868    373910.6
#> 10:   male      Total savings 1159816  69780107  69780584.2
#> 11:      m      Total savings 1159816  69780107  69780584.2
#> 12:   male age_group1 savings  517660  30924542  30918712.5
#> 13:      m age_group1 savings  517660  30924542  30918712.5
#> 14:   male       ag1a savings  517660  30924542  30918712.5
#> 15:      m       ag1a savings  517660  30924542  30918712.5
#> 16:   male age_group2 savings  280923  16764259  16769565.2
#> 17:      m age_group2 savings  280923  16764259  16769565.2
#> 18:   male       ag2a savings  280923  16764259  16769565.2
#> 19:      m       ag2a savings  280923  16764259  16769565.2
#> 20:   male age_group3 savings  214970  13215429  13212058.7
#> 21:      m age_group3 savings  214970  13215429  13212058.7
#> 22:   male age_group4 savings   99420   6136850   6130485.9
#> 23:      m age_group4 savings   99420   6136850   6130485.9
#> 24:   male age_group5 savings   43233   2539966   2538344.2
#> 25:      m age_group5 savings   43233   2539966   2538344.2
#> 26:   male age_group6 savings    3610    199061    206418.4
#> 27:      m age_group6 savings    3610    199061    206418.4
#> 28: female      Total savings 1113716  67688547  67685808.5
#> 29:      f      Total savings 1113716  67688547  67685808.5
#> 30: female age_group1 savings  464726  27806923  27808972.5
#> 31:      f age_group1 savings  464726  27806923  27808972.5
#> 32: female       ag1a savings  464726  27806923  27808972.5
#> 33:      f       ag1a savings  464726  27806923  27808972.5
#> 34: female age_group2 savings  271413  16441432  16441732.0
#> 35:      f age_group2 savings  271413  16441432  16441732.0
#> 36: female       ag2a savings  271413  16441432  16441732.0
#> 37:      f       ag2a savings  271413  16441432  16441732.0
#> 38: female age_group3 savings  222131  13840714  13839752.3
#> 39:      f age_group3 savings  222131  13840714  13839752.3
#> 40: female age_group4 savings  115241   7051929   7051539.8
#> 41:      f age_group4 savings  115241   7051929   7051539.8
#> 42: female age_group5 savings   37218   2371742   2370729.5
#> 43:      f age_group5 savings   37218   2371742   2370729.5
#> 44: female age_group6 savings    2987    175807    171089.5
#> 45:      f age_group6 savings    2987    175807    171089.5
#>        sex        age   vname     uws        ws         pws

# results can be resetted, too
tab$reset_cntvars(v = "cnt_males")

# we can then set other parameters and perturb again
tab$params_cnts_set(val = p_cnts1, v = "cnt_males")
#> --> setting perturbation parameters for variable 'cnt_males'

tab$perturb(v = "cnt_males")
#> Count variable 'cnt_males' was perturbed.

# write results to a .csv file
tab$freqtab(
  v = c("total", "cnt_males"),
  path = file.path(tempdir(), "outtab.csv")
)
#> File '/tmp/RtmpAtvqZz/outtab.csv' successfully written to disk.
#> NULL

# show results containing weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))
#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 276703 4579 276642.5845
#>  2:  Total age_group1     total 1969 118641 1968 118580.7456
#>  3:  Total       ag1a     total 1969 118641 1968 118580.7456
#>  4:  Total age_group2     total 1143  68573 1142  68513.0061
#>  5:  Total       ag2a     total 1143  68573 1142  68513.0061
#>  6:  Total age_group3     total  864  52725  863  52663.9757
#>  7:  Total age_group4     total  423  25729  423  25729.0000
#>  8:  Total age_group5     total  168  10350  170  10473.2143
#>  9:  Total age_group6     total   13    685   14    737.6923
#> 10:   male      Total     total 2296 138307 2298 138427.4765
#> 11:      m      Total     total 2296 138307 2298 138427.4765
#> 12:   male age_group1     total 1015  60859 1013  60739.0808
#> 13:      m age_group1     total 1015  60859 1013  60739.0808
#> 14:   male       ag1a     total 1015  60859 1013  60739.0808
#> 15:      m       ag1a     total 1015  60859 1013  60739.0808
#> 16:   male age_group2     total  571  34232  571  34232.0000
#> 17:      m age_group2     total  571  34232  571  34232.0000
#> 18:   male       ag2a     total  571  34232  571  34232.0000
#> 19:      m       ag2a     total  571  34232  571  34232.0000
#> 20:   male age_group3     total  424  25681  425  25741.5684
#> 21:      m age_group3     total  424  25681  425  25741.5684
#> 22:   male age_group4     total  195  12114  193  11989.7538
#> 23:      m age_group4     total  195  12114  193  11989.7538
#> 24:   male age_group5     total   84   5051   86   5171.2619
#> 25:      m age_group5     total   84   5051   86   5171.2619
#> 26:   male age_group6     total    7    370    8    422.8571
#> 27:      m age_group6     total    7    370    8    422.8571
#> 28: female      Total     total 2284 138396 2283 138335.4063
#> 29:      f      Total     total 2284 138396 2283 138335.4063
#> 30: female age_group1     total  954  57782  953  57721.4319
#> 31:      f age_group1     total  954  57782  953  57721.4319
#> 32: female       ag1a     total  954  57782  953  57721.4319
#> 33:      f       ag1a     total  954  57782  953  57721.4319
#> 34: female age_group2     total  572  34341  572  34341.0000
#> 35:      f age_group2     total  572  34341  572  34341.0000
#> 36: female       ag2a     total  572  34341  572  34341.0000
#> 37:      f       ag2a     total  572  34341  572  34341.0000
#> 38: female age_group3     total  440  27044  440  27044.0000
#> 39:      f age_group3     total  440  27044  440  27044.0000
#> 40: female age_group4     total  228  13615  228  13615.0000
#> 41:      f age_group4     total  228  13615  228  13615.0000
#> 42: female age_group5     total   84   5299   82   5172.8333
#> 43:      f age_group5     total   84   5299   82   5172.8333
#> 44: female age_group6     total    6    315    5    262.5000
#> 45:      f age_group6     total    6    315    5    262.5000
#> 46:  Total      Total cnt_males 2296 138307 2298 138427.4765
#> 47:  Total age_group1 cnt_males 1015  60859 1013  60739.0808
#> 48:  Total       ag1a cnt_males 1015  60859 1013  60739.0808
#> 49:  Total age_group2 cnt_males  571  34232  571  34232.0000
#> 50:  Total       ag2a cnt_males  571  34232  571  34232.0000
#> 51:  Total age_group3 cnt_males  424  25681  425  25741.5684
#> 52:  Total age_group4 cnt_males  195  12114  193  11989.7538
#> 53:  Total age_group5 cnt_males   84   5051   86   5171.2619
#> 54:  Total age_group6 cnt_males    7    370    8    422.8571
#> 55:   male      Total cnt_males 2296 138307 2298 138427.4765
#> 56:      m      Total cnt_males 2296 138307 2298 138427.4765
#> 57:   male age_group1 cnt_males 1015  60859 1013  60739.0808
#> 58:      m age_group1 cnt_males 1015  60859 1013  60739.0808
#> 59:   male       ag1a cnt_males 1015  60859 1013  60739.0808
#> 60:      m       ag1a cnt_males 1015  60859 1013  60739.0808
#> 61:   male age_group2 cnt_males  571  34232  571  34232.0000
#> 62:      m age_group2 cnt_males  571  34232  571  34232.0000
#> 63:   male       ag2a cnt_males  571  34232  571  34232.0000
#> 64:      m       ag2a cnt_males  571  34232  571  34232.0000
#> 65:   male age_group3 cnt_males  424  25681  425  25741.5684
#> 66:      m age_group3 cnt_males  424  25681  425  25741.5684
#> 67:   male age_group4 cnt_males  195  12114  193  11989.7538
#> 68:      m age_group4 cnt_males  195  12114  193  11989.7538
#> 69:   male age_group5 cnt_males   84   5051   86   5171.2619
#> 70:      m age_group5 cnt_males   84   5051   86   5171.2619
#> 71:   male age_group6 cnt_males    7    370    8    422.8571
#> 72:      m age_group6 cnt_males    7    370    8    422.8571
#> 73: female      Total cnt_males    0      0    0      0.0000
#> 74:      f      Total cnt_males    0      0    0      0.0000
#> 75: female age_group1 cnt_males    0      0    0      0.0000
#> 76:      f age_group1 cnt_males    0      0    0      0.0000
#> 77: female       ag1a cnt_males    0      0    0      0.0000
#> 78:      f       ag1a cnt_males    0      0    0      0.0000
#> 79: female age_group2 cnt_males    0      0    0      0.0000
#> 80:      f age_group2 cnt_males    0      0    0      0.0000
#> 81: female       ag2a cnt_males    0      0    0      0.0000
#> 82:      f       ag2a cnt_males    0      0    0      0.0000
#> 83: female age_group3 cnt_males    0      0    0      0.0000
#> 84:      f age_group3 cnt_males    0      0    0      0.0000
#> 85: female age_group4 cnt_males    0      0    0      0.0000
#> 86:      f age_group4 cnt_males    0      0    0      0.0000
#> 87: female age_group5 cnt_males    0      0    0      0.0000
#> 88:      f age_group5 cnt_males    0      0    0      0.0000
#> 89: female age_group6 cnt_males    0      0    0      0.0000
#> 90:      f age_group6 cnt_males    0      0    0      0.0000
#>        sex        age     vname  uwc     wc puwc         pwc

# utility measures for a count variable
tab$measures_cnts(v = "total", exclude_zeros = TRUE)
#> $overview
#>    noise cnt       pct
#> 1:    -2   5 0.1111111
#> 2:    -1   5 0.1111111
#> 3:     0  13 0.2888889
#> 4:     1  14 0.3111111
#> 5:     2   8 0.1777778
#> 
#> $measures
#>       what d1    d2    d3
#>  1:    Min  0 0.000 0.000
#>  2:    Q10  0 0.000 0.000
#>  3:    Q20  0 0.000 0.000
#>  4:    Q30  1 0.000 0.008
#>  5:    Q40  1 0.001 0.013
#>  6:   Mean  1 0.019 0.044
#>  7: Median  1 0.001 0.016
#>  8:    Q60  1 0.001 0.022
#>  9:    Q70  1 0.002 0.031
#> 10:    Q80  2 0.014 0.083
#> 11:    Q90  2 0.056 0.126
#> 12:    Q95  2 0.143 0.183
#> 13:    Q99  2 0.167 0.213
#> 14:    Max  2 0.167 0.213
#> 
#> $cumdistr_d1
#>    cat cnt       pct
#> 1:   0  13 0.2888889
#> 2:   1  32 0.7111111
#> 3:   2  45 1.0000000
#> 
#> $cumdistr_d2
#>            cat cnt       pct
#> 1:    [0,0.02]  36 0.8000000
#> 2: (0.02,0.05]  40 0.8888889
#> 3:  (0.05,0.1]  41 0.9111111
#> 4:   (0.1,0.2]  45 1.0000000
#> 5:   (0.2,0.3]  45 1.0000000
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $cumdistr_d3
#>            cat cnt       pct
#> 1:    [0,0.02]  25 0.5555556
#> 2: (0.02,0.05]  33 0.7333333
#> 3:  (0.05,0.1]  36 0.8000000
#> 4:   (0.1,0.2]  43 0.9555556
#> 5:   (0.2,0.3]  45 1.0000000
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $false_zero
#> [1] 0
#> 
#> $false_nonzero
#> [1] 0
#> 
#> $exclude_zeros
#> [1] TRUE
#> 

# modifications for perturbed count variables
tab$mod_cnts()
#>         sex        age row_nr pert       ckey  countvar
#>   1:  Total      Total     14   -1 0.09885913     total
#>   2:  Total age_group1     14   -1 0.29393464     total
#>   3:  Total       ag1a     14   -1 0.29393464     total
#>   4:  Total age_group2     14   -1 0.15531047     total
#>   5:  Total       ag2a     14   -1 0.15531047     total
#>  ---                                                   
#> 131:      f age_group4     -1    0 0.00000000 cnt_males
#> 132: female age_group5     -1    0 0.00000000 cnt_males
#> 133:      f age_group5     -1    0 0.00000000 cnt_males
#> 134: female age_group6     -1    0 0.00000000 cnt_males
#> 135:      f age_group6     -1    0 0.00000000 cnt_males

# display a summary about utility measures
tab$summary()
#> ┌──────────────────────────────────────────────┐
#> │Utility measures for perturbed count variables│
#> └──────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#>          countvar Min Q10  Q20 Q30 Q40   Mean Median Q60 Q70 Q80 Q90 Q95 Q99
#> 1:          total  -2  -2 -1.0  -1  -1 -0.333      0   0   0   1 1.6   2   2
#> 2: cnt_highincome  -3  -2 -1.0  -1  -1 -0.378      0   0   0   1 1.6   2   2
#> 3:      cnt_males  -2  -2 -0.4   0   0  0.000      0   0   0   1 2.0   2   2
#>    Max
#> 1:   2
#> 2:   2
#> 3:   2
#> 
#> ── Distance-based measures ─────────────────────────────────────────────────────
#> ✔ Variable: 'total'
#> 
#>       what d1    d2    d3
#>  1:    Min  0 0.000 0.000
#>  2:    Q10  0 0.000 0.000
#>  3:    Q20  0 0.000 0.000
#>  4:    Q30  1 0.000 0.008
#>  5:    Q40  1 0.001 0.013
#>  6:   Mean  1 0.019 0.044
#>  7: Median  1 0.001 0.016
#>  8:    Q60  1 0.001 0.022
#>  9:    Q70  1 0.002 0.031
#> 10:    Q80  2 0.014 0.083
#> 11:    Q90  2 0.056 0.126
#> 12:    Q95  2 0.143 0.183
#> 13:    Q99  2 0.167 0.213
#> 14:    Max  2 0.167 0.213
#> 
#> ✔ Variable: 'cnt_males'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.200 0.000 0.004
#>  4:    Q30 1.000 0.001 0.021
#>  5:    Q40 1.000 0.002 0.024
#>  6:   Mean 1.333 0.020 0.052
#>  7: Median 2.000 0.002 0.031
#>  8:    Q60 2.000 0.002 0.031
#>  9:    Q70 2.000 0.010 0.072
#> 10:    Q80 2.000 0.021 0.101
#> 11:    Q90 2.000 0.071 0.138
#> 12:    Q95 2.000 0.143 0.183
#> 13:    Q99 2.000 0.143 0.183
#> 14:    Max 2.000 0.143 0.183
#> 
#> ✔ Variable: 'cnt_highincome'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.800 0.002 0.019
#>  4:    Q30 1.000 0.009 0.053
#>  5:    Q40 1.000 0.011 0.062
#>  6:   Mean 1.175 0.032 0.084
#>  7: Median 1.000 0.015 0.067
#>  8:    Q60 1.000 0.018 0.068
#>  9:    Q70 2.000 0.020 0.089
#> 10:    Q80 2.000 0.024 0.101
#> 11:    Q90 2.000 0.075 0.142
#> 12:    Q95 2.050 0.114 0.244
#> 13:    Q99 3.000 0.286 0.410
#> 14:    Max 3.000 0.286 0.410
#> 
#> ┌──────────────────────────────────────────────────┐
#> │Utility measures for perturbed numerical variables│
#> └──────────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#>      vname         Min         Q10        Q20       Q30       Q40       Mean
#> 1:  expend         Inf          NA         NA        NA        NA        NaN
#> 2:  income -141236.091 -104219.411 -62925.584 -12349.32 -4367.517 -10576.470
#> 3: savings   -6364.113   -5829.503  -3370.276  -1904.19  -982.025   -358.036
#> 4:   mixed         Inf          NA         NA        NA        NA        NaN
#>       Median      Q60       Q70       Q80      Q90      Q95      Q99      Max
#> 1:        NA       NA        NA        NA       NA       NA       NA     -Inf
#> 2: -2599.854 8067.930 17380.074 47436.309 56707.21 63381.74 78218.60 78218.60
#> 3:  -467.077  299.973   642.276  2049.528  5306.18  5306.18  7357.37  7357.37
#> 4:        NA       NA        NA        NA       NA       NA       NA     -Inf
# }