R6 Class defining statistical tables that can be perturbed

This class allows to define statistical tables and perturb both count and numerical variables.

ck_setup(x, rkey, dims, w = NULL, countvars = NULL, numvars = NULL)

Arguments

x

an object coercible to a data.frame

rkey

either a column name within x referring to a variable containing record keys or a single integer(ish) number > 5 that referns to the number of digits for record keys that will be generated internally.

dims

a list containing slots for each variable that should be tabulated. Each slot consists should be created/modified using sdcHierarchies::hier_create(), sdcHierarchies::hier_add() and other functionality from package sdcHierarchies.

w

(character) a scalar character referring to a variable in x holding sampling weights. If w is NULL (the default), all weights are assumed to be 1

countvars

(character) an optional vector containing names of binary (0/1 coded) variables withing x that should be included in the problem instance. These variables can later be perturbed.

numvars

(character) an optional vector of numerical variables that can later be tabulated.

Details

Such objects are typically generated using ck_setup().

Methods

Public methods

ck_class$new()

ck_class$perturb()

ck_class$freqtab()

ck_class$numtab()

ck_class$measures_cnts()

ck_class$measures_nums()

ck_class$allvars()

ck_class$cntvars()

ck_class$numvars()

ck_class$hierarchy_info()

ck_class$mod_cnts()

ck_class$mod_nums()

ck_class$supp_freq()

ck_class$supp_val()

ck_class$supp_cells()

ck_class$supp_p()

ck_class$supp_pq()

ck_class$supp_nk()

ck_class$params_cnts_get()

ck_class$params_cnts_set()

ck_class$reset_cntvars()

ck_class$reset_numvars()

ck_class$reset_allvars()

ck_class$params_nums_get()

ck_class$params_nums_set()

ck_class$summary()

ck_class$print()

Method new()

Create a new table instance

Usage

ck_class$new(x, rkey, dims, w = NULL, countvars = NULL, numvars = NULL)

Arguments

x

an object coercible to a data.frame

rkey

either a column name within x referring to a variable containing record keys or a single integer(ish) number > 5 that referns to the number of digits for record keys that will be generated internally.

dims

a list containing slots for each variable that should be tabulated. Each slot consists should be created/modified using sdcHierarchies::hier_create(), sdcHierarchies::hier_add() and other functionality from package sdcHierarchies.

w

(character) a scalar character referring to a variable in x holding sampling weights. If w is NULL (the default), all weights are assumed to be 1

countvars

(character) an optional vector containing names of binary (0/1 coded) variables withing x that should be included in the problem instance. These variables can later be perturbed.

numvars

(character) an optional vector of numerical variables that can later be tabulated.

Returns

A new cellkey_obj object.

Method perturb()

Perturb a count- or magnitude variable

Usage

ck_class$perturb(v)

Arguments

v

name(s) of count- or magnitude variables that should be perturbed.

Method freqtab()

Extract results from already perturbed count variables as a data.table

Usage

ck_class$freqtab(v = NULL, path = NULL)

Arguments

v

a vector of variable names for count variables. If NULL (the default), the results are returned for all available count variables. For variables that have not yet perturbed, columns puwc and pwc are filled with NA.

path

if not NULL, a scalar character defining a (relative or absolute) path to which the result table should be written. A csv file will be generated and, if specified, path must have ".csv" as file-ending

Returns

This method returns a data.table containing all combinations of the dimensional variables in the first n columns. Additionally, the following columns are shown:

vname: name of the perturbed variable

uwc: unweighted counts

wc: weighted counts

puwc: perturbed unweighted counts or NA if vname was not yet perturbed

pwc: perturbed weighted counts or NA if vname was not yet perturbed

Method numtab()

Extract results from already perturbed continuous variables as a data.table.

Usage

ck_class$numtab(v = NULL, mean_before_sum = FALSE, path = NULL)

Arguments

v

a vector of variable names of continuous variables. If NULL (the default), the results are returned for all available numeric variables.

mean_before_sum

(logical); if TRUE, the perturbed values are adjusted by a factor ((n+p))⁄n with

n: the original weighted cell value

p: the perturbed cell value

This makes sense if the the accuracy of the variable mean is considered to be more important than accuracy of sums of the variable. The default value is FALSE (no adjustment is done)

path

if not NULL, a scalar character defining a (relative or absolute) path to which the result table should be written. A csv file will be generated and, if specified, path must have ".csv" as file-ending

Returns

This method returns a data.table containing all combinations of the dimensional variables in the first n columns. Additionally, the following columns are shown:

vname: name of the perturbed variable

uws: unweighted sum of the given variable

ws: weighted cellsum

pws: perturbed weighted sum of the given cell or NA if vname has not not perturbed

Method measures_cnts()

Utility measures for perturbed count variables

Usage

ck_class$measures_cnts(v, exclude_zeros = TRUE)

Arguments

v

name of a count variable for which utility measures should be computed.

exclude_zeros

should empty (zero) cells in the original values be excluded when computing distance measures

Returns

This method returns a list containing a set of utility measures based on some distance functions. For a detailed description of the computed measures, see ck_cnt_measures()

Method measures_nums()

Utility measures for continuous variables (not yet implemented)

Usage

ck_class$measures_nums(v)

Arguments

v

name of a continuosu variable for which utility measures should be computed.

Method allvars()

returns a list with the following two elements:

cntvars: character vector with names of available count variables for perturbation

numvars: character vector with names of available numerical variables for perturbation

Usage

ck_class$allvars()

Method cntvars()

character vector with names of count variables that could be perturbed

Usage

ck_class$cntvars()

Method numvars()

character vector with names of continuous variables that could be perturbed

Usage

ck_class$numvars()

Method hierarchy_info()

Information about hierarchies

Usage

ck_class$hierarchy_info()

Returns

a list (for each dimensional variable) with information on the hierarchies. This may be used to restrict output tables to specific levels or codes. Each list element is a data.table containing the following variables:

code: the name of a code within the hierarchy

level: number defining the level of the code; the higher the number, the lower the hierarchy with 1 being the overall total

is_leaf: if TRUE, this code is a leaf node which means no other codes contribute to it

parent: name of the parent code

Method mod_cnts()

Modifications applied to count variables

Usage

ck_class$mod_cnts()

Returns

a data.table containing modifications applied to count variables

Method mod_nums()

Modifications applied to numerical variables

Usage

ck_class$mod_nums()

Returns

a data.table containing modifications applied to numerical variables

Method supp_freq()

Identify sensitive cells based on minimum frequency rule

Usage

ck_class$supp_freq(v, n, weighted = TRUE)

Arguments

v

a single variable name of a continuous variable (see method numvars())

n

a number defining the threshold. All cells <= n are considered as unsafe.

weighted

if TRUE, the weighted number of contributors to a cell are compared to the threshold specified in n (default); else the unweighted number of contributors is used.

Method supp_val()

Identify sensitive cells based on weighted or unweighted cell value

Usage

ck_class$supp_val(v, n, weighted = TRUE)

Arguments

v

a single variable name of a continuous variable (see method numvars())

n

a number defining the threshold. All cells <= n are considered as unsafe.

weighted

if TRUE, the weighted cell value of variable v is compared to the threshold specified in n (default); else the unweighted number is used.

Method supp_cells()

Identify sensitive cells based on their names

Usage

ck_class$supp_cells(v, inp)

Arguments

v

a single variable name of a continuous variable (see method numvars())

inp

a data.frame where each colum represents a dimensional variable. Each row of this input is then used to compute the relevant cells to be identified as sensitive where NA-values are possible and used to match any characteristics of the dimensional variable.

Method supp_p()

Identify sensitive cells based on the p%-rule rule. Please note that this rule can only be applied to positive-only variables.

Usage

ck_class$supp_p(v, p)

Arguments

v

a single variable name of a continuous variable (see method numvars())

p

a number defining a percentage between 1 and 99.

Method supp_pq()

Identify sensitive cells based on the pq-rule. Please note that this rule can only be applied to positive-only variables.

Usage

ck_class$supp_pq(v, p, q)

Arguments

v

a single variable name of a continuous variable (see method numvars())

p

a number defining a percentage between 1 and 99.

q

a number defining a percentage between 1 and 99. This value must be larger than p.

Method supp_nk()

Identify sensitive cells based on the nk-dominance rule. Please note that this rule can only be applied to positive-only variables.

Usage

ck_class$supp_nk(v, n, k)

Arguments

v

a single variable name of a continuous variable (see method numvars())

n

an integerish number >= 2

k

a number defining a percentage between 1 and 99. All cells to which the top n contributers contribute more than k% is considered unsafe

Method params_cnts_get()

Return perturbation parameters of count variables

Usage

ck_class$params_cnts_get()

Returns

a named list in which each list-element contains the active perturbation parameters for the specific count variable defined by the list-name.

Method params_cnts_set()

Set perturbation parameters for count variables

Usage

ck_class$params_cnts_set(val, v = NULL)

Arguments

val

a perturbation object created with ck_params_cnts()

v

a character vector (or NULL). If NULL (the default), the perturbation parameters provided in val are set for all count variables; otherwise one may specify the names of the count variables for which the parameters should be set.

Method reset_cntvars()

reset results and parameters for already perturbed count variables

Usage

ck_class$reset_cntvars(v = NULL)

Arguments

v

if v equals NULL (the default), the results are reset for all perturbed count variables; otherwise it is possible to specify the names of already perturbed count variables.

Method reset_numvars()

reset results and parameters for already perturbed numerical variables

Usage

ck_class$reset_numvars(v = NULL)

Arguments

v

if v equals NULL (the default), the results are reset for all perturbed numerical variables; otherwise it is possible to specify the names of already perturbed continuous variables.

Method reset_allvars()

reset results and parameters for all already perturbed variables.

Usage

ck_class$reset_allvars()

Method params_nums_get()

Return perturbation parameters of continuous variables

Usage

ck_class$params_nums_get()

Returns

a named list in which each list-element contains the active perturbation parameters for the specific continuous variable defined by the list-name.

Method params_nums_set()

set perturbation parameters for continuous variables.

Usage

ck_class$params_nums_set(val, v = NULL)

Arguments

val

a perturbation object created with ck_params_nums()

v

a character vector (or NULL); if NULL (the default), the perturbation parameters provided in val are set for all continuous variables; otherwise one may specify the names of the numeric variables for which the parameters should be set.

Method summary()

some aggregated summary statistics about perturbed variables

Usage

ck_class$summary()

Method print()

prints information about the current table

Usage

ck_class$print()

Examples

x <- ck_create_testdata()

# create some 0/1 variables that should be perturbed later
x[, cnt_females := ifelse(sex == "male", 0, 1)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              93          25.00000           0
#>    2:   2530      28       1              94          25.00000           1
#>    3:   6920     550       1              65          25.00000           0
#>    4:   7960     870       1              35          25.00000           0
#>    5:   9030      20       2              79          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              79          16.66667           1
#> 4577:   1420     987    1000              32          16.66667           0
#> 4578:   8900     684    1000              55          16.66667           0
#> 4579:   3880     294    1000              30          16.66667           1
#> 4580:   4830     911    1000              36          16.66667           0
x[, cnt_males := ifelse(sex == "male", 1, 0)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              93          25.00000           0
#>    2:   2530      28       1              94          25.00000           1
#>    3:   6920     550       1              65          25.00000           0
#>    4:   7960     870       1              35          25.00000           0
#>    5:   9030      20       2              79          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              79          16.66667           1
#> 4577:   1420     987    1000              32          16.66667           0
#> 4578:   8900     684    1000              55          16.66667           0
#> 4579:   3880     294    1000              30          16.66667           1
#> 4580:   4830     911    1000              36          16.66667           0
#>       cnt_males
#>    1:         1
#>    2:         0
#>    3:         1
#>    4:         1
#>    5:         1
#>   ---          
#> 4576:         0
#> 4577:         1
#> 4578:         1
#> 4579:         0
#> 4580:         1
x[, cnt_highincome := ifelse(income >= 9000, 1, 0)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              93          25.00000           0
#>    2:   2530      28       1              94          25.00000           1
#>    3:   6920     550       1              65          25.00000           0
#>    4:   7960     870       1              35          25.00000           0
#>    5:   9030      20       2              79          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              79          16.66667           1
#> 4577:   1420     987    1000              32          16.66667           0
#> 4578:   8900     684    1000              55          16.66667           0
#> 4579:   3880     294    1000              30          16.66667           1
#> 4580:   4830     911    1000              36          16.66667           0
#>       cnt_males cnt_highincome
#>    1:         1              0
#>    2:         0              0
#>    3:         1              0
#>    4:         1              0
#>    5:         1              1
#>   ---                         
#> 4576:         0              0
#> 4577:         1              0
#> 4578:         1              0
#> 4579:         0              0
#> 4580:         1              0
# a variable with positive and negative contributions
x[, mixed := sample(-10:10, nrow(x), replace = TRUE)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              93          25.00000           0
#>    2:   2530      28       1              94          25.00000           1
#>    3:   6920     550       1              65          25.00000           0
#>    4:   7960     870       1              35          25.00000           0
#>    5:   9030      20       2              79          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              79          16.66667           1
#> 4577:   1420     987    1000              32          16.66667           0
#> 4578:   8900     684    1000              55          16.66667           0
#> 4579:   3880     294    1000              30          16.66667           1
#> 4580:   4830     911    1000              36          16.66667           0
#>       cnt_males cnt_highincome mixed
#>    1:         1              0     5
#>    2:         0              0    -2
#>    3:         1              0     8
#>    4:         1              0     4
#>    5:         1              1    -5
#>   ---                               
#> 4576:         0              0     7
#> 4577:         1              0     1
#> 4578:         1              0     1
#> 4579:         0              0    -4
#> 4580:         1              0    -1

# create record keys
x$rkey <- ck_generate_rkeys(dat = x)

# define required inputs

# hierarchy with some bogus codes
d_sex <- hier_create(root = "Total", nodes = c("male", "female"))
d_sex <- hier_add(d_sex, root = "female", "f")
d_sex <- hier_add(d_sex, root = "male", "m")

d_age <- hier_create(root = "Total", nodes = paste0("age_group", 1:6))
d_age <- hier_add(d_age, root = "age_group1", "ag1a")
d_age <- hier_add(d_age, root = "age_group2", "ag2a")

# define the cell key object
countvars <- c("cnt_females", "cnt_males", "cnt_highincome")
numvars <- c("expend", "income", "savings", "mixed")
tab <- ck_setup(
  x = x,
  rkey = "rkey",
  dims = list(sex = d_sex, age = d_age),
  w = "sampling_weight",
  countvars = countvars,
  numvars = numvars)

# show some information about this table instance
tab$print() # identical with print(tab)
#> ── Table Information ───────────────────────────────────────────────────────────
#> ✔ 45 cells in 2 dimensions ('sex', 'age')
#> ✔ weights: yes
#> ── Tabulated / Perturbed countvars ─────────────────────────────────────────────
#> ☐ 'total'
#> ☐ 'cnt_females'
#> ☐ 'cnt_males'
#> ☐ 'cnt_highincome'
#> ── Tabulated / Perturbed numvars ───────────────────────────────────────────────
#> ☐ 'expend'
#> ☐ 'income'
#> ☐ 'savings'
#> ☐ 'mixed'

# information about the hierarchies
tab$hierarchy_info()
#> $sex
#>      code level is_leaf parent
#> 1:  Total     1   FALSE  Total
#> 2:   male     2   FALSE  Total
#> 3:      m     3    TRUE   male
#> 4: female     2   FALSE  Total
#> 5:      f     3    TRUE female
#> 
#> $age
#>          code level is_leaf     parent
#> 1:      Total     1   FALSE      Total
#> 2: age_group1     2   FALSE      Total
#> 3:       ag1a     3    TRUE age_group1
#> 4: age_group2     2   FALSE      Total
#> 5:       ag2a     3    TRUE age_group2
#> 6: age_group3     2    TRUE      Total
#> 7: age_group4     2    TRUE      Total
#> 8: age_group5     2    TRUE      Total
#> 9: age_group6     2    TRUE      Total
#> 

# which variables have been defined?
tab$allvars()
#> $cntvars
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"
#> 
#> $numvars
#> [1] "expend"  "income"  "savings" "mixed"  
#> 

# count variables
tab$cntvars()
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"

# continuous variables
tab$numvars()
#> [1] "expend"  "income"  "savings" "mixed"  

# create perturbation parameters for "total" variable and
# write to yaml-file

# create a ptable using functionality from the ptable-pkg
f_yaml <- tempfile(fileext = ".yaml")
p_cnts1 <- ck_params_cnts(
  ptab = ptable::pt_ex_cnts(),
  path = f_yaml)
#> yaml configuration '/tmp/RtmpSL8ysM/file1e595e7d348a.yaml' successfully written.

# read parameters from yaml-file and set them for variable `"total"`
p_cnts1 <- ck_read_yaml(path = f_yaml)

tab$params_cnts_set(val = p_cnts1, v = "total")
#> --> setting perturbation parameters for variable 'total'

# create alternative perturbation parameters by specifying parameters
para2 <- ptable::create_cnt_ptable(
  D = 8, V = 3, js = 2, create = FALSE)

p_cnts2 <- ck_params_cnts(ptab = para2)

# use these ptable it for the remaining variables
tab$params_cnts_set(val = p_cnts2, v = countvars)
#> --> setting perturbation parameters for variable 'cnt_females'
#> --> setting perturbation parameters for variable 'cnt_males'
#> --> setting perturbation parameters for variable 'cnt_highincome'

# perturb a variable
tab$perturb(v = "total")
#> Count variable 'total' was perturbed.

# multiple variables can be perturbed as well
tab$perturb(v = c("cnt_males", "cnt_highincome"))
#> Count variable 'cnt_males' was perturbed.
#> Count variable 'cnt_highincome' was perturbed.

# return weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))
#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 272147 4581 272206.4207
#>  2:  Total age_group1     total 1969 117159 1971 117278.0036
#>  3:  Total       ag1a     total 1969 117159 1971 117278.0036
#>  4:  Total age_group2     total 1143  67810 1144  67869.3263
#>  5:  Total       ag2a     total 1143  67810 1144  67869.3263
#>  6:  Total age_group3     total  864  51802  863  51742.0440
#>  7:  Total age_group4     total  423  24598  424  24656.1513
#>  8:  Total age_group5     total  168  10057  168  10057.0000
#>  9:  Total age_group6     total   13    721   14    776.4615
#> 10:   male      Total     total 2296 137650 2296 137650.0000
#> 11:      m      Total     total 2296 137650 2296 137650.0000
#> 12:   male age_group1     total 1015  60860 1013  60740.0788
#> 13:      m age_group1     total 1015  60860 1013  60740.0788
#> 14:   male       ag1a     total 1015  60860 1013  60740.0788
#> 15:      m       ag1a     total 1015  60860 1013  60740.0788
#> 16:   male age_group2     total  571  34435  571  34435.0000
#> 17:      m age_group2     total  571  34435  571  34435.0000
#> 18:   male       ag2a     total  571  34435  571  34435.0000
#> 19:      m       ag2a     total  571  34435  571  34435.0000
#> 20:   male age_group3     total  424  25559  423  25498.7193
#> 21:      m age_group3     total  424  25559  423  25498.7193
#> 22:   male age_group4     total  195  11488  194  11429.0872
#> 23:      m age_group4     total  195  11488  194  11429.0872
#> 24:   male age_group5     total   84   4962   86   5080.1429
#> 25:      m age_group5     total   84   4962   86   5080.1429
#> 26:   male age_group6     total    7    346    8    395.4286
#> 27:      m age_group6     total    7    346    8    395.4286
#> 28: female      Total     total 2284 134497 2283 134438.1134
#> 29:      f      Total     total 2284 134497 2283 134438.1134
#> 30: female age_group1     total  954  56299  956  56417.0273
#> 31:      f age_group1     total  954  56299  956  56417.0273
#> 32: female       ag1a     total  954  56299  956  56417.0273
#> 33:      f       ag1a     total  954  56299  956  56417.0273
#> 34: female age_group2     total  572  33375  571  33316.6521
#> 35:      f age_group2     total  572  33375  571  33316.6521
#> 36: female       ag2a     total  572  33375  571  33316.6521
#> 37:      f       ag2a     total  572  33375  571  33316.6521
#> 38: female age_group3     total  440  26243  439  26183.3568
#> 39:      f age_group3     total  440  26243  439  26183.3568
#> 40: female age_group4     total  228  13110  228  13110.0000
#> 41:      f age_group4     total  228  13110  228  13110.0000
#> 42: female age_group5     total   84   5095   84   5095.0000
#> 43:      f age_group5     total   84   5095   84   5095.0000
#> 44: female age_group6     total    6    375    4    250.0000
#> 45:      f age_group6     total    6    375    4    250.0000
#> 46:  Total      Total cnt_males 2296 137650 2296 137650.0000
#> 47:  Total age_group1 cnt_males 1015  60860 1012  60680.1182
#> 48:  Total       ag1a cnt_males 1015  60860 1012  60680.1182
#> 49:  Total age_group2 cnt_males  571  34435  571  34435.0000
#> 50:  Total       ag2a cnt_males  571  34435  571  34435.0000
#> 51:  Total age_group3 cnt_males  424  25559  422  25438.4387
#> 52:  Total age_group4 cnt_males  195  11488  193  11370.1744
#> 53:  Total age_group5 cnt_males   84   4962   88   5198.2857
#> 54:  Total age_group6 cnt_males    7    346    9    444.8571
#> 55:   male      Total cnt_males 2296 137650 2296 137650.0000
#> 56:      m      Total cnt_males 2296 137650 2296 137650.0000
#> 57:   male age_group1 cnt_males 1015  60860 1012  60680.1182
#> 58:      m age_group1 cnt_males 1015  60860 1012  60680.1182
#> 59:   male       ag1a cnt_males 1015  60860 1012  60680.1182
#> 60:      m       ag1a cnt_males 1015  60860 1012  60680.1182
#> 61:   male age_group2 cnt_males  571  34435  571  34435.0000
#> 62:      m age_group2 cnt_males  571  34435  571  34435.0000
#> 63:   male       ag2a cnt_males  571  34435  571  34435.0000
#> 64:      m       ag2a cnt_males  571  34435  571  34435.0000
#> 65:   male age_group3 cnt_males  424  25559  422  25438.4387
#> 66:      m age_group3 cnt_males  424  25559  422  25438.4387
#> 67:   male age_group4 cnt_males  195  11488  193  11370.1744
#> 68:      m age_group4 cnt_males  195  11488  193  11370.1744
#> 69:   male age_group5 cnt_males   84   4962   88   5198.2857
#> 70:      m age_group5 cnt_males   84   4962   88   5198.2857
#> 71:   male age_group6 cnt_males    7    346    9    444.8571
#> 72:      m age_group6 cnt_males    7    346    9    444.8571
#> 73: female      Total cnt_males    0      0    0      0.0000
#> 74:      f      Total cnt_males    0      0    0      0.0000
#> 75: female age_group1 cnt_males    0      0    0      0.0000
#> 76:      f age_group1 cnt_males    0      0    0      0.0000
#> 77: female       ag1a cnt_males    0      0    0      0.0000
#> 78:      f       ag1a cnt_males    0      0    0      0.0000
#> 79: female age_group2 cnt_males    0      0    0      0.0000
#> 80:      f age_group2 cnt_males    0      0    0      0.0000
#> 81: female       ag2a cnt_males    0      0    0      0.0000
#> 82:      f       ag2a cnt_males    0      0    0      0.0000
#> 83: female age_group3 cnt_males    0      0    0      0.0000
#> 84:      f age_group3 cnt_males    0      0    0      0.0000
#> 85: female age_group4 cnt_males    0      0    0      0.0000
#> 86:      f age_group4 cnt_males    0      0    0      0.0000
#> 87: female age_group5 cnt_males    0      0    0      0.0000
#> 88:      f age_group5 cnt_males    0      0    0      0.0000
#> 89: female age_group6 cnt_males    0      0    0      0.0000
#> 90:      f age_group6 cnt_males    0      0    0      0.0000
#>        sex        age     vname  uwc     wc puwc         pwc

# numerical variables (positive variables using flex-function)
# we also write the config to a yaml file
f_yaml <- tempfile(fileext = ".yaml")

# create a ptable using functionality from the ptable-pkg
# a single ptable for all cells
ptab1 <- ptable::pt_ex_nums(parity = TRUE, separation = FALSE)

# a single ptab for all cells except for very small ones
ptab2 <- ptable::pt_ex_nums(parity = TRUE, separation = TRUE)

# different ptables for cells with even/odd number of contributors
# and very small cells
ptab3 <- ptable::pt_ex_nums(parity = FALSE, separation = TRUE)

p_nums1 <- ck_params_nums(
  ptab = ptab1,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.30, 0.03),
    epsilon = c(1, 0.5, 0.2),
    q = 3),
  mu_c = 2,
  same_key = FALSE,
  use_zero_rkeys = FALSE,
  path = f_yaml)
#> yaml configuration '/tmp/RtmpSL8ysM/file1e59449a0c5d.yaml' successfully written.

# we read the parameters from the yaml-file
p_nums1 <- ck_read_yaml(path = f_yaml)

# for variables with positive and negative values
p_nums2 <- ck_params_nums(
  ptab = ptab2,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.15, 0.02),
    epsilon = c(1, 0.4, 0.15),
    q = 3),
  mu_c = 2,
  same_key = FALSE)

# simple perturbation parameters (not using the flex-function approach)
p_nums3 <- ck_params_nums(
  ptab = ptab3,
  type = "mean",
  mult_params = ck_simpleparams(p = 0.25),
  mu_c = 2,
  same_key = FALSE)

# use `p_nums1` for all variables
tab$params_nums_set(p_nums1, c("savings", "income", "expend"))
#> --> setting perturbation parameters for variable 'savings'
#> --> setting perturbation parameters for variable 'income'
#> --> setting perturbation parameters for variable 'expend'

# use different parameters for variable `mixed`
tab$params_nums_set(p_nums2, v = "mixed")
#> --> setting perturbation parameters for variable 'mixed'

# identify sensitive cells to which extra protection (`mu_c`) is added.
tab$supp_p(v = "income", p = 85)
#> computing contributing indices | rawdata <--> table; this might take a while
#> p%-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_pq(v = "income", p = 85, q = 90)
#> computing contributing indices | rawdata <--> table; this might take a while
#> pq-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_nk(v = "income", n = 2, k = 90)
#> computing contributing indices | rawdata <--> table; this might take a while
#> nk-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_freq(v = "income", n = 14, weighted = FALSE)
#> freq-rule: 5 new sensitive cells (incl. duplicates) found (total: 5)
tab$supp_val(v = "income", n = 10000, weighted = TRUE)
#> val-rule: 0 new sensitive cells (incl. duplicates) found (total: 5)
tab$supp_cells(
  v = "income",
  inp = data.frame(
    sex = c("female", "female"),
    "age" = c("age_group1", "age_group3")
  )
)
#> cell-rule: 2 new sensitive cells (incl. duplicates) found (total: 7)

# perturb variables
tab$perturb(v = c("income", "savings"))
#> Numeric variable 'income' was perturbed.
#> Numeric variable 'savings' was perturbed.

# extract results
tab$numtab("income", mean_before_sum = TRUE)
#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1363363866 1363378898
#>  2:  Total age_group1 income  9810547  587662360  587678488
#>  3:  Total       ag1a income  9810547  587662360  587678488
#>  4:  Total age_group2 income  5692119  335328195  335386383
#>  5:  Total       ag2a income  5692119  335328195  335386383
#>  6:  Total age_group3 income  4406946  264368004  264262925
#>  7:  Total age_group4 income  2133543  120673584  120687172
#>  8:  Total age_group5 income   848151   51730091   51688146
#>  9:  Total age_group6 income    61672    3601632    3714859
#> 10:   male      Total income 11262049  675641341  675437097
#> 11:      m      Total income 11262049  675641341  675437097
#> 12:   male age_group1 income  4877164  294617873  294596885
#> 13:      m age_group1 income  4877164  294617873  294596885
#> 14:   male       ag1a income  4877164  294617873  294596885
#> 15:      m       ag1a income  4877164  294617873  294596885
#> 16:   male age_group2 income  2811379  168430924  168601476
#> 17:      m age_group2 income  2811379  168430924  168601476
#> 18:   male       ag2a income  2811379  168430924  168601476
#> 19:      m       ag2a income  2811379  168430924  168601476
#> 20:   male age_group3 income  2168169  130634402  130764152
#> 21:      m age_group3 income  2168169  130634402  130764152
#> 22:   male age_group4 income   978510   56259275   56238490
#> 23:      m age_group4 income   978510   56259275   56238490
#> 24:   male age_group5 income   393134   24011397   23971433
#> 25:      m age_group5 income   393134   24011397   23971433
#> 26:   male age_group6 income    33693    1687470    1717224
#> 27:      m age_group6 income    33693    1687470    1717224
#> 28: female      Total income 11690929  687722525  687727920
#> 29:      f      Total income 11690929  687722525  687727920
#> 30: female age_group1 income  4933383  293044487  293062462
#> 31:      f age_group1 income  4933383  293044487  293062462
#> 32: female       ag1a income  4933383  293044487  293062462
#> 33:      f       ag1a income  4933383  293044487  293062462
#> 34: female age_group2 income  2880740  166897271  166872458
#> 35:      f age_group2 income  2880740  166897271  166872458
#> 36: female       ag2a income  2880740  166897271  166872458
#> 37:      f       ag2a income  2880740  166897271  166872458
#> 38: female age_group3 income  2238777  133733602  133687660
#> 39:      f age_group3 income  2238777  133733602  133687660
#> 40: female age_group4 income  1155033   64414309   64371420
#> 41:      f age_group4 income  1155033   64414309   64371420
#> 42: female age_group5 income   455017   27718694   27694675
#> 43:      f age_group5 income   455017   27718694   27694675
#> 44: female age_group6 income    27979    1914162    1618918
#> 45:      f age_group6 income    27979    1914162    1618918
#>        sex        age  vname      uws         ws        pws
tab$numtab("income", mean_before_sum = FALSE)
#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1363363866 1363371382
#>  2:  Total age_group1 income  9810547  587662360  587670424
#>  3:  Total       ag1a income  9810547  587662360  587670424
#>  4:  Total age_group2 income  5692119  335328195  335357288
#>  5:  Total       ag2a income  5692119  335328195  335357288
#>  6:  Total age_group3 income  4406946  264368004  264315459
#>  7:  Total age_group4 income  2133543  120673584  120680378
#>  8:  Total age_group5 income   848151   51730091   51709114
#>  9:  Total age_group6 income    61672    3601632    3657808
#> 10:   male      Total income 11262049  675641341  675539212
#> 11:      m      Total income 11262049  675641341  675539212
#> 12:   male age_group1 income  4877164  294617873  294607379
#> 13:      m age_group1 income  4877164  294617873  294607379
#> 14:   male       ag1a income  4877164  294617873  294607379
#> 15:      m       ag1a income  4877164  294617873  294607379
#> 16:   male age_group2 income  2811379  168430924  168516179
#> 17:      m age_group2 income  2811379  168430924  168516179
#> 18:   male       ag2a income  2811379  168430924  168516179
#> 19:      m       ag2a income  2811379  168430924  168516179
#> 20:   male age_group3 income  2168169  130634402  130699261
#> 21:      m age_group3 income  2168169  130634402  130699261
#> 22:   male age_group4 income   978510   56259275   56248881
#> 23:      m age_group4 income   978510   56259275   56248881
#> 24:   male age_group5 income   393134   24011397   23991406
#> 25:      m age_group5 income   393134   24011397   23991406
#> 26:   male age_group6 income    33693    1687470    1702282
#> 27:      m age_group6 income    33693    1687470    1702282
#> 28: female      Total income 11690929  687722525  687725223
#> 29:      f      Total income 11690929  687722525  687725223
#> 30: female age_group1 income  4933383  293044487  293053474
#> 31:      f age_group1 income  4933383  293044487  293053474
#> 32: female       ag1a income  4933383  293044487  293053474
#> 33:      f       ag1a income  4933383  293044487  293053474
#> 34: female age_group2 income  2880740  166897271  166884864
#> 35:      f age_group2 income  2880740  166897271  166884864
#> 36: female       ag2a income  2880740  166897271  166884864
#> 37:      f       ag2a income  2880740  166897271  166884864
#> 38: female age_group3 income  2238777  133733602  133710629
#> 39:      f age_group3 income  2238777  133733602  133710629
#> 40: female age_group4 income  1155033   64414309   64392861
#> 41:      f age_group4 income  1155033   64414309   64392861
#> 42: female age_group5 income   455017   27718694   27706682
#> 43:      f age_group5 income   455017   27718694   27706682
#> 44: female age_group6 income    27979    1914162    1760361
#> 45:      f age_group6 income    27979    1914162    1760361
#>        sex        age  vname      uws         ws        pws
tab$numtab("savings")
#>        sex        age   vname     uws        ws         pws
#>  1:  Total      Total savings 2273532 135121655 135122266.2
#>  2:  Total age_group1 savings  982386  58638786  58635063.4
#>  3:  Total       ag1a savings  982386  58638786  58635063.4
#>  4:  Total age_group2 savings  552336  32730875  32732897.1
#>  5:  Total       ag2a savings  552336  32730875  32732897.1
#>  6:  Total age_group3 savings  437101  26406244  26398831.7
#>  7:  Total age_group4 savings  214661  12286510  12287454.1
#>  8:  Total age_group5 savings   80451   4662873   4658876.4
#>  9:  Total age_group6 savings    6597    396367    398623.4
#> 10:   male      Total savings 1159816  69585480  69572878.5
#> 11:      m      Total savings 1159816  69585480  69572878.5
#> 12:   male age_group1 savings  517660  30845782  30844551.2
#> 13:      m age_group1 savings  517660  30845782  30844551.2
#> 14:   male       ag1a savings  517660  30845782  30844551.2
#> 15:      m       ag1a savings  517660  30845782  30844551.2
#> 16:   male age_group2 savings  280923  17201683  17210542.2
#> 17:      m age_group2 savings  280923  17201683  17210542.2
#> 18:   male       ag2a savings  280923  17201683  17210542.2
#> 19:      m       ag2a savings  280923  17201683  17210542.2
#> 20:   male age_group3 savings  214970  13003251  13009743.0
#> 21:      m age_group3 savings  214970  13003251  13009743.0
#> 22:   male age_group4 savings   99420   5796927   5796435.9
#> 23:      m age_group4 savings   99420   5796927   5796435.9
#> 24:   male age_group5 savings   43233   2558547   2561948.0
#> 25:      m age_group5 savings   43233   2558547   2561948.0
#> 26:   male age_group6 savings    3610    179290    178854.5
#> 27:      m age_group6 savings    3610    179290    178854.5
#> 28: female      Total savings 1113716  65536175  65534781.0
#> 29:      f      Total savings 1113716  65536175  65534781.0
#> 30: female age_group1 savings  464726  27793004  27792311.7
#> 31:      f age_group1 savings  464726  27793004  27792311.7
#> 32: female       ag1a savings  464726  27793004  27792311.7
#> 33:      f       ag1a savings  464726  27793004  27792311.7
#> 34: female age_group2 savings  271413  15529192  15528536.7
#> 35:      f age_group2 savings  271413  15529192  15528536.7
#> 36: female       ag2a savings  271413  15529192  15528536.7
#> 37:      f       ag2a savings  271413  15529192  15528536.7
#> 38: female age_group3 savings  222131  13402993  13396829.4
#> 39:      f age_group3 savings  222131  13402993  13396829.4
#> 40: female age_group4 savings  115241   6489583   6489508.5
#> 41:      f age_group4 savings  115241   6489583   6489508.5
#> 42: female age_group5 savings   37218   2104326   2103137.1
#> 43:      f age_group5 savings   37218   2104326   2103137.1
#> 44: female age_group6 savings    2987    217077    205695.6
#> 45:      f age_group6 savings    2987    217077    205695.6
#>        sex        age   vname     uws        ws         pws

# results can be resetted, too
tab$reset_cntvars(v = "cnt_males")

# we can then set other parameters and perturb again
tab$params_cnts_set(val = p_cnts1, v = "cnt_males")
#> --> setting perturbation parameters for variable 'cnt_males'
tab$perturb(v = "cnt_males")
#> Count variable 'cnt_males' was perturbed.

# write to a file "outtab.csv" (.csv is automatically added to the path)
if (FALSE) {
tab$freqtab(v = c("total", "cnt_males"), path = "outtab.csv")
}

# show results containing weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))
#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 272147 4581 272206.4207
#>  2:  Total age_group1     total 1969 117159 1971 117278.0036
#>  3:  Total       ag1a     total 1969 117159 1971 117278.0036
#>  4:  Total age_group2     total 1143  67810 1144  67869.3263
#>  5:  Total       ag2a     total 1143  67810 1144  67869.3263
#>  6:  Total age_group3     total  864  51802  863  51742.0440
#>  7:  Total age_group4     total  423  24598  424  24656.1513
#>  8:  Total age_group5     total  168  10057  168  10057.0000
#>  9:  Total age_group6     total   13    721   14    776.4615
#> 10:   male      Total     total 2296 137650 2296 137650.0000
#> 11:      m      Total     total 2296 137650 2296 137650.0000
#> 12:   male age_group1     total 1015  60860 1013  60740.0788
#> 13:      m age_group1     total 1015  60860 1013  60740.0788
#> 14:   male       ag1a     total 1015  60860 1013  60740.0788
#> 15:      m       ag1a     total 1015  60860 1013  60740.0788
#> 16:   male age_group2     total  571  34435  571  34435.0000
#> 17:      m age_group2     total  571  34435  571  34435.0000
#> 18:   male       ag2a     total  571  34435  571  34435.0000
#> 19:      m       ag2a     total  571  34435  571  34435.0000
#> 20:   male age_group3     total  424  25559  423  25498.7193
#> 21:      m age_group3     total  424  25559  423  25498.7193
#> 22:   male age_group4     total  195  11488  194  11429.0872
#> 23:      m age_group4     total  195  11488  194  11429.0872
#> 24:   male age_group5     total   84   4962   86   5080.1429
#> 25:      m age_group5     total   84   4962   86   5080.1429
#> 26:   male age_group6     total    7    346    8    395.4286
#> 27:      m age_group6     total    7    346    8    395.4286
#> 28: female      Total     total 2284 134497 2283 134438.1134
#> 29:      f      Total     total 2284 134497 2283 134438.1134
#> 30: female age_group1     total  954  56299  956  56417.0273
#> 31:      f age_group1     total  954  56299  956  56417.0273
#> 32: female       ag1a     total  954  56299  956  56417.0273
#> 33:      f       ag1a     total  954  56299  956  56417.0273
#> 34: female age_group2     total  572  33375  571  33316.6521
#> 35:      f age_group2     total  572  33375  571  33316.6521
#> 36: female       ag2a     total  572  33375  571  33316.6521
#> 37:      f       ag2a     total  572  33375  571  33316.6521
#> 38: female age_group3     total  440  26243  439  26183.3568
#> 39:      f age_group3     total  440  26243  439  26183.3568
#> 40: female age_group4     total  228  13110  228  13110.0000
#> 41:      f age_group4     total  228  13110  228  13110.0000
#> 42: female age_group5     total   84   5095   84   5095.0000
#> 43:      f age_group5     total   84   5095   84   5095.0000
#> 44: female age_group6     total    6    375    4    250.0000
#> 45:      f age_group6     total    6    375    4    250.0000
#> 46:  Total      Total cnt_males 2296 137650 2296 137650.0000
#> 47:  Total age_group1 cnt_males 1015  60860 1013  60740.0788
#> 48:  Total       ag1a cnt_males 1015  60860 1013  60740.0788
#> 49:  Total age_group2 cnt_males  571  34435  571  34435.0000
#> 50:  Total       ag2a cnt_males  571  34435  571  34435.0000
#> 51:  Total age_group3 cnt_males  424  25559  423  25498.7193
#> 52:  Total age_group4 cnt_males  195  11488  194  11429.0872
#> 53:  Total age_group5 cnt_males   84   4962   86   5080.1429
#> 54:  Total age_group6 cnt_males    7    346    8    395.4286
#> 55:   male      Total cnt_males 2296 137650 2296 137650.0000
#> 56:      m      Total cnt_males 2296 137650 2296 137650.0000
#> 57:   male age_group1 cnt_males 1015  60860 1013  60740.0788
#> 58:      m age_group1 cnt_males 1015  60860 1013  60740.0788
#> 59:   male       ag1a cnt_males 1015  60860 1013  60740.0788
#> 60:      m       ag1a cnt_males 1015  60860 1013  60740.0788
#> 61:   male age_group2 cnt_males  571  34435  571  34435.0000
#> 62:      m age_group2 cnt_males  571  34435  571  34435.0000
#> 63:   male       ag2a cnt_males  571  34435  571  34435.0000
#> 64:      m       ag2a cnt_males  571  34435  571  34435.0000
#> 65:   male age_group3 cnt_males  424  25559  423  25498.7193
#> 66:      m age_group3 cnt_males  424  25559  423  25498.7193
#> 67:   male age_group4 cnt_males  195  11488  194  11429.0872
#> 68:      m age_group4 cnt_males  195  11488  194  11429.0872
#> 69:   male age_group5 cnt_males   84   4962   86   5080.1429
#> 70:      m age_group5 cnt_males   84   4962   86   5080.1429
#> 71:   male age_group6 cnt_males    7    346    8    395.4286
#> 72:      m age_group6 cnt_males    7    346    8    395.4286
#> 73: female      Total cnt_males    0      0    0      0.0000
#> 74:      f      Total cnt_males    0      0    0      0.0000
#> 75: female age_group1 cnt_males    0      0    0      0.0000
#> 76:      f age_group1 cnt_males    0      0    0      0.0000
#> 77: female       ag1a cnt_males    0      0    0      0.0000
#> 78:      f       ag1a cnt_males    0      0    0      0.0000
#> 79: female age_group2 cnt_males    0      0    0      0.0000
#> 80:      f age_group2 cnt_males    0      0    0      0.0000
#> 81: female       ag2a cnt_males    0      0    0      0.0000
#> 82:      f       ag2a cnt_males    0      0    0      0.0000
#> 83: female age_group3 cnt_males    0      0    0      0.0000
#> 84:      f age_group3 cnt_males    0      0    0      0.0000
#> 85: female age_group4 cnt_males    0      0    0      0.0000
#> 86:      f age_group4 cnt_males    0      0    0      0.0000
#> 87: female age_group5 cnt_males    0      0    0      0.0000
#> 88:      f age_group5 cnt_males    0      0    0      0.0000
#> 89: female age_group6 cnt_males    0      0    0      0.0000
#> 90:      f age_group6 cnt_males    0      0    0      0.0000
#>        sex        age     vname  uwc     wc puwc         pwc

# utility measures for a count variable
tab$measures_cnts(v = "total", exclude_zeros = TRUE)
#> $overview
#>    noise cnt       pct
#> 1:    -2   8 0.1777778
#> 2:    -1   7 0.1555556
#> 3:     0  11 0.2444444
#> 4:     1  13 0.2888889
#> 5:     2   6 0.1333333
#> 
#> $measures
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 1.000 0.001 0.011
#>  5:    Q40 1.000 0.001 0.021
#>  6:   Mean 1.067 0.025 0.050
#>  7: Median 1.000 0.002 0.023
#>  8:    Q60 1.000 0.002 0.024
#>  9:    Q70 1.800 0.002 0.031
#> 10:    Q80 2.000 0.003 0.033
#> 11:    Q90 2.000 0.056 0.125
#> 12:    Q95 2.000 0.143 0.183
#> 13:    Q99 2.000 0.333 0.449
#> 14:    Max 2.000 0.333 0.449
#> 
#> $cumdistr_d1
#>    cat cnt       pct
#> 1:   0  11 0.2444444
#> 2:   1  31 0.6888889
#> 3:   2  45 1.0000000
#> 
#> $cumdistr_d2
#>            cat cnt       pct
#> 1:    [0,0.02]  38 0.8444444
#> 2: (0.02,0.05]  40 0.8888889
#> 3:  (0.05,0.1]  41 0.9111111
#> 4:   (0.1,0.2]  43 0.9555556
#> 5:   (0.2,0.3]  43 0.9555556
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $cumdistr_d3
#>            cat cnt       pct
#> 1:    [0,0.02]  17 0.3777778
#> 2: (0.02,0.05]  38 0.8444444
#> 3:  (0.05,0.1]  38 0.8444444
#> 4:   (0.1,0.2]  43 0.9555556
#> 5:   (0.2,0.3]  43 0.9555556
#> 6:   (0.3,0.4]  43 0.9555556
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $false_zero
#> [1] 0
#> 
#> $false_nonzero
#> [1] 0
#> 
#> $exclude_zeros
#> [1] TRUE
#> 

# modifications for perturbed count variables
tab$mod_cnts()
#>         sex        age row_nr pert      ckey  countvar
#>   1:  Total      Total     16    1 0.8479095     total
#>   2:  Total age_group1     17    2 0.9821003     total
#>   3:  Total       ag1a     17    2 0.9821003     total
#>   4:  Total age_group2     16    1 0.7389549     total
#>   5:  Total       ag2a     16    1 0.7389549     total
#>  ---                                                  
#> 131:      f age_group4     -1    0 0.0000000 cnt_males
#> 132: female age_group5     -1    0 0.0000000 cnt_males
#> 133:      f age_group5     -1    0 0.0000000 cnt_males
#> 134: female age_group6     -1    0 0.0000000 cnt_males
#> 135:      f age_group6     -1    0 0.0000000 cnt_males

# display a summary about utility measures
tab$summary()
#> ┌──────────────────────────────────────────────┐
#> │Utility measures for perturbed count variables│
#> └──────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#>          countvar Min Q10 Q20 Q30 Q40   Mean Median Q60 Q70 Q80 Q90 Q95 Q99 Max
#> 1:          total  -2  -2  -1  -1  -1 -0.044      0   0   1   1   2 2.0   2   2
#> 2: cnt_highincome  -4  -3  -2  -1  -1 -0.756     -1   0   0   0   1 1.8   2   2
#> 3:      cnt_males  -2  -2  -1   0   0 -0.200      0   0   0   0   1 1.8   2   2
#> 
#> ── Distance-based measures ─────────────────────────────────────────────────────
#> ✔ Variable: 'total'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 1.000 0.001 0.011
#>  5:    Q40 1.000 0.001 0.021
#>  6:   Mean 1.067 0.025 0.050
#>  7: Median 1.000 0.002 0.023
#>  8:    Q60 1.000 0.002 0.024
#>  9:    Q70 1.800 0.002 0.031
#> 10:    Q80 2.000 0.003 0.033
#> 11:    Q90 2.000 0.056 0.125
#> 12:    Q95 2.000 0.143 0.183
#> 13:    Q99 2.000 0.333 0.449
#> 14:    Max 2.000 0.333 0.449
#> 
#> ✔ Variable: 'cnt_males'
#> 
#>       what d1    d2    d3
#>  1:    Min  0 0.000 0.000
#>  2:    Q10  0 0.000 0.000
#>  3:    Q20  0 0.000 0.000
#>  4:    Q30  0 0.000 0.000
#>  5:    Q40  1 0.002 0.024
#>  6:   Mean  1 0.020 0.046
#>  7: Median  1 0.002 0.031
#>  8:    Q60  1 0.002 0.031
#>  9:    Q70  2 0.005 0.036
#> 10:    Q80  2 0.020 0.094
#> 11:    Q90  2 0.071 0.138
#> 12:    Q95  2 0.143 0.183
#> 13:    Q99  2 0.143 0.183
#> 14:    Max  2 0.143 0.183
#> 
#> ✔ Variable: 'cnt_highincome'
#> 
#>       what   d1    d2    d3
#>  1:    Min 0.00 0.000 0.000
#>  2:    Q10 0.00 0.000 0.000
#>  3:    Q20 0.00 0.000 0.000
#>  4:    Q30 1.00 0.005 0.034
#>  5:    Q40 1.00 0.011 0.053
#>  6:   Mean 1.35 0.045 0.112
#>  7: Median 1.00 0.029 0.131
#>  8:    Q60 1.40 0.035 0.134
#>  9:    Q70 2.00 0.049 0.150
#> 10:    Q80 2.00 0.073 0.183
#> 11:    Q90 3.00 0.143 0.236
#> 12:    Q95 3.00 0.143 0.239
#> 13:    Q99 3.61 0.230 0.447
#> 14:    Max 4.00 0.286 0.579
#> 
#> ┌──────────────────────────────────────────────────┐
#> │Utility measures for perturbed numerical variables│
#> └──────────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#>      vname       Min        Q10        Q20       Q30        Q40      Mean
#> 1:  expend       Inf         NA         NA        NA         NA       NaN
#> 2:  income -153801.1 -40715.807 -21071.000 -12406.88 -11101.552 -3647.073
#> 3: savings  -12601.5  -6912.843  -3722.592  -1230.81   -890.966  -745.558
#> 4:   mixed       Inf         NA         NA        NA         NA       NaN
#>        Median      Q60      Q70       Q80       Q90       Q95       Q99
#> 1:         NA       NA       NA        NA        NA        NA        NA
#> 2: -10393.523 7082.747 8987.189 17668.140 64858.752 85254.672 85254.672
#> 3:   -655.313 -491.141  -74.463  2068.967  6492.035  8859.205  8859.205
#> 4:         NA       NA       NA        NA        NA        NA        NA
#>          Max
#> 1:      -Inf
#> 2: 85254.672
#> 3:  8859.205
#> 4:      -Inf