Set parameters required to perturb numeric variables using a flex function

ck_flexparams() allows to define a flex function that is used to lookup perturbation magnitudes (percentages) used when perturbing continuous variables.

ck_flexparams(fp, p = c(0.25, 0.05), epsilon = 1, q = 3)

Arguments

fp

(numeric scalar); at which point should the noise coefficient function reaches its desired maximum (defined by the first element of p)

p

a numeric vector of length 2 where both elements specify a percentage. The first value refers to the desired maximum perturbation percentage for small cells (depending on fp) while the second element refers to the desired maximum perturbation percentage for large cells. Both values must be between 0 and 1 and need to be in descending order.

epsilon

a numeric vector in descending order with all values >= 0 and <= 1 with the first element forced to equal 1. The length of this vector must correspond with the number top_k specified in ck_params_nums() when creating parameters for type == "top_contr" which is checked at runtime. This setting allows to use different flex-functions for the largest top_k contributors.

q

(numeric scalar); Parameter of the function; q needs to be >= 1

Value

an object suitable as input for ck_params_nums().

Details

details about the flex function can be found in Deliverable D4.2, Part I in SGA "Open Source tools for perturbative confidentiality methods"

See also

Examples

# \donttest{
x <- ck_create_testdata()

# create some 0/1 variables that should be perturbed later
x[, cnt_females := ifelse(sex == "male", 0, 1)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              62          25.00000           0
#>    2:   2530      28       1              47          25.00000           1
#>    3:   6920     550       1              35          25.00000           0
#>    4:   7960     870       1              53          25.00000           0
#>    5:   9030      20       2              43          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              73          16.66667           1
#> 4577:   1420     987    1000              36          16.66667           0
#> 4578:   8900     684    1000              23          16.66667           0
#> 4579:   3880     294    1000              48          16.66667           1
#> 4580:   4830     911    1000              99          16.66667           0
x[, cnt_males := ifelse(sex == "male", 1, 0)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              62          25.00000           0
#>    2:   2530      28       1              47          25.00000           1
#>    3:   6920     550       1              35          25.00000           0
#>    4:   7960     870       1              53          25.00000           0
#>    5:   9030      20       2              43          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              73          16.66667           1
#> 4577:   1420     987    1000              36          16.66667           0
#> 4578:   8900     684    1000              23          16.66667           0
#> 4579:   3880     294    1000              48          16.66667           1
#> 4580:   4830     911    1000              99          16.66667           0
#>       cnt_males
#>    1:         1
#>    2:         0
#>    3:         1
#>    4:         1
#>    5:         1
#>   ---          
#> 4576:         0
#> 4577:         1
#> 4578:         1
#> 4579:         0
#> 4580:         1
x[, cnt_highincome := ifelse(income >= 9000, 1, 0)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              62          25.00000           0
#>    2:   2530      28       1              47          25.00000           1
#>    3:   6920     550       1              35          25.00000           0
#>    4:   7960     870       1              53          25.00000           0
#>    5:   9030      20       2              43          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              73          16.66667           1
#> 4577:   1420     987    1000              36          16.66667           0
#> 4578:   8900     684    1000              23          16.66667           0
#> 4579:   3880     294    1000              48          16.66667           1
#> 4580:   4830     911    1000              99          16.66667           0
#>       cnt_males cnt_highincome
#>    1:         1              0
#>    2:         0              0
#>    3:         1              0
#>    4:         1              0
#>    5:         1              1
#>   ---                         
#> 4576:         0              0
#> 4577:         1              0
#> 4578:         1              0
#> 4579:         0              0
#> 4580:         1              0
# a variable with positive and negative contributions
x[, mixed := sample(-10:10, nrow(x), replace = TRUE)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              62          25.00000           0
#>    2:   2530      28       1              47          25.00000           1
#>    3:   6920     550       1              35          25.00000           0
#>    4:   7960     870       1              53          25.00000           0
#>    5:   9030      20       2              43          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              73          16.66667           1
#> 4577:   1420     987    1000              36          16.66667           0
#> 4578:   8900     684    1000              23          16.66667           0
#> 4579:   3880     294    1000              48          16.66667           1
#> 4580:   4830     911    1000              99          16.66667           0
#>       cnt_males cnt_highincome mixed
#>    1:         1              0    -3
#>    2:         0              0    -4
#>    3:         1              0     0
#>    4:         1              0    -3
#>    5:         1              1    -4
#>   ---                               
#> 4576:         0              0    10
#> 4577:         1              0    -4
#> 4578:         1              0    -8
#> 4579:         0              0    -8
#> 4580:         1              0     3

# create record keys
x$rkey <- ck_generate_rkeys(dat = x)

# define required inputs

# hierarchy with some bogus codes
d_sex <- hier_create(root = "Total", nodes = c("male", "female"))
d_sex <- hier_add(d_sex, root = "female", "f")
d_sex <- hier_add(d_sex, root = "male", "m")

d_age <- hier_create(root = "Total", nodes = paste0("age_group", 1:6))
d_age <- hier_add(d_age, root = "age_group1", "ag1a")
d_age <- hier_add(d_age, root = "age_group2", "ag2a")

# define the cell key object
countvars <- c("cnt_females", "cnt_males", "cnt_highincome")
numvars <- c("expend", "income", "savings", "mixed")
tab <- ck_setup(
  x = x,
  rkey = "rkey",
  dims = list(sex = d_sex, age = d_age),
  w = "sampling_weight",
  countvars = countvars,
  numvars = numvars)

# show some information about this table instance
tab$print() # identical with print(tab)
#> ── Table Information ───────────────────────────────────────────────────────────
#> ✔ 45 cells in 2 dimensions ('sex', 'age')
#> ✔ weights: yes
#> ── Tabulated / Perturbed countvars ─────────────────────────────────────────────
#> ☐ 'total'
#> ☐ 'cnt_females'
#> ☐ 'cnt_males'
#> ☐ 'cnt_highincome'
#> ── Tabulated / Perturbed numvars ───────────────────────────────────────────────
#> ☐ 'expend'
#> ☐ 'income'
#> ☐ 'savings'
#> ☐ 'mixed'

# information about the hierarchies
tab$hierarchy_info()
#> $sex
#>      code level is_leaf parent
#> 1:  Total     1   FALSE  Total
#> 2:   male     2   FALSE  Total
#> 3:      m     3    TRUE   male
#> 4: female     2   FALSE  Total
#> 5:      f     3    TRUE female
#> 
#> $age
#>          code level is_leaf     parent
#> 1:      Total     1   FALSE      Total
#> 2: age_group1     2   FALSE      Total
#> 3:       ag1a     3    TRUE age_group1
#> 4: age_group2     2   FALSE      Total
#> 5:       ag2a     3    TRUE age_group2
#> 6: age_group3     2    TRUE      Total
#> 7: age_group4     2    TRUE      Total
#> 8: age_group5     2    TRUE      Total
#> 9: age_group6     2    TRUE      Total
#> 

# which variables have been defined?
tab$allvars()
#> $cntvars
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"
#> 
#> $numvars
#> [1] "expend"  "income"  "savings" "mixed"  
#> 

# count variables
tab$cntvars()
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"

# continuous variables
tab$numvars()
#> [1] "expend"  "income"  "savings" "mixed"  

# create perturbation parameters for "total" variable and
# write to yaml-file

# create a ptable using functionality from the ptable-pkg
f_yaml <- tempfile(fileext = ".yaml")
p_cnts1 <- ck_params_cnts(
  ptab = ptable::pt_ex_cnts(),
  path = f_yaml)
#> yaml configuration '/tmp/RtmpojDFXK/file19cd68c43145.yaml' successfully written.

# read parameters from yaml-file and set them for variable `"total"`
p_cnts1 <- ck_read_yaml(path = f_yaml)

tab$params_cnts_set(val = p_cnts1, v = "total")
#> --> setting perturbation parameters for variable 'total'

# create alternative perturbation parameters by specifying parameters
para2 <- ptable::create_cnt_ptable(
  D = 8, V = 3, js = 2, create = FALSE)

p_cnts2 <- ck_params_cnts(ptab = para2)

# use these ptable it for the remaining variables
tab$params_cnts_set(val = p_cnts2, v = countvars)
#> --> setting perturbation parameters for variable 'cnt_females'
#> --> setting perturbation parameters for variable 'cnt_males'
#> --> setting perturbation parameters for variable 'cnt_highincome'

# perturb a variable
tab$perturb(v = "total")
#> Count variable 'total' was perturbed.

# multiple variables can be perturbed as well
tab$perturb(v = c("cnt_males", "cnt_highincome"))
#> Count variable 'cnt_males' was perturbed.
#> Count variable 'cnt_highincome' was perturbed.

# return weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))
#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 275178 4580 275178.0000
#>  2:  Total age_group1     total 1969 117866 1969 117866.0000
#>  3:  Total       ag1a     total 1969 117866 1969 117866.0000
#>  4:  Total age_group2     total 1143  67628 1142  67568.8329
#>  5:  Total       ag2a     total 1143  67628 1142  67568.8329
#>  6:  Total age_group3     total  864  52445  863  52384.2998
#>  7:  Total age_group4     total  423  25806  423  25806.0000
#>  8:  Total age_group5     total  168  10734  168  10734.0000
#>  9:  Total age_group6     total   13    699   14    752.7692
#> 10:   male      Total     total 2296 137975 2297 138035.0936
#> 11:      m      Total     total 2296 137975 2297 138035.0936
#> 12:   male age_group1     total 1015  60806 1016  60865.9074
#> 13:      m age_group1     total 1015  60806 1016  60865.9074
#> 14:   male       ag1a     total 1015  60806 1016  60865.9074
#> 15:      m       ag1a     total 1015  60806 1016  60865.9074
#> 16:   male age_group2     total  571  33874  570  33814.6760
#> 17:      m age_group2     total  571  33874  570  33814.6760
#> 18:   male       ag2a     total  571  33874  570  33814.6760
#> 19:      m       ag2a     total  571  33874  570  33814.6760
#> 20:   male age_group3     total  424  26023  425  26084.3750
#> 21:      m age_group3     total  424  26023  425  26084.3750
#> 22:   male age_group4     total  195  11644  195  11644.0000
#> 23:      m age_group4     total  195  11644  195  11644.0000
#> 24:   male age_group5     total   84   5303   84   5303.0000
#> 25:      m age_group5     total   84   5303   84   5303.0000
#> 26:   male age_group6     total    7    325    6    278.5714
#> 27:      m age_group6     total    7    325    6    278.5714
#> 28: female      Total     total 2284 137203 2285 137263.0714
#> 29:      f      Total     total 2284 137203 2285 137263.0714
#> 30: female age_group1     total  954  57060  955  57119.8113
#> 31:      f age_group1     total  954  57060  955  57119.8113
#> 32: female       ag1a     total  954  57060  955  57119.8113
#> 33:      f       ag1a     total  954  57060  955  57119.8113
#> 34: female age_group2     total  572  33754  573  33813.0105
#> 35:      f age_group2     total  572  33754  573  33813.0105
#> 36: female       ag2a     total  572  33754  573  33813.0105
#> 37:      f       ag2a     total  572  33754  573  33813.0105
#> 38: female age_group3     total  440  26422  439  26361.9500
#> 39:      f age_group3     total  440  26422  439  26361.9500
#> 40: female age_group4     total  228  14162  227  14099.8860
#> 41:      f age_group4     total  228  14162  227  14099.8860
#> 42: female age_group5     total   84   5431   82   5301.6905
#> 43:      f age_group5     total   84   5431   82   5301.6905
#> 44: female age_group6     total    6    374    6    374.0000
#> 45:      f age_group6     total    6    374    6    374.0000
#> 46:  Total      Total cnt_males 2296 137975 2298 138095.1873
#> 47:  Total age_group1 cnt_males 1015  60806 1016  60865.9074
#> 48:  Total       ag1a cnt_males 1015  60806 1016  60865.9074
#> 49:  Total age_group2 cnt_males  571  33874  570  33814.6760
#> 50:  Total       ag2a cnt_males  571  33874  570  33814.6760
#> 51:  Total age_group3 cnt_males  424  26023  426  26145.7500
#> 52:  Total age_group4 cnt_males  195  11644  194  11584.2872
#> 53:  Total age_group5 cnt_males   84   5303   84   5303.0000
#> 54:  Total age_group6 cnt_males    7    325    6    278.5714
#> 55:   male      Total cnt_males 2296 137975 2298 138095.1873
#> 56:      m      Total cnt_males 2296 137975 2298 138095.1873
#> 57:   male age_group1 cnt_males 1015  60806 1016  60865.9074
#> 58:      m age_group1 cnt_males 1015  60806 1016  60865.9074
#> 59:   male       ag1a cnt_males 1015  60806 1016  60865.9074
#> 60:      m       ag1a cnt_males 1015  60806 1016  60865.9074
#> 61:   male age_group2 cnt_males  571  33874  570  33814.6760
#> 62:      m age_group2 cnt_males  571  33874  570  33814.6760
#> 63:   male       ag2a cnt_males  571  33874  570  33814.6760
#> 64:      m       ag2a cnt_males  571  33874  570  33814.6760
#> 65:   male age_group3 cnt_males  424  26023  426  26145.7500
#> 66:      m age_group3 cnt_males  424  26023  426  26145.7500
#> 67:   male age_group4 cnt_males  195  11644  194  11584.2872
#> 68:      m age_group4 cnt_males  195  11644  194  11584.2872
#> 69:   male age_group5 cnt_males   84   5303   84   5303.0000
#> 70:      m age_group5 cnt_males   84   5303   84   5303.0000
#> 71:   male age_group6 cnt_males    7    325    6    278.5714
#> 72:      m age_group6 cnt_males    7    325    6    278.5714
#> 73: female      Total cnt_males    0      0    0      0.0000
#> 74:      f      Total cnt_males    0      0    0      0.0000
#> 75: female age_group1 cnt_males    0      0    0      0.0000
#> 76:      f age_group1 cnt_males    0      0    0      0.0000
#> 77: female       ag1a cnt_males    0      0    0      0.0000
#> 78:      f       ag1a cnt_males    0      0    0      0.0000
#> 79: female age_group2 cnt_males    0      0    0      0.0000
#> 80:      f age_group2 cnt_males    0      0    0      0.0000
#> 81: female       ag2a cnt_males    0      0    0      0.0000
#> 82:      f       ag2a cnt_males    0      0    0      0.0000
#> 83: female age_group3 cnt_males    0      0    0      0.0000
#> 84:      f age_group3 cnt_males    0      0    0      0.0000
#> 85: female age_group4 cnt_males    0      0    0      0.0000
#> 86:      f age_group4 cnt_males    0      0    0      0.0000
#> 87: female age_group5 cnt_males    0      0    0      0.0000
#> 88:      f age_group5 cnt_males    0      0    0      0.0000
#> 89: female age_group6 cnt_males    0      0    0      0.0000
#> 90:      f age_group6 cnt_males    0      0    0      0.0000
#>        sex        age     vname  uwc     wc puwc         pwc

# numerical variables (positive variables using flex-function)
# we also write the config to a yaml file
f_yaml <- tempfile(fileext = ".yaml")

# create a ptable using functionality from the ptable-pkg
# a single ptable for all cells
ptab1 <- ptable::pt_ex_nums(parity = TRUE, separation = FALSE)

# a single ptab for all cells except for very small ones
ptab2 <- ptable::pt_ex_nums(parity = TRUE, separation = TRUE)

# different ptables for cells with even/odd number of contributors
# and very small cells
ptab3 <- ptable::pt_ex_nums(parity = FALSE, separation = TRUE)

p_nums1 <- ck_params_nums(
  ptab = ptab1,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.30, 0.03),
    epsilon = c(1, 0.5, 0.2),
    q = 3),
  mu_c = 2,
  same_key = FALSE,
  use_zero_rkeys = FALSE,
  path = f_yaml)
#> yaml configuration '/tmp/RtmpojDFXK/file19cd25150452.yaml' successfully written.

# we read the parameters from the yaml-file
p_nums1 <- ck_read_yaml(path = f_yaml)

# for variables with positive and negative values
p_nums2 <- ck_params_nums(
  ptab = ptab2,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.15, 0.02),
    epsilon = c(1, 0.4, 0.15),
    q = 3),
  mu_c = 2,
  same_key = FALSE)

# simple perturbation parameters (not using the flex-function approach)
p_nums3 <- ck_params_nums(
  ptab = ptab3,
  type = "mean",
  mult_params = ck_simpleparams(p = 0.25),
  mu_c = 2,
  same_key = FALSE)

# use `p_nums1` for all variables
tab$params_nums_set(p_nums1, c("savings", "income", "expend"))
#> --> setting perturbation parameters for variable 'savings'
#> --> setting perturbation parameters for variable 'income'
#> --> setting perturbation parameters for variable 'expend'

# use different parameters for variable `mixed`
tab$params_nums_set(p_nums2, v = "mixed")
#> --> setting perturbation parameters for variable 'mixed'

# identify sensitive cells to which extra protection (`mu_c`) is added.
tab$supp_p(v = "income", p = 85)
#> computing contributing indices | rawdata <--> table; this might take a while
#> p%-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_pq(v = "income", p = 85, q = 90)
#> computing contributing indices | rawdata <--> table; this might take a while
#> pq-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_nk(v = "income", n = 2, k = 90)
#> computing contributing indices | rawdata <--> table; this might take a while
#> nk-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_freq(v = "income", n = 14, weighted = FALSE)
#> freq-rule: 5 new sensitive cells (incl. duplicates) found (total: 5)
tab$supp_val(v = "income", n = 10000, weighted = TRUE)
#> val-rule: 0 new sensitive cells (incl. duplicates) found (total: 5)
tab$supp_cells(
  v = "income",
  inp = data.frame(
    sex = c("female", "female"),
    "age" = c("age_group1", "age_group3")
  )
)
#> cell-rule: 2 new sensitive cells (incl. duplicates) found (total: 7)

# perturb variables
tab$perturb(v = c("income", "savings"))
#> Numeric variable 'income' was perturbed.
#> Numeric variable 'savings' was perturbed.

# extract results
tab$numtab("income", mean_before_sum = TRUE)
#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1382153102 1382085438
#>  2:  Total age_group1 income  9810547  589272531  589323653
#>  3:  Total       ag1a income  9810547  589272531  589323653
#>  4:  Total age_group2 income  5692119  339090898  339038466
#>  5:  Total       ag2a income  5692119  339090898  339038466
#>  6:  Total age_group3 income  4406946  267431771  267290918
#>  7:  Total age_group4 income  2133543  128281393  128324986
#>  8:  Total age_group5 income   848151   54412310   54436642
#>  9:  Total age_group6 income    61672    3664199    3785875
#> 10:   male      Total income 11262049  676669465  676561158
#> 11:      m      Total income 11262049  676669465  676561158
#> 12:   male age_group1 income  4877164  292288160  292295063
#> 13:      m age_group1 income  4877164  292288160  292295063
#> 14:   male       ag1a income  4877164  292288160  292295063
#> 15:      m       ag1a income  4877164  292288160  292295063
#> 16:   male age_group2 income  2811379  168022325  167969793
#> 17:      m age_group2 income  2811379  168022325  167969793
#> 18:   male       ag2a income  2811379  168022325  167969793
#> 19:      m       ag2a income  2811379  168022325  167969793
#> 20:   male age_group3 income  2168169  132296584  132358588
#> 21:      m age_group3 income  2168169  132296584  132358588
#> 22:   male age_group4 income   978510   57911967   57922048
#> 23:      m age_group4 income   978510   57911967   57922048
#> 24:   male age_group5 income   393134   24535061   24554844
#> 25:      m age_group5 income   393134   24535061   24554844
#> 26:   male age_group6 income    33693    1615368    1668982
#> 27:      m age_group6 income    33693    1615368    1668982
#> 28: female      Total income 11690929  705483637  705392168
#> 29:      f      Total income 11690929  705483637  705392168
#> 30: female age_group1 income  4933383  296984371  296960555
#> 31:      f age_group1 income  4933383  296984371  296960555
#> 32: female       ag1a income  4933383  296984371  296960555
#> 33:      f       ag1a income  4933383  296984371  296960555
#> 34: female age_group2 income  2880740  171068573  171063194
#> 35:      f age_group2 income  2880740  171068573  171063194
#> 36: female       ag2a income  2880740  171068573  171063194
#> 37:      f       ag2a income  2880740  171068573  171063194
#> 38: female age_group3 income  2238777  135135187  135121694
#> 39:      f age_group3 income  2238777  135135187  135121694
#> 40: female age_group4 income  1155033   70369426   70345639
#> 41:      f age_group4 income  1155033   70369426   70345639
#> 42: female age_group5 income   455017   29877249   29883865
#> 43:      f age_group5 income   455017   29877249   29883865
#> 44: female age_group6 income    27979    2048831    2113115
#> 45:      f age_group6 income    27979    2048831    2113115
#>        sex        age  vname      uws         ws        pws
tab$numtab("income", mean_before_sum = FALSE)
#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1382153102 1382119270
#>  2:  Total age_group1 income  9810547  589272531  589298092
#>  3:  Total       ag1a income  9810547  589272531  589298092
#>  4:  Total age_group2 income  5692119  339090898  339064681
#>  5:  Total       ag2a income  5692119  339090898  339064681
#>  6:  Total age_group3 income  4406946  267431771  267361335
#>  7:  Total age_group4 income  2133543  128281393  128303188
#>  8:  Total age_group5 income   848151   54412310   54424475
#>  9:  Total age_group6 income    61672    3664199    3724540
#> 10:   male      Total income 11262049  676669465  676615310
#> 11:      m      Total income 11262049  676669465  676615310
#> 12:   male age_group1 income  4877164  292288160  292291612
#> 13:      m age_group1 income  4877164  292288160  292291612
#> 14:   male       ag1a income  4877164  292288160  292291612
#> 15:      m       ag1a income  4877164  292288160  292291612
#> 16:   male age_group2 income  2811379  168022325  167996057
#> 17:      m age_group2 income  2811379  168022325  167996057
#> 18:   male       ag2a income  2811379  168022325  167996057
#> 19:      m       ag2a income  2811379  168022325  167996057
#> 20:   male age_group3 income  2168169  132296584  132327582
#> 21:      m age_group3 income  2168169  132296584  132327582
#> 22:   male age_group4 income   978510   57911967   57917008
#> 23:      m age_group4 income   978510   57911967   57917008
#> 24:   male age_group5 income   393134   24535061   24544951
#> 25:      m age_group5 income   393134   24535061   24544951
#> 26:   male age_group6 income    33693    1615368    1641956
#> 27:      m age_group6 income    33693    1615368    1641956
#> 28: female      Total income 11690929  705483637  705437901
#> 29:      f      Total income 11690929  705483637  705437901
#> 30: female age_group1 income  4933383  296984371  296972463
#> 31:      f age_group1 income  4933383  296984371  296972463
#> 32: female       ag1a income  4933383  296984371  296972463
#> 33:      f       ag1a income  4933383  296984371  296972463
#> 34: female age_group2 income  2880740  171068573  171065883
#> 35:      f age_group2 income  2880740  171068573  171065883
#> 36: female       ag2a income  2880740  171068573  171065883
#> 37:      f       ag2a income  2880740  171068573  171065883
#> 38: female age_group3 income  2238777  135135187  135128440
#> 39:      f age_group3 income  2238777  135135187  135128440
#> 40: female age_group4 income  1155033   70369426   70357531
#> 41:      f age_group4 income  1155033   70369426   70357531
#> 42: female age_group5 income   455017   29877249   29880557
#> 43:      f age_group5 income   455017   29877249   29880557
#> 44: female age_group6 income    27979    2048831    2080725
#> 45:      f age_group6 income    27979    2048831    2080725
#>        sex        age  vname      uws         ws        pws
tab$numtab("savings")
#>        sex        age   vname     uws        ws         pws
#>  1:  Total      Total savings 2273532 136408428 136411647.8
#>  2:  Total age_group1 savings  982386  58679394  58679247.7
#>  3:  Total       ag1a savings  982386  58679394  58679247.7
#>  4:  Total age_group2 savings  552336  32715268  32713233.3
#>  5:  Total       ag2a savings  552336  32715268  32713233.3
#>  6:  Total age_group3 savings  437101  26584077  26589845.2
#>  7:  Total age_group4 savings  214661  13037920  13037445.9
#>  8:  Total age_group5 savings   80451   4997787   4999178.2
#>  9:  Total age_group6 savings    6597    393982    395262.0
#> 10:   male      Total savings 1159816  69758773  69763832.9
#> 11:      m      Total savings 1159816  69758773  69763832.9
#> 12:   male age_group1 savings  517660  31073904  31074894.7
#> 13:      m age_group1 savings  517660  31073904  31074894.7
#> 14:   male       ag1a savings  517660  31073904  31074894.7
#> 15:      m       ag1a savings  517660  31073904  31074894.7
#> 16:   male age_group2 savings  280923  16744540  16743379.8
#> 17:      m age_group2 savings  280923  16744540  16743379.8
#> 18:   male       ag2a savings  280923  16744540  16743379.8
#> 19:      m       ag2a savings  280923  16744540  16743379.8
#> 20:   male age_group3 savings  214970  13148152  13145955.0
#> 21:      m age_group3 savings  214970  13148152  13145955.0
#> 22:   male age_group4 savings   99420   5982325   5983946.2
#> 23:      m age_group4 savings   99420   5982325   5983946.2
#> 24:   male age_group5 savings   43233   2624407   2627548.1
#> 25:      m age_group5 savings   43233   2624407   2627548.1
#> 26:   male age_group6 savings    3610    185445    187274.3
#> 27:      m age_group6 savings    3610    185445    187274.3
#> 28: female      Total savings 1113716  66649655  66651808.7
#> 29:      f      Total savings 1113716  66649655  66651808.7
#> 30: female age_group1 savings  464726  27605490  27607926.3
#> 31:      f age_group1 savings  464726  27605490  27607926.3
#> 32: female       ag1a savings  464726  27605490  27607926.3
#> 33:      f       ag1a savings  464726  27605490  27607926.3
#> 34: female age_group2 savings  271413  15970728  15963862.2
#> 35:      f age_group2 savings  271413  15970728  15963862.2
#> 36: female       ag2a savings  271413  15970728  15963862.2
#> 37:      f       ag2a savings  271413  15970728  15963862.2
#> 38: female age_group3 savings  222131  13435925  13433994.5
#> 39:      f age_group3 savings  222131  13435925  13433994.5
#> 40: female age_group4 savings  115241   7055595   7054363.0
#> 41:      f age_group4 savings  115241   7055595   7054363.0
#> 42: female age_group5 savings   37218   2373380   2374492.6
#> 43:      f age_group5 savings   37218   2373380   2374492.6
#> 44: female age_group6 savings    2987    208537    210009.2
#> 45:      f age_group6 savings    2987    208537    210009.2
#>        sex        age   vname     uws        ws         pws

# results can be resetted, too
tab$reset_cntvars(v = "cnt_males")

# we can then set other parameters and perturb again
tab$params_cnts_set(val = p_cnts1, v = "cnt_males")
#> --> setting perturbation parameters for variable 'cnt_males'

tab$perturb(v = "cnt_males")
#> Count variable 'cnt_males' was perturbed.

# write results to a .csv file
tab$freqtab(
  v = c("total", "cnt_males"),
  path = file.path(tempdir(), "outtab.csv")
)
#> File '/tmp/RtmpojDFXK/outtab.csv' successfully written to disk.
#> NULL

# show results containing weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))
#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 275178 4580 275178.0000
#>  2:  Total age_group1     total 1969 117866 1969 117866.0000
#>  3:  Total       ag1a     total 1969 117866 1969 117866.0000
#>  4:  Total age_group2     total 1143  67628 1142  67568.8329
#>  5:  Total       ag2a     total 1143  67628 1142  67568.8329
#>  6:  Total age_group3     total  864  52445  863  52384.2998
#>  7:  Total age_group4     total  423  25806  423  25806.0000
#>  8:  Total age_group5     total  168  10734  168  10734.0000
#>  9:  Total age_group6     total   13    699   14    752.7692
#> 10:   male      Total     total 2296 137975 2297 138035.0936
#> 11:      m      Total     total 2296 137975 2297 138035.0936
#> 12:   male age_group1     total 1015  60806 1016  60865.9074
#> 13:      m age_group1     total 1015  60806 1016  60865.9074
#> 14:   male       ag1a     total 1015  60806 1016  60865.9074
#> 15:      m       ag1a     total 1015  60806 1016  60865.9074
#> 16:   male age_group2     total  571  33874  570  33814.6760
#> 17:      m age_group2     total  571  33874  570  33814.6760
#> 18:   male       ag2a     total  571  33874  570  33814.6760
#> 19:      m       ag2a     total  571  33874  570  33814.6760
#> 20:   male age_group3     total  424  26023  425  26084.3750
#> 21:      m age_group3     total  424  26023  425  26084.3750
#> 22:   male age_group4     total  195  11644  195  11644.0000
#> 23:      m age_group4     total  195  11644  195  11644.0000
#> 24:   male age_group5     total   84   5303   84   5303.0000
#> 25:      m age_group5     total   84   5303   84   5303.0000
#> 26:   male age_group6     total    7    325    6    278.5714
#> 27:      m age_group6     total    7    325    6    278.5714
#> 28: female      Total     total 2284 137203 2285 137263.0714
#> 29:      f      Total     total 2284 137203 2285 137263.0714
#> 30: female age_group1     total  954  57060  955  57119.8113
#> 31:      f age_group1     total  954  57060  955  57119.8113
#> 32: female       ag1a     total  954  57060  955  57119.8113
#> 33:      f       ag1a     total  954  57060  955  57119.8113
#> 34: female age_group2     total  572  33754  573  33813.0105
#> 35:      f age_group2     total  572  33754  573  33813.0105
#> 36: female       ag2a     total  572  33754  573  33813.0105
#> 37:      f       ag2a     total  572  33754  573  33813.0105
#> 38: female age_group3     total  440  26422  439  26361.9500
#> 39:      f age_group3     total  440  26422  439  26361.9500
#> 40: female age_group4     total  228  14162  227  14099.8860
#> 41:      f age_group4     total  228  14162  227  14099.8860
#> 42: female age_group5     total   84   5431   82   5301.6905
#> 43:      f age_group5     total   84   5431   82   5301.6905
#> 44: female age_group6     total    6    374    6    374.0000
#> 45:      f age_group6     total    6    374    6    374.0000
#> 46:  Total      Total cnt_males 2296 137975 2297 138035.0936
#> 47:  Total age_group1 cnt_males 1015  60806 1016  60865.9074
#> 48:  Total       ag1a cnt_males 1015  60806 1016  60865.9074
#> 49:  Total age_group2 cnt_males  571  33874  570  33814.6760
#> 50:  Total       ag2a cnt_males  571  33874  570  33814.6760
#> 51:  Total age_group3 cnt_males  424  26023  425  26084.3750
#> 52:  Total age_group4 cnt_males  195  11644  195  11644.0000
#> 53:  Total age_group5 cnt_males   84   5303   84   5303.0000
#> 54:  Total age_group6 cnt_males    7    325    6    278.5714
#> 55:   male      Total cnt_males 2296 137975 2297 138035.0936
#> 56:      m      Total cnt_males 2296 137975 2297 138035.0936
#> 57:   male age_group1 cnt_males 1015  60806 1016  60865.9074
#> 58:      m age_group1 cnt_males 1015  60806 1016  60865.9074
#> 59:   male       ag1a cnt_males 1015  60806 1016  60865.9074
#> 60:      m       ag1a cnt_males 1015  60806 1016  60865.9074
#> 61:   male age_group2 cnt_males  571  33874  570  33814.6760
#> 62:      m age_group2 cnt_males  571  33874  570  33814.6760
#> 63:   male       ag2a cnt_males  571  33874  570  33814.6760
#> 64:      m       ag2a cnt_males  571  33874  570  33814.6760
#> 65:   male age_group3 cnt_males  424  26023  425  26084.3750
#> 66:      m age_group3 cnt_males  424  26023  425  26084.3750
#> 67:   male age_group4 cnt_males  195  11644  195  11644.0000
#> 68:      m age_group4 cnt_males  195  11644  195  11644.0000
#> 69:   male age_group5 cnt_males   84   5303   84   5303.0000
#> 70:      m age_group5 cnt_males   84   5303   84   5303.0000
#> 71:   male age_group6 cnt_males    7    325    6    278.5714
#> 72:      m age_group6 cnt_males    7    325    6    278.5714
#> 73: female      Total cnt_males    0      0    0      0.0000
#> 74:      f      Total cnt_males    0      0    0      0.0000
#> 75: female age_group1 cnt_males    0      0    0      0.0000
#> 76:      f age_group1 cnt_males    0      0    0      0.0000
#> 77: female       ag1a cnt_males    0      0    0      0.0000
#> 78:      f       ag1a cnt_males    0      0    0      0.0000
#> 79: female age_group2 cnt_males    0      0    0      0.0000
#> 80:      f age_group2 cnt_males    0      0    0      0.0000
#> 81: female       ag2a cnt_males    0      0    0      0.0000
#> 82:      f       ag2a cnt_males    0      0    0      0.0000
#> 83: female age_group3 cnt_males    0      0    0      0.0000
#> 84:      f age_group3 cnt_males    0      0    0      0.0000
#> 85: female age_group4 cnt_males    0      0    0      0.0000
#> 86:      f age_group4 cnt_males    0      0    0      0.0000
#> 87: female age_group5 cnt_males    0      0    0      0.0000
#> 88:      f age_group5 cnt_males    0      0    0      0.0000
#> 89: female age_group6 cnt_males    0      0    0      0.0000
#> 90:      f age_group6 cnt_males    0      0    0      0.0000
#>        sex        age     vname  uwc     wc puwc         pwc

# utility measures for a count variable
tab$measures_cnts(v = "total", exclude_zeros = TRUE)
#> $overview
#>    noise cnt        pct
#> 1:    -1  19 0.42222222
#> 2:     0  11 0.24444444
#> 3:     1  13 0.28888889
#> 4:     2   2 0.04444444
#> 
#> $measures
#>       what  d1    d2    d3
#>  1:    Min 0.0 0.000 0.000
#>  2:    Q10 0.0 0.000 0.000
#>  3:    Q20 0.0 0.000 0.000
#>  4:    Q30 1.0 0.000 0.010
#>  5:    Q40 1.0 0.001 0.016
#>  6:   Mean 0.8 0.010 0.029
#>  7: Median 1.0 0.001 0.016
#>  8:    Q60 1.0 0.002 0.021
#>  9:    Q70 1.0 0.002 0.021
#> 10:    Q80 1.0 0.002 0.024
#> 11:    Q90 1.0 0.016 0.079
#> 12:    Q95 1.0 0.066 0.131
#> 13:    Q99 2.0 0.143 0.196
#> 14:    Max 2.0 0.143 0.196
#> 
#> $cumdistr_d1
#>    cat cnt       pct
#> 1:   0  11 0.2444444
#> 2:   1  43 0.9555556
#> 3:   2  45 1.0000000
#> 
#> $cumdistr_d2
#>            cat cnt       pct
#> 1:    [0,0.02]  40 0.8888889
#> 2: (0.02,0.05]  42 0.9333333
#> 3:  (0.05,0.1]  43 0.9555556
#> 4:   (0.1,0.2]  45 1.0000000
#> 5:   (0.2,0.3]  45 1.0000000
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $cumdistr_d3
#>            cat cnt       pct
#> 1:    [0,0.02]  26 0.5777778
#> 2: (0.02,0.05]  40 0.8888889
#> 3:  (0.05,0.1]  40 0.8888889
#> 4:   (0.1,0.2]  45 1.0000000
#> 5:   (0.2,0.3]  45 1.0000000
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $false_zero
#> [1] 0
#> 
#> $false_nonzero
#> [1] 0
#> 
#> $exclude_zeros
#> [1] TRUE
#> 

# modifications for perturbed count variables
tab$mod_cnts()
#>         sex        age row_nr pert      ckey  countvar
#>   1:  Total      Total     15    0 0.6154048     total
#>   2:  Total age_group1     15    0 0.4825659     total
#>   3:  Total       ag1a     15    0 0.4825659     total
#>   4:  Total age_group2     14   -1 0.1506082     total
#>   5:  Total       ag2a     14   -1 0.1506082     total
#>  ---                                                  
#> 131:      f age_group4     -1    0 0.0000000 cnt_males
#> 132: female age_group5     -1    0 0.0000000 cnt_males
#> 133:      f age_group5     -1    0 0.0000000 cnt_males
#> 134: female age_group6     -1    0 0.0000000 cnt_males
#> 135:      f age_group6     -1    0 0.0000000 cnt_males

# display a summary about utility measures
tab$summary()
#> ┌──────────────────────────────────────────────┐
#> │Utility measures for perturbed count variables│
#> └──────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#>          countvar Min Q10  Q20  Q30 Q40  Mean Median Q60 Q70 Q80 Q90 Q95 Q99
#> 1:          total  -2  -1 -1.0 -1.0   0 0.044      0   1   1   1   1   1   1
#> 2: cnt_highincome  -3  -3 -2.0 -0.8   0 0.156      0   1   1   2   2   2   3
#> 3:      cnt_males  -1  -1 -0.2  0.0   0 0.067      0   0   0   1   1   1   1
#>    Max
#> 1:   1
#> 2:   3
#> 3:   1
#> 
#> ── Distance-based measures ─────────────────────────────────────────────────────
#> ✔ Variable: 'total'
#> 
#>       what  d1    d2    d3
#>  1:    Min 0.0 0.000 0.000
#>  2:    Q10 0.0 0.000 0.000
#>  3:    Q20 0.0 0.000 0.000
#>  4:    Q30 1.0 0.000 0.010
#>  5:    Q40 1.0 0.001 0.016
#>  6:   Mean 0.8 0.010 0.029
#>  7: Median 1.0 0.001 0.016
#>  8:    Q60 1.0 0.002 0.021
#>  9:    Q70 1.0 0.002 0.021
#> 10:    Q80 1.0 0.002 0.024
#> 11:    Q90 1.0 0.016 0.079
#> 12:    Q95 1.0 0.066 0.131
#> 13:    Q99 2.0 0.143 0.196
#> 14:    Max 2.0 0.143 0.196
#> 
#> ✔ Variable: 'cnt_males'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.200 0.000 0.002
#>  4:    Q30 1.000 0.000 0.010
#>  5:    Q40 1.000 0.001 0.016
#>  6:   Mean 0.778 0.017 0.034
#>  7: Median 1.000 0.001 0.016
#>  8:    Q60 1.000 0.002 0.021
#>  9:    Q70 1.000 0.002 0.021
#> 10:    Q80 1.000 0.002 0.024
#> 11:    Q90 1.000 0.059 0.093
#> 12:    Q95 1.000 0.143 0.196
#> 13:    Q99 1.000 0.143 0.196
#> 14:    Max 1.000 0.143 0.196
#> 
#> ✔ Variable: 'cnt_highincome'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.900 0.004 0.030
#>  3:    Q20 1.000 0.009 0.053
#>  4:    Q30 1.000 0.011 0.064
#>  5:    Q40 1.000 0.020 0.099
#>  6:   Mean 1.675 0.050 0.126
#>  7: Median 2.000 0.030 0.122
#>  8:    Q60 2.000 0.039 0.131
#>  9:    Q70 2.000 0.048 0.146
#> 10:    Q80 2.200 0.067 0.187
#> 11:    Q90 3.000 0.143 0.239
#> 12:    Q95 3.000 0.150 0.266
#> 13:    Q99 3.000 0.286 0.410
#> 14:    Max 3.000 0.286 0.410
#> 
#> ┌──────────────────────────────────────────────────┐
#> │Utility measures for perturbed numerical variables│
#> └──────────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#>      vname        Min        Q10        Q20        Q30       Q40      Mean
#> 1:  expend        Inf         NA         NA         NA        NA       NaN
#> 2:  income -70435.798 -40974.431 -26268.055 -11908.099 -8806.026 -4057.032
#> 3: savings  -6865.842  -2197.046  -1930.534  -1160.201  -146.301   233.061
#> 4:   mixed        Inf         NA         NA         NA        NA       NaN
#>       Median      Q60      Q70       Q80      Q90       Q95       Q99       Max
#> 1:        NA       NA       NA        NA       NA        NA        NA      -Inf
#> 2: -2689.767 3451.703 5040.519 22548.039 29234.15 31714.760 47824.239 60340.942
#> 3:   990.692 1423.589 1787.648  2436.291  3141.12  4691.915  5456.575  5768.216
#> 4:        NA       NA       NA        NA       NA        NA        NA      -Inf
# }