Set parameters required to perturb numeric variables using a flex function

ck_flexparams() allows to define a flex function that is used to lookup perturbation magnitudes (percentages) used when perturbing continuous variables.

ck_flexparams(fp, p = c(0.25, 0.05), epsilon = 1, q = 3)

Arguments

fp

(numeric scalar); at which point should the noise coefficient function reaches its desired maximum (defined by the first element of p)

p

a numeric vector of length 2 where both elements specify a percentage. The first value refers to the desired maximum perturbation percentage for small cells (depending on fp) while the second element refers to the desired maximum perturbation percentage for large cells. Both values must be between 0 and 1 and need to be in descending order.

epsilon

a numeric vector in descending order with all values >= 0 and <= 1 with the first element forced to equal 1. The length of this vector must correspond with the number top_k specified in ck_params_nums() when creating parameters for type == "top_contr" which is checked at runtime. This setting allows to use different flex-functions for the largest top_k contributors.

q

(numeric scalar); Parameter of the function; q needs to be >= 1

Value

an object suitable as input for ck_params_nums().

Details

details about the flex function can be found in Deliverable D4.2, Part I in SGA "Open Source tools for perturbative confidentiality methods"

See also

Examples

# \donttest{
x <- ck_create_testdata()

# create some 0/1 variables that should be perturbed later
x[, cnt_females := ifelse(sex == "male", 0, 1)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              28          25.00000           0
#>    2:   2530      28       1              87          25.00000           1
#>    3:   6920     550       1              21          25.00000           0
#>    4:   7960     870       1              37          25.00000           0
#>    5:   9030      20       2              33          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              82          16.66667           1
#> 4577:   1420     987    1000              63          16.66667           0
#> 4578:   8900     684    1000              41          16.66667           0
#> 4579:   3880     294    1000              43          16.66667           1
#> 4580:   4830     911    1000              93          16.66667           0
x[, cnt_males := ifelse(sex == "male", 1, 0)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              28          25.00000           0
#>    2:   2530      28       1              87          25.00000           1
#>    3:   6920     550       1              21          25.00000           0
#>    4:   7960     870       1              37          25.00000           0
#>    5:   9030      20       2              33          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              82          16.66667           1
#> 4577:   1420     987    1000              63          16.66667           0
#> 4578:   8900     684    1000              41          16.66667           0
#> 4579:   3880     294    1000              43          16.66667           1
#> 4580:   4830     911    1000              93          16.66667           0
#>       cnt_males
#>    1:         1
#>    2:         0
#>    3:         1
#>    4:         1
#>    5:         1
#>   ---          
#> 4576:         0
#> 4577:         1
#> 4578:         1
#> 4579:         0
#> 4580:         1
x[, cnt_highincome := ifelse(income >= 9000, 1, 0)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              28          25.00000           0
#>    2:   2530      28       1              87          25.00000           1
#>    3:   6920     550       1              21          25.00000           0
#>    4:   7960     870       1              37          25.00000           0
#>    5:   9030      20       2              33          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              82          16.66667           1
#> 4577:   1420     987    1000              63          16.66667           0
#> 4578:   8900     684    1000              41          16.66667           0
#> 4579:   3880     294    1000              43          16.66667           1
#> 4580:   4830     911    1000              93          16.66667           0
#>       cnt_males cnt_highincome
#>    1:         1              0
#>    2:         0              0
#>    3:         1              0
#>    4:         1              0
#>    5:         1              1
#>   ---                         
#> 4576:         0              0
#> 4577:         1              0
#> 4578:         1              0
#> 4579:         0              0
#> 4580:         1              0
# a variable with positive and negative contributions
x[, mixed := sample(-10:10, nrow(x), replace = TRUE)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              28          25.00000           0
#>    2:   2530      28       1              87          25.00000           1
#>    3:   6920     550       1              21          25.00000           0
#>    4:   7960     870       1              37          25.00000           0
#>    5:   9030      20       2              33          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              82          16.66667           1
#> 4577:   1420     987    1000              63          16.66667           0
#> 4578:   8900     684    1000              41          16.66667           0
#> 4579:   3880     294    1000              43          16.66667           1
#> 4580:   4830     911    1000              93          16.66667           0
#>       cnt_males cnt_highincome mixed
#>    1:         1              0   -10
#>    2:         0              0    -8
#>    3:         1              0     0
#>    4:         1              0     9
#>    5:         1              1    -2
#>   ---                               
#> 4576:         0              0    -4
#> 4577:         1              0    -8
#> 4578:         1              0   -10
#> 4579:         0              0    -8
#> 4580:         1              0     1

# create record keys
x$rkey <- ck_generate_rkeys(dat = x)

# define required inputs

# hierarchy with some bogus codes
d_sex <- hier_create(root = "Total", nodes = c("male", "female"))
d_sex <- hier_add(d_sex, root = "female", "f")
d_sex <- hier_add(d_sex, root = "male", "m")

d_age <- hier_create(root = "Total", nodes = paste0("age_group", 1:6))
d_age <- hier_add(d_age, root = "age_group1", "ag1a")
d_age <- hier_add(d_age, root = "age_group2", "ag2a")

# define the cell key object
countvars <- c("cnt_females", "cnt_males", "cnt_highincome")
numvars <- c("expend", "income", "savings", "mixed")
tab <- ck_setup(
  x = x,
  rkey = "rkey",
  dims = list(sex = d_sex, age = d_age),
  w = "sampling_weight",
  countvars = countvars,
  numvars = numvars)

# show some information about this table instance
tab$print() # identical with print(tab)
#> ── Table Information ───────────────────────────────────────────────────────────
#> ✔ 45 cells in 2 dimensions ('sex', 'age')
#> ✔ weights: yes
#> ── Tabulated / Perturbed countvars ─────────────────────────────────────────────
#> ☐ 'total'
#> ☐ 'cnt_females'
#> ☐ 'cnt_males'
#> ☐ 'cnt_highincome'
#> ── Tabulated / Perturbed numvars ───────────────────────────────────────────────
#> ☐ 'expend'
#> ☐ 'income'
#> ☐ 'savings'
#> ☐ 'mixed'

# information about the hierarchies
tab$hierarchy_info()
#> $sex
#>      code level is_leaf parent
#> 1:  Total     1   FALSE  Total
#> 2:   male     2   FALSE  Total
#> 3:      m     3    TRUE   male
#> 4: female     2   FALSE  Total
#> 5:      f     3    TRUE female
#> 
#> $age
#>          code level is_leaf     parent
#> 1:      Total     1   FALSE      Total
#> 2: age_group1     2   FALSE      Total
#> 3:       ag1a     3    TRUE age_group1
#> 4: age_group2     2   FALSE      Total
#> 5:       ag2a     3    TRUE age_group2
#> 6: age_group3     2    TRUE      Total
#> 7: age_group4     2    TRUE      Total
#> 8: age_group5     2    TRUE      Total
#> 9: age_group6     2    TRUE      Total
#> 

# which variables have been defined?
tab$allvars()
#> $cntvars
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"
#> 
#> $numvars
#> [1] "expend"  "income"  "savings" "mixed"  
#> 

# count variables
tab$cntvars()
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"

# continuous variables
tab$numvars()
#> [1] "expend"  "income"  "savings" "mixed"  

# create perturbation parameters for "total" variable and
# write to yaml-file

# create a ptable using functionality from the ptable-pkg
f_yaml <- tempfile(fileext = ".yaml")
p_cnts1 <- ck_params_cnts(
  ptab = ptable::pt_ex_cnts(),
  path = f_yaml)
#> yaml configuration '/tmp/RtmpCjMH4Z/file17d3454bf843.yaml' successfully written.

# read parameters from yaml-file and set them for variable `"total"`
p_cnts1 <- ck_read_yaml(path = f_yaml)

tab$params_cnts_set(val = p_cnts1, v = "total")
#> --> setting perturbation parameters for variable 'total'

# create alternative perturbation parameters by specifying parameters
para2 <- ptable::create_cnt_ptable(
  D = 8, V = 3, js = 2, create = FALSE)

p_cnts2 <- ck_params_cnts(ptab = para2)

# use these ptable it for the remaining variables
tab$params_cnts_set(val = p_cnts2, v = countvars)
#> --> setting perturbation parameters for variable 'cnt_females'
#> --> setting perturbation parameters for variable 'cnt_males'
#> --> setting perturbation parameters for variable 'cnt_highincome'

# perturb a variable
tab$perturb(v = "total")
#> Count variable 'total' was perturbed.

# multiple variables can be perturbed as well
tab$perturb(v = c("cnt_males", "cnt_highincome"))
#> Count variable 'cnt_males' was perturbed.
#> Count variable 'cnt_highincome' was perturbed.

# return weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))
#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 274937 4580 274937.0000
#>  2:  Total age_group1     total 1969 117416 1969 117416.0000
#>  3:  Total       ag1a     total 1969 117416 1969 117416.0000
#>  4:  Total age_group2     total 1143  69237 1142  69176.4252
#>  5:  Total       ag2a     total 1143  69237 1142  69176.4252
#>  6:  Total age_group3     total  864  52246  865  52306.4699
#>  7:  Total age_group4     total  423  25007  422  24947.8818
#>  8:  Total age_group5     total  168  10226  167  10165.1310
#>  9:  Total age_group6     total   13    805   13    805.0000
#> 10:   male      Total     total 2296 139534 2296 139534.0000
#> 11:      m      Total     total 2296 139534 2296 139534.0000
#> 12:   male age_group1     total 1015  60753 1015  60753.0000
#> 13:      m age_group1     total 1015  60753 1015  60753.0000
#> 14:   male       ag1a     total 1015  60753 1015  60753.0000
#> 15:      m       ag1a     total 1015  60753 1015  60753.0000
#> 16:   male age_group2     total  571  34840  573  34962.0315
#> 17:      m age_group2     total  571  34840  573  34962.0315
#> 18:   male       ag2a     total  571  34840  573  34962.0315
#> 19:      m       ag2a     total  571  34840  573  34962.0315
#> 20:   male age_group3     total  424  26694  425  26756.9575
#> 21:      m age_group3     total  424  26694  425  26756.9575
#> 22:   male age_group4     total  195  11330  196  11388.1026
#> 23:      m age_group4     total  195  11330  196  11388.1026
#> 24:   male age_group5     total   84   5489   84   5489.0000
#> 25:      m age_group5     total   84   5489   84   5489.0000
#> 26:   male age_group6     total    7    428    6    366.8571
#> 27:      m age_group6     total    7    428    6    366.8571
#> 28: female      Total     total 2284 135403 2285 135462.2833
#> 29:      f      Total     total 2284 135403 2285 135462.2833
#> 30: female age_group1     total  954  56663  952  56544.2096
#> 31:      f age_group1     total  954  56663  952  56544.2096
#> 32: female       ag1a     total  954  56663  952  56544.2096
#> 33:      f       ag1a     total  954  56663  952  56544.2096
#> 34: female age_group2     total  572  34397  571  34336.8654
#> 35:      f age_group2     total  572  34397  571  34336.8654
#> 36: female       ag2a     total  572  34397  571  34336.8654
#> 37:      f       ag2a     total  572  34397  571  34336.8654
#> 38: female age_group3     total  440  25552  439  25493.9273
#> 39:      f age_group3     total  440  25552  439  25493.9273
#> 40: female age_group4     total  228  13677  228  13677.0000
#> 41:      f age_group4     total  228  13677  228  13677.0000
#> 42: female age_group5     total   84   4737   85   4793.3929
#> 43:      f age_group5     total   84   4737   85   4793.3929
#> 44: female age_group6     total    6    377    5    314.1667
#> 45:      f age_group6     total    6    377    5    314.1667
#> 46:  Total      Total cnt_males 2296 139534 2296 139534.0000
#> 47:  Total age_group1 cnt_males 1015  60753 1014  60693.1448
#> 48:  Total       ag1a cnt_males 1015  60753 1014  60693.1448
#> 49:  Total age_group2 cnt_males  571  34840  575  35084.0630
#> 50:  Total       ag2a cnt_males  571  34840  575  35084.0630
#> 51:  Total age_group3 cnt_males  424  26694  425  26756.9575
#> 52:  Total age_group4 cnt_males  195  11330  197  11446.2051
#> 53:  Total age_group5 cnt_males   84   5489   84   5489.0000
#> 54:  Total age_group6 cnt_males    7    428    5    305.7143
#> 55:   male      Total cnt_males 2296 139534 2296 139534.0000
#> 56:      m      Total cnt_males 2296 139534 2296 139534.0000
#> 57:   male age_group1 cnt_males 1015  60753 1014  60693.1448
#> 58:      m age_group1 cnt_males 1015  60753 1014  60693.1448
#> 59:   male       ag1a cnt_males 1015  60753 1014  60693.1448
#> 60:      m       ag1a cnt_males 1015  60753 1014  60693.1448
#> 61:   male age_group2 cnt_males  571  34840  575  35084.0630
#> 62:      m age_group2 cnt_males  571  34840  575  35084.0630
#> 63:   male       ag2a cnt_males  571  34840  575  35084.0630
#> 64:      m       ag2a cnt_males  571  34840  575  35084.0630
#> 65:   male age_group3 cnt_males  424  26694  425  26756.9575
#> 66:      m age_group3 cnt_males  424  26694  425  26756.9575
#> 67:   male age_group4 cnt_males  195  11330  197  11446.2051
#> 68:      m age_group4 cnt_males  195  11330  197  11446.2051
#> 69:   male age_group5 cnt_males   84   5489   84   5489.0000
#> 70:      m age_group5 cnt_males   84   5489   84   5489.0000
#> 71:   male age_group6 cnt_males    7    428    5    305.7143
#> 72:      m age_group6 cnt_males    7    428    5    305.7143
#> 73: female      Total cnt_males    0      0    0      0.0000
#> 74:      f      Total cnt_males    0      0    0      0.0000
#> 75: female age_group1 cnt_males    0      0    0      0.0000
#> 76:      f age_group1 cnt_males    0      0    0      0.0000
#> 77: female       ag1a cnt_males    0      0    0      0.0000
#> 78:      f       ag1a cnt_males    0      0    0      0.0000
#> 79: female age_group2 cnt_males    0      0    0      0.0000
#> 80:      f age_group2 cnt_males    0      0    0      0.0000
#> 81: female       ag2a cnt_males    0      0    0      0.0000
#> 82:      f       ag2a cnt_males    0      0    0      0.0000
#> 83: female age_group3 cnt_males    0      0    0      0.0000
#> 84:      f age_group3 cnt_males    0      0    0      0.0000
#> 85: female age_group4 cnt_males    0      0    0      0.0000
#> 86:      f age_group4 cnt_males    0      0    0      0.0000
#> 87: female age_group5 cnt_males    0      0    0      0.0000
#> 88:      f age_group5 cnt_males    0      0    0      0.0000
#> 89: female age_group6 cnt_males    0      0    0      0.0000
#> 90:      f age_group6 cnt_males    0      0    0      0.0000
#>        sex        age     vname  uwc     wc puwc         pwc

# numerical variables (positive variables using flex-function)
# we also write the config to a yaml file
f_yaml <- tempfile(fileext = ".yaml")

# create a ptable using functionality from the ptable-pkg
# a single ptable for all cells
ptab1 <- ptable::pt_ex_nums(parity = TRUE, separation = FALSE)

# a single ptab for all cells except for very small ones
ptab2 <- ptable::pt_ex_nums(parity = TRUE, separation = TRUE)

# different ptables for cells with even/odd number of contributors
# and very small cells
ptab3 <- ptable::pt_ex_nums(parity = FALSE, separation = TRUE)

p_nums1 <- ck_params_nums(
  ptab = ptab1,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.30, 0.03),
    epsilon = c(1, 0.5, 0.2),
    q = 3),
  mu_c = 2,
  same_key = FALSE,
  use_zero_rkeys = FALSE,
  path = f_yaml)
#> yaml configuration '/tmp/RtmpCjMH4Z/file17d3ef88075.yaml' successfully written.

# we read the parameters from the yaml-file
p_nums1 <- ck_read_yaml(path = f_yaml)

# for variables with positive and negative values
p_nums2 <- ck_params_nums(
  ptab = ptab2,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.15, 0.02),
    epsilon = c(1, 0.4, 0.15),
    q = 3),
  mu_c = 2,
  same_key = FALSE)

# simple perturbation parameters (not using the flex-function approach)
p_nums3 <- ck_params_nums(
  ptab = ptab3,
  type = "mean",
  mult_params = ck_simpleparams(p = 0.25),
  mu_c = 2,
  same_key = FALSE)

# use `p_nums1` for all variables
tab$params_nums_set(p_nums1, c("savings", "income", "expend"))
#> --> setting perturbation parameters for variable 'savings'
#> --> setting perturbation parameters for variable 'income'
#> --> setting perturbation parameters for variable 'expend'

# use different parameters for variable `mixed`
tab$params_nums_set(p_nums2, v = "mixed")
#> --> setting perturbation parameters for variable 'mixed'

# identify sensitive cells to which extra protection (`mu_c`) is added.
tab$supp_p(v = "income", p = 85)
#> computing contributing indices | rawdata <--> table; this might take a while
#> p%-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_pq(v = "income", p = 85, q = 90)
#> computing contributing indices | rawdata <--> table; this might take a while
#> pq-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_nk(v = "income", n = 2, k = 90)
#> computing contributing indices | rawdata <--> table; this might take a while
#> nk-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_freq(v = "income", n = 14, weighted = FALSE)
#> freq-rule: 5 new sensitive cells (incl. duplicates) found (total: 5)
tab$supp_val(v = "income", n = 10000, weighted = TRUE)
#> val-rule: 0 new sensitive cells (incl. duplicates) found (total: 5)
tab$supp_cells(
  v = "income",
  inp = data.frame(
    sex = c("female", "female"),
    "age" = c("age_group1", "age_group3")
  )
)
#> cell-rule: 2 new sensitive cells (incl. duplicates) found (total: 7)

# perturb variables
tab$perturb(v = c("income", "savings"))
#> Numeric variable 'income' was perturbed.
#> Numeric variable 'savings' was perturbed.

# extract results
tab$numtab("income", mean_before_sum = TRUE)
#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1369498249 1369499413
#>  2:  Total age_group1 income  9810547  579463097  579501534
#>  3:  Total       ag1a income  9810547  579463097  579501534
#>  4:  Total age_group2 income  5692119  341246342  341198368
#>  5:  Total       ag2a income  5692119  341246342  341198368
#>  6:  Total age_group3 income  4406946  264065926  263964930
#>  7:  Total age_group4 income  2133543  127367290  127266287
#>  8:  Total age_group5 income   848151   53316198   53264044
#>  9:  Total age_group6 income    61672    4039396    3960828
#> 10:   male      Total income 11262049  679542643  679488728
#> 11:      m      Total income 11262049  679542643  679488728
#> 12:   male age_group1 income  4877164  289440139  289394543
#> 13:      m age_group1 income  4877164  289440139  289394543
#> 14:   male       ag1a income  4877164  289440139  289394543
#> 15:      m       ag1a income  4877164  289440139  289394543
#> 16:   male age_group2 income  2811379  169851527  169986331
#> 17:      m age_group2 income  2811379  169851527  169986331
#> 18:   male       ag2a income  2811379  169851527  169986331
#> 19:      m       ag2a income  2811379  169851527  169986331
#> 20:   male age_group3 income  2168169  134715840  134730043
#> 21:      m age_group3 income  2168169  134715840  134730043
#> 22:   male age_group4 income   978510   56936235   56909187
#> 23:      m age_group4 income   978510   56936235   56909187
#> 24:   male age_group5 income   393134   26355889   26446162
#> 25:      m age_group5 income   393134   26355889   26446162
#> 26:   male age_group6 income    33693    2243013    2360506
#> 27:      m age_group6 income    33693    2243013    2360506
#> 28: female      Total income 11690929  689955606  689920030
#> 29:      f      Total income 11690929  689955606  689920030
#> 30: female age_group1 income  4933383  290022958  290005462
#> 31:      f age_group1 income  4933383  290022958  290005462
#> 32: female       ag1a income  4933383  290022958  290005462
#> 33:      f       ag1a income  4933383  290022958  290005462
#> 34: female age_group2 income  2880740  171394815  171369561
#> 35:      f age_group2 income  2880740  171394815  171369561
#> 36: female       ag2a income  2880740  171394815  171369561
#> 37:      f       ag2a income  2880740  171394815  171369561
#> 38: female age_group3 income  2238777  129350086  129293913
#> 39:      f age_group3 income  2238777  129350086  129293913
#> 40: female age_group4 income  1155033   70431055   70472095
#> 41:      f age_group4 income  1155033   70431055   70472095
#> 42: female age_group5 income   455017   26960309   27017903
#> 43:      f age_group5 income   455017   26960309   27017903
#> 44: female age_group6 income    27979    1796383    1794200
#> 45:      f age_group6 income    27979    1796383    1794200
#>        sex        age  vname      uws         ws        pws
tab$numtab("income", mean_before_sum = FALSE)
#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1369498249 1369498831
#>  2:  Total age_group1 income  9810547  579463097  579482315
#>  3:  Total       ag1a income  9810547  579463097  579482315
#>  4:  Total age_group2 income  5692119  341246342  341222354
#>  5:  Total       ag2a income  5692119  341246342  341222354
#>  6:  Total age_group3 income  4406946  264065926  264015423
#>  7:  Total age_group4 income  2133543  127367290  127316779
#>  8:  Total age_group5 income   848151   53316198   53290115
#>  9:  Total age_group6 income    61672    4039396    3999919
#> 10:   male      Total income 11262049  679542643  679515685
#> 11:      m      Total income 11262049  679542643  679515685
#> 12:   male age_group1 income  4877164  289440139  289417340
#> 13:      m age_group1 income  4877164  289440139  289417340
#> 14:   male       ag1a income  4877164  289440139  289417340
#> 15:      m       ag1a income  4877164  289440139  289417340
#> 16:   male age_group2 income  2811379  169851527  169918916
#> 17:      m age_group2 income  2811379  169851527  169918916
#> 18:   male       ag2a income  2811379  169851527  169918916
#> 19:      m       ag2a income  2811379  169851527  169918916
#> 20:   male age_group3 income  2168169  134715840  134722941
#> 21:      m age_group3 income  2168169  134715840  134722941
#> 22:   male age_group4 income   978510   56936235   56922709
#> 23:      m age_group4 income   978510   56936235   56922709
#> 24:   male age_group5 income   393134   26355889   26400987
#> 25:      m age_group5 income   393134   26355889   26400987
#> 26:   male age_group6 income    33693    2243013    2301010
#> 27:      m age_group6 income    33693    2243013    2301010
#> 28: female      Total income 11690929  689955606  689937818
#> 29:      f      Total income 11690929  689955606  689937818
#> 30: female age_group1 income  4933383  290022958  290014210
#> 31:      f age_group1 income  4933383  290022958  290014210
#> 32: female       ag1a income  4933383  290022958  290014210
#> 33:      f       ag1a income  4933383  290022958  290014210
#> 34: female age_group2 income  2880740  171394815  171382188
#> 35:      f age_group2 income  2880740  171394815  171382188
#> 36: female       ag2a income  2880740  171394815  171382188
#> 37:      f       ag2a income  2880740  171394815  171382188
#> 38: female age_group3 income  2238777  129350086  129321996
#> 39:      f age_group3 income  2238777  129350086  129321996
#> 40: female age_group4 income  1155033   70431055   70451572
#> 41:      f age_group4 income  1155033   70431055   70451572
#> 42: female age_group5 income   455017   26960309   26989091
#> 43:      f age_group5 income   455017   26960309   26989091
#> 44: female age_group6 income    27979    1796383    1795291
#> 45:      f age_group6 income    27979    1796383    1795291
#>        sex        age  vname      uws         ws        pws
tab$numtab("savings")
#>        sex        age   vname     uws        ws         pws
#>  1:  Total      Total savings 2273532 136637225 136636469.9
#>  2:  Total age_group1 savings  982386  58478956  58476607.1
#>  3:  Total       ag1a savings  982386  58478956  58476607.1
#>  4:  Total age_group2 savings  552336  33497417  33493875.1
#>  5:  Total       ag2a savings  552336  33497417  33493875.1
#>  6:  Total age_group3 savings  437101  26336817  26335277.7
#>  7:  Total age_group4 savings  214661  12928329  12924257.8
#>  8:  Total age_group5 savings   80451   4999038   4997790.7
#>  9:  Total age_group6 savings    6597    396668    394557.8
#> 10:   male      Total savings 1159816  70502575  70501866.1
#> 11:      m      Total savings 1159816  70502575  70501866.1
#> 12:   male age_group1 savings  517660  30965207  30960917.9
#> 13:      m age_group1 savings  517660  30965207  30960917.9
#> 14:   male       ag1a savings  517660  30965207  30960917.9
#> 15:      m       ag1a savings  517660  30965207  30960917.9
#> 16:   male age_group2 savings  280923  17130426  17135182.6
#> 17:      m age_group2 savings  280923  17130426  17135182.6
#> 18:   male       ag2a savings  280923  17130426  17135182.6
#> 19:      m       ag2a savings  280923  17130426  17135182.6
#> 20:   male age_group3 savings  214970  13418026  13411491.4
#> 21:      m age_group3 savings  214970  13418026  13411491.4
#> 22:   male age_group4 savings   99420   5923398   5922560.2
#> 23:      m age_group4 savings   99420   5923398   5922560.2
#> 24:   male age_group5 savings   43233   2841375   2845876.0
#> 25:      m age_group5 savings   43233   2841375   2845876.0
#> 26:   male age_group6 savings    3610    224143    224312.8
#> 27:      m age_group6 savings    3610    224143    224312.8
#> 28: female      Total savings 1113716  66134650  66133445.7
#> 29:      f      Total savings 1113716  66134650  66133445.7
#> 30: female age_group1 savings  464726  27513749  27508082.0
#> 31:      f age_group1 savings  464726  27513749  27508082.0
#> 32: female       ag1a savings  464726  27513749  27508082.0
#> 33:      f       ag1a savings  464726  27513749  27508082.0
#> 34: female age_group2 savings  271413  16366991  16364665.3
#> 35:      f age_group2 savings  271413  16366991  16364665.3
#> 36: female       ag2a savings  271413  16366991  16364665.3
#> 37:      f       ag2a savings  271413  16366991  16364665.3
#> 38: female age_group3 savings  222131  12918791  12914341.5
#> 39:      f age_group3 savings  222131  12918791  12914341.5
#> 40: female age_group4 savings  115241   7004931   7007067.8
#> 41:      f age_group4 savings  115241   7004931   7007067.8
#> 42: female age_group5 savings   37218   2157663   2161133.0
#> 43:      f age_group5 savings   37218   2157663   2161133.0
#> 44: female age_group6 savings    2987    172525    171985.8
#> 45:      f age_group6 savings    2987    172525    171985.8
#>        sex        age   vname     uws        ws         pws

# results can be resetted, too
tab$reset_cntvars(v = "cnt_males")

# we can then set other parameters and perturb again
tab$params_cnts_set(val = p_cnts1, v = "cnt_males")
#> --> setting perturbation parameters for variable 'cnt_males'

tab$perturb(v = "cnt_males")
#> Count variable 'cnt_males' was perturbed.

# write results to a .csv file
tab$freqtab(
  v = c("total", "cnt_males"),
  path = file.path(tempdir(), "outtab.csv")
)
#> File '/tmp/RtmpCjMH4Z/outtab.csv' successfully written to disk.
#> NULL

# show results containing weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))
#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 274937 4580 274937.0000
#>  2:  Total age_group1     total 1969 117416 1969 117416.0000
#>  3:  Total       ag1a     total 1969 117416 1969 117416.0000
#>  4:  Total age_group2     total 1143  69237 1142  69176.4252
#>  5:  Total       ag2a     total 1143  69237 1142  69176.4252
#>  6:  Total age_group3     total  864  52246  865  52306.4699
#>  7:  Total age_group4     total  423  25007  422  24947.8818
#>  8:  Total age_group5     total  168  10226  167  10165.1310
#>  9:  Total age_group6     total   13    805   13    805.0000
#> 10:   male      Total     total 2296 139534 2296 139534.0000
#> 11:      m      Total     total 2296 139534 2296 139534.0000
#> 12:   male age_group1     total 1015  60753 1015  60753.0000
#> 13:      m age_group1     total 1015  60753 1015  60753.0000
#> 14:   male       ag1a     total 1015  60753 1015  60753.0000
#> 15:      m       ag1a     total 1015  60753 1015  60753.0000
#> 16:   male age_group2     total  571  34840  573  34962.0315
#> 17:      m age_group2     total  571  34840  573  34962.0315
#> 18:   male       ag2a     total  571  34840  573  34962.0315
#> 19:      m       ag2a     total  571  34840  573  34962.0315
#> 20:   male age_group3     total  424  26694  425  26756.9575
#> 21:      m age_group3     total  424  26694  425  26756.9575
#> 22:   male age_group4     total  195  11330  196  11388.1026
#> 23:      m age_group4     total  195  11330  196  11388.1026
#> 24:   male age_group5     total   84   5489   84   5489.0000
#> 25:      m age_group5     total   84   5489   84   5489.0000
#> 26:   male age_group6     total    7    428    6    366.8571
#> 27:      m age_group6     total    7    428    6    366.8571
#> 28: female      Total     total 2284 135403 2285 135462.2833
#> 29:      f      Total     total 2284 135403 2285 135462.2833
#> 30: female age_group1     total  954  56663  952  56544.2096
#> 31:      f age_group1     total  954  56663  952  56544.2096
#> 32: female       ag1a     total  954  56663  952  56544.2096
#> 33:      f       ag1a     total  954  56663  952  56544.2096
#> 34: female age_group2     total  572  34397  571  34336.8654
#> 35:      f age_group2     total  572  34397  571  34336.8654
#> 36: female       ag2a     total  572  34397  571  34336.8654
#> 37:      f       ag2a     total  572  34397  571  34336.8654
#> 38: female age_group3     total  440  25552  439  25493.9273
#> 39:      f age_group3     total  440  25552  439  25493.9273
#> 40: female age_group4     total  228  13677  228  13677.0000
#> 41:      f age_group4     total  228  13677  228  13677.0000
#> 42: female age_group5     total   84   4737   85   4793.3929
#> 43:      f age_group5     total   84   4737   85   4793.3929
#> 44: female age_group6     total    6    377    5    314.1667
#> 45:      f age_group6     total    6    377    5    314.1667
#> 46:  Total      Total cnt_males 2296 139534 2296 139534.0000
#> 47:  Total age_group1 cnt_males 1015  60753 1015  60753.0000
#> 48:  Total       ag1a cnt_males 1015  60753 1015  60753.0000
#> 49:  Total age_group2 cnt_males  571  34840  573  34962.0315
#> 50:  Total       ag2a cnt_males  571  34840  573  34962.0315
#> 51:  Total age_group3 cnt_males  424  26694  425  26756.9575
#> 52:  Total age_group4 cnt_males  195  11330  196  11388.1026
#> 53:  Total age_group5 cnt_males   84   5489   84   5489.0000
#> 54:  Total age_group6 cnt_males    7    428    6    366.8571
#> 55:   male      Total cnt_males 2296 139534 2296 139534.0000
#> 56:      m      Total cnt_males 2296 139534 2296 139534.0000
#> 57:   male age_group1 cnt_males 1015  60753 1015  60753.0000
#> 58:      m age_group1 cnt_males 1015  60753 1015  60753.0000
#> 59:   male       ag1a cnt_males 1015  60753 1015  60753.0000
#> 60:      m       ag1a cnt_males 1015  60753 1015  60753.0000
#> 61:   male age_group2 cnt_males  571  34840  573  34962.0315
#> 62:      m age_group2 cnt_males  571  34840  573  34962.0315
#> 63:   male       ag2a cnt_males  571  34840  573  34962.0315
#> 64:      m       ag2a cnt_males  571  34840  573  34962.0315
#> 65:   male age_group3 cnt_males  424  26694  425  26756.9575
#> 66:      m age_group3 cnt_males  424  26694  425  26756.9575
#> 67:   male age_group4 cnt_males  195  11330  196  11388.1026
#> 68:      m age_group4 cnt_males  195  11330  196  11388.1026
#> 69:   male age_group5 cnt_males   84   5489   84   5489.0000
#> 70:      m age_group5 cnt_males   84   5489   84   5489.0000
#> 71:   male age_group6 cnt_males    7    428    6    366.8571
#> 72:      m age_group6 cnt_males    7    428    6    366.8571
#> 73: female      Total cnt_males    0      0    0      0.0000
#> 74:      f      Total cnt_males    0      0    0      0.0000
#> 75: female age_group1 cnt_males    0      0    0      0.0000
#> 76:      f age_group1 cnt_males    0      0    0      0.0000
#> 77: female       ag1a cnt_males    0      0    0      0.0000
#> 78:      f       ag1a cnt_males    0      0    0      0.0000
#> 79: female age_group2 cnt_males    0      0    0      0.0000
#> 80:      f age_group2 cnt_males    0      0    0      0.0000
#> 81: female       ag2a cnt_males    0      0    0      0.0000
#> 82:      f       ag2a cnt_males    0      0    0      0.0000
#> 83: female age_group3 cnt_males    0      0    0      0.0000
#> 84:      f age_group3 cnt_males    0      0    0      0.0000
#> 85: female age_group4 cnt_males    0      0    0      0.0000
#> 86:      f age_group4 cnt_males    0      0    0      0.0000
#> 87: female age_group5 cnt_males    0      0    0      0.0000
#> 88:      f age_group5 cnt_males    0      0    0      0.0000
#> 89: female age_group6 cnt_males    0      0    0      0.0000
#> 90:      f age_group6 cnt_males    0      0    0      0.0000
#>        sex        age     vname  uwc     wc puwc         pwc

# utility measures for a count variable
tab$measures_cnts(v = "total", exclude_zeros = TRUE)
#> $overview
#>    noise cnt        pct
#> 1:    -2   4 0.08888889
#> 2:    -1   9 0.20000000
#> 3:     0  14 0.31111111
#> 4:     1  14 0.31111111
#> 5:     2   4 0.08888889
#> 
#> $measures
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.200 0.000 0.002
#>  5:    Q40 1.000 0.001 0.016
#>  6:   Mean 0.867 0.016 0.036
#>  7: Median 1.000 0.002 0.021
#>  8:    Q60 1.000 0.002 0.024
#>  9:    Q70 1.000 0.002 0.032
#> 10:    Q80 1.000 0.004 0.042
#> 11:    Q90 2.000 0.012 0.054
#> 12:    Q95 2.000 0.143 0.196
#> 13:    Q99 2.000 0.167 0.213
#> 14:    Max 2.000 0.167 0.213
#> 
#> $cumdistr_d1
#>    cat cnt       pct
#> 1:   0  14 0.3111111
#> 2:   1  37 0.8222222
#> 3:   2  45 1.0000000
#> 
#> $cumdistr_d2
#>            cat cnt       pct
#> 1:    [0,0.02]  41 0.9111111
#> 2: (0.02,0.05]  41 0.9111111
#> 3:  (0.05,0.1]  41 0.9111111
#> 4:   (0.1,0.2]  45 1.0000000
#> 5:   (0.2,0.3]  45 1.0000000
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $cumdistr_d3
#>            cat cnt       pct
#> 1:    [0,0.02]  19 0.4222222
#> 2: (0.02,0.05]  39 0.8666667
#> 3:  (0.05,0.1]  41 0.9111111
#> 4:   (0.1,0.2]  43 0.9555556
#> 5:   (0.2,0.3]  45 1.0000000
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $false_zero
#> [1] 0
#> 
#> $false_nonzero
#> [1] 0
#> 
#> $exclude_zeros
#> [1] TRUE
#> 

# modifications for perturbed count variables
tab$mod_cnts()
#>         sex        age row_nr pert      ckey  countvar
#>   1:  Total      Total     15    0 0.3692195     total
#>   2:  Total age_group1     15    0 0.3595446     total
#>   3:  Total       ag1a     15    0 0.3595446     total
#>   4:  Total age_group2     14   -1 0.1503608     total
#>   5:  Total       ag2a     14   -1 0.1503608     total
#>  ---                                                  
#> 131:      f age_group4     -1    0 0.0000000 cnt_males
#> 132: female age_group5     -1    0 0.0000000 cnt_males
#> 133:      f age_group5     -1    0 0.0000000 cnt_males
#> 134: female age_group6     -1    0 0.0000000 cnt_males
#> 135:      f age_group6     -1    0 0.0000000 cnt_males

# display a summary about utility measures
tab$summary()
#> ┌──────────────────────────────────────────────┐
#> │Utility measures for perturbed count variables│
#> └──────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#>          countvar Min Q10  Q20 Q30  Q40   Mean Median Q60 Q70 Q80 Q90 Q95 Q99
#> 1:          total  -2  -1 -1.0  -1 -0.4 -0.111      0 0.0   0   1   1   2   2
#> 2: cnt_highincome  -2  -2 -1.2   0  0.0  0.267      0 0.4   1   1   3   3   3
#> 3:      cnt_males  -1   0  0.0   0  0.0  0.333      0 0.0   0   1   2   2   2
#>    Max
#> 1:   2
#> 2:   3
#> 3:   2
#> 
#> ── Distance-based measures ─────────────────────────────────────────────────────
#> ✔ Variable: 'total'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.200 0.000 0.002
#>  5:    Q40 1.000 0.001 0.016
#>  6:   Mean 0.867 0.016 0.036
#>  7: Median 1.000 0.002 0.021
#>  8:    Q60 1.000 0.002 0.024
#>  9:    Q70 1.000 0.002 0.032
#> 10:    Q80 1.000 0.004 0.042
#> 11:    Q90 2.000 0.012 0.054
#> 12:    Q95 2.000 0.143 0.196
#> 13:    Q99 2.000 0.167 0.213
#> 14:    Max 2.000 0.167 0.213
#> 
#> ✔ Variable: 'cnt_males'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 0.000 0.000 0.000
#>  6:   Mean 0.778 0.017 0.038
#>  7: Median 1.000 0.002 0.024
#>  8:    Q60 1.000 0.004 0.036
#>  9:    Q70 1.000 0.004 0.042
#> 10:    Q80 1.800 0.005 0.042
#> 11:    Q90 2.000 0.060 0.104
#> 12:    Q95 2.000 0.143 0.196
#> 13:    Q99 2.000 0.143 0.196
#> 14:    Max 2.000 0.143 0.196
#> 
#> ✔ Variable: 'cnt_highincome'
#> 
#>       what  d1    d2    d3
#>  1:    Min 0.0 0.000 0.000
#>  2:    Q10 0.0 0.000 0.000
#>  3:    Q20 0.0 0.000 0.000
#>  4:    Q30 1.0 0.004 0.032
#>  5:    Q40 1.0 0.011 0.061
#>  6:   Mean 1.3 0.047 0.107
#>  7: Median 1.0 0.015 0.068
#>  8:    Q60 2.0 0.022 0.106
#>  9:    Q70 2.0 0.053 0.196
#> 10:    Q80 2.0 0.073 0.196
#> 11:    Q90 3.0 0.143 0.267
#> 12:    Q95 3.0 0.150 0.281
#> 13:    Q99 3.0 0.286 0.354
#> 14:    Max 3.0 0.286 0.354
#> 
#> ┌──────────────────────────────────────────────────┐
#> │Utility measures for perturbed numerical variables│
#> └──────────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#>      vname        Min        Q10        Q20        Q30        Q40      Mean
#> 1:  expend        Inf         NA         NA         NA         NA       NaN
#> 2:  income -50511.299 -27637.086 -23987.638 -21796.870 -12986.632  1364.590
#> 3: savings  -6534.562  -5667.012  -4289.122  -3541.906  -2325.667 -1324.415
#> 4:   mixed        Inf         NA         NA         NA         NA       NaN
#>       Median       Q60       Q70       Q80      Q90       Q95       Q99
#> 1:        NA        NA        NA        NA       NA        NA        NA
#> 2: -8748.106 -1091.758 16794.728 28781.562 57996.51 67388.595 67388.595
#> 3: -1539.324  -837.824  -573.105  2136.838  4501.00  4756.603  4756.603
#> 4:        NA        NA        NA        NA       NA        NA        NA
#>          Max
#> 1:      -Inf
#> 2: 67388.595
#> 3:  4756.603
#> 4:      -Inf
# }