Set parameters required to perturb numeric variables using a flex function

ck_flexparams() allows to define a flex function that is used to lookup perturbation magnitudes (percentages) used when perturbing continuous variables.

ck_flexparams(fp, p = c(0.25, 0.05), epsilon = 1, q = 3)

Arguments

fp
(numeric scalar); at which point should the noise coefficient function reaches its desired maximum (defined by the first element of p)

p
a numeric vector of length 2 where both elements specify a percentage. The first value refers to the desired maximum perturbation percentage for small cells (depending on fp) while the second element refers to the desired maximum perturbation percentage for large cells. Both values must be between 0 and 1 and need to be in descending order.

epsilon
a numeric vector in descending order with all values >= 0 and <= 1 with the first element forced to equal 1. The length of this vector must correspond with the number top_k specified in ck_params_nums() when creating parameters for type == "top_contr" which is checked at runtime. This setting allows to use different flex-functions for the largest top_k contributors.

q
(numeric scalar); Parameter of the function; q needs to be >= 1

Value

an object suitable as input for ck_params_nums().

Details

details about the flex function can be found in Deliverable D4.2, Part I in SGA "Open Source tools for perturbative confidentiality methods"

See also

Examples

x <- ck_create_testdata()

# create some 0/1 variables that should be perturbed later
x[, cnt_females := ifelse(sex == "male", 0, 1)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              97          25.00000           0
#>    2:   2530      28       1              57          25.00000           1
#>    3:   6920     550       1              38          25.00000           0
#>    4:   7960     870       1              53          25.00000           0
#>    5:   9030      20       2              44          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              43          16.66667           1
#> 4577:   1420     987    1000              61          16.66667           0
#> 4578:   8900     684    1000              78          16.66667           0
#> 4579:   3880     294    1000              30          16.66667           1
#> 4580:   4830     911    1000              56          16.66667           0

x[, cnt_males := ifelse(sex == "male", 1, 0)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              97          25.00000           0
#>    2:   2530      28       1              57          25.00000           1
#>    3:   6920     550       1              38          25.00000           0
#>    4:   7960     870       1              53          25.00000           0
#>    5:   9030      20       2              44          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              43          16.66667           1
#> 4577:   1420     987    1000              61          16.66667           0
#> 4578:   8900     684    1000              78          16.66667           0
#> 4579:   3880     294    1000              30          16.66667           1
#> 4580:   4830     911    1000              56          16.66667           0
#>       cnt_males
#>    1:         1
#>    2:         0
#>    3:         1
#>    4:         1
#>    5:         1
#>   ---          
#> 4576:         0
#> 4577:         1
#> 4578:         1
#> 4579:         0
#> 4580:         1

x[, cnt_highincome := ifelse(income >= 9000, 1, 0)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              97          25.00000           0
#>    2:   2530      28       1              57          25.00000           1
#>    3:   6920     550       1              38          25.00000           0
#>    4:   7960     870       1              53          25.00000           0
#>    5:   9030      20       2              44          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              43          16.66667           1
#> 4577:   1420     987    1000              61          16.66667           0
#> 4578:   8900     684    1000              78          16.66667           0
#> 4579:   3880     294    1000              30          16.66667           1
#> 4580:   4830     911    1000              56          16.66667           0
#>       cnt_males cnt_highincome
#>    1:         1              0
#>    2:         0              0
#>    3:         1              0
#>    4:         1              0
#>    5:         1              1
#>   ---                         
#> 4576:         0              0
#> 4577:         1              0
#> 4578:         1              0
#> 4579:         0              0
#> 4580:         1              0

# a variable with positive and negative contributions
x[, mixed := sample(-10:10, nrow(x), replace = TRUE)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              97          25.00000           0
#>    2:   2530      28       1              57          25.00000           1
#>    3:   6920     550       1              38          25.00000           0
#>    4:   7960     870       1              53          25.00000           0
#>    5:   9030      20       2              44          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              43          16.66667           1
#> 4577:   1420     987    1000              61          16.66667           0
#> 4578:   8900     684    1000              78          16.66667           0
#> 4579:   3880     294    1000              30          16.66667           1
#> 4580:   4830     911    1000              56          16.66667           0
#>       cnt_males cnt_highincome mixed
#>    1:         1              0    -1
#>    2:         0              0    10
#>    3:         1              0     8
#>    4:         1              0     6
#>    5:         1              1    -9
#>   ---                               
#> 4576:         0              0    10
#> 4577:         1              0     8
#> 4578:         1              0     9
#> 4579:         0              0    10
#> 4580:         1              0     9

# create record keys
x$rkey <- ck_generate_rkeys(dat = x)

# define required inputs

# hierarchy with some bogus codes
d_sex <- hier_create(root = "Total", nodes = c("male", "female"))
d_sex <- hier_add(d_sex, root = "female", "f")
d_sex <- hier_add(d_sex, root = "male", "m")

d_age <- hier_create(root = "Total", nodes = paste0("age_group", 1:6))
d_age <- hier_add(d_age, root = "age_group1", "ag1a")
d_age <- hier_add(d_age, root = "age_group2", "ag2a")

# define the cell key object
countvars <- c("cnt_females", "cnt_males", "cnt_highincome")
numvars <- c("expend", "income", "savings", "mixed")
tab <- ck_setup(
  x = x,
  rkey = "rkey",
  dims = list(sex = d_sex, age = d_age),
  w = "sampling_weight",
  countvars = countvars,
  numvars = numvars)

# show some information about this table instance
tab$print() # identical with print(tab)

#> ── Table Information ───────────────────────────────────────────────────────────
#> ✔ 45 cells in 2 dimensions ('sex', 'age')
#> ✔ weights: yes
#> ── Tabulated / Perturbed countvars ─────────────────────────────────────────────
#> ☐ 'total'
#> ☐ 'cnt_females'
#> ☐ 'cnt_males'
#> ☐ 'cnt_highincome'
#> ── Tabulated / Perturbed numvars ───────────────────────────────────────────────
#> ☐ 'expend'
#> ☐ 'income'
#> ☐ 'savings'
#> ☐ 'mixed'

# information about the hierarchies
tab$hierarchy_info()

#> $sex
#>      code level is_leaf parent
#> 1:  Total     1   FALSE   <NA>
#> 2:   male     2   FALSE  Total
#> 3: female     2   FALSE  Total
#> 4:      f     3    TRUE female
#> 5:      m     3    TRUE   male
#> 
#> $age
#>          code level is_leaf     parent
#> 1:      Total     1   FALSE       <NA>
#> 2: age_group1     2   FALSE      Total
#> 3: age_group2     2   FALSE      Total
#> 4: age_group3     2    TRUE      Total
#> 5: age_group4     2    TRUE      Total
#> 6: age_group5     2    TRUE      Total
#> 7: age_group6     2    TRUE      Total
#> 8:       ag1a     3    TRUE age_group1
#> 9:       ag2a     3    TRUE age_group2
#> 

# which variables have been defined?
tab$allvars()

#> $cntvars
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"
#> 
#> $numvars
#> [1] "expend"  "income"  "savings" "mixed"  
#> 

# count variables
tab$cntvars()

#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"

# continuous variables
tab$numvars()

#> [1] "expend"  "income"  "savings" "mixed"  

# create perturbation parameters for "total" variable and
# write to yaml-file

# create a ptable using functionality from the ptable-pkg
f_yaml <- tempfile(fileext = ".yaml")
p_cnts1 <- ck_params_cnts(
  ptab = ptable::pt_ex_cnts(),
  path = f_yaml)

#> yaml configuration '/tmp/RtmpNQSP5C/file3ec225d9bbe.yaml' successfully written.

# read parameters from yaml-file and set them for variable `"total"`
p_cnts1 <- ck_read_yaml(path = f_yaml)

tab$params_cnts_set(val = p_cnts1, v = "total")

#> --> setting perturbation parameters for variable 'total'

# create alternative perturbation parameters by specifying parameters
para2 <- ptable::pt_create_pParams(
  D = 8, V = 3, js = 2, table = "cnts")

#> Ignoring argument `step` for freqency tables

p_cnts2 <- ck_params_cnts(ptab = para2)

# use these ptable it for the remaining variables
tab$params_cnts_set(val = p_cnts2, v = countvars)

#> --> setting perturbation parameters for variable 'cnt_females'

#> --> setting perturbation parameters for variable 'cnt_males'

#> --> setting perturbation parameters for variable 'cnt_highincome'

# perturb a variable
tab$perturb(v = "total")

#> Count variable 'total' was perturbed.

# multiple variables can be perturbed as well
tab$perturb(v = c("cnt_males", "cnt_highincome"))

#> Count variable 'cnt_males' was perturbed.

#> Count variable 'cnt_highincome' was perturbed.

# return weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))

#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 278847 4580 278847.0000
#>  2:  Total age_group1     total 1969 120025 1968 119964.0427
#>  3:  Total       ag1a     total 1969 120025 1968 119964.0427
#>  4:  Total age_group2     total 1143  68867 1143  68867.0000
#>  5:  Total       ag2a     total 1143  68867 1143  68867.0000
#>  6:  Total age_group3     total  864  52558  865  52618.8310
#>  7:  Total age_group4     total  423  26224  423  26224.0000
#>  8:  Total age_group5     total  168  10353  170  10476.2500
#>  9:  Total age_group6     total   13    820   11    693.8462
#> 10: female      Total     total 2284 137871 2284 137871.0000
#> 11:      f      Total     total 2284 137871 2284 137871.0000
#> 12: female age_group1     total  954  58008  954  58008.0000
#> 13:      f age_group1     total  954  58008  954  58008.0000
#> 14: female       ag1a     total  954  58008  954  58008.0000
#> 15:      f       ag1a     total  954  58008  954  58008.0000
#> 16: female age_group2     total  572  33641  573  33699.8129
#> 17:      f age_group2     total  572  33641  573  33699.8129
#> 18: female       ag2a     total  572  33641  573  33699.8129
#> 19:      f       ag2a     total  572  33641  573  33699.8129
#> 20: female age_group3     total  440  26817  441  26877.9477
#> 21:      f age_group3     total  440  26817  441  26877.9477
#> 22: female age_group4     total  228  13877  226  13755.2719
#> 23:      f age_group4     total  228  13877  226  13755.2719
#> 24: female age_group5     total   84   5138   85   5199.1667
#> 25:      f age_group5     total   84   5138   85   5199.1667
#> 26: female age_group6     total    6    390    5    325.0000
#> 27:      f age_group6     total    6    390    5    325.0000
#> 28:   male      Total     total 2296 140976 2295 140914.5993
#> 29:      m      Total     total 2296 140976 2295 140914.5993
#> 30:   male age_group1     total 1015  62017 1015  62017.0000
#> 31:      m age_group1     total 1015  62017 1015  62017.0000
#> 32:   male       ag1a     total 1015  62017 1015  62017.0000
#> 33:      m       ag1a     total 1015  62017 1015  62017.0000
#> 34:   male age_group2     total  571  35226  572  35287.6918
#> 35:      m age_group2     total  571  35226  572  35287.6918
#> 36:   male       ag2a     total  571  35226  572  35287.6918
#> 37:      m       ag2a     total  571  35226  572  35287.6918
#> 38:   male age_group3     total  424  25741  422  25619.5802
#> 39:      m age_group3     total  424  25741  422  25619.5802
#> 40:   male age_group4     total  195  12347  194  12283.6821
#> 41:      m age_group4     total  195  12347  194  12283.6821
#> 42:   male age_group5     total   84   5215   82   5090.8333
#> 43:      m age_group5     total   84   5215   82   5090.8333
#> 44:   male age_group6     total    7    430    8    491.4286
#> 45:      m age_group6     total    7    430    8    491.4286
#> 46:  Total      Total cnt_males 2296 140976 2295 140914.5993
#> 47:  Total age_group1 cnt_males 1015  62017 1015  62017.0000
#> 48:  Total       ag1a cnt_males 1015  62017 1015  62017.0000
#> 49:  Total age_group2 cnt_males  571  35226  572  35287.6918
#> 50:  Total       ag2a cnt_males  571  35226  572  35287.6918
#> 51:  Total age_group3 cnt_males  424  25741  420  25498.1604
#> 52:  Total age_group4 cnt_males  195  12347  194  12283.6821
#> 53:  Total age_group5 cnt_males   84   5215   81   5028.7500
#> 54:  Total age_group6 cnt_males    7    430    8    491.4286
#> 55: female      Total cnt_males    0      0    0      0.0000
#> 56:      f      Total cnt_males    0      0    0      0.0000
#> 57: female age_group1 cnt_males    0      0    0      0.0000
#> 58:      f age_group1 cnt_males    0      0    0      0.0000
#> 59: female       ag1a cnt_males    0      0    0      0.0000
#> 60:      f       ag1a cnt_males    0      0    0      0.0000
#> 61: female age_group2 cnt_males    0      0    0      0.0000
#> 62:      f age_group2 cnt_males    0      0    0      0.0000
#> 63: female       ag2a cnt_males    0      0    0      0.0000
#> 64:      f       ag2a cnt_males    0      0    0      0.0000
#> 65: female age_group3 cnt_males    0      0    0      0.0000
#> 66:      f age_group3 cnt_males    0      0    0      0.0000
#> 67: female age_group4 cnt_males    0      0    0      0.0000
#> 68:      f age_group4 cnt_males    0      0    0      0.0000
#> 69: female age_group5 cnt_males    0      0    0      0.0000
#> 70:      f age_group5 cnt_males    0      0    0      0.0000
#> 71: female age_group6 cnt_males    0      0    0      0.0000
#> 72:      f age_group6 cnt_males    0      0    0      0.0000
#> 73:   male      Total cnt_males 2296 140976 2295 140914.5993
#> 74:      m      Total cnt_males 2296 140976 2295 140914.5993
#> 75:   male age_group1 cnt_males 1015  62017 1015  62017.0000
#> 76:      m age_group1 cnt_males 1015  62017 1015  62017.0000
#> 77:   male       ag1a cnt_males 1015  62017 1015  62017.0000
#> 78:      m       ag1a cnt_males 1015  62017 1015  62017.0000
#> 79:   male age_group2 cnt_males  571  35226  572  35287.6918
#> 80:      m age_group2 cnt_males  571  35226  572  35287.6918
#> 81:   male       ag2a cnt_males  571  35226  572  35287.6918
#> 82:      m       ag2a cnt_males  571  35226  572  35287.6918
#> 83:   male age_group3 cnt_males  424  25741  420  25498.1604
#> 84:      m age_group3 cnt_males  424  25741  420  25498.1604
#> 85:   male age_group4 cnt_males  195  12347  194  12283.6821
#> 86:      m age_group4 cnt_males  195  12347  194  12283.6821
#> 87:   male age_group5 cnt_males   84   5215   81   5028.7500
#> 88:      m age_group5 cnt_males   84   5215   81   5028.7500
#> 89:   male age_group6 cnt_males    7    430    8    491.4286
#> 90:      m age_group6 cnt_males    7    430    8    491.4286
#>        sex        age     vname  uwc     wc puwc         pwc

# numerical variables (positive variables using flex-function)
# we also write the config to a yaml file
f_yaml <- tempfile(fileext = ".yaml")

# create a ptable using functionality from the ptable-pkg
# a single ptable for all cells
ptab1 <- ptable::pt_ex_nums(parity = TRUE, separation = FALSE)

# a single ptab for all cells except for very small ones
ptab2 <- ptable::pt_ex_nums(parity = TRUE, separation = TRUE)

# different ptables for cells with even/odd number of contributors
# and very small cells
ptab3 <- ptable::pt_ex_nums(parity = FALSE, separation = TRUE)

p_nums1 <- ck_params_nums(
  ptab = ptab1,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.30, 0.03),
    epsilon = c(1, 0.5, 0.2),
    q = 3),
  mu_c = 2,
  same_key = FALSE,
  use_zero_rkeys = FALSE,
  path = f_yaml)

#> yaml configuration '/tmp/RtmpNQSP5C/file3ec242ce595.yaml' successfully written.

# we read the parameters from the yaml-file
p_nums1 <- ck_read_yaml(path = f_yaml)

# for variables with positive and negative values
p_nums2 <- ck_params_nums(
  ptab = ptab2,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.15, 0.02),
    epsilon = c(1, 0.4, 0.15),
    q = 3),
  mu_c = 2,
  same_key = FALSE)

# simple perturbation parameters (not using the flex-function approach)
p_nums3 <- ck_params_nums(
  ptab = ptab3,
  type = "mean",
  mult_params = ck_simpleparams(p = 0.25),
  mu_c = 2,
  same_key = FALSE)

# use `p_nums1` for all variables
tab$params_nums_set(p_nums1, c("savings", "income", "expend"))

#> --> setting perturbation parameters for variable 'savings'

#> --> setting perturbation parameters for variable 'income'

#> --> setting perturbation parameters for variable 'expend'

# use different parameters for variable `mixed`
tab$params_nums_set(p_nums2, v = "mixed")

#> --> setting perturbation parameters for variable 'mixed'

# identify sensitive cells to which extra protection (`mu_c`) is added.
tab$supp_p(v = "income", p = 85)

#> p%-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)

tab$supp_pq(v = "income", p = 85, q = 90)

#> pq-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)

tab$supp_nk(v = "income", n = 2, k = 90)

#> nk-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)

tab$supp_freq(v = "income", n = 14, weighted = FALSE)

#> freq-rule: 5 new sensitive cells (incl. duplicates) found (total: 5)

tab$supp_val(v = "income", n = 10000, weighted = TRUE)

#> val-rule: 0 new sensitive cells (incl. duplicates) found (total: 5)

# perturb variables
tab$perturb(v = c("income", "savings"))

#> Numeric variable 'income' was perturbed.

#> Numeric variable 'savings' was perturbed.

# extract results
tab$numtab("income", mean_before_sum = TRUE)

#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1405528860 1405538067
#>  2:  Total age_group1 income  9810547  599706128  599630710
#>  3:  Total       ag1a income  9810547  599706128  599630710
#>  4:  Total age_group2 income  5692119  342451169  342409927
#>  5:  Total       ag2a income  5692119  342451169  342409927
#>  6:  Total age_group3 income  4406946  271311756  271221464
#>  7:  Total age_group4 income  2133543  134387256  134440345
#>  8:  Total age_group5 income   848151   53679018   53733121
#>  9:  Total age_group6 income    61672    3993533    3923840
#> 10: female      Total income 11690929  709736064  709778892
#> 11:      f      Total income 11690929  709736064  709778892
#> 12: female age_group1 income  4933383  302171000  302180834
#> 13:      f age_group1 income  4933383  302171000  302180834
#> 14: female       ag1a income  4933383  302171000  302180834
#> 15:      f       ag1a income  4933383  302171000  302180834
#> 16: female age_group2 income  2880740  167235298  167222271
#> 17:      f age_group2 income  2880740  167235298  167222271
#> 18: female       ag2a income  2880740  167235298  167222271
#> 19:      f       ag2a income  2880740  167235298  167222271
#> 20: female age_group3 income  2238777  138784433  138684105
#> 21:      f age_group3 income  2238777  138784433  138684105
#> 22: female age_group4 income  1155033   71163028   71229933
#> 23:      f age_group4 income  1155033   71163028   71229933
#> 24: female age_group5 income   455017   28241696   28353451
#> 25:      f age_group5 income   455017   28241696   28353451
#> 26: female age_group6 income    27979    2140609    2471848
#> 27:      f age_group6 income    27979    2140609    2471848
#> 28:   male      Total income 11262049  695792796  695789174
#> 29:      m      Total income 11262049  695792796  695789174
#> 30:   male age_group1 income  4877164  297535128  297639613
#> 31:      m age_group1 income  4877164  297535128  297639613
#> 32:   male       ag1a income  4877164  297535128  297639613
#> 33:      m       ag1a income  4877164  297535128  297639613
#> 34:   male age_group2 income  2811379  175215871  175376723
#> 35:      m age_group2 income  2811379  175215871  175376723
#> 36:   male       ag2a income  2811379  175215871  175376723
#> 37:      m       ag2a income  2811379  175215871  175376723
#> 38:   male age_group3 income  2168169  132527323  132658520
#> 39:      m age_group3 income  2168169  132527323  132658520
#> 40:   male age_group4 income   978510   63224228   63167801
#> 41:      m age_group4 income   978510   63224228   63167801
#> 42:   male age_group5 income   393134   25437322   25458680
#> 43:      m age_group5 income   393134   25437322   25458680
#> 44:   male age_group6 income    33693    1852924    1791655
#> 45:      m age_group6 income    33693    1852924    1791655
#>        sex        age  vname      uws         ws        pws

tab$numtab("income", mean_before_sum = FALSE)

#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1405528860 1405533464
#>  2:  Total age_group1 income  9810547  599706128  599668418
#>  3:  Total       ag1a income  9810547  599706128  599668418
#>  4:  Total age_group2 income  5692119  342451169  342430547
#>  5:  Total       ag2a income  5692119  342451169  342430547
#>  6:  Total age_group3 income  4406946  271311756  271266606
#>  7:  Total age_group4 income  2133543  134387256  134413798
#>  8:  Total age_group5 income   848151   53679018   53706063
#>  9:  Total age_group6 income    61672    3993533    3958533
#> 10: female      Total income 11690929  709736064  709757478
#> 11:      f      Total income 11690929  709736064  709757478
#> 12: female age_group1 income  4933383  302171000  302175917
#> 13:      f age_group1 income  4933383  302171000  302175917
#> 14: female       ag1a income  4933383  302171000  302175917
#> 15:      f       ag1a income  4933383  302171000  302175917
#> 16: female age_group2 income  2880740  167235298  167228784
#> 17:      f age_group2 income  2880740  167235298  167228784
#> 18: female       ag2a income  2880740  167235298  167228784
#> 19:      f       ag2a income  2880740  167235298  167228784
#> 20: female age_group3 income  2238777  138784433  138734260
#> 21:      f age_group3 income  2238777  138784433  138734260
#> 22: female age_group4 income  1155033   71163028   71196472
#> 23:      f age_group4 income  1155033   71163028   71196472
#> 24: female age_group5 income   455017   28241696   28297518
#> 25:      f age_group5 income   455017   28241696   28297518
#> 26: female age_group6 income    27979    2140609    2300274
#> 27:      f age_group6 income    27979    2140609    2300274
#> 28:   male      Total income 11262049  695792796  695790985
#> 29:      m      Total income 11262049  695792796  695790985
#> 30:   male age_group1 income  4877164  297535128  297587366
#> 31:      m age_group1 income  4877164  297535128  297587366
#> 32:   male       ag1a income  4877164  297535128  297587366
#> 33:      m       ag1a income  4877164  297535128  297587366
#> 34:   male age_group2 income  2811379  175215871  175296279
#> 35:      m age_group2 income  2811379  175215871  175296279
#> 36:   male       ag2a income  2811379  175215871  175296279
#> 37:      m       ag2a income  2811379  175215871  175296279
#> 38:   male age_group3 income  2168169  132527323  132592905
#> 39:      m age_group3 income  2168169  132527323  132592905
#> 40:   male age_group4 income   978510   63224228   63196008
#> 41:      m age_group4 income   978510   63224228   63196008
#> 42:   male age_group5 income   393134   25437322   25447999
#> 43:      m age_group5 income   393134   25437322   25447999
#> 44:   male age_group6 income    33693    1852924    1822032
#> 45:      m age_group6 income    33693    1852924    1822032
#>        sex        age  vname      uws         ws        pws

tab$numtab("savings")

#>        sex        age   vname     uws        ws         pws
#>  1:  Total      Total savings 2273532 138721646 138718658.7
#>  2:  Total age_group1 savings  982386  59859223  59860176.4
#>  3:  Total       ag1a savings  982386  59859223  59860176.4
#>  4:  Total age_group2 savings  552336  33389971  33386441.9
#>  5:  Total       ag2a savings  552336  33389971  33386441.9
#>  6:  Total age_group3 savings  437101  26708009  26704684.5
#>  7:  Total age_group4 savings  214661  13355292  13357173.9
#>  8:  Total age_group5 savings   80451   4965604   4967078.5
#>  9:  Total age_group6 savings    6597    443547    437511.5
#> 10: female      Total savings 1113716  67340067  67346165.8
#> 11:      f      Total savings 1113716  67340067  67346165.8
#> 12: female age_group1 savings  464726  28226195  28231830.6
#> 13:      f age_group1 savings  464726  28226195  28231830.6
#> 14: female       ag1a savings  464726  28226195  28231830.6
#> 15:      f       ag1a savings  464726  28226195  28231830.6
#> 16: female age_group2 savings  271413  15972400  15970713.5
#> 17:      f age_group2 savings  271413  15972400  15970713.5
#> 18: female       ag2a savings  271413  15972400  15970713.5
#> 19:      f       ag2a savings  271413  15972400  15970713.5
#> 20: female age_group3 savings  222131  13535588  13528878.8
#> 21:      f age_group3 savings  222131  13535588  13528878.8
#> 22: female age_group4 savings  115241   7032794   7034950.1
#> 23:      f age_group4 savings  115241   7032794   7034950.1
#> 24: female age_group5 savings   37218   2339807   2343890.0
#> 25:      f age_group5 savings   37218   2339807   2343890.0
#> 26: female age_group6 savings    2987    233283    243655.8
#> 27:      f age_group6 savings    2987    233283    243655.8
#> 28:   male      Total savings 1159816  71381579  71382166.7
#> 29:      m      Total savings 1159816  71381579  71382166.7
#> 30:   male age_group1 savings  517660  31633028  31637388.8
#> 31:      m age_group1 savings  517660  31633028  31637388.8
#> 32:   male       ag1a savings  517660  31633028  31637388.8
#> 33:      m       ag1a savings  517660  31633028  31637388.8
#> 34:   male age_group2 savings  280923  17417571  17424414.2
#> 35:      m age_group2 savings  280923  17417571  17424414.2
#> 36:   male       ag2a savings  280923  17417571  17424414.2
#> 37:      m       ag2a savings  280923  17417571  17424414.2
#> 38:   male age_group3 savings  214970  13172421  13163208.9
#> 39:      m age_group3 savings  214970  13172421  13163208.9
#> 40:   male age_group4 savings   99420   6322498   6319841.3
#> 41:      m age_group4 savings   99420   6322498   6319841.3
#> 42:   male age_group5 savings   43233   2625797   2627381.6
#> 43:      m age_group5 savings   43233   2625797   2627381.6
#> 44:   male age_group6 savings    3610    210264    212259.0
#> 45:      m age_group6 savings    3610    210264    212259.0
#>        sex        age   vname     uws        ws         pws

# results can be resetted, too
tab$reset_cntvars(v = "cnt_males")

# we can then set other parameters and perturb again
tab$params_cnts_set(val = p_cnts1, v = "cnt_males")

#> --> setting perturbation parameters for variable 'cnt_males'

tab$perturb(v = "cnt_males")

#> Count variable 'cnt_males' was perturbed.

# write to a file "outtab.csv" (.csv is automatically added to the path)
if (FALSE) {
tab$freqtab(v = c("total", "cnt_males"), path = "outtab.csv")
}

# show results containing weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))

#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 278847 4580 278847.0000
#>  2:  Total age_group1     total 1969 120025 1968 119964.0427
#>  3:  Total       ag1a     total 1969 120025 1968 119964.0427
#>  4:  Total age_group2     total 1143  68867 1143  68867.0000
#>  5:  Total       ag2a     total 1143  68867 1143  68867.0000
#>  6:  Total age_group3     total  864  52558  865  52618.8310
#>  7:  Total age_group4     total  423  26224  423  26224.0000
#>  8:  Total age_group5     total  168  10353  170  10476.2500
#>  9:  Total age_group6     total   13    820   11    693.8462
#> 10: female      Total     total 2284 137871 2284 137871.0000
#> 11:      f      Total     total 2284 137871 2284 137871.0000
#> 12: female age_group1     total  954  58008  954  58008.0000
#> 13:      f age_group1     total  954  58008  954  58008.0000
#> 14: female       ag1a     total  954  58008  954  58008.0000
#> 15:      f       ag1a     total  954  58008  954  58008.0000
#> 16: female age_group2     total  572  33641  573  33699.8129
#> 17:      f age_group2     total  572  33641  573  33699.8129
#> 18: female       ag2a     total  572  33641  573  33699.8129
#> 19:      f       ag2a     total  572  33641  573  33699.8129
#> 20: female age_group3     total  440  26817  441  26877.9477
#> 21:      f age_group3     total  440  26817  441  26877.9477
#> 22: female age_group4     total  228  13877  226  13755.2719
#> 23:      f age_group4     total  228  13877  226  13755.2719
#> 24: female age_group5     total   84   5138   85   5199.1667
#> 25:      f age_group5     total   84   5138   85   5199.1667
#> 26: female age_group6     total    6    390    5    325.0000
#> 27:      f age_group6     total    6    390    5    325.0000
#> 28:   male      Total     total 2296 140976 2295 140914.5993
#> 29:      m      Total     total 2296 140976 2295 140914.5993
#> 30:   male age_group1     total 1015  62017 1015  62017.0000
#> 31:      m age_group1     total 1015  62017 1015  62017.0000
#> 32:   male       ag1a     total 1015  62017 1015  62017.0000
#> 33:      m       ag1a     total 1015  62017 1015  62017.0000
#> 34:   male age_group2     total  571  35226  572  35287.6918
#> 35:      m age_group2     total  571  35226  572  35287.6918
#> 36:   male       ag2a     total  571  35226  572  35287.6918
#> 37:      m       ag2a     total  571  35226  572  35287.6918
#> 38:   male age_group3     total  424  25741  422  25619.5802
#> 39:      m age_group3     total  424  25741  422  25619.5802
#> 40:   male age_group4     total  195  12347  194  12283.6821
#> 41:      m age_group4     total  195  12347  194  12283.6821
#> 42:   male age_group5     total   84   5215   82   5090.8333
#> 43:      m age_group5     total   84   5215   82   5090.8333
#> 44:   male age_group6     total    7    430    8    491.4286
#> 45:      m age_group6     total    7    430    8    491.4286
#> 46:  Total      Total cnt_males 2296 140976 2295 140914.5993
#> 47:  Total age_group1 cnt_males 1015  62017 1015  62017.0000
#> 48:  Total       ag1a cnt_males 1015  62017 1015  62017.0000
#> 49:  Total age_group2 cnt_males  571  35226  572  35287.6918
#> 50:  Total       ag2a cnt_males  571  35226  572  35287.6918
#> 51:  Total age_group3 cnt_males  424  25741  422  25619.5802
#> 52:  Total age_group4 cnt_males  195  12347  194  12283.6821
#> 53:  Total age_group5 cnt_males   84   5215   82   5090.8333
#> 54:  Total age_group6 cnt_males    7    430    8    491.4286
#> 55: female      Total cnt_males    0      0    0      0.0000
#> 56:      f      Total cnt_males    0      0    0      0.0000
#> 57: female age_group1 cnt_males    0      0    0      0.0000
#> 58:      f age_group1 cnt_males    0      0    0      0.0000
#> 59: female       ag1a cnt_males    0      0    0      0.0000
#> 60:      f       ag1a cnt_males    0      0    0      0.0000
#> 61: female age_group2 cnt_males    0      0    0      0.0000
#> 62:      f age_group2 cnt_males    0      0    0      0.0000
#> 63: female       ag2a cnt_males    0      0    0      0.0000
#> 64:      f       ag2a cnt_males    0      0    0      0.0000
#> 65: female age_group3 cnt_males    0      0    0      0.0000
#> 66:      f age_group3 cnt_males    0      0    0      0.0000
#> 67: female age_group4 cnt_males    0      0    0      0.0000
#> 68:      f age_group4 cnt_males    0      0    0      0.0000
#> 69: female age_group5 cnt_males    0      0    0      0.0000
#> 70:      f age_group5 cnt_males    0      0    0      0.0000
#> 71: female age_group6 cnt_males    0      0    0      0.0000
#> 72:      f age_group6 cnt_males    0      0    0      0.0000
#> 73:   male      Total cnt_males 2296 140976 2295 140914.5993
#> 74:      m      Total cnt_males 2296 140976 2295 140914.5993
#> 75:   male age_group1 cnt_males 1015  62017 1015  62017.0000
#> 76:      m age_group1 cnt_males 1015  62017 1015  62017.0000
#> 77:   male       ag1a cnt_males 1015  62017 1015  62017.0000
#> 78:      m       ag1a cnt_males 1015  62017 1015  62017.0000
#> 79:   male age_group2 cnt_males  571  35226  572  35287.6918
#> 80:      m age_group2 cnt_males  571  35226  572  35287.6918
#> 81:   male       ag2a cnt_males  571  35226  572  35287.6918
#> 82:      m       ag2a cnt_males  571  35226  572  35287.6918
#> 83:   male age_group3 cnt_males  424  25741  422  25619.5802
#> 84:      m age_group3 cnt_males  424  25741  422  25619.5802
#> 85:   male age_group4 cnt_males  195  12347  194  12283.6821
#> 86:      m age_group4 cnt_males  195  12347  194  12283.6821
#> 87:   male age_group5 cnt_males   84   5215   82   5090.8333
#> 88:      m age_group5 cnt_males   84   5215   82   5090.8333
#> 89:   male age_group6 cnt_males    7    430    8    491.4286
#> 90:      m age_group6 cnt_males    7    430    8    491.4286
#>        sex        age     vname  uwc     wc puwc         pwc

# utility measures for a count variable
tab$measures_cnts(v = "total", exclude_zeros = TRUE)

#> $overview
#>    noise cnt        pct
#> 1:    -2   1 0.02222222
#> 2:    -1  15 0.33333333
#> 3:     0  14 0.31111111
#> 4:     1   8 0.17777778
#> 5:     2   7 0.15555556
#> 
#> $measures
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.200 0.000 0.002
#>  5:    Q40 1.000 0.001 0.015
#>  6:   Mean 0.867 0.020 0.046
#>  7: Median 1.000 0.002 0.021
#>  8:    Q60 1.000 0.002 0.022
#>  9:    Q70 1.000 0.005 0.046
#> 10:    Q80 1.000 0.012 0.066
#> 11:    Q90 2.000 0.095 0.154
#> 12:    Q95 2.000 0.152 0.207
#> 13:    Q99 2.000 0.167 0.256
#> 14:    Max 2.000 0.167 0.289
#> 
#> $cumdistr_d1
#>    cat cnt       pct
#> 1:   0  14 0.3111111
#> 2:   1  37 0.8222222
#> 3:   2  45 1.0000000
#> 
#> $cumdistr_d2
#>            cat cnt       pct
#> 1:    [0,0.02]  38 0.8444444
#> 2: (0.02,0.05]  40 0.8888889
#> 3:  (0.05,0.1]  40 0.8888889
#> 4:   (0.1,0.2]  45 1.0000000
#> 5:   (0.2,0.3]  45 1.0000000
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $cumdistr_d3
#>            cat cnt       pct
#> 1:    [0,0.02]  19 0.4222222
#> 2: (0.02,0.05]  33 0.7333333
#> 3:  (0.05,0.1]  38 0.8444444
#> 4:   (0.1,0.2]  42 0.9333333
#> 5:   (0.2,0.3]  45 1.0000000
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $false_zero
#> [1] 0
#> 
#> $false_nonzero
#> [1] 0
#> 
#> $exclude_zeros
#> [1] TRUE
#> 

# modifications for perturbed count variables
tab$mod_cnts()

#>        sex        age row_nr pert       ckey  countvar
#>   1: Total      Total     15    0 0.57648978     total
#>   2: Total age_group1     14   -1 0.11635891     total
#>   3: Total       ag1a     14   -1 0.11635891     total
#>   4: Total age_group2     15    0 0.41702907     total
#>   5: Total       ag2a     15    0 0.41702907     total
#>  ---                                                  
#> 131:     m age_group4     14   -1 0.30721556 cnt_males
#> 132:  male age_group5     13   -2 0.05398588 cnt_males
#> 133:     m age_group5     13   -2 0.05398588 cnt_males
#> 134:  male age_group6     16    1 0.70984926 cnt_males
#> 135:     m age_group6     16    1 0.70984926 cnt_males

# display a summary about utility measures
tab$summary()

#> ┌──────────────────────────────────────────────┐
#> │Utility measures for perturbed count variables│
#> └──────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#>          countvar Min Q10 Q20 Q30 Q40   Mean Median Q60 Q70 Q80 Q90 Q95  Q99
#> 1:          total  -2  -2  -1  -1   0 -0.111      0   0   1 1.0   1   1 1.56
#> 2: cnt_highincome  -5  -3  -2  -2   0 -0.511      0   0   1 1.0   2   2 2.00
#> 3:      cnt_males  -2  -2  -1   0   0 -0.200      0   0   0 0.2   1   1 1.00
#>    Max
#> 1:   2
#> 2:   2
#> 3:   1
#> 
#> ── Distance-based measures ─────────────────────────────────────────────────────
#> ✔ Variable: 'total'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.200 0.000 0.002
#>  5:    Q40 1.000 0.001 0.015
#>  6:   Mean 0.867 0.020 0.046
#>  7: Median 1.000 0.002 0.021
#>  8:    Q60 1.000 0.002 0.022
#>  9:    Q70 1.000 0.005 0.046
#> 10:    Q80 1.000 0.012 0.066
#> 11:    Q90 2.000 0.095 0.154
#> 12:    Q95 2.000 0.152 0.207
#> 13:    Q99 2.000 0.167 0.256
#> 14:    Max 2.000 0.167 0.289
#> 
#> ✔ Variable: 'cnt_males'
#> 
#>       what  d1    d2    d3
#>  1:    Min 0.0 0.000 0.000
#>  2:    Q10 0.0 0.000 0.000
#>  3:    Q20 0.2 0.000 0.002
#>  4:    Q30 1.0 0.000 0.010
#>  5:    Q40 1.0 0.002 0.021
#>  6:   Mean 1.0 0.020 0.048
#>  7: Median 1.0 0.002 0.021
#>  8:    Q60 1.0 0.005 0.036
#>  9:    Q70 1.0 0.005 0.049
#> 10:    Q80 1.8 0.020 0.098
#> 11:    Q90 2.0 0.071 0.139
#> 12:    Q95 2.0 0.143 0.183
#> 13:    Q99 2.0 0.143 0.183
#> 14:    Max 2.0 0.143 0.183
#> 
#> ✔ Variable: 'cnt_highincome'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.800 0.002 0.019
#>  4:    Q30 1.000 0.009 0.063
#>  5:    Q40 1.000 0.018 0.075
#>  6:   Mean 1.775 0.065 0.150
#>  7: Median 2.000 0.023 0.105
#>  8:    Q60 2.000 0.035 0.127
#>  9:    Q70 2.000 0.049 0.134
#> 10:    Q80 2.200 0.067 0.251
#> 11:    Q90 3.200 0.171 0.364
#> 12:    Q95 5.000 0.293 0.421
#> 13:    Q99 5.000 0.429 0.646
#> 14:    Max 5.000 0.429 0.646
#> 
#> ┌──────────────────────────────────────────────────┐
#> │Utility measures for perturbed numerical variables│
#> └──────────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────

#> Warning: no non-missing arguments to min; returning Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to min; returning Inf

#> Warning: no non-missing arguments to max; returning -Inf

#>      vname        Min        Q10        Q20       Q30      Q40      Mean
#> 1:  expend        Inf         NA         NA        NA       NA       NaN
#> 2:  income -50173.049 -36626.160 -28219.926 -6513.629 2037.675 19035.290
#> 3: savings  -9212.091  -5032.945  -2722.789 -1686.492  953.365  1401.554
#> 4:   mixed        Inf         NA         NA        NA       NA       NaN
#>      Median       Q60       Q70       Q80       Q90       Q95       Q99
#> 1:       NA        NA        NA        NA        NA        NA        NA
#> 2: 4917.139 23464.955 48479.140 55822.295 80407.704 80407.704 159664.96
#> 3: 1881.930  2926.881  4360.827  5635.571  6843.179  6843.179  10372.77
#> 4:       NA        NA        NA        NA        NA        NA        NA
#>          Max
#> 1:      -Inf
#> 2: 159664.96
#> 3:  10372.77
#> 4:      -Inf