Set parameters required to perturb numeric variables using a flex function

ck_flexparams() allows to define a flex function that is used to lookup perturbation magnitudes (percentages) used when perturbing continuous variables.

ck_flexparams(fp, p = c(0.25, 0.05), epsilon = 1, q = 3)

Arguments

fp

(numeric scalar); at which point should the noise coefficient function reaches its desired maximum (defined by the first element of p)

p

a numeric vector of length 2 where both elements specify a percentage. The first value refers to the desired maximum perturbation percentage for small cells (depending on fp) while the second element refers to the desired maximum perturbation percentage for large cells. Both values must be between 0 and 1 and need to be in descending order.

epsilon

a numeric vector in descending order with all values >= 0 and <= 1 with the first element forced to equal 1. The length of this vector must correspond with the number top_k specified in ck_params_nums() when creating parameters for type == "top_contr" which is checked at runtime. This setting allows to use different flex-functions for the largest top_k contributors.

q

(numeric scalar); Parameter of the function; q needs to be >= 1

Value

an object suitable as input for ck_params_nums().

Details

details about the flex function can be found in Deliverable D4.2, Part I in SGA "Open Source tools for perturbative confidentiality methods"

See also

Examples

# \donttest{
x <- ck_create_testdata()

# create some 0/1 variables that should be perturbed later
x[, cnt_females := ifelse(sex == "male", 0, 1)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              20          25.00000           0
#>    2:   2530      28       1              90          25.00000           1
#>    3:   6920     550       1              87          25.00000           0
#>    4:   7960     870       1              63          25.00000           0
#>    5:   9030      20       2              82          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              51          16.66667           1
#> 4577:   1420     987    1000              60          16.66667           0
#> 4578:   8900     684    1000              92          16.66667           0
#> 4579:   3880     294    1000              64          16.66667           1
#> 4580:   4830     911    1000             100          16.66667           0
x[, cnt_males := ifelse(sex == "male", 1, 0)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              20          25.00000           0
#>    2:   2530      28       1              90          25.00000           1
#>    3:   6920     550       1              87          25.00000           0
#>    4:   7960     870       1              63          25.00000           0
#>    5:   9030      20       2              82          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              51          16.66667           1
#> 4577:   1420     987    1000              60          16.66667           0
#> 4578:   8900     684    1000              92          16.66667           0
#> 4579:   3880     294    1000              64          16.66667           1
#> 4580:   4830     911    1000             100          16.66667           0
#>       cnt_males
#>    1:         1
#>    2:         0
#>    3:         1
#>    4:         1
#>    5:         1
#>   ---          
#> 4576:         0
#> 4577:         1
#> 4578:         1
#> 4579:         0
#> 4580:         1
x[, cnt_highincome := ifelse(income >= 9000, 1, 0)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              20          25.00000           0
#>    2:   2530      28       1              90          25.00000           1
#>    3:   6920     550       1              87          25.00000           0
#>    4:   7960     870       1              63          25.00000           0
#>    5:   9030      20       2              82          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              51          16.66667           1
#> 4577:   1420     987    1000              60          16.66667           0
#> 4578:   8900     684    1000              92          16.66667           0
#> 4579:   3880     294    1000              64          16.66667           1
#> 4580:   4830     911    1000             100          16.66667           0
#>       cnt_males cnt_highincome
#>    1:         1              0
#>    2:         0              0
#>    3:         1              0
#>    4:         1              0
#>    5:         1              1
#>   ---                         
#> 4576:         0              0
#> 4577:         1              0
#> 4578:         1              0
#> 4579:         0              0
#> 4580:         1              0
# a variable with positive and negative contributions
x[, mixed := sample(-10:10, nrow(x), replace = TRUE)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              20          25.00000           0
#>    2:   2530      28       1              90          25.00000           1
#>    3:   6920     550       1              87          25.00000           0
#>    4:   7960     870       1              63          25.00000           0
#>    5:   9030      20       2              82          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              51          16.66667           1
#> 4577:   1420     987    1000              60          16.66667           0
#> 4578:   8900     684    1000              92          16.66667           0
#> 4579:   3880     294    1000              64          16.66667           1
#> 4580:   4830     911    1000             100          16.66667           0
#>       cnt_males cnt_highincome mixed
#>    1:         1              0     5
#>    2:         0              0     5
#>    3:         1              0    -3
#>    4:         1              0    -2
#>    5:         1              1    10
#>   ---                               
#> 4576:         0              0   -10
#> 4577:         1              0     1
#> 4578:         1              0    -4
#> 4579:         0              0     5
#> 4580:         1              0     2

# create record keys
x$rkey <- ck_generate_rkeys(dat = x)

# define required inputs

# hierarchy with some bogus codes
d_sex <- hier_create(root = "Total", nodes = c("male", "female"))
d_sex <- hier_add(d_sex, root = "female", "f")
d_sex <- hier_add(d_sex, root = "male", "m")

d_age <- hier_create(root = "Total", nodes = paste0("age_group", 1:6))
d_age <- hier_add(d_age, root = "age_group1", "ag1a")
d_age <- hier_add(d_age, root = "age_group2", "ag2a")

# define the cell key object
countvars <- c("cnt_females", "cnt_males", "cnt_highincome")
numvars <- c("expend", "income", "savings", "mixed")
tab <- ck_setup(
  x = x,
  rkey = "rkey",
  dims = list(sex = d_sex, age = d_age),
  w = "sampling_weight",
  countvars = countvars,
  numvars = numvars)

# show some information about this table instance
tab$print() # identical with print(tab)
#> ── Table Information ───────────────────────────────────────────────────────────
#> ✔ 45 cells in 2 dimensions ('sex', 'age')
#> ✔ weights: yes
#> ── Tabulated / Perturbed countvars ─────────────────────────────────────────────
#> ☐ 'total'
#> ☐ 'cnt_females'
#> ☐ 'cnt_males'
#> ☐ 'cnt_highincome'
#> ── Tabulated / Perturbed numvars ───────────────────────────────────────────────
#> ☐ 'expend'
#> ☐ 'income'
#> ☐ 'savings'
#> ☐ 'mixed'

# information about the hierarchies
tab$hierarchy_info()
#> $sex
#>      code level is_leaf parent
#> 1:  Total     1   FALSE  Total
#> 2:   male     2   FALSE  Total
#> 3:      m     3    TRUE   male
#> 4: female     2   FALSE  Total
#> 5:      f     3    TRUE female
#> 
#> $age
#>          code level is_leaf     parent
#> 1:      Total     1   FALSE      Total
#> 2: age_group1     2   FALSE      Total
#> 3:       ag1a     3    TRUE age_group1
#> 4: age_group2     2   FALSE      Total
#> 5:       ag2a     3    TRUE age_group2
#> 6: age_group3     2    TRUE      Total
#> 7: age_group4     2    TRUE      Total
#> 8: age_group5     2    TRUE      Total
#> 9: age_group6     2    TRUE      Total
#> 

# which variables have been defined?
tab$allvars()
#> $cntvars
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"
#> 
#> $numvars
#> [1] "expend"  "income"  "savings" "mixed"  
#> 

# count variables
tab$cntvars()
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"

# continuous variables
tab$numvars()
#> [1] "expend"  "income"  "savings" "mixed"  

# create perturbation parameters for "total" variable and
# write to yaml-file

# create a ptable using functionality from the ptable-pkg
f_yaml <- tempfile(fileext = ".yaml")
p_cnts1 <- ck_params_cnts(
  ptab = ptable::pt_ex_cnts(),
  path = f_yaml)
#> yaml configuration '/tmp/RtmpCEVK6P/file1f0aae45a12.yaml' successfully written.

# read parameters from yaml-file and set them for variable `"total"`
p_cnts1 <- ck_read_yaml(path = f_yaml)

tab$params_cnts_set(val = p_cnts1, v = "total")
#> --> setting perturbation parameters for variable 'total'

# create alternative perturbation parameters by specifying parameters
para2 <- ptable::create_cnt_ptable(
  D = 8, V = 3, js = 2, create = FALSE)

p_cnts2 <- ck_params_cnts(ptab = para2)

# use these ptable it for the remaining variables
tab$params_cnts_set(val = p_cnts2, v = countvars)
#> --> setting perturbation parameters for variable 'cnt_females'
#> --> setting perturbation parameters for variable 'cnt_males'
#> --> setting perturbation parameters for variable 'cnt_highincome'

# perturb a variable
tab$perturb(v = "total")
#> Count variable 'total' was perturbed.

# multiple variables can be perturbed as well
tab$perturb(v = c("cnt_males", "cnt_highincome"))
#> Count variable 'cnt_males' was perturbed.
#> Count variable 'cnt_highincome' was perturbed.

# return weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))
#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 273986 4582 274105.6445
#>  2:  Total age_group1     total 1969 116053 1969 116053.0000
#>  3:  Total       ag1a     total 1969 116053 1969 116053.0000
#>  4:  Total age_group2     total 1143  69223 1144  69283.5626
#>  5:  Total       ag2a     total 1143  69223 1144  69283.5626
#>  6:  Total age_group3     total  864  53012  865  53073.3565
#>  7:  Total age_group4     total  423  25231  423  25231.0000
#>  8:  Total age_group5     total  168   9872  168   9872.0000
#>  9:  Total age_group6     total   13    595   14    640.7692
#> 10:   male      Total     total 2296 136535 2298 136653.9329
#> 11:      m      Total     total 2296 136535 2298 136653.9329
#> 12:   male age_group1     total 1015  59889 1014  59829.9961
#> 13:      m age_group1     total 1015  59889 1014  59829.9961
#> 14:   male       ag1a     total 1015  59889 1014  59829.9961
#> 15:      m       ag1a     total 1015  59889 1014  59829.9961
#> 16:   male age_group2     total  571  34012  571  34012.0000
#> 17:      m age_group2     total  571  34012  571  34012.0000
#> 18:   male       ag2a     total  571  34012  571  34012.0000
#> 19:      m       ag2a     total  571  34012  571  34012.0000
#> 20:   male age_group3     total  424  26150  423  26088.3255
#> 21:      m age_group3     total  424  26150  423  26088.3255
#> 22:   male age_group4     total  195  11314  195  11314.0000
#> 23:      m age_group4     total  195  11314  195  11314.0000
#> 24:   male age_group5     total   84   4906   84   4906.0000
#> 25:      m age_group5     total   84   4906   84   4906.0000
#> 26:   male age_group6     total    7    264    8    301.7143
#> 27:      m age_group6     total    7    264    8    301.7143
#> 28: female      Total     total 2284 137451 2282 137330.6401
#> 29:      f      Total     total 2284 137451 2282 137330.6401
#> 30: female age_group1     total  954  56164  953  56105.1279
#> 31:      f age_group1     total  954  56164  953  56105.1279
#> 32: female       ag1a     total  954  56164  953  56105.1279
#> 33:      f       ag1a     total  954  56164  953  56105.1279
#> 34: female age_group2     total  572  35211  572  35211.0000
#> 35:      f age_group2     total  572  35211  572  35211.0000
#> 36: female       ag2a     total  572  35211  572  35211.0000
#> 37:      f       ag2a     total  572  35211  572  35211.0000
#> 38: female age_group3     total  440  26862  440  26862.0000
#> 39:      f age_group3     total  440  26862  440  26862.0000
#> 40: female age_group4     total  228  13917  230  14039.0789
#> 41:      f age_group4     total  228  13917  230  14039.0789
#> 42: female age_group5     total   84   4966   85   5025.1190
#> 43:      f age_group5     total   84   4966   85   5025.1190
#> 44: female age_group6     total    6    331    7    386.1667
#> 45:      f age_group6     total    6    331    7    386.1667
#> 46:  Total      Total cnt_males 2296 136535 2299 136713.3994
#> 47:  Total age_group1 cnt_males 1015  59889 1014  59829.9961
#> 48:  Total       ag1a cnt_males 1015  59889 1014  59829.9961
#> 49:  Total age_group2 cnt_males  571  34012  571  34012.0000
#> 50:  Total       ag2a cnt_males  571  34012  571  34012.0000
#> 51:  Total age_group3 cnt_males  424  26150  423  26088.3255
#> 52:  Total age_group4 cnt_males  195  11314  195  11314.0000
#> 53:  Total age_group5 cnt_males   84   4906   84   4906.0000
#> 54:  Total age_group6 cnt_males    7    264    9    339.4286
#> 55:   male      Total cnt_males 2296 136535 2299 136713.3994
#> 56:      m      Total cnt_males 2296 136535 2299 136713.3994
#> 57:   male age_group1 cnt_males 1015  59889 1014  59829.9961
#> 58:      m age_group1 cnt_males 1015  59889 1014  59829.9961
#> 59:   male       ag1a cnt_males 1015  59889 1014  59829.9961
#> 60:      m       ag1a cnt_males 1015  59889 1014  59829.9961
#> 61:   male age_group2 cnt_males  571  34012  571  34012.0000
#> 62:      m age_group2 cnt_males  571  34012  571  34012.0000
#> 63:   male       ag2a cnt_males  571  34012  571  34012.0000
#> 64:      m       ag2a cnt_males  571  34012  571  34012.0000
#> 65:   male age_group3 cnt_males  424  26150  423  26088.3255
#> 66:      m age_group3 cnt_males  424  26150  423  26088.3255
#> 67:   male age_group4 cnt_males  195  11314  195  11314.0000
#> 68:      m age_group4 cnt_males  195  11314  195  11314.0000
#> 69:   male age_group5 cnt_males   84   4906   84   4906.0000
#> 70:      m age_group5 cnt_males   84   4906   84   4906.0000
#> 71:   male age_group6 cnt_males    7    264    9    339.4286
#> 72:      m age_group6 cnt_males    7    264    9    339.4286
#> 73: female      Total cnt_males    0      0    0      0.0000
#> 74:      f      Total cnt_males    0      0    0      0.0000
#> 75: female age_group1 cnt_males    0      0    0      0.0000
#> 76:      f age_group1 cnt_males    0      0    0      0.0000
#> 77: female       ag1a cnt_males    0      0    0      0.0000
#> 78:      f       ag1a cnt_males    0      0    0      0.0000
#> 79: female age_group2 cnt_males    0      0    0      0.0000
#> 80:      f age_group2 cnt_males    0      0    0      0.0000
#> 81: female       ag2a cnt_males    0      0    0      0.0000
#> 82:      f       ag2a cnt_males    0      0    0      0.0000
#> 83: female age_group3 cnt_males    0      0    0      0.0000
#> 84:      f age_group3 cnt_males    0      0    0      0.0000
#> 85: female age_group4 cnt_males    0      0    0      0.0000
#> 86:      f age_group4 cnt_males    0      0    0      0.0000
#> 87: female age_group5 cnt_males    0      0    0      0.0000
#> 88:      f age_group5 cnt_males    0      0    0      0.0000
#> 89: female age_group6 cnt_males    0      0    0      0.0000
#> 90:      f age_group6 cnt_males    0      0    0      0.0000
#>        sex        age     vname  uwc     wc puwc         pwc

# numerical variables (positive variables using flex-function)
# we also write the config to a yaml file
f_yaml <- tempfile(fileext = ".yaml")

# create a ptable using functionality from the ptable-pkg
# a single ptable for all cells
ptab1 <- ptable::pt_ex_nums(parity = TRUE, separation = FALSE)

# a single ptab for all cells except for very small ones
ptab2 <- ptable::pt_ex_nums(parity = TRUE, separation = TRUE)

# different ptables for cells with even/odd number of contributors
# and very small cells
ptab3 <- ptable::pt_ex_nums(parity = FALSE, separation = TRUE)

p_nums1 <- ck_params_nums(
  ptab = ptab1,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.30, 0.03),
    epsilon = c(1, 0.5, 0.2),
    q = 3),
  mu_c = 2,
  same_key = FALSE,
  use_zero_rkeys = FALSE,
  path = f_yaml)
#> yaml configuration '/tmp/RtmpCEVK6P/file1f0a4b0c477c.yaml' successfully written.

# we read the parameters from the yaml-file
p_nums1 <- ck_read_yaml(path = f_yaml)

# for variables with positive and negative values
p_nums2 <- ck_params_nums(
  ptab = ptab2,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.15, 0.02),
    epsilon = c(1, 0.4, 0.15),
    q = 3),
  mu_c = 2,
  same_key = FALSE)

# simple perturbation parameters (not using the flex-function approach)
p_nums3 <- ck_params_nums(
  ptab = ptab3,
  type = "mean",
  mult_params = ck_simpleparams(p = 0.25),
  mu_c = 2,
  same_key = FALSE)

# use `p_nums1` for all variables
tab$params_nums_set(p_nums1, c("savings", "income", "expend"))
#> --> setting perturbation parameters for variable 'savings'
#> --> setting perturbation parameters for variable 'income'
#> --> setting perturbation parameters for variable 'expend'

# use different parameters for variable `mixed`
tab$params_nums_set(p_nums2, v = "mixed")
#> --> setting perturbation parameters for variable 'mixed'

# identify sensitive cells to which extra protection (`mu_c`) is added.
tab$supp_p(v = "income", p = 85)
#> computing contributing indices | rawdata <--> table; this might take a while
#> p%-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_pq(v = "income", p = 85, q = 90)
#> computing contributing indices | rawdata <--> table; this might take a while
#> pq-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_nk(v = "income", n = 2, k = 90)
#> computing contributing indices | rawdata <--> table; this might take a while
#> nk-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_freq(v = "income", n = 14, weighted = FALSE)
#> freq-rule: 5 new sensitive cells (incl. duplicates) found (total: 5)
tab$supp_val(v = "income", n = 10000, weighted = TRUE)
#> val-rule: 0 new sensitive cells (incl. duplicates) found (total: 5)
tab$supp_cells(
  v = "income",
  inp = data.frame(
    sex = c("female", "female"),
    "age" = c("age_group1", "age_group3")
  )
)
#> cell-rule: 2 new sensitive cells (incl. duplicates) found (total: 7)

# perturb variables
tab$perturb(v = c("income", "savings"))
#> Numeric variable 'income' was perturbed.
#> Numeric variable 'savings' was perturbed.

# extract results
tab$numtab("income", mean_before_sum = TRUE)
#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1375035831 1374930222
#>  2:  Total age_group1 income  9810547  576497557  576445805
#>  3:  Total       ag1a income  9810547  576497557  576445805
#>  4:  Total age_group2 income  5692119  346480414  346529259
#>  5:  Total       ag2a income  5692119  346480414  346529259
#>  6:  Total age_group3 income  4406946  270752453  270878272
#>  7:  Total age_group4 income  2133543  128397873  128403746
#>  8:  Total age_group5 income   848151   49937475   50062800
#>  9:  Total age_group6 income    61672    2970059    3088265
#> 10:   male      Total income 11262049  669022896  669001590
#> 11:      m      Total income 11262049  669022896  669001590
#> 12:   male age_group1 income  4877164  287410800  287467549
#> 13:      m age_group1 income  4877164  287410800  287467549
#> 14:   male       ag1a income  4877164  287410800  287467549
#> 15:      m       ag1a income  4877164  287410800  287467549
#> 16:   male age_group2 income  2811379  166599566  166760973
#> 17:      m age_group2 income  2811379  166599566  166760973
#> 18:   male       ag2a income  2811379  166599566  166760973
#> 19:      m       ag2a income  2811379  166599566  166760973
#> 20:   male age_group3 income  2168169  133872961  133989927
#> 21:      m age_group3 income  2168169  133872961  133989927
#> 22:   male age_group4 income   978510   56892175   56938642
#> 23:      m age_group4 income   978510   56892175   56938642
#> 24:   male age_group5 income   393134   23042325   23058778
#> 25:      m age_group5 income   393134   23042325   23058778
#> 26:   male age_group6 income    33693    1205069    1151989
#> 27:      m age_group6 income    33693    1205069    1151989
#> 28: female      Total income 11690929  706012935  705919423
#> 29:      f      Total income 11690929  706012935  705919423
#> 30: female age_group1 income  4933383  289086757  289208125
#> 31:      f age_group1 income  4933383  289086757  289208125
#> 32: female       ag1a income  4933383  289086757  289208125
#> 33:      f       ag1a income  4933383  289086757  289208125
#> 34: female age_group2 income  2880740  179880848  179853808
#> 35:      f age_group2 income  2880740  179880848  179853808
#> 36: female       ag2a income  2880740  179880848  179853808
#> 37:      f       ag2a income  2880740  179880848  179853808
#> 38: female age_group3 income  2238777  136879492  136689332
#> 39:      f age_group3 income  2238777  136879492  136689332
#> 40: female age_group4 income  1155033   71505698   71508386
#> 41:      f age_group4 income  1155033   71505698   71508386
#> 42: female age_group5 income   455017   26895150   26770910
#> 43:      f age_group5 income   455017   26895150   26770910
#> 44: female age_group6 income    27979    1764990    1855487
#> 45:      f age_group6 income    27979    1764990    1855487
#>        sex        age  vname      uws         ws        pws
tab$numtab("income", mean_before_sum = FALSE)
#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1375035831 1374983026
#>  2:  Total age_group1 income  9810547  576497557  576471680
#>  3:  Total       ag1a income  9810547  576497557  576471680
#>  4:  Total age_group2 income  5692119  346480414  346504836
#>  5:  Total       ag2a income  5692119  346480414  346504836
#>  6:  Total age_group3 income  4406946  270752453  270815355
#>  7:  Total age_group4 income  2133543  128397873  128400810
#>  8:  Total age_group5 income   848151   49937475   50000098
#>  9:  Total age_group6 income    61672    2970059    3028585
#> 10:   male      Total income 11262049  669022896  669012243
#> 11:      m      Total income 11262049  669022896  669012243
#> 12:   male age_group1 income  4877164  287410800  287439173
#> 13:      m age_group1 income  4877164  287410800  287439173
#> 14:   male       ag1a income  4877164  287410800  287439173
#> 15:      m       ag1a income  4877164  287410800  287439173
#> 16:   male age_group2 income  2811379  166599566  166680250
#> 17:      m age_group2 income  2811379  166599566  166680250
#> 18:   male       ag2a income  2811379  166599566  166680250
#> 19:      m       ag2a income  2811379  166599566  166680250
#> 20:   male age_group3 income  2168169  133872961  133931431
#> 21:      m age_group3 income  2168169  133872961  133931431
#> 22:   male age_group4 income   978510   56892175   56915404
#> 23:      m age_group4 income   978510   56892175   56915404
#> 24:   male age_group5 income   393134   23042325   23050550
#> 25:      m age_group5 income   393134   23042325   23050550
#> 26:   male age_group6 income    33693    1205069    1178230
#> 27:      m age_group6 income    33693    1205069    1178230
#> 28: female      Total income 11690929  706012935  705966178
#> 29:      f      Total income 11690929  706012935  705966178
#> 30: female age_group1 income  4933383  289086757  289147435
#> 31:      f age_group1 income  4933383  289086757  289147435
#> 32: female       ag1a income  4933383  289086757  289147435
#> 33:      f       ag1a income  4933383  289086757  289147435
#> 34: female age_group2 income  2880740  179880848  179867328
#> 35:      f age_group2 income  2880740  179880848  179867328
#> 36: female       ag2a income  2880740  179880848  179867328
#> 37:      f       ag2a income  2880740  179880848  179867328
#> 38: female age_group3 income  2238777  136879492  136784379
#> 39:      f age_group3 income  2238777  136879492  136784379
#> 40: female age_group4 income  1155033   71505698   71507042
#> 41:      f age_group4 income  1155033   71505698   71507042
#> 42: female age_group5 income   455017   26895150   26832958
#> 43:      f age_group5 income   455017   26895150   26832958
#> 44: female age_group6 income    27979    1764990    1809673
#> 45:      f age_group6 income    27979    1764990    1809673
#>        sex        age  vname      uws         ws        pws
tab$numtab("savings")
#>        sex        age   vname     uws        ws         pws
#>  1:  Total      Total savings 2273532 136284177 136284891.3
#>  2:  Total age_group1 savings  982386  58040842  58041868.8
#>  3:  Total       ag1a savings  982386  58040842  58041868.8
#>  4:  Total age_group2 savings  552336  33657819  33660469.6
#>  5:  Total       ag2a savings  552336  33657819  33660469.6
#>  6:  Total age_group3 savings  437101  26804687  26798572.2
#>  7:  Total age_group4 savings  214661  12759773  12760711.7
#>  8:  Total age_group5 savings   80451   4693281   4705398.7
#>  9:  Total age_group6 savings    6597    327775    332053.9
#> 10:   male      Total savings 1159816  69269600  69269600.0
#> 11:      m      Total savings 1159816  69269600  69269600.0
#> 12:   male age_group1 savings  517660  30664347  30667523.3
#> 13:      m age_group1 savings  517660  30664347  30667523.3
#> 14:   male       ag1a savings  517660  30664347  30667523.3
#> 15:      m       ag1a savings  517660  30664347  30667523.3
#> 16:   male age_group2 savings  280923  16988776  16995327.7
#> 17:      m age_group2 savings  280923  16988776  16995327.7
#> 18:   male       ag2a savings  280923  16988776  16995327.7
#> 19:      m       ag2a savings  280923  16988776  16995327.7
#> 20:   male age_group3 savings  214970  13113850  13106755.4
#> 21:      m age_group3 savings  214970  13113850  13106755.4
#> 22:   male age_group4 savings   99420   5849364   5852227.0
#> 23:      m age_group4 savings   99420   5849364   5852227.0
#> 24:   male age_group5 savings   43233   2508684   2511568.6
#> 25:      m age_group5 savings   43233   2508684   2511568.6
#> 26:   male age_group6 savings    3610    144579    143598.5
#> 27:      m age_group6 savings    3610    144579    143598.5
#> 28: female      Total savings 1113716  67014577  67010077.7
#> 29:      f      Total savings 1113716  67014577  67010077.7
#> 30: female age_group1 savings  464726  27376495  27376126.6
#> 31:      f age_group1 savings  464726  27376495  27376126.6
#> 32: female       ag1a savings  464726  27376495  27376126.6
#> 33:      f       ag1a savings  464726  27376495  27376126.6
#> 34: female age_group2 savings  271413  16669043  16666573.1
#> 35:      f age_group2 savings  271413  16669043  16666573.1
#> 36: female       ag2a savings  271413  16669043  16666573.1
#> 37:      f       ag2a savings  271413  16669043  16666573.1
#> 38: female age_group3 savings  222131  13690837  13679166.8
#> 39:      f age_group3 savings  222131  13690837  13679166.8
#> 40: female age_group4 savings  115241   6910409   6913117.6
#> 41:      f age_group4 savings  115241   6910409   6913117.6
#> 42: female age_group5 savings   37218   2184597   2180187.8
#> 43:      f age_group5 savings   37218   2184597   2180187.8
#> 44: female age_group6 savings    2987    183196    184886.8
#> 45:      f age_group6 savings    2987    183196    184886.8
#>        sex        age   vname     uws        ws         pws

# results can be resetted, too
tab$reset_cntvars(v = "cnt_males")

# we can then set other parameters and perturb again
tab$params_cnts_set(val = p_cnts1, v = "cnt_males")
#> --> setting perturbation parameters for variable 'cnt_males'

tab$perturb(v = "cnt_males")
#> Count variable 'cnt_males' was perturbed.

# write results to a .csv file
tab$freqtab(
  v = c("total", "cnt_males"),
  path = file.path(tempdir(), "outtab.csv")
)
#> File '/tmp/RtmpCEVK6P/outtab.csv' successfully written to disk.
#> NULL

# show results containing weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))
#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 273986 4582 274105.6445
#>  2:  Total age_group1     total 1969 116053 1969 116053.0000
#>  3:  Total       ag1a     total 1969 116053 1969 116053.0000
#>  4:  Total age_group2     total 1143  69223 1144  69283.5626
#>  5:  Total       ag2a     total 1143  69223 1144  69283.5626
#>  6:  Total age_group3     total  864  53012  865  53073.3565
#>  7:  Total age_group4     total  423  25231  423  25231.0000
#>  8:  Total age_group5     total  168   9872  168   9872.0000
#>  9:  Total age_group6     total   13    595   14    640.7692
#> 10:   male      Total     total 2296 136535 2298 136653.9329
#> 11:      m      Total     total 2296 136535 2298 136653.9329
#> 12:   male age_group1     total 1015  59889 1014  59829.9961
#> 13:      m age_group1     total 1015  59889 1014  59829.9961
#> 14:   male       ag1a     total 1015  59889 1014  59829.9961
#> 15:      m       ag1a     total 1015  59889 1014  59829.9961
#> 16:   male age_group2     total  571  34012  571  34012.0000
#> 17:      m age_group2     total  571  34012  571  34012.0000
#> 18:   male       ag2a     total  571  34012  571  34012.0000
#> 19:      m       ag2a     total  571  34012  571  34012.0000
#> 20:   male age_group3     total  424  26150  423  26088.3255
#> 21:      m age_group3     total  424  26150  423  26088.3255
#> 22:   male age_group4     total  195  11314  195  11314.0000
#> 23:      m age_group4     total  195  11314  195  11314.0000
#> 24:   male age_group5     total   84   4906   84   4906.0000
#> 25:      m age_group5     total   84   4906   84   4906.0000
#> 26:   male age_group6     total    7    264    8    301.7143
#> 27:      m age_group6     total    7    264    8    301.7143
#> 28: female      Total     total 2284 137451 2282 137330.6401
#> 29:      f      Total     total 2284 137451 2282 137330.6401
#> 30: female age_group1     total  954  56164  953  56105.1279
#> 31:      f age_group1     total  954  56164  953  56105.1279
#> 32: female       ag1a     total  954  56164  953  56105.1279
#> 33:      f       ag1a     total  954  56164  953  56105.1279
#> 34: female age_group2     total  572  35211  572  35211.0000
#> 35:      f age_group2     total  572  35211  572  35211.0000
#> 36: female       ag2a     total  572  35211  572  35211.0000
#> 37:      f       ag2a     total  572  35211  572  35211.0000
#> 38: female age_group3     total  440  26862  440  26862.0000
#> 39:      f age_group3     total  440  26862  440  26862.0000
#> 40: female age_group4     total  228  13917  230  14039.0789
#> 41:      f age_group4     total  228  13917  230  14039.0789
#> 42: female age_group5     total   84   4966   85   5025.1190
#> 43:      f age_group5     total   84   4966   85   5025.1190
#> 44: female age_group6     total    6    331    7    386.1667
#> 45:      f age_group6     total    6    331    7    386.1667
#> 46:  Total      Total cnt_males 2296 136535 2298 136653.9329
#> 47:  Total age_group1 cnt_males 1015  59889 1014  59829.9961
#> 48:  Total       ag1a cnt_males 1015  59889 1014  59829.9961
#> 49:  Total age_group2 cnt_males  571  34012  571  34012.0000
#> 50:  Total       ag2a cnt_males  571  34012  571  34012.0000
#> 51:  Total age_group3 cnt_males  424  26150  423  26088.3255
#> 52:  Total age_group4 cnt_males  195  11314  195  11314.0000
#> 53:  Total age_group5 cnt_males   84   4906   84   4906.0000
#> 54:  Total age_group6 cnt_males    7    264    8    301.7143
#> 55:   male      Total cnt_males 2296 136535 2298 136653.9329
#> 56:      m      Total cnt_males 2296 136535 2298 136653.9329
#> 57:   male age_group1 cnt_males 1015  59889 1014  59829.9961
#> 58:      m age_group1 cnt_males 1015  59889 1014  59829.9961
#> 59:   male       ag1a cnt_males 1015  59889 1014  59829.9961
#> 60:      m       ag1a cnt_males 1015  59889 1014  59829.9961
#> 61:   male age_group2 cnt_males  571  34012  571  34012.0000
#> 62:      m age_group2 cnt_males  571  34012  571  34012.0000
#> 63:   male       ag2a cnt_males  571  34012  571  34012.0000
#> 64:      m       ag2a cnt_males  571  34012  571  34012.0000
#> 65:   male age_group3 cnt_males  424  26150  423  26088.3255
#> 66:      m age_group3 cnt_males  424  26150  423  26088.3255
#> 67:   male age_group4 cnt_males  195  11314  195  11314.0000
#> 68:      m age_group4 cnt_males  195  11314  195  11314.0000
#> 69:   male age_group5 cnt_males   84   4906   84   4906.0000
#> 70:      m age_group5 cnt_males   84   4906   84   4906.0000
#> 71:   male age_group6 cnt_males    7    264    8    301.7143
#> 72:      m age_group6 cnt_males    7    264    8    301.7143
#> 73: female      Total cnt_males    0      0    0      0.0000
#> 74:      f      Total cnt_males    0      0    0      0.0000
#> 75: female age_group1 cnt_males    0      0    0      0.0000
#> 76:      f age_group1 cnt_males    0      0    0      0.0000
#> 77: female       ag1a cnt_males    0      0    0      0.0000
#> 78:      f       ag1a cnt_males    0      0    0      0.0000
#> 79: female age_group2 cnt_males    0      0    0      0.0000
#> 80:      f age_group2 cnt_males    0      0    0      0.0000
#> 81: female       ag2a cnt_males    0      0    0      0.0000
#> 82:      f       ag2a cnt_males    0      0    0      0.0000
#> 83: female age_group3 cnt_males    0      0    0      0.0000
#> 84:      f age_group3 cnt_males    0      0    0      0.0000
#> 85: female age_group4 cnt_males    0      0    0      0.0000
#> 86:      f age_group4 cnt_males    0      0    0      0.0000
#> 87: female age_group5 cnt_males    0      0    0      0.0000
#> 88:      f age_group5 cnt_males    0      0    0      0.0000
#> 89: female age_group6 cnt_males    0      0    0      0.0000
#> 90:      f age_group6 cnt_males    0      0    0      0.0000
#>        sex        age     vname  uwc     wc puwc         pwc

# utility measures for a count variable
tab$measures_cnts(v = "total", exclude_zeros = TRUE)
#> $overview
#>    noise cnt        pct
#> 1:    -2   5 0.11111111
#> 2:    -1  10 0.22222222
#> 3:     0  18 0.40000000
#> 4:     1  10 0.22222222
#> 5:     2   2 0.04444444
#> 
#> $measures
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 0.600 0.000 0.009
#>  6:   Mean 0.756 0.017 0.032
#>  7: Median 1.000 0.001 0.016
#>  8:    Q60 1.000 0.001 0.016
#>  9:    Q70 1.000 0.001 0.021
#> 10:    Q80 1.000 0.004 0.030
#> 11:    Q90 2.000 0.051 0.108
#> 12:    Q95 2.000 0.143 0.183
#> 13:    Q99 2.000 0.167 0.196
#> 14:    Max 2.000 0.167 0.196
#> 
#> $cumdistr_d1
#>    cat cnt       pct
#> 1:   0  18 0.4000000
#> 2:   1  38 0.8444444
#> 3:   2  45 1.0000000
#> 
#> $cumdistr_d2
#>            cat cnt       pct
#> 1:    [0,0.02]  40 0.8888889
#> 2: (0.02,0.05]  40 0.8888889
#> 3:  (0.05,0.1]  41 0.9111111
#> 4:   (0.1,0.2]  45 1.0000000
#> 5:   (0.2,0.3]  45 1.0000000
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $cumdistr_d3
#>            cat cnt       pct
#> 1:    [0,0.02]  30 0.6666667
#> 2: (0.02,0.05]  36 0.8000000
#> 3:  (0.05,0.1]  40 0.8888889
#> 4:   (0.1,0.2]  45 1.0000000
#> 5:   (0.2,0.3]  45 1.0000000
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $false_zero
#> [1] 0
#> 
#> $false_nonzero
#> [1] 0
#> 
#> $exclude_zeros
#> [1] TRUE
#> 

# modifications for perturbed count variables
tab$mod_cnts()
#>         sex        age row_nr pert      ckey  countvar
#>   1:  Total      Total     17    2 0.9876037     total
#>   2:  Total age_group1     15    0 0.5610423     total
#>   3:  Total       ag1a     15    0 0.5610423     total
#>   4:  Total age_group2     16    1 0.7590566     total
#>   5:  Total       ag2a     16    1 0.7590566     total
#>  ---                                                  
#> 131:      f age_group4     -1    0 0.0000000 cnt_males
#> 132: female age_group5     -1    0 0.0000000 cnt_males
#> 133:      f age_group5     -1    0 0.0000000 cnt_males
#> 134: female age_group6     -1    0 0.0000000 cnt_males
#> 135:      f age_group6     -1    0 0.0000000 cnt_males

# display a summary about utility measures
tab$summary()
#> ┌──────────────────────────────────────────────┐
#> │Utility measures for perturbed count variables│
#> └──────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#>          countvar Min Q10  Q20 Q30 Q40   Mean Median Q60 Q70 Q80 Q90 Q95  Q99
#> 1:          total  -2  -1 -1.0   0   0  0.133      0   0 1.0   1 1.6 2.0 2.00
#> 2: cnt_highincome  -4  -3 -2.0  -1  -1 -0.356     -1   0 0.8   2 2.0 2.0 2.56
#> 3:      cnt_males  -1  -1 -0.2   0   0  0.000      0   0 0.0   0 1.0 1.8 2.00
#>    Max
#> 1:   2
#> 2:   3
#> 3:   2
#> 
#> ── Distance-based measures ─────────────────────────────────────────────────────
#> ✔ Variable: 'total'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 0.600 0.000 0.009
#>  6:   Mean 0.756 0.017 0.032
#>  7: Median 1.000 0.001 0.016
#>  8:    Q60 1.000 0.001 0.016
#>  9:    Q70 1.000 0.001 0.021
#> 10:    Q80 1.000 0.004 0.030
#> 11:    Q90 2.000 0.051 0.108
#> 12:    Q95 2.000 0.143 0.183
#> 13:    Q99 2.000 0.167 0.196
#> 14:    Max 2.000 0.167 0.196
#> 
#> ✔ Variable: 'cnt_males'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 0.000 0.000 0.000
#>  6:   Mean 0.667 0.016 0.029
#>  7: Median 1.000 0.001 0.016
#>  8:    Q60 1.000 0.001 0.016
#>  9:    Q70 1.000 0.001 0.021
#> 10:    Q80 1.000 0.002 0.024
#> 11:    Q90 1.400 0.059 0.088
#> 12:    Q95 2.000 0.143 0.183
#> 13:    Q99 2.000 0.143 0.183
#> 14:    Max 2.000 0.143 0.183
#> 
#> ✔ Variable: 'cnt_highincome'
#> 
#>       what   d1    d2    d3
#>  1:    Min 0.00 0.000 0.000
#>  2:    Q10 1.00 0.005 0.034
#>  3:    Q20 1.00 0.010 0.062
#>  4:    Q30 1.00 0.015 0.067
#>  5:    Q40 1.00 0.018 0.072
#>  6:   Mean 1.75 0.054 0.132
#>  7: Median 2.00 0.020 0.089
#>  8:    Q60 2.00 0.032 0.114
#>  9:    Q70 2.00 0.046 0.177
#> 10:    Q80 2.00 0.084 0.213
#> 11:    Q90 3.10 0.144 0.261
#> 12:    Q95 4.00 0.164 0.354
#> 13:    Q99 4.00 0.286 0.359
#> 14:    Max 4.00 0.286 0.359
#> 
#> ┌──────────────────────────────────────────────────┐
#> │Utility measures for perturbed numerical variables│
#> └──────────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#>      vname       Min       Q10        Q20        Q30      Q40      Mean
#> 1:  expend       Inf        NA         NA         NA       NA       NaN
#> 2:  income -95113.28 -50386.21 -26069.115 -13520.320 1344.135 12109.440
#> 3: savings -11670.21  -5468.56  -2857.809   -980.548 -368.448   218.541
#> 4:   mixed       Inf        NA         NA         NA       NA       NaN
#>       Median       Q60       Q70       Q80       Q90       Q95      Q99
#> 1:        NA        NA        NA        NA        NA        NA       NA
#> 2: 23228.965 28373.089 44682.945 60677.671 62790.784 80684.110 80684.11
#> 3:   938.655  2074.738  2832.092  3176.338  5642.591  6551.686  9668.64
#> 4:        NA        NA        NA        NA        NA        NA       NA
#>         Max
#> 1:     -Inf
#> 2: 80684.11
#> 3: 12117.68
#> 4:     -Inf
# }