Set parameters required to perturb numeric variables using a flex function

ck_flexparams() allows to define a flex function that is used to lookup perturbation magnitudes (percentages) used when perturbing continuous variables.

ck_flexparams(fp, p = c(0.25, 0.05), epsilon = 1, q = 3)

Arguments

fp
(numeric scalar); at which point should the noise coefficient function reaches its desired maximum (defined by the first element of p)

p
a numeric vector of length 2 where both elements specify a percentage. The first value refers to the desired maximum perturbation percentage for small cells (depending on fp) while the second element refers to the desired maximum perturbation percentage for large cells. Both values must be between 0 and 1 and need to be in descending order.

epsilon
a numeric vector in descending order with all values >= 0 and <= 1 with the first element forced to equal 1. The length of this vector must correspond with the number top_k specified in ck_params_nums() when creating parameters for type == "top_contr" which is checked at runtime. This setting allows to use different flex-functions for the largest top_k contributors.

q
(numeric scalar); Parameter of the function; q needs to be >= 1

Value

an object suitable as input for ck_params_nums().

Details

details about the flex function can be found in Deliverable D4.2, Part I in SGA "Open Source tools for perturbative confidentiality methods"

See also

Examples

x <- ck_create_testdata()

# create some 0/1 variables that should be perturbed later
x[, cnt_females := ifelse(sex == "male", 0, 1)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              35          25.00000           0
#>    2:   2530      28       1              47          25.00000           1
#>    3:   6920     550       1              83          25.00000           0
#>    4:   7960     870       1              44          25.00000           0
#>    5:   9030      20       2              27          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              20          16.66667           1
#> 4577:   1420     987    1000              70          16.66667           0
#> 4578:   8900     684    1000              41          16.66667           0
#> 4579:   3880     294    1000              53          16.66667           1
#> 4580:   4830     911    1000              88          16.66667           0

x[, cnt_males := ifelse(sex == "male", 1, 0)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              35          25.00000           0
#>    2:   2530      28       1              47          25.00000           1
#>    3:   6920     550       1              83          25.00000           0
#>    4:   7960     870       1              44          25.00000           0
#>    5:   9030      20       2              27          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              20          16.66667           1
#> 4577:   1420     987    1000              70          16.66667           0
#> 4578:   8900     684    1000              41          16.66667           0
#> 4579:   3880     294    1000              53          16.66667           1
#> 4580:   4830     911    1000              88          16.66667           0
#>       cnt_males
#>    1:         1
#>    2:         0
#>    3:         1
#>    4:         1
#>    5:         1
#>   ---          
#> 4576:         0
#> 4577:         1
#> 4578:         1
#> 4579:         0
#> 4580:         1

x[, cnt_highincome := ifelse(income >= 9000, 1, 0)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              35          25.00000           0
#>    2:   2530      28       1              47          25.00000           1
#>    3:   6920     550       1              83          25.00000           0
#>    4:   7960     870       1              44          25.00000           0
#>    5:   9030      20       2              27          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              20          16.66667           1
#> 4577:   1420     987    1000              70          16.66667           0
#> 4578:   8900     684    1000              41          16.66667           0
#> 4579:   3880     294    1000              53          16.66667           1
#> 4580:   4830     911    1000              88          16.66667           0
#>       cnt_males cnt_highincome
#>    1:         1              0
#>    2:         0              0
#>    3:         1              0
#>    4:         1              0
#>    5:         1              1
#>   ---                         
#> 4576:         0              0
#> 4577:         1              0
#> 4578:         1              0
#> 4579:         0              0
#> 4580:         1              0

# a variable with positive and negative contributions
x[, mixed := sample(-10:10, nrow(x), replace = TRUE)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              35          25.00000           0
#>    2:   2530      28       1              47          25.00000           1
#>    3:   6920     550       1              83          25.00000           0
#>    4:   7960     870       1              44          25.00000           0
#>    5:   9030      20       2              27          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              20          16.66667           1
#> 4577:   1420     987    1000              70          16.66667           0
#> 4578:   8900     684    1000              41          16.66667           0
#> 4579:   3880     294    1000              53          16.66667           1
#> 4580:   4830     911    1000              88          16.66667           0
#>       cnt_males cnt_highincome mixed
#>    1:         1              0    -8
#>    2:         0              0     1
#>    3:         1              0     1
#>    4:         1              0    -3
#>    5:         1              1     1
#>   ---                               
#> 4576:         0              0    -7
#> 4577:         1              0     9
#> 4578:         1              0     6
#> 4579:         0              0    -3
#> 4580:         1              0    -2

# create record keys
x$rkey <- ck_generate_rkeys(dat = x)

# define required inputs

# hierarchy with some bogus codes
d_sex <- hier_create(root = "Total", nodes = c("male", "female"))
d_sex <- hier_add(d_sex, root = "female", "f")
d_sex <- hier_add(d_sex, root = "male", "m")

d_age <- hier_create(root = "Total", nodes = paste0("age_group", 1:6))
d_age <- hier_add(d_age, root = "age_group1", "ag1a")
d_age <- hier_add(d_age, root = "age_group2", "ag2a")

# define the cell key object
countvars <- c("cnt_females", "cnt_males", "cnt_highincome")
numvars <- c("expend", "income", "savings", "mixed")
tab <- ck_setup(
  x = x,
  rkey = "rkey",
  dims = list(sex = d_sex, age = d_age),
  w = "sampling_weight",
  countvars = countvars,
  numvars = numvars)

# show some information about this table instance
tab$print() # identical with print(tab)

#> ── Table Information ───────────────────────────────────────────────────────────
#> ✔ 45 cells in 2 dimensions ('sex', 'age')
#> ✔ weights: yes
#> ── Tabulated / Perturbed countvars ─────────────────────────────────────────────
#> ☐ 'total'
#> ☐ 'cnt_females'
#> ☐ 'cnt_males'
#> ☐ 'cnt_highincome'
#> ── Tabulated / Perturbed numvars ───────────────────────────────────────────────
#> ☐ 'expend'
#> ☐ 'income'
#> ☐ 'savings'
#> ☐ 'mixed'

# information about the hierarchies
tab$hierarchy_info()

#> $sex
#>      code level is_leaf parent
#> 1:  Total     1   FALSE   <NA>
#> 2:   male     2   FALSE  Total
#> 3: female     2   FALSE  Total
#> 4:      f     3    TRUE female
#> 5:      m     3    TRUE   male
#> 
#> $age
#>          code level is_leaf     parent
#> 1:      Total     1   FALSE       <NA>
#> 2: age_group1     2   FALSE      Total
#> 3: age_group2     2   FALSE      Total
#> 4: age_group3     2    TRUE      Total
#> 5: age_group4     2    TRUE      Total
#> 6: age_group5     2    TRUE      Total
#> 7: age_group6     2    TRUE      Total
#> 8:       ag1a     3    TRUE age_group1
#> 9:       ag2a     3    TRUE age_group2
#> 

# which variables have been defined?
tab$allvars()

#> $cntvars
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"
#> 
#> $numvars
#> [1] "expend"  "income"  "savings" "mixed"  
#> 

# count variables
tab$cntvars()

#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"

# continuous variables
tab$numvars()

#> [1] "expend"  "income"  "savings" "mixed"  

# create perturbation parameters for "total" variable and
# write to yaml-file

# create a ptable using functionality from the ptable-pkg
f_yaml <- tempfile(fileext = ".yaml")
p_cnts1 <- ck_params_cnts(
  ptab = ptable::pt_ex_cnts(),
  path = f_yaml)

#> yaml configuration '/tmp/RtmpNgukcO/file37dd76e5f8b9.yaml' successfully written.

# read parameters from yaml-file and set them for variable `"total"`
p_cnts1 <- ck_read_yaml(path = f_yaml)

tab$params_cnts_set(val = p_cnts1, v = "total")

#> --> setting perturbation parameters for variable 'total'

# create alternative perturbation parameters by specifying parameters
para2 <- ptable::pt_create_pParams(
  D = 8, V = 3, js = 2, table = "cnts")

#> Ignoring argument `step` for freqency tables

p_cnts2 <- ck_params_cnts(ptab = para2)

# use these ptable it for the remaining variables
tab$params_cnts_set(val = p_cnts2, v = countvars)

#> --> setting perturbation parameters for variable 'cnt_females'

#> --> setting perturbation parameters for variable 'cnt_males'

#> --> setting perturbation parameters for variable 'cnt_highincome'

# perturb a variable
tab$perturb(v = "total")

#> Count variable 'total' was perturbed.

# multiple variables can be perturbed as well
tab$perturb(v = c("cnt_males", "cnt_highincome"))

#> Count variable 'cnt_males' was perturbed.

#> Count variable 'cnt_highincome' was perturbed.

# return weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))

#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 274465 4581 274524.9269
#>  2:  Total age_group1     total 1969 116795 1969 116795.0000
#>  3:  Total       ag1a     total 1969 116795 1969 116795.0000
#>  4:  Total age_group2     total 1143  69363 1143  69363.0000
#>  5:  Total       ag2a     total 1143  69363 1143  69363.0000
#>  6:  Total age_group3     total  864  52731  864  52731.0000
#>  7:  Total age_group4     total  423  25184  424  25243.5366
#>  8:  Total age_group5     total  168   9625  169   9682.2917
#>  9:  Total age_group6     total   13    767   13    767.0000
#> 10: female      Total     total 2284 137121 2285 137181.0355
#> 11:      f      Total     total 2284 137121 2285 137181.0355
#> 12: female age_group1     total  954  57051  954  57051.0000
#> 13:      f age_group1     total  954  57051  954  57051.0000
#> 14: female       ag1a     total  954  57051  954  57051.0000
#> 15:      f       ag1a     total  954  57051  954  57051.0000
#> 16: female age_group2     total  572  35019  572  35019.0000
#> 17:      f age_group2     total  572  35019  572  35019.0000
#> 18: female       ag2a     total  572  35019  572  35019.0000
#> 19:      f       ag2a     total  572  35019  572  35019.0000
#> 20: female age_group3     total  440  26536  440  26536.0000
#> 21:      f age_group3     total  440  26536  440  26536.0000
#> 22: female age_group4     total  228  13299  230  13415.6579
#> 23:      f age_group4     total  228  13299  230  13415.6579
#> 24: female age_group5     total   84   4795   83   4737.9167
#> 25:      f age_group5     total   84   4795   83   4737.9167
#> 26: female age_group6     total    6    421    7    491.1667
#> 27:      f age_group6     total    6    421    7    491.1667
#> 28:   male      Total     total 2296 137344 2297 137403.8188
#> 29:      m      Total     total 2296 137344 2297 137403.8188
#> 30:   male age_group1     total 1015  59744 1015  59744.0000
#> 31:      m age_group1     total 1015  59744 1015  59744.0000
#> 32:   male       ag1a     total 1015  59744 1015  59744.0000
#> 33:      m       ag1a     total 1015  59744 1015  59744.0000
#> 34:   male age_group2     total  571  34344  570  34283.8529
#> 35:      m age_group2     total  571  34344  570  34283.8529
#> 36:   male       ag2a     total  571  34344  570  34283.8529
#> 37:      m       ag2a     total  571  34344  570  34283.8529
#> 38:   male age_group3     total  424  26195  426  26318.5613
#> 39:      m age_group3     total  424  26195  426  26318.5613
#> 40:   male age_group4     total  195  11885  196  11945.9487
#> 41:      m age_group4     total  195  11885  196  11945.9487
#> 42:   male age_group5     total   84   4830   85   4887.5000
#> 43:      m age_group5     total   84   4830   85   4887.5000
#> 44:   male age_group6     total    7    346    8    395.4286
#> 45:      m age_group6     total    7    346    8    395.4286
#> 46:  Total      Total cnt_males 2296 137344 2298 137463.6376
#> 47:  Total age_group1 cnt_males 1015  59744 1016  59802.8611
#> 48:  Total       ag1a cnt_males 1015  59744 1016  59802.8611
#> 49:  Total age_group2 cnt_males  571  34344  569  34223.7058
#> 50:  Total       ag2a cnt_males  571  34344  569  34223.7058
#> 51:  Total age_group3 cnt_males  424  26195  427  26380.3420
#> 52:  Total age_group4 cnt_males  195  11885  196  11945.9487
#> 53:  Total age_group5 cnt_males   84   4830   85   4887.5000
#> 54:  Total age_group6 cnt_males    7    346    8    395.4286
#> 55: female      Total cnt_males    0      0    0      0.0000
#> 56:      f      Total cnt_males    0      0    0      0.0000
#> 57: female age_group1 cnt_males    0      0    0      0.0000
#> 58:      f age_group1 cnt_males    0      0    0      0.0000
#> 59: female       ag1a cnt_males    0      0    0      0.0000
#> 60:      f       ag1a cnt_males    0      0    0      0.0000
#> 61: female age_group2 cnt_males    0      0    0      0.0000
#> 62:      f age_group2 cnt_males    0      0    0      0.0000
#> 63: female       ag2a cnt_males    0      0    0      0.0000
#> 64:      f       ag2a cnt_males    0      0    0      0.0000
#> 65: female age_group3 cnt_males    0      0    0      0.0000
#> 66:      f age_group3 cnt_males    0      0    0      0.0000
#> 67: female age_group4 cnt_males    0      0    0      0.0000
#> 68:      f age_group4 cnt_males    0      0    0      0.0000
#> 69: female age_group5 cnt_males    0      0    0      0.0000
#> 70:      f age_group5 cnt_males    0      0    0      0.0000
#> 71: female age_group6 cnt_males    0      0    0      0.0000
#> 72:      f age_group6 cnt_males    0      0    0      0.0000
#> 73:   male      Total cnt_males 2296 137344 2298 137463.6376
#> 74:      m      Total cnt_males 2296 137344 2298 137463.6376
#> 75:   male age_group1 cnt_males 1015  59744 1016  59802.8611
#> 76:      m age_group1 cnt_males 1015  59744 1016  59802.8611
#> 77:   male       ag1a cnt_males 1015  59744 1016  59802.8611
#> 78:      m       ag1a cnt_males 1015  59744 1016  59802.8611
#> 79:   male age_group2 cnt_males  571  34344  569  34223.7058
#> 80:      m age_group2 cnt_males  571  34344  569  34223.7058
#> 81:   male       ag2a cnt_males  571  34344  569  34223.7058
#> 82:      m       ag2a cnt_males  571  34344  569  34223.7058
#> 83:   male age_group3 cnt_males  424  26195  427  26380.3420
#> 84:      m age_group3 cnt_males  424  26195  427  26380.3420
#> 85:   male age_group4 cnt_males  195  11885  196  11945.9487
#> 86:      m age_group4 cnt_males  195  11885  196  11945.9487
#> 87:   male age_group5 cnt_males   84   4830   85   4887.5000
#> 88:      m age_group5 cnt_males   84   4830   85   4887.5000
#> 89:   male age_group6 cnt_males    7    346    8    395.4286
#> 90:      m age_group6 cnt_males    7    346    8    395.4286
#>        sex        age     vname  uwc     wc puwc         pwc

# numerical variables (positive variables using flex-function)
# we also write the config to a yaml file
f_yaml <- tempfile(fileext = ".yaml")

# create a ptable using functionality from the ptable-pkg
# a single ptable for all cells
ptab1 <- ptable::pt_ex_nums(parity = TRUE, separation = FALSE)

# a single ptab for all cells except for very small ones
ptab2 <- ptable::pt_ex_nums(parity = TRUE, separation = TRUE)

# different ptables for cells with even/odd number of contributors
# and very small cells
ptab3 <- ptable::pt_ex_nums(parity = FALSE, separation = TRUE)

p_nums1 <- ck_params_nums(
  ptab = ptab1,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.30, 0.03),
    epsilon = c(1, 0.5, 0.2),
    q = 3),
  mu_c = 2,
  same_key = FALSE,
  use_zero_rkeys = FALSE,
  path = f_yaml)

#> yaml configuration '/tmp/RtmpNgukcO/file37dd166debef.yaml' successfully written.

# we read the parameters from the yaml-file
p_nums1 <- ck_read_yaml(path = f_yaml)

# for variables with positive and negative values
p_nums2 <- ck_params_nums(
  ptab = ptab2,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.15, 0.02),
    epsilon = c(1, 0.4, 0.15),
    q = 3),
  mu_c = 2,
  same_key = FALSE)

# simple perturbation parameters (not using the flex-function approach)
p_nums3 <- ck_params_nums(
  ptab = ptab3,
  type = "mean",
  mult_params = ck_simpleparams(p = 0.25),
  mu_c = 2,
  same_key = FALSE)

# use `p_nums1` for all variables
tab$params_nums_set(p_nums1, c("savings", "income", "expend"))

#> --> setting perturbation parameters for variable 'savings'

#> --> setting perturbation parameters for variable 'income'

#> --> setting perturbation parameters for variable 'expend'

# use different parameters for variable `mixed`
tab$params_nums_set(p_nums2, v = "mixed")

#> --> setting perturbation parameters for variable 'mixed'

# identify sensitive cells to which extra protection (`mu_c`) is added.
tab$supp_p(v = "income", p = 85)

#> p%-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)

tab$supp_pq(v = "income", p = 85, q = 90)

#> pq-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)

tab$supp_nk(v = "income", n = 2, k = 90)

#> nk-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)

tab$supp_freq(v = "income", n = 14, weighted = FALSE)

#> freq-rule: 5 new sensitive cells (incl. duplicates) found (total: 5)

tab$supp_val(v = "income", n = 10000, weighted = TRUE)

#> val-rule: 0 new sensitive cells (incl. duplicates) found (total: 5)

# perturb variables
tab$perturb(v = c("income", "savings"))

#> Numeric variable 'income' was perturbed.

#> Numeric variable 'savings' was perturbed.

# extract results
tab$numtab("income", mean_before_sum = TRUE)

#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1380343174 1380335604
#>  2:  Total age_group1 income  9810547  583371885  583399557
#>  3:  Total       ag1a income  9810547  583371885  583399557
#>  4:  Total age_group2 income  5692119  348230950  348251186
#>  5:  Total       ag2a income  5692119  348230950  348251186
#>  6:  Total age_group3 income  4406946  271419101  271359454
#>  7:  Total age_group4 income  2133543  124817955  124730892
#>  8:  Total age_group5 income   848151   48665777   48610897
#>  9:  Total age_group6 income    61672    3837506    3738510
#> 10: female      Total income 11690929  702395312  702404924
#> 11:      f      Total income 11690929  702395312  702404924
#> 12: female age_group1 income  4933383  296210793  296272001
#> 13:      f age_group1 income  4933383  296210793  296272001
#> 14: female       ag1a income  4933383  296210793  296272001
#> 15:      f       ag1a income  4933383  296210793  296272001
#> 16: female age_group2 income  2880740  175833631  175908909
#> 17:      f age_group2 income  2880740  175833631  175908909
#> 18: female       ag2a income  2880740  175833631  175908909
#> 19:      f       ag2a income  2880740  175833631  175908909
#> 20: female age_group3 income  2238777  136520932  136545867
#> 21:      f age_group3 income  2238777  136520932  136545867
#> 22: female age_group4 income  1155033   66064250   66088828
#> 23:      f age_group4 income  1155033   66064250   66088828
#> 24: female age_group5 income   455017   25669046   25678903
#> 25:      f age_group5 income   455017   25669046   25678903
#> 26: female age_group6 income    27979    2096660    2029792
#> 27:      f age_group6 income    27979    2096660    2029792
#> 28:   male      Total income 11262049  677947862  678148672
#> 29:      m      Total income 11262049  677947862  678148672
#> 30:   male age_group1 income  4877164  287161092  287177909
#> 31:      m age_group1 income  4877164  287161092  287177909
#> 32:   male       ag1a income  4877164  287161092  287177909
#> 33:      m       ag1a income  4877164  287161092  287177909
#> 34:   male age_group2 income  2811379  172397319  172500091
#> 35:      m age_group2 income  2811379  172397319  172500091
#> 36:   male       ag2a income  2811379  172397319  172500091
#> 37:      m       ag2a income  2811379  172397319  172500091
#> 38:   male age_group3 income  2168169  134898169  134930798
#> 39:      m age_group3 income  2168169  134898169  134930798
#> 40:   male age_group4 income   978510   58753705   58788777
#> 41:      m age_group4 income   978510   58753705   58788777
#> 42:   male age_group5 income   393134   22996731   23023571
#> 43:      m age_group5 income   393134   22996731   23023571
#> 44:   male age_group6 income    33693    1740846    1825547
#> 45:      m age_group6 income    33693    1740846    1825547
#>        sex        age  vname      uws         ws        pws

tab$numtab("income", mean_before_sum = FALSE)

#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1380343174 1380339389
#>  2:  Total age_group1 income  9810547  583371885  583385721
#>  3:  Total       ag1a income  9810547  583371885  583385721
#>  4:  Total age_group2 income  5692119  348230950  348241068
#>  5:  Total       ag2a income  5692119  348230950  348241068
#>  6:  Total age_group3 income  4406946  271419101  271389276
#>  7:  Total age_group4 income  2133543  124817955  124774416
#>  8:  Total age_group5 income   848151   48665777   48638329
#>  9:  Total age_group6 income    61672    3837506    3787685
#> 10: female      Total income 11690929  702395312  702400118
#> 11:      f      Total income 11690929  702395312  702400118
#> 12: female age_group1 income  4933383  296210793  296241396
#> 13:      f age_group1 income  4933383  296210793  296241396
#> 14: female       ag1a income  4933383  296210793  296241396
#> 15:      f       ag1a income  4933383  296210793  296241396
#> 16: female age_group2 income  2880740  175833631  175871266
#> 17:      f age_group2 income  2880740  175833631  175871266
#> 18: female       ag2a income  2880740  175833631  175871266
#> 19:      f       ag2a income  2880740  175833631  175871266
#> 20: female age_group3 income  2238777  136520932  136533399
#> 21:      f age_group3 income  2238777  136520932  136533399
#> 22: female age_group4 income  1155033   66064250   66076538
#> 23:      f age_group4 income  1155033   66064250   66076538
#> 24: female age_group5 income   455017   25669046   25673974
#> 25:      f age_group5 income   455017   25669046   25673974
#> 26: female age_group6 income    27979    2096660    2062955
#> 27:      f age_group6 income    27979    2096660    2062955
#> 28:   male      Total income 11262049  677947862  678048260
#> 29:      m      Total income 11262049  677947862  678048260
#> 30:   male age_group1 income  4877164  287161092  287169500
#> 31:      m age_group1 income  4877164  287161092  287169500
#> 32:   male       ag1a income  4877164  287161092  287169500
#> 33:      m       ag1a income  4877164  287161092  287169500
#> 34:   male age_group2 income  2811379  172397319  172448697
#> 35:      m age_group2 income  2811379  172397319  172448697
#> 36:   male       ag2a income  2811379  172397319  172448697
#> 37:      m       ag2a income  2811379  172397319  172448697
#> 38:   male age_group3 income  2168169  134898169  134914482
#> 39:      m age_group3 income  2168169  134898169  134914482
#> 40:   male age_group4 income   978510   58753705   58771238
#> 41:      m age_group4 income   978510   58753705   58771238
#> 42:   male age_group5 income   393134   22996731   23010147
#> 43:      m age_group5 income   393134   22996731   23010147
#> 44:   male age_group6 income    33693    1740846    1782693
#> 45:      m age_group6 income    33693    1740846    1782693
#>        sex        age  vname      uws         ws        pws

tab$numtab("savings")

#>        sex        age   vname     uws        ws         pws
#>  1:  Total      Total savings 2273532 136020945 136015539.8
#>  2:  Total age_group1 savings  982386  57768819  57768663.7
#>  3:  Total       ag1a savings  982386  57768819  57768663.7
#>  4:  Total age_group2 savings  552336  33586720  33587450.8
#>  5:  Total       ag2a savings  552336  33586720  33587450.8
#>  6:  Total age_group3 savings  437101  26838230  26847907.2
#>  7:  Total age_group4 savings  214661  12773384  12768231.2
#>  8:  Total age_group5 savings   80451   4647387   4643029.2
#>  9:  Total age_group6 savings    6597    406405    402727.0
#> 10: female      Total savings 1113716  66908937  66910527.3
#> 11:      f      Total savings 1113716  66908937  66910527.3
#> 12: female age_group1 savings  464726  27490233  27489599.9
#> 13:      f age_group1 savings  464726  27490233  27489599.9
#> 14: female       ag1a savings  464726  27490233  27489599.9
#> 15:      f       ag1a savings  464726  27490233  27489599.9
#> 16: female age_group2 savings  271413  16745796  16749024.3
#> 17:      f age_group2 savings  271413  16745796  16749024.3
#> 18: female       ag2a savings  271413  16745796  16749024.3
#> 19:      f       ag2a savings  271413  16745796  16749024.3
#> 20: female age_group3 savings  222131  13637472  13636970.3
#> 21:      f age_group3 savings  222131  13637472  13636970.3
#> 22: female age_group4 savings  115241   6655694   6656398.0
#> 23:      f age_group4 savings  115241   6655694   6656398.0
#> 24: female age_group5 savings   37218   2146801   2147977.8
#> 25:      f age_group5 savings   37218   2146801   2147977.8
#> 26: female age_group6 savings    2987    232941    231706.2
#> 27:      f age_group6 savings    2987    232941    231706.2
#> 28:   male      Total savings 1159816  69112008  69118477.6
#> 29:      m      Total savings 1159816  69112008  69118477.6
#> 30:   male age_group1 savings  517660  30278586  30280539.9
#> 31:      m age_group1 savings  517660  30278586  30280539.9
#> 32:   male       ag1a savings  517660  30278586  30280539.9
#> 33:      m       ag1a savings  517660  30278586  30280539.9
#> 34:   male age_group2 savings  280923  16840924  16844641.4
#> 35:      m age_group2 savings  280923  16840924  16844641.4
#> 36:   male       ag2a savings  280923  16840924  16844641.4
#> 37:      m       ag2a savings  280923  16840924  16844641.4
#> 38:   male age_group3 savings  214970  13200758  13204976.6
#> 39:      m age_group3 savings  214970  13200758  13204976.6
#> 40:   male age_group4 savings   99420   6117690   6119190.5
#> 41:      m age_group4 savings   99420   6117690   6119190.5
#> 42:   male age_group5 savings   43233   2500586   2503487.6
#> 43:      m age_group5 savings   43233   2500586   2503487.6
#> 44:   male age_group6 savings    3610    173464    173794.6
#> 45:      m age_group6 savings    3610    173464    173794.6
#>        sex        age   vname     uws        ws         pws

# results can be resetted, too
tab$reset_cntvars(v = "cnt_males")

# we can then set other parameters and perturb again
tab$params_cnts_set(val = p_cnts1, v = "cnt_males")

#> --> setting perturbation parameters for variable 'cnt_males'

tab$perturb(v = "cnt_males")

#> Count variable 'cnt_males' was perturbed.

# write to a file "outtab.csv" (.csv is automatically added to the path)
if (FALSE) {
tab$freqtab(v = c("total", "cnt_males"), path = "outtab.csv")
}

# show results containing weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))

#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 274465 4581 274524.9269
#>  2:  Total age_group1     total 1969 116795 1969 116795.0000
#>  3:  Total       ag1a     total 1969 116795 1969 116795.0000
#>  4:  Total age_group2     total 1143  69363 1143  69363.0000
#>  5:  Total       ag2a     total 1143  69363 1143  69363.0000
#>  6:  Total age_group3     total  864  52731  864  52731.0000
#>  7:  Total age_group4     total  423  25184  424  25243.5366
#>  8:  Total age_group5     total  168   9625  169   9682.2917
#>  9:  Total age_group6     total   13    767   13    767.0000
#> 10: female      Total     total 2284 137121 2285 137181.0355
#> 11:      f      Total     total 2284 137121 2285 137181.0355
#> 12: female age_group1     total  954  57051  954  57051.0000
#> 13:      f age_group1     total  954  57051  954  57051.0000
#> 14: female       ag1a     total  954  57051  954  57051.0000
#> 15:      f       ag1a     total  954  57051  954  57051.0000
#> 16: female age_group2     total  572  35019  572  35019.0000
#> 17:      f age_group2     total  572  35019  572  35019.0000
#> 18: female       ag2a     total  572  35019  572  35019.0000
#> 19:      f       ag2a     total  572  35019  572  35019.0000
#> 20: female age_group3     total  440  26536  440  26536.0000
#> 21:      f age_group3     total  440  26536  440  26536.0000
#> 22: female age_group4     total  228  13299  230  13415.6579
#> 23:      f age_group4     total  228  13299  230  13415.6579
#> 24: female age_group5     total   84   4795   83   4737.9167
#> 25:      f age_group5     total   84   4795   83   4737.9167
#> 26: female age_group6     total    6    421    7    491.1667
#> 27:      f age_group6     total    6    421    7    491.1667
#> 28:   male      Total     total 2296 137344 2297 137403.8188
#> 29:      m      Total     total 2296 137344 2297 137403.8188
#> 30:   male age_group1     total 1015  59744 1015  59744.0000
#> 31:      m age_group1     total 1015  59744 1015  59744.0000
#> 32:   male       ag1a     total 1015  59744 1015  59744.0000
#> 33:      m       ag1a     total 1015  59744 1015  59744.0000
#> 34:   male age_group2     total  571  34344  570  34283.8529
#> 35:      m age_group2     total  571  34344  570  34283.8529
#> 36:   male       ag2a     total  571  34344  570  34283.8529
#> 37:      m       ag2a     total  571  34344  570  34283.8529
#> 38:   male age_group3     total  424  26195  426  26318.5613
#> 39:      m age_group3     total  424  26195  426  26318.5613
#> 40:   male age_group4     total  195  11885  196  11945.9487
#> 41:      m age_group4     total  195  11885  196  11945.9487
#> 42:   male age_group5     total   84   4830   85   4887.5000
#> 43:      m age_group5     total   84   4830   85   4887.5000
#> 44:   male age_group6     total    7    346    8    395.4286
#> 45:      m age_group6     total    7    346    8    395.4286
#> 46:  Total      Total cnt_males 2296 137344 2297 137403.8188
#> 47:  Total age_group1 cnt_males 1015  59744 1015  59744.0000
#> 48:  Total       ag1a cnt_males 1015  59744 1015  59744.0000
#> 49:  Total age_group2 cnt_males  571  34344  570  34283.8529
#> 50:  Total       ag2a cnt_males  571  34344  570  34283.8529
#> 51:  Total age_group3 cnt_males  424  26195  426  26318.5613
#> 52:  Total age_group4 cnt_males  195  11885  196  11945.9487
#> 53:  Total age_group5 cnt_males   84   4830   85   4887.5000
#> 54:  Total age_group6 cnt_males    7    346    8    395.4286
#> 55: female      Total cnt_males    0      0    0      0.0000
#> 56:      f      Total cnt_males    0      0    0      0.0000
#> 57: female age_group1 cnt_males    0      0    0      0.0000
#> 58:      f age_group1 cnt_males    0      0    0      0.0000
#> 59: female       ag1a cnt_males    0      0    0      0.0000
#> 60:      f       ag1a cnt_males    0      0    0      0.0000
#> 61: female age_group2 cnt_males    0      0    0      0.0000
#> 62:      f age_group2 cnt_males    0      0    0      0.0000
#> 63: female       ag2a cnt_males    0      0    0      0.0000
#> 64:      f       ag2a cnt_males    0      0    0      0.0000
#> 65: female age_group3 cnt_males    0      0    0      0.0000
#> 66:      f age_group3 cnt_males    0      0    0      0.0000
#> 67: female age_group4 cnt_males    0      0    0      0.0000
#> 68:      f age_group4 cnt_males    0      0    0      0.0000
#> 69: female age_group5 cnt_males    0      0    0      0.0000
#> 70:      f age_group5 cnt_males    0      0    0      0.0000
#> 71: female age_group6 cnt_males    0      0    0      0.0000
#> 72:      f age_group6 cnt_males    0      0    0      0.0000
#> 73:   male      Total cnt_males 2296 137344 2297 137403.8188
#> 74:      m      Total cnt_males 2296 137344 2297 137403.8188
#> 75:   male age_group1 cnt_males 1015  59744 1015  59744.0000
#> 76:      m age_group1 cnt_males 1015  59744 1015  59744.0000
#> 77:   male       ag1a cnt_males 1015  59744 1015  59744.0000
#> 78:      m       ag1a cnt_males 1015  59744 1015  59744.0000
#> 79:   male age_group2 cnt_males  571  34344  570  34283.8529
#> 80:      m age_group2 cnt_males  571  34344  570  34283.8529
#> 81:   male       ag2a cnt_males  571  34344  570  34283.8529
#> 82:      m       ag2a cnt_males  571  34344  570  34283.8529
#> 83:   male age_group3 cnt_males  424  26195  426  26318.5613
#> 84:      m age_group3 cnt_males  424  26195  426  26318.5613
#> 85:   male age_group4 cnt_males  195  11885  196  11945.9487
#> 86:      m age_group4 cnt_males  195  11885  196  11945.9487
#> 87:   male age_group5 cnt_males   84   4830   85   4887.5000
#> 88:      m age_group5 cnt_males   84   4830   85   4887.5000
#> 89:   male age_group6 cnt_males    7    346    8    395.4286
#> 90:      m age_group6 cnt_males    7    346    8    395.4286
#>        sex        age     vname  uwc     wc puwc         pwc

# utility measures for a count variable
tab$measures_cnts(v = "total", exclude_zeros = TRUE)

#> $overview
#>    noise cnt        pct
#> 1:    -2   4 0.08888889
#> 2:    -1  15 0.33333333
#> 3:     0  20 0.44444444
#> 4:     1   6 0.13333333
#> 
#> $measures
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 0.000 0.000 0.000
#>  6:   Mean 0.644 0.016 0.033
#>  7: Median 1.000 0.000 0.010
#>  8:    Q60 1.000 0.002 0.021
#>  9:    Q70 1.000 0.005 0.036
#> 10:    Q80 1.000 0.009 0.054
#> 11:    Q90 1.000 0.012 0.066
#> 12:    Q95 2.000 0.143 0.183
#> 13:    Q99 2.000 0.167 0.196
#> 14:    Max 2.000 0.167 0.196
#> 
#> $cumdistr_d1
#>    cat cnt       pct
#> 1:   0  20 0.4444444
#> 2:   1  41 0.9111111
#> 3:   2  45 1.0000000
#> 
#> $cumdistr_d2
#>            cat cnt       pct
#> 1:    [0,0.02]  41 0.9111111
#> 2: (0.02,0.05]  41 0.9111111
#> 3:  (0.05,0.1]  41 0.9111111
#> 4:   (0.1,0.2]  45 1.0000000
#> 5:   (0.2,0.3]  45 1.0000000
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $cumdistr_d3
#>            cat cnt       pct
#> 1:    [0,0.02]  25 0.5555556
#> 2: (0.02,0.05]  35 0.7777778
#> 3:  (0.05,0.1]  41 0.9111111
#> 4:   (0.1,0.2]  45 1.0000000
#> 5:   (0.2,0.3]  45 1.0000000
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $false_zero
#> [1] 0
#> 
#> $false_nonzero
#> [1] 0
#> 
#> $exclude_zeros
#> [1] TRUE
#> 

# modifications for perturbed count variables
tab$mod_cnts()

#>        sex        age row_nr pert      ckey  countvar
#>   1: Total      Total     16    1 0.7397504     total
#>   2: Total age_group1     15    0 0.3519898     total
#>   3: Total       ag1a     15    0 0.3519898     total
#>   4: Total age_group2     15    0 0.6456544     total
#>   5: Total       ag2a     15    0 0.6456544     total
#>  ---                                                 
#> 131:     m age_group4     16    1 0.7724867 cnt_males
#> 132:  male age_group5     16    1 0.6920779 cnt_males
#> 133:     m age_group5     16    1 0.6920779 cnt_males
#> 134:  male age_group6     16    1 0.6902029 cnt_males
#> 135:     m age_group6     16    1 0.6902029 cnt_males

# display a summary about utility measures
tab$summary()

#> ┌──────────────────────────────────────────────┐
#> │Utility measures for perturbed count variables│
#> └──────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#>          countvar Min Q10 Q20 Q30  Q40   Mean Median Q60 Q70 Q80 Q90 Q95 Q99
#> 1:          total  -1  -1   0   0  0.0  0.378      0   1   1   1   1 2.0   2
#> 2: cnt_highincome  -3  -3  -1  -1 -0.4 -0.356      0   0   0   0   2 2.0   5
#> 3:      cnt_males  -1  -1   0   0  0.0  0.267      0   0   1   1   1 1.8   2
#>    Max
#> 1:   2
#> 2:   5
#> 3:   2
#> 
#> ── Distance-based measures ─────────────────────────────────────────────────────
#> ✔ Variable: 'total'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 0.000 0.000 0.000
#>  6:   Mean 0.644 0.016 0.033
#>  7: Median 1.000 0.000 0.010
#>  8:    Q60 1.000 0.002 0.021
#>  9:    Q70 1.000 0.005 0.036
#> 10:    Q80 1.000 0.009 0.054
#> 11:    Q90 1.000 0.012 0.066
#> 12:    Q95 2.000 0.143 0.183
#> 13:    Q99 2.000 0.167 0.196
#> 14:    Max 2.000 0.167 0.196
#> 
#> ✔ Variable: 'cnt_males'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.200 0.000 0.002
#>  4:    Q30 1.000 0.000 0.010
#>  5:    Q40 1.000 0.002 0.021
#>  6:   Mean 0.889 0.019 0.042
#>  7: Median 1.000 0.002 0.021
#>  8:    Q60 1.000 0.005 0.036
#>  9:    Q70 1.000 0.005 0.049
#> 10:    Q80 1.000 0.011 0.053
#> 11:    Q90 1.400 0.064 0.106
#> 12:    Q95 2.000 0.143 0.183
#> 13:    Q99 2.000 0.143 0.183
#> 14:    Max 2.000 0.143 0.183
#> 
#> ✔ Variable: 'cnt_highincome'
#> 
#>       what  d1    d2    d3
#>  1:    Min 0.0 0.000 0.000
#>  2:    Q10 0.0 0.000 0.000
#>  3:    Q20 0.0 0.000 0.000
#>  4:    Q30 0.0 0.000 0.000
#>  5:    Q40 0.6 0.005 0.040
#>  6:   Mean 1.3 0.054 0.121
#>  7: Median 1.0 0.017 0.067
#>  8:    Q60 1.0 0.022 0.090
#>  9:    Q70 2.0 0.029 0.140
#> 10:    Q80 3.0 0.082 0.159
#> 11:    Q90 3.0 0.144 0.361
#> 12:    Q95 3.1 0.171 0.393
#> 13:    Q99 5.0 0.429 0.646
#> 14:    Max 5.0 0.429 0.646
#> 
#> ┌──────────────────────────────────────────────────┐
#> │Utility measures for perturbed numerical variables│
#> └──────────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────

#> Warning: no non-missing arguments to min; returning Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to min; returning Inf

#> Warning: no non-missing arguments to max; returning -Inf

#>      vname        Min       Q10      Q20      Q30      Q40      Mean    Median
#> 1:  expend        Inf        NA       NA       NA       NA       NaN        NA
#> 2:  income -49821.407 -28874.10 4903.582 8408.477 12287.64 17470.388 13416.196
#> 3: savings  -5405.224  -1234.82 -633.073  -58.102   720.07  1324.699  1500.499
#> 4:   mixed        Inf        NA       NA       NA       NA       NaN        NA
#>          Q60      Q70       Q80       Q90       Q95        Q99        Max
#> 1:        NA       NA        NA        NA        NA         NA       -Inf
#> 2: 16801.382 30602.65 37634.749 51378.167 51378.167 100397.693 100397.693
#> 3:  1953.854  2901.60  3326.157  4018.139  6019.372   8265.846   9677.213
#> 4:        NA       NA        NA        NA        NA         NA       -Inf