Create perturbation parameters for count variables

This function allows to generate required perturbation parameters that are used to perturb count variables.

ck_params_cnts(ptab, path = NULL)

Arguments

ptab

an object created with ptable::create_ptable(), or ptable::create_cnt_ptable()

path

a scalar character specifying a path to which the parameters created with this functions should be written to (in yaml format)

Value

an object suitable as input to method $params_cnts_set() for the perturbation of counts and frequencies.

See also

Examples

x <- ck_create_testdata()

# create some 0/1 variables that should be perturbed later
x[, cnt_females := ifelse(sex == "male", 0, 1)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              21          25.00000           0
#>    2:   2530      28       1              65          25.00000           1
#>    3:   6920     550       1              94          25.00000           0
#>    4:   7960     870       1              37          25.00000           0
#>    5:   9030      20       2              49          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              94          16.66667           1
#> 4577:   1420     987    1000              29          16.66667           0
#> 4578:   8900     684    1000              40          16.66667           0
#> 4579:   3880     294    1000              32          16.66667           1
#> 4580:   4830     911    1000              62          16.66667           0
x[, cnt_males := ifelse(sex == "male", 1, 0)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              21          25.00000           0
#>    2:   2530      28       1              65          25.00000           1
#>    3:   6920     550       1              94          25.00000           0
#>    4:   7960     870       1              37          25.00000           0
#>    5:   9030      20       2              49          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              94          16.66667           1
#> 4577:   1420     987    1000              29          16.66667           0
#> 4578:   8900     684    1000              40          16.66667           0
#> 4579:   3880     294    1000              32          16.66667           1
#> 4580:   4830     911    1000              62          16.66667           0
#>       cnt_males
#>    1:         1
#>    2:         0
#>    3:         1
#>    4:         1
#>    5:         1
#>   ---          
#> 4576:         0
#> 4577:         1
#> 4578:         1
#> 4579:         0
#> 4580:         1
x[, cnt_highincome := ifelse(income >= 9000, 1, 0)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              21          25.00000           0
#>    2:   2530      28       1              65          25.00000           1
#>    3:   6920     550       1              94          25.00000           0
#>    4:   7960     870       1              37          25.00000           0
#>    5:   9030      20       2              49          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              94          16.66667           1
#> 4577:   1420     987    1000              29          16.66667           0
#> 4578:   8900     684    1000              40          16.66667           0
#> 4579:   3880     294    1000              32          16.66667           1
#> 4580:   4830     911    1000              62          16.66667           0
#>       cnt_males cnt_highincome
#>    1:         1              0
#>    2:         0              0
#>    3:         1              0
#>    4:         1              0
#>    5:         1              1
#>   ---                         
#> 4576:         0              0
#> 4577:         1              0
#> 4578:         1              0
#> 4579:         0              0
#> 4580:         1              0
# a variable with positive and negative contributions
x[, mixed := sample(-10:10, nrow(x), replace = TRUE)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              21          25.00000           0
#>    2:   2530      28       1              65          25.00000           1
#>    3:   6920     550       1              94          25.00000           0
#>    4:   7960     870       1              37          25.00000           0
#>    5:   9030      20       2              49          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              94          16.66667           1
#> 4577:   1420     987    1000              29          16.66667           0
#> 4578:   8900     684    1000              40          16.66667           0
#> 4579:   3880     294    1000              32          16.66667           1
#> 4580:   4830     911    1000              62          16.66667           0
#>       cnt_males cnt_highincome mixed
#>    1:         1              0     0
#>    2:         0              0    -9
#>    3:         1              0     0
#>    4:         1              0    -8
#>    5:         1              1     3
#>   ---                               
#> 4576:         0              0    -3
#> 4577:         1              0    -8
#> 4578:         1              0     5
#> 4579:         0              0     0
#> 4580:         1              0    -9

# create record keys
x$rkey <- ck_generate_rkeys(dat = x)

# define required inputs

# hierarchy with some bogus codes
d_sex <- hier_create(root = "Total", nodes = c("male", "female"))
d_sex <- hier_add(d_sex, root = "female", "f")
d_sex <- hier_add(d_sex, root = "male", "m")

d_age <- hier_create(root = "Total", nodes = paste0("age_group", 1:6))
d_age <- hier_add(d_age, root = "age_group1", "ag1a")
d_age <- hier_add(d_age, root = "age_group2", "ag2a")

# define the cell key object
countvars <- c("cnt_females", "cnt_males", "cnt_highincome")
numvars <- c("expend", "income", "savings", "mixed")
tab <- ck_setup(
  x = x,
  rkey = "rkey",
  dims = list(sex = d_sex, age = d_age),
  w = "sampling_weight",
  countvars = countvars,
  numvars = numvars)

# show some information about this table instance
tab$print() # identical with print(tab)
#> ── Table Information ───────────────────────────────────────────────────────────
#> ✔ 45 cells in 2 dimensions ('sex', 'age')
#> ✔ weights: yes
#> ── Tabulated / Perturbed countvars ─────────────────────────────────────────────
#> ☐ 'total'
#> ☐ 'cnt_females'
#> ☐ 'cnt_males'
#> ☐ 'cnt_highincome'
#> ── Tabulated / Perturbed numvars ───────────────────────────────────────────────
#> ☐ 'expend'
#> ☐ 'income'
#> ☐ 'savings'
#> ☐ 'mixed'

# information about the hierarchies
tab$hierarchy_info()
#> $sex
#>      code level is_leaf parent
#> 1:  Total     1   FALSE  Total
#> 2:   male     2   FALSE  Total
#> 3:      m     3    TRUE   male
#> 4: female     2   FALSE  Total
#> 5:      f     3    TRUE female
#> 
#> $age
#>          code level is_leaf     parent
#> 1:      Total     1   FALSE      Total
#> 2: age_group1     2   FALSE      Total
#> 3:       ag1a     3    TRUE age_group1
#> 4: age_group2     2   FALSE      Total
#> 5:       ag2a     3    TRUE age_group2
#> 6: age_group3     2    TRUE      Total
#> 7: age_group4     2    TRUE      Total
#> 8: age_group5     2    TRUE      Total
#> 9: age_group6     2    TRUE      Total
#> 

# which variables have been defined?
tab$allvars()
#> $cntvars
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"
#> 
#> $numvars
#> [1] "expend"  "income"  "savings" "mixed"  
#> 

# count variables
tab$cntvars()
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"

# continuous variables
tab$numvars()
#> [1] "expend"  "income"  "savings" "mixed"  

# create perturbation parameters for "total" variable and
# write to yaml-file

# create a ptable using functionality from the ptable-pkg
f_yaml <- tempfile(fileext = ".yaml")
p_cnts1 <- ck_params_cnts(
  ptab = ptable::pt_ex_cnts(),
  path = f_yaml)
#> yaml configuration '/tmp/RtmpRLPm23/file1e1b5d02cd6f.yaml' successfully written.

# read parameters from yaml-file and set them for variable `"total"`
p_cnts1 <- ck_read_yaml(path = f_yaml)

tab$params_cnts_set(val = p_cnts1, v = "total")
#> --> setting perturbation parameters for variable 'total'

# create alternative perturbation parameters by specifying parameters
para2 <- ptable::create_num_ptable(
  D = 8, V = 3, js = 2, create = FALSE)
#> Error in ptable::create_num_ptable(D = 8, V = 3, js = 2, create = FALSE): unused argument (js = 2)

p_cnts2 <- ck_params_cnts(ptab = para2)
#> Error in ck_params_cnts(ptab = para2): object 'para2' not found

# use these ptable it for the remaining variables
tab$params_cnts_set(val = p_cnts2, v = countvars)
#> Error in tab$params_cnts_set(val = p_cnts2, v = countvars): object 'p_cnts2' not found

# perturb a variable
tab$perturb(v = "total")
#> Count variable 'total' was perturbed.

# multiple variables can be perturbed as well
tab$perturb(v = c("cnt_males", "cnt_highincome"))
#> Error: No perturbation parameters found for variable 'cnt_males' - Please use method `$params_cnts_set().

# return weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))
#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 274411 4581 274470.9151
#>  2:  Total age_group1     total 1969 117928 1969 117928.0000
#>  3:  Total       ag1a     total 1969 117928 1969 117928.0000
#>  4:  Total age_group2     total 1143  68875 1144  68935.2581
#>  5:  Total       ag2a     total 1143  68875 1144  68935.2581
#>  6:  Total age_group3     total  864  51634  863  51574.2384
#>  7:  Total age_group4     total  423  24781  423  24781.0000
#>  8:  Total age_group5     total  168  10398  168  10398.0000
#>  9:  Total age_group6     total   13    795   13    795.0000
#> 10:   male      Total     total 2296 138591 2297 138651.3619
#> 11:      m      Total     total 2296 138591 2297 138651.3619
#> 12:   male age_group1     total 1015  60953 1015  60953.0000
#> 13:      m age_group1     total 1015  60953 1015  60953.0000
#> 14:   male       ag1a     total 1015  60953 1015  60953.0000
#> 15:      m       ag1a     total 1015  60953 1015  60953.0000
#> 16:   male age_group2     total  571  35520  570  35457.7933
#> 17:      m age_group2     total  571  35520  570  35457.7933
#> 18:   male       ag2a     total  571  35520  570  35457.7933
#> 19:      m       ag2a     total  571  35520  570  35457.7933
#> 20:   male age_group3     total  424  25445  422  25324.9764
#> 21:      m age_group3     total  424  25445  422  25324.9764
#> 22:   male age_group4     total  195  11088  197  11201.7231
#> 23:      m age_group4     total  195  11088  197  11201.7231
#> 24:   male age_group5     total   84   5171   85   5232.5595
#> 25:      m age_group5     total   84   5171   85   5232.5595
#> 26:   male age_group6     total    7    414    6    354.8571
#> 27:      m age_group6     total    7    414    6    354.8571
#> 28: female      Total     total 2284 135820 2282 135701.0683
#> 29:      f      Total     total 2284 135820 2282 135701.0683
#> 30: female age_group1     total  954  56975  955  57034.7222
#> 31:      f age_group1     total  954  56975  955  57034.7222
#> 32: female       ag1a     total  954  56975  955  57034.7222
#> 33:      f       ag1a     total  954  56975  955  57034.7222
#> 34: female age_group2     total  572  33355  572  33355.0000
#> 35:      f age_group2     total  572  33355  572  33355.0000
#> 36: female       ag2a     total  572  33355  572  33355.0000
#> 37:      f       ag2a     total  572  33355  572  33355.0000
#> 38: female age_group3     total  440  26189  439  26129.4795
#> 39:      f age_group3     total  440  26189  439  26129.4795
#> 40: female age_group4     total  228  13693  228  13693.0000
#> 41:      f age_group4     total  228  13693  228  13693.0000
#> 42: female age_group5     total   84   5227   84   5227.0000
#> 43:      f age_group5     total   84   5227   84   5227.0000
#> 44: female age_group6     total    6    381    6    381.0000
#> 45:      f age_group6     total    6    381    6    381.0000
#> 46:  Total      Total cnt_males 2296 138591   NA          NA
#> 47:  Total age_group1 cnt_males 1015  60953   NA          NA
#> 48:  Total       ag1a cnt_males 1015  60953   NA          NA
#> 49:  Total age_group2 cnt_males  571  35520   NA          NA
#> 50:  Total       ag2a cnt_males  571  35520   NA          NA
#> 51:  Total age_group3 cnt_males  424  25445   NA          NA
#> 52:  Total age_group4 cnt_males  195  11088   NA          NA
#> 53:  Total age_group5 cnt_males   84   5171   NA          NA
#> 54:  Total age_group6 cnt_males    7    414   NA          NA
#> 55:   male      Total cnt_males 2296 138591   NA          NA
#> 56:      m      Total cnt_males 2296 138591   NA          NA
#> 57:   male age_group1 cnt_males 1015  60953   NA          NA
#> 58:      m age_group1 cnt_males 1015  60953   NA          NA
#> 59:   male       ag1a cnt_males 1015  60953   NA          NA
#> 60:      m       ag1a cnt_males 1015  60953   NA          NA
#> 61:   male age_group2 cnt_males  571  35520   NA          NA
#> 62:      m age_group2 cnt_males  571  35520   NA          NA
#> 63:   male       ag2a cnt_males  571  35520   NA          NA
#> 64:      m       ag2a cnt_males  571  35520   NA          NA
#> 65:   male age_group3 cnt_males  424  25445   NA          NA
#> 66:      m age_group3 cnt_males  424  25445   NA          NA
#> 67:   male age_group4 cnt_males  195  11088   NA          NA
#> 68:      m age_group4 cnt_males  195  11088   NA          NA
#> 69:   male age_group5 cnt_males   84   5171   NA          NA
#> 70:      m age_group5 cnt_males   84   5171   NA          NA
#> 71:   male age_group6 cnt_males    7    414   NA          NA
#> 72:      m age_group6 cnt_males    7    414   NA          NA
#> 73: female      Total cnt_males    0      0   NA          NA
#> 74:      f      Total cnt_males    0      0   NA          NA
#> 75: female age_group1 cnt_males    0      0   NA          NA
#> 76:      f age_group1 cnt_males    0      0   NA          NA
#> 77: female       ag1a cnt_males    0      0   NA          NA
#> 78:      f       ag1a cnt_males    0      0   NA          NA
#> 79: female age_group2 cnt_males    0      0   NA          NA
#> 80:      f age_group2 cnt_males    0      0   NA          NA
#> 81: female       ag2a cnt_males    0      0   NA          NA
#> 82:      f       ag2a cnt_males    0      0   NA          NA
#> 83: female age_group3 cnt_males    0      0   NA          NA
#> 84:      f age_group3 cnt_males    0      0   NA          NA
#> 85: female age_group4 cnt_males    0      0   NA          NA
#> 86:      f age_group4 cnt_males    0      0   NA          NA
#> 87: female age_group5 cnt_males    0      0   NA          NA
#> 88:      f age_group5 cnt_males    0      0   NA          NA
#> 89: female age_group6 cnt_males    0      0   NA          NA
#> 90:      f age_group6 cnt_males    0      0   NA          NA
#>        sex        age     vname  uwc     wc puwc         pwc

# numerical variables (positive variables using flex-function)
# we also write the config to a yaml file
f_yaml <- tempfile(fileext = ".yaml")

# create a ptable using functionality from the ptable-pkg
# a single ptable for all cells
ptab1 <- ptable::pt_ex_nums(parity = TRUE, separation = FALSE)

# a single ptab for all cells except for very small ones
ptab2 <- ptable::pt_ex_nums(parity = TRUE, separation = TRUE)

# different ptables for cells with even/odd number of contributors
# and very small cells
ptab3 <- ptable::pt_ex_nums(parity = FALSE, separation = TRUE)

p_nums1 <- ck_params_nums(
  ptab = ptab1,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.30, 0.03),
    epsilon = c(1, 0.5, 0.2),
    q = 3),
  mu_c = 2,
  same_key = FALSE,
  use_zero_rkeys = FALSE,
  path = f_yaml)
#> yaml configuration '/tmp/RtmpRLPm23/file1e1b280127f8.yaml' successfully written.

# we read the parameters from the yaml-file
p_nums1 <- ck_read_yaml(path = f_yaml)

# for variables with positive and negative values
p_nums2 <- ck_params_nums(
  ptab = ptab2,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.15, 0.02),
    epsilon = c(1, 0.4, 0.15),
    q = 3),
  mu_c = 2,
  same_key = FALSE)

# simple perturbation parameters (not using the flex-function approach)
p_nums3 <- ck_params_nums(
  ptab = ptab3,
  type = "mean",
  mult_params = ck_simpleparams(p = 0.25),
  mu_c = 2,
  same_key = FALSE)

# use `p_nums1` for all variables
tab$params_nums_set(p_nums1, c("savings", "income", "expend"))
#> --> setting perturbation parameters for variable 'savings'
#> --> setting perturbation parameters for variable 'income'
#> --> setting perturbation parameters for variable 'expend'

# use different parameters for variable `mixed`
tab$params_nums_set(p_nums2, v = "mixed")
#> --> setting perturbation parameters for variable 'mixed'

# identify sensitive cells to which extra protection (`mu_c`) is added.
tab$supp_p(v = "income", p = 85)
#> computing contributing indices | rawdata <--> table; this might take a while
#> p%-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_pq(v = "income", p = 85, q = 90)
#> computing contributing indices | rawdata <--> table; this might take a while
#> pq-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_nk(v = "income", n = 2, k = 90)
#> computing contributing indices | rawdata <--> table; this might take a while
#> nk-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_freq(v = "income", n = 14, weighted = FALSE)
#> freq-rule: 5 new sensitive cells (incl. duplicates) found (total: 5)
tab$supp_val(v = "income", n = 10000, weighted = TRUE)
#> val-rule: 0 new sensitive cells (incl. duplicates) found (total: 5)
tab$supp_cells(
  v = "income",
  inp = data.frame(
    sex = c("female", "female"),
    "age" = c("age_group1", "age_group3")
  )
)
#> cell-rule: 2 new sensitive cells (incl. duplicates) found (total: 7)

# perturb variables
tab$perturb(v = c("income", "savings"))
#> Numeric variable 'income' was perturbed.
#> Numeric variable 'savings' was perturbed.

# extract results
tab$numtab("income", mean_before_sum = TRUE)
#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1371928688 1371848576
#>  2:  Total age_group1 income  9810547  584839258  584817119
#>  3:  Total       ag1a income  9810547  584839258  584817119
#>  4:  Total age_group2 income  5692119  342834108  342997095
#>  5:  Total       ag2a income  5692119  342834108  342997095
#>  6:  Total age_group3 income  4406946  262376581  262477948
#>  7:  Total age_group4 income  2133543  126427053  126441596
#>  8:  Total age_group5 income   848151   51616634   51504792
#>  9:  Total age_group6 income    61672    3835054    3951647
#> 10:   male      Total income 11262049  679676705  679974489
#> 11:      m      Total income 11262049  679676705  679974489
#> 12:   male age_group1 income  4877164  293559490  293649440
#> 13:      m age_group1 income  4877164  293559490  293649440
#> 14:   male       ag1a income  4877164  293559490  293649440
#> 15:      m       ag1a income  4877164  293559490  293649440
#> 16:   male age_group2 income  2811379  174772302  174666665
#> 17:      m age_group2 income  2811379  174772302  174666665
#> 18:   male       ag2a income  2811379  174772302  174666665
#> 19:      m       ag2a income  2811379  174772302  174666665
#> 20:   male age_group3 income  2168169  128413237  128406697
#> 21:      m age_group3 income  2168169  128413237  128406697
#> 22:   male age_group4 income   978510   56898674   56849156
#> 23:      m age_group4 income   978510   56898674   56849156
#> 24:   male age_group5 income   393134   24065735   24090458
#> 25:      m age_group5 income   393134   24065735   24090458
#> 26:   male age_group6 income    33693    1967267    1760307
#> 27:      m age_group6 income    33693    1967267    1760307
#> 28: female      Total income 11690929  692251983  692184130
#> 29:      f      Total income 11690929  692251983  692184130
#> 30: female age_group1 income  4933383  291279768  291304839
#> 31:      f age_group1 income  4933383  291279768  291304839
#> 32: female       ag1a income  4933383  291279768  291304839
#> 33:      f       ag1a income  4933383  291279768  291304839
#> 34: female age_group2 income  2880740  168061806  168184627
#> 35:      f age_group2 income  2880740  168061806  168184627
#> 36: female       ag2a income  2880740  168061806  168184627
#> 37:      f       ag2a income  2880740  168061806  168184627
#> 38: female age_group3 income  2238777  133963344  133842001
#> 39:      f age_group3 income  2238777  133963344  133842001
#> 40: female age_group4 income  1155033   69528379   69559660
#> 41:      f age_group4 income  1155033   69528379   69559660
#> 42: female age_group5 income   455017   27550899   27569309
#> 43:      f age_group5 income   455017   27550899   27569309
#> 44: female age_group6 income    27979    1867787    1871693
#> 45:      f age_group6 income    27979    1867787    1871693
#>        sex        age  vname      uws         ws        pws
tab$numtab("income", mean_before_sum = FALSE)
#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1371928688 1371888631
#>  2:  Total age_group1 income  9810547  584839258  584828188
#>  3:  Total       ag1a income  9810547  584839258  584828188
#>  4:  Total age_group2 income  5692119  342834108  342915592
#>  5:  Total       ag2a income  5692119  342834108  342915592
#>  6:  Total age_group3 income  4406946  262376581  262427260
#>  7:  Total age_group4 income  2133543  126427053  126434324
#>  8:  Total age_group5 income   848151   51616634   51560683
#>  9:  Total age_group6 income    61672    3835054    3892914
#> 10:   male      Total income 11262049  679676705  679825581
#> 11:      m      Total income 11262049  679676705  679825581
#> 12:   male age_group1 income  4877164  293559490  293604462
#> 13:      m age_group1 income  4877164  293559490  293604462
#> 14:   male       ag1a income  4877164  293559490  293604462
#> 15:      m       ag1a income  4877164  293559490  293604462
#> 16:   male age_group2 income  2811379  174772302  174719475
#> 17:      m age_group2 income  2811379  174772302  174719475
#> 18:   male       ag2a income  2811379  174772302  174719475
#> 19:      m       ag2a income  2811379  174772302  174719475
#> 20:   male age_group3 income  2168169  128413237  128409967
#> 21:      m age_group3 income  2168169  128413237  128409967
#> 22:   male age_group4 income   978510   56898674   56873909
#> 23:      m age_group4 income   978510   56898674   56873909
#> 24:   male age_group5 income   393134   24065735   24078093
#> 25:      m age_group5 income   393134   24065735   24078093
#> 26:   male age_group6 income    33693    1967267    1860912
#> 27:      m age_group6 income    33693    1967267    1860912
#> 28: female      Total income 11690929  692251983  692218056
#> 29:      f      Total income 11690929  692251983  692218056
#> 30: female age_group1 income  4933383  291279768  291292303
#> 31:      f age_group1 income  4933383  291279768  291292303
#> 32: female       ag1a income  4933383  291279768  291292303
#> 33:      f       ag1a income  4933383  291279768  291292303
#> 34: female age_group2 income  2880740  168061806  168123205
#> 35:      f age_group2 income  2880740  168061806  168123205
#> 36: female       ag2a income  2880740  168061806  168123205
#> 37:      f       ag2a income  2880740  168061806  168123205
#> 38: female age_group3 income  2238777  133963344  133902659
#> 39:      f age_group3 income  2238777  133963344  133902659
#> 40: female age_group4 income  1155033   69528379   69544017
#> 41:      f age_group4 income  1155033   69528379   69544017
#> 42: female age_group5 income   455017   27550899   27560102
#> 43:      f age_group5 income   455017   27550899   27560102
#> 44: female age_group6 income    27979    1867787    1869739
#> 45:      f age_group6 income    27979    1867787    1869739
#>        sex        age  vname      uws         ws        pws
tab$numtab("savings")
#>        sex        age   vname     uws        ws         pws
#>  1:  Total      Total savings 2273532 135938975 135941337.2
#>  2:  Total age_group1 savings  982386  58642146  58644115.8
#>  3:  Total       ag1a savings  982386  58642146  58644115.8
#>  4:  Total age_group2 savings  552336  33407455  33412651.9
#>  5:  Total       ag2a savings  552336  33407455  33412651.9
#>  6:  Total age_group3 savings  437101  26171097  26172983.2
#>  7:  Total age_group4 savings  214661  12336923  12336107.5
#>  8:  Total age_group5 savings   80451   4985855   4975747.3
#>  9:  Total age_group6 savings    6597    395499    397646.2
#> 10:   male      Total savings 1159816  70029298  70023452.0
#> 11:      m      Total savings 1159816  70029298  70023452.0
#> 12:   male age_group1 savings  517660  30906007  30908919.2
#> 13:      m age_group1 savings  517660  30906007  30908919.2
#> 14:   male       ag1a savings  517660  30906007  30908919.2
#> 15:      m       ag1a savings  517660  30906007  30908919.2
#> 16:   male age_group2 savings  280923  17818295  17814880.7
#> 17:      m age_group2 savings  280923  17818295  17814880.7
#> 18:   male       ag2a savings  280923  17818295  17814880.7
#> 19:      m       ag2a savings  280923  17818295  17814880.7
#> 20:   male age_group3 savings  214970  12913860  12908957.5
#> 21:      m age_group3 savings  214970  12913860  12908957.5
#> 22:   male age_group4 savings   99420   5516106   5515719.6
#> 23:      m age_group4 savings   99420   5516106   5515719.6
#> 24:   male age_group5 savings   43233   2679941   2681101.5
#> 25:      m age_group5 savings   43233   2679941   2681101.5
#> 26:   male age_group6 savings    3610    195089    188492.5
#> 27:      m age_group6 savings    3610    195089    188492.5
#> 28: female      Total savings 1113716  65909677  65908619.4
#> 29:      f      Total savings 1113716  65909677  65908619.4
#> 30: female age_group1 savings  464726  27736139  27740594.1
#> 31:      f age_group1 savings  464726  27736139  27740594.1
#> 32: female       ag1a savings  464726  27736139  27740594.1
#> 33:      f       ag1a savings  464726  27736139  27740594.1
#> 34: female age_group2 savings  271413  15589160  15595034.0
#> 35:      f age_group2 savings  271413  15589160  15595034.0
#> 36: female       ag2a savings  271413  15589160  15595034.0
#> 37:      f       ag2a savings  271413  15589160  15595034.0
#> 38: female age_group3 savings  222131  13257237  13251476.0
#> 39:      f age_group3 savings  222131  13257237  13251476.0
#> 40: female age_group4 savings  115241   6820817   6822357.0
#> 41:      f age_group4 savings  115241   6820817   6822357.0
#> 42: female age_group5 savings   37218   2305914   2305311.5
#> 43:      f age_group5 savings   37218   2305914   2305311.5
#> 44: female age_group6 savings    2987    200410    201240.7
#> 45:      f age_group6 savings    2987    200410    201240.7
#>        sex        age   vname     uws        ws         pws

# results can be resetted, too
tab$reset_cntvars(v = "cnt_males")
#> Error: Invalid variables specified in `v`: Possible choices are:
#>  'total'

# we can then set other parameters and perturb again
tab$params_cnts_set(val = p_cnts1, v = "cnt_males")
#> --> setting perturbation parameters for variable 'cnt_males'
tab$perturb(v = "cnt_males")
#> Count variable 'cnt_males' was perturbed.

# write to a file "outtab.csv" (.csv is automatically added to the path)
if (FALSE) {
tab$freqtab(v = c("total", "cnt_males"), path = "outtab.csv")
}

# show results containing weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))
#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 274411 4581 274470.9151
#>  2:  Total age_group1     total 1969 117928 1969 117928.0000
#>  3:  Total       ag1a     total 1969 117928 1969 117928.0000
#>  4:  Total age_group2     total 1143  68875 1144  68935.2581
#>  5:  Total       ag2a     total 1143  68875 1144  68935.2581
#>  6:  Total age_group3     total  864  51634  863  51574.2384
#>  7:  Total age_group4     total  423  24781  423  24781.0000
#>  8:  Total age_group5     total  168  10398  168  10398.0000
#>  9:  Total age_group6     total   13    795   13    795.0000
#> 10:   male      Total     total 2296 138591 2297 138651.3619
#> 11:      m      Total     total 2296 138591 2297 138651.3619
#> 12:   male age_group1     total 1015  60953 1015  60953.0000
#> 13:      m age_group1     total 1015  60953 1015  60953.0000
#> 14:   male       ag1a     total 1015  60953 1015  60953.0000
#> 15:      m       ag1a     total 1015  60953 1015  60953.0000
#> 16:   male age_group2     total  571  35520  570  35457.7933
#> 17:      m age_group2     total  571  35520  570  35457.7933
#> 18:   male       ag2a     total  571  35520  570  35457.7933
#> 19:      m       ag2a     total  571  35520  570  35457.7933
#> 20:   male age_group3     total  424  25445  422  25324.9764
#> 21:      m age_group3     total  424  25445  422  25324.9764
#> 22:   male age_group4     total  195  11088  197  11201.7231
#> 23:      m age_group4     total  195  11088  197  11201.7231
#> 24:   male age_group5     total   84   5171   85   5232.5595
#> 25:      m age_group5     total   84   5171   85   5232.5595
#> 26:   male age_group6     total    7    414    6    354.8571
#> 27:      m age_group6     total    7    414    6    354.8571
#> 28: female      Total     total 2284 135820 2282 135701.0683
#> 29:      f      Total     total 2284 135820 2282 135701.0683
#> 30: female age_group1     total  954  56975  955  57034.7222
#> 31:      f age_group1     total  954  56975  955  57034.7222
#> 32: female       ag1a     total  954  56975  955  57034.7222
#> 33:      f       ag1a     total  954  56975  955  57034.7222
#> 34: female age_group2     total  572  33355  572  33355.0000
#> 35:      f age_group2     total  572  33355  572  33355.0000
#> 36: female       ag2a     total  572  33355  572  33355.0000
#> 37:      f       ag2a     total  572  33355  572  33355.0000
#> 38: female age_group3     total  440  26189  439  26129.4795
#> 39:      f age_group3     total  440  26189  439  26129.4795
#> 40: female age_group4     total  228  13693  228  13693.0000
#> 41:      f age_group4     total  228  13693  228  13693.0000
#> 42: female age_group5     total   84   5227   84   5227.0000
#> 43:      f age_group5     total   84   5227   84   5227.0000
#> 44: female age_group6     total    6    381    6    381.0000
#> 45:      f age_group6     total    6    381    6    381.0000
#> 46:  Total      Total cnt_males 2296 138591 2297 138651.3619
#> 47:  Total age_group1 cnt_males 1015  60953 1015  60953.0000
#> 48:  Total       ag1a cnt_males 1015  60953 1015  60953.0000
#> 49:  Total age_group2 cnt_males  571  35520  570  35457.7933
#> 50:  Total       ag2a cnt_males  571  35520  570  35457.7933
#> 51:  Total age_group3 cnt_males  424  25445  422  25324.9764
#> 52:  Total age_group4 cnt_males  195  11088  197  11201.7231
#> 53:  Total age_group5 cnt_males   84   5171   85   5232.5595
#> 54:  Total age_group6 cnt_males    7    414    6    354.8571
#> 55:   male      Total cnt_males 2296 138591 2297 138651.3619
#> 56:      m      Total cnt_males 2296 138591 2297 138651.3619
#> 57:   male age_group1 cnt_males 1015  60953 1015  60953.0000
#> 58:      m age_group1 cnt_males 1015  60953 1015  60953.0000
#> 59:   male       ag1a cnt_males 1015  60953 1015  60953.0000
#> 60:      m       ag1a cnt_males 1015  60953 1015  60953.0000
#> 61:   male age_group2 cnt_males  571  35520  570  35457.7933
#> 62:      m age_group2 cnt_males  571  35520  570  35457.7933
#> 63:   male       ag2a cnt_males  571  35520  570  35457.7933
#> 64:      m       ag2a cnt_males  571  35520  570  35457.7933
#> 65:   male age_group3 cnt_males  424  25445  422  25324.9764
#> 66:      m age_group3 cnt_males  424  25445  422  25324.9764
#> 67:   male age_group4 cnt_males  195  11088  197  11201.7231
#> 68:      m age_group4 cnt_males  195  11088  197  11201.7231
#> 69:   male age_group5 cnt_males   84   5171   85   5232.5595
#> 70:      m age_group5 cnt_males   84   5171   85   5232.5595
#> 71:   male age_group6 cnt_males    7    414    6    354.8571
#> 72:      m age_group6 cnt_males    7    414    6    354.8571
#> 73: female      Total cnt_males    0      0    0      0.0000
#> 74:      f      Total cnt_males    0      0    0      0.0000
#> 75: female age_group1 cnt_males    0      0    0      0.0000
#> 76:      f age_group1 cnt_males    0      0    0      0.0000
#> 77: female       ag1a cnt_males    0      0    0      0.0000
#> 78:      f       ag1a cnt_males    0      0    0      0.0000
#> 79: female age_group2 cnt_males    0      0    0      0.0000
#> 80:      f age_group2 cnt_males    0      0    0      0.0000
#> 81: female       ag2a cnt_males    0      0    0      0.0000
#> 82:      f       ag2a cnt_males    0      0    0      0.0000
#> 83: female age_group3 cnt_males    0      0    0      0.0000
#> 84:      f age_group3 cnt_males    0      0    0      0.0000
#> 85: female age_group4 cnt_males    0      0    0      0.0000
#> 86:      f age_group4 cnt_males    0      0    0      0.0000
#> 87: female age_group5 cnt_males    0      0    0      0.0000
#> 88:      f age_group5 cnt_males    0      0    0      0.0000
#> 89: female age_group6 cnt_males    0      0    0      0.0000
#> 90:      f age_group6 cnt_males    0      0    0      0.0000
#>        sex        age     vname  uwc     wc puwc         pwc

# utility measures for a count variable
tab$measures_cnts(v = "total", exclude_zeros = TRUE)
#> $overview
#>    noise cnt        pct
#> 1:    -2   2 0.04444444
#> 2:    -1  11 0.24444444
#> 3:     0  19 0.42222222
#> 4:     1   9 0.20000000
#> 5:     2   4 0.08888889
#> 
#> $measures
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 0.000 0.000 0.000
#>  6:   Mean 0.711 0.008 0.023
#>  7: Median 1.000 0.001 0.015
#>  8:    Q60 1.000 0.001 0.016
#>  9:    Q70 1.000 0.002 0.021
#> 10:    Q80 1.000 0.002 0.024
#> 11:    Q90 2.000 0.010 0.054
#> 12:    Q95 2.000 0.012 0.071
#> 13:    Q99 2.000 0.143 0.196
#> 14:    Max 2.000 0.143 0.196
#> 
#> $cumdistr_d1
#>    cat cnt       pct
#> 1:   0  19 0.4222222
#> 2:   1  39 0.8666667
#> 3:   2  45 1.0000000
#> 
#> $cumdistr_d2
#>            cat cnt       pct
#> 1:    [0,0.02]  43 0.9555556
#> 2: (0.02,0.05]  43 0.9555556
#> 3:  (0.05,0.1]  43 0.9555556
#> 4:   (0.1,0.2]  45 1.0000000
#> 5:   (0.2,0.3]  45 1.0000000
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $cumdistr_d3
#>            cat cnt       pct
#> 1:    [0,0.02]  29 0.6444444
#> 2: (0.02,0.05]  39 0.8666667
#> 3:  (0.05,0.1]  43 0.9555556
#> 4:   (0.1,0.2]  45 1.0000000
#> 5:   (0.2,0.3]  45 1.0000000
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $false_zero
#> [1] 0
#> 
#> $false_nonzero
#> [1] 0
#> 
#> $exclude_zeros
#> [1] TRUE
#> 

# modifications for perturbed count variables
tab$mod_cnts()
#>        sex        age row_nr pert       ckey  countvar
#>  1:  Total      Total     16    1 0.77291236     total
#>  2:  Total age_group1     15    0 0.35236126     total
#>  3:  Total       ag1a     15    0 0.35236126     total
#>  4:  Total age_group2     16    1 0.79610146     total
#>  5:  Total       ag2a     16    1 0.79610146     total
#>  6:  Total age_group3     14   -1 0.13878487     total
#>  7:  Total age_group4     15    0 0.52695132     total
#>  8:  Total age_group5     15    0 0.40035003     total
#>  9:  Total age_group6     15    0 0.55836342     total
#> 10:   male      Total     16    1 0.75012904     total
#> 11:      m      Total     16    1 0.75012904     total
#> 12:   male age_group1     15    0 0.55934594     total
#> 13:      m age_group1     15    0 0.55934594     total
#> 14:   male       ag1a     15    0 0.55934594     total
#> 15:      m       ag1a     15    0 0.55934594     total
#> 16:   male age_group2     14   -1 0.20698017     total
#> 17:      m age_group2     14   -1 0.20698017     total
#> 18:   male       ag2a     14   -1 0.20698017     total
#> 19:      m       ag2a     14   -1 0.20698017     total
#> 20:   male age_group3     13   -2 0.03066701     total
#> 21:      m age_group3     13   -2 0.03066701     total
#> 22:   male age_group4     17    2 0.96121125     total
#> 23:      m age_group4     17    2 0.96121125     total
#> 24:   male age_group5     16    1 0.77185442     total
#> 25:      m age_group5     16    1 0.77185442     total
#> 26:   male age_group6     14   -1 0.22007025     total
#> 27:      m age_group6     14   -1 0.22007025     total
#> 28: female      Total     13   -2 0.02278332     total
#> 29:      f      Total     13   -2 0.02278332     total
#> 30: female age_group1     16    1 0.79301532     total
#> 31:      f age_group1     16    1 0.79301532     total
#> 32: female       ag1a     16    1 0.79301532     total
#> 33:      f       ag1a     16    1 0.79301532     total
#> 34: female age_group2     15    0 0.58912129     total
#> 35:      f age_group2     15    0 0.58912129     total
#> 36: female       ag2a     15    0 0.58912129     total
#> 37:      f       ag2a     15    0 0.58912129     total
#> 38: female age_group3     14   -1 0.10811786     total
#> 39:      f age_group3     14   -1 0.10811786     total
#> 40: female age_group4     15    0 0.56574007     total
#> 41:      f age_group4     15    0 0.56574007     total
#> 42: female age_group5     15    0 0.62849561     total
#> 43:      f age_group5     15    0 0.62849561     total
#> 44: female age_group6     15    0 0.33829317     total
#> 45:      f age_group6     15    0 0.33829317     total
#> 46:  Total      Total     16    1 0.75012904 cnt_males
#> 47:  Total age_group1     15    0 0.55934594 cnt_males
#> 48:  Total       ag1a     15    0 0.55934594 cnt_males
#> 49:  Total age_group2     14   -1 0.20698017 cnt_males
#> 50:  Total       ag2a     14   -1 0.20698017 cnt_males
#> 51:  Total age_group3     13   -2 0.03066701 cnt_males
#> 52:  Total age_group4     17    2 0.96121125 cnt_males
#> 53:  Total age_group5     16    1 0.77185442 cnt_males
#> 54:  Total age_group6     14   -1 0.22007025 cnt_males
#> 55:   male      Total     16    1 0.75012904 cnt_males
#> 56:      m      Total     16    1 0.75012904 cnt_males
#> 57:   male age_group1     15    0 0.55934594 cnt_males
#> 58:      m age_group1     15    0 0.55934594 cnt_males
#> 59:   male       ag1a     15    0 0.55934594 cnt_males
#> 60:      m       ag1a     15    0 0.55934594 cnt_males
#> 61:   male age_group2     14   -1 0.20698017 cnt_males
#> 62:      m age_group2     14   -1 0.20698017 cnt_males
#> 63:   male       ag2a     14   -1 0.20698017 cnt_males
#> 64:      m       ag2a     14   -1 0.20698017 cnt_males
#> 65:   male age_group3     13   -2 0.03066701 cnt_males
#> 66:      m age_group3     13   -2 0.03066701 cnt_males
#> 67:   male age_group4     17    2 0.96121125 cnt_males
#> 68:      m age_group4     17    2 0.96121125 cnt_males
#> 69:   male age_group5     16    1 0.77185442 cnt_males
#> 70:      m age_group5     16    1 0.77185442 cnt_males
#> 71:   male age_group6     14   -1 0.22007025 cnt_males
#> 72:      m age_group6     14   -1 0.22007025 cnt_males
#> 73: female      Total     -1    0 0.00000000 cnt_males
#> 74:      f      Total     -1    0 0.00000000 cnt_males
#> 75: female age_group1     -1    0 0.00000000 cnt_males
#> 76:      f age_group1     -1    0 0.00000000 cnt_males
#> 77: female       ag1a     -1    0 0.00000000 cnt_males
#> 78:      f       ag1a     -1    0 0.00000000 cnt_males
#> 79: female age_group2     -1    0 0.00000000 cnt_males
#> 80:      f age_group2     -1    0 0.00000000 cnt_males
#> 81: female       ag2a     -1    0 0.00000000 cnt_males
#> 82:      f       ag2a     -1    0 0.00000000 cnt_males
#> 83: female age_group3     -1    0 0.00000000 cnt_males
#> 84:      f age_group3     -1    0 0.00000000 cnt_males
#> 85: female age_group4     -1    0 0.00000000 cnt_males
#> 86:      f age_group4     -1    0 0.00000000 cnt_males
#> 87: female age_group5     -1    0 0.00000000 cnt_males
#> 88:      f age_group5     -1    0 0.00000000 cnt_males
#> 89: female age_group6     -1    0 0.00000000 cnt_males
#> 90:      f age_group6     -1    0 0.00000000 cnt_males
#>        sex        age row_nr pert       ckey  countvar

# display a summary about utility measures
tab$summary()
#> ┌──────────────────────────────────────────────┐
#> │Utility measures for perturbed count variables│
#> └──────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#>     countvar Min Q10 Q20 Q30 Q40   Mean Median Q60 Q70 Q80 Q90 Q95 Q99 Max
#> 1:     total  -2  -1  -1   0   0 -0.044      0   0   0 1.0   1 1.0   2   2
#> 2: cnt_males  -2  -1  -1   0   0 -0.067      0   0   0 0.2   1 1.8   2   2
#> 
#> ── Distance-based measures ─────────────────────────────────────────────────────
#> ✔ Variable: 'total'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 0.000 0.000 0.000
#>  6:   Mean 0.711 0.008 0.023
#>  7: Median 1.000 0.001 0.015
#>  8:    Q60 1.000 0.001 0.016
#>  9:    Q70 1.000 0.002 0.021
#> 10:    Q80 1.000 0.002 0.024
#> 11:    Q90 2.000 0.010 0.054
#> 12:    Q95 2.000 0.012 0.071
#> 13:    Q99 2.000 0.143 0.196
#> 14:    Max 2.000 0.143 0.196
#> 
#> ✔ Variable: 'cnt_males'
#> 
#>       what  d1    d2    d3
#>  1:    Min 0.0 0.000 0.000
#>  2:    Q10 0.0 0.000 0.000
#>  3:    Q20 0.2 0.000 0.002
#>  4:    Q30 1.0 0.000 0.010
#>  5:    Q40 1.0 0.002 0.021
#>  6:   Mean 1.0 0.019 0.047
#>  7: Median 1.0 0.002 0.021
#>  8:    Q60 1.0 0.005 0.049
#>  9:    Q70 1.0 0.010 0.054
#> 10:    Q80 1.8 0.012 0.068
#> 11:    Q90 2.0 0.064 0.121
#> 12:    Q95 2.0 0.143 0.196
#> 13:    Q99 2.0 0.143 0.196
#> 14:    Max 2.0 0.143 0.196
#> 
#> ┌──────────────────────────────────────────────────┐
#> │Utility measures for perturbed numerical variables│
#> └──────────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#>      vname       Min        Q10        Q20        Q30      Q40     Mean
#> 1:  expend       Inf         NA         NA         NA       NA      NaN
#> 2:  income -106355.1 -54701.346 -42610.573 -22025.548 -136.984 7622.237
#> 3: savings  -10107.7  -5812.003  -3711.919  -1057.555 -472.819  130.476
#> 4:   mixed       Inf         NA         NA         NA       NA      NaN
#>      Median       Q60       Q70       Q80       Q90       Q95        Q99
#> 1:       NA        NA        NA        NA        NA        NA         NA
#> 2: 9203.400 12535.217 39105.073 52114.813 61399.466 81483.942 148875.911
#> 3: 1160.505  1919.663  2802.196  4455.082  5196.876  5874.015   5874.015
#> 4:       NA        NA        NA        NA        NA        NA         NA
#>           Max
#> 1:       -Inf
#> 2: 148875.911
#> 3:   5874.015
#> 4:       -Inf