Create perturbation parameters for count variables

This function allows to generate required perturbation parameters that are used to perturb count variables.

ck_params_cnts(ptab, path = NULL)

Arguments

ptab
an object created with ptable::pt_create_pParams(), ptable::pt_create_pTable() or ptable::create_cnt_ptable()

path
a scalar character specifying a path to which the parameters created with this functions should be written to (in yaml format)

Value

an object suitable as input to method $params_cnts_set() for the perturbation of counts and frequencies.

See also

Examples

x <- ck_create_testdata()

# create some 0/1 variables that should be perturbed later
x[, cnt_females := ifelse(sex == "male", 0, 1)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              76          25.00000           0
#>    2:   2530      28       1              39          25.00000           1
#>    3:   6920     550       1              29          25.00000           0
#>    4:   7960     870       1              75          25.00000           0
#>    5:   9030      20       2              89          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              51          16.66667           1
#> 4577:   1420     987    1000              87          16.66667           0
#> 4578:   8900     684    1000              23          16.66667           0
#> 4579:   3880     294    1000              52          16.66667           1
#> 4580:   4830     911    1000              75          16.66667           0

x[, cnt_males := ifelse(sex == "male", 1, 0)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              76          25.00000           0
#>    2:   2530      28       1              39          25.00000           1
#>    3:   6920     550       1              29          25.00000           0
#>    4:   7960     870       1              75          25.00000           0
#>    5:   9030      20       2              89          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              51          16.66667           1
#> 4577:   1420     987    1000              87          16.66667           0
#> 4578:   8900     684    1000              23          16.66667           0
#> 4579:   3880     294    1000              52          16.66667           1
#> 4580:   4830     911    1000              75          16.66667           0
#>       cnt_males
#>    1:         1
#>    2:         0
#>    3:         1
#>    4:         1
#>    5:         1
#>   ---          
#> 4576:         0
#> 4577:         1
#> 4578:         1
#> 4579:         0
#> 4580:         1

x[, cnt_highincome := ifelse(income >= 9000, 1, 0)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              76          25.00000           0
#>    2:   2530      28       1              39          25.00000           1
#>    3:   6920     550       1              29          25.00000           0
#>    4:   7960     870       1              75          25.00000           0
#>    5:   9030      20       2              89          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              51          16.66667           1
#> 4577:   1420     987    1000              87          16.66667           0
#> 4578:   8900     684    1000              23          16.66667           0
#> 4579:   3880     294    1000              52          16.66667           1
#> 4580:   4830     911    1000              75          16.66667           0
#>       cnt_males cnt_highincome
#>    1:         1              0
#>    2:         0              0
#>    3:         1              0
#>    4:         1              0
#>    5:         1              1
#>   ---                         
#> 4576:         0              0
#> 4577:         1              0
#> 4578:         1              0
#> 4579:         0              0
#> 4580:         1              0

# a variable with positive and negative contributions
x[, mixed := sample(-10:10, nrow(x), replace = TRUE)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              76          25.00000           0
#>    2:   2530      28       1              39          25.00000           1
#>    3:   6920     550       1              29          25.00000           0
#>    4:   7960     870       1              75          25.00000           0
#>    5:   9030      20       2              89          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              51          16.66667           1
#> 4577:   1420     987    1000              87          16.66667           0
#> 4578:   8900     684    1000              23          16.66667           0
#> 4579:   3880     294    1000              52          16.66667           1
#> 4580:   4830     911    1000              75          16.66667           0
#>       cnt_males cnt_highincome mixed
#>    1:         1              0    -6
#>    2:         0              0    10
#>    3:         1              0     9
#>    4:         1              0     0
#>    5:         1              1     7
#>   ---                               
#> 4576:         0              0     5
#> 4577:         1              0    -4
#> 4578:         1              0    10
#> 4579:         0              0     0
#> 4580:         1              0    10

# create record keys
x$rkey <- ck_generate_rkeys(dat = x)

# define required inputs

# hierarchy with some bogus codes
d_sex <- hier_create(root = "Total", nodes = c("male", "female"))
d_sex <- hier_add(d_sex, root = "female", "f")
d_sex <- hier_add(d_sex, root = "male", "m")

d_age <- hier_create(root = "Total", nodes = paste0("age_group", 1:6))
d_age <- hier_add(d_age, root = "age_group1", "ag1a")
d_age <- hier_add(d_age, root = "age_group2", "ag2a")

# define the cell key object
countvars <- c("cnt_females", "cnt_males", "cnt_highincome")
numvars <- c("expend", "income", "savings", "mixed")
tab <- ck_setup(
  x = x,
  rkey = "rkey",
  dims = list(sex = d_sex, age = d_age),
  w = "sampling_weight",
  countvars = countvars,
  numvars = numvars)

# show some information about this table instance
tab$print() # identical with print(tab)

#> ── Table Information ───────────────────────────────────────────────────────────
#> ✔ 45 cells in 2 dimensions ('sex', 'age')
#> ✔ weights: yes
#> ── Tabulated / Perturbed countvars ─────────────────────────────────────────────
#> ☐ 'total'
#> ☐ 'cnt_females'
#> ☐ 'cnt_males'
#> ☐ 'cnt_highincome'
#> ── Tabulated / Perturbed numvars ───────────────────────────────────────────────
#> ☐ 'expend'
#> ☐ 'income'
#> ☐ 'savings'
#> ☐ 'mixed'

# information about the hierarchies
tab$hierarchy_info()

#> $sex
#>      code level is_leaf parent
#> 1:  Total     1   FALSE   <NA>
#> 2:   male     2   FALSE  Total
#> 3: female     2   FALSE  Total
#> 4:      f     3    TRUE female
#> 5:      m     3    TRUE   male
#> 
#> $age
#>          code level is_leaf     parent
#> 1:      Total     1   FALSE       <NA>
#> 2: age_group1     2   FALSE      Total
#> 3: age_group2     2   FALSE      Total
#> 4: age_group3     2    TRUE      Total
#> 5: age_group4     2    TRUE      Total
#> 6: age_group5     2    TRUE      Total
#> 7: age_group6     2    TRUE      Total
#> 8:       ag1a     3    TRUE age_group1
#> 9:       ag2a     3    TRUE age_group2
#> 

# which variables have been defined?
tab$allvars()

#> $cntvars
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"
#> 
#> $numvars
#> [1] "expend"  "income"  "savings" "mixed"  
#> 

# count variables
tab$cntvars()

#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"

# continuous variables
tab$numvars()

#> [1] "expend"  "income"  "savings" "mixed"  

# create perturbation parameters for "total" variable and
# write to yaml-file

# create a ptable using functionality from the ptable-pkg
f_yaml <- tempfile(fileext = ".yaml")
p_cnts1 <- ck_params_cnts(
  ptab = ptable::pt_ex_cnts(),
  path = f_yaml)

#> yaml configuration '/tmp/Rtmp09UBN4/file3bd45cc3483c.yaml' successfully written.

# read parameters from yaml-file and set them for variable `"total"`
p_cnts1 <- ck_read_yaml(path = f_yaml)

tab$params_cnts_set(val = p_cnts1, v = "total")

#> --> setting perturbation parameters for variable 'total'

# create alternative perturbation parameters by specifying parameters
para2 <- ptable::pt_create_pParams(
  D = 8, V = 3, js = 2, table = "cnts")

#> Ignoring argument `step` for freqency tables

p_cnts2 <- ck_params_cnts(ptab = para2)

# use these ptable it for the remaining variables
tab$params_cnts_set(val = p_cnts2, v = countvars)

#> --> setting perturbation parameters for variable 'cnt_females'

#> --> setting perturbation parameters for variable 'cnt_males'

#> --> setting perturbation parameters for variable 'cnt_highincome'

# perturb a variable
tab$perturb(v = "total")

#> Count variable 'total' was perturbed.

# multiple variables can be perturbed as well
tab$perturb(v = c("cnt_males", "cnt_highincome"))

#> Count variable 'cnt_males' was perturbed.

#> Count variable 'cnt_highincome' was perturbed.

# return weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))

#>        sex        age     vname  uwc     wc puwc        pwc
#>  1:  Total      Total     total 4580 272971 4579 272911.399
#>  2:  Total age_group1     total 1969 118179 1970 118239.020
#>  3:  Total       ag1a     total 1969 118179 1970 118239.020
#>  4:  Total age_group2     total 1143  67260 1145  67377.690
#>  5:  Total       ag2a     total 1143  67260 1145  67377.690
#>  6:  Total age_group3     total  864  51608  865  51667.731
#>  7:  Total age_group4     total  423  24845  424  24903.735
#>  8:  Total age_group5     total  168  10250  170  10372.024
#>  9:  Total age_group6     total   13    829   13    829.000
#> 10: female      Total     total 2284 136454 2284 136454.000
#> 11:      f      Total     total 2284 136454 2284 136454.000
#> 12: female age_group1     total  954  57080  953  57020.168
#> 13:      f age_group1     total  954  57080  953  57020.168
#> 14: female       ag1a     total  954  57080  953  57020.168
#> 15:      f       ag1a     total  954  57080  953  57020.168
#> 16: female age_group2     total  572  34218  573  34277.822
#> 17:      f age_group2     total  572  34218  573  34277.822
#> 18: female       ag2a     total  572  34218  573  34277.822
#> 19:      f       ag2a     total  572  34218  573  34277.822
#> 20: female age_group3     total  440  26277  440  26277.000
#> 21:      f age_group3     total  440  26277  440  26277.000
#> 22: female age_group4     total  228  13374  227  13315.342
#> 23:      f age_group4     total  228  13374  227  13315.342
#> 24: female age_group5     total   84   5145   83   5083.750
#> 25:      f age_group5     total   84   5145   83   5083.750
#> 26: female age_group6     total    6    360    7    420.000
#> 27:      f age_group6     total    6    360    7    420.000
#> 28:   male      Total     total 2296 136517 2296 136517.000
#> 29:      m      Total     total 2296 136517 2296 136517.000
#> 30:   male age_group1     total 1015  61099 1016  61159.196
#> 31:      m age_group1     total 1015  61099 1016  61159.196
#> 32:   male       ag1a     total 1015  61099 1016  61159.196
#> 33:      m       ag1a     total 1015  61099 1016  61159.196
#> 34:   male age_group2     total  571  33042  570  32984.133
#> 35:      m age_group2     total  571  33042  570  32984.133
#> 36:   male       ag2a     total  571  33042  570  32984.133
#> 37:      m       ag2a     total  571  33042  570  32984.133
#> 38:   male age_group3     total  424  25331  424  25331.000
#> 39:      m age_group3     total  424  25331  424  25331.000
#> 40:   male age_group4     total  195  11471  196  11529.826
#> 41:      m age_group4     total  195  11471  196  11529.826
#> 42:   male age_group5     total   84   5105   85   5165.774
#> 43:      m age_group5     total   84   5105   85   5165.774
#> 44:   male age_group6     total    7    469    8    536.000
#> 45:      m age_group6     total    7    469    8    536.000
#> 46:  Total      Total cnt_males 2296 136517 2297 136576.459
#> 47:  Total age_group1 cnt_males 1015  61099 1016  61159.196
#> 48:  Total       ag1a cnt_males 1015  61099 1016  61159.196
#> 49:  Total age_group2 cnt_males  571  33042  569  32926.266
#> 50:  Total       ag2a cnt_males  571  33042  569  32926.266
#> 51:  Total age_group3 cnt_males  424  25331  424  25331.000
#> 52:  Total age_group4 cnt_males  195  11471  196  11529.826
#> 53:  Total age_group5 cnt_males   84   5105   86   5226.548
#> 54:  Total age_group6 cnt_males    7    469    8    536.000
#> 55: female      Total cnt_males    0      0    0      0.000
#> 56:      f      Total cnt_males    0      0    0      0.000
#> 57: female age_group1 cnt_males    0      0    0      0.000
#> 58:      f age_group1 cnt_males    0      0    0      0.000
#> 59: female       ag1a cnt_males    0      0    0      0.000
#> 60:      f       ag1a cnt_males    0      0    0      0.000
#> 61: female age_group2 cnt_males    0      0    0      0.000
#> 62:      f age_group2 cnt_males    0      0    0      0.000
#> 63: female       ag2a cnt_males    0      0    0      0.000
#> 64:      f       ag2a cnt_males    0      0    0      0.000
#> 65: female age_group3 cnt_males    0      0    0      0.000
#> 66:      f age_group3 cnt_males    0      0    0      0.000
#> 67: female age_group4 cnt_males    0      0    0      0.000
#> 68:      f age_group4 cnt_males    0      0    0      0.000
#> 69: female age_group5 cnt_males    0      0    0      0.000
#> 70:      f age_group5 cnt_males    0      0    0      0.000
#> 71: female age_group6 cnt_males    0      0    0      0.000
#> 72:      f age_group6 cnt_males    0      0    0      0.000
#> 73:   male      Total cnt_males 2296 136517 2297 136576.459
#> 74:      m      Total cnt_males 2296 136517 2297 136576.459
#> 75:   male age_group1 cnt_males 1015  61099 1016  61159.196
#> 76:      m age_group1 cnt_males 1015  61099 1016  61159.196
#> 77:   male       ag1a cnt_males 1015  61099 1016  61159.196
#> 78:      m       ag1a cnt_males 1015  61099 1016  61159.196
#> 79:   male age_group2 cnt_males  571  33042  569  32926.266
#> 80:      m age_group2 cnt_males  571  33042  569  32926.266
#> 81:   male       ag2a cnt_males  571  33042  569  32926.266
#> 82:      m       ag2a cnt_males  571  33042  569  32926.266
#> 83:   male age_group3 cnt_males  424  25331  424  25331.000
#> 84:      m age_group3 cnt_males  424  25331  424  25331.000
#> 85:   male age_group4 cnt_males  195  11471  196  11529.826
#> 86:      m age_group4 cnt_males  195  11471  196  11529.826
#> 87:   male age_group5 cnt_males   84   5105   86   5226.548
#> 88:      m age_group5 cnt_males   84   5105   86   5226.548
#> 89:   male age_group6 cnt_males    7    469    8    536.000
#> 90:      m age_group6 cnt_males    7    469    8    536.000
#>        sex        age     vname  uwc     wc puwc        pwc

# numerical variables (positive variables using flex-function)
# we also write the config to a yaml file
f_yaml <- tempfile(fileext = ".yaml")

# create a ptable using functionality from the ptable-pkg
# a single ptable for all cells
ptab1 <- ptable::pt_ex_nums(parity = TRUE, separation = FALSE)

# a single ptab for all cells except for very small ones
ptab2 <- ptable::pt_ex_nums(parity = TRUE, separation = TRUE)

# different ptables for cells with even/odd number of contributors
# and very small cells
ptab3 <- ptable::pt_ex_nums(parity = FALSE, separation = TRUE)

p_nums1 <- ck_params_nums(
  ptab = ptab1,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.30, 0.03),
    epsilon = c(1, 0.5, 0.2),
    q = 3),
  mu_c = 2,
  same_key = FALSE,
  use_zero_rkeys = FALSE,
  path = f_yaml)

#> yaml configuration '/tmp/Rtmp09UBN4/file3bd44426d978.yaml' successfully written.

# we read the parameters from the yaml-file
p_nums1 <- ck_read_yaml(path = f_yaml)

# for variables with positive and negative values
p_nums2 <- ck_params_nums(
  ptab = ptab2,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.15, 0.02),
    epsilon = c(1, 0.4, 0.15),
    q = 3),
  mu_c = 2,
  same_key = FALSE)

# simple perturbation parameters (not using the flex-function approach)
p_nums3 <- ck_params_nums(
  ptab = ptab3,
  type = "mean",
  mult_params = ck_simpleparams(p = 0.25),
  mu_c = 2,
  same_key = FALSE)

# use `p_nums1` for all variables
tab$params_nums_set(p_nums1, c("savings", "income", "expend"))

#> --> setting perturbation parameters for variable 'savings'

#> --> setting perturbation parameters for variable 'income'

#> --> setting perturbation parameters for variable 'expend'

# use different parameters for variable `mixed`
tab$params_nums_set(p_nums2, v = "mixed")

#> --> setting perturbation parameters for variable 'mixed'

# identify sensitive cells to which extra protection (`mu_c`) is added.
tab$supp_p(v = "income", p = 85)

#> p%-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)

tab$supp_pq(v = "income", p = 85, q = 90)

#> pq-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)

tab$supp_nk(v = "income", n = 2, k = 90)

#> nk-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)

tab$supp_freq(v = "income", n = 14, weighted = FALSE)

#> freq-rule: 5 new sensitive cells (incl. duplicates) found (total: 5)

tab$supp_val(v = "income", n = 10000, weighted = TRUE)

#> val-rule: 0 new sensitive cells (incl. duplicates) found (total: 5)

# perturb variables
tab$perturb(v = c("income", "savings"))

#> Numeric variable 'income' was perturbed.

#> Numeric variable 'savings' was perturbed.

# extract results
tab$numtab("income", mean_before_sum = TRUE)

#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1371638432 1371695139
#>  2:  Total age_group1 income  9810547  586633532  586691779
#>  3:  Total       ag1a income  9810547  586633532  586691779
#>  4:  Total age_group2 income  5692119  336714461  336740882
#>  5:  Total       ag2a income  5692119  336714461  336740882
#>  6:  Total age_group3 income  4406946  264659634  264542706
#>  7:  Total age_group4 income  2133543  127077769  127201441
#>  8:  Total age_group5 income   848151   52585564   52588366
#>  9:  Total age_group6 income    61672    3967472    3880936
#> 10: female      Total income 11690929  700814534  700695219
#> 11:      f      Total income 11690929  700814534  700695219
#> 12: female age_group1 income  4933383  294104115  293977437
#> 13:      f age_group1 income  4933383  294104115  293977437
#> 14: female       ag1a income  4933383  294104115  293977437
#> 15:      f       ag1a income  4933383  294104115  293977437
#> 16: female age_group2 income  2880740  173955500  173931564
#> 17:      f age_group2 income  2880740  173955500  173931564
#> 18: female       ag2a income  2880740  173955500  173931564
#> 19:      f       ag2a income  2880740  173955500  173931564
#> 20: female age_group3 income  2238777  134826570  134881878
#> 21:      f age_group3 income  2238777  134826570  134881878
#> 22: female age_group4 income  1155033   68529559   68493971
#> 23:      f age_group4 income  1155033   68529559   68493971
#> 24: female age_group5 income   455017   27618184   27507531
#> 25:      f age_group5 income   455017   27618184   27507531
#> 26: female age_group6 income    27979    1780606    1737452
#> 27:      f age_group6 income    27979    1780606    1737452
#> 28:   male      Total income 11262049  670823898  670766071
#> 29:      m      Total income 11262049  670823898  670766071
#> 30:   male age_group1 income  4877164  292529417  292547211
#> 31:      m age_group1 income  4877164  292529417  292547211
#> 32:   male       ag1a income  4877164  292529417  292547211
#> 33:      m       ag1a income  4877164  292529417  292547211
#> 34:   male age_group2 income  2811379  162758961  162605994
#> 35:      m age_group2 income  2811379  162758961  162605994
#> 36:   male       ag2a income  2811379  162758961  162605994
#> 37:      m       ag2a income  2811379  162758961  162605994
#> 38:   male age_group3 income  2168169  129833064  129731892
#> 39:      m age_group3 income  2168169  129833064  129731892
#> 40:   male age_group4 income   978510   58548210   58518430
#> 41:      m age_group4 income   978510   58548210   58518430
#> 42:   male age_group5 income   393134   24967380   25014647
#> 43:      m age_group5 income   393134   24967380   25014647
#> 44:   male age_group6 income    33693    2186866    1960967
#> 45:      m age_group6 income    33693    2186866    1960967
#>        sex        age  vname      uws         ws        pws

tab$numtab("income", mean_before_sum = FALSE)

#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1371638432 1371666785
#>  2:  Total age_group1 income  9810547  586633532  586662655
#>  3:  Total       ag1a income  9810547  586633532  586662655
#>  4:  Total age_group2 income  5692119  336714461  336727671
#>  5:  Total       ag2a income  5692119  336714461  336727671
#>  6:  Total age_group3 income  4406946  264659634  264601163
#>  7:  Total age_group4 income  2133543  127077769  127139590
#>  8:  Total age_group5 income   848151   52585564   52586965
#>  9:  Total age_group6 income    61672    3967472    3923965
#> 10: female      Total income 11690929  700814534  700754874
#> 11:      f      Total income 11690929  700814534  700754874
#> 12: female age_group1 income  4933383  294104115  294040769
#> 13:      f age_group1 income  4933383  294104115  294040769
#> 14: female       ag1a income  4933383  294104115  294040769
#> 15:      f       ag1a income  4933383  294104115  294040769
#> 16: female age_group2 income  2880740  173955500  173943532
#> 17:      f age_group2 income  2880740  173955500  173943532
#> 18: female       ag2a income  2880740  173955500  173943532
#> 19:      f       ag2a income  2880740  173955500  173943532
#> 20: female age_group3 income  2238777  134826570  134854221
#> 21:      f age_group3 income  2238777  134826570  134854221
#> 22: female age_group4 income  1155033   68529559   68511763
#> 23:      f age_group4 income  1155033   68529559   68511763
#> 24: female age_group5 income   455017   27618184   27562802
#> 25:      f age_group5 income   455017   27618184   27562802
#> 26: female age_group6 income    27979    1780606    1758897
#> 27:      f age_group6 income    27979    1780606    1758897
#> 28:   male      Total income 11262049  670823898  670794984
#> 29:      m      Total income 11262049  670823898  670794984
#> 30:   male age_group1 income  4877164  292529417  292538314
#> 31:      m age_group1 income  4877164  292529417  292538314
#> 32:   male       ag1a income  4877164  292529417  292538314
#> 33:      m       ag1a income  4877164  292529417  292538314
#> 34:   male age_group2 income  2811379  162758961  162682459
#> 35:      m age_group2 income  2811379  162758961  162682459
#> 36:   male       ag2a income  2811379  162758961  162682459
#> 37:      m       ag2a income  2811379  162758961  162682459
#> 38:   male age_group3 income  2168169  129833064  129782468
#> 39:      m age_group3 income  2168169  129833064  129782468
#> 40:   male age_group4 income   978510   58548210   58533318
#> 41:      m age_group4 income   978510   58548210   58533318
#> 42:   male age_group5 income   393134   24967380   24991002
#> 43:      m age_group5 income   393134   24967380   24991002
#> 44:   male age_group6 income    33693    2186866    2070839
#> 45:      m age_group6 income    33693    2186866    2070839
#>        sex        age  vname      uws         ws        pws

tab$numtab("savings")

#>        sex        age   vname     uws        ws         pws
#>  1:  Total      Total savings 2273532 135507254 135504592.3
#>  2:  Total age_group1 savings  982386  59106062  59109282.4
#>  3:  Total       ag1a savings  982386  59106062  59109282.4
#>  4:  Total age_group2 savings  552336  32351896  32358207.4
#>  5:  Total       ag2a savings  552336  32351896  32358207.4
#>  6:  Total age_group3 savings  437101  26152047  26153129.5
#>  7:  Total age_group4 savings  214661  12602844  12605554.8
#>  8:  Total age_group5 savings   80451   4836239   4836680.9
#>  9:  Total age_group6 savings    6597    458166    456611.7
#> 10: female      Total savings 1113716  66502088  66499940.2
#> 11:      f      Total savings 1113716  66502088  66499940.2
#> 12: female age_group1 savings  464726  27757278  27754601.5
#> 13:      f age_group1 savings  464726  27757278  27754601.5
#> 14: female       ag1a savings  464726  27757278  27754601.5
#> 15:      f       ag1a savings  464726  27757278  27754601.5
#> 16: female age_group2 savings  271413  16242443  16242398.8
#> 17:      f age_group2 savings  271413  16242443  16242398.8
#> 18: female       ag2a savings  271413  16242443  16242398.8
#> 19:      f       ag2a savings  271413  16242443  16242398.8
#> 20: female age_group3 savings  222131  13335509  13337175.8
#> 21:      f age_group3 savings  222131  13335509  13337175.8
#> 22: female age_group4 savings  115241   6686628   6682987.3
#> 23:      f age_group4 savings  115241   6686628   6682987.3
#> 24: female age_group5 savings   37218   2272773   2270709.0
#> 25:      f age_group5 savings   37218   2272773   2270709.0
#> 26: female age_group6 savings    2987    207457    205958.9
#> 27:      f age_group6 savings    2987    207457    205958.9
#> 28:   male      Total savings 1159816  69005166  69005571.7
#> 29:      m      Total savings 1159816  69005166  69005571.7
#> 30:   male age_group1 savings  517660  31348784  31348853.0
#> 31:      m age_group1 savings  517660  31348784  31348853.0
#> 32:   male       ag1a savings  517660  31348784  31348853.0
#> 33:      m       ag1a savings  517660  31348784  31348853.0
#> 34:   male age_group2 savings  280923  16109453  16104698.6
#> 35:      m age_group2 savings  280923  16109453  16104698.6
#> 36:   male       ag2a savings  280923  16109453  16104698.6
#> 37:      m       ag2a savings  280923  16109453  16104698.6
#> 38:   male age_group3 savings  214970  12816538  12817856.6
#> 39:      m age_group3 savings  214970  12816538  12817856.6
#> 40:   male age_group4 savings   99420   5916216   5916010.8
#> 41:      m age_group4 savings   99420   5916216   5916010.8
#> 42:   male age_group5 savings   43233   2563466   2567008.8
#> 43:      m age_group5 savings   43233   2563466   2567008.8
#> 44:   male age_group6 savings    3610    250709    240434.1
#> 45:      m age_group6 savings    3610    250709    240434.1
#>        sex        age   vname     uws        ws         pws

# results can be resetted, too
tab$reset_cntvars(v = "cnt_males")

# we can then set other parameters and perturb again
tab$params_cnts_set(val = p_cnts1, v = "cnt_males")

#> --> setting perturbation parameters for variable 'cnt_males'

tab$perturb(v = "cnt_males")

#> Count variable 'cnt_males' was perturbed.

# write to a file "outtab.csv" (.csv is automatically added to the path)
if (FALSE) {
tab$freqtab(v = c("total", "cnt_males"), path = "outtab.csv")
}

# show results containing weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))

#>        sex        age     vname  uwc     wc puwc        pwc
#>  1:  Total      Total     total 4580 272971 4579 272911.399
#>  2:  Total age_group1     total 1969 118179 1970 118239.020
#>  3:  Total       ag1a     total 1969 118179 1970 118239.020
#>  4:  Total age_group2     total 1143  67260 1145  67377.690
#>  5:  Total       ag2a     total 1143  67260 1145  67377.690
#>  6:  Total age_group3     total  864  51608  865  51667.731
#>  7:  Total age_group4     total  423  24845  424  24903.735
#>  8:  Total age_group5     total  168  10250  170  10372.024
#>  9:  Total age_group6     total   13    829   13    829.000
#> 10: female      Total     total 2284 136454 2284 136454.000
#> 11:      f      Total     total 2284 136454 2284 136454.000
#> 12: female age_group1     total  954  57080  953  57020.168
#> 13:      f age_group1     total  954  57080  953  57020.168
#> 14: female       ag1a     total  954  57080  953  57020.168
#> 15:      f       ag1a     total  954  57080  953  57020.168
#> 16: female age_group2     total  572  34218  573  34277.822
#> 17:      f age_group2     total  572  34218  573  34277.822
#> 18: female       ag2a     total  572  34218  573  34277.822
#> 19:      f       ag2a     total  572  34218  573  34277.822
#> 20: female age_group3     total  440  26277  440  26277.000
#> 21:      f age_group3     total  440  26277  440  26277.000
#> 22: female age_group4     total  228  13374  227  13315.342
#> 23:      f age_group4     total  228  13374  227  13315.342
#> 24: female age_group5     total   84   5145   83   5083.750
#> 25:      f age_group5     total   84   5145   83   5083.750
#> 26: female age_group6     total    6    360    7    420.000
#> 27:      f age_group6     total    6    360    7    420.000
#> 28:   male      Total     total 2296 136517 2296 136517.000
#> 29:      m      Total     total 2296 136517 2296 136517.000
#> 30:   male age_group1     total 1015  61099 1016  61159.196
#> 31:      m age_group1     total 1015  61099 1016  61159.196
#> 32:   male       ag1a     total 1015  61099 1016  61159.196
#> 33:      m       ag1a     total 1015  61099 1016  61159.196
#> 34:   male age_group2     total  571  33042  570  32984.133
#> 35:      m age_group2     total  571  33042  570  32984.133
#> 36:   male       ag2a     total  571  33042  570  32984.133
#> 37:      m       ag2a     total  571  33042  570  32984.133
#> 38:   male age_group3     total  424  25331  424  25331.000
#> 39:      m age_group3     total  424  25331  424  25331.000
#> 40:   male age_group4     total  195  11471  196  11529.826
#> 41:      m age_group4     total  195  11471  196  11529.826
#> 42:   male age_group5     total   84   5105   85   5165.774
#> 43:      m age_group5     total   84   5105   85   5165.774
#> 44:   male age_group6     total    7    469    8    536.000
#> 45:      m age_group6     total    7    469    8    536.000
#> 46:  Total      Total cnt_males 2296 136517 2296 136517.000
#> 47:  Total age_group1 cnt_males 1015  61099 1016  61159.196
#> 48:  Total       ag1a cnt_males 1015  61099 1016  61159.196
#> 49:  Total age_group2 cnt_males  571  33042  570  32984.133
#> 50:  Total       ag2a cnt_males  571  33042  570  32984.133
#> 51:  Total age_group3 cnt_males  424  25331  424  25331.000
#> 52:  Total age_group4 cnt_males  195  11471  196  11529.826
#> 53:  Total age_group5 cnt_males   84   5105   85   5165.774
#> 54:  Total age_group6 cnt_males    7    469    8    536.000
#> 55: female      Total cnt_males    0      0    0      0.000
#> 56:      f      Total cnt_males    0      0    0      0.000
#> 57: female age_group1 cnt_males    0      0    0      0.000
#> 58:      f age_group1 cnt_males    0      0    0      0.000
#> 59: female       ag1a cnt_males    0      0    0      0.000
#> 60:      f       ag1a cnt_males    0      0    0      0.000
#> 61: female age_group2 cnt_males    0      0    0      0.000
#> 62:      f age_group2 cnt_males    0      0    0      0.000
#> 63: female       ag2a cnt_males    0      0    0      0.000
#> 64:      f       ag2a cnt_males    0      0    0      0.000
#> 65: female age_group3 cnt_males    0      0    0      0.000
#> 66:      f age_group3 cnt_males    0      0    0      0.000
#> 67: female age_group4 cnt_males    0      0    0      0.000
#> 68:      f age_group4 cnt_males    0      0    0      0.000
#> 69: female age_group5 cnt_males    0      0    0      0.000
#> 70:      f age_group5 cnt_males    0      0    0      0.000
#> 71: female age_group6 cnt_males    0      0    0      0.000
#> 72:      f age_group6 cnt_males    0      0    0      0.000
#> 73:   male      Total cnt_males 2296 136517 2296 136517.000
#> 74:      m      Total cnt_males 2296 136517 2296 136517.000
#> 75:   male age_group1 cnt_males 1015  61099 1016  61159.196
#> 76:      m age_group1 cnt_males 1015  61099 1016  61159.196
#> 77:   male       ag1a cnt_males 1015  61099 1016  61159.196
#> 78:      m       ag1a cnt_males 1015  61099 1016  61159.196
#> 79:   male age_group2 cnt_males  571  33042  570  32984.133
#> 80:      m age_group2 cnt_males  571  33042  570  32984.133
#> 81:   male       ag2a cnt_males  571  33042  570  32984.133
#> 82:      m       ag2a cnt_males  571  33042  570  32984.133
#> 83:   male age_group3 cnt_males  424  25331  424  25331.000
#> 84:      m age_group3 cnt_males  424  25331  424  25331.000
#> 85:   male age_group4 cnt_males  195  11471  196  11529.826
#> 86:      m age_group4 cnt_males  195  11471  196  11529.826
#> 87:   male age_group5 cnt_males   84   5105   85   5165.774
#> 88:      m age_group5 cnt_males   84   5105   85   5165.774
#> 89:   male age_group6 cnt_males    7    469    8    536.000
#> 90:      m age_group6 cnt_males    7    469    8    536.000
#>        sex        age     vname  uwc     wc puwc        pwc

# utility measures for a count variable
tab$measures_cnts(v = "total", exclude_zeros = TRUE)

#> $overview
#>    noise cnt        pct
#> 1:    -2   3 0.06666667
#> 2:    -1  20 0.44444444
#> 3:     0   9 0.20000000
#> 4:     1  13 0.28888889
#> 
#> $measures
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.800 0.000 0.006
#>  4:    Q30 1.000 0.001 0.016
#>  5:    Q40 1.000 0.001 0.016
#>  6:   Mean 0.867 0.016 0.036
#>  7: Median 1.000 0.002 0.021
#>  8:    Q60 1.000 0.002 0.021
#>  9:    Q70 1.000 0.002 0.030
#> 10:    Q80 1.000 0.006 0.039
#> 11:    Q90 1.000 0.012 0.068
#> 12:    Q95 1.800 0.143 0.183
#> 13:    Q99 2.000 0.167 0.196
#> 14:    Max 2.000 0.167 0.196
#> 
#> $cumdistr_d1
#>    cat cnt       pct
#> 1:   0   9 0.2000000
#> 2:   1  42 0.9333333
#> 3:   2  45 1.0000000
#> 
#> $cumdistr_d2
#>            cat cnt       pct
#> 1:    [0,0.02]  41 0.9111111
#> 2: (0.02,0.05]  41 0.9111111
#> 3:  (0.05,0.1]  41 0.9111111
#> 4:   (0.1,0.2]  45 1.0000000
#> 5:   (0.2,0.3]  45 1.0000000
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $cumdistr_d3
#>            cat cnt       pct
#> 1:    [0,0.02]  21 0.4666667
#> 2: (0.02,0.05]  36 0.8000000
#> 3:  (0.05,0.1]  41 0.9111111
#> 4:   (0.1,0.2]  45 1.0000000
#> 5:   (0.2,0.3]  45 1.0000000
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $false_zero
#> [1] 0
#> 
#> $false_nonzero
#> [1] 0
#> 
#> $exclude_zeros
#> [1] TRUE
#> 

# modifications for perturbed count variables
tab$mod_cnts()

#>        sex        age row_nr pert      ckey  countvar
#>   1: Total      Total     14   -1 0.1213574     total
#>   2: Total age_group1     16    1 0.8850270     total
#>   3: Total       ag1a     16    1 0.8850270     total
#>   4: Total age_group2     17    2 0.9847088     total
#>   5: Total       ag2a     17    2 0.9847088     total
#>  ---                                                 
#> 131:     m age_group4     16    1 0.7337113 cnt_males
#> 132:  male age_group5     16    1 0.8585011 cnt_males
#> 133:     m age_group5     16    1 0.8585011 cnt_males
#> 134:  male age_group6     16    1 0.7007744 cnt_males
#> 135:     m age_group6     16    1 0.7007744 cnt_males

# display a summary about utility measures
tab$summary()

#> ┌──────────────────────────────────────────────┐
#> │Utility measures for perturbed count variables│
#> └──────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#>          countvar Min  Q10 Q20 Q30 Q40  Mean Median Q60 Q70 Q80 Q90 Q95  Q99
#> 1:          total  -1 -1.0  -1   0   0 0.289      1   1 1.0 1.0   1 1.8 2.00
#> 2: cnt_highincome  -2 -0.6   0   0   1 1.022      1   1 1.8 2.2   3 3.0 3.56
#> 3:      cnt_males  -1 -1.0   0   0   0 0.200      0   0 1.0 1.0   1 1.0 1.00
#>    Max
#> 1:   2
#> 2:   4
#> 3:   1
#> 
#> ── Distance-based measures ─────────────────────────────────────────────────────
#> ✔ Variable: 'total'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.800 0.000 0.006
#>  4:    Q30 1.000 0.001 0.016
#>  5:    Q40 1.000 0.001 0.016
#>  6:   Mean 0.867 0.016 0.036
#>  7: Median 1.000 0.002 0.021
#>  8:    Q60 1.000 0.002 0.021
#>  9:    Q70 1.000 0.002 0.030
#> 10:    Q80 1.000 0.006 0.039
#> 11:    Q90 1.000 0.012 0.068
#> 12:    Q95 1.800 0.143 0.183
#> 13:    Q99 2.000 0.167 0.196
#> 14:    Max 2.000 0.167 0.196
#> 
#> ✔ Variable: 'cnt_males'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.200 0.000 0.003
#>  4:    Q30 1.000 0.001 0.016
#>  5:    Q40 1.000 0.001 0.016
#>  6:   Mean 0.778 0.018 0.038
#>  7: Median 1.000 0.002 0.021
#>  8:    Q60 1.000 0.002 0.021
#>  9:    Q70 1.000 0.005 0.036
#> 10:    Q80 1.000 0.011 0.051
#> 11:    Q90 1.000 0.064 0.106
#> 12:    Q95 1.000 0.143 0.183
#> 13:    Q99 1.000 0.143 0.183
#> 14:    Max 1.000 0.143 0.183
#> 
#> ✔ Variable: 'cnt_highincome'
#> 
#>       what   d1    d2    d3
#>  1:    Min 0.00 0.000 0.000
#>  2:    Q10 0.00 0.000 0.000
#>  3:    Q20 1.00 0.005 0.034
#>  4:    Q30 1.00 0.014 0.061
#>  5:    Q40 1.00 0.017 0.066
#>  6:   Mean 1.55 0.044 0.111
#>  7: Median 1.00 0.021 0.078
#>  8:    Q60 2.00 0.029 0.135
#>  9:    Q70 2.00 0.033 0.150
#> 10:    Q80 3.00 0.051 0.157
#> 11:    Q90 3.00 0.143 0.183
#> 12:    Q95 3.00 0.150 0.266
#> 13:    Q99 3.61 0.286 0.410
#> 14:    Max 4.00 0.286 0.410
#> 
#> ┌──────────────────────────────────────────────────┐
#> │Utility measures for perturbed numerical variables│
#> └──────────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────

#> Warning: no non-missing arguments to min; returning Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to min; returning Inf

#> Warning: no non-missing arguments to max; returning -Inf

#>      vname        Min        Q10        Q20        Q30        Q40       Mean
#> 1:  expend        Inf         NA         NA         NA         NA        NaN
#> 2:  income -116027.24 -76501.501 -63345.812 -55381.918 -34751.196 -24995.915
#> 3: savings  -10274.86  -4754.388  -2676.514  -2147.757  -1520.573   -807.446
#> 4:   mixed        Inf         NA         NA         NA         NA        NaN
#>        Median        Q60      Q70       Q80       Q90       Q95       Q99
#> 1:         NA         NA       NA        NA        NA        NA        NA
#> 2: -17796.199 -11968.410 7397.756 13210.425 27651.224 28969.081 47433.803
#> 3:    -44.161     68.981  405.667  1388.218  3220.426  3542.809  6311.375
#> 4:         NA         NA       NA        NA        NA        NA        NA
#>          Max
#> 1:      -Inf
#> 2: 61820.851
#> 3:  6311.375
#> 4:      -Inf