Create perturbation parameters for count variables

This function allows to generate required perturbation parameters that are used to perturb count variables.

ck_params_cnts(ptab, path = NULL)

Arguments

ptab

an object created with ptable::create_ptable(), or ptable::create_cnt_ptable()

path

a scalar character specifying a path to which the parameters created with this functions should be written to (in yaml format)

Value

an object suitable as input to method $params_cnts_set() for the perturbation of counts and frequencies.

See also

Examples

# \donttest{
x <- ck_create_testdata()

# create some 0/1 variables that should be perturbed later
x[, cnt_females := ifelse(sex == "male", 0, 1)]
#>       urbrur  roof walls water electcon relat    sex        age hhcivil expend
#>        <int> <int> <int> <int>    <int> <int> <fctr>     <fctr>   <int>  <num>
#>    1:      2     4     3     3        1     1   male age_group3       2   9093
#>    2:      2     4     3     3        1     2 female age_group3       2   2734
#>    3:      2     4     3     3        1     3   male age_group1       1   2652
#>    4:      2     4     3     3        1     3   male age_group1       1   1807
#>    5:      2     4     2     3        1     1   male age_group4       2    671
#>   ---                                                                         
#> 4576:      2     4     3     4        1     2 female age_group3       2   3696
#> 4577:      2     4     3     4        1     3   male age_group1       1    282
#> 4578:      2     4     3     4        1     3   male age_group1       1    840
#> 4579:      2     4     3     4        1     3 female age_group1       1   6258
#> 4580:      2     4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>        <num>   <num>   <int>           <int>             <num>       <num>
#>    1:   5780      12       1              45          25.00000           0
#>    2:   2530      28       1              68          25.00000           1
#>    3:   6920     550       1              63          25.00000           0
#>    4:   7960     870       1              98          25.00000           0
#>    5:   9030      20       2              61          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              74          16.66667           1
#> 4577:   1420     987    1000              87          16.66667           0
#> 4578:   8900     684    1000              40          16.66667           0
#> 4579:   3880     294    1000              48          16.66667           1
#> 4580:   4830     911    1000              53          16.66667           0
x[, cnt_males := ifelse(sex == "male", 1, 0)]
#>       urbrur  roof walls water electcon relat    sex        age hhcivil expend
#>        <int> <int> <int> <int>    <int> <int> <fctr>     <fctr>   <int>  <num>
#>    1:      2     4     3     3        1     1   male age_group3       2   9093
#>    2:      2     4     3     3        1     2 female age_group3       2   2734
#>    3:      2     4     3     3        1     3   male age_group1       1   2652
#>    4:      2     4     3     3        1     3   male age_group1       1   1807
#>    5:      2     4     2     3        1     1   male age_group4       2    671
#>   ---                                                                         
#> 4576:      2     4     3     4        1     2 female age_group3       2   3696
#> 4577:      2     4     3     4        1     3   male age_group1       1    282
#> 4578:      2     4     3     4        1     3   male age_group1       1    840
#> 4579:      2     4     3     4        1     3 female age_group1       1   6258
#> 4580:      2     4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>        <num>   <num>   <int>           <int>             <num>       <num>
#>    1:   5780      12       1              45          25.00000           0
#>    2:   2530      28       1              68          25.00000           1
#>    3:   6920     550       1              63          25.00000           0
#>    4:   7960     870       1              98          25.00000           0
#>    5:   9030      20       2              61          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              74          16.66667           1
#> 4577:   1420     987    1000              87          16.66667           0
#> 4578:   8900     684    1000              40          16.66667           0
#> 4579:   3880     294    1000              48          16.66667           1
#> 4580:   4830     911    1000              53          16.66667           0
#>       cnt_males
#>           <num>
#>    1:         1
#>    2:         0
#>    3:         1
#>    4:         1
#>    5:         1
#>   ---          
#> 4576:         0
#> 4577:         1
#> 4578:         1
#> 4579:         0
#> 4580:         1
x[, cnt_highincome := ifelse(income >= 9000, 1, 0)]
#>       urbrur  roof walls water electcon relat    sex        age hhcivil expend
#>        <int> <int> <int> <int>    <int> <int> <fctr>     <fctr>   <int>  <num>
#>    1:      2     4     3     3        1     1   male age_group3       2   9093
#>    2:      2     4     3     3        1     2 female age_group3       2   2734
#>    3:      2     4     3     3        1     3   male age_group1       1   2652
#>    4:      2     4     3     3        1     3   male age_group1       1   1807
#>    5:      2     4     2     3        1     1   male age_group4       2    671
#>   ---                                                                         
#> 4576:      2     4     3     4        1     2 female age_group3       2   3696
#> 4577:      2     4     3     4        1     3   male age_group1       1    282
#> 4578:      2     4     3     4        1     3   male age_group1       1    840
#> 4579:      2     4     3     4        1     3 female age_group1       1   6258
#> 4580:      2     4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>        <num>   <num>   <int>           <int>             <num>       <num>
#>    1:   5780      12       1              45          25.00000           0
#>    2:   2530      28       1              68          25.00000           1
#>    3:   6920     550       1              63          25.00000           0
#>    4:   7960     870       1              98          25.00000           0
#>    5:   9030      20       2              61          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              74          16.66667           1
#> 4577:   1420     987    1000              87          16.66667           0
#> 4578:   8900     684    1000              40          16.66667           0
#> 4579:   3880     294    1000              48          16.66667           1
#> 4580:   4830     911    1000              53          16.66667           0
#>       cnt_males cnt_highincome
#>           <num>          <num>
#>    1:         1              0
#>    2:         0              0
#>    3:         1              0
#>    4:         1              0
#>    5:         1              1
#>   ---                         
#> 4576:         0              0
#> 4577:         1              0
#> 4578:         1              0
#> 4579:         0              0
#> 4580:         1              0
# a variable with positive and negative contributions
x[, mixed := sample(-10:10, nrow(x), replace = TRUE)]
#>       urbrur  roof walls water electcon relat    sex        age hhcivil expend
#>        <int> <int> <int> <int>    <int> <int> <fctr>     <fctr>   <int>  <num>
#>    1:      2     4     3     3        1     1   male age_group3       2   9093
#>    2:      2     4     3     3        1     2 female age_group3       2   2734
#>    3:      2     4     3     3        1     3   male age_group1       1   2652
#>    4:      2     4     3     3        1     3   male age_group1       1   1807
#>    5:      2     4     2     3        1     1   male age_group4       2    671
#>   ---                                                                         
#> 4576:      2     4     3     4        1     2 female age_group3       2   3696
#> 4577:      2     4     3     4        1     3   male age_group1       1    282
#> 4578:      2     4     3     4        1     3   male age_group1       1    840
#> 4579:      2     4     3     4        1     3 female age_group1       1   6258
#> 4580:      2     4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>        <num>   <num>   <int>           <int>             <num>       <num>
#>    1:   5780      12       1              45          25.00000           0
#>    2:   2530      28       1              68          25.00000           1
#>    3:   6920     550       1              63          25.00000           0
#>    4:   7960     870       1              98          25.00000           0
#>    5:   9030      20       2              61          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              74          16.66667           1
#> 4577:   1420     987    1000              87          16.66667           0
#> 4578:   8900     684    1000              40          16.66667           0
#> 4579:   3880     294    1000              48          16.66667           1
#> 4580:   4830     911    1000              53          16.66667           0
#>       cnt_males cnt_highincome mixed
#>           <num>          <num> <int>
#>    1:         1              0    10
#>    2:         0              0    -1
#>    3:         1              0    -9
#>    4:         1              0    -2
#>    5:         1              1    -1
#>   ---                               
#> 4576:         0              0     7
#> 4577:         1              0    -2
#> 4578:         1              0     7
#> 4579:         0              0    -7
#> 4580:         1              0     8

# create record keys
x$rkey <- ck_generate_rkeys(dat = x)

# define required inputs

# hierarchy with some bogus codes
d_sex <- hier_create(root = "Total", nodes = c("male", "female"))
d_sex <- hier_add(d_sex, root = "female", "f")
d_sex <- hier_add(d_sex, root = "male", "m")

d_age <- hier_create(root = "Total", nodes = paste0("age_group", 1:6))
d_age <- hier_add(d_age, root = "age_group1", "ag1a")
d_age <- hier_add(d_age, root = "age_group2", "ag2a")

# define the cell key object
countvars <- c("cnt_females", "cnt_males", "cnt_highincome")
numvars <- c("expend", "income", "savings", "mixed")
tab <- ck_setup(
  x = x,
  rkey = "rkey",
  dims = list(sex = d_sex, age = d_age),
  w = "sampling_weight",
  countvars = countvars,
  numvars = numvars)
#> computing contributing indices | rawdata <--> table; this might take a while

# show some information about this table instance
tab$print() # identical with print(tab)
#> ── Table Information ───────────────────────────────────────────────────────────
#> ✔ 45 cells in 2 dimensions ('sex', 'age')
#> ✔ weights: yes
#> ── Tabulated / Perturbed countvars ─────────────────────────────────────────────
#> ☐ 'total'
#> ☐ 'cnt_females'
#> ☐ 'cnt_males'
#> ☐ 'cnt_highincome'
#> ── Tabulated / Perturbed numvars ───────────────────────────────────────────────
#> ☐ 'expend'
#> ☐ 'income'
#> ☐ 'savings'
#> ☐ 'mixed'

# information about the hierarchies
tab$hierarchy_info()
#> $sex
#>      code level is_leaf parent
#>    <char> <int>  <lgcl> <char>
#> 1:  Total     1   FALSE  Total
#> 2:   male     2   FALSE  Total
#> 3:      m     3    TRUE   male
#> 4: female     2   FALSE  Total
#> 5:      f     3    TRUE female
#> 
#> $age
#>          code level is_leaf     parent
#>        <char> <int>  <lgcl>     <char>
#> 1:      Total     1   FALSE      Total
#> 2: age_group1     2   FALSE      Total
#> 3:       ag1a     3    TRUE age_group1
#> 4: age_group2     2   FALSE      Total
#> 5:       ag2a     3    TRUE age_group2
#> 6: age_group3     2    TRUE      Total
#> 7: age_group4     2    TRUE      Total
#> 8: age_group5     2    TRUE      Total
#> 9: age_group6     2    TRUE      Total
#> 

# which variables have been defined?
tab$allvars()
#> $cntvars
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"
#> 
#> $numvars
#> [1] "expend"  "income"  "savings" "mixed"  
#> 

# count variables
tab$cntvars()
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"

# continuous variables
tab$numvars()
#> [1] "expend"  "income"  "savings" "mixed"  

# create perturbation parameters for "total" variable and
# write to yaml-file

# create a ptable using functionality from the ptable-pkg
f_yaml <- tempfile(fileext = ".yaml")
p_cnts1 <- ck_params_cnts(
  ptab = ptable::pt_ex_cnts(),
  path = f_yaml)
#> yaml configuration '/tmp/Rtmpu511Tl/file1fa5466b6274.yaml' successfully written.

# read parameters from yaml-file and set them for variable `"total"`
p_cnts1 <- ck_read_yaml(path = f_yaml)

tab$params_cnts_set(val = p_cnts1, v = "total")
#> --> setting perturbation parameters for variable 'total'

# create alternative perturbation parameters by specifying parameters
para2 <- ptable::create_cnt_ptable(
  D = 8, V = 3, js = 2, create = FALSE)

p_cnts2 <- ck_params_cnts(ptab = para2)

# use these ptable it for the remaining variables
tab$params_cnts_set(val = p_cnts2, v = countvars)
#> --> setting perturbation parameters for variable 'cnt_females'
#> --> setting perturbation parameters for variable 'cnt_males'
#> --> setting perturbation parameters for variable 'cnt_highincome'

# perturb a variable
tab$perturb(v = "total")
#> Count variable 'total' was perturbed.

# multiple variables can be perturbed as well
tab$perturb(v = c("cnt_males", "cnt_highincome"))
#> Count variable 'cnt_males' was perturbed.
#> Count variable 'cnt_highincome' was perturbed.

# return weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))
#>        sex        age     vname   uwc     wc  puwc         pwc
#>     <char>     <char>    <char> <num>  <num> <num>       <num>
#>  1:  Total      Total     total  4580 277416  4580 277416.0000
#>  2:  Total age_group1     total  1969 117568  1969 117568.0000
#>  3:  Total       ag1a     total  1969 117568  1969 117568.0000
#>  4:  Total age_group2     total  1143  69592  1143  69592.0000
#>  5:  Total       ag2a     total  1143  69592  1143  69592.0000
#>  6:  Total age_group3     total   864  53968   863  53905.5370
#>  7:  Total age_group4     total   423  25741   423  25741.0000
#>  8:  Total age_group5     total   168   9790   168   9790.0000
#>  9:  Total age_group6     total    13    757    13    757.0000
#> 10:   male      Total     total  2296 138830  2297 138890.4660
#> 11:      m      Total     total  2296 138830  2297 138890.4660
#> 12:   male age_group1     total  1015  60383  1015  60383.0000
#> 13:      m age_group1     total  1015  60383  1015  60383.0000
#> 14:   male       ag1a     total  1015  60383  1015  60383.0000
#> 15:      m       ag1a     total  1015  60383  1015  60383.0000
#> 16:   male age_group2     total   571  34594   570  34533.4151
#> 17:      m age_group2     total   571  34594   570  34533.4151
#> 18:   male       ag2a     total   571  34594   570  34533.4151
#> 19:      m       ag2a     total   571  34594   570  34533.4151
#> 20:   male age_group3     total   424  26496   422  26371.0189
#> 21:      m age_group3     total   424  26496   422  26371.0189
#> 22:   male age_group4     total   195  11891   194  11830.0205
#> 23:      m age_group4     total   195  11891   194  11830.0205
#> 24:   male age_group5     total    84   5070    83   5009.6429
#> 25:      m age_group5     total    84   5070    83   5009.6429
#> 26:   male age_group6     total     7    396     9    509.1429
#> 27:      m age_group6     total     7    396     9    509.1429
#> 28: female      Total     total  2284 138586  2284 138586.0000
#> 29:      f      Total     total  2284 138586  2284 138586.0000
#> 30: female age_group1     total   954  57185   952  57065.1153
#> 31:      f age_group1     total   954  57185   952  57065.1153
#> 32: female       ag1a     total   954  57185   952  57065.1153
#> 33:      f       ag1a     total   954  57185   952  57065.1153
#> 34: female age_group2     total   572  34998   571  34936.8147
#> 35:      f age_group2     total   572  34998   571  34936.8147
#> 36: female       ag2a     total   572  34998   571  34936.8147
#> 37:      f       ag2a     total   572  34998   571  34936.8147
#> 38: female age_group3     total   440  27472   439  27409.5636
#> 39:      f age_group3     total   440  27472   439  27409.5636
#> 40: female age_group4     total   228  13850   227  13789.2544
#> 41:      f age_group4     total   228  13850   227  13789.2544
#> 42: female age_group5     total    84   4720    84   4720.0000
#> 43:      f age_group5     total    84   4720    84   4720.0000
#> 44: female age_group6     total     6    361     6    361.0000
#> 45:      f age_group6     total     6    361     6    361.0000
#> 46:  Total      Total cnt_males  2296 138830  2298 138950.9321
#> 47:  Total age_group1 cnt_males  1015  60383  1015  60383.0000
#> 48:  Total       ag1a cnt_males  1015  60383  1015  60383.0000
#> 49:  Total age_group2 cnt_males   571  34594   570  34533.4151
#> 50:  Total       ag2a cnt_males   571  34594   570  34533.4151
#> 51:  Total age_group3 cnt_males   424  26496   420  26246.0377
#> 52:  Total age_group4 cnt_males   195  11891   194  11830.0205
#> 53:  Total age_group5 cnt_males    84   5070    82   4949.2857
#> 54:  Total age_group6 cnt_males     7    396    10    565.7143
#> 55:   male      Total cnt_males  2296 138830  2298 138950.9321
#> 56:      m      Total cnt_males  2296 138830  2298 138950.9321
#> 57:   male age_group1 cnt_males  1015  60383  1015  60383.0000
#> 58:      m age_group1 cnt_males  1015  60383  1015  60383.0000
#> 59:   male       ag1a cnt_males  1015  60383  1015  60383.0000
#> 60:      m       ag1a cnt_males  1015  60383  1015  60383.0000
#> 61:   male age_group2 cnt_males   571  34594   570  34533.4151
#> 62:      m age_group2 cnt_males   571  34594   570  34533.4151
#> 63:   male       ag2a cnt_males   571  34594   570  34533.4151
#> 64:      m       ag2a cnt_males   571  34594   570  34533.4151
#> 65:   male age_group3 cnt_males   424  26496   420  26246.0377
#> 66:      m age_group3 cnt_males   424  26496   420  26246.0377
#> 67:   male age_group4 cnt_males   195  11891   194  11830.0205
#> 68:      m age_group4 cnt_males   195  11891   194  11830.0205
#> 69:   male age_group5 cnt_males    84   5070    82   4949.2857
#> 70:      m age_group5 cnt_males    84   5070    82   4949.2857
#> 71:   male age_group6 cnt_males     7    396    10    565.7143
#> 72:      m age_group6 cnt_males     7    396    10    565.7143
#> 73: female      Total cnt_males     0      0     0      0.0000
#> 74:      f      Total cnt_males     0      0     0      0.0000
#> 75: female age_group1 cnt_males     0      0     0      0.0000
#> 76:      f age_group1 cnt_males     0      0     0      0.0000
#> 77: female       ag1a cnt_males     0      0     0      0.0000
#> 78:      f       ag1a cnt_males     0      0     0      0.0000
#> 79: female age_group2 cnt_males     0      0     0      0.0000
#> 80:      f age_group2 cnt_males     0      0     0      0.0000
#> 81: female       ag2a cnt_males     0      0     0      0.0000
#> 82:      f       ag2a cnt_males     0      0     0      0.0000
#> 83: female age_group3 cnt_males     0      0     0      0.0000
#> 84:      f age_group3 cnt_males     0      0     0      0.0000
#> 85: female age_group4 cnt_males     0      0     0      0.0000
#> 86:      f age_group4 cnt_males     0      0     0      0.0000
#> 87: female age_group5 cnt_males     0      0     0      0.0000
#> 88:      f age_group5 cnt_males     0      0     0      0.0000
#> 89: female age_group6 cnt_males     0      0     0      0.0000
#> 90:      f age_group6 cnt_males     0      0     0      0.0000
#>        sex        age     vname   uwc     wc  puwc         pwc

# numerical variables (positive variables using flex-function)
# we also write the config to a yaml file
f_yaml <- tempfile(fileext = ".yaml")

# create a ptable using functionality from the ptable-pkg
# a single ptable for all cells
ptab1 <- ptable::pt_ex_nums(parity = TRUE, separation = FALSE)

# a single ptab for all cells except for very small ones
ptab2 <- ptable::pt_ex_nums(parity = TRUE, separation = TRUE)

# different ptables for cells with even/odd number of contributors
# and very small cells
ptab3 <- ptable::pt_ex_nums(parity = FALSE, separation = TRUE)

p_nums1 <- ck_params_nums(
  ptab = ptab1,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.30, 0.03),
    epsilon = c(1, 0.5, 0.2),
    q = 3),
  mu_c = 2,
  same_key = FALSE,
  use_zero_rkeys = FALSE,
  path = f_yaml)
#> yaml configuration '/tmp/Rtmpu511Tl/file1fa519e16914.yaml' successfully written.

# we read the parameters from the yaml-file
p_nums1 <- ck_read_yaml(path = f_yaml)

# for variables with positive and negative values
p_nums2 <- ck_params_nums(
  ptab = ptab2,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.15, 0.02),
    epsilon = c(1, 0.4, 0.15),
    q = 3),
  mu_c = 2,
  same_key = FALSE)

# simple perturbation parameters (not using the flex-function approach)
p_nums3 <- ck_params_nums(
  ptab = ptab3,
  type = "mean",
  mult_params = ck_simpleparams(p = 0.25),
  mu_c = 2,
  same_key = FALSE)

# use `p_nums1` for all variables
tab$params_nums_set(p_nums1, c("savings", "income", "expend"))
#> --> setting perturbation parameters for variable 'savings'
#> --> setting perturbation parameters for variable 'income'
#> --> setting perturbation parameters for variable 'expend'

# use different parameters for variable `mixed`
tab$params_nums_set(p_nums2, v = "mixed")
#> --> setting perturbation parameters for variable 'mixed'

# identify sensitive cells to which extra protection (`mu_c`) is added.
tab$supp_p(v = "income", p = 85)
#> computing contributing indices | rawdata <--> table; this might take a while
#> p%-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_pq(v = "income", p = 85, q = 90)
#> computing contributing indices | rawdata <--> table; this might take a while
#> pq-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_nk(v = "income", n = 2, k = 90)
#> computing contributing indices | rawdata <--> table; this might take a while
#> nk-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_freq(v = "income", n = 14, weighted = FALSE)
#> freq-rule: 5 new sensitive cells (incl. duplicates) found (total: 5)
tab$supp_val(v = "income", n = 10000, weighted = TRUE)
#> val-rule: 0 new sensitive cells (incl. duplicates) found (total: 5)
tab$supp_cells(
  v = "income",
  inp = data.frame(
    sex = c("female", "female"),
    "age" = c("age_group1", "age_group3")
  )
)
#> cell-rule: 2 new sensitive cells (incl. duplicates) found (total: 7)

# perturb variables
tab$perturb(v = c("income", "savings"))
#> Numeric variable 'income' was perturbed.
#> Numeric variable 'savings' was perturbed.

# extract results
tab$numtab("income", mean_before_sum = TRUE)
#>        sex        age  vname      uws         ws        pws
#>     <char>     <char> <char>    <num>      <num>      <num>
#>  1:  Total      Total income 22952978 1389725230 1389700927
#>  2:  Total age_group1 income  9810547  589228354  589263472
#>  3:  Total       ag1a income  9810547  589228354  589263472
#>  4:  Total age_group2 income  5692119  344985537  345045877
#>  5:  Total       ag2a income  5692119  344985537  345045877
#>  6:  Total age_group3 income  4406946  276593268  276390942
#>  7:  Total age_group4 income  2133543  127479059  127489902
#>  8:  Total age_group5 income   848151   48061868   48134370
#>  9:  Total age_group6 income    61672    3377144    3314947
#> 10:   male      Total income 11262049  677621008  677614627
#> 11:      m      Total income 11262049  677621008  677614627
#> 12:   male age_group1 income  4877164  289787701  289696339
#> 13:      m age_group1 income  4877164  289787701  289696339
#> 14:   male       ag1a income  4877164  289787701  289696339
#> 15:      m       ag1a income  4877164  289787701  289696339
#> 16:   male age_group2 income  2811379  168679656  168678828
#> 17:      m age_group2 income  2811379  168679656  168678828
#> 18:   male       ag2a income  2811379  168679656  168678828
#> 19:      m       ag2a income  2811379  168679656  168678828
#> 20:   male age_group3 income  2168169  136694033  136745521
#> 21:      m age_group3 income  2168169  136694033  136745521
#> 22:   male age_group4 income   978510   56908935   57162307
#> 23:      m age_group4 income   978510   56908935   57162307
#> 24:   male age_group5 income   393134   23616752   23681313
#> 25:      m age_group5 income   393134   23616752   23681313
#> 26:   male age_group6 income    33693    1933931    1968472
#> 27:      m age_group6 income    33693    1933931    1968472
#> 28: female      Total income 11690929  712104222  712120750
#> 29:      f      Total income 11690929  712104222  712120750
#> 30: female age_group1 income  4933383  299440653  299449608
#> 31:      f age_group1 income  4933383  299440653  299449608
#> 32: female       ag1a income  4933383  299440653  299449608
#> 33:      f       ag1a income  4933383  299440653  299449608
#> 34: female age_group2 income  2880740  176305881  176265360
#> 35:      f age_group2 income  2880740  176305881  176265360
#> 36: female       ag2a income  2880740  176305881  176265360
#> 37:      f       ag2a income  2880740  176305881  176265360
#> 38: female age_group3 income  2238777  139899235  139843340
#> 39:      f age_group3 income  2238777  139899235  139843340
#> 40: female age_group4 income  1155033   70570124   70618717
#> 41:      f age_group4 income  1155033   70570124   70618717
#> 42: female age_group5 income   455017   24445116   24320383
#> 43:      f age_group5 income   455017   24445116   24320383
#> 44: female age_group6 income    27979    1443213    1475128
#> 45:      f age_group6 income    27979    1443213    1475128
#>        sex        age  vname      uws         ws        pws
tab$numtab("income", mean_before_sum = FALSE)
#>        sex        age  vname      uws         ws        pws
#>     <char>     <char> <char>    <num>      <num>      <num>
#>  1:  Total      Total income 22952978 1389725230 1389713079
#>  2:  Total age_group1 income  9810547  589228354  589245913
#>  3:  Total       ag1a income  9810547  589228354  589245913
#>  4:  Total age_group2 income  5692119  344985537  345015706
#>  5:  Total       ag2a income  5692119  344985537  345015706
#>  6:  Total age_group3 income  4406946  276593268  276492087
#>  7:  Total age_group4 income  2133543  127479059  127484480
#>  8:  Total age_group5 income   848151   48061868   48098105
#>  9:  Total age_group6 income    61672    3377144    3345901
#> 10:   male      Total income 11262049  677621008  677617817
#> 11:      m      Total income 11262049  677621008  677617817
#> 12:   male age_group1 income  4877164  289787701  289742016
#> 13:      m age_group1 income  4877164  289787701  289742016
#> 14:   male       ag1a income  4877164  289787701  289742016
#> 15:      m       ag1a income  4877164  289787701  289742016
#> 16:   male age_group2 income  2811379  168679656  168679242
#> 17:      m age_group2 income  2811379  168679656  168679242
#> 18:   male       ag2a income  2811379  168679656  168679242
#> 19:      m       ag2a income  2811379  168679656  168679242
#> 20:   male age_group3 income  2168169  136694033  136719774
#> 21:      m age_group3 income  2168169  136694033  136719774
#> 22:   male age_group4 income   978510   56908935   57035480
#> 23:      m age_group4 income   978510   56908935   57035480
#> 24:   male age_group5 income   393134   23616752   23649010
#> 25:      m age_group5 income   393134   23616752   23649010
#> 26:   male age_group6 income    33693    1933931    1951125
#> 27:      m age_group6 income    33693    1933931    1951125
#> 28: female      Total income 11690929  712104222  712112486
#> 29:      f      Total income 11690929  712104222  712112486
#> 30: female age_group1 income  4933383  299440653  299445130
#> 31:      f age_group1 income  4933383  299440653  299445130
#> 32: female       ag1a income  4933383  299440653  299445130
#> 33:      f       ag1a income  4933383  299440653  299445130
#> 34: female age_group2 income  2880740  176305881  176285619
#> 35:      f age_group2 income  2880740  176305881  176285619
#> 36: female       ag2a income  2880740  176305881  176285619
#> 37:      f       ag2a income  2880740  176305881  176285619
#> 38: female age_group3 income  2238777  139899235  139871284
#> 39:      f age_group3 income  2238777  139899235  139871284
#> 40: female age_group4 income  1155033   70570124   70594416
#> 41:      f age_group4 income  1155033   70570124   70594416
#> 42: female age_group5 income   455017   24445116   24382669
#> 43:      f age_group5 income   455017   24445116   24382669
#> 44: female age_group6 income    27979    1443213    1459083
#> 45:      f age_group6 income    27979    1443213    1459083
#>        sex        age  vname      uws         ws        pws
tab$numtab("savings")
#>        sex        age   vname     uws        ws         pws
#>     <char>     <char>  <char>   <num>     <num>       <num>
#>  1:  Total      Total savings 2273532 138283213 138284165.7
#>  2:  Total age_group1 savings  982386  58884843  58887323.4
#>  3:  Total       ag1a savings  982386  58884843  58887323.4
#>  4:  Total age_group2 savings  552336  33987490  33989266.8
#>  5:  Total       ag2a savings  552336  33987490  33989266.8
#>  6:  Total age_group3 savings  437101  27383520  27379344.4
#>  7:  Total age_group4 savings  214661  12974149  12973913.5
#>  8:  Total age_group5 savings   80451   4728107   4729951.8
#>  9:  Total age_group6 savings    6597    325104    325640.3
#> 10:   male      Total savings 1159816  70527274  70513785.5
#> 11:      m      Total savings 1159816  70527274  70513785.5
#> 12:   male age_group1 savings  517660  30982810  30977910.4
#> 13:      m age_group1 savings  517660  30982810  30977910.4
#> 14:   male       ag1a savings  517660  30982810  30977910.4
#> 15:      m       ag1a savings  517660  30982810  30977910.4
#> 16:   male age_group2 savings  280923  17180681  17181733.0
#> 17:      m age_group2 savings  280923  17180681  17181733.0
#> 18:   male       ag2a savings  280923  17180681  17181733.0
#> 19:      m       ag2a savings  280923  17180681  17181733.0
#> 20:   male age_group3 savings  214970  13489005  13484987.4
#> 21:      m age_group3 savings  214970  13489005  13484987.4
#> 22:   male age_group4 savings   99420   6034849   6044109.4
#> 23:      m age_group4 savings   99420   6034849   6044109.4
#> 24:   male age_group5 savings   43233   2661256   2663576.4
#> 25:      m age_group5 savings   43233   2661256   2663576.4
#> 26:   male age_group6 savings    3610    178673    178237.7
#> 27:      m age_group6 savings    3610    178673    178237.7
#> 28: female      Total savings 1113716  67755939  67756442.4
#> 29:      f      Total savings 1113716  67755939  67756442.4
#> 30: female age_group1 savings  464726  27902033  27902443.2
#> 31:      f age_group1 savings  464726  27902033  27902443.2
#> 32: female       ag1a savings  464726  27902033  27902443.2
#> 33:      f       ag1a savings  464726  27902033  27902443.2
#> 34: female age_group2 savings  271413  16806809  16804984.5
#> 35:      f age_group2 savings  271413  16806809  16804984.5
#> 36: female       ag2a savings  271413  16806809  16804984.5
#> 37:      f       ag2a savings  271413  16806809  16804984.5
#> 38: female age_group3 savings  222131  13894515  13894126.6
#> 39:      f age_group3 savings  222131  13894515  13894126.6
#> 40: female age_group4 savings  115241   6939300   6941903.8
#> 41:      f age_group4 savings  115241   6939300   6941903.8
#> 42: female age_group5 savings   37218   2066851   2063812.9
#> 43:      f age_group5 savings   37218   2066851   2063812.9
#> 44: female age_group6 savings    2987    146431    146590.6
#> 45:      f age_group6 savings    2987    146431    146590.6
#>        sex        age   vname     uws        ws         pws

# results can be resetted, too
tab$reset_cntvars(v = "cnt_males")

# we can then set other parameters and perturb again
tab$params_cnts_set(val = p_cnts1, v = "cnt_males")
#> --> setting perturbation parameters for variable 'cnt_males'

tab$perturb(v = "cnt_males")
#> Count variable 'cnt_males' was perturbed.

# write results to a .csv file
tab$freqtab(
  v = c("total", "cnt_males"),
  path = file.path(tempdir(), "outtab.csv")
)
#> File '/tmp/Rtmpu511Tl/outtab.csv' successfully written to disk.
#> NULL

# show results containing weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))
#>        sex        age     vname   uwc     wc  puwc         pwc
#>     <char>     <char>    <char> <num>  <num> <num>       <num>
#>  1:  Total      Total     total  4580 277416  4580 277416.0000
#>  2:  Total age_group1     total  1969 117568  1969 117568.0000
#>  3:  Total       ag1a     total  1969 117568  1969 117568.0000
#>  4:  Total age_group2     total  1143  69592  1143  69592.0000
#>  5:  Total       ag2a     total  1143  69592  1143  69592.0000
#>  6:  Total age_group3     total   864  53968   863  53905.5370
#>  7:  Total age_group4     total   423  25741   423  25741.0000
#>  8:  Total age_group5     total   168   9790   168   9790.0000
#>  9:  Total age_group6     total    13    757    13    757.0000
#> 10:   male      Total     total  2296 138830  2297 138890.4660
#> 11:      m      Total     total  2296 138830  2297 138890.4660
#> 12:   male age_group1     total  1015  60383  1015  60383.0000
#> 13:      m age_group1     total  1015  60383  1015  60383.0000
#> 14:   male       ag1a     total  1015  60383  1015  60383.0000
#> 15:      m       ag1a     total  1015  60383  1015  60383.0000
#> 16:   male age_group2     total   571  34594   570  34533.4151
#> 17:      m age_group2     total   571  34594   570  34533.4151
#> 18:   male       ag2a     total   571  34594   570  34533.4151
#> 19:      m       ag2a     total   571  34594   570  34533.4151
#> 20:   male age_group3     total   424  26496   422  26371.0189
#> 21:      m age_group3     total   424  26496   422  26371.0189
#> 22:   male age_group4     total   195  11891   194  11830.0205
#> 23:      m age_group4     total   195  11891   194  11830.0205
#> 24:   male age_group5     total    84   5070    83   5009.6429
#> 25:      m age_group5     total    84   5070    83   5009.6429
#> 26:   male age_group6     total     7    396     9    509.1429
#> 27:      m age_group6     total     7    396     9    509.1429
#> 28: female      Total     total  2284 138586  2284 138586.0000
#> 29:      f      Total     total  2284 138586  2284 138586.0000
#> 30: female age_group1     total   954  57185   952  57065.1153
#> 31:      f age_group1     total   954  57185   952  57065.1153
#> 32: female       ag1a     total   954  57185   952  57065.1153
#> 33:      f       ag1a     total   954  57185   952  57065.1153
#> 34: female age_group2     total   572  34998   571  34936.8147
#> 35:      f age_group2     total   572  34998   571  34936.8147
#> 36: female       ag2a     total   572  34998   571  34936.8147
#> 37:      f       ag2a     total   572  34998   571  34936.8147
#> 38: female age_group3     total   440  27472   439  27409.5636
#> 39:      f age_group3     total   440  27472   439  27409.5636
#> 40: female age_group4     total   228  13850   227  13789.2544
#> 41:      f age_group4     total   228  13850   227  13789.2544
#> 42: female age_group5     total    84   4720    84   4720.0000
#> 43:      f age_group5     total    84   4720    84   4720.0000
#> 44: female age_group6     total     6    361     6    361.0000
#> 45:      f age_group6     total     6    361     6    361.0000
#> 46:  Total      Total cnt_males  2296 138830  2297 138890.4660
#> 47:  Total age_group1 cnt_males  1015  60383  1015  60383.0000
#> 48:  Total       ag1a cnt_males  1015  60383  1015  60383.0000
#> 49:  Total age_group2 cnt_males   571  34594   570  34533.4151
#> 50:  Total       ag2a cnt_males   571  34594   570  34533.4151
#> 51:  Total age_group3 cnt_males   424  26496   422  26371.0189
#> 52:  Total age_group4 cnt_males   195  11891   194  11830.0205
#> 53:  Total age_group5 cnt_males    84   5070    83   5009.6429
#> 54:  Total age_group6 cnt_males     7    396     9    509.1429
#> 55:   male      Total cnt_males  2296 138830  2297 138890.4660
#> 56:      m      Total cnt_males  2296 138830  2297 138890.4660
#> 57:   male age_group1 cnt_males  1015  60383  1015  60383.0000
#> 58:      m age_group1 cnt_males  1015  60383  1015  60383.0000
#> 59:   male       ag1a cnt_males  1015  60383  1015  60383.0000
#> 60:      m       ag1a cnt_males  1015  60383  1015  60383.0000
#> 61:   male age_group2 cnt_males   571  34594   570  34533.4151
#> 62:      m age_group2 cnt_males   571  34594   570  34533.4151
#> 63:   male       ag2a cnt_males   571  34594   570  34533.4151
#> 64:      m       ag2a cnt_males   571  34594   570  34533.4151
#> 65:   male age_group3 cnt_males   424  26496   422  26371.0189
#> 66:      m age_group3 cnt_males   424  26496   422  26371.0189
#> 67:   male age_group4 cnt_males   195  11891   194  11830.0205
#> 68:      m age_group4 cnt_males   195  11891   194  11830.0205
#> 69:   male age_group5 cnt_males    84   5070    83   5009.6429
#> 70:      m age_group5 cnt_males    84   5070    83   5009.6429
#> 71:   male age_group6 cnt_males     7    396     9    509.1429
#> 72:      m age_group6 cnt_males     7    396     9    509.1429
#> 73: female      Total cnt_males     0      0     0      0.0000
#> 74:      f      Total cnt_males     0      0     0      0.0000
#> 75: female age_group1 cnt_males     0      0     0      0.0000
#> 76:      f age_group1 cnt_males     0      0     0      0.0000
#> 77: female       ag1a cnt_males     0      0     0      0.0000
#> 78:      f       ag1a cnt_males     0      0     0      0.0000
#> 79: female age_group2 cnt_males     0      0     0      0.0000
#> 80:      f age_group2 cnt_males     0      0     0      0.0000
#> 81: female       ag2a cnt_males     0      0     0      0.0000
#> 82:      f       ag2a cnt_males     0      0     0      0.0000
#> 83: female age_group3 cnt_males     0      0     0      0.0000
#> 84:      f age_group3 cnt_males     0      0     0      0.0000
#> 85: female age_group4 cnt_males     0      0     0      0.0000
#> 86:      f age_group4 cnt_males     0      0     0      0.0000
#> 87: female age_group5 cnt_males     0      0     0      0.0000
#> 88:      f age_group5 cnt_males     0      0     0      0.0000
#> 89: female age_group6 cnt_males     0      0     0      0.0000
#> 90:      f age_group6 cnt_males     0      0     0      0.0000
#>        sex        age     vname   uwc     wc  puwc         pwc

# utility measures for a count variable
tab$measures_cnts(v = "total", exclude_zeros = TRUE)
#> $overview
#>     noise   cnt        pct
#>    <fctr> <int>      <num>
#> 1:     -2     2 0.04444444
#> 2:     -1     2 0.04444444
#> 3:      0    18 0.40000000
#> 4:      1    17 0.37777778
#> 5:      2     6 0.13333333
#> 
#> $measures
#>       what    d1    d2    d3
#>     <char> <num> <num> <num>
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 0.600 0.000 0.006
#>  6:   Mean 0.778 0.015 0.032
#>  7: Median 1.000 0.002 0.021
#>  8:    Q60 1.000 0.002 0.021
#>  9:    Q70 1.000 0.002 0.031
#> 10:    Q80 1.000 0.004 0.033
#> 11:    Q90 2.000 0.005 0.049
#> 12:    Q95 2.000 0.012 0.055
#> 13:    Q99 2.000 0.286 0.354
#> 14:    Max 2.000 0.286 0.354
#> 
#> $cumdistr_d1
#>       cat   cnt       pct
#>    <char> <int>     <num>
#> 1:      0    18 0.4000000
#> 2:      1    37 0.8222222
#> 3:      2    45 1.0000000
#> 
#> $cumdistr_d2
#>            cat   cnt       pct
#>         <char> <int>     <num>
#> 1:    [0,0.02]    43 0.9555556
#> 2: (0.02,0.05]    43 0.9555556
#> 3:  (0.05,0.1]    43 0.9555556
#> 4:   (0.1,0.2]    43 0.9555556
#> 5:   (0.2,0.3]    45 1.0000000
#> 6:   (0.3,0.4]    45 1.0000000
#> 7:   (0.4,0.5]    45 1.0000000
#> 8:   (0.5,Inf]    45 1.0000000
#> 
#> $cumdistr_d3
#>            cat   cnt       pct
#>         <char> <int>     <num>
#> 1:    [0,0.02]    21 0.4666667
#> 2: (0.02,0.05]    41 0.9111111
#> 3:  (0.05,0.1]    43 0.9555556
#> 4:   (0.1,0.2]    43 0.9555556
#> 5:   (0.2,0.3]    43 0.9555556
#> 6:   (0.3,0.4]    45 1.0000000
#> 7:   (0.4,0.5]    45 1.0000000
#> 8:   (0.5,Inf]    45 1.0000000
#> 
#> $false_zero
#> [1] 0
#> 
#> $false_nonzero
#> [1] 0
#> 
#> $exclude_zeros
#> [1] TRUE
#> 

# modifications for perturbed count variables
tab$mod_cnts()
#>         sex        age row_nr  pert      ckey  countvar
#>      <char>     <char>  <num> <int>     <num>    <char>
#>   1:  Total      Total     15     0 0.3723341     total
#>   2:  Total age_group1     15     0 0.4751934     total
#>   3:  Total       ag1a     15     0 0.4751934     total
#>   4:  Total age_group2     15     0 0.3809729     total
#>   5:  Total       ag2a     15     0 0.3809729     total
#>  ---                                                   
#> 131:      f age_group4     -1     0 0.0000000 cnt_males
#> 132: female age_group5     -1     0 0.0000000 cnt_males
#> 133:      f age_group5     -1     0 0.0000000 cnt_males
#> 134: female age_group6     -1     0 0.0000000 cnt_males
#> 135:      f age_group6     -1     0 0.0000000 cnt_males

# display a summary about utility measures
tab$summary()
#> ┌──────────────────────────────────────────────┐
#> │Utility measures for perturbed count variables│
#> └──────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#>          countvar   Min   Q10   Q20   Q30   Q40   Mean Median   Q60   Q70   Q80
#>            <char> <num> <num> <num> <num> <num>  <num>  <num> <num> <num> <num>
#> 1:          total    -2    -2    -1    -1    -1 -0.511     -1   0.0     0     0
#> 2: cnt_highincome    -4    -1    -1     0     0  0.489      0   0.4     1     2
#> 3:      cnt_males    -2    -1    -1    -1     0 -0.200      0   0.0     0     0
#>      Q90   Q95   Q99   Max
#>    <num> <num> <num> <num>
#> 1:     0   1.0     2     2
#> 2:     2   3.0     3     3
#> 3:     1   1.8     2     2
#> 
#> ── Distance-based measures ─────────────────────────────────────────────────────
#> ✔ Variable: 'total'
#> 
#>       what    d1    d2    d3
#>     <char> <num> <num> <num>
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 0.600 0.000 0.006
#>  6:   Mean 0.778 0.015 0.032
#>  7: Median 1.000 0.002 0.021
#>  8:    Q60 1.000 0.002 0.021
#>  9:    Q70 1.000 0.002 0.031
#> 10:    Q80 1.000 0.004 0.033
#> 11:    Q90 2.000 0.005 0.049
#> 12:    Q95 2.000 0.012 0.055
#> 13:    Q99 2.000 0.286 0.354
#> 14:    Max 2.000 0.286 0.354
#> 
#> ✔ Variable: 'cnt_males'
#> 
#>       what    d1    d2    d3
#>     <char> <num> <num> <num>
#>  1:    Min   0.0 0.000 0.000
#>  2:    Q10   0.0 0.000 0.000
#>  3:    Q20   0.2 0.000 0.002
#>  4:    Q30   1.0 0.000 0.010
#>  5:    Q40   1.0 0.002 0.021
#>  6:   Mean   1.0 0.035 0.061
#>  7: Median   1.0 0.002 0.021
#>  8:    Q60   1.0 0.005 0.036
#>  9:    Q70   1.0 0.005 0.049
#> 10:    Q80   1.8 0.011 0.053
#> 11:    Q90   2.0 0.121 0.175
#> 12:    Q95   2.0 0.286 0.354
#> 13:    Q99   2.0 0.286 0.354
#> 14:    Max   2.0 0.286 0.354
#> 
#> ✔ Variable: 'cnt_highincome'
#> 
#>       what    d1    d2    d3
#>     <char> <num> <num> <num>
#>  1:    Min  0.00 0.000 0.000
#>  2:    Q10  0.00 0.000 0.000
#>  3:    Q20  0.00 0.000 0.000
#>  4:    Q30  0.70 0.004 0.025
#>  5:    Q40  1.00 0.009 0.051
#>  6:   Mean  1.20 0.023 0.074
#>  7: Median  1.00 0.018 0.067
#>  8:    Q60  1.00 0.020 0.096
#>  9:    Q70  2.00 0.022 0.100
#> 10:    Q80  2.00 0.029 0.110
#> 11:    Q90  3.00 0.047 0.183
#> 12:    Q95  3.00 0.075 0.183
#> 13:    Q99  3.61 0.143 0.196
#> 14:    Max  4.00 0.143 0.196
#> 
#> ┌──────────────────────────────────────────────────┐
#> │Utility measures for perturbed numerical variables│
#> └──────────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#>      vname        Min        Q10        Q20        Q30      Q40     Mean
#>     <char>      <num>      <num>      <num>      <num>    <num>    <num>
#> 1:  expend        Inf         NA         NA         NA       NA      NaN
#> 2:  income -101181.49 -45684.822 -27950.535 -18639.757 -413.967 1292.433
#> 3: savings  -13488.49  -4899.629  -3233.991  -1824.499 -388.419 -592.252
#> 4:   mixed        Inf         NA         NA         NA       NA      NaN
#>      Median      Q60       Q70       Q80       Q90       Q95        Q99
#>       <num>    <num>     <num>     <num>     <num>     <num>      <num>
#> 1:       NA       NA        NA        NA        NA        NA         NA
#> 2: 4477.308 8263.919 17194.197 24582.073 31422.468 35441.554 126545.085
#> 3:  410.197  503.351  1051.984  1790.391  2480.401  2603.802   9260.386
#> 4:       NA       NA        NA        NA        NA        NA         NA
#>           Max
#>         <num>
#> 1:       -Inf
#> 2: 126545.085
#> 3:   9260.386
#> 4:       -Inf
# }