Create perturbation parameters for count variables

This function allows to generate required perturbation parameters that are used to perturb count variables.

ck_params_cnts(ptab, path = NULL)

Arguments

ptab

an object created with ptable::create_ptable(), or ptable::create_cnt_ptable()

path

a scalar character specifying a path to which the parameters created with this functions should be written to (in yaml format)

Value

an object suitable as input to method $params_cnts_set() for the perturbation of counts and frequencies.

See also

Examples

# \donttest{
x <- ck_create_testdata()

# create some 0/1 variables that should be perturbed later
x[, cnt_females := ifelse(sex == "male", 0, 1)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              44          25.00000           0
#>    2:   2530      28       1              49          25.00000           1
#>    3:   6920     550       1              22          25.00000           0
#>    4:   7960     870       1              79          25.00000           0
#>    5:   9030      20       2              57          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              35          16.66667           1
#> 4577:   1420     987    1000              39          16.66667           0
#> 4578:   8900     684    1000              81          16.66667           0
#> 4579:   3880     294    1000              79          16.66667           1
#> 4580:   4830     911    1000              63          16.66667           0
x[, cnt_males := ifelse(sex == "male", 1, 0)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              44          25.00000           0
#>    2:   2530      28       1              49          25.00000           1
#>    3:   6920     550       1              22          25.00000           0
#>    4:   7960     870       1              79          25.00000           0
#>    5:   9030      20       2              57          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              35          16.66667           1
#> 4577:   1420     987    1000              39          16.66667           0
#> 4578:   8900     684    1000              81          16.66667           0
#> 4579:   3880     294    1000              79          16.66667           1
#> 4580:   4830     911    1000              63          16.66667           0
#>       cnt_males
#>    1:         1
#>    2:         0
#>    3:         1
#>    4:         1
#>    5:         1
#>   ---          
#> 4576:         0
#> 4577:         1
#> 4578:         1
#> 4579:         0
#> 4580:         1
x[, cnt_highincome := ifelse(income >= 9000, 1, 0)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              44          25.00000           0
#>    2:   2530      28       1              49          25.00000           1
#>    3:   6920     550       1              22          25.00000           0
#>    4:   7960     870       1              79          25.00000           0
#>    5:   9030      20       2              57          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              35          16.66667           1
#> 4577:   1420     987    1000              39          16.66667           0
#> 4578:   8900     684    1000              81          16.66667           0
#> 4579:   3880     294    1000              79          16.66667           1
#> 4580:   4830     911    1000              63          16.66667           0
#>       cnt_males cnt_highincome
#>    1:         1              0
#>    2:         0              0
#>    3:         1              0
#>    4:         1              0
#>    5:         1              1
#>   ---                         
#> 4576:         0              0
#> 4577:         1              0
#> 4578:         1              0
#> 4579:         0              0
#> 4580:         1              0
# a variable with positive and negative contributions
x[, mixed := sample(-10:10, nrow(x), replace = TRUE)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              44          25.00000           0
#>    2:   2530      28       1              49          25.00000           1
#>    3:   6920     550       1              22          25.00000           0
#>    4:   7960     870       1              79          25.00000           0
#>    5:   9030      20       2              57          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              35          16.66667           1
#> 4577:   1420     987    1000              39          16.66667           0
#> 4578:   8900     684    1000              81          16.66667           0
#> 4579:   3880     294    1000              79          16.66667           1
#> 4580:   4830     911    1000              63          16.66667           0
#>       cnt_males cnt_highincome mixed
#>    1:         1              0    10
#>    2:         0              0    -1
#>    3:         1              0    -5
#>    4:         1              0     2
#>    5:         1              1    -9
#>   ---                               
#> 4576:         0              0     7
#> 4577:         1              0    -9
#> 4578:         1              0     1
#> 4579:         0              0     5
#> 4580:         1              0    -3

# create record keys
x$rkey <- ck_generate_rkeys(dat = x)

# define required inputs

# hierarchy with some bogus codes
d_sex <- hier_create(root = "Total", nodes = c("male", "female"))
d_sex <- hier_add(d_sex, root = "female", "f")
d_sex <- hier_add(d_sex, root = "male", "m")

d_age <- hier_create(root = "Total", nodes = paste0("age_group", 1:6))
d_age <- hier_add(d_age, root = "age_group1", "ag1a")
d_age <- hier_add(d_age, root = "age_group2", "ag2a")

# define the cell key object
countvars <- c("cnt_females", "cnt_males", "cnt_highincome")
numvars <- c("expend", "income", "savings", "mixed")
tab <- ck_setup(
  x = x,
  rkey = "rkey",
  dims = list(sex = d_sex, age = d_age),
  w = "sampling_weight",
  countvars = countvars,
  numvars = numvars)

# show some information about this table instance
tab$print() # identical with print(tab)
#> ── Table Information ───────────────────────────────────────────────────────────
#> ✔ 45 cells in 2 dimensions ('sex', 'age')
#> ✔ weights: yes
#> ── Tabulated / Perturbed countvars ─────────────────────────────────────────────
#> ☐ 'total'
#> ☐ 'cnt_females'
#> ☐ 'cnt_males'
#> ☐ 'cnt_highincome'
#> ── Tabulated / Perturbed numvars ───────────────────────────────────────────────
#> ☐ 'expend'
#> ☐ 'income'
#> ☐ 'savings'
#> ☐ 'mixed'

# information about the hierarchies
tab$hierarchy_info()
#> $sex
#>      code level is_leaf parent
#> 1:  Total     1   FALSE  Total
#> 2:   male     2   FALSE  Total
#> 3:      m     3    TRUE   male
#> 4: female     2   FALSE  Total
#> 5:      f     3    TRUE female
#> 
#> $age
#>          code level is_leaf     parent
#> 1:      Total     1   FALSE      Total
#> 2: age_group1     2   FALSE      Total
#> 3:       ag1a     3    TRUE age_group1
#> 4: age_group2     2   FALSE      Total
#> 5:       ag2a     3    TRUE age_group2
#> 6: age_group3     2    TRUE      Total
#> 7: age_group4     2    TRUE      Total
#> 8: age_group5     2    TRUE      Total
#> 9: age_group6     2    TRUE      Total
#> 

# which variables have been defined?
tab$allvars()
#> $cntvars
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"
#> 
#> $numvars
#> [1] "expend"  "income"  "savings" "mixed"  
#> 

# count variables
tab$cntvars()
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"

# continuous variables
tab$numvars()
#> [1] "expend"  "income"  "savings" "mixed"  

# create perturbation parameters for "total" variable and
# write to yaml-file

# create a ptable using functionality from the ptable-pkg
f_yaml <- tempfile(fileext = ".yaml")
p_cnts1 <- ck_params_cnts(
  ptab = ptable::pt_ex_cnts(),
  path = f_yaml)
#> yaml configuration '/tmp/RtmpCjMH4Z/file17d3400ca02d.yaml' successfully written.

# read parameters from yaml-file and set them for variable `"total"`
p_cnts1 <- ck_read_yaml(path = f_yaml)

tab$params_cnts_set(val = p_cnts1, v = "total")
#> --> setting perturbation parameters for variable 'total'

# create alternative perturbation parameters by specifying parameters
para2 <- ptable::create_cnt_ptable(
  D = 8, V = 3, js = 2, create = FALSE)

p_cnts2 <- ck_params_cnts(ptab = para2)

# use these ptable it for the remaining variables
tab$params_cnts_set(val = p_cnts2, v = countvars)
#> --> setting perturbation parameters for variable 'cnt_females'
#> --> setting perturbation parameters for variable 'cnt_males'
#> --> setting perturbation parameters for variable 'cnt_highincome'

# perturb a variable
tab$perturb(v = "total")
#> Count variable 'total' was perturbed.

# multiple variables can be perturbed as well
tab$perturb(v = c("cnt_males", "cnt_highincome"))
#> Count variable 'cnt_males' was perturbed.
#> Count variable 'cnt_highincome' was perturbed.

# return weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))
#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 274012 4579 273952.1721
#>  2:  Total age_group1     total 1969 118018 1970 118077.9380
#>  3:  Total       ag1a     total 1969 118018 1970 118077.9380
#>  4:  Total age_group2     total 1143  67604 1143  67604.0000
#>  5:  Total       ag2a     total 1143  67604 1143  67604.0000
#>  6:  Total age_group3     total  864  52253  863  52192.5220
#>  7:  Total age_group4     total  423  24847  423  24847.0000
#>  8:  Total age_group5     total  168  10542  166  10416.5000
#>  9:  Total age_group6     total   13    748   12    690.4615
#> 10:   male      Total     total 2296 138247 2296 138247.0000
#> 11:      m      Total     total 2296 138247 2296 138247.0000
#> 12:   male age_group1     total 1015  60759 1015  60759.0000
#> 13:      m age_group1     total 1015  60759 1015  60759.0000
#> 14:   male       ag1a     total 1015  60759 1015  60759.0000
#> 15:      m       ag1a     total 1015  60759 1015  60759.0000
#> 16:   male age_group2     total  571  33696  572  33755.0123
#> 17:      m age_group2     total  571  33696  572  33755.0123
#> 18:   male       ag2a     total  571  33696  572  33755.0123
#> 19:      m       ag2a     total  571  33696  572  33755.0123
#> 20:   male age_group3     total  424  26295  424  26295.0000
#> 21:      m age_group3     total  424  26295  424  26295.0000
#> 22:   male age_group4     total  195  11730  195  11730.0000
#> 23:      m age_group4     total  195  11730  195  11730.0000
#> 24:   male age_group5     total   84   5326   83   5262.5952
#> 25:      m age_group5     total   84   5326   83   5262.5952
#> 26:   male age_group6     total    7    441    6    378.0000
#> 27:      m age_group6     total    7    441    6    378.0000
#> 28: female      Total     total 2284 135765 2284 135765.0000
#> 29:      f      Total     total 2284 135765 2284 135765.0000
#> 30: female age_group1     total  954  57259  953  57198.9801
#> 31:      f age_group1     total  954  57259  953  57198.9801
#> 32: female       ag1a     total  954  57259  953  57198.9801
#> 33:      f       ag1a     total  954  57259  953  57198.9801
#> 34: female age_group2     total  572  33908  572  33908.0000
#> 35:      f age_group2     total  572  33908  572  33908.0000
#> 36: female       ag2a     total  572  33908  572  33908.0000
#> 37:      f       ag2a     total  572  33908  572  33908.0000
#> 38: female age_group3     total  440  25958  440  25958.0000
#> 39:      f age_group3     total  440  25958  440  25958.0000
#> 40: female age_group4     total  228  13117  227  13059.4693
#> 41:      f age_group4     total  228  13117  227  13059.4693
#> 42: female age_group5     total   84   5216   85   5278.0952
#> 43:      f age_group5     total   84   5216   85   5278.0952
#> 44: female age_group6     total    6    307    4    204.6667
#> 45:      f age_group6     total    6    307    4    204.6667
#> 46:  Total      Total cnt_males 2296 138247 2297 138307.2121
#> 47:  Total age_group1 cnt_males 1015  60759 1016  60818.8611
#> 48:  Total       ag1a cnt_males 1015  60759 1016  60818.8611
#> 49:  Total age_group2 cnt_males  571  33696  573  33814.0245
#> 50:  Total       ag2a cnt_males  571  33696  573  33814.0245
#> 51:  Total age_group3 cnt_males  424  26295  425  26357.0165
#> 52:  Total age_group4 cnt_males  195  11730  194  11669.8462
#> 53:  Total age_group5 cnt_males   84   5326   82   5199.1905
#> 54:  Total age_group6 cnt_males    7    441    5    315.0000
#> 55:   male      Total cnt_males 2296 138247 2297 138307.2121
#> 56:      m      Total cnt_males 2296 138247 2297 138307.2121
#> 57:   male age_group1 cnt_males 1015  60759 1016  60818.8611
#> 58:      m age_group1 cnt_males 1015  60759 1016  60818.8611
#> 59:   male       ag1a cnt_males 1015  60759 1016  60818.8611
#> 60:      m       ag1a cnt_males 1015  60759 1016  60818.8611
#> 61:   male age_group2 cnt_males  571  33696  573  33814.0245
#> 62:      m age_group2 cnt_males  571  33696  573  33814.0245
#> 63:   male       ag2a cnt_males  571  33696  573  33814.0245
#> 64:      m       ag2a cnt_males  571  33696  573  33814.0245
#> 65:   male age_group3 cnt_males  424  26295  425  26357.0165
#> 66:      m age_group3 cnt_males  424  26295  425  26357.0165
#> 67:   male age_group4 cnt_males  195  11730  194  11669.8462
#> 68:      m age_group4 cnt_males  195  11730  194  11669.8462
#> 69:   male age_group5 cnt_males   84   5326   82   5199.1905
#> 70:      m age_group5 cnt_males   84   5326   82   5199.1905
#> 71:   male age_group6 cnt_males    7    441    5    315.0000
#> 72:      m age_group6 cnt_males    7    441    5    315.0000
#> 73: female      Total cnt_males    0      0    0      0.0000
#> 74:      f      Total cnt_males    0      0    0      0.0000
#> 75: female age_group1 cnt_males    0      0    0      0.0000
#> 76:      f age_group1 cnt_males    0      0    0      0.0000
#> 77: female       ag1a cnt_males    0      0    0      0.0000
#> 78:      f       ag1a cnt_males    0      0    0      0.0000
#> 79: female age_group2 cnt_males    0      0    0      0.0000
#> 80:      f age_group2 cnt_males    0      0    0      0.0000
#> 81: female       ag2a cnt_males    0      0    0      0.0000
#> 82:      f       ag2a cnt_males    0      0    0      0.0000
#> 83: female age_group3 cnt_males    0      0    0      0.0000
#> 84:      f age_group3 cnt_males    0      0    0      0.0000
#> 85: female age_group4 cnt_males    0      0    0      0.0000
#> 86:      f age_group4 cnt_males    0      0    0      0.0000
#> 87: female age_group5 cnt_males    0      0    0      0.0000
#> 88:      f age_group5 cnt_males    0      0    0      0.0000
#> 89: female age_group6 cnt_males    0      0    0      0.0000
#> 90:      f age_group6 cnt_males    0      0    0      0.0000
#>        sex        age     vname  uwc     wc puwc         pwc

# numerical variables (positive variables using flex-function)
# we also write the config to a yaml file
f_yaml <- tempfile(fileext = ".yaml")

# create a ptable using functionality from the ptable-pkg
# a single ptable for all cells
ptab1 <- ptable::pt_ex_nums(parity = TRUE, separation = FALSE)

# a single ptab for all cells except for very small ones
ptab2 <- ptable::pt_ex_nums(parity = TRUE, separation = TRUE)

# different ptables for cells with even/odd number of contributors
# and very small cells
ptab3 <- ptable::pt_ex_nums(parity = FALSE, separation = TRUE)

p_nums1 <- ck_params_nums(
  ptab = ptab1,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.30, 0.03),
    epsilon = c(1, 0.5, 0.2),
    q = 3),
  mu_c = 2,
  same_key = FALSE,
  use_zero_rkeys = FALSE,
  path = f_yaml)
#> yaml configuration '/tmp/RtmpCjMH4Z/file17d35d43ec70.yaml' successfully written.

# we read the parameters from the yaml-file
p_nums1 <- ck_read_yaml(path = f_yaml)

# for variables with positive and negative values
p_nums2 <- ck_params_nums(
  ptab = ptab2,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.15, 0.02),
    epsilon = c(1, 0.4, 0.15),
    q = 3),
  mu_c = 2,
  same_key = FALSE)

# simple perturbation parameters (not using the flex-function approach)
p_nums3 <- ck_params_nums(
  ptab = ptab3,
  type = "mean",
  mult_params = ck_simpleparams(p = 0.25),
  mu_c = 2,
  same_key = FALSE)

# use `p_nums1` for all variables
tab$params_nums_set(p_nums1, c("savings", "income", "expend"))
#> --> setting perturbation parameters for variable 'savings'
#> --> setting perturbation parameters for variable 'income'
#> --> setting perturbation parameters for variable 'expend'

# use different parameters for variable `mixed`
tab$params_nums_set(p_nums2, v = "mixed")
#> --> setting perturbation parameters for variable 'mixed'

# identify sensitive cells to which extra protection (`mu_c`) is added.
tab$supp_p(v = "income", p = 85)
#> computing contributing indices | rawdata <--> table; this might take a while
#> p%-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_pq(v = "income", p = 85, q = 90)
#> computing contributing indices | rawdata <--> table; this might take a while
#> pq-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_nk(v = "income", n = 2, k = 90)
#> computing contributing indices | rawdata <--> table; this might take a while
#> nk-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_freq(v = "income", n = 14, weighted = FALSE)
#> freq-rule: 5 new sensitive cells (incl. duplicates) found (total: 5)
tab$supp_val(v = "income", n = 10000, weighted = TRUE)
#> val-rule: 0 new sensitive cells (incl. duplicates) found (total: 5)
tab$supp_cells(
  v = "income",
  inp = data.frame(
    sex = c("female", "female"),
    "age" = c("age_group1", "age_group3")
  )
)
#> cell-rule: 2 new sensitive cells (incl. duplicates) found (total: 7)

# perturb variables
tab$perturb(v = c("income", "savings"))
#> Numeric variable 'income' was perturbed.
#> Numeric variable 'savings' was perturbed.

# extract results
tab$numtab("income", mean_before_sum = TRUE)
#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1371982329 1372004685
#>  2:  Total age_group1 income  9810547  589269465  589289752
#>  3:  Total       ag1a income  9810547  589269465  589289752
#>  4:  Total age_group2 income  5692119  334492243  334541647
#>  5:  Total       ag2a income  5692119  334492243  334541647
#>  6:  Total age_group3 income  4406946  265612006  265617191
#>  7:  Total age_group4 income  2133543  124291539  124276945
#>  8:  Total age_group5 income   848151   54654207   54575248
#>  9:  Total age_group6 income    61672    3662869    3642184
#> 10:   male      Total income 11262049  677348661  677477635
#> 11:      m      Total income 11262049  677348661  677477635
#> 12:   male age_group1 income  4877164  291090250  291038963
#> 13:      m age_group1 income  4877164  291090250  291038963
#> 14:   male       ag1a income  4877164  291090250  291038963
#> 15:      m       ag1a income  4877164  291090250  291038963
#> 16:   male age_group2 income  2811379  165158540  164990878
#> 17:      m age_group2 income  2811379  165158540  164990878
#> 18:   male       ag2a income  2811379  165158540  164990878
#> 19:      m       ag2a income  2811379  165158540  164990878
#> 20:   male age_group3 income  2168169  133766353  133896198
#> 21:      m age_group3 income  2168169  133766353  133896198
#> 22:   male age_group4 income   978510   59444099   59376503
#> 23:      m age_group4 income   978510   59444099   59376503
#> 24:   male age_group5 income   393134   25778429   25901125
#> 25:      m age_group5 income   393134   25778429   25901125
#> 26:   male age_group6 income    33693    2110990    1897422
#> 27:      m age_group6 income    33693    2110990    1897422
#> 28: female      Total income 11690929  694633668  694699285
#> 29:      f      Total income 11690929  694633668  694699285
#> 30: female age_group1 income  4933383  298179215  298130458
#> 31:      f age_group1 income  4933383  298179215  298130458
#> 32: female       ag1a income  4933383  298179215  298130458
#> 33:      f       ag1a income  4933383  298179215  298130458
#> 34: female age_group2 income  2880740  169333703  169312379
#> 35:      f age_group2 income  2880740  169333703  169312379
#> 36: female       ag2a income  2880740  169333703  169312379
#> 37:      f       ag2a income  2880740  169333703  169312379
#> 38: female age_group3 income  2238777  131845653  131899350
#> 39:      f age_group3 income  2238777  131845653  131899350
#> 40: female age_group4 income  1155033   64847440   64820850
#> 41:      f age_group4 income  1155033   64847440   64820850
#> 42: female age_group5 income   455017   28875778   28856469
#> 43:      f age_group5 income   455017   28875778   28856469
#> 44: female age_group6 income    27979    1551879    1549871
#> 45:      f age_group6 income    27979    1551879    1549871
#>        sex        age  vname      uws         ws        pws
tab$numtab("income", mean_before_sum = FALSE)
#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1371982329 1371993507
#>  2:  Total age_group1 income  9810547  589269465  589279609
#>  3:  Total       ag1a income  9810547  589269465  589279609
#>  4:  Total age_group2 income  5692119  334492243  334516944
#>  5:  Total       ag2a income  5692119  334492243  334516944
#>  6:  Total age_group3 income  4406946  265612006  265614599
#>  7:  Total age_group4 income  2133543  124291539  124284242
#>  8:  Total age_group5 income   848151   54654207   54614713
#>  9:  Total age_group6 income    61672    3662869    3652512
#> 10:   male      Total income 11262049  677348661  677413145
#> 11:      m      Total income 11262049  677348661  677413145
#> 12:   male age_group1 income  4877164  291090250  291064605
#> 13:      m age_group1 income  4877164  291090250  291064605
#> 14:   male       ag1a income  4877164  291090250  291064605
#> 15:      m       ag1a income  4877164  291090250  291064605
#> 16:   male age_group2 income  2811379  165158540  165074688
#> 17:      m age_group2 income  2811379  165158540  165074688
#> 18:   male       ag2a income  2811379  165158540  165074688
#> 19:      m       ag2a income  2811379  165158540  165074688
#> 20:   male age_group3 income  2168169  133766353  133831260
#> 21:      m age_group3 income  2168169  133766353  133831260
#> 22:   male age_group4 income   978510   59444099   59410291
#> 23:      m age_group4 income   978510   59444099   59410291
#> 24:   male age_group5 income   393134   25778429   25839704
#> 25:      m age_group5 income   393134   25778429   25839704
#> 26:   male age_group6 income    33693    2110990    2001359
#> 27:      m age_group6 income    33693    2110990    2001359
#> 28: female      Total income 11690929  694633668  694666476
#> 29:      f      Total income 11690929  694633668  694666476
#> 30: female age_group1 income  4933383  298179215  298154836
#> 31:      f age_group1 income  4933383  298179215  298154836
#> 32: female       ag1a income  4933383  298179215  298154836
#> 33:      f       ag1a income  4933383  298179215  298154836
#> 34: female age_group2 income  2880740  169333703  169323040
#> 35:      f age_group2 income  2880740  169333703  169323040
#> 36: female       ag2a income  2880740  169333703  169323040
#> 37:      f       ag2a income  2880740  169333703  169323040
#> 38: female age_group3 income  2238777  131845653  131872499
#> 39:      f age_group3 income  2238777  131845653  131872499
#> 40: female age_group4 income  1155033   64847440   64834144
#> 41:      f age_group4 income  1155033   64847440   64834144
#> 42: female age_group5 income   455017   28875778   28866122
#> 43:      f age_group5 income   455017   28875778   28866122
#> 44: female age_group6 income    27979    1551879    1550875
#> 45:      f age_group6 income    27979    1551879    1550875
#>        sex        age  vname      uws         ws        pws
tab$numtab("savings")
#>        sex        age   vname     uws        ws         pws
#>  1:  Total      Total savings 2273532 136051811 136051965.6
#>  2:  Total age_group1 savings  982386  58856562  58862789.6
#>  3:  Total       ag1a savings  982386  58856562  58862789.6
#>  4:  Total age_group2 savings  552336  32652530  32653986.7
#>  5:  Total       ag2a savings  552336  32652530  32653986.7
#>  6:  Total age_group3 savings  437101  26331758  26331238.4
#>  7:  Total age_group4 savings  214661  12682803  12682306.5
#>  8:  Total age_group5 savings   80451   5058107   5055610.4
#>  9:  Total age_group6 savings    6597    470051    470641.5
#> 10:   male      Total savings 1159816  69605585  69600858.5
#> 11:      m      Total savings 1159816  69605585  69600858.5
#> 12:   male age_group1 savings  517660  30833206  30830959.4
#> 13:      m age_group1 savings  517660  30833206  30830959.4
#> 14:   male       ag1a savings  517660  30833206  30830959.4
#> 15:      m       ag1a savings  517660  30833206  30830959.4
#> 16:   male age_group2 savings  280923  16650754  16642758.6
#> 17:      m age_group2 savings  280923  16650754  16642758.6
#> 18:   male       ag2a savings  280923  16650754  16642758.6
#> 19:      m       ag2a savings  280923  16650754  16642758.6
#> 20:   male age_group3 savings  214970  13073983  13073802.7
#> 21:      m age_group3 savings  214970  13073983  13073802.7
#> 22:   male age_group4 savings   99420   6004391   6001375.4
#> 23:      m age_group4 savings   99420   6004391   6001375.4
#> 24:   male age_group5 savings   43233   2773310   2777224.6
#> 25:      m age_group5 savings   43233   2773310   2777224.6
#> 26:   male age_group6 savings    3610    269941    259489.9
#> 27:      m age_group6 savings    3610    269941    259489.9
#> 28: female      Total savings 1113716  66446226  66454327.8
#> 29:      f      Total savings 1113716  66446226  66454327.8
#> 30: female age_group1 savings  464726  28023356  28022353.7
#> 31:      f age_group1 savings  464726  28023356  28022353.7
#> 32: female       ag1a savings  464726  28023356  28022353.7
#> 33:      f       ag1a savings  464726  28023356  28022353.7
#> 34: female age_group2 savings  271413  16001776  16000857.1
#> 35:      f age_group2 savings  271413  16001776  16000857.1
#> 36: female       ag2a savings  271413  16001776  16000857.1
#> 37:      f       ag2a savings  271413  16001776  16000857.1
#> 38: female age_group3 savings  222131  13257775  13260230.6
#> 39:      f age_group3 savings  222131  13257775  13260230.6
#> 40: female age_group4 savings  115241   6678412   6678877.2
#> 41:      f age_group4 savings  115241   6678412   6678877.2
#> 42: female age_group5 savings   37218   2284797   2284657.6
#> 43:      f age_group5 savings   37218   2284797   2284657.6
#> 44: female age_group6 savings    2987    200110    198548.5
#> 45:      f age_group6 savings    2987    200110    198548.5
#>        sex        age   vname     uws        ws         pws

# results can be resetted, too
tab$reset_cntvars(v = "cnt_males")

# we can then set other parameters and perturb again
tab$params_cnts_set(val = p_cnts1, v = "cnt_males")
#> --> setting perturbation parameters for variable 'cnt_males'

tab$perturb(v = "cnt_males")
#> Count variable 'cnt_males' was perturbed.

# write results to a .csv file
tab$freqtab(
  v = c("total", "cnt_males"),
  path = file.path(tempdir(), "outtab.csv")
)
#> File '/tmp/RtmpCjMH4Z/outtab.csv' successfully written to disk.
#> NULL

# show results containing weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))
#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 274012 4579 273952.1721
#>  2:  Total age_group1     total 1969 118018 1970 118077.9380
#>  3:  Total       ag1a     total 1969 118018 1970 118077.9380
#>  4:  Total age_group2     total 1143  67604 1143  67604.0000
#>  5:  Total       ag2a     total 1143  67604 1143  67604.0000
#>  6:  Total age_group3     total  864  52253  863  52192.5220
#>  7:  Total age_group4     total  423  24847  423  24847.0000
#>  8:  Total age_group5     total  168  10542  166  10416.5000
#>  9:  Total age_group6     total   13    748   12    690.4615
#> 10:   male      Total     total 2296 138247 2296 138247.0000
#> 11:      m      Total     total 2296 138247 2296 138247.0000
#> 12:   male age_group1     total 1015  60759 1015  60759.0000
#> 13:      m age_group1     total 1015  60759 1015  60759.0000
#> 14:   male       ag1a     total 1015  60759 1015  60759.0000
#> 15:      m       ag1a     total 1015  60759 1015  60759.0000
#> 16:   male age_group2     total  571  33696  572  33755.0123
#> 17:      m age_group2     total  571  33696  572  33755.0123
#> 18:   male       ag2a     total  571  33696  572  33755.0123
#> 19:      m       ag2a     total  571  33696  572  33755.0123
#> 20:   male age_group3     total  424  26295  424  26295.0000
#> 21:      m age_group3     total  424  26295  424  26295.0000
#> 22:   male age_group4     total  195  11730  195  11730.0000
#> 23:      m age_group4     total  195  11730  195  11730.0000
#> 24:   male age_group5     total   84   5326   83   5262.5952
#> 25:      m age_group5     total   84   5326   83   5262.5952
#> 26:   male age_group6     total    7    441    6    378.0000
#> 27:      m age_group6     total    7    441    6    378.0000
#> 28: female      Total     total 2284 135765 2284 135765.0000
#> 29:      f      Total     total 2284 135765 2284 135765.0000
#> 30: female age_group1     total  954  57259  953  57198.9801
#> 31:      f age_group1     total  954  57259  953  57198.9801
#> 32: female       ag1a     total  954  57259  953  57198.9801
#> 33:      f       ag1a     total  954  57259  953  57198.9801
#> 34: female age_group2     total  572  33908  572  33908.0000
#> 35:      f age_group2     total  572  33908  572  33908.0000
#> 36: female       ag2a     total  572  33908  572  33908.0000
#> 37:      f       ag2a     total  572  33908  572  33908.0000
#> 38: female age_group3     total  440  25958  440  25958.0000
#> 39:      f age_group3     total  440  25958  440  25958.0000
#> 40: female age_group4     total  228  13117  227  13059.4693
#> 41:      f age_group4     total  228  13117  227  13059.4693
#> 42: female age_group5     total   84   5216   85   5278.0952
#> 43:      f age_group5     total   84   5216   85   5278.0952
#> 44: female age_group6     total    6    307    4    204.6667
#> 45:      f age_group6     total    6    307    4    204.6667
#> 46:  Total      Total cnt_males 2296 138247 2296 138247.0000
#> 47:  Total age_group1 cnt_males 1015  60759 1015  60759.0000
#> 48:  Total       ag1a cnt_males 1015  60759 1015  60759.0000
#> 49:  Total age_group2 cnt_males  571  33696  572  33755.0123
#> 50:  Total       ag2a cnt_males  571  33696  572  33755.0123
#> 51:  Total age_group3 cnt_males  424  26295  424  26295.0000
#> 52:  Total age_group4 cnt_males  195  11730  195  11730.0000
#> 53:  Total age_group5 cnt_males   84   5326   83   5262.5952
#> 54:  Total age_group6 cnt_males    7    441    6    378.0000
#> 55:   male      Total cnt_males 2296 138247 2296 138247.0000
#> 56:      m      Total cnt_males 2296 138247 2296 138247.0000
#> 57:   male age_group1 cnt_males 1015  60759 1015  60759.0000
#> 58:      m age_group1 cnt_males 1015  60759 1015  60759.0000
#> 59:   male       ag1a cnt_males 1015  60759 1015  60759.0000
#> 60:      m       ag1a cnt_males 1015  60759 1015  60759.0000
#> 61:   male age_group2 cnt_males  571  33696  572  33755.0123
#> 62:      m age_group2 cnt_males  571  33696  572  33755.0123
#> 63:   male       ag2a cnt_males  571  33696  572  33755.0123
#> 64:      m       ag2a cnt_males  571  33696  572  33755.0123
#> 65:   male age_group3 cnt_males  424  26295  424  26295.0000
#> 66:      m age_group3 cnt_males  424  26295  424  26295.0000
#> 67:   male age_group4 cnt_males  195  11730  195  11730.0000
#> 68:      m age_group4 cnt_males  195  11730  195  11730.0000
#> 69:   male age_group5 cnt_males   84   5326   83   5262.5952
#> 70:      m age_group5 cnt_males   84   5326   83   5262.5952
#> 71:   male age_group6 cnt_males    7    441    6    378.0000
#> 72:      m age_group6 cnt_males    7    441    6    378.0000
#> 73: female      Total cnt_males    0      0    0      0.0000
#> 74:      f      Total cnt_males    0      0    0      0.0000
#> 75: female age_group1 cnt_males    0      0    0      0.0000
#> 76:      f age_group1 cnt_males    0      0    0      0.0000
#> 77: female       ag1a cnt_males    0      0    0      0.0000
#> 78:      f       ag1a cnt_males    0      0    0      0.0000
#> 79: female age_group2 cnt_males    0      0    0      0.0000
#> 80:      f age_group2 cnt_males    0      0    0      0.0000
#> 81: female       ag2a cnt_males    0      0    0      0.0000
#> 82:      f       ag2a cnt_males    0      0    0      0.0000
#> 83: female age_group3 cnt_males    0      0    0      0.0000
#> 84:      f age_group3 cnt_males    0      0    0      0.0000
#> 85: female age_group4 cnt_males    0      0    0      0.0000
#> 86:      f age_group4 cnt_males    0      0    0      0.0000
#> 87: female age_group5 cnt_males    0      0    0      0.0000
#> 88:      f age_group5 cnt_males    0      0    0      0.0000
#> 89: female age_group6 cnt_males    0      0    0      0.0000
#> 90:      f age_group6 cnt_males    0      0    0      0.0000
#>        sex        age     vname  uwc     wc puwc         pwc

# utility measures for a count variable
tab$measures_cnts(v = "total", exclude_zeros = TRUE)
#> $overview
#>    noise cnt        pct
#> 1:    -1   8 0.17777778
#> 2:     0  21 0.46666667
#> 3:     1  13 0.28888889
#> 4:     2   3 0.06666667
#> 
#> $measures
#>       what  d1    d2    d3
#>  1:    Min 0.0 0.000 0.000
#>  2:    Q10 0.0 0.000 0.000
#>  3:    Q20 0.0 0.000 0.000
#>  4:    Q30 0.0 0.000 0.000
#>  5:    Q40 0.0 0.000 0.000
#>  6:   Mean 0.6 0.025 0.044
#>  7: Median 1.0 0.001 0.011
#>  8:    Q60 1.0 0.001 0.016
#>  9:    Q70 1.0 0.002 0.021
#> 10:    Q80 1.0 0.012 0.054
#> 11:    Q90 1.0 0.051 0.116
#> 12:    Q95 1.8 0.143 0.196
#> 13:    Q99 2.0 0.333 0.449
#> 14:    Max 2.0 0.333 0.449
#> 
#> $cumdistr_d1
#>    cat cnt       pct
#> 1:   0  21 0.4666667
#> 2:   1  42 0.9333333
#> 3:   2  45 1.0000000
#> 
#> $cumdistr_d2
#>            cat cnt       pct
#> 1:    [0,0.02]  40 0.8888889
#> 2: (0.02,0.05]  40 0.8888889
#> 3:  (0.05,0.1]  41 0.9111111
#> 4:   (0.1,0.2]  43 0.9555556
#> 5:   (0.2,0.3]  43 0.9555556
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $cumdistr_d3
#>            cat cnt       pct
#> 1:    [0,0.02]  29 0.6444444
#> 2: (0.02,0.05]  35 0.7777778
#> 3:  (0.05,0.1]  40 0.8888889
#> 4:   (0.1,0.2]  43 0.9555556
#> 5:   (0.2,0.3]  43 0.9555556
#> 6:   (0.3,0.4]  43 0.9555556
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $false_zero
#> [1] 0
#> 
#> $false_nonzero
#> [1] 0
#> 
#> $exclude_zeros
#> [1] TRUE
#> 

# modifications for perturbed count variables
tab$mod_cnts()
#>         sex        age row_nr pert      ckey  countvar
#>   1:  Total      Total     14   -1 0.1183030     total
#>   2:  Total age_group1     16    1 0.7652018     total
#>   3:  Total       ag1a     16    1 0.7652018     total
#>   4:  Total age_group2     15    0 0.3815609     total
#>   5:  Total       ag2a     15    0 0.3815609     total
#>  ---                                                  
#> 131:      f age_group4     -1    0 0.0000000 cnt_males
#> 132: female age_group5     -1    0 0.0000000 cnt_males
#> 133:      f age_group5     -1    0 0.0000000 cnt_males
#> 134: female age_group6     -1    0 0.0000000 cnt_males
#> 135:      f age_group6     -1    0 0.0000000 cnt_males

# display a summary about utility measures
tab$summary()
#> ┌──────────────────────────────────────────────┐
#> │Utility measures for perturbed count variables│
#> └──────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#>          countvar Min Q10 Q20 Q30 Q40   Mean Median Q60 Q70 Q80 Q90 Q95 Q99 Max
#> 1:          total  -2  -1  -1  -1   0 -0.244      0   0 0.0   0   1   1   1   1
#> 2: cnt_highincome  -2  -1  -1   0   0  0.489      0   1 1.8   2   2   2   3   3
#> 3:      cnt_males  -1  -1   0   0   0  0.000      0   0 0.0   0   1   1   1   1
#> 
#> ── Distance-based measures ─────────────────────────────────────────────────────
#> ✔ Variable: 'total'
#> 
#>       what  d1    d2    d3
#>  1:    Min 0.0 0.000 0.000
#>  2:    Q10 0.0 0.000 0.000
#>  3:    Q20 0.0 0.000 0.000
#>  4:    Q30 0.0 0.000 0.000
#>  5:    Q40 0.0 0.000 0.000
#>  6:   Mean 0.6 0.025 0.044
#>  7: Median 1.0 0.001 0.011
#>  8:    Q60 1.0 0.001 0.016
#>  9:    Q70 1.0 0.002 0.021
#> 10:    Q80 1.0 0.012 0.054
#> 11:    Q90 1.0 0.051 0.116
#> 12:    Q95 1.8 0.143 0.196
#> 13:    Q99 2.0 0.333 0.449
#> 14:    Max 2.0 0.333 0.449
#> 
#> ✔ Variable: 'cnt_males'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 0.000 0.000 0.000
#>  6:   Mean 0.444 0.018 0.033
#>  7: Median 0.000 0.000 0.000
#>  8:    Q60 1.000 0.002 0.021
#>  9:    Q70 1.000 0.002 0.021
#> 10:    Q80 1.000 0.010 0.048
#> 11:    Q90 1.000 0.064 0.111
#> 12:    Q95 1.000 0.143 0.196
#> 13:    Q99 1.000 0.143 0.196
#> 14:    Max 1.000 0.143 0.196
#> 
#> ✔ Variable: 'cnt_highincome'
#> 
#>       what   d1    d2    d3
#>  1:    Min 0.00 0.000 0.000
#>  2:    Q10 0.00 0.000 0.000
#>  3:    Q20 0.00 0.000 0.000
#>  4:    Q30 1.00 0.004 0.033
#>  5:    Q40 1.00 0.009 0.059
#>  6:   Mean 1.20 0.025 0.079
#>  7: Median 1.00 0.015 0.065
#>  8:    Q60 1.40 0.019 0.096
#>  9:    Q70 2.00 0.026 0.107
#> 10:    Q80 2.00 0.038 0.131
#> 11:    Q90 2.00 0.054 0.134
#> 12:    Q95 2.05 0.068 0.159
#> 13:    Q99 3.00 0.165 0.331
#> 14:    Max 3.00 0.214 0.381
#> 
#> ┌──────────────────────────────────────────────────┐
#> │Utility measures for perturbed numerical variables│
#> └──────────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#>      vname        Min        Q10        Q20        Q30        Q40      Mean
#> 1:  expend        Inf         NA         NA         NA         NA       NaN
#> 2:  income -109630.54 -83852.236 -27277.298 -24379.487 -13296.437 -8577.673
#> 3: savings  -10451.07  -7995.418  -3015.612  -2246.639  -1002.324 -1029.476
#> 4:   mixed        Inf         NA         NA         NA         NA       NaN
#>        Median       Q60       Q70       Q80       Q90      Q95       Q99
#> 1:         NA        NA        NA        NA        NA       NA        NA
#> 2: -10662.513 -4780.027 10143.606 26845.617 61275.242 64484.06 64906.827
#> 3:   -918.898  -370.017    95.754  1456.663  3914.559  6227.62  8101.841
#> 4:         NA        NA        NA        NA        NA       NA        NA
#>          Max
#> 1:      -Inf
#> 2: 64906.827
#> 3:  8101.841
#> 4:      -Inf
# }