Create perturbation parameters for count variables

This function allows to generate required perturbation parameters that are used to perturb count variables.

ck_params_cnts(D, V, js = 0, pstay = NULL, optim = 1, mono = TRUE,
  epsilon = 0.0000001)

Arguments

D
perturbation parameter for maximum perturbation (scalar or vector)

V
perturbation parameter for variance (scalar)

js
treshold value for blocking of small frequencies (i.e. there won't occur positive target frequencies below the treshold value)

pstay
optional parameter to set

optim
optimization parameter: 1 standard approach (default)

mono
(logical) vector specifying optimization parameter for monotony condition

epsilon
(double)

Value

an object suitable as input to cellkey_pkg() for the perturbation of counts and frequencies.

See also

Examples

x <- ck_create_testdata()

# create some 0/1 variables that should be perturbed later
x[, cnt_females := ifelse(sex == "male", 0, 1)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              85          25.00000           0
#>    2:   2530      28       1              80          25.00000           1
#>    3:   6920     550       1              34          25.00000           0
#>    4:   7960     870       1              34          25.00000           0
#>    5:   9030      20       2              99          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              42          16.66667           1
#> 4577:   1420     987    1000              40          16.66667           0
#> 4578:   8900     684    1000              20          16.66667           0
#> 4579:   3880     294    1000              55          16.66667           1
#> 4580:   4830     911    1000              93          16.66667           0

x[, cnt_males := ifelse(sex == "male", 1, 0)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              85          25.00000           0
#>    2:   2530      28       1              80          25.00000           1
#>    3:   6920     550       1              34          25.00000           0
#>    4:   7960     870       1              34          25.00000           0
#>    5:   9030      20       2              99          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              42          16.66667           1
#> 4577:   1420     987    1000              40          16.66667           0
#> 4578:   8900     684    1000              20          16.66667           0
#> 4579:   3880     294    1000              55          16.66667           1
#> 4580:   4830     911    1000              93          16.66667           0
#>       cnt_males
#>    1:         1
#>    2:         0
#>    3:         1
#>    4:         1
#>    5:         1
#>   ---          
#> 4576:         0
#> 4577:         1
#> 4578:         1
#> 4579:         0
#> 4580:         1

x[, cnt_highincome := ifelse(income >= 9000, 1, 0)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              85          25.00000           0
#>    2:   2530      28       1              80          25.00000           1
#>    3:   6920     550       1              34          25.00000           0
#>    4:   7960     870       1              34          25.00000           0
#>    5:   9030      20       2              99          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              42          16.66667           1
#> 4577:   1420     987    1000              40          16.66667           0
#> 4578:   8900     684    1000              20          16.66667           0
#> 4579:   3880     294    1000              55          16.66667           1
#> 4580:   4830     911    1000              93          16.66667           0
#>       cnt_males cnt_highincome
#>    1:         1              0
#>    2:         0              0
#>    3:         1              0
#>    4:         1              0
#>    5:         1              1
#>   ---                         
#> 4576:         0              0
#> 4577:         1              0
#> 4578:         1              0
#> 4579:         0              0
#> 4580:         1              0

# create record keys
x$rkey <- ck_generate_rkeys(dat = x)

# define required inputs

# hierarchy with some bogus codes
d_sex <- hier_create(root = "Total", nodes = c("male", "female"))
d_sex <- hier_add(d_sex, root = "female", "f")
d_sex <- hier_add(d_sex, root = "male", "m")

d_age <- hier_create(root = "Total", nodes = paste0("age_group", 1:6))
d_age <- hier_add(d_age, root = "age_group1", "ag1a")
d_age <- hier_add(d_age, root = "age_group2", "ag2a")

# create a named list
dims <- list(sex = d_sex, age = d_age)
w <- "sampling_weight"
countvars <- c("cnt_females", "cnt_males", "cnt_highincome")
rkey <- "rkey"

# perturbation parameters for count variables
params_cnts <- ck_params_cnts(
  D = 5,
  V = 3,
  js = 2,
  pstay = 0.5,
  optim = 1,
  mono = TRUE
)

# define the cell key object
tab <- ck_setup(
  x = x,
  rkey = 5,
  dims = dims,
  w = w,
  countvars = countvars,
  params_cnts = params_cnts
)

# show some information about this table instance
print(tab) # identical with tab$print()

#> ── Table Information ───────────────────────────────────────────────────────────
#> ✔ 45 cells in 2 dimensions ('sex', 'age')
#> ✔ weights: yes
#> ── Tabulated / Perturbed countvars ─────────────────────────────────────────────
#> ☐ 'total'
#> ☐ 'cnt_females'
#> ☐ 'cnt_males'
#> ☐ 'cnt_highincome'

# perturb a single variable
tab$perturb(v = "total")

#> Count variable 'total' was perturbed.

# modify the perturbation parameters
# create alternative perturbation parameters
params_cnts_alt <- ck_params_cnts(
  D = 8,
  V = 3,
  js = 2,
  pstay = 0.5,
  optim = 1,
  mono = TRUE
)
# we creplace the current parameters with the alternatives
# they will be used from now on
tab$params_cnts(params_cnts_alt)

#> Perturbation parameters for count variables were modified.

# multiple variables can be perturbed as well
tab$perturb(v = c("cnt_males", "cnt_highincome"))

#> Count variable 'cnt_males' was perturbed.

#> Count variable 'cnt_highincome' was perturbed.

# return results (weighted and unweighted)
tab$freqtab(v = c("total", "cnt_males"))

#>        sex        age     vname  uwc     wc puwc        pwc
#>  1:  Total      Total     total 4580 276667 4581 4641.40764
#>  2:  Total age_group1     total 1969 119983 1966 2026.93601
#>  3:  Total       ag1a     total 1969 119983 1966 2026.93601
#>  4:  Total age_group2     total 1143  67768 1144 1203.28959
#>  5:  Total       ag2a     total 1143  67768 1144 1203.28959
#>  6:  Total age_group3     total  864  51533  865  924.64468
#>  7:  Total age_group4     total  423  26067  423  484.62411
#>  8:  Total age_group5     total  168  10643  168  231.35119
#>  9:  Total age_group6     total   13    673   11   62.76923
#> 10: female      Total     total 2284 138219 2284 2344.51620
#> 11:      f      Total     total 2284 138219 2284 2344.51620
#> 12: female age_group1     total  954  58122  958 1018.92453
#> 13:      f age_group1     total  954  58122  958 1018.92453
#> 14: female       ag1a     total  954  58122  958 1018.92453
#> 15:      f       ag1a     total  954  58122  958 1018.92453
#> 16: female age_group2     total  572  34313  572  631.98776
#> 17:      f age_group2     total  572  34313  572  631.98776
#> 18: female       ag2a     total  572  34313  572  631.98776
#> 19:      f       ag2a     total  572  34313  572  631.98776
#> 20: female age_group3     total  440  26374  440  499.94091
#> 21:      f age_group3     total  440  26374  440  499.94091
#> 22: female age_group4     total  228  13990  228  289.35965
#> 23:      f age_group4     total  228  13990  228  289.35965
#> 24: female age_group5     total   84   5170   87  148.54762
#> 25:      f age_group5     total   84   5170   87  148.54762
#> 26: female age_group6     total    6    250    4   45.66667
#> 27:      f age_group6     total    6    250    4   45.66667
#> 28:   male      Total     total 2296 138448 2296 2356.29965
#> 29:      m      Total     total 2296 138448 2296 2356.29965
#> 30:   male age_group1     total 1015  61861 1013 1073.94680
#> 31:      m age_group1     total 1015  61861 1013 1073.94680
#> 32:   male       ag1a     total 1015  61861 1013 1073.94680
#> 33:      m       ag1a     total 1015  61861 1013 1073.94680
#> 34:   male age_group2     total  571  33455  570  628.59019
#> 35:      m age_group2     total  571  33455  570  628.59019
#> 36:   male       ag2a     total  571  33455  570  628.59019
#> 37:      m       ag2a     total  571  33455  570  628.59019
#> 38:   male age_group3     total  424  25159  424  483.33726
#> 39:      m age_group3     total  424  25159  424  483.33726
#> 40:   male age_group4     total  195  12077  197  258.93333
#> 41:      m age_group4     total  195  12077  197  258.93333
#> 42:   male age_group5     total   84   5473   84  149.15476
#> 43:      m age_group5     total   84   5473   84  149.15476
#> 44:   male age_group6     total    7    423   11   71.42857
#> 45:      m age_group6     total    7    423   11   71.42857
#> 46:  Total      Total cnt_males 2296 138448 2296 2356.29965
#> 47:  Total age_group1 cnt_males 1015  61861 1013 1073.94680
#> 48:  Total       ag1a cnt_males 1015  61861 1013 1073.94680
#> 49:  Total age_group2 cnt_males  571  33455  570  628.59019
#> 50:  Total       ag2a cnt_males  571  33455  570  628.59019
#> 51:  Total age_group3 cnt_males  424  25159  424  483.33726
#> 52:  Total age_group4 cnt_males  195  12077  197  258.93333
#> 53:  Total age_group5 cnt_males   84   5473   84  149.15476
#> 54:  Total age_group6 cnt_males    7    423   11   71.42857
#> 55: female      Total cnt_males    0      0    0    0.00000
#> 56:      f      Total cnt_males    0      0    0    0.00000
#> 57: female age_group1 cnt_males    0      0    0    0.00000
#> 58:      f age_group1 cnt_males    0      0    0    0.00000
#> 59: female       ag1a cnt_males    0      0    0    0.00000
#> 60:      f       ag1a cnt_males    0      0    0    0.00000
#> 61: female age_group2 cnt_males    0      0    0    0.00000
#> 62:      f age_group2 cnt_males    0      0    0    0.00000
#> 63: female       ag2a cnt_males    0      0    0    0.00000
#> 64:      f       ag2a cnt_males    0      0    0    0.00000
#> 65: female age_group3 cnt_males    0      0    0    0.00000
#> 66:      f age_group3 cnt_males    0      0    0    0.00000
#> 67: female age_group4 cnt_males    0      0    0    0.00000
#> 68:      f age_group4 cnt_males    0      0    0    0.00000
#> 69: female age_group5 cnt_males    0      0    0    0.00000
#> 70:      f age_group5 cnt_males    0      0    0    0.00000
#> 71: female age_group6 cnt_males    0      0    0    0.00000
#> 72:      f age_group6 cnt_males    0      0    0    0.00000
#> 73:   male      Total cnt_males 2296 138448 2296 2356.29965
#> 74:      m      Total cnt_males 2296 138448 2296 2356.29965
#> 75:   male age_group1 cnt_males 1015  61861 1013 1073.94680
#> 76:      m age_group1 cnt_males 1015  61861 1013 1073.94680
#> 77:   male       ag1a cnt_males 1015  61861 1013 1073.94680
#> 78:      m       ag1a cnt_males 1015  61861 1013 1073.94680
#> 79:   male age_group2 cnt_males  571  33455  570  628.59019
#> 80:      m age_group2 cnt_males  571  33455  570  628.59019
#> 81:   male       ag2a cnt_males  571  33455  570  628.59019
#> 82:      m       ag2a cnt_males  571  33455  570  628.59019
#> 83:   male age_group3 cnt_males  424  25159  424  483.33726
#> 84:      m age_group3 cnt_males  424  25159  424  483.33726
#> 85:   male age_group4 cnt_males  195  12077  197  258.93333
#> 86:      m age_group4 cnt_males  195  12077  197  258.93333
#> 87:   male age_group5 cnt_males   84   5473   84  149.15476
#> 88:      m age_group5 cnt_males   84   5473   84  149.15476
#> 89:   male age_group6 cnt_males    7    423   11   71.42857
#> 90:      m age_group6 cnt_males    7    423   11   71.42857
#>        sex        age     vname  uwc     wc puwc        pwc

# only unweighted and perturbed unweighted counts
tab$freqtab(v = c("total", "cnt_males"), type = "unweighted")

#>        sex        age     vname  uwc puwc
#>  1:  Total      Total     total 4580 4581
#>  2:  Total age_group1     total 1969 1966
#>  3:  Total       ag1a     total 1969 1966
#>  4:  Total age_group2     total 1143 1144
#>  5:  Total       ag2a     total 1143 1144
#>  6:  Total age_group3     total  864  865
#>  7:  Total age_group4     total  423  423
#>  8:  Total age_group5     total  168  168
#>  9:  Total age_group6     total   13   11
#> 10: female      Total     total 2284 2284
#> 11:      f      Total     total 2284 2284
#> 12: female age_group1     total  954  958
#> 13:      f age_group1     total  954  958
#> 14: female       ag1a     total  954  958
#> 15:      f       ag1a     total  954  958
#> 16: female age_group2     total  572  572
#> 17:      f age_group2     total  572  572
#> 18: female       ag2a     total  572  572
#> 19:      f       ag2a     total  572  572
#> 20: female age_group3     total  440  440
#> 21:      f age_group3     total  440  440
#> 22: female age_group4     total  228  228
#> 23:      f age_group4     total  228  228
#> 24: female age_group5     total   84   87
#> 25:      f age_group5     total   84   87
#> 26: female age_group6     total    6    4
#> 27:      f age_group6     total    6    4
#> 28:   male      Total     total 2296 2296
#> 29:      m      Total     total 2296 2296
#> 30:   male age_group1     total 1015 1013
#> 31:      m age_group1     total 1015 1013
#> 32:   male       ag1a     total 1015 1013
#> 33:      m       ag1a     total 1015 1013
#> 34:   male age_group2     total  571  570
#> 35:      m age_group2     total  571  570
#> 36:   male       ag2a     total  571  570
#> 37:      m       ag2a     total  571  570
#> 38:   male age_group3     total  424  424
#> 39:      m age_group3     total  424  424
#> 40:   male age_group4     total  195  197
#> 41:      m age_group4     total  195  197
#> 42:   male age_group5     total   84   84
#> 43:      m age_group5     total   84   84
#> 44:   male age_group6     total    7   11
#> 45:      m age_group6     total    7   11
#> 46:  Total      Total cnt_males 2296 2296
#> 47:  Total age_group1 cnt_males 1015 1013
#> 48:  Total       ag1a cnt_males 1015 1013
#> 49:  Total age_group2 cnt_males  571  570
#> 50:  Total       ag2a cnt_males  571  570
#> 51:  Total age_group3 cnt_males  424  424
#> 52:  Total age_group4 cnt_males  195  197
#> 53:  Total age_group5 cnt_males   84   84
#> 54:  Total age_group6 cnt_males    7   11
#> 55: female      Total cnt_males    0    0
#> 56:      f      Total cnt_males    0    0
#> 57: female age_group1 cnt_males    0    0
#> 58:      f age_group1 cnt_males    0    0
#> 59: female       ag1a cnt_males    0    0
#> 60:      f       ag1a cnt_males    0    0
#> 61: female age_group2 cnt_males    0    0
#> 62:      f age_group2 cnt_males    0    0
#> 63: female       ag2a cnt_males    0    0
#> 64:      f       ag2a cnt_males    0    0
#> 65: female age_group3 cnt_males    0    0
#> 66:      f age_group3 cnt_males    0    0
#> 67: female age_group4 cnt_males    0    0
#> 68:      f age_group4 cnt_males    0    0
#> 69: female age_group5 cnt_males    0    0
#> 70:      f age_group5 cnt_males    0    0
#> 71: female age_group6 cnt_males    0    0
#> 72:      f age_group6 cnt_males    0    0
#> 73:   male      Total cnt_males 2296 2296
#> 74:      m      Total cnt_males 2296 2296
#> 75:   male age_group1 cnt_males 1015 1013
#> 76:      m age_group1 cnt_males 1015 1013
#> 77:   male       ag1a cnt_males 1015 1013
#> 78:      m       ag1a cnt_males 1015 1013
#> 79:   male age_group2 cnt_males  571  570
#> 80:      m age_group2 cnt_males  571  570
#> 81:   male       ag2a cnt_males  571  570
#> 82:      m       ag2a cnt_males  571  570
#> 83:   male age_group3 cnt_males  424  424
#> 84:      m age_group3 cnt_males  424  424
#> 85:   male age_group4 cnt_males  195  197
#> 86:      m age_group4 cnt_males  195  197
#> 87:   male age_group5 cnt_males   84   84
#> 88:      m age_group5 cnt_males   84   84
#> 89:   male age_group6 cnt_males    7   11
#> 90:      m age_group6 cnt_males    7   11
#>        sex        age     vname  uwc puwc

# write to a file "outtab.csv" (.csv is automatically added to the path)

# NOT RUN {
tab$freqtab(v = c("total", "cnt_males"), path = "outtab")
# }

# only weighted and perturbed weighted counts
tab$freqtab(v = c("total", "cnt_males"), type = "weighted")

#>        sex        age     vname     wc        pwc
#>  1:  Total      Total     total 276667 4641.40764
#>  2:  Total age_group1     total 119983 2026.93601
#>  3:  Total       ag1a     total 119983 2026.93601
#>  4:  Total age_group2     total  67768 1203.28959
#>  5:  Total       ag2a     total  67768 1203.28959
#>  6:  Total age_group3     total  51533  924.64468
#>  7:  Total age_group4     total  26067  484.62411
#>  8:  Total age_group5     total  10643  231.35119
#>  9:  Total age_group6     total    673   62.76923
#> 10: female      Total     total 138219 2344.51620
#> 11:      f      Total     total 138219 2344.51620
#> 12: female age_group1     total  58122 1018.92453
#> 13:      f age_group1     total  58122 1018.92453
#> 14: female       ag1a     total  58122 1018.92453
#> 15:      f       ag1a     total  58122 1018.92453
#> 16: female age_group2     total  34313  631.98776
#> 17:      f age_group2     total  34313  631.98776
#> 18: female       ag2a     total  34313  631.98776
#> 19:      f       ag2a     total  34313  631.98776
#> 20: female age_group3     total  26374  499.94091
#> 21:      f age_group3     total  26374  499.94091
#> 22: female age_group4     total  13990  289.35965
#> 23:      f age_group4     total  13990  289.35965
#> 24: female age_group5     total   5170  148.54762
#> 25:      f age_group5     total   5170  148.54762
#> 26: female age_group6     total    250   45.66667
#> 27:      f age_group6     total    250   45.66667
#> 28:   male      Total     total 138448 2356.29965
#> 29:      m      Total     total 138448 2356.29965
#> 30:   male age_group1     total  61861 1073.94680
#> 31:      m age_group1     total  61861 1073.94680
#> 32:   male       ag1a     total  61861 1073.94680
#> 33:      m       ag1a     total  61861 1073.94680
#> 34:   male age_group2     total  33455  628.59019
#> 35:      m age_group2     total  33455  628.59019
#> 36:   male       ag2a     total  33455  628.59019
#> 37:      m       ag2a     total  33455  628.59019
#> 38:   male age_group3     total  25159  483.33726
#> 39:      m age_group3     total  25159  483.33726
#> 40:   male age_group4     total  12077  258.93333
#> 41:      m age_group4     total  12077  258.93333
#> 42:   male age_group5     total   5473  149.15476
#> 43:      m age_group5     total   5473  149.15476
#> 44:   male age_group6     total    423   71.42857
#> 45:      m age_group6     total    423   71.42857
#> 46:  Total      Total cnt_males 138448 2356.29965
#> 47:  Total age_group1 cnt_males  61861 1073.94680
#> 48:  Total       ag1a cnt_males  61861 1073.94680
#> 49:  Total age_group2 cnt_males  33455  628.59019
#> 50:  Total       ag2a cnt_males  33455  628.59019
#> 51:  Total age_group3 cnt_males  25159  483.33726
#> 52:  Total age_group4 cnt_males  12077  258.93333
#> 53:  Total age_group5 cnt_males   5473  149.15476
#> 54:  Total age_group6 cnt_males    423   71.42857
#> 55: female      Total cnt_males      0    0.00000
#> 56:      f      Total cnt_males      0    0.00000
#> 57: female age_group1 cnt_males      0    0.00000
#> 58:      f age_group1 cnt_males      0    0.00000
#> 59: female       ag1a cnt_males      0    0.00000
#> 60:      f       ag1a cnt_males      0    0.00000
#> 61: female age_group2 cnt_males      0    0.00000
#> 62:      f age_group2 cnt_males      0    0.00000
#> 63: female       ag2a cnt_males      0    0.00000
#> 64:      f       ag2a cnt_males      0    0.00000
#> 65: female age_group3 cnt_males      0    0.00000
#> 66:      f age_group3 cnt_males      0    0.00000
#> 67: female age_group4 cnt_males      0    0.00000
#> 68:      f age_group4 cnt_males      0    0.00000
#> 69: female age_group5 cnt_males      0    0.00000
#> 70:      f age_group5 cnt_males      0    0.00000
#> 71: female age_group6 cnt_males      0    0.00000
#> 72:      f age_group6 cnt_males      0    0.00000
#> 73:   male      Total cnt_males 138448 2356.29965
#> 74:      m      Total cnt_males 138448 2356.29965
#> 75:   male age_group1 cnt_males  61861 1073.94680
#> 76:      m age_group1 cnt_males  61861 1073.94680
#> 77:   male       ag1a cnt_males  61861 1073.94680
#> 78:      m       ag1a cnt_males  61861 1073.94680
#> 79:   male age_group2 cnt_males  33455  628.59019
#> 80:      m age_group2 cnt_males  33455  628.59019
#> 81:   male       ag2a cnt_males  33455  628.59019
#> 82:      m       ag2a cnt_males  33455  628.59019
#> 83:   male age_group3 cnt_males  25159  483.33726
#> 84:      m age_group3 cnt_males  25159  483.33726
#> 85:   male age_group4 cnt_males  12077  258.93333
#> 86:      m age_group4 cnt_males  12077  258.93333
#> 87:   male age_group5 cnt_males   5473  149.15476
#> 88:      m age_group5 cnt_males   5473  149.15476
#> 89:   male age_group6 cnt_males    423   71.42857
#> 90:      m age_group6 cnt_males    423   71.42857
#>        sex        age     vname     wc        pwc

# utility measures for a count variable
tab$measures_cnts(v = "total", exclude_zeros = TRUE)

#> $overview
#>    noise cnt        pct
#> 1:    -4   6 0.13333333
#> 2:    -3   2 0.04444444
#> 3:    -2   2 0.04444444
#> 4:    -1   4 0.08888889
#> 5:     0  18 0.40000000
#> 6:     1   4 0.08888889
#> 7:     2   7 0.15555556
#> 8:     3   2 0.04444444
#> 
#> $measures
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 0.600 0.000 0.004
#>  6:   Mean 1.378 0.047 0.080
#>  7: Median 1.000 0.002 0.021
#>  8:    Q60 2.000 0.002 0.031
#>  9:    Q70 2.000 0.002 0.034
#> 10:    Q80 3.000 0.005 0.066
#> 11:    Q90 4.000 0.107 0.238
#> 12:    Q95 4.000 0.333 0.449
#> 13:    Q99 4.000 0.571 0.671
#> 14:    Max 4.000 0.571 0.671
#> 
#> $cumdistr_d1
#>    cat cnt       pct
#> 1:   0  18 0.4000000
#> 2:   1  26 0.5777778
#> 3:   2  35 0.7777778
#> 4:   3  39 0.8666667
#> 5:   4  45 1.0000000
#> 
#> $cumdistr_d2
#>            cat cnt       pct
#> 1:    [0,0.02]  38 0.8444444
#> 2: (0.02,0.05]  40 0.8888889
#> 3:  (0.05,0.1]  40 0.8888889
#> 4:   (0.1,0.2]  41 0.9111111
#> 5:   (0.2,0.3]  41 0.9111111
#> 6:   (0.3,0.4]  43 0.9555556
#> 7:   (0.4,0.5]  43 0.9555556
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $cumdistr_d3
#>            cat cnt       pct
#> 1:    [0,0.02]  22 0.4888889
#> 2: (0.02,0.05]  32 0.7111111
#> 3:  (0.05,0.1]  38 0.8444444
#> 4:   (0.1,0.2]  40 0.8888889
#> 5:   (0.2,0.3]  41 0.9111111
#> 6:   (0.3,0.4]  41 0.9111111
#> 7:   (0.4,0.5]  43 0.9555556
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $false_zero
#> [1] 0
#> 
#> $false_nonzero
#> [1] 0
#> 
#> $exclude_zeros
#> [1] TRUE
#> 

# modifications for perturbed count variables
tab$mod_cnts()

#>        sex        age row_nr pert    ckey       countvar
#>   1: Total      Total     62    1 0.75966          total
#>   2: Total age_group1     58   -3 0.07863          total
#>   3: Total       ag1a     58   -3 0.07863          total
#>   4: Total age_group2     62    1 0.78649          total
#>   5: Total       ag2a     62    1 0.78649          total
#>  ---                                                    
#> 131:     m age_group4    133    0 0.42521 cnt_highincome
#> 132:  male age_group5     70    0 0.36031 cnt_highincome
#> 133:     m age_group5     70    0 0.36031 cnt_highincome
#> 134:  male age_group6     -1    0 0.00000 cnt_highincome
#> 135:     m age_group6     -1    0 0.00000 cnt_highincome

# display a summary about utility measures
tab$summary()

#> ┌──────────────────────────────────────────────┐
#> │Utility measures for perturbed count variables│
#> └──────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#>          countvar Min  Q10  Q20 Q30 Q40   Mean Median Q60 Q70 Q80 Q90 Q95 Q99
#> 1:          total  -3 -2.0 -1.2   0   0  0.311      0   0 0.8   2   4 4.0   4
#> 2:      cnt_males  -2 -2.0 -1.0   0   0  0.000      0   0 0.0   0   2 3.6   4
#> 3: cnt_highincome  -4 -2.2 -1.0   0   0 -0.089      0   0 0.0   0   1 4.0   4
#>    Max
#> 1:   4
#> 2:   4
#> 3:   4
#> 
#> ── Distance-based measures ─────────────────────────────────────────────────────
#> ✔ Variable: 'total'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 0.600 0.000 0.004
#>  6:   Mean 1.378 0.047 0.080
#>  7: Median 1.000 0.002 0.021
#>  8:    Q60 2.000 0.002 0.031
#>  9:    Q70 2.000 0.002 0.034
#> 10:    Q80 3.000 0.005 0.066
#> 11:    Q90 4.000 0.107 0.238
#> 12:    Q95 4.000 0.333 0.449
#> 13:    Q99 4.000 0.571 0.671
#> 14:    Max 4.000 0.571 0.671
#> 
#> ✔ Variable: 'cnt_males'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 1.000 0.002 0.021
#>  6:   Mean 1.333 0.065 0.094
#>  7: Median 1.000 0.002 0.021
#>  8:    Q60 2.000 0.002 0.031
#>  9:    Q70 2.000 0.002 0.031
#> 10:    Q80 2.000 0.009 0.063
#> 11:    Q90 2.800 0.235 0.311
#> 12:    Q95 4.000 0.571 0.671
#> 13:    Q99 4.000 0.571 0.671
#> 14:    Max 4.000 0.571 0.671
#> 
#> ✔ Variable: 'cnt_highincome'
#> 
#>       what  d1    d2    d3
#>  1:    Min 0.0 0.000 0.000
#>  2:    Q10 0.0 0.000 0.000
#>  3:    Q20 0.0 0.000 0.000
#>  4:    Q30 0.0 0.000 0.000
#>  5:    Q40 0.0 0.000 0.000
#>  6:   Mean 1.1 0.029 0.083
#>  7: Median 0.0 0.000 0.000
#>  8:    Q60 1.0 0.008 0.042
#>  9:    Q70 1.0 0.018 0.066
#> 10:    Q80 3.2 0.044 0.213
#> 11:    Q90 4.0 0.098 0.305
#> 12:    Q95 4.0 0.211 0.426
#> 13:    Q99 4.0 0.213 0.437
#> 14:    Max 4.0 0.214 0.437
#>