Create perturbation parameters for count variables

This function allows to generate required perturbation parameters that are used to perturb count variables.

ck_params_cnts(ptab, path = NULL)

Arguments

ptab
an object created with ptable::pt_create_pTable()

path
a scalar character specifying a path to which the parameters created with this functions should be written to (in yaml format)

Value

an object suitable as input to method $params_cnts_set() for the perturbation of counts and frequencies.

See also

Examples

x <- ck_create_testdata()

# create some 0/1 variables that should be perturbed later
x[, cnt_females := ifelse(sex == "male", 0, 1)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              76          25.00000           0
#>    2:   2530      28       1              39          25.00000           1
#>    3:   6920     550       1              29          25.00000           0
#>    4:   7960     870       1              75          25.00000           0
#>    5:   9030      20       2              89          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              51          16.66667           1
#> 4577:   1420     987    1000              87          16.66667           0
#> 4578:   8900     684    1000              23          16.66667           0
#> 4579:   3880     294    1000              52          16.66667           1
#> 4580:   4830     911    1000              75          16.66667           0

x[, cnt_males := ifelse(sex == "male", 1, 0)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              76          25.00000           0
#>    2:   2530      28       1              39          25.00000           1
#>    3:   6920     550       1              29          25.00000           0
#>    4:   7960     870       1              75          25.00000           0
#>    5:   9030      20       2              89          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              51          16.66667           1
#> 4577:   1420     987    1000              87          16.66667           0
#> 4578:   8900     684    1000              23          16.66667           0
#> 4579:   3880     294    1000              52          16.66667           1
#> 4580:   4830     911    1000              75          16.66667           0
#>       cnt_males
#>    1:         1
#>    2:         0
#>    3:         1
#>    4:         1
#>    5:         1
#>   ---          
#> 4576:         0
#> 4577:         1
#> 4578:         1
#> 4579:         0
#> 4580:         1

x[, cnt_highincome := ifelse(income >= 9000, 1, 0)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              76          25.00000           0
#>    2:   2530      28       1              39          25.00000           1
#>    3:   6920     550       1              29          25.00000           0
#>    4:   7960     870       1              75          25.00000           0
#>    5:   9030      20       2              89          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              51          16.66667           1
#> 4577:   1420     987    1000              87          16.66667           0
#> 4578:   8900     684    1000              23          16.66667           0
#> 4579:   3880     294    1000              52          16.66667           1
#> 4580:   4830     911    1000              75          16.66667           0
#>       cnt_males cnt_highincome
#>    1:         1              0
#>    2:         0              0
#>    3:         1              0
#>    4:         1              0
#>    5:         1              1
#>   ---                         
#> 4576:         0              0
#> 4577:         1              0
#> 4578:         1              0
#> 4579:         0              0
#> 4580:         1              0

# a variable with positive and negative contributions
x[, mixed := sample(-10:10, nrow(x), replace = TRUE)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              76          25.00000           0
#>    2:   2530      28       1              39          25.00000           1
#>    3:   6920     550       1              29          25.00000           0
#>    4:   7960     870       1              75          25.00000           0
#>    5:   9030      20       2              89          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              51          16.66667           1
#> 4577:   1420     987    1000              87          16.66667           0
#> 4578:   8900     684    1000              23          16.66667           0
#> 4579:   3880     294    1000              52          16.66667           1
#> 4580:   4830     911    1000              75          16.66667           0
#>       cnt_males cnt_highincome mixed
#>    1:         1              0    -6
#>    2:         0              0    10
#>    3:         1              0     9
#>    4:         1              0     0
#>    5:         1              1     7
#>   ---                               
#> 4576:         0              0     5
#> 4577:         1              0    -4
#> 4578:         1              0    10
#> 4579:         0              0     0
#> 4580:         1              0    10

# create record keys
x$rkey <- ck_generate_rkeys(dat = x)

# define required inputs

# hierarchy with some bogus codes
d_sex <- hier_create(root = "Total", nodes = c("male", "female"))
d_sex <- hier_add(d_sex, root = "female", "f")
d_sex <- hier_add(d_sex, root = "male", "m")

d_age <- hier_create(root = "Total", nodes = paste0("age_group", 1:6))
d_age <- hier_add(d_age, root = "age_group1", "ag1a")
d_age <- hier_add(d_age, root = "age_group2", "ag2a")

# define the cell key object
countvars <- c("cnt_females", "cnt_males", "cnt_highincome")
numvars <- c("expend", "income", "savings", "mixed")
tab <- ck_setup(
  x = x,
  rkey = "rkey",
  dims = list(sex = d_sex, age = d_age),
  w = "sampling_weight",
  countvars = countvars,
  numvars = numvars)

# show some information about this table instance
tab$print() # identical with print(tab)

#> ── Table Information ───────────────────────────────────────────────────────────
#> ✔ 45 cells in 2 dimensions ('sex', 'age')
#> ✔ weights: yes
#> ── Tabulated / Perturbed countvars ─────────────────────────────────────────────
#> ☐ 'total'
#> ☐ 'cnt_females'
#> ☐ 'cnt_males'
#> ☐ 'cnt_highincome'
#> ── Tabulated / Perturbed numvars ───────────────────────────────────────────────
#> ☐ 'expend'
#> ☐ 'income'
#> ☐ 'savings'
#> ☐ 'mixed'

# which variables have been defined?
tab$allvars()

#> $cntvars
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"
#> 
#> $numvars
#> [1] "expend"  "income"  "savings" "mixed"  
#> 

# count-variables
tab$cntvars()

#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"

# continuous variables
tab$numvars()

#> [1] "expend"  "income"  "savings" "mixed"  

# create perturbation parameters for "total" variable and
# write to yaml-file

# create a ptable using functionality
# from the ptable-pkg
para1 <- ptable::pt_create_pParams(
  D = 5, V = 1, table = "cnts")

#> Since type of table is frequency table (argument table is set to 'cnts'), the input parameters 'step' and 'icat' will be ignored.

# create suitable parameter inputs
f_yaml <- tempfile(fileext = ".yaml")
p_cnts1 <- ck_params_cnts(
  ptab = ptable::pt_create_pTable(para1),
  path = f_yaml)

#> yaml configuration '/tmp/RtmpKWoyEE/file76d465eb37f7.yaml' successfully written.

# read parameters from yaml-file and set them for variable `"total"`
p_cnts1 <- ck_read_yaml(path = f_yaml)

tab$params_cnts_set(val = p_cnts1, v = "total")

#> --> setting perturbation parameters for variable 'total'

# create alternative perturbation parameters
para2 <- ptable::pt_create_pParams(
  D = 8, V = 3, js = 2, table = "cnts")

#> Since type of table is frequency table (argument table is set to 'cnts'), the input parameters 'step' and 'icat' will be ignored.

p_cnts2 <- ck_params_cnts(
  ptab = ptable::pt_create_pTable(para1))

# use it for the remaining variables
tab$params_cnts_set(val = p_cnts2, v = countvars)

#> --> setting perturbation parameters for variable 'cnt_females'

#> --> setting perturbation parameters for variable 'cnt_males'

#> --> setting perturbation parameters for variable 'cnt_highincome'

# perturb a variable
tab$perturb(v = "total")

#> Count variable 'total' was perturbed.

# multiple variables can be perturbed as well
tab$perturb(v = c("cnt_males", "cnt_highincome"))

#> Count variable 'cnt_males' was perturbed.

#> Count variable 'cnt_highincome' was perturbed.

# return weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))

#>        sex        age     vname  uwc     wc puwc        pwc
#>  1:  Total      Total     total 4580 272971 4579 272911.399
#>  2:  Total age_group1     total 1969 118179 1970 118239.020
#>  3:  Total       ag1a     total 1969 118179 1970 118239.020
#>  4:  Total age_group2     total 1143  67260 1145  67377.690
#>  5:  Total       ag2a     total 1143  67260 1145  67377.690
#>  6:  Total age_group3     total  864  51608  865  51667.731
#>  7:  Total age_group4     total  423  24845  424  24903.735
#>  8:  Total age_group5     total  168  10250  170  10372.024
#>  9:  Total age_group6     total   13    829   13    829.000
#> 10: female      Total     total 2284 136454 2284 136454.000
#> 11:      f      Total     total 2284 136454 2284 136454.000
#> 12: female age_group1     total  954  57080  953  57020.168
#> 13:      f age_group1     total  954  57080  953  57020.168
#> 14: female       ag1a     total  954  57080  953  57020.168
#> 15:      f       ag1a     total  954  57080  953  57020.168
#> 16: female age_group2     total  572  34218  573  34277.822
#> 17:      f age_group2     total  572  34218  573  34277.822
#> 18: female       ag2a     total  572  34218  573  34277.822
#> 19:      f       ag2a     total  572  34218  573  34277.822
#> 20: female age_group3     total  440  26277  440  26277.000
#> 21:      f age_group3     total  440  26277  440  26277.000
#> 22: female age_group4     total  228  13374  227  13315.342
#> 23:      f age_group4     total  228  13374  227  13315.342
#> 24: female age_group5     total   84   5145   83   5083.750
#> 25:      f age_group5     total   84   5145   83   5083.750
#> 26: female age_group6     total    6    360    7    420.000
#> 27:      f age_group6     total    6    360    7    420.000
#> 28:   male      Total     total 2296 136517 2296 136517.000
#> 29:      m      Total     total 2296 136517 2296 136517.000
#> 30:   male age_group1     total 1015  61099 1016  61159.196
#> 31:      m age_group1     total 1015  61099 1016  61159.196
#> 32:   male       ag1a     total 1015  61099 1016  61159.196
#> 33:      m       ag1a     total 1015  61099 1016  61159.196
#> 34:   male age_group2     total  571  33042  570  32984.133
#> 35:      m age_group2     total  571  33042  570  32984.133
#> 36:   male       ag2a     total  571  33042  570  32984.133
#> 37:      m       ag2a     total  571  33042  570  32984.133
#> 38:   male age_group3     total  424  25331  424  25331.000
#> 39:      m age_group3     total  424  25331  424  25331.000
#> 40:   male age_group4     total  195  11471  196  11529.826
#> 41:      m age_group4     total  195  11471  196  11529.826
#> 42:   male age_group5     total   84   5105   85   5165.774
#> 43:      m age_group5     total   84   5105   85   5165.774
#> 44:   male age_group6     total    7    469    8    536.000
#> 45:      m age_group6     total    7    469    8    536.000
#> 46:  Total      Total cnt_males 2296 136517 2296 136517.000
#> 47:  Total age_group1 cnt_males 1015  61099 1016  61159.196
#> 48:  Total       ag1a cnt_males 1015  61099 1016  61159.196
#> 49:  Total age_group2 cnt_males  571  33042  570  32984.133
#> 50:  Total       ag2a cnt_males  571  33042  570  32984.133
#> 51:  Total age_group3 cnt_males  424  25331  424  25331.000
#> 52:  Total age_group4 cnt_males  195  11471  196  11529.826
#> 53:  Total age_group5 cnt_males   84   5105   85   5165.774
#> 54:  Total age_group6 cnt_males    7    469    8    536.000
#> 55: female      Total cnt_males    0      0    0      0.000
#> 56:      f      Total cnt_males    0      0    0      0.000
#> 57: female age_group1 cnt_males    0      0    0      0.000
#> 58:      f age_group1 cnt_males    0      0    0      0.000
#> 59: female       ag1a cnt_males    0      0    0      0.000
#> 60:      f       ag1a cnt_males    0      0    0      0.000
#> 61: female age_group2 cnt_males    0      0    0      0.000
#> 62:      f age_group2 cnt_males    0      0    0      0.000
#> 63: female       ag2a cnt_males    0      0    0      0.000
#> 64:      f       ag2a cnt_males    0      0    0      0.000
#> 65: female age_group3 cnt_males    0      0    0      0.000
#> 66:      f age_group3 cnt_males    0      0    0      0.000
#> 67: female age_group4 cnt_males    0      0    0      0.000
#> 68:      f age_group4 cnt_males    0      0    0      0.000
#> 69: female age_group5 cnt_males    0      0    0      0.000
#> 70:      f age_group5 cnt_males    0      0    0      0.000
#> 71: female age_group6 cnt_males    0      0    0      0.000
#> 72:      f age_group6 cnt_males    0      0    0      0.000
#> 73:   male      Total cnt_males 2296 136517 2296 136517.000
#> 74:      m      Total cnt_males 2296 136517 2296 136517.000
#> 75:   male age_group1 cnt_males 1015  61099 1016  61159.196
#> 76:      m age_group1 cnt_males 1015  61099 1016  61159.196
#> 77:   male       ag1a cnt_males 1015  61099 1016  61159.196
#> 78:      m       ag1a cnt_males 1015  61099 1016  61159.196
#> 79:   male age_group2 cnt_males  571  33042  570  32984.133
#> 80:      m age_group2 cnt_males  571  33042  570  32984.133
#> 81:   male       ag2a cnt_males  571  33042  570  32984.133
#> 82:      m       ag2a cnt_males  571  33042  570  32984.133
#> 83:   male age_group3 cnt_males  424  25331  424  25331.000
#> 84:      m age_group3 cnt_males  424  25331  424  25331.000
#> 85:   male age_group4 cnt_males  195  11471  196  11529.826
#> 86:      m age_group4 cnt_males  195  11471  196  11529.826
#> 87:   male age_group5 cnt_males   84   5105   85   5165.774
#> 88:      m age_group5 cnt_males   84   5105   85   5165.774
#> 89:   male age_group6 cnt_males    7    469    8    536.000
#> 90:      m age_group6 cnt_males    7    469    8    536.000
#>        sex        age     vname  uwc     wc puwc        pwc

# numerical variables (positive variables using flex-function)
# we also write the config to a yaml file
f_yaml <- tempfile(fileext = ".yaml")

# create a ptable using functionality
# from the ptable-pkg
para <- ptable::pt_create_pParams(
  D = 5,
  V = 1,
  table = "nums",
  step = 0.2,
  icat = c(1, 5),
  type = "all")
ptab <- ptable::pt_create_pTable(para)

p_nums1 <- ck_params_nums(
  ptab = ptab,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.30, 0.03),
    epsilon = c(1, 0.5, 0.2),
    q = 3),
  mu_c = 2,
  same_key = FALSE,
  use_zero_rkeys = FALSE,
  path = f_yaml)

#> yaml configuration '/tmp/RtmpKWoyEE/file76d47475009d.yaml' successfully written.

# we read the parameters from the yaml-file
p_nums1 <- ck_read_yaml(path = f_yaml)

# another set of parameters
para <- ptable::pt_create_pParams(
  D = 10,
  V = 1,
  table = "nums",
  step = 0.5,
  icat = c(1, 5, 10),
  type = "all")
ptab <- ptable::pt_create_pTable(para)

# for variables with positive and negative values
p_nums2 <- ck_params_nums(
  ptab = ptab,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.15, 0.02),
    epsilon = c(1, 0.4, 0.15),
    q = 3),
  mu_c = 2,
  same_key = FALSE)

# simple perturbation parameters (not using the flex-function approach)
p_nums3 <- ck_params_nums(
  ptab = ptab,
  type = "mean",
  mult_params = ck_simpleparams(p = 0.25),
  mu_c = 2,
  same_key = FALSE)

# use `p_nums1` for all variables
tab$params_nums_set(p_nums1, c("savings", "income", "expend"))

#> --> setting perturbation parameters for variable 'savings'

#> --> setting perturbation parameters for variable 'income'

#> --> setting perturbation parameters for variable 'expend'

# use different parameters for variable `mixed`
tab$params_nums_set(p_nums2, v = "mixed")

#> --> setting perturbation parameters for variable 'mixed'

# identify sensitive cells to which extra protection (`mu_c`) is added.
tab$supp_p(v = "income", p = 85)

#> p%-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)

tab$supp_pq(v = "income", p = 85, q = 90)

#> pq-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)

tab$supp_nk(v = "income", n = 2, k = 90)

#> nk-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)

tab$supp_freq(v = "income", n = 14, weighted = FALSE)

#> freq-rule: 5 new sensitive cells (incl. duplicates) found (total: 5)

tab$supp_val(v = "income", n = 10000, weighted = TRUE)

#> val-rule: 0 new sensitive cells (incl. duplicates) found (total: 5)

# perturb variables
tab$perturb(v = c("income", "savings"))

#> Numeric variable 'income' was perturbed.

#> Numeric variable 'savings' was perturbed.

# extract results
tab$numtab("income", mean_before_sum = TRUE)

#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1371638432 1371668916
#>  2:  Total age_group1 income  9810547  586633532  586659958
#>  3:  Total       ag1a income  9810547  586633532  586659958
#>  4:  Total age_group2 income  5692119  336714461  336727911
#>  5:  Total       ag2a income  5692119  336714461  336727911
#>  6:  Total age_group3 income  4406946  264659634  264597269
#>  7:  Total age_group4 income  2133543  127077769  127140462
#>  8:  Total age_group5 income   848151   52585564   52589139
#>  9:  Total age_group6 income    61672    3967472    3892329
#> 10: female      Total income 11690929  700814534  700750141
#> 11:      f      Total income 11690929  700814534  700750141
#> 12: female age_group1 income  4933383  294104115  294034146
#> 13:      f age_group1 income  4933383  294104115  294034146
#> 14: female       ag1a income  4933383  294104115  294034146
#> 15:      f       ag1a income  4933383  294104115  294034146
#> 16: female age_group2 income  2880740  173955500  173937781
#> 17:      f age_group2 income  2880740  173955500  173937781
#> 18: female       ag2a income  2880740  173955500  173937781
#> 19:      f       ag2a income  2880740  173955500  173937781
#> 20: female age_group3 income  2238777  134826570  134861213
#> 21:      f age_group3 income  2238777  134826570  134861213
#> 22: female age_group4 income  1155033   68529559   68508205
#> 23:      f age_group4 income  1155033   68529559   68508205
#> 24: female age_group5 income   455017   27618184   27555155
#> 25:      f age_group5 income   455017   27618184   27555155
#> 26: female age_group6 income    27979    1780606    1744009
#> 27:      f age_group6 income    27979    1780606    1744009
#> 28:   male      Total income 11262049  670823898  670781761
#> 29:      m      Total income 11262049  670823898  670781761
#> 30:   male age_group1 income  4877164  292529417  292538079
#> 31:      m age_group1 income  4877164  292529417  292538079
#> 32:   male       ag1a income  4877164  292529417  292538079
#> 33:      m       ag1a income  4877164  292529417  292538079
#> 34:   male age_group2 income  2811379  162758961  162674689
#> 35:      m age_group2 income  2811379  162758961  162674689
#> 36:   male       ag2a income  2811379  162758961  162674689
#> 37:      m       ag2a income  2811379  162758961  162674689
#> 38:   male age_group3 income  2168169  129833064  129774246
#> 39:      m age_group3 income  2168169  129833064  129774246
#> 40:   male age_group4 income   978510   58548210   58523323
#> 41:      m age_group4 income   978510   58548210   58523323
#> 42:   male age_group5 income   393134   24967380   24990227
#> 43:      m age_group5 income   393134   24967380   24990227
#> 44:   male age_group6 income    33693    2186866    2019461
#> 45:      m age_group6 income    33693    2186866    2019461
#>        sex        age  vname      uws         ws        pws

tab$numtab("income", mean_before_sum = FALSE)

#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1371638432 1371653674
#>  2:  Total age_group1 income  9810547  586633532  586646745
#>  3:  Total       ag1a income  9810547  586633532  586646745
#>  4:  Total age_group2 income  5692119  336714461  336721186
#>  5:  Total       ag2a income  5692119  336714461  336721186
#>  6:  Total age_group3 income  4406946  264659634  264628450
#>  7:  Total age_group4 income  2133543  127077769  127109112
#>  8:  Total age_group5 income   848151   52585564   52587351
#>  9:  Total age_group6 income    61672    3967472    3929721
#> 10: female      Total income 11690929  700814534  700782337
#> 11:      f      Total income 11690929  700814534  700782337
#> 12: female age_group1 income  4933383  294104115  294069129
#> 13:      f age_group1 income  4933383  294104115  294069129
#> 14: female       ag1a income  4933383  294104115  294069129
#> 15:      f       ag1a income  4933383  294104115  294069129
#> 16: female age_group2 income  2880740  173955500  173946640
#> 17:      f age_group2 income  2880740  173955500  173946640
#> 18: female       ag2a income  2880740  173955500  173946640
#> 19:      f       ag2a income  2880740  173955500  173946640
#> 20: female age_group3 income  2238777  134826570  134843890
#> 21:      f age_group3 income  2238777  134826570  134843890
#> 22: female age_group4 income  1155033   68529559   68518881
#> 23:      f age_group4 income  1155033   68529559   68518881
#> 24: female age_group5 income   455017   27618184   27586651
#> 25:      f age_group5 income   455017   27618184   27586651
#> 26: female age_group6 income    27979    1780606    1762212
#> 27:      f age_group6 income    27979    1780606    1762212
#> 28:   male      Total income 11262049  670823898  670802829
#> 29:      m      Total income 11262049  670823898  670802829
#> 30:   male age_group1 income  4877164  292529417  292533748
#> 31:      m age_group1 income  4877164  292529417  292533748
#> 32:   male       ag1a income  4877164  292529417  292533748
#> 33:      m       ag1a income  4877164  292529417  292533748
#> 34:   male age_group2 income  2811379  162758961  162716819
#> 35:      m age_group2 income  2811379  162758961  162716819
#> 36:   male       ag2a income  2811379  162758961  162716819
#> 37:      m       ag2a income  2811379  162758961  162716819
#> 38:   male age_group3 income  2168169  129833064  129803652
#> 39:      m age_group3 income  2168169  129833064  129803652
#> 40:   male age_group4 income   978510   58548210   58535765
#> 41:      m age_group4 income   978510   58548210   58535765
#> 42:   male age_group5 income   393134   24967380   24978801
#> 43:      m age_group5 income   393134   24967380   24978801
#> 44:   male age_group6 income    33693    2186866    2101497
#> 45:      m age_group6 income    33693    2186866    2101497
#>        sex        age  vname      uws         ws        pws

tab$numtab("savings")

#>        sex        age   vname     uws        ws         pws
#>  1:  Total      Total savings 2273532 135507254 135505485.6
#>  2:  Total age_group1 savings  982386  59106062  59108039.9
#>  3:  Total       ag1a savings  982386  59106062  59108039.9
#>  4:  Total age_group2 savings  552336  32351896  32355096.0
#>  5:  Total       ag2a savings  552336  32351896  32355096.0
#>  6:  Total age_group3 savings  437101  26152047  26152679.1
#>  7:  Total age_group4 savings  214661  12602844  12604336.6
#>  8:  Total age_group5 savings   80451   4836239   4836790.0
#>  9:  Total age_group6 savings    6597    458166    456930.1
#> 10: female      Total savings 1113716  66502088  66500872.6
#> 11:      f      Total savings 1113716  66502088  66500872.6
#> 12: female age_group1 savings  464726  27757278  27755633.3
#> 13:      f age_group1 savings  464726  27757278  27755633.3
#> 14: female       ag1a savings  464726  27757278  27755633.3
#> 15:      f       ag1a savings  464726  27757278  27755633.3
#> 16: female age_group2 savings  271413  16242443  16241987.0
#> 17:      f age_group2 savings  271413  16242443  16241987.0
#> 18: female       ag2a savings  271413  16242443  16241987.0
#> 19:      f       ag2a savings  271413  16242443  16241987.0
#> 20: female age_group3 savings  222131  13335509  13336597.7
#> 21:      f age_group3 savings  222131  13335509  13336597.7
#> 22: female age_group4 savings  115241   6686628   6684443.6
#> 23:      f age_group4 savings  115241   6686628   6684443.6
#> 24: female age_group5 savings   37218   2272773   2271513.1
#> 25:      f age_group5 savings   37218   2272773   2271513.1
#> 26: female age_group6 savings    2987    207457    206406.2
#> 27:      f age_group6 savings    2987    207457    206406.2
#> 28:   male      Total savings 1159816  69005166  69005292.4
#> 29:      m      Total savings 1159816  69005166  69005292.4
#> 30:   male age_group1 savings  517660  31348784  31348765.2
#> 31:      m age_group1 savings  517660  31348784  31348765.2
#> 32:   male       ag1a savings  517660  31348784  31348765.2
#> 33:      m       ag1a savings  517660  31348784  31348765.2
#> 34:   male age_group2 savings  280923  16109453  16106864.2
#> 35:      m age_group2 savings  280923  16109453  16106864.2
#> 36:   male       ag2a savings  280923  16109453  16106864.2
#> 37:      m       ag2a savings  280923  16109453  16106864.2
#> 38:   male age_group3 savings  214970  12816538  12817481.0
#> 39:      m age_group3 savings  214970  12816538  12817481.0
#> 40:   male age_group4 savings   99420   5916216   5915933.1
#> 41:      m age_group4 savings   99420   5916216   5915933.1
#> 42:   male age_group5 savings   43233   2563466   2565119.2
#> 43:      m age_group5 savings   43233   2563466   2565119.2
#> 44:   male age_group6 savings    3610    250709    244636.4
#> 45:      m age_group6 savings    3610    250709    244636.4
#>        sex        age   vname     uws        ws         pws

# results can be resetted, too
tab$reset_cntvars(v = "cnt_males")

# we can then set other parameters and perturb again
tab$params_cnts_set(val = p_cnts1, v = "cnt_males")

#> --> setting perturbation parameters for variable 'cnt_males'

tab$perturb(v = "cnt_males")

#> Count variable 'cnt_males' was perturbed.

# write to a file "outtab.csv" (.csv is automatically added to the path)
if (FALSE) {
tab$freqtab(v = c("total", "cnt_males"), path = "outtab.csv")
}

# show results containing weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))

#>        sex        age     vname  uwc     wc puwc        pwc
#>  1:  Total      Total     total 4580 272971 4579 272911.399
#>  2:  Total age_group1     total 1969 118179 1970 118239.020
#>  3:  Total       ag1a     total 1969 118179 1970 118239.020
#>  4:  Total age_group2     total 1143  67260 1145  67377.690
#>  5:  Total       ag2a     total 1143  67260 1145  67377.690
#>  6:  Total age_group3     total  864  51608  865  51667.731
#>  7:  Total age_group4     total  423  24845  424  24903.735
#>  8:  Total age_group5     total  168  10250  170  10372.024
#>  9:  Total age_group6     total   13    829   13    829.000
#> 10: female      Total     total 2284 136454 2284 136454.000
#> 11:      f      Total     total 2284 136454 2284 136454.000
#> 12: female age_group1     total  954  57080  953  57020.168
#> 13:      f age_group1     total  954  57080  953  57020.168
#> 14: female       ag1a     total  954  57080  953  57020.168
#> 15:      f       ag1a     total  954  57080  953  57020.168
#> 16: female age_group2     total  572  34218  573  34277.822
#> 17:      f age_group2     total  572  34218  573  34277.822
#> 18: female       ag2a     total  572  34218  573  34277.822
#> 19:      f       ag2a     total  572  34218  573  34277.822
#> 20: female age_group3     total  440  26277  440  26277.000
#> 21:      f age_group3     total  440  26277  440  26277.000
#> 22: female age_group4     total  228  13374  227  13315.342
#> 23:      f age_group4     total  228  13374  227  13315.342
#> 24: female age_group5     total   84   5145   83   5083.750
#> 25:      f age_group5     total   84   5145   83   5083.750
#> 26: female age_group6     total    6    360    7    420.000
#> 27:      f age_group6     total    6    360    7    420.000
#> 28:   male      Total     total 2296 136517 2296 136517.000
#> 29:      m      Total     total 2296 136517 2296 136517.000
#> 30:   male age_group1     total 1015  61099 1016  61159.196
#> 31:      m age_group1     total 1015  61099 1016  61159.196
#> 32:   male       ag1a     total 1015  61099 1016  61159.196
#> 33:      m       ag1a     total 1015  61099 1016  61159.196
#> 34:   male age_group2     total  571  33042  570  32984.133
#> 35:      m age_group2     total  571  33042  570  32984.133
#> 36:   male       ag2a     total  571  33042  570  32984.133
#> 37:      m       ag2a     total  571  33042  570  32984.133
#> 38:   male age_group3     total  424  25331  424  25331.000
#> 39:      m age_group3     total  424  25331  424  25331.000
#> 40:   male age_group4     total  195  11471  196  11529.826
#> 41:      m age_group4     total  195  11471  196  11529.826
#> 42:   male age_group5     total   84   5105   85   5165.774
#> 43:      m age_group5     total   84   5105   85   5165.774
#> 44:   male age_group6     total    7    469    8    536.000
#> 45:      m age_group6     total    7    469    8    536.000
#> 46:  Total      Total cnt_males 2296 136517 2296 136517.000
#> 47:  Total age_group1 cnt_males 1015  61099 1016  61159.196
#> 48:  Total       ag1a cnt_males 1015  61099 1016  61159.196
#> 49:  Total age_group2 cnt_males  571  33042  570  32984.133
#> 50:  Total       ag2a cnt_males  571  33042  570  32984.133
#> 51:  Total age_group3 cnt_males  424  25331  424  25331.000
#> 52:  Total age_group4 cnt_males  195  11471  196  11529.826
#> 53:  Total age_group5 cnt_males   84   5105   85   5165.774
#> 54:  Total age_group6 cnt_males    7    469    8    536.000
#> 55: female      Total cnt_males    0      0    0      0.000
#> 56:      f      Total cnt_males    0      0    0      0.000
#> 57: female age_group1 cnt_males    0      0    0      0.000
#> 58:      f age_group1 cnt_males    0      0    0      0.000
#> 59: female       ag1a cnt_males    0      0    0      0.000
#> 60:      f       ag1a cnt_males    0      0    0      0.000
#> 61: female age_group2 cnt_males    0      0    0      0.000
#> 62:      f age_group2 cnt_males    0      0    0      0.000
#> 63: female       ag2a cnt_males    0      0    0      0.000
#> 64:      f       ag2a cnt_males    0      0    0      0.000
#> 65: female age_group3 cnt_males    0      0    0      0.000
#> 66:      f age_group3 cnt_males    0      0    0      0.000
#> 67: female age_group4 cnt_males    0      0    0      0.000
#> 68:      f age_group4 cnt_males    0      0    0      0.000
#> 69: female age_group5 cnt_males    0      0    0      0.000
#> 70:      f age_group5 cnt_males    0      0    0      0.000
#> 71: female age_group6 cnt_males    0      0    0      0.000
#> 72:      f age_group6 cnt_males    0      0    0      0.000
#> 73:   male      Total cnt_males 2296 136517 2296 136517.000
#> 74:      m      Total cnt_males 2296 136517 2296 136517.000
#> 75:   male age_group1 cnt_males 1015  61099 1016  61159.196
#> 76:      m age_group1 cnt_males 1015  61099 1016  61159.196
#> 77:   male       ag1a cnt_males 1015  61099 1016  61159.196
#> 78:      m       ag1a cnt_males 1015  61099 1016  61159.196
#> 79:   male age_group2 cnt_males  571  33042  570  32984.133
#> 80:      m age_group2 cnt_males  571  33042  570  32984.133
#> 81:   male       ag2a cnt_males  571  33042  570  32984.133
#> 82:      m       ag2a cnt_males  571  33042  570  32984.133
#> 83:   male age_group3 cnt_males  424  25331  424  25331.000
#> 84:      m age_group3 cnt_males  424  25331  424  25331.000
#> 85:   male age_group4 cnt_males  195  11471  196  11529.826
#> 86:      m age_group4 cnt_males  195  11471  196  11529.826
#> 87:   male age_group5 cnt_males   84   5105   85   5165.774
#> 88:      m age_group5 cnt_males   84   5105   85   5165.774
#> 89:   male age_group6 cnt_males    7    469    8    536.000
#> 90:      m age_group6 cnt_males    7    469    8    536.000
#>        sex        age     vname  uwc     wc puwc        pwc

# utility measures for a count variable
tab$measures_cnts(v = "total", exclude_zeros = TRUE)

#> $overview
#>    noise cnt        pct
#> 1:    -2   3 0.06666667
#> 2:    -1  20 0.44444444
#> 3:     0   9 0.20000000
#> 4:     1  13 0.28888889
#> 
#> $measures
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.800 0.000 0.006
#>  4:    Q30 1.000 0.001 0.016
#>  5:    Q40 1.000 0.001 0.016
#>  6:   Mean 0.867 0.016 0.036
#>  7: Median 1.000 0.002 0.021
#>  8:    Q60 1.000 0.002 0.021
#>  9:    Q70 1.000 0.002 0.030
#> 10:    Q80 1.000 0.006 0.039
#> 11:    Q90 1.000 0.012 0.068
#> 12:    Q95 1.800 0.143 0.183
#> 13:    Q99 2.000 0.167 0.196
#> 14:    Max 2.000 0.167 0.196
#> 
#> $cumdistr_d1
#>    cat cnt       pct
#> 1:   0   9 0.2000000
#> 2:   1  42 0.9333333
#> 3:   2  45 1.0000000
#> 
#> $cumdistr_d2
#>            cat cnt       pct
#> 1:    [0,0.02]  41 0.9111111
#> 2: (0.02,0.05]  41 0.9111111
#> 3:  (0.05,0.1]  41 0.9111111
#> 4:   (0.1,0.2]  45 1.0000000
#> 5:   (0.2,0.3]  45 1.0000000
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $cumdistr_d3
#>            cat cnt       pct
#> 1:    [0,0.02]  21 0.4666667
#> 2: (0.02,0.05]  36 0.8000000
#> 3:  (0.05,0.1]  41 0.9111111
#> 4:   (0.1,0.2]  45 1.0000000
#> 5:   (0.2,0.3]  45 1.0000000
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $false_zero
#> [1] 0
#> 
#> $false_nonzero
#> [1] 0
#> 
#> $exclude_zeros
#> [1] TRUE
#> 

# modifications for perturbed count variables
tab$mod_cnts()

#>        sex        age row_nr pert      ckey  countvar
#>   1: Total      Total     40   -1 0.1213574     total
#>   2: Total age_group1     42    1 0.8850270     total
#>   3: Total       ag1a     42    1 0.8850270     total
#>   4: Total age_group2     43    2 0.9847088     total
#>   5: Total       ag2a     43    2 0.9847088     total
#>  ---                                                 
#> 131:     m age_group4     42    1 0.7337113 cnt_males
#> 132:  male age_group5     42    1 0.8585011 cnt_males
#> 133:     m age_group5     42    1 0.8585011 cnt_males
#> 134:  male age_group6     42    1 0.7007744 cnt_males
#> 135:     m age_group6     42    1 0.7007744 cnt_males

# display a summary about utility measures
tab$summary()

#> ┌──────────────────────────────────────────────┐
#> │Utility measures for perturbed count variables│
#> └──────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#>          countvar Min  Q10 Q20 Q30 Q40  Mean Median Q60 Q70 Q80 Q90 Q95 Q99 Max
#> 1:          total  -1 -1.0  -1   0   0 0.289      1   1   1 1.0   1 1.8   2   2
#> 2: cnt_highincome  -1 -0.6   0   0   0 0.667      1   1   1 1.2   2 2.0   2   2
#> 3:      cnt_males  -1 -1.0   0   0   0 0.200      0   0   1 1.0   1 1.0   1   1
#> 
#> ── Distance-based measures ─────────────────────────────────────────────────────
#> ✔ Variable: 'total'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.800 0.000 0.006
#>  4:    Q30 1.000 0.001 0.016
#>  5:    Q40 1.000 0.001 0.016
#>  6:   Mean 0.867 0.016 0.036
#>  7: Median 1.000 0.002 0.021
#>  8:    Q60 1.000 0.002 0.021
#>  9:    Q70 1.000 0.002 0.030
#> 10:    Q80 1.000 0.006 0.039
#> 11:    Q90 1.000 0.012 0.068
#> 12:    Q95 1.800 0.143 0.183
#> 13:    Q99 2.000 0.167 0.196
#> 14:    Max 2.000 0.167 0.196
#> 
#> ✔ Variable: 'cnt_males'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.200 0.000 0.003
#>  4:    Q30 1.000 0.001 0.016
#>  5:    Q40 1.000 0.001 0.016
#>  6:   Mean 0.778 0.018 0.038
#>  7: Median 1.000 0.002 0.021
#>  8:    Q60 1.000 0.002 0.021
#>  9:    Q70 1.000 0.005 0.036
#> 10:    Q80 1.000 0.011 0.051
#> 11:    Q90 1.000 0.064 0.106
#> 12:    Q95 1.000 0.143 0.183
#> 13:    Q99 1.000 0.143 0.183
#> 14:    Max 1.000 0.143 0.183
#> 
#> ✔ Variable: 'cnt_highincome'
#> 
#>       what d1    d2    d3
#>  1:    Min  0 0.000 0.000
#>  2:    Q10  0 0.000 0.000
#>  3:    Q20  0 0.000 0.000
#>  4:    Q30  1 0.005 0.034
#>  5:    Q40  1 0.011 0.056
#>  6:   Mean  1 0.026 0.068
#>  7: Median  1 0.016 0.064
#>  8:    Q60  1 0.018 0.071
#>  9:    Q70  1 0.020 0.099
#> 10:    Q80  2 0.023 0.105
#> 11:    Q90  2 0.079 0.136
#> 12:    Q95  2 0.143 0.183
#> 13:    Q99  2 0.143 0.196
#> 14:    Max  2 0.143 0.196
#> 
#> ┌──────────────────────────────────────────────────┐
#> │Utility measures for perturbed numerical variables│
#> └──────────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#>      vname        Min        Q10        Q20        Q30        Q40       Mean
#> 1:  income -85368.594 -42141.611 -34986.481 -31532.563 -24406.199 -16267.152
#> 2: savings  -6072.642  -2588.847  -1669.441  -1259.904  -1116.629   -562.573
#>        Median       Q60      Q70      Q80       Q90       Q95       Q99
#> 1: -12444.666 -8859.635 3822.360 6725.090 13212.724 16904.574 25172.881
#> 2:   -456.048   -18.775  126.429  972.129  1653.157  1977.851  3200.032
#>          Max
#> 1: 31342.793
#> 2:  3200.032