Create perturbation parameters for continuous variables

This function allows to define perturbation parameters used to perturb cells in magnitude tables.

ck_params_nums(P = "top_contr", top_k = NULL, D, l, same_key = TRUE,
  use_zero_rkeys = FALSE)

Arguments

P
a character value defining the way to identify the magnifier, e.g which contributions/values in a cell should be the used in the perturbation procedure. Possible choices are:

top_contr: the k largest contributions are used. In this case, it is also required to specify argument top_k

mean: the (weighted) cellmean is used as starting point

range: the difference between largest and smallest contribution is used.

sum: the (weighted) cellvalue itself is used as starting point

top_k
it is ignored if P is different from top_contr. Otherwise,

D
maximum perturbation value

l
stepwidth parameter for computation of perturbation tables a integerish number >= 1 specifying the number of top contributions whose values should be perturbed.

same_key
(logical) should original cell key (TRUE) used for for finding perturbation values of the largest contributor to a cell or should a perturbation of the cellkey itself (FALSE) take place.

use_zero_rkeys
(logical) skalar defining if record keys of units not contributing to a specific numeric variables should be used (TRUE) or ignored (FALSE) when computing cell keys.

Value

an object suitable as input to ck_setup() for the perturbation of magnitude tables

Examples

ck_params_nums(D = 3, l = 0.5, P = "top_contr", top_k = 3)

#> $params
#> $params$P
#> [1] "top_contr"
#> 
#> $params$top_k
#> [1] 3
#> 
#> $params$stab
#>     i   j           p    kum_p_u    kum_p_o diff
#>  1: 0 0.0 1.000000000 0.00000000 1.00000000  0.0
#>  2: 1 0.0 0.012336985 0.00000000 0.01233699 -1.0
#>  3: 1 0.5 0.010059579 0.01233699 0.02239656 -0.5
#>  4: 1 1.0 0.006458873 0.02239656 0.02885544  0.0
#>  5: 1 1.5 0.246789435 0.02885544 0.27564487  0.5
#>  6: 1 2.0 0.149631830 0.27564487 0.42527670  1.0
#>  7: 1 2.5 0.011465385 0.42527670 0.43674209  1.5
#>  8: 1 3.0 0.290107711 0.43674209 0.72684980  2.0
#>  9: 1 3.5 0.048947298 0.72684980 0.77579710  2.5
#> 10: 1 4.0 0.038093121 0.77579710 1.00000000  3.0
#> 11: 3 0.0 0.061281256 0.00000000 0.06128126 -3.0
#> 12: 3 0.5 0.014317899 0.06128126 0.07559916 -2.5
#> 13: 3 1.0 0.067825180 0.07559916 0.14342434 -2.0
#> 14: 3 1.5 0.176131292 0.14342434 0.31955563 -1.5
#> 15: 3 2.0 0.002714366 0.31955563 0.32226999 -1.0
#> 16: 3 2.5 0.047496737 0.32226999 0.36976673 -0.5
#> 17: 3 3.0 0.026446960 0.36976673 0.39621369  0.0
#> 18: 3 3.5 0.015602201 0.39621369 0.41181589  0.5
#> 19: 3 4.0 0.030445301 0.41181589 0.44226119  1.0
#> 20: 3 4.5 0.062980696 0.44226119 0.50524189  1.5
#> 21: 3 5.0 0.136690256 0.50524189 0.64193214  2.0
#> 22: 3 5.5 0.065322216 0.64193214 0.70725436  2.5
#> 23: 3 6.0 0.044637735 0.70725436 1.00000000  3.0
#>     i   j           p    kum_p_u    kum_p_o diff
#> 
#> $params$same_key
#> [1] TRUE
#> 
#> $params$use_zero_rkeys
#> [1] FALSE
#> 
#> 
#> $type
#> [1] "nums"
#> 
#> attr(,"class")
#> [1] "ck_params"

ck_params_nums(D = 10, l = .5, P = "mean")

#> $params
#> $params$P
#> [1] "mean"
#> 
#> $params$top_k
#> [1] 1
#> 
#> $params$stab
#>      i    j             p     kum_p_u     kum_p_o  diff
#>  1:  0  0.0 1.00000000000 0.000000000 1.000000000   0.0
#>  2:  1  0.0 0.00946512143 0.000000000 0.009465121  -1.0
#>  3:  1  0.5 0.02434970415 0.009465121 0.033814826  -0.5
#>  4:  1  1.0 0.08212723816 0.033814826 0.115942064   0.0
#>  5:  1  1.5 0.06181406206 0.115942064 0.177756126   0.5
#>  6:  1  2.0 0.00790845393 0.177756126 0.185664580   1.0
#>  7:  1  2.5 0.03050883440 0.185664580 0.216173414   1.5
#>  8:  1  3.0 0.02711114264 0.216173414 0.243284557   2.0
#>  9:  1  3.5 0.06806718092 0.243284557 0.311351738   2.5
#> 10:  1  4.0 0.03150079376 0.311351738 0.342852531   3.0
#> 11:  1  4.5 0.00149163464 0.342852531 0.344344166   3.5
#> 12:  1  5.0 0.04455465823 0.344344166 0.388898824   4.0
#> 13:  1  5.5 0.02548224013 0.388898824 0.414381064   4.5
#> 14:  1  6.0 0.01803789916 0.414381064 0.432418964   5.0
#> 15:  1  6.5 0.03450042033 0.432418964 0.466919384   5.5
#> 16:  1  7.0 0.00986653008 0.466919384 0.476785914   6.0
#> 17:  1  7.5 0.07185763563 0.476785914 0.548643550   6.5
#> 18:  1  8.0 0.02151482110 0.548643550 0.570158371   7.0
#> 19:  1  8.5 0.04534539673 0.570158371 0.615503768   7.5
#> 20:  1  9.0 0.01495505753 0.615503768 0.630458825   8.0
#> 21:  1  9.5 0.13905808283 0.630458825 0.769516908   8.5
#> 22:  1 10.0 0.00966093969 0.769516908 0.779177848   9.0
#> 23:  1 10.5 0.08307129983 0.779177848 0.862249147   9.5
#> 24:  1 11.0 0.01538467919 0.862249147 1.000000000  10.0
#> 25: 10  0.0 0.01200617524 0.000000000 0.012006175 -10.0
#> 26: 10  0.5 0.07023551152 0.012006175 0.082241687  -9.5
#> 27: 10  1.0 0.03044322319 0.082241687 0.112684910  -9.0
#> 28: 10  1.5 0.02809083462 0.112684910 0.140775745  -8.5
#> 29: 10  2.0 0.02254869300 0.140775745 0.163324438  -8.0
#> 30: 10  2.5 0.02263938403 0.163324438 0.185963822  -7.5
#> 31: 10  3.0 0.00744617777 0.185963822 0.193409999  -7.0
#> 32: 10  3.5 0.00531985587 0.193409999 0.198729855  -6.5
#> 33: 10  4.0 0.06330197817 0.198729855 0.262031833  -6.0
#> 34: 10  4.5 0.03596905922 0.262031833 0.298000893  -5.5
#> 35: 10  5.0 0.00353252213 0.298000893 0.301533415  -5.0
#> 36: 10  5.5 0.02143245633 0.301533415 0.322965871  -4.5
#> 37: 10  6.0 0.05715760775 0.322965871 0.380123479  -4.0
#> 38: 10  6.5 0.05277850898 0.380123479 0.432901988  -3.5
#> 39: 10  7.0 0.01630745316 0.432901988 0.449209441  -3.0
#> 40: 10  7.5 0.02163640666 0.449209441 0.470845848  -2.5
#> 41: 10  8.0 0.00006356835 0.470845848 0.470909416  -2.0
#> 42: 10  8.5 0.00172258844 0.470909416 0.472632004  -1.5
#> 43: 10  9.0 0.01569885225 0.472632004 0.488330857  -1.0
#> 44: 10  9.5 0.00785621628 0.488330857 0.496187073  -0.5
#> 45: 10 10.0 0.00073237601 0.496187073 0.496919449   0.0
#> 46: 10 10.5 0.01080902596 0.496919449 0.507728475   0.5
#> 47: 10 11.0 0.02607845026 0.507728475 0.533806925   1.0
#> 48: 10 11.5 0.02021673718 0.533806925 0.554023662   1.5
#> 49: 10 12.0 0.00295717246 0.554023662 0.556980835   2.0
#> 50: 10 12.5 0.05118112732 0.556980835 0.608161962   2.5
#> 51: 10 13.0 0.04934103368 0.608161962 0.657502996   3.0
#> 52: 10 13.5 0.02211043751 0.657502996 0.679613433   3.5
#> 53: 10 14.0 0.01836216496 0.679613433 0.697975598   4.0
#> 54: 10 14.5 0.01325686532 0.697975598 0.711232464   4.5
#> 55: 10 15.0 0.03025941434 0.711232464 0.741491878   5.0
#> 56: 10 15.5 0.00006768084 0.741491878 0.741559559   5.5
#> 57: 10 16.0 0.00718053733 0.741559559 0.748740096   6.0
#> 58: 10 16.5 0.06878424506 0.748740096 0.817524341   6.5
#> 59: 10 17.0 0.00733272452 0.817524341 0.824857066   7.0
#> 60: 10 17.5 0.03990906943 0.824857066 0.864766135   7.5
#> 61: 10 18.0 0.03496617870 0.864766135 0.899732314   8.0
#> 62: 10 18.5 0.01047131838 0.899732314 0.910203632   8.5
#> 63: 10 19.0 0.00014137221 0.910203632 0.910345004   9.0
#> 64: 10 19.5 0.00307888468 0.910345004 0.913423889   9.5
#> 65: 10 20.0 0.00826106546 0.913423889 1.000000000  10.0
#>      i    j             p     kum_p_u     kum_p_o  diff
#> 
#> $params$same_key
#> [1] TRUE
#> 
#> $params$use_zero_rkeys
#> [1] FALSE
#> 
#> 
#> $type
#> [1] "nums"
#> 
#> attr(,"class")
#> [1] "ck_params"