Read perturbation parameters from yaml-files

ck_read_yaml() allows to create perturbation parameter inputs from yaml-files that were previously created using ck_params_cnts() or ck_params_nums().

ck_read_yaml(path)

Arguments

path

a path to a yaml-input file

Value

an object object suitable as input to method $params_nums_set() for the perturbation of continous variables in case path was created using ck_params_nums() or an object suitable as input for $params_cnts_set() for the perturbation of counts and frequencies if the input file was generated using ck_params_cnts().

Examples

# \donttest{
x <- ck_create_testdata()

# create some 0/1 variables that should be perturbed later
x[, cnt_females := ifelse(sex == "male", 0, 1)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              23          25.00000           0
#>    2:   2530      28       1              28          25.00000           1
#>    3:   6920     550       1             100          25.00000           0
#>    4:   7960     870       1              75          25.00000           0
#>    5:   9030      20       2              39          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              55          16.66667           1
#> 4577:   1420     987    1000              97          16.66667           0
#> 4578:   8900     684    1000              49          16.66667           0
#> 4579:   3880     294    1000              63          16.66667           1
#> 4580:   4830     911    1000              23          16.66667           0
x[, cnt_males := ifelse(sex == "male", 1, 0)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              23          25.00000           0
#>    2:   2530      28       1              28          25.00000           1
#>    3:   6920     550       1             100          25.00000           0
#>    4:   7960     870       1              75          25.00000           0
#>    5:   9030      20       2              39          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              55          16.66667           1
#> 4577:   1420     987    1000              97          16.66667           0
#> 4578:   8900     684    1000              49          16.66667           0
#> 4579:   3880     294    1000              63          16.66667           1
#> 4580:   4830     911    1000              23          16.66667           0
#>       cnt_males
#>    1:         1
#>    2:         0
#>    3:         1
#>    4:         1
#>    5:         1
#>   ---          
#> 4576:         0
#> 4577:         1
#> 4578:         1
#> 4579:         0
#> 4580:         1
x[, cnt_highincome := ifelse(income >= 9000, 1, 0)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              23          25.00000           0
#>    2:   2530      28       1              28          25.00000           1
#>    3:   6920     550       1             100          25.00000           0
#>    4:   7960     870       1              75          25.00000           0
#>    5:   9030      20       2              39          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              55          16.66667           1
#> 4577:   1420     987    1000              97          16.66667           0
#> 4578:   8900     684    1000              49          16.66667           0
#> 4579:   3880     294    1000              63          16.66667           1
#> 4580:   4830     911    1000              23          16.66667           0
#>       cnt_males cnt_highincome
#>    1:         1              0
#>    2:         0              0
#>    3:         1              0
#>    4:         1              0
#>    5:         1              1
#>   ---                         
#> 4576:         0              0
#> 4577:         1              0
#> 4578:         1              0
#> 4579:         0              0
#> 4580:         1              0
# a variable with positive and negative contributions
x[, mixed := sample(-10:10, nrow(x), replace = TRUE)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              23          25.00000           0
#>    2:   2530      28       1              28          25.00000           1
#>    3:   6920     550       1             100          25.00000           0
#>    4:   7960     870       1              75          25.00000           0
#>    5:   9030      20       2              39          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              55          16.66667           1
#> 4577:   1420     987    1000              97          16.66667           0
#> 4578:   8900     684    1000              49          16.66667           0
#> 4579:   3880     294    1000              63          16.66667           1
#> 4580:   4830     911    1000              23          16.66667           0
#>       cnt_males cnt_highincome mixed
#>    1:         1              0     9
#>    2:         0              0    -6
#>    3:         1              0     2
#>    4:         1              0    -8
#>    5:         1              1     3
#>   ---                               
#> 4576:         0              0     2
#> 4577:         1              0     4
#> 4578:         1              0    -9
#> 4579:         0              0    10
#> 4580:         1              0     8

# create record keys
x$rkey <- ck_generate_rkeys(dat = x)

# define required inputs

# hierarchy with some bogus codes
d_sex <- hier_create(root = "Total", nodes = c("male", "female"))
d_sex <- hier_add(d_sex, root = "female", "f")
d_sex <- hier_add(d_sex, root = "male", "m")

d_age <- hier_create(root = "Total", nodes = paste0("age_group", 1:6))
d_age <- hier_add(d_age, root = "age_group1", "ag1a")
d_age <- hier_add(d_age, root = "age_group2", "ag2a")

# define the cell key object
countvars <- c("cnt_females", "cnt_males", "cnt_highincome")
numvars <- c("expend", "income", "savings", "mixed")
tab <- ck_setup(
  x = x,
  rkey = "rkey",
  dims = list(sex = d_sex, age = d_age),
  w = "sampling_weight",
  countvars = countvars,
  numvars = numvars)

# show some information about this table instance
tab$print() # identical with print(tab)
#> ── Table Information ───────────────────────────────────────────────────────────
#> ✔ 45 cells in 2 dimensions ('sex', 'age')
#> ✔ weights: yes
#> ── Tabulated / Perturbed countvars ─────────────────────────────────────────────
#> ☐ 'total'
#> ☐ 'cnt_females'
#> ☐ 'cnt_males'
#> ☐ 'cnt_highincome'
#> ── Tabulated / Perturbed numvars ───────────────────────────────────────────────
#> ☐ 'expend'
#> ☐ 'income'
#> ☐ 'savings'
#> ☐ 'mixed'

# information about the hierarchies
tab$hierarchy_info()
#> $sex
#>      code level is_leaf parent
#> 1:  Total     1   FALSE  Total
#> 2:   male     2   FALSE  Total
#> 3:      m     3    TRUE   male
#> 4: female     2   FALSE  Total
#> 5:      f     3    TRUE female
#> 
#> $age
#>          code level is_leaf     parent
#> 1:      Total     1   FALSE      Total
#> 2: age_group1     2   FALSE      Total
#> 3:       ag1a     3    TRUE age_group1
#> 4: age_group2     2   FALSE      Total
#> 5:       ag2a     3    TRUE age_group2
#> 6: age_group3     2    TRUE      Total
#> 7: age_group4     2    TRUE      Total
#> 8: age_group5     2    TRUE      Total
#> 9: age_group6     2    TRUE      Total
#> 

# which variables have been defined?
tab$allvars()
#> $cntvars
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"
#> 
#> $numvars
#> [1] "expend"  "income"  "savings" "mixed"  
#> 

# count variables
tab$cntvars()
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"

# continuous variables
tab$numvars()
#> [1] "expend"  "income"  "savings" "mixed"  

# create perturbation parameters for "total" variable and
# write to yaml-file

# create a ptable using functionality from the ptable-pkg
f_yaml <- tempfile(fileext = ".yaml")
p_cnts1 <- ck_params_cnts(
  ptab = ptable::pt_ex_cnts(),
  path = f_yaml)
#> yaml configuration '/tmp/Rtmp2sERaX/file1a9061ec9eb2.yaml' successfully written.

# read parameters from yaml-file and set them for variable `"total"`
p_cnts1 <- ck_read_yaml(path = f_yaml)

tab$params_cnts_set(val = p_cnts1, v = "total")
#> --> setting perturbation parameters for variable 'total'

# create alternative perturbation parameters by specifying parameters
para2 <- ptable::create_cnt_ptable(
  D = 8, V = 3, js = 2, create = FALSE)

p_cnts2 <- ck_params_cnts(ptab = para2)

# use these ptable it for the remaining variables
tab$params_cnts_set(val = p_cnts2, v = countvars)
#> --> setting perturbation parameters for variable 'cnt_females'
#> --> setting perturbation parameters for variable 'cnt_males'
#> --> setting perturbation parameters for variable 'cnt_highincome'

# perturb a variable
tab$perturb(v = "total")
#> Count variable 'total' was perturbed.

# multiple variables can be perturbed as well
tab$perturb(v = c("cnt_males", "cnt_highincome"))
#> Count variable 'cnt_males' was perturbed.
#> Count variable 'cnt_highincome' was perturbed.

# return weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))
#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 272748 4581 272807.5520
#>  2:  Total age_group1     total 1969 117507 1968 117447.3215
#>  3:  Total       ag1a     total 1969 117507 1968 117447.3215
#>  4:  Total age_group2     total 1143  66906 1143  66906.0000
#>  5:  Total       ag2a     total 1143  66906 1143  66906.0000
#>  6:  Total age_group3     total  864  51939  864  51939.0000
#>  7:  Total age_group4     total  423  25348  424  25407.9243
#>  8:  Total age_group5     total  168  10177  169  10237.5774
#>  9:  Total age_group6     total   13    871   14    938.0000
#> 10:   male      Total     total 2296 136255 2296 136255.0000
#> 11:      m      Total     total 2296 136255 2296 136255.0000
#> 12:   male age_group1     total 1015  60141 1015  60141.0000
#> 13:      m age_group1     total 1015  60141 1015  60141.0000
#> 14:   male       ag1a     total 1015  60141 1015  60141.0000
#> 15:      m       ag1a     total 1015  60141 1015  60141.0000
#> 16:   male age_group2     total  571  33580  572  33638.8091
#> 17:      m age_group2     total  571  33580  572  33638.8091
#> 18:   male       ag2a     total  571  33580  572  33638.8091
#> 19:      m       ag2a     total  571  33580  572  33638.8091
#> 20:   male age_group3     total  424  25774  425  25834.7877
#> 21:      m age_group3     total  424  25774  425  25834.7877
#> 22:   male age_group4     total  195  11232  195  11232.0000
#> 23:      m age_group4     total  195  11232  195  11232.0000
#> 24:   male age_group5     total   84   5033   84   5033.0000
#> 25:      m age_group5     total   84   5033   84   5033.0000
#> 26:   male age_group6     total    7    495    6    424.2857
#> 27:      m age_group6     total    7    495    6    424.2857
#> 28: female      Total     total 2284 136493 2284 136493.0000
#> 29:      f      Total     total 2284 136493 2284 136493.0000
#> 30: female age_group1     total  954  57366  954  57366.0000
#> 31:      f age_group1     total  954  57366  954  57366.0000
#> 32: female       ag1a     total  954  57366  954  57366.0000
#> 33:      f       ag1a     total  954  57366  954  57366.0000
#> 34: female age_group2     total  572  33326  573  33384.2622
#> 35:      f age_group2     total  572  33326  573  33384.2622
#> 36: female       ag2a     total  572  33326  573  33384.2622
#> 37:      f       ag2a     total  572  33326  573  33384.2622
#> 38: female age_group3     total  440  26165  441  26224.4659
#> 39:      f age_group3     total  440  26165  441  26224.4659
#> 40: female age_group4     total  228  14116  228  14116.0000
#> 41:      f age_group4     total  228  14116  228  14116.0000
#> 42: female age_group5     total   84   5144   83   5082.7619
#> 43:      f age_group5     total   84   5144   83   5082.7619
#> 44: female age_group6     total    6    376    7    438.6667
#> 45:      f age_group6     total    6    376    7    438.6667
#> 46:  Total      Total cnt_males 2296 136255 2295 136195.6555
#> 47:  Total age_group1 cnt_males 1015  60141 1015  60141.0000
#> 48:  Total       ag1a cnt_males 1015  60141 1015  60141.0000
#> 49:  Total age_group2 cnt_males  571  33580  572  33638.8091
#> 50:  Total       ag2a cnt_males  571  33580  572  33638.8091
#> 51:  Total age_group3 cnt_males  424  25774  425  25834.7877
#> 52:  Total age_group4 cnt_males  195  11232  195  11232.0000
#> 53:  Total age_group5 cnt_males   84   5033   85   5092.9167
#> 54:  Total age_group6 cnt_males    7    495    5    353.5714
#> 55:   male      Total cnt_males 2296 136255 2295 136195.6555
#> 56:      m      Total cnt_males 2296 136255 2295 136195.6555
#> 57:   male age_group1 cnt_males 1015  60141 1015  60141.0000
#> 58:      m age_group1 cnt_males 1015  60141 1015  60141.0000
#> 59:   male       ag1a cnt_males 1015  60141 1015  60141.0000
#> 60:      m       ag1a cnt_males 1015  60141 1015  60141.0000
#> 61:   male age_group2 cnt_males  571  33580  572  33638.8091
#> 62:      m age_group2 cnt_males  571  33580  572  33638.8091
#> 63:   male       ag2a cnt_males  571  33580  572  33638.8091
#> 64:      m       ag2a cnt_males  571  33580  572  33638.8091
#> 65:   male age_group3 cnt_males  424  25774  425  25834.7877
#> 66:      m age_group3 cnt_males  424  25774  425  25834.7877
#> 67:   male age_group4 cnt_males  195  11232  195  11232.0000
#> 68:      m age_group4 cnt_males  195  11232  195  11232.0000
#> 69:   male age_group5 cnt_males   84   5033   85   5092.9167
#> 70:      m age_group5 cnt_males   84   5033   85   5092.9167
#> 71:   male age_group6 cnt_males    7    495    5    353.5714
#> 72:      m age_group6 cnt_males    7    495    5    353.5714
#> 73: female      Total cnt_males    0      0    0      0.0000
#> 74:      f      Total cnt_males    0      0    0      0.0000
#> 75: female age_group1 cnt_males    0      0    0      0.0000
#> 76:      f age_group1 cnt_males    0      0    0      0.0000
#> 77: female       ag1a cnt_males    0      0    0      0.0000
#> 78:      f       ag1a cnt_males    0      0    0      0.0000
#> 79: female age_group2 cnt_males    0      0    0      0.0000
#> 80:      f age_group2 cnt_males    0      0    0      0.0000
#> 81: female       ag2a cnt_males    0      0    0      0.0000
#> 82:      f       ag2a cnt_males    0      0    0      0.0000
#> 83: female age_group3 cnt_males    0      0    0      0.0000
#> 84:      f age_group3 cnt_males    0      0    0      0.0000
#> 85: female age_group4 cnt_males    0      0    0      0.0000
#> 86:      f age_group4 cnt_males    0      0    0      0.0000
#> 87: female age_group5 cnt_males    0      0    0      0.0000
#> 88:      f age_group5 cnt_males    0      0    0      0.0000
#> 89: female age_group6 cnt_males    0      0    0      0.0000
#> 90:      f age_group6 cnt_males    0      0    0      0.0000
#>        sex        age     vname  uwc     wc puwc         pwc

# numerical variables (positive variables using flex-function)
# we also write the config to a yaml file
f_yaml <- tempfile(fileext = ".yaml")

# create a ptable using functionality from the ptable-pkg
# a single ptable for all cells
ptab1 <- ptable::pt_ex_nums(parity = TRUE, separation = FALSE)

# a single ptab for all cells except for very small ones
ptab2 <- ptable::pt_ex_nums(parity = TRUE, separation = TRUE)

# different ptables for cells with even/odd number of contributors
# and very small cells
ptab3 <- ptable::pt_ex_nums(parity = FALSE, separation = TRUE)

p_nums1 <- ck_params_nums(
  ptab = ptab1,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.30, 0.03),
    epsilon = c(1, 0.5, 0.2),
    q = 3),
  mu_c = 2,
  same_key = FALSE,
  use_zero_rkeys = FALSE,
  path = f_yaml)
#> yaml configuration '/tmp/Rtmp2sERaX/file1a906cc1e577.yaml' successfully written.

# we read the parameters from the yaml-file
p_nums1 <- ck_read_yaml(path = f_yaml)

# for variables with positive and negative values
p_nums2 <- ck_params_nums(
  ptab = ptab2,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.15, 0.02),
    epsilon = c(1, 0.4, 0.15),
    q = 3),
  mu_c = 2,
  same_key = FALSE)

# simple perturbation parameters (not using the flex-function approach)
p_nums3 <- ck_params_nums(
  ptab = ptab3,
  type = "mean",
  mult_params = ck_simpleparams(p = 0.25),
  mu_c = 2,
  same_key = FALSE)

# use `p_nums1` for all variables
tab$params_nums_set(p_nums1, c("savings", "income", "expend"))
#> --> setting perturbation parameters for variable 'savings'
#> --> setting perturbation parameters for variable 'income'
#> --> setting perturbation parameters for variable 'expend'

# use different parameters for variable `mixed`
tab$params_nums_set(p_nums2, v = "mixed")
#> --> setting perturbation parameters for variable 'mixed'

# identify sensitive cells to which extra protection (`mu_c`) is added.
tab$supp_p(v = "income", p = 85)
#> computing contributing indices | rawdata <--> table; this might take a while
#> p%-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_pq(v = "income", p = 85, q = 90)
#> computing contributing indices | rawdata <--> table; this might take a while
#> pq-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_nk(v = "income", n = 2, k = 90)
#> computing contributing indices | rawdata <--> table; this might take a while
#> nk-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_freq(v = "income", n = 14, weighted = FALSE)
#> freq-rule: 5 new sensitive cells (incl. duplicates) found (total: 5)
tab$supp_val(v = "income", n = 10000, weighted = TRUE)
#> val-rule: 0 new sensitive cells (incl. duplicates) found (total: 5)
tab$supp_cells(
  v = "income",
  inp = data.frame(
    sex = c("female", "female"),
    "age" = c("age_group1", "age_group3")
  )
)
#> cell-rule: 2 new sensitive cells (incl. duplicates) found (total: 7)

# perturb variables
tab$perturb(v = c("income", "savings"))
#> Numeric variable 'income' was perturbed.
#> Numeric variable 'savings' was perturbed.

# extract results
tab$numtab("income", mean_before_sum = TRUE)
#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1362142589 1362101808
#>  2:  Total age_group1 income  9810547  583430699  583446705
#>  3:  Total       ag1a income  9810547  583430699  583446705
#>  4:  Total age_group2 income  5692119  334828596  334912036
#>  5:  Total       ag2a income  5692119  334828596  334912036
#>  6:  Total age_group3 income  4406946  262217500  262366939
#>  7:  Total age_group4 income  2133543  126244387  126240156
#>  8:  Total age_group5 income   848151   50900107   50802995
#>  9:  Total age_group6 income    61672    4521300    4403391
#> 10:   male      Total income 11262049  663792331  663807949
#> 11:      m      Total income 11262049  663792331  663807949
#> 12:   male age_group1 income  4877164  285321518  285247115
#> 13:      m age_group1 income  4877164  285321518  285247115
#> 14:   male       ag1a income  4877164  285321518  285247115
#> 15:      m       ag1a income  4877164  285321518  285247115
#> 16:   male age_group2 income  2811379  166083105  166062429
#> 17:      m age_group2 income  2811379  166083105  166062429
#> 18:   male       ag2a income  2811379  166083105  166062429
#> 19:      m       ag2a income  2811379  166083105  166062429
#> 20:   male age_group3 income  2168169  130587382  130606069
#> 21:      m age_group3 income  2168169  130587382  130606069
#> 22:   male age_group4 income   978510   55664025   55531528
#> 23:      m age_group4 income   978510   55664025   55531528
#> 24:   male age_group5 income   393134   23556263   23597647
#> 25:      m age_group5 income   393134   23556263   23597647
#> 26:   male age_group6 income    33693    2580038    2668855
#> 27:      m age_group6 income    33693    2580038    2668855
#> 28: female      Total income 11690929  698350258  698383872
#> 29:      f      Total income 11690929  698350258  698383872
#> 30: female age_group1 income  4933383  298109181  298076494
#> 31:      f age_group1 income  4933383  298109181  298076494
#> 32: female       ag1a income  4933383  298109181  298076494
#> 33:      f       ag1a income  4933383  298109181  298076494
#> 34: female age_group2 income  2880740  168745491  168818168
#> 35:      f age_group2 income  2880740  168745491  168818168
#> 36: female       ag2a income  2880740  168745491  168818168
#> 37:      f       ag2a income  2880740  168745491  168818168
#> 38: female age_group3 income  2238777  131630118  131623799
#> 39:      f age_group3 income  2238777  131630118  131623799
#> 40: female age_group4 income  1155033   70580362   70645210
#> 41:      f age_group4 income  1155033   70580362   70645210
#> 42: female age_group5 income   455017   27343844   27337272
#> 43:      f age_group5 income   455017   27343844   27337272
#> 44: female age_group6 income    27979    1941262    1882182
#> 45:      f age_group6 income    27979    1941262    1882182
#>        sex        age  vname      uws         ws        pws
tab$numtab("income", mean_before_sum = FALSE)
#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1362142589 1362122199
#>  2:  Total age_group1 income  9810547  583430699  583438702
#>  3:  Total       ag1a income  9810547  583430699  583438702
#>  4:  Total age_group2 income  5692119  334828596  334870313
#>  5:  Total       ag2a income  5692119  334828596  334870313
#>  6:  Total age_group3 income  4406946  262217500  262292209
#>  7:  Total age_group4 income  2133543  126244387  126242272
#>  8:  Total age_group5 income   848151   50900107   50851528
#>  9:  Total age_group6 income    61672    4521300    4461956
#> 10:   male      Total income 11262049  663792331  663800140
#> 11:      m      Total income 11262049  663792331  663800140
#> 12:   male age_group1 income  4877164  285321518  285284314
#> 13:      m age_group1 income  4877164  285321518  285284314
#> 14:   male       ag1a income  4877164  285321518  285284314
#> 15:      m       ag1a income  4877164  285321518  285284314
#> 16:   male age_group2 income  2811379  166083105  166072767
#> 17:      m age_group2 income  2811379  166083105  166072767
#> 18:   male       ag2a income  2811379  166083105  166072767
#> 19:      m       ag2a income  2811379  166083105  166072767
#> 20:   male age_group3 income  2168169  130587382  130596725
#> 21:      m age_group3 income  2168169  130587382  130596725
#> 22:   male age_group4 income   978510   55664025   55597737
#> 23:      m age_group4 income   978510   55664025   55597737
#> 24:   male age_group5 income   393134   23556263   23576946
#> 25:      m age_group5 income   393134   23556263   23576946
#> 26:   male age_group6 income    33693    2580038    2624071
#> 27:      m age_group6 income    33693    2580038    2624071
#> 28: female      Total income 11690929  698350258  698367065
#> 29:      f      Total income 11690929  698350258  698367065
#> 30: female age_group1 income  4933383  298109181  298092837
#> 31:      f age_group1 income  4933383  298109181  298092837
#> 32: female       ag1a income  4933383  298109181  298092837
#> 33:      f       ag1a income  4933383  298109181  298092837
#> 34: female age_group2 income  2880740  168745491  168781825
#> 35:      f age_group2 income  2880740  168745491  168781825
#> 36: female       ag2a income  2880740  168745491  168781825
#> 37:      f       ag2a income  2880740  168745491  168781825
#> 38: female age_group3 income  2238777  131630118  131626959
#> 39:      f age_group3 income  2238777  131630118  131626959
#> 40: female age_group4 income  1155033   70580362   70612779
#> 41:      f age_group4 income  1155033   70580362   70612779
#> 42: female age_group5 income   455017   27343844   27340558
#> 43:      f age_group5 income   455017   27343844   27340558
#> 44: female age_group6 income    27979    1941262    1911494
#> 45:      f age_group6 income    27979    1941262    1911494
#>        sex        age  vname      uws         ws        pws
tab$numtab("savings")
#>        sex        age   vname     uws        ws         pws
#>  1:  Total      Total savings 2273532 135479776 135475132.2
#>  2:  Total age_group1 savings  982386  58609147  58602512.3
#>  3:  Total       ag1a savings  982386  58609147  58602512.3
#>  4:  Total age_group2 savings  552336  32074756  32080506.3
#>  5:  Total       ag2a savings  552336  32074756  32080506.3
#>  6:  Total age_group3 savings  437101  26552247  26549040.7
#>  7:  Total age_group4 savings  214661  12857391  12857320.9
#>  8:  Total age_group5 savings   80451   4954748   4949212.6
#>  9:  Total age_group6 savings    6597    431487    427999.1
#> 10:   male      Total savings 1159816  68673179  68675836.6
#> 11:      m      Total savings 1159816  68673179  68675836.6
#> 12:   male age_group1 savings  517660  30538860  30531506.9
#> 13:      m age_group1 savings  517660  30538860  30531506.9
#> 14:   male       ag1a savings  517660  30538860  30531506.9
#> 15:      m       ag1a savings  517660  30538860  30531506.9
#> 16:   male age_group2 savings  280923  16478407  16476410.2
#> 17:      m age_group2 savings  280923  16478407  16476410.2
#> 18:   male       ag2a savings  280923  16478407  16476410.2
#> 19:      m       ag2a savings  280923  16478407  16476410.2
#> 20:   male age_group3 savings  214970  13010098  13009634.2
#> 21:      m age_group3 savings  214970  13010098  13009634.2
#> 22:   male age_group4 savings   99420   5814193   5808446.3
#> 23:      m age_group4 savings   99420   5814193   5808446.3
#> 24:   male age_group5 savings   43233   2587304   2589669.0
#> 25:      m age_group5 savings   43233   2587304   2589669.0
#> 26:   male age_group6 savings    3610    244317    249476.8
#> 27:      m age_group6 savings    3610    244317    249476.8
#> 28: female      Total savings 1113716  66806597  66813847.4
#> 29:      f      Total savings 1113716  66806597  66813847.4
#> 30: female age_group1 savings  464726  28070287  28075841.1
#> 31:      f age_group1 savings  464726  28070287  28075841.1
#> 32: female       ag1a savings  464726  28070287  28075841.1
#> 33:      f       ag1a savings  464726  28070287  28075841.1
#> 34: female age_group2 savings  271413  15596349  15601663.3
#> 35:      f age_group2 savings  271413  15596349  15601663.3
#> 36: female       ag2a savings  271413  15596349  15601663.3
#> 37:      f       ag2a savings  271413  15596349  15601663.3
#> 38: female age_group3 savings  222131  13542149  13541307.7
#> 39:      f age_group3 savings  222131  13542149  13541307.7
#> 40: female age_group4 savings  115241   7043198   7043852.5
#> 41:      f age_group4 savings  115241   7043198   7043852.5
#> 42: female age_group5 savings   37218   2367444   2366950.9
#> 43:      f age_group5 savings   37218   2367444   2366950.9
#> 44: female age_group6 savings    2987    187170    185513.5
#> 45:      f age_group6 savings    2987    187170    185513.5
#>        sex        age   vname     uws        ws         pws

# results can be resetted, too
tab$reset_cntvars(v = "cnt_males")

# we can then set other parameters and perturb again
tab$params_cnts_set(val = p_cnts1, v = "cnt_males")
#> --> setting perturbation parameters for variable 'cnt_males'

tab$perturb(v = "cnt_males")
#> Count variable 'cnt_males' was perturbed.

# write results to a .csv file
tab$freqtab(
  v = c("total", "cnt_males"),
  path = file.path(tempdir(), "outtab.csv")
)
#> File '/tmp/Rtmp2sERaX/outtab.csv' successfully written to disk.
#> NULL

# show results containing weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))
#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 272748 4581 272807.5520
#>  2:  Total age_group1     total 1969 117507 1968 117447.3215
#>  3:  Total       ag1a     total 1969 117507 1968 117447.3215
#>  4:  Total age_group2     total 1143  66906 1143  66906.0000
#>  5:  Total       ag2a     total 1143  66906 1143  66906.0000
#>  6:  Total age_group3     total  864  51939  864  51939.0000
#>  7:  Total age_group4     total  423  25348  424  25407.9243
#>  8:  Total age_group5     total  168  10177  169  10237.5774
#>  9:  Total age_group6     total   13    871   14    938.0000
#> 10:   male      Total     total 2296 136255 2296 136255.0000
#> 11:      m      Total     total 2296 136255 2296 136255.0000
#> 12:   male age_group1     total 1015  60141 1015  60141.0000
#> 13:      m age_group1     total 1015  60141 1015  60141.0000
#> 14:   male       ag1a     total 1015  60141 1015  60141.0000
#> 15:      m       ag1a     total 1015  60141 1015  60141.0000
#> 16:   male age_group2     total  571  33580  572  33638.8091
#> 17:      m age_group2     total  571  33580  572  33638.8091
#> 18:   male       ag2a     total  571  33580  572  33638.8091
#> 19:      m       ag2a     total  571  33580  572  33638.8091
#> 20:   male age_group3     total  424  25774  425  25834.7877
#> 21:      m age_group3     total  424  25774  425  25834.7877
#> 22:   male age_group4     total  195  11232  195  11232.0000
#> 23:      m age_group4     total  195  11232  195  11232.0000
#> 24:   male age_group5     total   84   5033   84   5033.0000
#> 25:      m age_group5     total   84   5033   84   5033.0000
#> 26:   male age_group6     total    7    495    6    424.2857
#> 27:      m age_group6     total    7    495    6    424.2857
#> 28: female      Total     total 2284 136493 2284 136493.0000
#> 29:      f      Total     total 2284 136493 2284 136493.0000
#> 30: female age_group1     total  954  57366  954  57366.0000
#> 31:      f age_group1     total  954  57366  954  57366.0000
#> 32: female       ag1a     total  954  57366  954  57366.0000
#> 33:      f       ag1a     total  954  57366  954  57366.0000
#> 34: female age_group2     total  572  33326  573  33384.2622
#> 35:      f age_group2     total  572  33326  573  33384.2622
#> 36: female       ag2a     total  572  33326  573  33384.2622
#> 37:      f       ag2a     total  572  33326  573  33384.2622
#> 38: female age_group3     total  440  26165  441  26224.4659
#> 39:      f age_group3     total  440  26165  441  26224.4659
#> 40: female age_group4     total  228  14116  228  14116.0000
#> 41:      f age_group4     total  228  14116  228  14116.0000
#> 42: female age_group5     total   84   5144   83   5082.7619
#> 43:      f age_group5     total   84   5144   83   5082.7619
#> 44: female age_group6     total    6    376    7    438.6667
#> 45:      f age_group6     total    6    376    7    438.6667
#> 46:  Total      Total cnt_males 2296 136255 2296 136255.0000
#> 47:  Total age_group1 cnt_males 1015  60141 1015  60141.0000
#> 48:  Total       ag1a cnt_males 1015  60141 1015  60141.0000
#> 49:  Total age_group2 cnt_males  571  33580  572  33638.8091
#> 50:  Total       ag2a cnt_males  571  33580  572  33638.8091
#> 51:  Total age_group3 cnt_males  424  25774  425  25834.7877
#> 52:  Total age_group4 cnt_males  195  11232  195  11232.0000
#> 53:  Total age_group5 cnt_males   84   5033   84   5033.0000
#> 54:  Total age_group6 cnt_males    7    495    6    424.2857
#> 55:   male      Total cnt_males 2296 136255 2296 136255.0000
#> 56:      m      Total cnt_males 2296 136255 2296 136255.0000
#> 57:   male age_group1 cnt_males 1015  60141 1015  60141.0000
#> 58:      m age_group1 cnt_males 1015  60141 1015  60141.0000
#> 59:   male       ag1a cnt_males 1015  60141 1015  60141.0000
#> 60:      m       ag1a cnt_males 1015  60141 1015  60141.0000
#> 61:   male age_group2 cnt_males  571  33580  572  33638.8091
#> 62:      m age_group2 cnt_males  571  33580  572  33638.8091
#> 63:   male       ag2a cnt_males  571  33580  572  33638.8091
#> 64:      m       ag2a cnt_males  571  33580  572  33638.8091
#> 65:   male age_group3 cnt_males  424  25774  425  25834.7877
#> 66:      m age_group3 cnt_males  424  25774  425  25834.7877
#> 67:   male age_group4 cnt_males  195  11232  195  11232.0000
#> 68:      m age_group4 cnt_males  195  11232  195  11232.0000
#> 69:   male age_group5 cnt_males   84   5033   84   5033.0000
#> 70:      m age_group5 cnt_males   84   5033   84   5033.0000
#> 71:   male age_group6 cnt_males    7    495    6    424.2857
#> 72:      m age_group6 cnt_males    7    495    6    424.2857
#> 73: female      Total cnt_males    0      0    0      0.0000
#> 74:      f      Total cnt_males    0      0    0      0.0000
#> 75: female age_group1 cnt_males    0      0    0      0.0000
#> 76:      f age_group1 cnt_males    0      0    0      0.0000
#> 77: female       ag1a cnt_males    0      0    0      0.0000
#> 78:      f       ag1a cnt_males    0      0    0      0.0000
#> 79: female age_group2 cnt_males    0      0    0      0.0000
#> 80:      f age_group2 cnt_males    0      0    0      0.0000
#> 81: female       ag2a cnt_males    0      0    0      0.0000
#> 82:      f       ag2a cnt_males    0      0    0      0.0000
#> 83: female age_group3 cnt_males    0      0    0      0.0000
#> 84:      f age_group3 cnt_males    0      0    0      0.0000
#> 85: female age_group4 cnt_males    0      0    0      0.0000
#> 86:      f age_group4 cnt_males    0      0    0      0.0000
#> 87: female age_group5 cnt_males    0      0    0      0.0000
#> 88:      f age_group5 cnt_males    0      0    0      0.0000
#> 89: female age_group6 cnt_males    0      0    0      0.0000
#> 90:      f age_group6 cnt_males    0      0    0      0.0000
#>        sex        age     vname  uwc     wc puwc         pwc

# utility measures for a count variable
tab$measures_cnts(v = "total", exclude_zeros = TRUE)
#> $overview
#>    noise cnt       pct
#> 1:    -1  18 0.4000000
#> 2:     0  21 0.4666667
#> 3:     1   6 0.1333333
#> 
#> $measures
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 0.000 0.000 0.000
#>  6:   Mean 0.533 0.017 0.031
#>  7: Median 1.000 0.001 0.011
#>  8:    Q60 1.000 0.002 0.021
#>  9:    Q70 1.000 0.002 0.021
#> 10:    Q80 1.000 0.002 0.024
#> 11:    Q90 1.000 0.051 0.104
#> 12:    Q95 1.000 0.143 0.196
#> 13:    Q99 1.000 0.167 0.196
#> 14:    Max 1.000 0.167 0.196
#> 
#> $cumdistr_d1
#>    cat cnt       pct
#> 1:   0  21 0.4666667
#> 2:   1  45 1.0000000
#> 
#> $cumdistr_d2
#>            cat cnt       pct
#> 1:    [0,0.02]  40 0.8888889
#> 2: (0.02,0.05]  40 0.8888889
#> 3:  (0.05,0.1]  41 0.9111111
#> 4:   (0.1,0.2]  45 1.0000000
#> 5:   (0.2,0.3]  45 1.0000000
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $cumdistr_d3
#>            cat cnt       pct
#> 1:    [0,0.02]  24 0.5333333
#> 2: (0.02,0.05]  38 0.8444444
#> 3:  (0.05,0.1]  40 0.8888889
#> 4:   (0.1,0.2]  45 1.0000000
#> 5:   (0.2,0.3]  45 1.0000000
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $false_zero
#> [1] 0
#> 
#> $false_nonzero
#> [1] 0
#> 
#> $exclude_zeros
#> [1] TRUE
#> 

# modifications for perturbed count variables
tab$mod_cnts()
#>         sex        age row_nr pert      ckey  countvar
#>   1:  Total      Total     16    1 0.7754114     total
#>   2:  Total age_group1     14   -1 0.2547845     total
#>   3:  Total       ag1a     14   -1 0.2547845     total
#>   4:  Total age_group2     15    0 0.5862558     total
#>   5:  Total       ag2a     15    0 0.5862558     total
#>  ---                                                  
#> 131:      f age_group4     -1    0 0.0000000 cnt_males
#> 132: female age_group5     -1    0 0.0000000 cnt_males
#> 133:      f age_group5     -1    0 0.0000000 cnt_males
#> 134: female age_group6     -1    0 0.0000000 cnt_males
#> 135:      f age_group6     -1    0 0.0000000 cnt_males

# display a summary about utility measures
tab$summary()
#> ┌──────────────────────────────────────────────┐
#> │Utility measures for perturbed count variables│
#> └──────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#>          countvar Min Q10 Q20 Q30 Q40  Mean Median Q60 Q70 Q80 Q90 Q95 Q99 Max
#> 1:          total  -1  -1   0   0   0 0.267      0 0.4   1 1.0   1   1   1   1
#> 2: cnt_highincome  -2  -2  -1   0   0 0.422      0 1.0   1 2.0   2   3   3   3
#> 3:      cnt_males  -1   0   0   0   0 0.133      0 0.0   0 0.2   1   1   1   1
#> 
#> ── Distance-based measures ─────────────────────────────────────────────────────
#> ✔ Variable: 'total'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 0.000 0.000 0.000
#>  6:   Mean 0.533 0.017 0.031
#>  7: Median 1.000 0.001 0.011
#>  8:    Q60 1.000 0.002 0.021
#>  9:    Q70 1.000 0.002 0.021
#> 10:    Q80 1.000 0.002 0.024
#> 11:    Q90 1.000 0.051 0.104
#> 12:    Q95 1.000 0.143 0.196
#> 13:    Q99 1.000 0.167 0.196
#> 14:    Max 1.000 0.167 0.196
#> 
#> ✔ Variable: 'cnt_males'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 0.000 0.000 0.000
#>  6:   Mean 0.444 0.017 0.029
#>  7: Median 0.000 0.000 0.000
#>  8:    Q60 1.000 0.002 0.021
#>  9:    Q70 1.000 0.002 0.021
#> 10:    Q80 1.000 0.002 0.024
#> 11:    Q90 1.000 0.059 0.093
#> 12:    Q95 1.000 0.143 0.196
#> 13:    Q99 1.000 0.143 0.196
#> 14:    Max 1.000 0.143 0.196
#> 
#> ✔ Variable: 'cnt_highincome'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 1.000 0.005 0.036
#>  4:    Q30 1.000 0.010 0.049
#>  5:    Q40 1.000 0.015 0.062
#>  6:   Mean 1.375 0.051 0.117
#>  7: Median 1.000 0.018 0.066
#>  8:    Q60 2.000 0.024 0.120
#>  9:    Q70 2.000 0.052 0.159
#> 10:    Q80 2.000 0.105 0.230
#> 11:    Q90 2.100 0.143 0.250
#> 12:    Q95 3.000 0.150 0.266
#> 13:    Q99 3.000 0.286 0.410
#> 14:    Max 3.000 0.286 0.410
#> 
#> ┌──────────────────────────────────────────────────┐
#> │Utility measures for perturbed numerical variables│
#> └──────────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#>      vname        Min       Q10        Q20        Q30        Q40     Mean
#> 1:  expend        Inf        NA         NA         NA         NA      NaN
#> 2:  income -66288.058 -37204.07 -29768.277 -16343.903 -10338.387 -206.899
#> 3: savings  -7353.075  -6634.70  -4822.105  -1996.767  -1167.345  114.082
#> 4:   mixed        Inf        NA         NA         NA         NA      NaN
#>       Median      Q60       Q70       Q80       Q90       Q95       Q99
#> 1:        NA       NA        NA        NA        NA        NA        NA
#> 2: -3159.364 8002.797 16806.604 33200.295 39564.094 43569.622 61211.233
#> 3:  -463.839 1338.677  4659.403  5314.281  5554.116  5750.268  7250.368
#> 4:        NA       NA        NA        NA        NA        NA        NA
#>          Max
#> 1:      -Inf
#> 2: 74708.642
#> 3:  7250.368
#> 4:      -Inf
# }