Read perturbation parameters from yaml-files

ck_read_yaml() allows to create perturbation parameter inputs from yaml-files that were previously created using ck_params_cnts() or ck_params_nums().

ck_read_yaml(path)

Arguments

path
a path to a yaml-input file

Value

an object object suitable as input to method $params_nums_set() for the perturbation of continous variables in case path was created using ck_params_nums() or an object suitable as input for $params_cnts_set() for the perturbation of counts and frequencies if the input file was generated using ck_params_cnts().

Examples

x <- ck_create_testdata()

# create some 0/1 variables that should be perturbed later
x[, cnt_females := ifelse(sex == "male", 0, 1)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              38          25.00000           0
#>    2:   2530      28       1              28          25.00000           1
#>    3:   6920     550       1              59          25.00000           0
#>    4:   7960     870       1              95          25.00000           0
#>    5:   9030      20       2              56          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              70          16.66667           1
#> 4577:   1420     987    1000              97          16.66667           0
#> 4578:   8900     684    1000              66          16.66667           0
#> 4579:   3880     294    1000              67          16.66667           1
#> 4580:   4830     911    1000              85          16.66667           0

x[, cnt_males := ifelse(sex == "male", 1, 0)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              38          25.00000           0
#>    2:   2530      28       1              28          25.00000           1
#>    3:   6920     550       1              59          25.00000           0
#>    4:   7960     870       1              95          25.00000           0
#>    5:   9030      20       2              56          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              70          16.66667           1
#> 4577:   1420     987    1000              97          16.66667           0
#> 4578:   8900     684    1000              66          16.66667           0
#> 4579:   3880     294    1000              67          16.66667           1
#> 4580:   4830     911    1000              85          16.66667           0
#>       cnt_males
#>    1:         1
#>    2:         0
#>    3:         1
#>    4:         1
#>    5:         1
#>   ---          
#> 4576:         0
#> 4577:         1
#> 4578:         1
#> 4579:         0
#> 4580:         1

x[, cnt_highincome := ifelse(income >= 9000, 1, 0)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              38          25.00000           0
#>    2:   2530      28       1              28          25.00000           1
#>    3:   6920     550       1              59          25.00000           0
#>    4:   7960     870       1              95          25.00000           0
#>    5:   9030      20       2              56          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              70          16.66667           1
#> 4577:   1420     987    1000              97          16.66667           0
#> 4578:   8900     684    1000              66          16.66667           0
#> 4579:   3880     294    1000              67          16.66667           1
#> 4580:   4830     911    1000              85          16.66667           0
#>       cnt_males cnt_highincome
#>    1:         1              0
#>    2:         0              0
#>    3:         1              0
#>    4:         1              0
#>    5:         1              1
#>   ---                         
#> 4576:         0              0
#> 4577:         1              0
#> 4578:         1              0
#> 4579:         0              0
#> 4580:         1              0

# a variable with positive and negative contributions
x[, mixed := sample(-10:10, nrow(x), replace = TRUE)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              38          25.00000           0
#>    2:   2530      28       1              28          25.00000           1
#>    3:   6920     550       1              59          25.00000           0
#>    4:   7960     870       1              95          25.00000           0
#>    5:   9030      20       2              56          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              70          16.66667           1
#> 4577:   1420     987    1000              97          16.66667           0
#> 4578:   8900     684    1000              66          16.66667           0
#> 4579:   3880     294    1000              67          16.66667           1
#> 4580:   4830     911    1000              85          16.66667           0
#>       cnt_males cnt_highincome mixed
#>    1:         1              0    -1
#>    2:         0              0    -1
#>    3:         1              0    -1
#>    4:         1              0     4
#>    5:         1              1     9
#>   ---                               
#> 4576:         0              0    -9
#> 4577:         1              0     1
#> 4578:         1              0     8
#> 4579:         0              0     4
#> 4580:         1              0     5

# create record keys
x$rkey <- ck_generate_rkeys(dat = x)

# define required inputs

# hierarchy with some bogus codes
d_sex <- hier_create(root = "Total", nodes = c("male", "female"))
d_sex <- hier_add(d_sex, root = "female", "f")
d_sex <- hier_add(d_sex, root = "male", "m")

d_age <- hier_create(root = "Total", nodes = paste0("age_group", 1:6))
d_age <- hier_add(d_age, root = "age_group1", "ag1a")
d_age <- hier_add(d_age, root = "age_group2", "ag2a")

# define the cell key object
countvars <- c("cnt_females", "cnt_males", "cnt_highincome")
numvars <- c("expend", "income", "savings", "mixed")
tab <- ck_setup(
  x = x,
  rkey = "rkey",
  dims = list(sex = d_sex, age = d_age),
  w = "sampling_weight",
  countvars = countvars,
  numvars = numvars)

# show some information about this table instance
tab$print() # identical with print(tab)

#> ── Table Information ───────────────────────────────────────────────────────────
#> ✔ 45 cells in 2 dimensions ('sex', 'age')
#> ✔ weights: yes
#> ── Tabulated / Perturbed countvars ─────────────────────────────────────────────
#> ☐ 'total'
#> ☐ 'cnt_females'
#> ☐ 'cnt_males'
#> ☐ 'cnt_highincome'
#> ── Tabulated / Perturbed numvars ───────────────────────────────────────────────
#> ☐ 'expend'
#> ☐ 'income'
#> ☐ 'savings'
#> ☐ 'mixed'

# information about the hierarchies
tab$hierarchy_info()

#> $sex
#>      code level is_leaf parent
#> 1:  Total     1   FALSE   <NA>
#> 2:   male     2   FALSE  Total
#> 3: female     2   FALSE  Total
#> 4:      f     3    TRUE female
#> 5:      m     3    TRUE   male
#> 
#> $age
#>          code level is_leaf     parent
#> 1:      Total     1   FALSE       <NA>
#> 2: age_group1     2   FALSE      Total
#> 3: age_group2     2   FALSE      Total
#> 4: age_group3     2    TRUE      Total
#> 5: age_group4     2    TRUE      Total
#> 6: age_group5     2    TRUE      Total
#> 7: age_group6     2    TRUE      Total
#> 8:       ag1a     3    TRUE age_group1
#> 9:       ag2a     3    TRUE age_group2
#> 

# which variables have been defined?
tab$allvars()

#> $cntvars
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"
#> 
#> $numvars
#> [1] "expend"  "income"  "savings" "mixed"  
#> 

# count variables
tab$cntvars()

#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"

# continuous variables
tab$numvars()

#> [1] "expend"  "income"  "savings" "mixed"  

# create perturbation parameters for "total" variable and
# write to yaml-file

# create a ptable using functionality from the ptable-pkg
f_yaml <- tempfile(fileext = ".yaml")
p_cnts1 <- ck_params_cnts(
  ptab = ptable::pt_ex_cnts(),
  path = f_yaml)

#> yaml configuration '/tmp/Rtmp3evMLZ/file3e5c5b3e1597.yaml' successfully written.

# read parameters from yaml-file and set them for variable `"total"`
p_cnts1 <- ck_read_yaml(path = f_yaml)

tab$params_cnts_set(val = p_cnts1, v = "total")

#> --> setting perturbation parameters for variable 'total'

# create alternative perturbation parameters by specifying parameters
para2 <- ptable::pt_create_pParams(
  D = 8, V = 3, js = 2, table = "cnts")

#> Ignoring argument `step` for freqency tables

p_cnts2 <- ck_params_cnts(ptab = para2)

# use these ptable it for the remaining variables
tab$params_cnts_set(val = p_cnts2, v = countvars)

#> --> setting perturbation parameters for variable 'cnt_females'

#> --> setting perturbation parameters for variable 'cnt_males'

#> --> setting perturbation parameters for variable 'cnt_highincome'

# perturb a variable
tab$perturb(v = "total")

#> Count variable 'total' was perturbed.

# multiple variables can be perturbed as well
tab$perturb(v = c("cnt_males", "cnt_highincome"))

#> Count variable 'cnt_males' was perturbed.

#> Count variable 'cnt_highincome' was perturbed.

# return weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))

#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 271805 4581 271864.3461
#>  2:  Total age_group1     total 1969 118133 1968 118073.0036
#>  3:  Total       ag1a     total 1969 118133 1968 118073.0036
#>  4:  Total age_group2     total 1143  67603 1142  67543.8548
#>  5:  Total       ag2a     total 1143  67603 1142  67543.8548
#>  6:  Total age_group3     total  864  50834  864  50834.0000
#>  7:  Total age_group4     total  423  24526  423  24526.0000
#>  8:  Total age_group5     total  168  10008  166   9888.8571
#>  9:  Total age_group6     total   13    701   13    701.0000
#> 10: female      Total     total 2284 136744 2282 136624.2592
#> 11:      f      Total     total 2284 136744 2282 136624.2592
#> 12: female age_group1     total  954  57691  954  57691.0000
#> 13:      f age_group1     total  954  57691  954  57691.0000
#> 14: female       ag1a     total  954  57691  954  57691.0000
#> 15:      f       ag1a     total  954  57691  954  57691.0000
#> 16: female age_group2     total  572  34041  573  34100.5122
#> 17:      f age_group2     total  572  34041  573  34100.5122
#> 18: female       ag2a     total  572  34041  573  34100.5122
#> 19:      f       ag2a     total  572  34041  573  34100.5122
#> 20: female age_group3     total  440  25930  441  25988.9318
#> 21:      f age_group3     total  440  25930  441  25988.9318
#> 22: female age_group4     total  228  13636  227  13576.1930
#> 23:      f age_group4     total  228  13636  227  13576.1930
#> 24: female age_group5     total   84   5135   83   5073.8690
#> 25:      f age_group5     total   84   5135   83   5073.8690
#> 26: female age_group6     total    6    311    7    362.8333
#> 27:      f age_group6     total    6    311    7    362.8333
#> 28:   male      Total     total 2296 135061 2297 135119.8245
#> 29:      m      Total     total 2296 135061 2297 135119.8245
#> 30:   male age_group1     total 1015  60442 1016  60501.5488
#> 31:      m age_group1     total 1015  60442 1016  60501.5488
#> 32:   male       ag1a     total 1015  60442 1016  60501.5488
#> 33:      m       ag1a     total 1015  60442 1016  60501.5488
#> 34:   male age_group2     total  571  33562  571  33562.0000
#> 35:      m age_group2     total  571  33562  571  33562.0000
#> 36:   male       ag2a     total  571  33562  571  33562.0000
#> 37:      m       ag2a     total  571  33562  571  33562.0000
#> 38:   male age_group3     total  424  24904  425  24962.7358
#> 39:      m age_group3     total  424  24904  425  24962.7358
#> 40:   male age_group4     total  195  10890  195  10890.0000
#> 41:      m age_group4     total  195  10890  195  10890.0000
#> 42:   male age_group5     total   84   4873   85   4931.0119
#> 43:      m age_group5     total   84   4873   85   4931.0119
#> 44:   male age_group6     total    7    390    8    445.7143
#> 45:      m age_group6     total    7    390    8    445.7143
#> 46:  Total      Total cnt_males 2296 135061 2298 135178.6490
#> 47:  Total age_group1 cnt_males 1015  60442 1016  60501.5488
#> 48:  Total       ag1a cnt_males 1015  60442 1016  60501.5488
#> 49:  Total age_group2 cnt_males  571  33562  570  33503.2224
#> 50:  Total       ag2a cnt_males  571  33562  570  33503.2224
#> 51:  Total age_group3 cnt_males  424  24904  425  24962.7358
#> 52:  Total age_group4 cnt_males  195  10890  194  10834.1538
#> 53:  Total age_group5 cnt_males   84   4873   86   4989.0238
#> 54:  Total age_group6 cnt_males    7    390    8    445.7143
#> 55: female      Total cnt_males    0      0    0      0.0000
#> 56:      f      Total cnt_males    0      0    0      0.0000
#> 57: female age_group1 cnt_males    0      0    0      0.0000
#> 58:      f age_group1 cnt_males    0      0    0      0.0000
#> 59: female       ag1a cnt_males    0      0    0      0.0000
#> 60:      f       ag1a cnt_males    0      0    0      0.0000
#> 61: female age_group2 cnt_males    0      0    0      0.0000
#> 62:      f age_group2 cnt_males    0      0    0      0.0000
#> 63: female       ag2a cnt_males    0      0    0      0.0000
#> 64:      f       ag2a cnt_males    0      0    0      0.0000
#> 65: female age_group3 cnt_males    0      0    0      0.0000
#> 66:      f age_group3 cnt_males    0      0    0      0.0000
#> 67: female age_group4 cnt_males    0      0    0      0.0000
#> 68:      f age_group4 cnt_males    0      0    0      0.0000
#> 69: female age_group5 cnt_males    0      0    0      0.0000
#> 70:      f age_group5 cnt_males    0      0    0      0.0000
#> 71: female age_group6 cnt_males    0      0    0      0.0000
#> 72:      f age_group6 cnt_males    0      0    0      0.0000
#> 73:   male      Total cnt_males 2296 135061 2298 135178.6490
#> 74:      m      Total cnt_males 2296 135061 2298 135178.6490
#> 75:   male age_group1 cnt_males 1015  60442 1016  60501.5488
#> 76:      m age_group1 cnt_males 1015  60442 1016  60501.5488
#> 77:   male       ag1a cnt_males 1015  60442 1016  60501.5488
#> 78:      m       ag1a cnt_males 1015  60442 1016  60501.5488
#> 79:   male age_group2 cnt_males  571  33562  570  33503.2224
#> 80:      m age_group2 cnt_males  571  33562  570  33503.2224
#> 81:   male       ag2a cnt_males  571  33562  570  33503.2224
#> 82:      m       ag2a cnt_males  571  33562  570  33503.2224
#> 83:   male age_group3 cnt_males  424  24904  425  24962.7358
#> 84:      m age_group3 cnt_males  424  24904  425  24962.7358
#> 85:   male age_group4 cnt_males  195  10890  194  10834.1538
#> 86:      m age_group4 cnt_males  195  10890  194  10834.1538
#> 87:   male age_group5 cnt_males   84   4873   86   4989.0238
#> 88:      m age_group5 cnt_males   84   4873   86   4989.0238
#> 89:   male age_group6 cnt_males    7    390    8    445.7143
#> 90:      m age_group6 cnt_males    7    390    8    445.7143
#>        sex        age     vname  uwc     wc puwc         pwc

# numerical variables (positive variables using flex-function)
# we also write the config to a yaml file
f_yaml <- tempfile(fileext = ".yaml")

# create a ptable using functionality from the ptable-pkg
# a single ptable for all cells
ptab1 <- ptable::pt_ex_nums(parity = TRUE, separation = FALSE)

# a single ptab for all cells except for very small ones
ptab2 <- ptable::pt_ex_nums(parity = TRUE, separation = TRUE)

# different ptables for cells with even/odd number of contributors
# and very small cells
ptab3 <- ptable::pt_ex_nums(parity = FALSE, separation = TRUE)

p_nums1 <- ck_params_nums(
  ptab = ptab1,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.30, 0.03),
    epsilon = c(1, 0.5, 0.2),
    q = 3),
  mu_c = 2,
  same_key = FALSE,
  use_zero_rkeys = FALSE,
  path = f_yaml)

#> yaml configuration '/tmp/Rtmp3evMLZ/file3e5c7b616b33.yaml' successfully written.

# we read the parameters from the yaml-file
p_nums1 <- ck_read_yaml(path = f_yaml)

# for variables with positive and negative values
p_nums2 <- ck_params_nums(
  ptab = ptab2,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.15, 0.02),
    epsilon = c(1, 0.4, 0.15),
    q = 3),
  mu_c = 2,
  same_key = FALSE)

# simple perturbation parameters (not using the flex-function approach)
p_nums3 <- ck_params_nums(
  ptab = ptab3,
  type = "mean",
  mult_params = ck_simpleparams(p = 0.25),
  mu_c = 2,
  same_key = FALSE)

# use `p_nums1` for all variables
tab$params_nums_set(p_nums1, c("savings", "income", "expend"))

#> --> setting perturbation parameters for variable 'savings'

#> --> setting perturbation parameters for variable 'income'

#> --> setting perturbation parameters for variable 'expend'

# use different parameters for variable `mixed`
tab$params_nums_set(p_nums2, v = "mixed")

#> --> setting perturbation parameters for variable 'mixed'

# identify sensitive cells to which extra protection (`mu_c`) is added.
tab$supp_p(v = "income", p = 85)

#> p%-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)

tab$supp_pq(v = "income", p = 85, q = 90)

#> pq-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)

tab$supp_nk(v = "income", n = 2, k = 90)

#> nk-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)

tab$supp_freq(v = "income", n = 14, weighted = FALSE)

#> freq-rule: 5 new sensitive cells (incl. duplicates) found (total: 5)

tab$supp_val(v = "income", n = 10000, weighted = TRUE)

#> val-rule: 0 new sensitive cells (incl. duplicates) found (total: 5)

# perturb variables
tab$perturb(v = c("income", "savings"))

#> Numeric variable 'income' was perturbed.

#> Numeric variable 'savings' was perturbed.

# extract results
tab$numtab("income", mean_before_sum = TRUE)

#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1361721621 1361741064
#>  2:  Total age_group1 income  9810547  588652164  588675165
#>  3:  Total       ag1a income  9810547  588652164  588675165
#>  4:  Total age_group2 income  5692119  337334479  337194403
#>  5:  Total       ag2a income  5692119  337334479  337194403
#>  6:  Total age_group3 income  4406946  260491052  260312756
#>  7:  Total age_group4 income  2133543  121474229  121466121
#>  8:  Total age_group5 income   848151   50095101   49956861
#>  9:  Total age_group6 income    61672    3674596    3945028
#> 10: female      Total income 11690929  701930760  701911570
#> 11:      f      Total income 11690929  701930760  701911570
#> 12: female age_group1 income  4933383  297254963  297295890
#> 13:      f age_group1 income  4933383  297254963  297295890
#> 14: female       ag1a income  4933383  297254963  297295890
#> 15:      f       ag1a income  4933383  297254963  297295890
#> 16: female age_group2 income  2880740  173000217  173061403
#> 17:      f age_group2 income  2880740  173000217  173061403
#> 18: female       ag2a income  2880740  173000217  173061403
#> 19:      f       ag2a income  2880740  173000217  173061403
#> 20: female age_group3 income  2238777  132715277  132728556
#> 21:      f age_group3 income  2238777  132715277  132728556
#> 22: female age_group4 income  1155033   69236510   69308924
#> 23:      f age_group4 income  1155033   69236510   69308924
#> 24: female age_group5 income   455017   28199231   28096327
#> 25:      f age_group5 income   455017   28199231   28096327
#> 26: female age_group6 income    27979    1524562    1580164
#> 27:      f age_group6 income    27979    1524562    1580164
#> 28:   male      Total income 11262049  659790861  659898650
#> 29:      m      Total income 11262049  659790861  659898650
#> 30:   male age_group1 income  4877164  291397201  291373758
#> 31:      m age_group1 income  4877164  291397201  291373758
#> 32:   male       ag1a income  4877164  291397201  291373758
#> 33:      m       ag1a income  4877164  291397201  291373758
#> 34:   male age_group2 income  2811379  164334262  164223117
#> 35:      m age_group2 income  2811379  164334262  164223117
#> 36:   male       ag2a income  2811379  164334262  164223117
#> 37:      m       ag2a income  2811379  164334262  164223117
#> 38:   male age_group3 income  2168169  127775775  127809895
#> 39:      m age_group3 income  2168169  127775775  127809895
#> 40:   male age_group4 income   978510   52237719   52232199
#> 41:      m age_group4 income   978510   52237719   52232199
#> 42:   male age_group5 income   393134   21895870   21872225
#> 43:      m age_group5 income   393134   21895870   21872225
#> 44:   male age_group6 income    33693    2150034    2097797
#> 45:      m age_group6 income    33693    2150034    2097797
#>        sex        age  vname      uws         ws        pws

tab$numtab("income", mean_before_sum = FALSE)

#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1361721621 1361731342
#>  2:  Total age_group1 income  9810547  588652164  588663664
#>  3:  Total       ag1a income  9810547  588652164  588663664
#>  4:  Total age_group2 income  5692119  337334479  337264434
#>  5:  Total       ag2a income  5692119  337334479  337264434
#>  6:  Total age_group3 income  4406946  260491052  260401889
#>  7:  Total age_group4 income  2133543  121474229  121470175
#>  8:  Total age_group5 income   848151   50095101   50025933
#>  9:  Total age_group6 income    61672    3674596    3807412
#> 10: female      Total income 11690929  701930760  701921165
#> 11:      f      Total income 11690929  701930760  701921165
#> 12: female age_group1 income  4933383  297254963  297275426
#> 13:      f age_group1 income  4933383  297254963  297275426
#> 14: female       ag1a income  4933383  297254963  297275426
#> 15:      f       ag1a income  4933383  297254963  297275426
#> 16: female age_group2 income  2880740  173000217  173030808
#> 17:      f age_group2 income  2880740  173000217  173030808
#> 18: female       ag2a income  2880740  173000217  173030808
#> 19:      f       ag2a income  2880740  173000217  173030808
#> 20: female age_group3 income  2238777  132715277  132721917
#> 21:      f age_group3 income  2238777  132715277  132721917
#> 22: female age_group4 income  1155033   69236510   69272708
#> 23:      f age_group4 income  1155033   69236510   69272708
#> 24: female age_group5 income   455017   28199231   28147732
#> 25:      f age_group5 income   455017   28199231   28147732
#> 26: female age_group6 income    27979    1524562    1552114
#> 27:      f age_group6 income    27979    1524562    1552114
#> 28:   male      Total income 11262049  659790861  659844753
#> 29:      m      Total income 11262049  659790861  659844753
#> 30:   male age_group1 income  4877164  291397201  291385479
#> 31:      m age_group1 income  4877164  291397201  291385479
#> 32:   male       ag1a income  4877164  291397201  291385479
#> 33:      m       ag1a income  4877164  291397201  291385479
#> 34:   male age_group2 income  2811379  164334262  164278680
#> 35:      m age_group2 income  2811379  164334262  164278680
#> 36:   male       ag2a income  2811379  164334262  164278680
#> 37:      m       ag2a income  2811379  164334262  164278680
#> 38:   male age_group3 income  2168169  127775775  127792834
#> 39:      m age_group3 income  2168169  127775775  127792834
#> 40:   male age_group4 income   978510   52237719   52234959
#> 41:      m age_group4 income   978510   52237719   52234959
#> 42:   male age_group5 income   393134   21895870   21884044
#> 43:      m age_group5 income   393134   21895870   21884044
#> 44:   male age_group6 income    33693    2150034    2123755
#> 45:      m age_group6 income    33693    2150034    2123755
#>        sex        age  vname      uws         ws        pws

tab$numtab("savings")

#>        sex        age   vname     uws        ws         pws
#>  1:  Total      Total savings 2273532 135267041 135265347.6
#>  2:  Total age_group1 savings  982386  59122198  59130161.3
#>  3:  Total       ag1a savings  982386  59122198  59130161.3
#>  4:  Total age_group2 savings  552336  32833221  32824739.8
#>  5:  Total       ag2a savings  552336  32833221  32824739.8
#>  6:  Total age_group3 savings  437101  25701815  25700631.9
#>  7:  Total age_group4 savings  214661  12385418  12385008.2
#>  8:  Total age_group5 savings   80451   4869334   4861503.8
#>  9:  Total age_group6 savings    6597    355055    361425.6
#> 10: female      Total savings 1113716  67050233  67051276.9
#> 11:      f      Total savings 1113716  67050233  67051276.9
#> 12: female age_group1 savings  464726  28363891  28366217.1
#> 13:      f age_group1 savings  464726  28363891  28366217.1
#> 14: female       ag1a savings  464726  28363891  28366217.1
#> 15:      f       ag1a savings  464726  28363891  28366217.1
#> 16: female age_group2 savings  271413  16330745  16334373.9
#> 17:      f age_group2 savings  271413  16330745  16334373.9
#> 18: female       ag2a savings  271413  16330745  16334373.9
#> 19:      f       ag2a savings  271413  16330745  16334373.9
#> 20: female age_group3 savings  222131  13005936  13006201.1
#> 21:      f age_group3 savings  222131  13005936  13006201.1
#> 22: female age_group4 savings  115241   6872566   6873326.4
#> 23:      f age_group4 savings  115241   6872566   6873326.4
#> 24: female age_group5 savings   37218   2314820   2308882.3
#> 25:      f age_group5 savings   37218   2314820   2308882.3
#> 26: female age_group6 savings    2987    162275    161968.1
#> 27:      f age_group6 savings    2987    162275    161968.1
#> 28:   male      Total savings 1159816  68216808  68215340.5
#> 29:      m      Total savings 1159816  68216808  68215340.5
#> 30:   male age_group1 savings  517660  30758307  30757173.9
#> 31:      m age_group1 savings  517660  30758307  30757173.9
#> 32:   male       ag1a savings  517660  30758307  30757173.9
#> 33:      m       ag1a savings  517660  30758307  30757173.9
#> 34:   male age_group2 savings  280923  16502476  16498865.7
#> 35:      m age_group2 savings  280923  16502476  16498865.7
#> 36:   male       ag2a savings  280923  16502476  16498865.7
#> 37:      m       ag2a savings  280923  16502476  16498865.7
#> 38:   male age_group3 savings  214970  12695879  12703360.4
#> 39:      m age_group3 savings  214970  12695879  12703360.4
#> 40:   male age_group4 savings   99420   5512852   5513415.1
#> 41:      m age_group4 savings   99420   5512852   5513415.1
#> 42:   male age_group5 savings   43233   2554514   2554413.8
#> 43:      m age_group5 savings   43233   2554514   2554413.8
#> 44:   male age_group6 savings    3610    192780    193188.4
#> 45:      m age_group6 savings    3610    192780    193188.4
#>        sex        age   vname     uws        ws         pws

# results can be resetted, too
tab$reset_cntvars(v = "cnt_males")

# we can then set other parameters and perturb again
tab$params_cnts_set(val = p_cnts1, v = "cnt_males")

#> --> setting perturbation parameters for variable 'cnt_males'

tab$perturb(v = "cnt_males")

#> Count variable 'cnt_males' was perturbed.

# write to a file "outtab.csv" (.csv is automatically added to the path)
if (FALSE) {
tab$freqtab(v = c("total", "cnt_males"), path = "outtab.csv")
}

# show results containing weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))

#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 271805 4581 271864.3461
#>  2:  Total age_group1     total 1969 118133 1968 118073.0036
#>  3:  Total       ag1a     total 1969 118133 1968 118073.0036
#>  4:  Total age_group2     total 1143  67603 1142  67543.8548
#>  5:  Total       ag2a     total 1143  67603 1142  67543.8548
#>  6:  Total age_group3     total  864  50834  864  50834.0000
#>  7:  Total age_group4     total  423  24526  423  24526.0000
#>  8:  Total age_group5     total  168  10008  166   9888.8571
#>  9:  Total age_group6     total   13    701   13    701.0000
#> 10: female      Total     total 2284 136744 2282 136624.2592
#> 11:      f      Total     total 2284 136744 2282 136624.2592
#> 12: female age_group1     total  954  57691  954  57691.0000
#> 13:      f age_group1     total  954  57691  954  57691.0000
#> 14: female       ag1a     total  954  57691  954  57691.0000
#> 15:      f       ag1a     total  954  57691  954  57691.0000
#> 16: female age_group2     total  572  34041  573  34100.5122
#> 17:      f age_group2     total  572  34041  573  34100.5122
#> 18: female       ag2a     total  572  34041  573  34100.5122
#> 19:      f       ag2a     total  572  34041  573  34100.5122
#> 20: female age_group3     total  440  25930  441  25988.9318
#> 21:      f age_group3     total  440  25930  441  25988.9318
#> 22: female age_group4     total  228  13636  227  13576.1930
#> 23:      f age_group4     total  228  13636  227  13576.1930
#> 24: female age_group5     total   84   5135   83   5073.8690
#> 25:      f age_group5     total   84   5135   83   5073.8690
#> 26: female age_group6     total    6    311    7    362.8333
#> 27:      f age_group6     total    6    311    7    362.8333
#> 28:   male      Total     total 2296 135061 2297 135119.8245
#> 29:      m      Total     total 2296 135061 2297 135119.8245
#> 30:   male age_group1     total 1015  60442 1016  60501.5488
#> 31:      m age_group1     total 1015  60442 1016  60501.5488
#> 32:   male       ag1a     total 1015  60442 1016  60501.5488
#> 33:      m       ag1a     total 1015  60442 1016  60501.5488
#> 34:   male age_group2     total  571  33562  571  33562.0000
#> 35:      m age_group2     total  571  33562  571  33562.0000
#> 36:   male       ag2a     total  571  33562  571  33562.0000
#> 37:      m       ag2a     total  571  33562  571  33562.0000
#> 38:   male age_group3     total  424  24904  425  24962.7358
#> 39:      m age_group3     total  424  24904  425  24962.7358
#> 40:   male age_group4     total  195  10890  195  10890.0000
#> 41:      m age_group4     total  195  10890  195  10890.0000
#> 42:   male age_group5     total   84   4873   85   4931.0119
#> 43:      m age_group5     total   84   4873   85   4931.0119
#> 44:   male age_group6     total    7    390    8    445.7143
#> 45:      m age_group6     total    7    390    8    445.7143
#> 46:  Total      Total cnt_males 2296 135061 2297 135119.8245
#> 47:  Total age_group1 cnt_males 1015  60442 1016  60501.5488
#> 48:  Total       ag1a cnt_males 1015  60442 1016  60501.5488
#> 49:  Total age_group2 cnt_males  571  33562  571  33562.0000
#> 50:  Total       ag2a cnt_males  571  33562  571  33562.0000
#> 51:  Total age_group3 cnt_males  424  24904  425  24962.7358
#> 52:  Total age_group4 cnt_males  195  10890  195  10890.0000
#> 53:  Total age_group5 cnt_males   84   4873   85   4931.0119
#> 54:  Total age_group6 cnt_males    7    390    8    445.7143
#> 55: female      Total cnt_males    0      0    0      0.0000
#> 56:      f      Total cnt_males    0      0    0      0.0000
#> 57: female age_group1 cnt_males    0      0    0      0.0000
#> 58:      f age_group1 cnt_males    0      0    0      0.0000
#> 59: female       ag1a cnt_males    0      0    0      0.0000
#> 60:      f       ag1a cnt_males    0      0    0      0.0000
#> 61: female age_group2 cnt_males    0      0    0      0.0000
#> 62:      f age_group2 cnt_males    0      0    0      0.0000
#> 63: female       ag2a cnt_males    0      0    0      0.0000
#> 64:      f       ag2a cnt_males    0      0    0      0.0000
#> 65: female age_group3 cnt_males    0      0    0      0.0000
#> 66:      f age_group3 cnt_males    0      0    0      0.0000
#> 67: female age_group4 cnt_males    0      0    0      0.0000
#> 68:      f age_group4 cnt_males    0      0    0      0.0000
#> 69: female age_group5 cnt_males    0      0    0      0.0000
#> 70:      f age_group5 cnt_males    0      0    0      0.0000
#> 71: female age_group6 cnt_males    0      0    0      0.0000
#> 72:      f age_group6 cnt_males    0      0    0      0.0000
#> 73:   male      Total cnt_males 2296 135061 2297 135119.8245
#> 74:      m      Total cnt_males 2296 135061 2297 135119.8245
#> 75:   male age_group1 cnt_males 1015  60442 1016  60501.5488
#> 76:      m age_group1 cnt_males 1015  60442 1016  60501.5488
#> 77:   male       ag1a cnt_males 1015  60442 1016  60501.5488
#> 78:      m       ag1a cnt_males 1015  60442 1016  60501.5488
#> 79:   male age_group2 cnt_males  571  33562  571  33562.0000
#> 80:      m age_group2 cnt_males  571  33562  571  33562.0000
#> 81:   male       ag2a cnt_males  571  33562  571  33562.0000
#> 82:      m       ag2a cnt_males  571  33562  571  33562.0000
#> 83:   male age_group3 cnt_males  424  24904  425  24962.7358
#> 84:      m age_group3 cnt_males  424  24904  425  24962.7358
#> 85:   male age_group4 cnt_males  195  10890  195  10890.0000
#> 86:      m age_group4 cnt_males  195  10890  195  10890.0000
#> 87:   male age_group5 cnt_males   84   4873   85   4931.0119
#> 88:      m age_group5 cnt_males   84   4873   85   4931.0119
#> 89:   male age_group6 cnt_males    7    390    8    445.7143
#> 90:      m age_group6 cnt_males    7    390    8    445.7143
#>        sex        age     vname  uwc     wc puwc         pwc

# utility measures for a count variable
tab$measures_cnts(v = "total", exclude_zeros = TRUE)

#> $overview
#>    noise cnt        pct
#> 1:    -1  21 0.46666667
#> 2:     0  13 0.28888889
#> 3:     1   8 0.17777778
#> 4:     2   3 0.06666667
#> 
#> $measures
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 1.000 0.000 0.008
#>  5:    Q40 1.000 0.001 0.013
#>  6:   Mean 0.778 0.016 0.033
#>  7: Median 1.000 0.001 0.016
#>  8:    Q60 1.000 0.002 0.021
#>  9:    Q70 1.000 0.002 0.024
#> 10:    Q80 1.000 0.006 0.037
#> 11:    Q90 1.000 0.012 0.068
#> 12:    Q95 1.800 0.143 0.183
#> 13:    Q99 2.000 0.167 0.196
#> 14:    Max 2.000 0.167 0.196
#> 
#> $cumdistr_d1
#>    cat cnt       pct
#> 1:   0  13 0.2888889
#> 2:   1  42 0.9333333
#> 3:   2  45 1.0000000
#> 
#> $cumdistr_d2
#>            cat cnt       pct
#> 1:    [0,0.02]  41 0.9111111
#> 2: (0.02,0.05]  41 0.9111111
#> 3:  (0.05,0.1]  41 0.9111111
#> 4:   (0.1,0.2]  45 1.0000000
#> 5:   (0.2,0.3]  45 1.0000000
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $cumdistr_d3
#>            cat cnt       pct
#> 1:    [0,0.02]  24 0.5333333
#> 2: (0.02,0.05]  36 0.8000000
#> 3:  (0.05,0.1]  41 0.9111111
#> 4:   (0.1,0.2]  45 1.0000000
#> 5:   (0.2,0.3]  45 1.0000000
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $false_zero
#> [1] 0
#> 
#> $false_nonzero
#> [1] 0
#> 
#> $exclude_zeros
#> [1] TRUE
#> 

# modifications for perturbed count variables
tab$mod_cnts()

#>        sex        age row_nr pert      ckey  countvar
#>   1: Total      Total     16    1 0.9083113     total
#>   2: Total age_group1     14   -1 0.2636681     total
#>   3: Total       ag1a     14   -1 0.2636681     total
#>   4: Total age_group2     14   -1 0.1012343     total
#>   5: Total       ag2a     14   -1 0.1012343     total
#>  ---                                                 
#> 131:     m age_group4     15    0 0.3617172 cnt_males
#> 132:  male age_group5     16    1 0.9172724 cnt_males
#> 133:     m age_group5     16    1 0.9172724 cnt_males
#> 134:  male age_group6     16    1 0.7359124 cnt_males
#> 135:     m age_group6     16    1 0.7359124 cnt_males

# display a summary about utility measures
tab$summary()

#> ┌──────────────────────────────────────────────┐
#> │Utility measures for perturbed count variables│
#> └──────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#>          countvar Min Q10 Q20 Q30 Q40   Mean Median Q60 Q70 Q80 Q90 Q95 Q99 Max
#> 1:          total  -2  -1  -1   0   0  0.156      0 1.0   1   1 1.0   1   1   1
#> 2: cnt_highincome  -3  -2  -2  -1   0 -0.267      0 0.0   0   1 1.6   2   2   2
#> 3:      cnt_males   0   0   0   0   0  0.400      0 0.4   1   1 1.0   1   1   1
#> 
#> ── Distance-based measures ─────────────────────────────────────────────────────
#> ✔ Variable: 'total'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 1.000 0.000 0.008
#>  5:    Q40 1.000 0.001 0.013
#>  6:   Mean 0.778 0.016 0.033
#>  7: Median 1.000 0.001 0.016
#>  8:    Q60 1.000 0.002 0.021
#>  9:    Q70 1.000 0.002 0.024
#> 10:    Q80 1.000 0.006 0.037
#> 11:    Q90 1.000 0.012 0.068
#> 12:    Q95 1.800 0.143 0.183
#> 13:    Q99 2.000 0.167 0.196
#> 14:    Max 2.000 0.167 0.196
#> 
#> ✔ Variable: 'cnt_males'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 1.000 0.000 0.010
#>  6:   Mean 0.667 0.018 0.034
#>  7: Median 1.000 0.001 0.016
#>  8:    Q60 1.000 0.001 0.016
#>  9:    Q70 1.000 0.002 0.024
#> 10:    Q80 1.000 0.010 0.048
#> 11:    Q90 1.000 0.064 0.106
#> 12:    Q95 1.000 0.143 0.183
#> 13:    Q99 1.000 0.143 0.183
#> 14:    Max 1.000 0.143 0.183
#> 
#> ✔ Variable: 'cnt_highincome'
#> 
#>       what   d1    d2    d3
#>  1:    Min 0.00 0.000 0.000
#>  2:    Q10 0.00 0.000 0.000
#>  3:    Q20 0.00 0.000 0.000
#>  4:    Q30 0.00 0.000 0.000
#>  5:    Q40 1.00 0.010 0.062
#>  6:   Mean 1.10 0.049 0.104
#>  7: Median 1.00 0.015 0.070
#>  8:    Q60 1.00 0.020 0.083
#>  9:    Q70 2.00 0.035 0.119
#> 10:    Q80 2.00 0.054 0.131
#> 11:    Q90 2.00 0.134 0.191
#> 12:    Q95 2.05 0.157 0.286
#> 13:    Q99 3.00 0.429 0.646
#> 14:    Max 3.00 0.429 0.646
#> 
#> ┌──────────────────────────────────────────────────┐
#> │Utility measures for perturbed numerical variables│
#> └──────────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────

#> Warning: no non-missing arguments to min; returning Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to min; returning Inf

#> Warning: no non-missing arguments to max; returning -Inf

#>      vname        Min       Q10       Q20        Q30        Q40      Mean
#> 1:  expend        Inf        NA        NA         NA         NA       NaN
#> 2:  income -89163.392 -55581.96 -51498.78 -11804.785 -10445.716 -2737.231
#> 3: savings  -8481.158  -5006.72  -2076.78  -1133.146   -348.055    99.669
#> 4:   mixed        Inf        NA        NA         NA         NA       NaN
#>       Median      Q60       Q70       Q80       Q90       Q95       Q99
#> 1:        NA       NA        NA        NA        NA        NA        NA
#> 2: -2760.145 11500.37 20462.711 28159.804 33954.775 50353.354 98089.415
#> 3:   265.120   563.08  1043.946  2586.683  5273.947  7481.375  7963.251
#> 4:        NA       NA        NA        NA        NA        NA        NA
#>           Max
#> 1:       -Inf
#> 2: 132815.730
#> 3:   7963.251
#> 4:       -Inf