Read perturbation parameters from yaml-files

ck_read_yaml() allows to create perturbation parameter inputs from yaml-files that were previously created using ck_params_cnts() or ck_params_nums().

ck_read_yaml(path)

Arguments

path

a path to a yaml-input file

Value

an object object suitable as input to method $params_nums_set() for the perturbation of continous variables in case path was created using ck_params_nums() or an object suitable as input for $params_cnts_set() for the perturbation of counts and frequencies if the input file was generated using ck_params_cnts().

Examples

x <- ck_create_testdata()

# create some 0/1 variables that should be perturbed later
x[, cnt_females := ifelse(sex == "male", 0, 1)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              88          25.00000           0
#>    2:   2530      28       1              70          25.00000           1
#>    3:   6920     550       1              74          25.00000           0
#>    4:   7960     870       1              35          25.00000           0
#>    5:   9030      20       2              50          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              23          16.66667           1
#> 4577:   1420     987    1000              59          16.66667           0
#> 4578:   8900     684    1000              64          16.66667           0
#> 4579:   3880     294    1000              43          16.66667           1
#> 4580:   4830     911    1000              28          16.66667           0
x[, cnt_males := ifelse(sex == "male", 1, 0)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              88          25.00000           0
#>    2:   2530      28       1              70          25.00000           1
#>    3:   6920     550       1              74          25.00000           0
#>    4:   7960     870       1              35          25.00000           0
#>    5:   9030      20       2              50          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              23          16.66667           1
#> 4577:   1420     987    1000              59          16.66667           0
#> 4578:   8900     684    1000              64          16.66667           0
#> 4579:   3880     294    1000              43          16.66667           1
#> 4580:   4830     911    1000              28          16.66667           0
#>       cnt_males
#>    1:         1
#>    2:         0
#>    3:         1
#>    4:         1
#>    5:         1
#>   ---          
#> 4576:         0
#> 4577:         1
#> 4578:         1
#> 4579:         0
#> 4580:         1
x[, cnt_highincome := ifelse(income >= 9000, 1, 0)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              88          25.00000           0
#>    2:   2530      28       1              70          25.00000           1
#>    3:   6920     550       1              74          25.00000           0
#>    4:   7960     870       1              35          25.00000           0
#>    5:   9030      20       2              50          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              23          16.66667           1
#> 4577:   1420     987    1000              59          16.66667           0
#> 4578:   8900     684    1000              64          16.66667           0
#> 4579:   3880     294    1000              43          16.66667           1
#> 4580:   4830     911    1000              28          16.66667           0
#>       cnt_males cnt_highincome
#>    1:         1              0
#>    2:         0              0
#>    3:         1              0
#>    4:         1              0
#>    5:         1              1
#>   ---                         
#> 4576:         0              0
#> 4577:         1              0
#> 4578:         1              0
#> 4579:         0              0
#> 4580:         1              0
# a variable with positive and negative contributions
x[, mixed := sample(-10:10, nrow(x), replace = TRUE)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              88          25.00000           0
#>    2:   2530      28       1              70          25.00000           1
#>    3:   6920     550       1              74          25.00000           0
#>    4:   7960     870       1              35          25.00000           0
#>    5:   9030      20       2              50          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              23          16.66667           1
#> 4577:   1420     987    1000              59          16.66667           0
#> 4578:   8900     684    1000              64          16.66667           0
#> 4579:   3880     294    1000              43          16.66667           1
#> 4580:   4830     911    1000              28          16.66667           0
#>       cnt_males cnt_highincome mixed
#>    1:         1              0     4
#>    2:         0              0   -10
#>    3:         1              0     6
#>    4:         1              0    -6
#>    5:         1              1    -6
#>   ---                               
#> 4576:         0              0    -8
#> 4577:         1              0     9
#> 4578:         1              0    -8
#> 4579:         0              0     0
#> 4580:         1              0     8

# create record keys
x$rkey <- ck_generate_rkeys(dat = x)

# define required inputs

# hierarchy with some bogus codes
d_sex <- hier_create(root = "Total", nodes = c("male", "female"))
d_sex <- hier_add(d_sex, root = "female", "f")
d_sex <- hier_add(d_sex, root = "male", "m")

d_age <- hier_create(root = "Total", nodes = paste0("age_group", 1:6))
d_age <- hier_add(d_age, root = "age_group1", "ag1a")
d_age <- hier_add(d_age, root = "age_group2", "ag2a")

# define the cell key object
countvars <- c("cnt_females", "cnt_males", "cnt_highincome")
numvars <- c("expend", "income", "savings", "mixed")
tab <- ck_setup(
  x = x,
  rkey = "rkey",
  dims = list(sex = d_sex, age = d_age),
  w = "sampling_weight",
  countvars = countvars,
  numvars = numvars)

# show some information about this table instance
tab$print() # identical with print(tab)
#> ── Table Information ───────────────────────────────────────────────────────────
#> ✔ 45 cells in 2 dimensions ('sex', 'age')
#> ✔ weights: yes
#> ── Tabulated / Perturbed countvars ─────────────────────────────────────────────
#> ☐ 'total'
#> ☐ 'cnt_females'
#> ☐ 'cnt_males'
#> ☐ 'cnt_highincome'
#> ── Tabulated / Perturbed numvars ───────────────────────────────────────────────
#> ☐ 'expend'
#> ☐ 'income'
#> ☐ 'savings'
#> ☐ 'mixed'

# information about the hierarchies
tab$hierarchy_info()
#> $sex
#>      code level is_leaf parent
#> 1:  Total     1   FALSE  Total
#> 2:   male     2   FALSE  Total
#> 3:      m     3    TRUE   male
#> 4: female     2   FALSE  Total
#> 5:      f     3    TRUE female
#> 
#> $age
#>          code level is_leaf     parent
#> 1:      Total     1   FALSE      Total
#> 2: age_group1     2   FALSE      Total
#> 3:       ag1a     3    TRUE age_group1
#> 4: age_group2     2   FALSE      Total
#> 5:       ag2a     3    TRUE age_group2
#> 6: age_group3     2    TRUE      Total
#> 7: age_group4     2    TRUE      Total
#> 8: age_group5     2    TRUE      Total
#> 9: age_group6     2    TRUE      Total
#> 

# which variables have been defined?
tab$allvars()
#> $cntvars
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"
#> 
#> $numvars
#> [1] "expend"  "income"  "savings" "mixed"  
#> 

# count variables
tab$cntvars()
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"

# continuous variables
tab$numvars()
#> [1] "expend"  "income"  "savings" "mixed"  

# create perturbation parameters for "total" variable and
# write to yaml-file

# create a ptable using functionality from the ptable-pkg
f_yaml <- tempfile(fileext = ".yaml")
p_cnts1 <- ck_params_cnts(
  ptab = ptable::pt_ex_cnts(),
  path = f_yaml)
#> Ignoring argument `step` for freqency tables
#> yaml configuration '/tmp/Rtmporiz8G/file35145e63067e.yaml' successfully written.

# read parameters from yaml-file and set them for variable `"total"`
p_cnts1 <- ck_read_yaml(path = f_yaml)

tab$params_cnts_set(val = p_cnts1, v = "total")
#> --> setting perturbation parameters for variable 'total'

# create alternative perturbation parameters by specifying parameters
para2 <- ptable::pt_create_pParams(
  D = 8, V = 3, js = 2, table = "cnts")
#> Ignoring argument `step` for freqency tables

p_cnts2 <- ck_params_cnts(ptab = para2)

# use these ptable it for the remaining variables
tab$params_cnts_set(val = p_cnts2, v = countvars)
#> --> setting perturbation parameters for variable 'cnt_females'
#> --> setting perturbation parameters for variable 'cnt_males'
#> --> setting perturbation parameters for variable 'cnt_highincome'

# perturb a variable
tab$perturb(v = "total")
#> Count variable 'total' was perturbed.

# multiple variables can be perturbed as well
tab$perturb(v = c("cnt_males", "cnt_highincome"))
#> Count variable 'cnt_males' was perturbed.
#> Count variable 'cnt_highincome' was perturbed.

# return weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))
#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 273985 4578 273865.3559
#>  2:  Total age_group1     total 1969 118952 1969 118952.0000
#>  3:  Total       ag1a     total 1969 118952 1969 118952.0000
#>  4:  Total age_group2     total 1143  68280 1142  68220.2625
#>  5:  Total       ag2a     total 1143  68280 1142  68220.2625
#>  6:  Total age_group3     total  864  51455  862  51335.8912
#>  7:  Total age_group4     total  423  24570  423  24570.0000
#>  8:  Total age_group5     total  168  10105  169  10165.1488
#>  9:  Total age_group6     total   13    623   12    575.0769
#> 10:   male      Total     total 2296 136415 2295 136355.5858
#> 11:      m      Total     total 2296 136415 2295 136355.5858
#> 12:   male age_group1     total 1015  60639 1015  60639.0000
#> 13:      m age_group1     total 1015  60639 1015  60639.0000
#> 14:   male       ag1a     total 1015  60639 1015  60639.0000
#> 15:      m       ag1a     total 1015  60639 1015  60639.0000
#> 16:   male age_group2     total  571  34196  573  34315.7758
#> 17:      m age_group2     total  571  34196  573  34315.7758
#> 18:   male       ag2a     total  571  34196  573  34315.7758
#> 19:      m       ag2a     total  571  34196  573  34315.7758
#> 20:   male age_group3     total  424  24985  424  24985.0000
#> 21:      m age_group3     total  424  24985  424  24985.0000
#> 22:   male age_group4     total  195  11210  197  11324.9744
#> 23:      m age_group4     total  195  11210  197  11324.9744
#> 24:   male age_group5     total   84   5116   86   5237.8095
#> 25:      m age_group5     total   84   5116   86   5237.8095
#> 26:   male age_group6     total    7    269    5    192.1429
#> 27:      m age_group6     total    7    269    5    192.1429
#> 28: female      Total     total 2284 137570 2285 137630.2320
#> 29:      f      Total     total 2284 137570 2285 137630.2320
#> 30: female age_group1     total  954  58313  956  58435.2495
#> 31:      f age_group1     total  954  58313  956  58435.2495
#> 32: female       ag1a     total  954  58313  956  58435.2495
#> 33:      f       ag1a     total  954  58313  956  58435.2495
#> 34: female age_group2     total  572  34084  571  34024.4126
#> 35:      f age_group2     total  572  34084  571  34024.4126
#> 36: female       ag2a     total  572  34084  571  34024.4126
#> 37:      f       ag2a     total  572  34084  571  34024.4126
#> 38: female age_group3     total  440  26470  440  26470.0000
#> 39:      f age_group3     total  440  26470  440  26470.0000
#> 40: female age_group4     total  228  13360  228  13360.0000
#> 41:      f age_group4     total  228  13360  228  13360.0000
#> 42: female age_group5     total   84   4989   85   5048.3929
#> 43:      f age_group5     total   84   4989   85   5048.3929
#> 44: female age_group6     total    6    354    5    295.0000
#> 45:      f age_group6     total    6    354    5    295.0000
#> 46:  Total      Total cnt_males 2296 136415 2295 136355.5858
#> 47:  Total age_group1 cnt_males 1015  60639 1016  60698.7429
#> 48:  Total       ag1a cnt_males 1015  60639 1016  60698.7429
#> 49:  Total age_group2 cnt_males  571  34196  575  34435.5517
#> 50:  Total       ag2a cnt_males  571  34196  575  34435.5517
#> 51:  Total age_group3 cnt_males  424  24985  425  25043.9269
#> 52:  Total age_group4 cnt_males  195  11210  198  11382.4615
#> 53:  Total age_group5 cnt_males   84   5116   87   5298.7143
#> 54:  Total age_group6 cnt_males    7    269    3    115.2857
#> 55:   male      Total cnt_males 2296 136415 2295 136355.5858
#> 56:      m      Total cnt_males 2296 136415 2295 136355.5858
#> 57:   male age_group1 cnt_males 1015  60639 1016  60698.7429
#> 58:      m age_group1 cnt_males 1015  60639 1016  60698.7429
#> 59:   male       ag1a cnt_males 1015  60639 1016  60698.7429
#> 60:      m       ag1a cnt_males 1015  60639 1016  60698.7429
#> 61:   male age_group2 cnt_males  571  34196  575  34435.5517
#> 62:      m age_group2 cnt_males  571  34196  575  34435.5517
#> 63:   male       ag2a cnt_males  571  34196  575  34435.5517
#> 64:      m       ag2a cnt_males  571  34196  575  34435.5517
#> 65:   male age_group3 cnt_males  424  24985  425  25043.9269
#> 66:      m age_group3 cnt_males  424  24985  425  25043.9269
#> 67:   male age_group4 cnt_males  195  11210  198  11382.4615
#> 68:      m age_group4 cnt_males  195  11210  198  11382.4615
#> 69:   male age_group5 cnt_males   84   5116   87   5298.7143
#> 70:      m age_group5 cnt_males   84   5116   87   5298.7143
#> 71:   male age_group6 cnt_males    7    269    3    115.2857
#> 72:      m age_group6 cnt_males    7    269    3    115.2857
#> 73: female      Total cnt_males    0      0    0      0.0000
#> 74:      f      Total cnt_males    0      0    0      0.0000
#> 75: female age_group1 cnt_males    0      0    0      0.0000
#> 76:      f age_group1 cnt_males    0      0    0      0.0000
#> 77: female       ag1a cnt_males    0      0    0      0.0000
#> 78:      f       ag1a cnt_males    0      0    0      0.0000
#> 79: female age_group2 cnt_males    0      0    0      0.0000
#> 80:      f age_group2 cnt_males    0      0    0      0.0000
#> 81: female       ag2a cnt_males    0      0    0      0.0000
#> 82:      f       ag2a cnt_males    0      0    0      0.0000
#> 83: female age_group3 cnt_males    0      0    0      0.0000
#> 84:      f age_group3 cnt_males    0      0    0      0.0000
#> 85: female age_group4 cnt_males    0      0    0      0.0000
#> 86:      f age_group4 cnt_males    0      0    0      0.0000
#> 87: female age_group5 cnt_males    0      0    0      0.0000
#> 88:      f age_group5 cnt_males    0      0    0      0.0000
#> 89: female age_group6 cnt_males    0      0    0      0.0000
#> 90:      f age_group6 cnt_males    0      0    0      0.0000
#>        sex        age     vname  uwc     wc puwc         pwc

# numerical variables (positive variables using flex-function)
# we also write the config to a yaml file
f_yaml <- tempfile(fileext = ".yaml")

# create a ptable using functionality from the ptable-pkg
# a single ptable for all cells
ptab1 <- ptable::pt_ex_nums(parity = TRUE, separation = FALSE)

# a single ptab for all cells except for very small ones
ptab2 <- ptable::pt_ex_nums(parity = TRUE, separation = TRUE)

# different ptables for cells with even/odd number of contributors
# and very small cells
ptab3 <- ptable::pt_ex_nums(parity = FALSE, separation = TRUE)

p_nums1 <- ck_params_nums(
  ptab = ptab1,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.30, 0.03),
    epsilon = c(1, 0.5, 0.2),
    q = 3),
  mu_c = 2,
  same_key = FALSE,
  use_zero_rkeys = FALSE,
  path = f_yaml)
#> yaml configuration '/tmp/Rtmporiz8G/file351467b65f02.yaml' successfully written.

# we read the parameters from the yaml-file
p_nums1 <- ck_read_yaml(path = f_yaml)

# for variables with positive and negative values
p_nums2 <- ck_params_nums(
  ptab = ptab2,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.15, 0.02),
    epsilon = c(1, 0.4, 0.15),
    q = 3),
  mu_c = 2,
  same_key = FALSE)

# simple perturbation parameters (not using the flex-function approach)
p_nums3 <- ck_params_nums(
  ptab = ptab3,
  type = "mean",
  mult_params = ck_simpleparams(p = 0.25),
  mu_c = 2,
  same_key = FALSE)

# use `p_nums1` for all variables
tab$params_nums_set(p_nums1, c("savings", "income", "expend"))
#> --> setting perturbation parameters for variable 'savings'
#> --> setting perturbation parameters for variable 'income'
#> --> setting perturbation parameters for variable 'expend'

# use different parameters for variable `mixed`
tab$params_nums_set(p_nums2, v = "mixed")
#> --> setting perturbation parameters for variable 'mixed'

# identify sensitive cells to which extra protection (`mu_c`) is added.
tab$supp_p(v = "income", p = 85)
#> computing contributing indices | rawdata <--> table; this might take a while
#> p%-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_pq(v = "income", p = 85, q = 90)
#> computing contributing indices | rawdata <--> table; this might take a while
#> pq-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_nk(v = "income", n = 2, k = 90)
#> computing contributing indices | rawdata <--> table; this might take a while
#> nk-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_freq(v = "income", n = 14, weighted = FALSE)
#> freq-rule: 5 new sensitive cells (incl. duplicates) found (total: 5)
tab$supp_val(v = "income", n = 10000, weighted = TRUE)
#> val-rule: 0 new sensitive cells (incl. duplicates) found (total: 5)
tab$supp_cells(
  v = "income",
  inp = data.frame(
    sex = c("female", "female"),
    "age" = c("age_group1", "age_group3")
  )
)
#> cell-rule: 2 new sensitive cells (incl. duplicates) found (total: 7)

# perturb variables
tab$perturb(v = c("income", "savings"))
#> Numeric variable 'income' was perturbed.
#> Numeric variable 'savings' was perturbed.

# extract results
tab$numtab("income", mean_before_sum = TRUE)
#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1371820922 1371951198
#>  2:  Total age_group1 income  9810547  591712474  591652914
#>  3:  Total       ag1a income  9810547  591712474  591652914
#>  4:  Total age_group2 income  5692119  339875381  339942546
#>  5:  Total       ag2a income  5692119  339875381  339942546
#>  6:  Total age_group3 income  4406946  260647962  260676562
#>  7:  Total age_group4 income  2133543  124946543  124890656
#>  8:  Total age_group5 income   848151   51577793   51543712
#>  9:  Total age_group6 income    61672    3060769    3110344
#> 10:   male      Total income 11262049  670131997  670177767
#> 11:      m      Total income 11262049  670131997  670177767
#> 12:   male age_group1 income  4877164  290152823  290177050
#> 13:      m age_group1 income  4877164  290152823  290177050
#> 14:   male       ag1a income  4877164  290152823  290177050
#> 15:      m       ag1a income  4877164  290152823  290177050
#> 16:   male age_group2 income  2811379  168329461  168469762
#> 17:      m age_group2 income  2811379  168329461  168469762
#> 18:   male       ag2a income  2811379  168329461  168469762
#> 19:      m       ag2a income  2811379  168329461  168469762
#> 20:   male age_group3 income  2168169  129121167  129068239
#> 21:      m age_group3 income  2168169  129121167  129068239
#> 22:   male age_group4 income   978510   57068365   57074513
#> 23:      m age_group4 income   978510   57068365   57074513
#> 24:   male age_group5 income   393134   24231349   24226580
#> 25:      m age_group5 income   393134   24231349   24226580
#> 26:   male age_group6 income    33693    1228832    1178718
#> 27:      m age_group6 income    33693    1228832    1178718
#> 28: female      Total income 11690929  701688925  701571471
#> 29:      f      Total income 11690929  701688925  701571471
#> 30: female age_group1 income  4933383  301559651  301687413
#> 31:      f age_group1 income  4933383  301559651  301687413
#> 32: female       ag1a income  4933383  301559651  301687413
#> 33:      f       ag1a income  4933383  301559651  301687413
#> 34: female age_group2 income  2880740  171545920  171496421
#> 35:      f age_group2 income  2880740  171545920  171496421
#> 36: female       ag2a income  2880740  171545920  171496421
#> 37:      f       ag2a income  2880740  171545920  171496421
#> 38: female age_group3 income  2238777  131526795  131544421
#> 39:      f age_group3 income  2238777  131526795  131544421
#> 40: female age_group4 income  1155033   67878178   67922977
#> 41:      f age_group4 income  1155033   67878178   67922977
#> 42: female age_group5 income   455017   27346444   27362776
#> 43:      f age_group5 income   455017   27346444   27362776
#> 44: female age_group6 income    27979    1831937    1699082
#> 45:      f age_group6 income    27979    1831937    1699082
#>        sex        age  vname      uws         ws        pws
tab$numtab("income", mean_before_sum = FALSE)
#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1371820922 1371886058
#>  2:  Total age_group1 income  9810547  591712474  591682693
#>  3:  Total       ag1a income  9810547  591712474  591682693
#>  4:  Total age_group2 income  5692119  339875381  339908962
#>  5:  Total       ag2a income  5692119  339875381  339908962
#>  6:  Total age_group3 income  4406946  260647962  260662261
#>  7:  Total age_group4 income  2133543  124946543  124918596
#>  8:  Total age_group5 income   848151   51577793   51560749
#>  9:  Total age_group6 income    61672    3060769    3085457
#> 10:   male      Total income 11262049  670131997  670154882
#> 11:      m      Total income 11262049  670131997  670154882
#> 12:   male age_group1 income  4877164  290152823  290164936
#> 13:      m age_group1 income  4877164  290152823  290164936
#> 14:   male       ag1a income  4877164  290152823  290164936
#> 15:      m       ag1a income  4877164  290152823  290164936
#> 16:   male age_group2 income  2811379  168329461  168399597
#> 17:      m age_group2 income  2811379  168329461  168399597
#> 18:   male       ag2a income  2811379  168329461  168399597
#> 19:      m       ag2a income  2811379  168329461  168399597
#> 20:   male age_group3 income  2168169  129121167  129094700
#> 21:      m age_group3 income  2168169  129121167  129094700
#> 22:   male age_group4 income   978510   57068365   57071439
#> 23:      m age_group4 income   978510   57068365   57071439
#> 24:   male age_group5 income   393134   24231349   24228964
#> 25:      m age_group5 income   393134   24231349   24228964
#> 26:   male age_group6 income    33693    1228832    1203514
#> 27:      m age_group6 income    33693    1228832    1203514
#> 28: female      Total income 11690929  701688925  701630195
#> 29:      f      Total income 11690929  701688925  701630195
#> 30: female age_group1 income  4933383  301559651  301623525
#> 31:      f age_group1 income  4933383  301559651  301623525
#> 32: female       ag1a income  4933383  301559651  301623525
#> 33:      f       ag1a income  4933383  301559651  301623525
#> 34: female age_group2 income  2880740  171545920  171521169
#> 35:      f age_group2 income  2880740  171545920  171521169
#> 36: female       ag2a income  2880740  171545920  171521169
#> 37:      f       ag2a income  2880740  171545920  171521169
#> 38: female age_group3 income  2238777  131526795  131535608
#> 39:      f age_group3 income  2238777  131526795  131535608
#> 40: female age_group4 income  1155033   67878178   67900574
#> 41:      f age_group4 income  1155033   67878178   67900574
#> 42: female age_group5 income   455017   27346444   27354609
#> 43:      f age_group5 income   455017   27346444   27354609
#> 44: female age_group6 income    27979    1831937    1764259
#> 45:      f age_group6 income    27979    1831937    1764259
#>        sex        age  vname      uws         ws        pws
tab$numtab("savings")
#>        sex        age   vname     uws        ws         pws
#>  1:  Total      Total savings 2273532 136033474 136034399.1
#>  2:  Total age_group1 savings  982386  59370923  59369855.1
#>  3:  Total       ag1a savings  982386  59370923  59369855.1
#>  4:  Total age_group2 savings  552336  33009687  33012351.5
#>  5:  Total       ag2a savings  552336  33009687  33012351.5
#>  6:  Total age_group3 savings  437101  25982012  25976115.5
#>  7:  Total age_group4 savings  214661  12501367  12498513.2
#>  8:  Total age_group5 savings   80451   4820256   4819028.1
#>  9:  Total age_group6 savings    6597    349229    350119.2
#> 10:   male      Total savings 1159816  68380466  68381049.5
#> 11:      m      Total savings 1159816  68380466  68381049.5
#> 12:   male age_group1 savings  517660  30727231  30726951.0
#> 13:      m age_group1 savings  517660  30727231  30726951.0
#> 14:   male       ag1a savings  517660  30727231  30726951.0
#> 15:      m       ag1a savings  517660  30727231  30726951.0
#> 16:   male age_group2 savings  280923  16807376  16813979.4
#> 17:      m age_group2 savings  280923  16807376  16813979.4
#> 18:   male       ag2a savings  280923  16807376  16813979.4
#> 19:      m       ag2a savings  280923  16807376  16813979.4
#> 20:   male age_group3 savings  214970  12511598  12513863.1
#> 21:      m age_group3 savings  214970  12511598  12513863.1
#> 22:   male age_group4 savings   99420   5583080   5583038.1
#> 23:      m age_group4 savings   99420   5583080   5583038.1
#> 24:   male age_group5 savings   43233   2604176   2602608.0
#> 25:      m age_group5 savings   43233   2604176   2602608.0
#> 26:   male age_group6 savings    3610    147005    145227.5
#> 27:      m age_group6 savings    3610    147005    145227.5
#> 28: female      Total savings 1113716  67653008  67651675.6
#> 29:      f      Total savings 1113716  67653008  67651675.6
#> 30: female age_group1 savings  464726  28643692  28650210.0
#> 31:      f age_group1 savings  464726  28643692  28650210.0
#> 32: female       ag1a savings  464726  28643692  28650210.0
#> 33:      f       ag1a savings  464726  28643692  28650210.0
#> 34: female age_group2 savings  271413  16202311  16200253.5
#> 35:      f age_group2 savings  271413  16202311  16200253.5
#> 36: female       ag2a savings  271413  16202311  16200253.5
#> 37:      f       ag2a savings  271413  16202311  16200253.5
#> 38: female age_group3 savings  222131  13470414  13470900.9
#> 39:      f age_group3 savings  222131  13470414  13470900.9
#> 40: female age_group4 savings  115241   6918287   6922846.7
#> 41:      f age_group4 savings  115241   6918287   6922846.7
#> 42: female age_group5 savings   37218   2216080   2216993.9
#> 43:      f age_group5 savings   37218   2216080   2216993.9
#> 44: female age_group6 savings    2987    202224    199540.3
#> 45:      f age_group6 savings    2987    202224    199540.3
#>        sex        age   vname     uws        ws         pws

# results can be resetted, too
tab$reset_cntvars(v = "cnt_males")

# we can then set other parameters and perturb again
tab$params_cnts_set(val = p_cnts1, v = "cnt_males")
#> --> setting perturbation parameters for variable 'cnt_males'
tab$perturb(v = "cnt_males")
#> Count variable 'cnt_males' was perturbed.

# write to a file "outtab.csv" (.csv is automatically added to the path)
if (FALSE) {
tab$freqtab(v = c("total", "cnt_males"), path = "outtab.csv")
}

# show results containing weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))
#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 273985 4578 273865.3559
#>  2:  Total age_group1     total 1969 118952 1969 118952.0000
#>  3:  Total       ag1a     total 1969 118952 1969 118952.0000
#>  4:  Total age_group2     total 1143  68280 1142  68220.2625
#>  5:  Total       ag2a     total 1143  68280 1142  68220.2625
#>  6:  Total age_group3     total  864  51455  862  51335.8912
#>  7:  Total age_group4     total  423  24570  423  24570.0000
#>  8:  Total age_group5     total  168  10105  169  10165.1488
#>  9:  Total age_group6     total   13    623   12    575.0769
#> 10:   male      Total     total 2296 136415 2295 136355.5858
#> 11:      m      Total     total 2296 136415 2295 136355.5858
#> 12:   male age_group1     total 1015  60639 1015  60639.0000
#> 13:      m age_group1     total 1015  60639 1015  60639.0000
#> 14:   male       ag1a     total 1015  60639 1015  60639.0000
#> 15:      m       ag1a     total 1015  60639 1015  60639.0000
#> 16:   male age_group2     total  571  34196  573  34315.7758
#> 17:      m age_group2     total  571  34196  573  34315.7758
#> 18:   male       ag2a     total  571  34196  573  34315.7758
#> 19:      m       ag2a     total  571  34196  573  34315.7758
#> 20:   male age_group3     total  424  24985  424  24985.0000
#> 21:      m age_group3     total  424  24985  424  24985.0000
#> 22:   male age_group4     total  195  11210  197  11324.9744
#> 23:      m age_group4     total  195  11210  197  11324.9744
#> 24:   male age_group5     total   84   5116   86   5237.8095
#> 25:      m age_group5     total   84   5116   86   5237.8095
#> 26:   male age_group6     total    7    269    5    192.1429
#> 27:      m age_group6     total    7    269    5    192.1429
#> 28: female      Total     total 2284 137570 2285 137630.2320
#> 29:      f      Total     total 2284 137570 2285 137630.2320
#> 30: female age_group1     total  954  58313  956  58435.2495
#> 31:      f age_group1     total  954  58313  956  58435.2495
#> 32: female       ag1a     total  954  58313  956  58435.2495
#> 33:      f       ag1a     total  954  58313  956  58435.2495
#> 34: female age_group2     total  572  34084  571  34024.4126
#> 35:      f age_group2     total  572  34084  571  34024.4126
#> 36: female       ag2a     total  572  34084  571  34024.4126
#> 37:      f       ag2a     total  572  34084  571  34024.4126
#> 38: female age_group3     total  440  26470  440  26470.0000
#> 39:      f age_group3     total  440  26470  440  26470.0000
#> 40: female age_group4     total  228  13360  228  13360.0000
#> 41:      f age_group4     total  228  13360  228  13360.0000
#> 42: female age_group5     total   84   4989   85   5048.3929
#> 43:      f age_group5     total   84   4989   85   5048.3929
#> 44: female age_group6     total    6    354    5    295.0000
#> 45:      f age_group6     total    6    354    5    295.0000
#> 46:  Total      Total cnt_males 2296 136415 2295 136355.5858
#> 47:  Total age_group1 cnt_males 1015  60639 1015  60639.0000
#> 48:  Total       ag1a cnt_males 1015  60639 1015  60639.0000
#> 49:  Total age_group2 cnt_males  571  34196  573  34315.7758
#> 50:  Total       ag2a cnt_males  571  34196  573  34315.7758
#> 51:  Total age_group3 cnt_males  424  24985  424  24985.0000
#> 52:  Total age_group4 cnt_males  195  11210  197  11324.9744
#> 53:  Total age_group5 cnt_males   84   5116   86   5237.8095
#> 54:  Total age_group6 cnt_males    7    269    5    192.1429
#> 55:   male      Total cnt_males 2296 136415 2295 136355.5858
#> 56:      m      Total cnt_males 2296 136415 2295 136355.5858
#> 57:   male age_group1 cnt_males 1015  60639 1015  60639.0000
#> 58:      m age_group1 cnt_males 1015  60639 1015  60639.0000
#> 59:   male       ag1a cnt_males 1015  60639 1015  60639.0000
#> 60:      m       ag1a cnt_males 1015  60639 1015  60639.0000
#> 61:   male age_group2 cnt_males  571  34196  573  34315.7758
#> 62:      m age_group2 cnt_males  571  34196  573  34315.7758
#> 63:   male       ag2a cnt_males  571  34196  573  34315.7758
#> 64:      m       ag2a cnt_males  571  34196  573  34315.7758
#> 65:   male age_group3 cnt_males  424  24985  424  24985.0000
#> 66:      m age_group3 cnt_males  424  24985  424  24985.0000
#> 67:   male age_group4 cnt_males  195  11210  197  11324.9744
#> 68:      m age_group4 cnt_males  195  11210  197  11324.9744
#> 69:   male age_group5 cnt_males   84   5116   86   5237.8095
#> 70:      m age_group5 cnt_males   84   5116   86   5237.8095
#> 71:   male age_group6 cnt_males    7    269    5    192.1429
#> 72:      m age_group6 cnt_males    7    269    5    192.1429
#> 73: female      Total cnt_males    0      0    0      0.0000
#> 74:      f      Total cnt_males    0      0    0      0.0000
#> 75: female age_group1 cnt_males    0      0    0      0.0000
#> 76:      f age_group1 cnt_males    0      0    0      0.0000
#> 77: female       ag1a cnt_males    0      0    0      0.0000
#> 78:      f       ag1a cnt_males    0      0    0      0.0000
#> 79: female age_group2 cnt_males    0      0    0      0.0000
#> 80:      f age_group2 cnt_males    0      0    0      0.0000
#> 81: female       ag2a cnt_males    0      0    0      0.0000
#> 82:      f       ag2a cnt_males    0      0    0      0.0000
#> 83: female age_group3 cnt_males    0      0    0      0.0000
#> 84:      f age_group3 cnt_males    0      0    0      0.0000
#> 85: female age_group4 cnt_males    0      0    0      0.0000
#> 86:      f age_group4 cnt_males    0      0    0      0.0000
#> 87: female age_group5 cnt_males    0      0    0      0.0000
#> 88:      f age_group5 cnt_males    0      0    0      0.0000
#> 89: female age_group6 cnt_males    0      0    0      0.0000
#> 90:      f age_group6 cnt_males    0      0    0      0.0000
#>        sex        age     vname  uwc     wc puwc         pwc

# utility measures for a count variable
tab$measures_cnts(v = "total", exclude_zeros = TRUE)
#> $overview
#>    noise cnt        pct
#> 1:    -2  12 0.26666667
#> 2:    -1   5 0.11111111
#> 3:     0  13 0.28888889
#> 4:     1  11 0.24444444
#> 5:     2   4 0.08888889
#> 
#> $measures
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 1.000 0.000 0.010
#>  5:    Q40 1.000 0.001 0.015
#>  6:   Mean 1.067 0.025 0.053
#>  7: Median 1.000 0.002 0.021
#>  8:    Q60 1.000 0.002 0.032
#>  9:    Q70 2.000 0.004 0.042
#> 10:    Q80 2.000 0.011 0.058
#> 11:    Q90 2.000 0.056 0.128
#> 12:    Q95 2.000 0.167 0.213
#> 13:    Q99 2.000 0.286 0.410
#> 14:    Max 2.000 0.286 0.410
#> 
#> $cumdistr_d1
#>    cat cnt       pct
#> 1:   0  13 0.2888889
#> 2:   1  29 0.6444444
#> 3:   2  45 1.0000000
#> 
#> $cumdistr_d2
#>            cat cnt       pct
#> 1:    [0,0.02]  38 0.8444444
#> 2: (0.02,0.05]  40 0.8888889
#> 3:  (0.05,0.1]  41 0.9111111
#> 4:   (0.1,0.2]  43 0.9555556
#> 5:   (0.2,0.3]  45 1.0000000
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $cumdistr_d3
#>            cat cnt       pct
#> 1:    [0,0.02]  20 0.4444444
#> 2: (0.02,0.05]  34 0.7555556
#> 3:  (0.05,0.1]  38 0.8444444
#> 4:   (0.1,0.2]  41 0.9111111
#> 5:   (0.2,0.3]  43 0.9555556
#> 6:   (0.3,0.4]  43 0.9555556
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $false_zero
#> [1] 0
#> 
#> $false_nonzero
#> [1] 0
#> 
#> $exclude_zeros
#> [1] TRUE
#> 

# modifications for perturbed count variables
tab$mod_cnts()
#>         sex        age row_nr pert       ckey  countvar
#>   1:  Total      Total     13   -2 0.05248231     total
#>   2:  Total age_group1     15    0 0.62359585     total
#>   3:  Total       ag1a     15    0 0.62359585     total
#>   4:  Total age_group2     14   -1 0.19354016     total
#>   5:  Total       ag2a     14   -1 0.19354016     total
#>  ---                                                   
#> 131:      f age_group4     -1    0 0.00000000 cnt_males
#> 132: female age_group5     -1    0 0.00000000 cnt_males
#> 133:      f age_group5     -1    0 0.00000000 cnt_males
#> 134: female age_group6     -1    0 0.00000000 cnt_males
#> 135:      f age_group6     -1    0 0.00000000 cnt_males

# display a summary about utility measures
tab$summary()
#> ┌──────────────────────────────────────────────┐
#> │Utility measures for perturbed count variables│
#> └──────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#>          countvar Min Q10 Q20 Q30 Q40   Mean Median Q60 Q70 Q80 Q90 Q95 Q99 Max
#> 1:          total  -2  -1  -1  -1   0  0.222      0   0   1   2   2 2.0   2   2
#> 2: cnt_highincome  -2  -1  -1  -1  -1 -0.067      0   0   1   1   1 1.8   3   3
#> 3:      cnt_males  -2  -1   0   0   0  0.333      0   0   0   2   2 2.0   2   2
#> 
#> ── Distance-based measures ─────────────────────────────────────────────────────
#> ✔ Variable: 'total'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 1.000 0.000 0.010
#>  5:    Q40 1.000 0.001 0.015
#>  6:   Mean 1.067 0.025 0.053
#>  7: Median 1.000 0.002 0.021
#>  8:    Q60 1.000 0.002 0.032
#>  9:    Q70 2.000 0.004 0.042
#> 10:    Q80 2.000 0.011 0.058
#> 11:    Q90 2.000 0.056 0.128
#> 12:    Q95 2.000 0.167 0.213
#> 13:    Q99 2.000 0.286 0.410
#> 14:    Max 2.000 0.286 0.410
#> 
#> ✔ Variable: 'cnt_males'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 1.000 0.000 0.010
#>  6:   Mean 1.222 0.036 0.076
#>  7: Median 2.000 0.004 0.042
#>  8:    Q60 2.000 0.004 0.042
#>  9:    Q70 2.000 0.010 0.071
#> 10:    Q80 2.000 0.021 0.101
#> 11:    Q90 2.000 0.129 0.229
#> 12:    Q95 2.000 0.286 0.410
#> 13:    Q99 2.000 0.286 0.410
#> 14:    Max 2.000 0.286 0.410
#> 
#> ✔ Variable: 'cnt_highincome'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 1.000 0.005 0.034
#>  4:    Q30 1.000 0.008 0.045
#>  5:    Q40 1.000 0.010 0.050
#>  6:   Mean 1.075 0.036 0.086
#>  7: Median 1.000 0.011 0.053
#>  8:    Q60 1.000 0.015 0.061
#>  9:    Q70 1.000 0.026 0.081
#> 10:    Q80 1.000 0.035 0.134
#> 11:    Q90 2.000 0.143 0.196
#> 12:    Q95 2.050 0.144 0.262
#> 13:    Q99 3.000 0.158 0.332
#> 14:    Max 3.000 0.158 0.332
#> 
#> ┌──────────────────────────────────────────────────┐
#> │Utility measures for perturbed numerical variables│
#> └──────────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#>      vname        Min        Q10        Q20        Q30      Q40     Mean
#> 1:  expend        Inf         NA         NA         NA       NA      NaN
#> 2:  income -67677.732 -29780.940 -25547.420 -24751.284  890.553 7149.647
#> 3: savings  -5896.513  -2057.477  -1777.527  -1311.495 -279.982  910.604
#> 4:   mixed        Inf         NA         NA         NA       NA      NaN
#>      Median      Q60       Q70       Q80       Q90       Q95       Q99
#> 1:       NA       NA        NA        NA        NA        NA        NA
#> 2: 8812.746 12113.19 22786.870 39639.752 64631.649 70135.854 70135.854
#> 3:  -41.909   706.19  1997.112  4559.661  6518.006  6603.441  6603.441
#> 4:       NA       NA        NA        NA        NA        NA        NA
#>          Max
#> 1:      -Inf
#> 2: 70135.854
#> 3:  6603.441
#> 4:      -Inf