Set parameters required to perturb numeric variables using a simple approach

ck_simpleparams() allows to define parameters for a simple perturbation approach based on a single magnitude parameter (m). The values of epsilon are used to "weight" parameter m in case type == "top_contr" is set in ck_params_nums().

ck_simpleparams(p, epsilon = 1)

Arguments

p
a percentage value used as magnitude for perturbation

epsilon
a numeric vector in descending order with all values >= 0 and <= 1 with the first element forced to equal 1. The length of this vector must correspond with the number top_k specified in ck_params_nums() when creating parameters for type == "top_contr" which is checked at runtime. This setting allows to use different flex-functions for the largest top_k contributors.

Value

an object suitable as input for ck_params_nums().

Details

details about the flex function can be found in Deliverable D4.2, Part I in SGA "Open Source tools for perturbative confidentiality methods"

See also

Examples

x <- ck_create_testdata()

# create some 0/1 variables that should be perturbed later
x[, cnt_females := ifelse(sex == "male", 0, 1)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              32          25.00000           0
#>    2:   2530      28       1              83          25.00000           1
#>    3:   6920     550       1              97          25.00000           0
#>    4:   7960     870       1              79          25.00000           0
#>    5:   9030      20       2              35          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              99          16.66667           1
#> 4577:   1420     987    1000              28          16.66667           0
#> 4578:   8900     684    1000              64          16.66667           0
#> 4579:   3880     294    1000             100          16.66667           1
#> 4580:   4830     911    1000              63          16.66667           0

x[, cnt_males := ifelse(sex == "male", 1, 0)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              32          25.00000           0
#>    2:   2530      28       1              83          25.00000           1
#>    3:   6920     550       1              97          25.00000           0
#>    4:   7960     870       1              79          25.00000           0
#>    5:   9030      20       2              35          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              99          16.66667           1
#> 4577:   1420     987    1000              28          16.66667           0
#> 4578:   8900     684    1000              64          16.66667           0
#> 4579:   3880     294    1000             100          16.66667           1
#> 4580:   4830     911    1000              63          16.66667           0
#>       cnt_males
#>    1:         1
#>    2:         0
#>    3:         1
#>    4:         1
#>    5:         1
#>   ---          
#> 4576:         0
#> 4577:         1
#> 4578:         1
#> 4579:         0
#> 4580:         1

x[, cnt_highincome := ifelse(income >= 9000, 1, 0)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              32          25.00000           0
#>    2:   2530      28       1              83          25.00000           1
#>    3:   6920     550       1              97          25.00000           0
#>    4:   7960     870       1              79          25.00000           0
#>    5:   9030      20       2              35          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              99          16.66667           1
#> 4577:   1420     987    1000              28          16.66667           0
#> 4578:   8900     684    1000              64          16.66667           0
#> 4579:   3880     294    1000             100          16.66667           1
#> 4580:   4830     911    1000              63          16.66667           0
#>       cnt_males cnt_highincome
#>    1:         1              0
#>    2:         0              0
#>    3:         1              0
#>    4:         1              0
#>    5:         1              1
#>   ---                         
#> 4576:         0              0
#> 4577:         1              0
#> 4578:         1              0
#> 4579:         0              0
#> 4580:         1              0

# a variable with positive and negative contributions
x[, mixed := sample(-10:10, nrow(x), replace = TRUE)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              32          25.00000           0
#>    2:   2530      28       1              83          25.00000           1
#>    3:   6920     550       1              97          25.00000           0
#>    4:   7960     870       1              79          25.00000           0
#>    5:   9030      20       2              35          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              99          16.66667           1
#> 4577:   1420     987    1000              28          16.66667           0
#> 4578:   8900     684    1000              64          16.66667           0
#> 4579:   3880     294    1000             100          16.66667           1
#> 4580:   4830     911    1000              63          16.66667           0
#>       cnt_males cnt_highincome mixed
#>    1:         1              0    -9
#>    2:         0              0     3
#>    3:         1              0    -3
#>    4:         1              0    -9
#>    5:         1              1   -10
#>   ---                               
#> 4576:         0              0     0
#> 4577:         1              0    -4
#> 4578:         1              0     0
#> 4579:         0              0     8
#> 4580:         1              0    -2

# create record keys
x$rkey <- ck_generate_rkeys(dat = x)

# define required inputs

# hierarchy with some bogus codes
d_sex <- hier_create(root = "Total", nodes = c("male", "female"))
d_sex <- hier_add(d_sex, root = "female", "f")
d_sex <- hier_add(d_sex, root = "male", "m")

d_age <- hier_create(root = "Total", nodes = paste0("age_group", 1:6))
d_age <- hier_add(d_age, root = "age_group1", "ag1a")
d_age <- hier_add(d_age, root = "age_group2", "ag2a")

# define the cell key object
countvars <- c("cnt_females", "cnt_males", "cnt_highincome")
numvars <- c("expend", "income", "savings", "mixed")
tab <- ck_setup(
  x = x,
  rkey = "rkey",
  dims = list(sex = d_sex, age = d_age),
  w = "sampling_weight",
  countvars = countvars,
  numvars = numvars)

# show some information about this table instance
tab$print() # identical with print(tab)

#> ── Table Information ───────────────────────────────────────────────────────────
#> ✔ 45 cells in 2 dimensions ('sex', 'age')
#> ✔ weights: yes
#> ── Tabulated / Perturbed countvars ─────────────────────────────────────────────
#> ☐ 'total'
#> ☐ 'cnt_females'
#> ☐ 'cnt_males'
#> ☐ 'cnt_highincome'
#> ── Tabulated / Perturbed numvars ───────────────────────────────────────────────
#> ☐ 'expend'
#> ☐ 'income'
#> ☐ 'savings'
#> ☐ 'mixed'

# information about the hierarchies
tab$hierarchy_info()

#> $sex
#>      code level is_leaf parent
#> 1:  Total     1   FALSE   <NA>
#> 2:   male     2   FALSE  Total
#> 3: female     2   FALSE  Total
#> 4:      f     3    TRUE female
#> 5:      m     3    TRUE   male
#> 
#> $age
#>          code level is_leaf     parent
#> 1:      Total     1   FALSE       <NA>
#> 2: age_group1     2   FALSE      Total
#> 3: age_group2     2   FALSE      Total
#> 4: age_group3     2    TRUE      Total
#> 5: age_group4     2    TRUE      Total
#> 6: age_group5     2    TRUE      Total
#> 7: age_group6     2    TRUE      Total
#> 8:       ag1a     3    TRUE age_group1
#> 9:       ag2a     3    TRUE age_group2
#> 

# which variables have been defined?
tab$allvars()

#> $cntvars
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"
#> 
#> $numvars
#> [1] "expend"  "income"  "savings" "mixed"  
#> 

# count variables
tab$cntvars()

#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"

# continuous variables
tab$numvars()

#> [1] "expend"  "income"  "savings" "mixed"  

# create perturbation parameters for "total" variable and
# write to yaml-file

# create a ptable using functionality from the ptable-pkg
f_yaml <- tempfile(fileext = ".yaml")
p_cnts1 <- ck_params_cnts(
  ptab = ptable::pt_ex_cnts(),
  path = f_yaml)

#> yaml configuration '/tmp/RtmpEtkipE/file42784e167786.yaml' successfully written.

# read parameters from yaml-file and set them for variable `"total"`
p_cnts1 <- ck_read_yaml(path = f_yaml)

tab$params_cnts_set(val = p_cnts1, v = "total")

#> --> setting perturbation parameters for variable 'total'

# create alternative perturbation parameters by specifying parameters
para2 <- ptable::pt_create_pParams(
  D = 8, V = 3, js = 2, table = "cnts")

#> Ignoring argument `step` for freqency tables

p_cnts2 <- ck_params_cnts(ptab = para2)

# use these ptable it for the remaining variables
tab$params_cnts_set(val = p_cnts2, v = countvars)

#> --> setting perturbation parameters for variable 'cnt_females'

#> --> setting perturbation parameters for variable 'cnt_males'

#> --> setting perturbation parameters for variable 'cnt_highincome'

# perturb a variable
tab$perturb(v = "total")

#> Count variable 'total' was perturbed.

# multiple variables can be perturbed as well
tab$perturb(v = c("cnt_males", "cnt_highincome"))

#> Count variable 'cnt_males' was perturbed.

#> Count variable 'cnt_highincome' was perturbed.

# return weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))

#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 273543 4581 273602.7255
#>  2:  Total age_group1     total 1969 117746 1970 117805.7999
#>  3:  Total       ag1a     total 1969 117746 1970 117805.7999
#>  4:  Total age_group2     total 1143  68205 1143  68205.0000
#>  5:  Total       ag2a     total 1143  68205 1143  68205.0000
#>  6:  Total age_group3     total  864  51813  865  51872.9688
#>  7:  Total age_group4     total  423  25062  422  25002.7518
#>  8:  Total age_group5     total  168   9965  168   9965.0000
#>  9:  Total age_group6     total   13    752   14    809.8462
#> 10: female      Total     total 2284 136736 2284 136736.0000
#> 11:      f      Total     total 2284 136736 2284 136736.0000
#> 12: female age_group1     total  954  57046  954  57046.0000
#> 13:      f age_group1     total  954  57046  954  57046.0000
#> 14: female       ag1a     total  954  57046  954  57046.0000
#> 15:      f       ag1a     total  954  57046  954  57046.0000
#> 16: female age_group2     total  572  34137  572  34137.0000
#> 17:      f age_group2     total  572  34137  572  34137.0000
#> 18: female       ag2a     total  572  34137  572  34137.0000
#> 19:      f       ag2a     total  572  34137  572  34137.0000
#> 20: female age_group3     total  440  26287  441  26346.7432
#> 21:      f age_group3     total  440  26287  441  26346.7432
#> 22: female age_group4     total  228  13765  228  13765.0000
#> 23:      f age_group4     total  228  13765  228  13765.0000
#> 24: female age_group5     total   84   5134   85   5195.1190
#> 25:      f age_group5     total   84   5134   85   5195.1190
#> 26: female age_group6     total    6    367    6    367.0000
#> 27:      f age_group6     total    6    367    6    367.0000
#> 28:   male      Total     total 2296 136807 2295 136747.4151
#> 29:      m      Total     total 2296 136807 2295 136747.4151
#> 30:   male age_group1     total 1015  60700 1015  60700.0000
#> 31:      m age_group1     total 1015  60700 1015  60700.0000
#> 32:   male       ag1a     total 1015  60700 1015  60700.0000
#> 33:      m       ag1a     total 1015  60700 1015  60700.0000
#> 34:   male age_group2     total  571  34068  572  34127.6637
#> 35:      m age_group2     total  571  34068  572  34127.6637
#> 36:   male       ag2a     total  571  34068  572  34127.6637
#> 37:      m       ag2a     total  571  34068  572  34127.6637
#> 38:   male age_group3     total  424  25526  422  25405.5943
#> 39:      m age_group3     total  424  25526  422  25405.5943
#> 40:   male age_group4     total  195  11297  196  11354.9333
#> 41:      m age_group4     total  195  11297  196  11354.9333
#> 42:   male age_group5     total   84   4831   85   4888.5119
#> 43:      m age_group5     total   84   4831   85   4888.5119
#> 44:   male age_group6     total    7    385    6    330.0000
#> 45:      m age_group6     total    7    385    6    330.0000
#> 46:  Total      Total cnt_males 2296 136807 2294 136687.8301
#> 47:  Total age_group1 cnt_males 1015  60700 1014  60640.1970
#> 48:  Total       ag1a cnt_males 1015  60700 1014  60640.1970
#> 49:  Total age_group2 cnt_males  571  34068  573  34187.3275
#> 50:  Total       ag2a cnt_males  571  34068  573  34187.3275
#> 51:  Total age_group3 cnt_males  424  25526  421  25345.3915
#> 52:  Total age_group4 cnt_males  195  11297  197  11412.8667
#> 53:  Total age_group5 cnt_males   84   4831   85   4888.5119
#> 54:  Total age_group6 cnt_males    7    385    6    330.0000
#> 55: female      Total cnt_males    0      0    0      0.0000
#> 56:      f      Total cnt_males    0      0    0      0.0000
#> 57: female age_group1 cnt_males    0      0    0      0.0000
#> 58:      f age_group1 cnt_males    0      0    0      0.0000
#> 59: female       ag1a cnt_males    0      0    0      0.0000
#> 60:      f       ag1a cnt_males    0      0    0      0.0000
#> 61: female age_group2 cnt_males    0      0    0      0.0000
#> 62:      f age_group2 cnt_males    0      0    0      0.0000
#> 63: female       ag2a cnt_males    0      0    0      0.0000
#> 64:      f       ag2a cnt_males    0      0    0      0.0000
#> 65: female age_group3 cnt_males    0      0    0      0.0000
#> 66:      f age_group3 cnt_males    0      0    0      0.0000
#> 67: female age_group4 cnt_males    0      0    0      0.0000
#> 68:      f age_group4 cnt_males    0      0    0      0.0000
#> 69: female age_group5 cnt_males    0      0    0      0.0000
#> 70:      f age_group5 cnt_males    0      0    0      0.0000
#> 71: female age_group6 cnt_males    0      0    0      0.0000
#> 72:      f age_group6 cnt_males    0      0    0      0.0000
#> 73:   male      Total cnt_males 2296 136807 2294 136687.8301
#> 74:      m      Total cnt_males 2296 136807 2294 136687.8301
#> 75:   male age_group1 cnt_males 1015  60700 1014  60640.1970
#> 76:      m age_group1 cnt_males 1015  60700 1014  60640.1970
#> 77:   male       ag1a cnt_males 1015  60700 1014  60640.1970
#> 78:      m       ag1a cnt_males 1015  60700 1014  60640.1970
#> 79:   male age_group2 cnt_males  571  34068  573  34187.3275
#> 80:      m age_group2 cnt_males  571  34068  573  34187.3275
#> 81:   male       ag2a cnt_males  571  34068  573  34187.3275
#> 82:      m       ag2a cnt_males  571  34068  573  34187.3275
#> 83:   male age_group3 cnt_males  424  25526  421  25345.3915
#> 84:      m age_group3 cnt_males  424  25526  421  25345.3915
#> 85:   male age_group4 cnt_males  195  11297  197  11412.8667
#> 86:      m age_group4 cnt_males  195  11297  197  11412.8667
#> 87:   male age_group5 cnt_males   84   4831   85   4888.5119
#> 88:      m age_group5 cnt_males   84   4831   85   4888.5119
#> 89:   male age_group6 cnt_males    7    385    6    330.0000
#> 90:      m age_group6 cnt_males    7    385    6    330.0000
#>        sex        age     vname  uwc     wc puwc         pwc

# numerical variables (positive variables using flex-function)
# we also write the config to a yaml file
f_yaml <- tempfile(fileext = ".yaml")

# create a ptable using functionality from the ptable-pkg
# a single ptable for all cells
ptab1 <- ptable::pt_ex_nums(parity = TRUE, separation = FALSE)

# a single ptab for all cells except for very small ones
ptab2 <- ptable::pt_ex_nums(parity = TRUE, separation = TRUE)

# different ptables for cells with even/odd number of contributors
# and very small cells
ptab3 <- ptable::pt_ex_nums(parity = FALSE, separation = TRUE)

p_nums1 <- ck_params_nums(
  ptab = ptab1,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.30, 0.03),
    epsilon = c(1, 0.5, 0.2),
    q = 3),
  mu_c = 2,
  same_key = FALSE,
  use_zero_rkeys = FALSE,
  path = f_yaml)

#> yaml configuration '/tmp/RtmpEtkipE/file4278396a4ebe.yaml' successfully written.

# we read the parameters from the yaml-file
p_nums1 <- ck_read_yaml(path = f_yaml)

# for variables with positive and negative values
p_nums2 <- ck_params_nums(
  ptab = ptab2,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.15, 0.02),
    epsilon = c(1, 0.4, 0.15),
    q = 3),
  mu_c = 2,
  same_key = FALSE)

# simple perturbation parameters (not using the flex-function approach)
p_nums3 <- ck_params_nums(
  ptab = ptab3,
  type = "mean",
  mult_params = ck_simpleparams(p = 0.25),
  mu_c = 2,
  same_key = FALSE)

# use `p_nums1` for all variables
tab$params_nums_set(p_nums1, c("savings", "income", "expend"))

#> --> setting perturbation parameters for variable 'savings'

#> --> setting perturbation parameters for variable 'income'

#> --> setting perturbation parameters for variable 'expend'

# use different parameters for variable `mixed`
tab$params_nums_set(p_nums2, v = "mixed")

#> --> setting perturbation parameters for variable 'mixed'

# identify sensitive cells to which extra protection (`mu_c`) is added.
tab$supp_p(v = "income", p = 85)

#> p%-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)

tab$supp_pq(v = "income", p = 85, q = 90)

#> pq-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)

tab$supp_nk(v = "income", n = 2, k = 90)

#> nk-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)

tab$supp_freq(v = "income", n = 14, weighted = FALSE)

#> freq-rule: 5 new sensitive cells (incl. duplicates) found (total: 5)

tab$supp_val(v = "income", n = 10000, weighted = TRUE)

#> val-rule: 0 new sensitive cells (incl. duplicates) found (total: 5)

# perturb variables
tab$perturb(v = c("income", "savings"))

#> Numeric variable 'income' was perturbed.

#> Numeric variable 'savings' was perturbed.

# extract results
tab$numtab("income", mean_before_sum = TRUE)

#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1371978292 1371988128
#>  2:  Total age_group1 income  9810547  590076088  590167288
#>  3:  Total       ag1a income  9810547  590076088  590167288
#>  4:  Total age_group2 income  5692119  335753112  335837105
#>  5:  Total       ag2a income  5692119  335753112  335837105
#>  6:  Total age_group3 income  4406946  266810871  266843552
#>  7:  Total age_group4 income  2133543  125977227  125931619
#>  8:  Total age_group5 income   848151   49933566   49965881
#>  9:  Total age_group6 income    61672    3427428    3234380
#> 10: female      Total income 11690929  700828427  700690875
#> 11:      f      Total income 11690929  700828427  700690875
#> 12: female age_group1 income  4933383  295813208  295809302
#> 13:      f age_group1 income  4933383  295813208  295809302
#> 14: female       ag1a income  4933383  295813208  295809302
#> 15:      f       ag1a income  4933383  295813208  295809302
#> 16: female age_group2 income  2880740  170387434  170454821
#> 17:      f age_group2 income  2880740  170387434  170454821
#> 18: female       ag2a income  2880740  170387434  170454821
#> 19:      f       ag2a income  2880740  170387434  170454821
#> 20: female age_group3 income  2238777  136921448  136835453
#> 21:      f age_group3 income  2238777  136921448  136835453
#> 22: female age_group4 income  1155033   68351160   68307434
#> 23:      f age_group4 income  1155033   68351160   68307434
#> 24: female age_group5 income   455017   27696208   27676755
#> 25:      f age_group5 income   455017   27696208   27676755
#> 26: female age_group6 income    27979    1658969    1755974
#> 27:      f age_group6 income    27979    1658969    1755974
#> 28:   male      Total income 11262049  671149865  671320148
#> 29:      m      Total income 11262049  671149865  671320148
#> 30:   male age_group1 income  4877164  294262880  294224244
#> 31:      m age_group1 income  4877164  294262880  294224244
#> 32:   male       ag1a income  4877164  294262880  294224244
#> 33:      m       ag1a income  4877164  294262880  294224244
#> 34:   male age_group2 income  2811379  165365678  165423062
#> 35:      m age_group2 income  2811379  165365678  165423062
#> 36:   male       ag2a income  2811379  165365678  165423062
#> 37:      m       ag2a income  2811379  165365678  165423062
#> 38:   male age_group3 income  2168169  129889423  129848237
#> 39:      m age_group3 income  2168169  129889423  129848237
#> 40:   male age_group4 income   978510   57626067   57593378
#> 41:      m age_group4 income   978510   57626067   57593378
#> 42:   male age_group5 income   393134   22237358   22234852
#> 43:      m age_group5 income   393134   22237358   22234852
#> 44:   male age_group6 income    33693    1768459    1735496
#> 45:      m age_group6 income    33693    1768459    1735496
#>        sex        age  vname      uws         ws        pws

tab$numtab("income", mean_before_sum = FALSE)

#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1371978292 1371983210
#>  2:  Total age_group1 income  9810547  590076088  590121686
#>  3:  Total       ag1a income  9810547  590076088  590121686
#>  4:  Total age_group2 income  5692119  335753112  335795106
#>  5:  Total       ag2a income  5692119  335753112  335795106
#>  6:  Total age_group3 income  4406946  266810871  266827211
#>  7:  Total age_group4 income  2133543  125977227  125954421
#>  8:  Total age_group5 income   848151   49933566   49949721
#>  9:  Total age_group6 income    61672    3427428    3329505
#> 10: female      Total income 11690929  700828427  700759648
#> 11:      f      Total income 11690929  700828427  700759648
#> 12: female age_group1 income  4933383  295813208  295811255
#> 13:      f age_group1 income  4933383  295813208  295811255
#> 14: female       ag1a income  4933383  295813208  295811255
#> 15:      f       ag1a income  4933383  295813208  295811255
#> 16: female age_group2 income  2880740  170387434  170421124
#> 17:      f age_group2 income  2880740  170387434  170421124
#> 18: female       ag2a income  2880740  170387434  170421124
#> 19:      f       ag2a income  2880740  170387434  170421124
#> 20: female age_group3 income  2238777  136921448  136878444
#> 21:      f age_group3 income  2238777  136921448  136878444
#> 22: female age_group4 income  1155033   68351160   68329294
#> 23:      f age_group4 income  1155033   68351160   68329294
#> 24: female age_group5 income   455017   27696208   27686480
#> 25:      f age_group5 income   455017   27696208   27686480
#> 26: female age_group6 income    27979    1658969    1706783
#> 27:      f age_group6 income    27979    1658969    1706783
#> 28:   male      Total income 11262049  671149865  671235001
#> 29:      m      Total income 11262049  671149865  671235001
#> 30:   male age_group1 income  4877164  294262880  294243561
#> 31:      m age_group1 income  4877164  294262880  294243561
#> 32:   male       ag1a income  4877164  294262880  294243561
#> 33:      m       ag1a income  4877164  294262880  294243561
#> 34:   male age_group2 income  2811379  165365678  165394367
#> 35:      m age_group2 income  2811379  165365678  165394367
#> 36:   male       ag2a income  2811379  165365678  165394367
#> 37:      m       ag2a income  2811379  165365678  165394367
#> 38:   male age_group3 income  2168169  129889423  129868828
#> 39:      m age_group3 income  2168169  129889423  129868828
#> 40:   male age_group4 income   978510   57626067   57609720
#> 41:      m age_group4 income   978510   57626067   57609720
#> 42:   male age_group5 income   393134   22237358   22236105
#> 43:      m age_group5 income   393134   22237358   22236105
#> 44:   male age_group6 income    33693    1768459    1751900
#> 45:      m age_group6 income    33693    1768459    1751900
#>        sex        age  vname      uws         ws        pws

tab$numtab("savings")

#>        sex        age   vname     uws        ws         pws
#>  1:  Total      Total savings 2273532 135506321 135504966.4
#>  2:  Total age_group1 savings  982386  58378293  58379974.7
#>  3:  Total       ag1a savings  982386  58378293  58379974.7
#>  4:  Total age_group2 savings  552336  33265287  33269666.4
#>  5:  Total       ag2a savings  552336  33265287  33269666.4
#>  6:  Total age_group3 savings  437101  26116184  26113668.6
#>  7:  Total age_group4 savings  214661  12683419  12680692.1
#>  8:  Total age_group5 savings   80451   4688687   4691144.5
#>  9:  Total age_group6 savings    6597    374451    366649.5
#> 10: female      Total savings 1113716  66497998  66496917.5
#> 11:      f      Total savings 1113716  66497998  66496917.5
#> 12: female age_group1 savings  464726  27495749  27495185.5
#> 13:      f age_group1 savings  464726  27495749  27495185.5
#> 14: female       ag1a savings  464726  27495749  27495185.5
#> 15:      f       ag1a savings  464726  27495749  27495185.5
#> 16: female age_group2 savings  271413  16500621  16503832.1
#> 17:      f age_group2 savings  271413  16500621  16503832.1
#> 18: female       ag2a savings  271413  16500621  16503832.1
#> 19:      f       ag2a savings  271413  16500621  16503832.1
#> 20: female age_group3 savings  222131  13164331  13157507.5
#> 21:      f age_group3 savings  222131  13164331  13157507.5
#> 22: female age_group4 savings  115241   6922207   6921293.3
#> 23:      f age_group4 savings  115241   6922207   6921293.3
#> 24: female age_group5 savings   37218   2241069   2241610.4
#> 25:      f age_group5 savings   37218   2241069   2241610.4
#> 26: female age_group6 savings    2987    174021    175797.1
#> 27:      f age_group6 savings    2987    174021    175797.1
#> 28:   male      Total savings 1159816  69008323  69009600.5
#> 29:      m      Total savings 1159816  69008323  69009600.5
#> 30:   male age_group1 savings  517660  30882544  30880545.7
#> 31:      m age_group1 savings  517660  30882544  30880545.7
#> 32:   male       ag1a savings  517660  30882544  30880545.7
#> 33:      m       ag1a savings  517660  30882544  30880545.7
#> 34:   male age_group2 savings  280923  16764666  16767292.3
#> 35:      m age_group2 savings  280923  16764666  16767292.3
#> 36:   male       ag2a savings  280923  16764666  16767292.3
#> 37:      m       ag2a savings  280923  16764666  16767292.3
#> 38:   male age_group3 savings  214970  12951853  12949654.0
#> 39:      m age_group3 savings  214970  12951853  12949654.0
#> 40:   male age_group4 savings   99420   5761212   5759895.0
#> 41:      m age_group4 savings   99420   5761212   5759895.0
#> 42:   male age_group5 savings   43233   2447618   2445757.9
#> 43:      m age_group5 savings   43233   2447618   2445757.9
#> 44:   male age_group6 savings    3610    200430    200039.2
#> 45:      m age_group6 savings    3610    200430    200039.2
#>        sex        age   vname     uws        ws         pws

# results can be resetted, too
tab$reset_cntvars(v = "cnt_males")

# we can then set other parameters and perturb again
tab$params_cnts_set(val = p_cnts1, v = "cnt_males")

#> --> setting perturbation parameters for variable 'cnt_males'

tab$perturb(v = "cnt_males")

#> Count variable 'cnt_males' was perturbed.

# write to a file "outtab.csv" (.csv is automatically added to the path)
if (FALSE) {
tab$freqtab(v = c("total", "cnt_males"), path = "outtab.csv")
}

# show results containing weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))

#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 273543 4581 273602.7255
#>  2:  Total age_group1     total 1969 117746 1970 117805.7999
#>  3:  Total       ag1a     total 1969 117746 1970 117805.7999
#>  4:  Total age_group2     total 1143  68205 1143  68205.0000
#>  5:  Total       ag2a     total 1143  68205 1143  68205.0000
#>  6:  Total age_group3     total  864  51813  865  51872.9688
#>  7:  Total age_group4     total  423  25062  422  25002.7518
#>  8:  Total age_group5     total  168   9965  168   9965.0000
#>  9:  Total age_group6     total   13    752   14    809.8462
#> 10: female      Total     total 2284 136736 2284 136736.0000
#> 11:      f      Total     total 2284 136736 2284 136736.0000
#> 12: female age_group1     total  954  57046  954  57046.0000
#> 13:      f age_group1     total  954  57046  954  57046.0000
#> 14: female       ag1a     total  954  57046  954  57046.0000
#> 15:      f       ag1a     total  954  57046  954  57046.0000
#> 16: female age_group2     total  572  34137  572  34137.0000
#> 17:      f age_group2     total  572  34137  572  34137.0000
#> 18: female       ag2a     total  572  34137  572  34137.0000
#> 19:      f       ag2a     total  572  34137  572  34137.0000
#> 20: female age_group3     total  440  26287  441  26346.7432
#> 21:      f age_group3     total  440  26287  441  26346.7432
#> 22: female age_group4     total  228  13765  228  13765.0000
#> 23:      f age_group4     total  228  13765  228  13765.0000
#> 24: female age_group5     total   84   5134   85   5195.1190
#> 25:      f age_group5     total   84   5134   85   5195.1190
#> 26: female age_group6     total    6    367    6    367.0000
#> 27:      f age_group6     total    6    367    6    367.0000
#> 28:   male      Total     total 2296 136807 2295 136747.4151
#> 29:      m      Total     total 2296 136807 2295 136747.4151
#> 30:   male age_group1     total 1015  60700 1015  60700.0000
#> 31:      m age_group1     total 1015  60700 1015  60700.0000
#> 32:   male       ag1a     total 1015  60700 1015  60700.0000
#> 33:      m       ag1a     total 1015  60700 1015  60700.0000
#> 34:   male age_group2     total  571  34068  572  34127.6637
#> 35:      m age_group2     total  571  34068  572  34127.6637
#> 36:   male       ag2a     total  571  34068  572  34127.6637
#> 37:      m       ag2a     total  571  34068  572  34127.6637
#> 38:   male age_group3     total  424  25526  422  25405.5943
#> 39:      m age_group3     total  424  25526  422  25405.5943
#> 40:   male age_group4     total  195  11297  196  11354.9333
#> 41:      m age_group4     total  195  11297  196  11354.9333
#> 42:   male age_group5     total   84   4831   85   4888.5119
#> 43:      m age_group5     total   84   4831   85   4888.5119
#> 44:   male age_group6     total    7    385    6    330.0000
#> 45:      m age_group6     total    7    385    6    330.0000
#> 46:  Total      Total cnt_males 2296 136807 2295 136747.4151
#> 47:  Total age_group1 cnt_males 1015  60700 1015  60700.0000
#> 48:  Total       ag1a cnt_males 1015  60700 1015  60700.0000
#> 49:  Total age_group2 cnt_males  571  34068  572  34127.6637
#> 50:  Total       ag2a cnt_males  571  34068  572  34127.6637
#> 51:  Total age_group3 cnt_males  424  25526  422  25405.5943
#> 52:  Total age_group4 cnt_males  195  11297  196  11354.9333
#> 53:  Total age_group5 cnt_males   84   4831   85   4888.5119
#> 54:  Total age_group6 cnt_males    7    385    6    330.0000
#> 55: female      Total cnt_males    0      0    0      0.0000
#> 56:      f      Total cnt_males    0      0    0      0.0000
#> 57: female age_group1 cnt_males    0      0    0      0.0000
#> 58:      f age_group1 cnt_males    0      0    0      0.0000
#> 59: female       ag1a cnt_males    0      0    0      0.0000
#> 60:      f       ag1a cnt_males    0      0    0      0.0000
#> 61: female age_group2 cnt_males    0      0    0      0.0000
#> 62:      f age_group2 cnt_males    0      0    0      0.0000
#> 63: female       ag2a cnt_males    0      0    0      0.0000
#> 64:      f       ag2a cnt_males    0      0    0      0.0000
#> 65: female age_group3 cnt_males    0      0    0      0.0000
#> 66:      f age_group3 cnt_males    0      0    0      0.0000
#> 67: female age_group4 cnt_males    0      0    0      0.0000
#> 68:      f age_group4 cnt_males    0      0    0      0.0000
#> 69: female age_group5 cnt_males    0      0    0      0.0000
#> 70:      f age_group5 cnt_males    0      0    0      0.0000
#> 71: female age_group6 cnt_males    0      0    0      0.0000
#> 72:      f age_group6 cnt_males    0      0    0      0.0000
#> 73:   male      Total cnt_males 2296 136807 2295 136747.4151
#> 74:      m      Total cnt_males 2296 136807 2295 136747.4151
#> 75:   male age_group1 cnt_males 1015  60700 1015  60700.0000
#> 76:      m age_group1 cnt_males 1015  60700 1015  60700.0000
#> 77:   male       ag1a cnt_males 1015  60700 1015  60700.0000
#> 78:      m       ag1a cnt_males 1015  60700 1015  60700.0000
#> 79:   male age_group2 cnt_males  571  34068  572  34127.6637
#> 80:      m age_group2 cnt_males  571  34068  572  34127.6637
#> 81:   male       ag2a cnt_males  571  34068  572  34127.6637
#> 82:      m       ag2a cnt_males  571  34068  572  34127.6637
#> 83:   male age_group3 cnt_males  424  25526  422  25405.5943
#> 84:      m age_group3 cnt_males  424  25526  422  25405.5943
#> 85:   male age_group4 cnt_males  195  11297  196  11354.9333
#> 86:      m age_group4 cnt_males  195  11297  196  11354.9333
#> 87:   male age_group5 cnt_males   84   4831   85   4888.5119
#> 88:      m age_group5 cnt_males   84   4831   85   4888.5119
#> 89:   male age_group6 cnt_males    7    385    6    330.0000
#> 90:      m age_group6 cnt_males    7    385    6    330.0000
#>        sex        age     vname  uwc     wc puwc         pwc

# utility measures for a count variable
tab$measures_cnts(v = "total", exclude_zeros = TRUE)

#> $overview
#>    noise cnt        pct
#> 1:    -1  17 0.37777778
#> 2:     0  21 0.46666667
#> 3:     1   5 0.11111111
#> 4:     2   2 0.04444444
#> 
#> $measures
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 0.000 0.000 0.000
#>  6:   Mean 0.578 0.010 0.025
#>  7: Median 1.000 0.000 0.010
#>  8:    Q60 1.000 0.001 0.019
#>  9:    Q70 1.000 0.002 0.023
#> 10:    Q80 1.000 0.005 0.038
#> 11:    Q90 1.000 0.012 0.054
#> 12:    Q95 1.000 0.064 0.120
#> 13:    Q99 2.000 0.143 0.196
#> 14:    Max 2.000 0.143 0.196
#> 
#> $cumdistr_d1
#>    cat cnt       pct
#> 1:   0  21 0.4666667
#> 2:   1  43 0.9555556
#> 3:   2  45 1.0000000
#> 
#> $cumdistr_d2
#>            cat cnt       pct
#> 1:    [0,0.02]  42 0.9333333
#> 2: (0.02,0.05]  42 0.9333333
#> 3:  (0.05,0.1]  43 0.9555556
#> 4:   (0.1,0.2]  45 1.0000000
#> 5:   (0.2,0.3]  45 1.0000000
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $cumdistr_d3
#>            cat cnt       pct
#> 1:    [0,0.02]  27 0.6000000
#> 2: (0.02,0.05]  38 0.8444444
#> 3:  (0.05,0.1]  42 0.9333333
#> 4:   (0.1,0.2]  45 1.0000000
#> 5:   (0.2,0.3]  45 1.0000000
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $false_zero
#> [1] 0
#> 
#> $false_nonzero
#> [1] 0
#> 
#> $exclude_zeros
#> [1] TRUE
#> 

# modifications for perturbed count variables
tab$mod_cnts()

#>        sex        age row_nr pert      ckey  countvar
#>   1: Total      Total     16    1 0.7605708     total
#>   2: Total age_group1     16    1 0.6890207     total
#>   3: Total       ag1a     16    1 0.6890207     total
#>   4: Total age_group2     15    0 0.4688946     total
#>   5: Total       ag2a     15    0 0.4688946     total
#>  ---                                                 
#> 131:     m age_group4     16    1 0.8328746 cnt_males
#> 132:  male age_group5     16    1 0.8069023 cnt_males
#> 133:     m age_group5     16    1 0.8069023 cnt_males
#> 134:  male age_group6     14   -1 0.2348190 cnt_males
#> 135:     m age_group6     14   -1 0.2348190 cnt_males

# display a summary about utility measures
tab$summary()

#> ┌──────────────────────────────────────────────┐
#> │Utility measures for perturbed count variables│
#> └──────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#>          countvar Min Q10  Q20 Q30 Q40   Mean Median Q60 Q70 Q80 Q90 Q95 Q99
#> 1:          total  -2  -1  0.0   0   0  0.178      0   0   1   1   1   1   1
#> 2: cnt_highincome  -3  -3 -3.0  -1  -1 -0.644      0   0   0   0   2   2   3
#> 3:      cnt_males  -2  -1 -0.2   0   0  0.000      0   0   0   1   1   1   1
#>    Max
#> 1:   1
#> 2:   3
#> 3:   1
#> 
#> ── Distance-based measures ─────────────────────────────────────────────────────
#> ✔ Variable: 'total'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 0.000 0.000 0.000
#>  6:   Mean 0.578 0.010 0.025
#>  7: Median 1.000 0.000 0.010
#>  8:    Q60 1.000 0.001 0.019
#>  9:    Q70 1.000 0.002 0.023
#> 10:    Q80 1.000 0.005 0.038
#> 11:    Q90 1.000 0.012 0.054
#> 12:    Q95 1.000 0.064 0.120
#> 13:    Q99 2.000 0.143 0.196
#> 14:    Max 2.000 0.143 0.196
#> 
#> ✔ Variable: 'cnt_males'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.200 0.000 0.002
#>  4:    Q30 1.000 0.000 0.010
#>  5:    Q40 1.000 0.002 0.021
#>  6:   Mean 0.889 0.019 0.043
#>  7: Median 1.000 0.002 0.021
#>  8:    Q60 1.000 0.005 0.036
#>  9:    Q70 1.000 0.005 0.049
#> 10:    Q80 1.000 0.011 0.053
#> 11:    Q90 1.400 0.064 0.111
#> 12:    Q95 2.000 0.143 0.196
#> 13:    Q99 2.000 0.143 0.196
#> 14:    Max 2.000 0.143 0.196
#> 
#> ✔ Variable: 'cnt_highincome'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.700 0.002 0.017
#>  5:    Q40 1.000 0.013 0.074
#>  6:   Mean 1.475 0.054 0.125
#>  7: Median 1.000 0.027 0.108
#>  8:    Q60 2.000 0.037 0.150
#>  9:    Q70 2.300 0.053 0.158
#> 10:    Q80 3.000 0.054 0.197
#> 11:    Q90 3.000 0.134 0.206
#> 12:    Q95 3.000 0.157 0.270
#> 13:    Q99 3.000 0.429 0.646
#> 14:    Max 3.000 0.429 0.646
#> 
#> ┌──────────────────────────────────────────────────┐
#> │Utility measures for perturbed numerical variables│
#> └──────────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────

#> Warning: no non-missing arguments to min; returning Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to min; returning Inf

#> Warning: no non-missing arguments to max; returning -Inf

#>      vname        Min        Q10        Q20        Q30        Q40     Mean
#> 1:  expend        Inf         NA         NA         NA         NA      NaN
#> 2:  income -97922.605 -34924.945 -20594.604 -18766.777 -12375.793 2798.593
#> 3: savings  -7801.543  -2388.876  -1998.291  -1347.078   -980.439 -193.242
#> 4:   mixed        Inf         NA         NA         NA         NA      NaN
#>       Median     Q60      Q70       Q80      Q90      Q95       Q99       Max
#> 1:        NA      NA       NA        NA       NA       NA        NA      -Inf
#> 2: -1953.064 9412.79 28689.31 33690.293 45598.29 47813.54 85136.048 85136.048
#> 3:  -563.536  541.40  1681.67  2626.272  3211.13  3211.13  4379.373  4379.373
#> 4:        NA      NA       NA        NA       NA       NA        NA      -Inf