Set parameters required to perturb numeric variables using a simple approach

ck_simpleparams() allows to define parameters for a simple perturbation approach based on a single magnitude parameter (m). The values of epsilon are used to "weight" parameter m in case type == "top_contr" is set in ck_params_nums().

ck_simpleparams(p, epsilon = 1)

Arguments

p

a percentage value used as magnitude for perturbation

epsilon

a numeric vector in descending order with all values >= 0 and <= 1 with the first element forced to equal 1. The length of this vector must correspond with the number top_k specified in ck_params_nums() when creating parameters for type == "top_contr" which is checked at runtime. This setting allows to use different flex-functions for the largest top_k contributors.

Value

an object suitable as input for ck_params_nums().

Details

details about the flex function can be found in Deliverable D4.2, Part I in SGA "Open Source tools for perturbative confidentiality methods"

See also

Examples

# \donttest{
x <- ck_create_testdata()

# create some 0/1 variables that should be perturbed later
x[, cnt_females := ifelse(sex == "male", 0, 1)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              33          25.00000           0
#>    2:   2530      28       1              74          25.00000           1
#>    3:   6920     550       1              72          25.00000           0
#>    4:   7960     870       1              65          25.00000           0
#>    5:   9030      20       2              71          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              82          16.66667           1
#> 4577:   1420     987    1000              77          16.66667           0
#> 4578:   8900     684    1000              43          16.66667           0
#> 4579:   3880     294    1000              41          16.66667           1
#> 4580:   4830     911    1000              47          16.66667           0
x[, cnt_males := ifelse(sex == "male", 1, 0)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              33          25.00000           0
#>    2:   2530      28       1              74          25.00000           1
#>    3:   6920     550       1              72          25.00000           0
#>    4:   7960     870       1              65          25.00000           0
#>    5:   9030      20       2              71          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              82          16.66667           1
#> 4577:   1420     987    1000              77          16.66667           0
#> 4578:   8900     684    1000              43          16.66667           0
#> 4579:   3880     294    1000              41          16.66667           1
#> 4580:   4830     911    1000              47          16.66667           0
#>       cnt_males
#>    1:         1
#>    2:         0
#>    3:         1
#>    4:         1
#>    5:         1
#>   ---          
#> 4576:         0
#> 4577:         1
#> 4578:         1
#> 4579:         0
#> 4580:         1
x[, cnt_highincome := ifelse(income >= 9000, 1, 0)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              33          25.00000           0
#>    2:   2530      28       1              74          25.00000           1
#>    3:   6920     550       1              72          25.00000           0
#>    4:   7960     870       1              65          25.00000           0
#>    5:   9030      20       2              71          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              82          16.66667           1
#> 4577:   1420     987    1000              77          16.66667           0
#> 4578:   8900     684    1000              43          16.66667           0
#> 4579:   3880     294    1000              41          16.66667           1
#> 4580:   4830     911    1000              47          16.66667           0
#>       cnt_males cnt_highincome
#>    1:         1              0
#>    2:         0              0
#>    3:         1              0
#>    4:         1              0
#>    5:         1              1
#>   ---                         
#> 4576:         0              0
#> 4577:         1              0
#> 4578:         1              0
#> 4579:         0              0
#> 4580:         1              0
# a variable with positive and negative contributions
x[, mixed := sample(-10:10, nrow(x), replace = TRUE)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              33          25.00000           0
#>    2:   2530      28       1              74          25.00000           1
#>    3:   6920     550       1              72          25.00000           0
#>    4:   7960     870       1              65          25.00000           0
#>    5:   9030      20       2              71          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              82          16.66667           1
#> 4577:   1420     987    1000              77          16.66667           0
#> 4578:   8900     684    1000              43          16.66667           0
#> 4579:   3880     294    1000              41          16.66667           1
#> 4580:   4830     911    1000              47          16.66667           0
#>       cnt_males cnt_highincome mixed
#>    1:         1              0     8
#>    2:         0              0    -7
#>    3:         1              0    -9
#>    4:         1              0     5
#>    5:         1              1   -10
#>   ---                               
#> 4576:         0              0    -5
#> 4577:         1              0     3
#> 4578:         1              0    -9
#> 4579:         0              0    -4
#> 4580:         1              0     6

# create record keys
x$rkey <- ck_generate_rkeys(dat = x)

# define required inputs

# hierarchy with some bogus codes
d_sex <- hier_create(root = "Total", nodes = c("male", "female"))
d_sex <- hier_add(d_sex, root = "female", "f")
d_sex <- hier_add(d_sex, root = "male", "m")

d_age <- hier_create(root = "Total", nodes = paste0("age_group", 1:6))
d_age <- hier_add(d_age, root = "age_group1", "ag1a")
d_age <- hier_add(d_age, root = "age_group2", "ag2a")

# define the cell key object
countvars <- c("cnt_females", "cnt_males", "cnt_highincome")
numvars <- c("expend", "income", "savings", "mixed")
tab <- ck_setup(
  x = x,
  rkey = "rkey",
  dims = list(sex = d_sex, age = d_age),
  w = "sampling_weight",
  countvars = countvars,
  numvars = numvars)

# show some information about this table instance
tab$print() # identical with print(tab)
#> ── Table Information ───────────────────────────────────────────────────────────
#> ✔ 45 cells in 2 dimensions ('sex', 'age')
#> ✔ weights: yes
#> ── Tabulated / Perturbed countvars ─────────────────────────────────────────────
#> ☐ 'total'
#> ☐ 'cnt_females'
#> ☐ 'cnt_males'
#> ☐ 'cnt_highincome'
#> ── Tabulated / Perturbed numvars ───────────────────────────────────────────────
#> ☐ 'expend'
#> ☐ 'income'
#> ☐ 'savings'
#> ☐ 'mixed'

# information about the hierarchies
tab$hierarchy_info()
#> $sex
#>      code level is_leaf parent
#> 1:  Total     1   FALSE  Total
#> 2:   male     2   FALSE  Total
#> 3:      m     3    TRUE   male
#> 4: female     2   FALSE  Total
#> 5:      f     3    TRUE female
#> 
#> $age
#>          code level is_leaf     parent
#> 1:      Total     1   FALSE      Total
#> 2: age_group1     2   FALSE      Total
#> 3:       ag1a     3    TRUE age_group1
#> 4: age_group2     2   FALSE      Total
#> 5:       ag2a     3    TRUE age_group2
#> 6: age_group3     2    TRUE      Total
#> 7: age_group4     2    TRUE      Total
#> 8: age_group5     2    TRUE      Total
#> 9: age_group6     2    TRUE      Total
#> 

# which variables have been defined?
tab$allvars()
#> $cntvars
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"
#> 
#> $numvars
#> [1] "expend"  "income"  "savings" "mixed"  
#> 

# count variables
tab$cntvars()
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"

# continuous variables
tab$numvars()
#> [1] "expend"  "income"  "savings" "mixed"  

# create perturbation parameters for "total" variable and
# write to yaml-file

# create a ptable using functionality from the ptable-pkg
f_yaml <- tempfile(fileext = ".yaml")
p_cnts1 <- ck_params_cnts(
  ptab = ptable::pt_ex_cnts(),
  path = f_yaml)
#> yaml configuration '/tmp/RtmpAtvqZz/file1a0a7a89ddef.yaml' successfully written.

# read parameters from yaml-file and set them for variable `"total"`
p_cnts1 <- ck_read_yaml(path = f_yaml)

tab$params_cnts_set(val = p_cnts1, v = "total")
#> --> setting perturbation parameters for variable 'total'

# create alternative perturbation parameters by specifying parameters
para2 <- ptable::create_cnt_ptable(
  D = 8, V = 3, js = 2, create = FALSE)

p_cnts2 <- ck_params_cnts(ptab = para2)

# use these ptable it for the remaining variables
tab$params_cnts_set(val = p_cnts2, v = countvars)
#> --> setting perturbation parameters for variable 'cnt_females'
#> --> setting perturbation parameters for variable 'cnt_males'
#> --> setting perturbation parameters for variable 'cnt_highincome'

# perturb a variable
tab$perturb(v = "total")
#> Count variable 'total' was perturbed.

# multiple variables can be perturbed as well
tab$perturb(v = c("cnt_males", "cnt_highincome"))
#> Count variable 'cnt_males' was perturbed.
#> Count variable 'cnt_highincome' was perturbed.

# return weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))
#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 274368 4581 274427.9057
#>  2:  Total age_group1     total 1969 118757 1970 118817.3134
#>  3:  Total       ag1a     total 1969 118757 1970 118817.3134
#>  4:  Total age_group2     total 1143  67801 1143  67801.0000
#>  5:  Total       ag2a     total 1143  67801 1143  67801.0000
#>  6:  Total age_group3     total  864  51113  865  51172.1586
#>  7:  Total age_group4     total  423  26106  422  26044.2837
#>  8:  Total age_group5     total  168   9741  169   9798.9821
#>  9:  Total age_group6     total   13    850   13    850.0000
#> 10:   male      Total     total 2296 137020 2298 137139.3554
#> 11:      m      Total     total 2296 137020 2298 137139.3554
#> 12:   male age_group1     total 1015  60912 1013  60791.9764
#> 13:      m age_group1     total 1015  60912 1013  60791.9764
#> 14:   male       ag1a     total 1015  60912 1013  60791.9764
#> 15:      m       ag1a     total 1015  60912 1013  60791.9764
#> 16:   male age_group2     total  571  34171  571  34171.0000
#> 17:      m age_group2     total  571  34171  571  34171.0000
#> 18:   male       ag2a     total  571  34171  571  34171.0000
#> 19:      m       ag2a     total  571  34171  571  34171.0000
#> 20:   male age_group3     total  424  24886  424  24886.0000
#> 21:      m age_group3     total  424  24886  424  24886.0000
#> 22:   male age_group4     total  195  11664  197  11783.6308
#> 23:      m age_group4     total  195  11664  197  11783.6308
#> 24:   male age_group5     total   84   4929   86   5046.3571
#> 25:      m age_group5     total   84   4929   86   5046.3571
#> 26:   male age_group6     total    7    458    5    327.1429
#> 27:      m age_group6     total    7    458    5    327.1429
#> 28: female      Total     total 2284 137348 2285 137408.1349
#> 29:      f      Total     total 2284 137348 2285 137408.1349
#> 30: female age_group1     total  954  57845  955  57905.6342
#> 31:      f age_group1     total  954  57845  955  57905.6342
#> 32: female       ag1a     total  954  57845  955  57905.6342
#> 33:      f       ag1a     total  954  57845  955  57905.6342
#> 34: female age_group2     total  572  33630  574  33747.5874
#> 35:      f age_group2     total  572  33630  574  33747.5874
#> 36: female       ag2a     total  572  33630  574  33747.5874
#> 37:      f       ag2a     total  572  33630  574  33747.5874
#> 38: female age_group3     total  440  26227  440  26227.0000
#> 39:      f age_group3     total  440  26227  440  26227.0000
#> 40: female age_group4     total  228  14442  227  14378.6579
#> 41:      f age_group4     total  228  14442  227  14378.6579
#> 42: female age_group5     total   84   4812   85   4869.2857
#> 43:      f age_group5     total   84   4812   85   4869.2857
#> 44: female age_group6     total    6    392    6    392.0000
#> 45:      f age_group6     total    6    392    6    392.0000
#> 46:  Total      Total cnt_males 2296 137020 2300 137258.7108
#> 47:  Total age_group1 cnt_males 1015  60912 1011  60671.9527
#> 48:  Total       ag1a cnt_males 1015  60912 1011  60671.9527
#> 49:  Total age_group2 cnt_males  571  34171  571  34171.0000
#> 50:  Total       ag2a cnt_males  571  34171  571  34171.0000
#> 51:  Total age_group3 cnt_males  424  24886  424  24886.0000
#> 52:  Total age_group4 cnt_males  195  11664  198  11843.4462
#> 53:  Total age_group5 cnt_males   84   4929   88   5163.7143
#> 54:  Total age_group6 cnt_males    7    458    4    261.7143
#> 55:   male      Total cnt_males 2296 137020 2300 137258.7108
#> 56:      m      Total cnt_males 2296 137020 2300 137258.7108
#> 57:   male age_group1 cnt_males 1015  60912 1011  60671.9527
#> 58:      m age_group1 cnt_males 1015  60912 1011  60671.9527
#> 59:   male       ag1a cnt_males 1015  60912 1011  60671.9527
#> 60:      m       ag1a cnt_males 1015  60912 1011  60671.9527
#> 61:   male age_group2 cnt_males  571  34171  571  34171.0000
#> 62:      m age_group2 cnt_males  571  34171  571  34171.0000
#> 63:   male       ag2a cnt_males  571  34171  571  34171.0000
#> 64:      m       ag2a cnt_males  571  34171  571  34171.0000
#> 65:   male age_group3 cnt_males  424  24886  424  24886.0000
#> 66:      m age_group3 cnt_males  424  24886  424  24886.0000
#> 67:   male age_group4 cnt_males  195  11664  198  11843.4462
#> 68:      m age_group4 cnt_males  195  11664  198  11843.4462
#> 69:   male age_group5 cnt_males   84   4929   88   5163.7143
#> 70:      m age_group5 cnt_males   84   4929   88   5163.7143
#> 71:   male age_group6 cnt_males    7    458    4    261.7143
#> 72:      m age_group6 cnt_males    7    458    4    261.7143
#> 73: female      Total cnt_males    0      0    0      0.0000
#> 74:      f      Total cnt_males    0      0    0      0.0000
#> 75: female age_group1 cnt_males    0      0    0      0.0000
#> 76:      f age_group1 cnt_males    0      0    0      0.0000
#> 77: female       ag1a cnt_males    0      0    0      0.0000
#> 78:      f       ag1a cnt_males    0      0    0      0.0000
#> 79: female age_group2 cnt_males    0      0    0      0.0000
#> 80:      f age_group2 cnt_males    0      0    0      0.0000
#> 81: female       ag2a cnt_males    0      0    0      0.0000
#> 82:      f       ag2a cnt_males    0      0    0      0.0000
#> 83: female age_group3 cnt_males    0      0    0      0.0000
#> 84:      f age_group3 cnt_males    0      0    0      0.0000
#> 85: female age_group4 cnt_males    0      0    0      0.0000
#> 86:      f age_group4 cnt_males    0      0    0      0.0000
#> 87: female age_group5 cnt_males    0      0    0      0.0000
#> 88:      f age_group5 cnt_males    0      0    0      0.0000
#> 89: female age_group6 cnt_males    0      0    0      0.0000
#> 90:      f age_group6 cnt_males    0      0    0      0.0000
#>        sex        age     vname  uwc     wc puwc         pwc

# numerical variables (positive variables using flex-function)
# we also write the config to a yaml file
f_yaml <- tempfile(fileext = ".yaml")

# create a ptable using functionality from the ptable-pkg
# a single ptable for all cells
ptab1 <- ptable::pt_ex_nums(parity = TRUE, separation = FALSE)

# a single ptab for all cells except for very small ones
ptab2 <- ptable::pt_ex_nums(parity = TRUE, separation = TRUE)

# different ptables for cells with even/odd number of contributors
# and very small cells
ptab3 <- ptable::pt_ex_nums(parity = FALSE, separation = TRUE)

p_nums1 <- ck_params_nums(
  ptab = ptab1,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.30, 0.03),
    epsilon = c(1, 0.5, 0.2),
    q = 3),
  mu_c = 2,
  same_key = FALSE,
  use_zero_rkeys = FALSE,
  path = f_yaml)
#> yaml configuration '/tmp/RtmpAtvqZz/file1a0a4261b281.yaml' successfully written.

# we read the parameters from the yaml-file
p_nums1 <- ck_read_yaml(path = f_yaml)

# for variables with positive and negative values
p_nums2 <- ck_params_nums(
  ptab = ptab2,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.15, 0.02),
    epsilon = c(1, 0.4, 0.15),
    q = 3),
  mu_c = 2,
  same_key = FALSE)

# simple perturbation parameters (not using the flex-function approach)
p_nums3 <- ck_params_nums(
  ptab = ptab3,
  type = "mean",
  mult_params = ck_simpleparams(p = 0.25),
  mu_c = 2,
  same_key = FALSE)

# use `p_nums1` for all variables
tab$params_nums_set(p_nums1, c("savings", "income", "expend"))
#> --> setting perturbation parameters for variable 'savings'
#> --> setting perturbation parameters for variable 'income'
#> --> setting perturbation parameters for variable 'expend'

# use different parameters for variable `mixed`
tab$params_nums_set(p_nums2, v = "mixed")
#> --> setting perturbation parameters for variable 'mixed'

# identify sensitive cells to which extra protection (`mu_c`) is added.
tab$supp_p(v = "income", p = 85)
#> computing contributing indices | rawdata <--> table; this might take a while
#> p%-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_pq(v = "income", p = 85, q = 90)
#> computing contributing indices | rawdata <--> table; this might take a while
#> pq-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_nk(v = "income", n = 2, k = 90)
#> computing contributing indices | rawdata <--> table; this might take a while
#> nk-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_freq(v = "income", n = 14, weighted = FALSE)
#> freq-rule: 5 new sensitive cells (incl. duplicates) found (total: 5)
tab$supp_val(v = "income", n = 10000, weighted = TRUE)
#> val-rule: 0 new sensitive cells (incl. duplicates) found (total: 5)
tab$supp_cells(
  v = "income",
  inp = data.frame(
    sex = c("female", "female"),
    "age" = c("age_group1", "age_group3")
  )
)
#> cell-rule: 2 new sensitive cells (incl. duplicates) found (total: 7)

# perturb variables
tab$perturb(v = c("income", "savings"))
#> Numeric variable 'income' was perturbed.
#> Numeric variable 'savings' was perturbed.

# extract results
tab$numtab("income", mean_before_sum = TRUE)
#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1377204600 1377281918
#>  2:  Total age_group1 income  9810547  595916327  595841773
#>  3:  Total       ag1a income  9810547  595916327  595841773
#>  4:  Total age_group2 income  5692119  333852364  333862485
#>  5:  Total       ag2a income  5692119  333852364  333862485
#>  6:  Total age_group3 income  4406946  261400981  261461176
#>  7:  Total age_group4 income  2133543  133299424  133377707
#>  8:  Total age_group5 income   848151   48901155   48936541
#>  9:  Total age_group6 income    61672    3834349    3817803
#> 10:   male      Total income 11262049  672506218  672268205
#> 11:      m      Total income 11262049  672506218  672268205
#> 12:   male age_group1 income  4877164  292813984  292707446
#> 13:      m age_group1 income  4877164  292813984  292707446
#> 14:   male       ag1a income  4877164  292813984  292707446
#> 15:      m       ag1a income  4877164  292813984  292707446
#> 16:   male age_group2 income  2811379  167328414  167377279
#> 17:      m age_group2 income  2811379  167328414  167377279
#> 18:   male       ag2a income  2811379  167328414  167377279
#> 19:      m       ag2a income  2811379  167328414  167377279
#> 20:   male age_group3 income  2168169  128838772  128774725
#> 21:      m age_group3 income  2168169  128838772  128774725
#> 22:   male age_group4 income   978510   59173457   59202775
#> 23:      m age_group4 income   978510   59173457   59202775
#> 24:   male age_group5 income   393134   22125902   22172021
#> 25:      m age_group5 income   393134   22125902   22172021
#> 26:   male age_group6 income    33693    2225689    2040622
#> 27:      m age_group6 income    33693    2225689    2040622
#> 28: female      Total income 11690929  704698382  704734606
#> 29:      f      Total income 11690929  704698382  704734606
#> 30: female age_group1 income  4933383  303102343  302973812
#> 31:      f age_group1 income  4933383  303102343  302973812
#> 32: female       ag1a income  4933383  303102343  302973812
#> 33:      f       ag1a income  4933383  303102343  302973812
#> 34: female age_group2 income  2880740  166523950  166537133
#> 35:      f age_group2 income  2880740  166523950  166537133
#> 36: female       ag2a income  2880740  166523950  166537133
#> 37:      f       ag2a income  2880740  166523950  166537133
#> 38: female age_group3 income  2238777  132562209  132675080
#> 39:      f age_group3 income  2238777  132562209  132675080
#> 40: female age_group4 income  1155033   74125967   74118018
#> 41:      f age_group4 income  1155033   74125967   74118018
#> 42: female age_group5 income   455017   26775253   26930557
#> 43:      f age_group5 income   455017   26775253   26930557
#> 44: female age_group6 income    27979    1608660    1652082
#> 45:      f age_group6 income    27979    1608660    1652082
#>        sex        age  vname      uws         ws        pws
tab$numtab("income", mean_before_sum = FALSE)
#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1377204600 1377243259
#>  2:  Total age_group1 income  9810547  595916327  595879049
#>  3:  Total       ag1a income  9810547  595916327  595879049
#>  4:  Total age_group2 income  5692119  333852364  333857424
#>  5:  Total       ag2a income  5692119  333852364  333857424
#>  6:  Total age_group3 income  4406946  261400981  261431077
#>  7:  Total age_group4 income  2133543  133299424  133338560
#>  8:  Total age_group5 income   848151   48901155   48918845
#>  9:  Total age_group6 income    61672    3834349    3826067
#> 10:   male      Total income 11262049  672506218  672387201
#> 11:      m      Total income 11262049  672506218  672387201
#> 12:   male age_group1 income  4877164  292813984  292760710
#> 13:      m age_group1 income  4877164  292813984  292760710
#> 14:   male       ag1a income  4877164  292813984  292760710
#> 15:      m       ag1a income  4877164  292813984  292760710
#> 16:   male age_group2 income  2811379  167328414  167352845
#> 17:      m age_group2 income  2811379  167328414  167352845
#> 18:   male       ag2a income  2811379  167328414  167352845
#> 19:      m       ag2a income  2811379  167328414  167352845
#> 20:   male age_group3 income  2168169  128838772  128806744
#> 21:      m age_group3 income  2168169  128838772  128806744
#> 22:   male age_group4 income   978510   59173457   59188114
#> 23:      m age_group4 income   978510   59173457   59188114
#> 24:   male age_group5 income   393134   22125902   22148950
#> 25:      m age_group5 income   393134   22125902   22148950
#> 26:   male age_group6 income    33693    2225689    2131148
#> 27:      m age_group6 income    33693    2225689    2131148
#> 28: female      Total income 11690929  704698382  704716494
#> 29:      f      Total income 11690929  704698382  704716494
#> 30: female age_group1 income  4933383  303102343  303038070
#> 31:      f age_group1 income  4933383  303102343  303038070
#> 32: female       ag1a income  4933383  303102343  303038070
#> 33:      f       ag1a income  4933383  303102343  303038070
#> 34: female age_group2 income  2880740  166523950  166530541
#> 35:      f age_group2 income  2880740  166523950  166530541
#> 36: female       ag2a income  2880740  166523950  166530541
#> 37:      f       ag2a income  2880740  166523950  166530541
#> 38: female age_group3 income  2238777  132562209  132618632
#> 39:      f age_group3 income  2238777  132562209  132618632
#> 40: female age_group4 income  1155033   74125967   74121993
#> 41:      f age_group4 income  1155033   74125967   74121993
#> 42: female age_group5 income   455017   26775253   26852793
#> 43:      f age_group5 income   455017   26775253   26852793
#> 44: female age_group6 income    27979    1608660    1630226
#> 45:      f age_group6 income    27979    1608660    1630226
#>        sex        age  vname      uws         ws        pws
tab$numtab("savings")
#>        sex        age   vname     uws        ws         pws
#>  1:  Total      Total savings 2273532 135344201 135347211.0
#>  2:  Total age_group1 savings  982386  58619942  58617978.2
#>  3:  Total       ag1a savings  982386  58619942  58617978.2
#>  4:  Total age_group2 savings  552336  32615044  32615954.0
#>  5:  Total       ag2a savings  552336  32615044  32615954.0
#>  6:  Total age_group3 savings  437101  25853809  25855849.5
#>  7:  Total age_group4 savings  214661  13328705  13331764.0
#>  8:  Total age_group5 savings   80451   4494320   4496919.9
#>  9:  Total age_group6 savings    6597    432381    428126.8
#> 10:   male      Total savings 1159816  68760099  68766653.2
#> 11:      m      Total savings 1159816  68760099  68766653.2
#> 12:   male age_group1 savings  517660  30890310  30889284.9
#> 13:      m age_group1 savings  517660  30890310  30889284.9
#> 14:   male       ag1a savings  517660  30890310  30889284.9
#> 15:      m       ag1a savings  517660  30890310  30889284.9
#> 16:   male age_group2 savings  280923  16711439  16714554.9
#> 17:      m age_group2 savings  280923  16711439  16714554.9
#> 18:   male       ag2a savings  280923  16711439  16714554.9
#> 19:      m       ag2a savings  280923  16711439  16714554.9
#> 20:   male age_group3 savings  214970  12618729  12609153.9
#> 21:      m age_group3 savings  214970  12618729  12609153.9
#> 22:   male age_group4 savings   99420   5933434   5934499.4
#> 23:      m age_group4 savings   99420   5933434   5934499.4
#> 24:   male age_group5 savings   43233   2358650   2362937.4
#> 25:      m age_group5 savings   43233   2358650   2362937.4
#> 26:   male age_group6 savings    3610    247537    244039.4
#> 27:      m age_group6 savings    3610    247537    244039.4
#> 28: female      Total savings 1113716  66584102  66586074.7
#> 29:      f      Total savings 1113716  66584102  66586074.7
#> 30: female age_group1 savings  464726  27729632  27726848.0
#> 31:      f age_group1 savings  464726  27729632  27726848.0
#> 32: female       ag1a savings  464726  27729632  27726848.0
#> 33:      f       ag1a savings  464726  27729632  27726848.0
#> 34: female age_group2 savings  271413  15903605  15905518.0
#> 35:      f age_group2 savings  271413  15903605  15905518.0
#> 36: female       ag2a savings  271413  15903605  15905518.0
#> 37:      f       ag2a savings  271413  15903605  15905518.0
#> 38: female age_group3 savings  222131  13235080  13243500.1
#> 39:      f age_group3 savings  222131  13235080  13243500.1
#> 40: female age_group4 savings  115241   7395271   7395050.4
#> 41:      f age_group4 savings  115241   7395271   7395050.4
#> 42: female age_group5 savings   37218   2135670   2140603.8
#> 43:      f age_group5 savings   37218   2135670   2140603.8
#> 44: female age_group6 savings    2987    184844    190540.4
#> 45:      f age_group6 savings    2987    184844    190540.4
#>        sex        age   vname     uws        ws         pws

# results can be resetted, too
tab$reset_cntvars(v = "cnt_males")

# we can then set other parameters and perturb again
tab$params_cnts_set(val = p_cnts1, v = "cnt_males")
#> --> setting perturbation parameters for variable 'cnt_males'

tab$perturb(v = "cnt_males")
#> Count variable 'cnt_males' was perturbed.

# write results to a .csv file
tab$freqtab(
  v = c("total", "cnt_males"),
  path = file.path(tempdir(), "outtab.csv")
)
#> File '/tmp/RtmpAtvqZz/outtab.csv' successfully written to disk.
#> NULL

# show results containing weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))
#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 274368 4581 274427.9057
#>  2:  Total age_group1     total 1969 118757 1970 118817.3134
#>  3:  Total       ag1a     total 1969 118757 1970 118817.3134
#>  4:  Total age_group2     total 1143  67801 1143  67801.0000
#>  5:  Total       ag2a     total 1143  67801 1143  67801.0000
#>  6:  Total age_group3     total  864  51113  865  51172.1586
#>  7:  Total age_group4     total  423  26106  422  26044.2837
#>  8:  Total age_group5     total  168   9741  169   9798.9821
#>  9:  Total age_group6     total   13    850   13    850.0000
#> 10:   male      Total     total 2296 137020 2298 137139.3554
#> 11:      m      Total     total 2296 137020 2298 137139.3554
#> 12:   male age_group1     total 1015  60912 1013  60791.9764
#> 13:      m age_group1     total 1015  60912 1013  60791.9764
#> 14:   male       ag1a     total 1015  60912 1013  60791.9764
#> 15:      m       ag1a     total 1015  60912 1013  60791.9764
#> 16:   male age_group2     total  571  34171  571  34171.0000
#> 17:      m age_group2     total  571  34171  571  34171.0000
#> 18:   male       ag2a     total  571  34171  571  34171.0000
#> 19:      m       ag2a     total  571  34171  571  34171.0000
#> 20:   male age_group3     total  424  24886  424  24886.0000
#> 21:      m age_group3     total  424  24886  424  24886.0000
#> 22:   male age_group4     total  195  11664  197  11783.6308
#> 23:      m age_group4     total  195  11664  197  11783.6308
#> 24:   male age_group5     total   84   4929   86   5046.3571
#> 25:      m age_group5     total   84   4929   86   5046.3571
#> 26:   male age_group6     total    7    458    5    327.1429
#> 27:      m age_group6     total    7    458    5    327.1429
#> 28: female      Total     total 2284 137348 2285 137408.1349
#> 29:      f      Total     total 2284 137348 2285 137408.1349
#> 30: female age_group1     total  954  57845  955  57905.6342
#> 31:      f age_group1     total  954  57845  955  57905.6342
#> 32: female       ag1a     total  954  57845  955  57905.6342
#> 33:      f       ag1a     total  954  57845  955  57905.6342
#> 34: female age_group2     total  572  33630  574  33747.5874
#> 35:      f age_group2     total  572  33630  574  33747.5874
#> 36: female       ag2a     total  572  33630  574  33747.5874
#> 37:      f       ag2a     total  572  33630  574  33747.5874
#> 38: female age_group3     total  440  26227  440  26227.0000
#> 39:      f age_group3     total  440  26227  440  26227.0000
#> 40: female age_group4     total  228  14442  227  14378.6579
#> 41:      f age_group4     total  228  14442  227  14378.6579
#> 42: female age_group5     total   84   4812   85   4869.2857
#> 43:      f age_group5     total   84   4812   85   4869.2857
#> 44: female age_group6     total    6    392    6    392.0000
#> 45:      f age_group6     total    6    392    6    392.0000
#> 46:  Total      Total cnt_males 2296 137020 2298 137139.3554
#> 47:  Total age_group1 cnt_males 1015  60912 1013  60791.9764
#> 48:  Total       ag1a cnt_males 1015  60912 1013  60791.9764
#> 49:  Total age_group2 cnt_males  571  34171  571  34171.0000
#> 50:  Total       ag2a cnt_males  571  34171  571  34171.0000
#> 51:  Total age_group3 cnt_males  424  24886  424  24886.0000
#> 52:  Total age_group4 cnt_males  195  11664  197  11783.6308
#> 53:  Total age_group5 cnt_males   84   4929   86   5046.3571
#> 54:  Total age_group6 cnt_males    7    458    5    327.1429
#> 55:   male      Total cnt_males 2296 137020 2298 137139.3554
#> 56:      m      Total cnt_males 2296 137020 2298 137139.3554
#> 57:   male age_group1 cnt_males 1015  60912 1013  60791.9764
#> 58:      m age_group1 cnt_males 1015  60912 1013  60791.9764
#> 59:   male       ag1a cnt_males 1015  60912 1013  60791.9764
#> 60:      m       ag1a cnt_males 1015  60912 1013  60791.9764
#> 61:   male age_group2 cnt_males  571  34171  571  34171.0000
#> 62:      m age_group2 cnt_males  571  34171  571  34171.0000
#> 63:   male       ag2a cnt_males  571  34171  571  34171.0000
#> 64:      m       ag2a cnt_males  571  34171  571  34171.0000
#> 65:   male age_group3 cnt_males  424  24886  424  24886.0000
#> 66:      m age_group3 cnt_males  424  24886  424  24886.0000
#> 67:   male age_group4 cnt_males  195  11664  197  11783.6308
#> 68:      m age_group4 cnt_males  195  11664  197  11783.6308
#> 69:   male age_group5 cnt_males   84   4929   86   5046.3571
#> 70:      m age_group5 cnt_males   84   4929   86   5046.3571
#> 71:   male age_group6 cnt_males    7    458    5    327.1429
#> 72:      m age_group6 cnt_males    7    458    5    327.1429
#> 73: female      Total cnt_males    0      0    0      0.0000
#> 74:      f      Total cnt_males    0      0    0      0.0000
#> 75: female age_group1 cnt_males    0      0    0      0.0000
#> 76:      f age_group1 cnt_males    0      0    0      0.0000
#> 77: female       ag1a cnt_males    0      0    0      0.0000
#> 78:      f       ag1a cnt_males    0      0    0      0.0000
#> 79: female age_group2 cnt_males    0      0    0      0.0000
#> 80:      f age_group2 cnt_males    0      0    0      0.0000
#> 81: female       ag2a cnt_males    0      0    0      0.0000
#> 82:      f       ag2a cnt_males    0      0    0      0.0000
#> 83: female age_group3 cnt_males    0      0    0      0.0000
#> 84:      f age_group3 cnt_males    0      0    0      0.0000
#> 85: female age_group4 cnt_males    0      0    0      0.0000
#> 86:      f age_group4 cnt_males    0      0    0      0.0000
#> 87: female age_group5 cnt_males    0      0    0      0.0000
#> 88:      f age_group5 cnt_males    0      0    0      0.0000
#> 89: female age_group6 cnt_males    0      0    0      0.0000
#> 90:      f age_group6 cnt_males    0      0    0      0.0000
#>        sex        age     vname  uwc     wc puwc         pwc

# utility measures for a count variable
tab$measures_cnts(v = "total", exclude_zeros = TRUE)
#> $overview
#>    noise cnt        pct
#> 1:    -2  10 0.22222222
#> 2:    -1  13 0.28888889
#> 3:     0  13 0.28888889
#> 4:     1   3 0.06666667
#> 5:     2   6 0.13333333
#> 
#> $measures
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 1.000 0.000 0.008
#>  5:    Q40 1.000 0.001 0.014
#>  6:   Mean 1.067 0.016 0.042
#>  7: Median 1.000 0.001 0.017
#>  8:    Q60 1.000 0.002 0.031
#>  9:    Q70 2.000 0.003 0.033
#> 10:    Q80 2.000 0.005 0.042
#> 11:    Q90 2.000 0.012 0.071
#> 12:    Q95 2.000 0.024 0.108
#> 13:    Q99 2.000 0.286 0.410
#> 14:    Max 2.000 0.286 0.410
#> 
#> $cumdistr_d1
#>    cat cnt       pct
#> 1:   0  13 0.2888889
#> 2:   1  29 0.6444444
#> 3:   2  45 1.0000000
#> 
#> $cumdistr_d2
#>            cat cnt       pct
#> 1:    [0,0.02]  41 0.9111111
#> 2: (0.02,0.05]  43 0.9555556
#> 3:  (0.05,0.1]  43 0.9555556
#> 4:   (0.1,0.2]  43 0.9555556
#> 5:   (0.2,0.3]  45 1.0000000
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $cumdistr_d3
#>            cat cnt       pct
#> 1:    [0,0.02]  23 0.5111111
#> 2: (0.02,0.05]  37 0.8222222
#> 3:  (0.05,0.1]  41 0.9111111
#> 4:   (0.1,0.2]  43 0.9555556
#> 5:   (0.2,0.3]  43 0.9555556
#> 6:   (0.3,0.4]  43 0.9555556
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $false_zero
#> [1] 0
#> 
#> $false_nonzero
#> [1] 0
#> 
#> $exclude_zeros
#> [1] TRUE
#> 

# modifications for perturbed count variables
tab$mod_cnts()
#>         sex        age row_nr pert      ckey  countvar
#>   1:  Total      Total     16    1 0.7741893     total
#>   2:  Total age_group1     16    1 0.8164757     total
#>   3:  Total       ag1a     16    1 0.8164757     total
#>   4:  Total age_group2     15    0 0.5457469     total
#>   5:  Total       ag2a     15    0 0.5457469     total
#>  ---                                                  
#> 131:      f age_group4     -1    0 0.0000000 cnt_males
#> 132: female age_group5     -1    0 0.0000000 cnt_males
#> 133:      f age_group5     -1    0 0.0000000 cnt_males
#> 134: female age_group6     -1    0 0.0000000 cnt_males
#> 135:      f age_group6     -1    0 0.0000000 cnt_males

# display a summary about utility measures
tab$summary()
#> ┌──────────────────────────────────────────────┐
#> │Utility measures for perturbed count variables│
#> └──────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#>          countvar Min  Q10  Q20 Q30 Q40   Mean Median Q60 Q70 Q80 Q90 Q95  Q99
#> 1:          total  -2 -2.0 -0.2   0   0  0.400      1   1   1 2.0   2   2 2.00
#> 2: cnt_highincome  -5 -2.6 -1.2  -1  -1 -0.222      0   0   1 1.0   2   2 2.56
#> 3:      cnt_males  -2 -2.0 -0.4   0   0  0.000      0   0   0 0.4   2   2 2.00
#>    Max
#> 1:   2
#> 2:   3
#> 3:   2
#> 
#> ── Distance-based measures ─────────────────────────────────────────────────────
#> ✔ Variable: 'total'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 1.000 0.000 0.008
#>  5:    Q40 1.000 0.001 0.014
#>  6:   Mean 1.067 0.016 0.042
#>  7: Median 1.000 0.001 0.017
#>  8:    Q60 1.000 0.002 0.031
#>  9:    Q70 2.000 0.003 0.033
#> 10:    Q80 2.000 0.005 0.042
#> 11:    Q90 2.000 0.012 0.071
#> 12:    Q95 2.000 0.024 0.108
#> 13:    Q99 2.000 0.286 0.410
#> 14:    Max 2.000 0.286 0.410
#> 
#> ✔ Variable: 'cnt_males'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 2.000 0.001 0.021
#>  6:   Mean 1.333 0.036 0.075
#>  7: Median 2.000 0.002 0.031
#>  8:    Q60 2.000 0.002 0.031
#>  9:    Q70 2.000 0.010 0.071
#> 10:    Q80 2.000 0.021 0.101
#> 11:    Q90 2.000 0.129 0.229
#> 12:    Q95 2.000 0.286 0.410
#> 13:    Q99 2.000 0.286 0.410
#> 14:    Max 2.000 0.286 0.410
#> 
#> ✔ Variable: 'cnt_highincome'
#> 
#>       what   d1    d2    d3
#>  1:    Min 0.00 0.000 0.000
#>  2:    Q10 0.90 0.004 0.030
#>  3:    Q20 1.00 0.005 0.043
#>  4:    Q30 1.00 0.011 0.053
#>  5:    Q40 1.00 0.013 0.062
#>  6:   Mean 1.50 0.065 0.139
#>  7: Median 1.00 0.015 0.067
#>  8:    Q60 1.40 0.018 0.078
#>  9:    Q70 2.00 0.032 0.124
#> 10:    Q80 2.00 0.105 0.229
#> 11:    Q90 3.00 0.286 0.409
#> 12:    Q95 3.00 0.286 0.410
#> 13:    Q99 4.22 0.329 0.612
#> 14:    Max 5.00 0.357 0.742
#> 
#> ┌──────────────────────────────────────────────────┐
#> │Utility measures for perturbed numerical variables│
#> └──────────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#>      vname        Min        Q10        Q20        Q30       Q40      Mean
#> 1:  expend        Inf         NA         NA         NA        NA       NaN
#> 2:  income -119017.17 -64272.516 -53273.663 -36227.910 -3974.389 -8214.718
#> 3: savings   -9575.13  -3212.173  -2127.832  -1025.129   910.047  1077.773
#> 4:   mixed        Inf         NA         NA         NA        NA       NaN
#>      Median       Q60       Q70       Q80       Q90       Q95       Q99
#> 1:       NA        NA        NA        NA        NA        NA        NA
#> 2: 6591.372 15870.270 21566.406 24430.766 38944.793 56423.374 77539.756
#> 3: 1913.032  1999.824  3104.482  4287.357  5696.438  6554.192  8420.095
#> 4:       NA        NA        NA        NA        NA        NA        NA
#>          Max
#> 1:      -Inf
#> 2: 77539.756
#> 3:  8420.095
#> 4:      -Inf
# }