Set parameters required to perturb numeric variables using a simple approach

ck_simpleparams() allows to define parameters for a simple perturbation approach based on a single magnitude parameter (m). The values of epsilon are used to "weight" parameter m in case type == "top_contr" is set in ck_params_nums().

ck_simpleparams(p, epsilon = 1)

Arguments

p
a percentage value used as magnitude for perturbation

epsilon
a numeric vector in descending order with all values >= 0 and <= 1 with the first element forced to equal 1. The length of this vector must correspond with the number top_k specified in ck_params_nums() when creating parameters for type == "top_contr" which is checked at runtime. This setting allows to use different flex-functions for the largest top_k contributors.

Value

an object suitable as input for ck_params_nums().

Details

details about the flex function can be found in Deliverable D4.2, Part I in SGA "Open Source tools for perturbative confidentiality methods"

See also

Examples

x <- ck_create_testdata()

# create some 0/1 variables that should be perturbed later
x[, cnt_females := ifelse(sex == "male", 0, 1)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              37          25.00000           0
#>    2:   2530      28       1              79          25.00000           1
#>    3:   6920     550       1              98          25.00000           0
#>    4:   7960     870       1              24          25.00000           0
#>    5:   9030      20       2              50          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              39          16.66667           1
#> 4577:   1420     987    1000              40          16.66667           0
#> 4578:   8900     684    1000              95          16.66667           0
#> 4579:   3880     294    1000              90          16.66667           1
#> 4580:   4830     911    1000              72          16.66667           0

x[, cnt_males := ifelse(sex == "male", 1, 0)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              37          25.00000           0
#>    2:   2530      28       1              79          25.00000           1
#>    3:   6920     550       1              98          25.00000           0
#>    4:   7960     870       1              24          25.00000           0
#>    5:   9030      20       2              50          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              39          16.66667           1
#> 4577:   1420     987    1000              40          16.66667           0
#> 4578:   8900     684    1000              95          16.66667           0
#> 4579:   3880     294    1000              90          16.66667           1
#> 4580:   4830     911    1000              72          16.66667           0
#>       cnt_males
#>    1:         1
#>    2:         0
#>    3:         1
#>    4:         1
#>    5:         1
#>   ---          
#> 4576:         0
#> 4577:         1
#> 4578:         1
#> 4579:         0
#> 4580:         1

x[, cnt_highincome := ifelse(income >= 9000, 1, 0)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              37          25.00000           0
#>    2:   2530      28       1              79          25.00000           1
#>    3:   6920     550       1              98          25.00000           0
#>    4:   7960     870       1              24          25.00000           0
#>    5:   9030      20       2              50          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              39          16.66667           1
#> 4577:   1420     987    1000              40          16.66667           0
#> 4578:   8900     684    1000              95          16.66667           0
#> 4579:   3880     294    1000              90          16.66667           1
#> 4580:   4830     911    1000              72          16.66667           0
#>       cnt_males cnt_highincome
#>    1:         1              0
#>    2:         0              0
#>    3:         1              0
#>    4:         1              0
#>    5:         1              1
#>   ---                         
#> 4576:         0              0
#> 4577:         1              0
#> 4578:         1              0
#> 4579:         0              0
#> 4580:         1              0

# a variable with positive and negative contributions
x[, mixed := sample(-10:10, nrow(x), replace = TRUE)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              37          25.00000           0
#>    2:   2530      28       1              79          25.00000           1
#>    3:   6920     550       1              98          25.00000           0
#>    4:   7960     870       1              24          25.00000           0
#>    5:   9030      20       2              50          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              39          16.66667           1
#> 4577:   1420     987    1000              40          16.66667           0
#> 4578:   8900     684    1000              95          16.66667           0
#> 4579:   3880     294    1000              90          16.66667           1
#> 4580:   4830     911    1000              72          16.66667           0
#>       cnt_males cnt_highincome mixed
#>    1:         1              0    -3
#>    2:         0              0    10
#>    3:         1              0    -3
#>    4:         1              0    -6
#>    5:         1              1    -8
#>   ---                               
#> 4576:         0              0    -1
#> 4577:         1              0   -10
#> 4578:         1              0     0
#> 4579:         0              0     5
#> 4580:         1              0   -10

# create record keys
x$rkey <- ck_generate_rkeys(dat = x)

# define required inputs

# hierarchy with some bogus codes
d_sex <- hier_create(root = "Total", nodes = c("male", "female"))
d_sex <- hier_add(d_sex, root = "female", "f")
d_sex <- hier_add(d_sex, root = "male", "m")

d_age <- hier_create(root = "Total", nodes = paste0("age_group", 1:6))
d_age <- hier_add(d_age, root = "age_group1", "ag1a")
d_age <- hier_add(d_age, root = "age_group2", "ag2a")

# define the cell key object
countvars <- c("cnt_females", "cnt_males", "cnt_highincome")
numvars <- c("expend", "income", "savings", "mixed")
tab <- ck_setup(
  x = x,
  rkey = "rkey",
  dims = list(sex = d_sex, age = d_age),
  w = "sampling_weight",
  countvars = countvars,
  numvars = numvars)

# show some information about this table instance
tab$print() # identical with print(tab)

#> ── Table Information ───────────────────────────────────────────────────────────
#> ✔ 45 cells in 2 dimensions ('sex', 'age')
#> ✔ weights: yes
#> ── Tabulated / Perturbed countvars ─────────────────────────────────────────────
#> ☐ 'total'
#> ☐ 'cnt_females'
#> ☐ 'cnt_males'
#> ☐ 'cnt_highincome'
#> ── Tabulated / Perturbed numvars ───────────────────────────────────────────────
#> ☐ 'expend'
#> ☐ 'income'
#> ☐ 'savings'
#> ☐ 'mixed'

# information about the hierarchies
tab$hierarchy_info()

#> $sex
#>      code level is_leaf parent
#> 1:  Total     1   FALSE   <NA>
#> 2:   male     2   FALSE  Total
#> 3: female     2   FALSE  Total
#> 4:      f     3    TRUE female
#> 5:      m     3    TRUE   male
#> 
#> $age
#>          code level is_leaf     parent
#> 1:      Total     1   FALSE       <NA>
#> 2: age_group1     2   FALSE      Total
#> 3: age_group2     2   FALSE      Total
#> 4: age_group3     2    TRUE      Total
#> 5: age_group4     2    TRUE      Total
#> 6: age_group5     2    TRUE      Total
#> 7: age_group6     2    TRUE      Total
#> 8:       ag1a     3    TRUE age_group1
#> 9:       ag2a     3    TRUE age_group2
#> 

# which variables have been defined?
tab$allvars()

#> $cntvars
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"
#> 
#> $numvars
#> [1] "expend"  "income"  "savings" "mixed"  
#> 

# count variables
tab$cntvars()

#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"

# continuous variables
tab$numvars()

#> [1] "expend"  "income"  "savings" "mixed"  

# create perturbation parameters for "total" variable and
# write to yaml-file

# create a ptable using functionality from the ptable-pkg
f_yaml <- tempfile(fileext = ".yaml")
p_cnts1 <- ck_params_cnts(
  ptab = ptable::pt_ex_cnts(),
  path = f_yaml)

#> yaml configuration '/tmp/Rtmp09UBN4/file3bd4534eb26.yaml' successfully written.

# read parameters from yaml-file and set them for variable `"total"`
p_cnts1 <- ck_read_yaml(path = f_yaml)

tab$params_cnts_set(val = p_cnts1, v = "total")

#> --> setting perturbation parameters for variable 'total'

# create alternative perturbation parameters by specifying parameters
para2 <- ptable::pt_create_pParams(
  D = 8, V = 3, js = 2, table = "cnts")

#> Ignoring argument `step` for freqency tables

p_cnts2 <- ck_params_cnts(ptab = para2)

# use these ptable it for the remaining variables
tab$params_cnts_set(val = p_cnts2, v = countvars)

#> --> setting perturbation parameters for variable 'cnt_females'

#> --> setting perturbation parameters for variable 'cnt_males'

#> --> setting perturbation parameters for variable 'cnt_highincome'

# perturb a variable
tab$perturb(v = "total")

#> Count variable 'total' was perturbed.

# multiple variables can be perturbed as well
tab$perturb(v = c("cnt_males", "cnt_highincome"))

#> Count variable 'cnt_males' was perturbed.

#> Count variable 'cnt_highincome' was perturbed.

# return weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))

#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 273086 4581 273145.6258
#>  2:  Total age_group1     total 1969 118141 1968 118080.9995
#>  3:  Total       ag1a     total 1969 118141 1968 118080.9995
#>  4:  Total age_group2     total 1143  68719 1143  68719.0000
#>  5:  Total       ag2a     total 1143  68719 1143  68719.0000
#>  6:  Total age_group3     total  864  50212  862  50095.7685
#>  7:  Total age_group4     total  423  25768  424  25828.9173
#>  8:  Total age_group5     total  168   9547  168   9547.0000
#>  9:  Total age_group6     total   13    699   13    699.0000
#> 10: female      Total     total 2284 135726 2284 135726.0000
#> 11:      f      Total     total 2284 135726 2284 135726.0000
#> 12: female age_group1     total  954  57024  955  57083.7736
#> 13:      f age_group1     total  954  57024  955  57083.7736
#> 14: female       ag1a     total  954  57024  955  57083.7736
#> 15:      f       ag1a     total  954  57024  955  57083.7736
#> 16: female age_group2     total  572  33785  574  33903.1294
#> 17:      f age_group2     total  572  33785  574  33903.1294
#> 18: female       ag2a     total  572  33785  574  33903.1294
#> 19:      f       ag2a     total  572  33785  574  33903.1294
#> 20: female age_group3     total  440  25885  440  25885.0000
#> 21:      f age_group3     total  440  25885  440  25885.0000
#> 22: female age_group4     total  228  13755  228  13755.0000
#> 23:      f age_group4     total  228  13755  228  13755.0000
#> 24: female age_group5     total   84   4978   85   5037.2619
#> 25:      f age_group5     total   84   4978   85   5037.2619
#> 26: female age_group6     total    6    299    8    398.6667
#> 27:      f age_group6     total    6    299    8    398.6667
#> 28:   male      Total     total 2296 137360 2295 137300.1742
#> 29:      m      Total     total 2296 137360 2295 137300.1742
#> 30:   male age_group1     total 1015  61117 1015  61117.0000
#> 31:      m age_group1     total 1015  61117 1015  61117.0000
#> 32:   male       ag1a     total 1015  61117 1015  61117.0000
#> 33:      m       ag1a     total 1015  61117 1015  61117.0000
#> 34:   male age_group2     total  571  34934  571  34934.0000
#> 35:      m age_group2     total  571  34934  571  34934.0000
#> 36:   male       ag2a     total  571  34934  571  34934.0000
#> 37:      m       ag2a     total  571  34934  571  34934.0000
#> 38:   male age_group3     total  424  24327  424  24327.0000
#> 39:      m age_group3     total  424  24327  424  24327.0000
#> 40:   male age_group4     total  195  12013  194  11951.3949
#> 41:      m age_group4     total  195  12013  194  11951.3949
#> 42:   male age_group5     total   84   4569   85   4623.3929
#> 43:      m age_group5     total   84   4569   85   4623.3929
#> 44:   male age_group6     total    7    400    7    400.0000
#> 45:      m age_group6     total    7    400    7    400.0000
#> 46:  Total      Total cnt_males 2296 137360 2294 137240.3484
#> 47:  Total age_group1 cnt_males 1015  61117 1015  61117.0000
#> 48:  Total       ag1a cnt_males 1015  61117 1015  61117.0000
#> 49:  Total age_group2 cnt_males  571  34934  572  34995.1804
#> 50:  Total       ag2a cnt_males  571  34934  572  34995.1804
#> 51:  Total age_group3 cnt_males  424  24327  424  24327.0000
#> 52:  Total age_group4 cnt_males  195  12013  194  11951.3949
#> 53:  Total age_group5 cnt_males   84   4569   85   4623.3929
#> 54:  Total age_group6 cnt_males    7    400    7    400.0000
#> 55: female      Total cnt_males    0      0    0      0.0000
#> 56:      f      Total cnt_males    0      0    0      0.0000
#> 57: female age_group1 cnt_males    0      0    0      0.0000
#> 58:      f age_group1 cnt_males    0      0    0      0.0000
#> 59: female       ag1a cnt_males    0      0    0      0.0000
#> 60:      f       ag1a cnt_males    0      0    0      0.0000
#> 61: female age_group2 cnt_males    0      0    0      0.0000
#> 62:      f age_group2 cnt_males    0      0    0      0.0000
#> 63: female       ag2a cnt_males    0      0    0      0.0000
#> 64:      f       ag2a cnt_males    0      0    0      0.0000
#> 65: female age_group3 cnt_males    0      0    0      0.0000
#> 66:      f age_group3 cnt_males    0      0    0      0.0000
#> 67: female age_group4 cnt_males    0      0    0      0.0000
#> 68:      f age_group4 cnt_males    0      0    0      0.0000
#> 69: female age_group5 cnt_males    0      0    0      0.0000
#> 70:      f age_group5 cnt_males    0      0    0      0.0000
#> 71: female age_group6 cnt_males    0      0    0      0.0000
#> 72:      f age_group6 cnt_males    0      0    0      0.0000
#> 73:   male      Total cnt_males 2296 137360 2294 137240.3484
#> 74:      m      Total cnt_males 2296 137360 2294 137240.3484
#> 75:   male age_group1 cnt_males 1015  61117 1015  61117.0000
#> 76:      m age_group1 cnt_males 1015  61117 1015  61117.0000
#> 77:   male       ag1a cnt_males 1015  61117 1015  61117.0000
#> 78:      m       ag1a cnt_males 1015  61117 1015  61117.0000
#> 79:   male age_group2 cnt_males  571  34934  572  34995.1804
#> 80:      m age_group2 cnt_males  571  34934  572  34995.1804
#> 81:   male       ag2a cnt_males  571  34934  572  34995.1804
#> 82:      m       ag2a cnt_males  571  34934  572  34995.1804
#> 83:   male age_group3 cnt_males  424  24327  424  24327.0000
#> 84:      m age_group3 cnt_males  424  24327  424  24327.0000
#> 85:   male age_group4 cnt_males  195  12013  194  11951.3949
#> 86:      m age_group4 cnt_males  195  12013  194  11951.3949
#> 87:   male age_group5 cnt_males   84   4569   85   4623.3929
#> 88:      m age_group5 cnt_males   84   4569   85   4623.3929
#> 89:   male age_group6 cnt_males    7    400    7    400.0000
#> 90:      m age_group6 cnt_males    7    400    7    400.0000
#>        sex        age     vname  uwc     wc puwc         pwc

# numerical variables (positive variables using flex-function)
# we also write the config to a yaml file
f_yaml <- tempfile(fileext = ".yaml")

# create a ptable using functionality from the ptable-pkg
# a single ptable for all cells
ptab1 <- ptable::pt_ex_nums(parity = TRUE, separation = FALSE)

# a single ptab for all cells except for very small ones
ptab2 <- ptable::pt_ex_nums(parity = TRUE, separation = TRUE)

# different ptables for cells with even/odd number of contributors
# and very small cells
ptab3 <- ptable::pt_ex_nums(parity = FALSE, separation = TRUE)

p_nums1 <- ck_params_nums(
  ptab = ptab1,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.30, 0.03),
    epsilon = c(1, 0.5, 0.2),
    q = 3),
  mu_c = 2,
  same_key = FALSE,
  use_zero_rkeys = FALSE,
  path = f_yaml)

#> yaml configuration '/tmp/Rtmp09UBN4/file3bd4b425bd7.yaml' successfully written.

# we read the parameters from the yaml-file
p_nums1 <- ck_read_yaml(path = f_yaml)

# for variables with positive and negative values
p_nums2 <- ck_params_nums(
  ptab = ptab2,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.15, 0.02),
    epsilon = c(1, 0.4, 0.15),
    q = 3),
  mu_c = 2,
  same_key = FALSE)

# simple perturbation parameters (not using the flex-function approach)
p_nums3 <- ck_params_nums(
  ptab = ptab3,
  type = "mean",
  mult_params = ck_simpleparams(p = 0.25),
  mu_c = 2,
  same_key = FALSE)

# use `p_nums1` for all variables
tab$params_nums_set(p_nums1, c("savings", "income", "expend"))

#> --> setting perturbation parameters for variable 'savings'

#> --> setting perturbation parameters for variable 'income'

#> --> setting perturbation parameters for variable 'expend'

# use different parameters for variable `mixed`
tab$params_nums_set(p_nums2, v = "mixed")

#> --> setting perturbation parameters for variable 'mixed'

# identify sensitive cells to which extra protection (`mu_c`) is added.
tab$supp_p(v = "income", p = 85)

#> p%-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)

tab$supp_pq(v = "income", p = 85, q = 90)

#> pq-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)

tab$supp_nk(v = "income", n = 2, k = 90)

#> nk-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)

tab$supp_freq(v = "income", n = 14, weighted = FALSE)

#> freq-rule: 5 new sensitive cells (incl. duplicates) found (total: 5)

tab$supp_val(v = "income", n = 10000, weighted = TRUE)

#> val-rule: 0 new sensitive cells (incl. duplicates) found (total: 5)

# perturb variables
tab$perturb(v = c("income", "savings"))

#> Numeric variable 'income' was perturbed.

#> Numeric variable 'savings' was perturbed.

# extract results
tab$numtab("income", mean_before_sum = TRUE)

#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1373397190 1373475807
#>  2:  Total age_group1 income  9810547  589228535  589314771
#>  3:  Total       ag1a income  9810547  589228535  589314771
#>  4:  Total age_group2 income  5692119  343138307  343061506
#>  5:  Total       ag2a income  5692119  343138307  343061506
#>  6:  Total age_group3 income  4406946  257320956  257293074
#>  7:  Total age_group4 income  2133543  131137156  131165266
#>  8:  Total age_group5 income   848151   49123042   49135667
#>  9:  Total age_group6 income    61672    3449194    3458108
#> 10: female      Total income 11690929  696249286  696260668
#> 11:      f      Total income 11690929  696249286  696260668
#> 12: female age_group1 income  4933383  294359263  294121844
#> 13:      f age_group1 income  4933383  294359263  294121844
#> 14: female       ag1a income  4933383  294359263  294121844
#> 15:      f       ag1a income  4933383  294359263  294121844
#> 16: female age_group2 income  2880740  169575918  169724900
#> 17:      f age_group2 income  2880740  169575918  169724900
#> 18: female       ag2a income  2880740  169575918  169724900
#> 19:      f       ag2a income  2880740  169575918  169724900
#> 20: female age_group3 income  2238777  132121545  132078831
#> 21:      f age_group3 income  2238777  132121545  132078831
#> 22: female age_group4 income  1155033   71134484   71095376
#> 23:      f age_group4 income  1155033   71134484   71095376
#> 24: female age_group5 income   455017   27581237   27502421
#> 25:      f age_group5 income   455017   27581237   27502421
#> 26: female age_group6 income    27979    1476839    1543745
#> 27:      f age_group6 income    27979    1476839    1543745
#> 28:   male      Total income 11262049  677147904  677239965
#> 29:      m      Total income 11262049  677147904  677239965
#> 30:   male age_group1 income  4877164  294869272  294921043
#> 31:      m age_group1 income  4877164  294869272  294921043
#> 32:   male       ag1a income  4877164  294869272  294921043
#> 33:      m       ag1a income  4877164  294869272  294921043
#> 34:   male age_group2 income  2811379  173562389  173531254
#> 35:      m age_group2 income  2811379  173562389  173531254
#> 36:   male       ag2a income  2811379  173562389  173531254
#> 37:      m       ag2a income  2811379  173562389  173531254
#> 38:   male age_group3 income  2168169  125199411  125206273
#> 39:      m age_group3 income  2168169  125199411  125206273
#> 40:   male age_group4 income   978510   60002672   60004016
#> 41:      m age_group4 income   978510   60002672   60004016
#> 42:   male age_group5 income   393134   21541805   21436382
#> 43:      m age_group5 income   393134   21541805   21436382
#> 44:   male age_group6 income    33693    1972355    2010641
#> 45:      m age_group6 income    33693    1972355    2010641
#>        sex        age  vname      uws         ws        pws

tab$numtab("income", mean_before_sum = FALSE)

#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1373397190 1373436498
#>  2:  Total age_group1 income  9810547  589228535  589271651
#>  3:  Total       ag1a income  9810547  589228535  589271651
#>  4:  Total age_group2 income  5692119  343138307  343099905
#>  5:  Total       ag2a income  5692119  343138307  343099905
#>  6:  Total age_group3 income  4406946  257320956  257307014
#>  7:  Total age_group4 income  2133543  131137156  131151210
#>  8:  Total age_group5 income   848151   49123042   49129354
#>  9:  Total age_group6 income    61672    3449194    3453648
#> 10: female      Total income 11690929  696249286  696254977
#> 11:      f      Total income 11690929  696249286  696254977
#> 12: female age_group1 income  4933383  294359263  294240529
#> 13:      f age_group1 income  4933383  294359263  294240529
#> 14: female       ag1a income  4933383  294359263  294240529
#> 15:      f       ag1a income  4933383  294359263  294240529
#> 16: female age_group2 income  2880740  169575918  169650393
#> 17:      f age_group2 income  2880740  169575918  169650393
#> 18: female       ag2a income  2880740  169575918  169650393
#> 19:      f       ag2a income  2880740  169575918  169650393
#> 20: female age_group3 income  2238777  132121545  132100186
#> 21:      f age_group3 income  2238777  132121545  132100186
#> 22: female age_group4 income  1155033   71134484   71114927
#> 23:      f age_group4 income  1155033   71134484   71114927
#> 24: female age_group5 income   455017   27581237   27541801
#> 25:      f age_group5 income   455017   27581237   27541801
#> 26: female age_group6 income    27979    1476839    1509921
#> 27:      f age_group6 income    27979    1476839    1509921
#> 28:   male      Total income 11262049  677147904  677193933
#> 29:      m      Total income 11262049  677147904  677193933
#> 30:   male age_group1 income  4877164  294869272  294895157
#> 31:      m age_group1 income  4877164  294869272  294895157
#> 32:   male       ag1a income  4877164  294869272  294895157
#> 33:      m       ag1a income  4877164  294869272  294895157
#> 34:   male age_group2 income  2811379  173562389  173546821
#> 35:      m age_group2 income  2811379  173562389  173546821
#> 36:   male       ag2a income  2811379  173562389  173546821
#> 37:      m       ag2a income  2811379  173562389  173546821
#> 38:   male age_group3 income  2168169  125199411  125202842
#> 39:      m age_group3 income  2168169  125199411  125202842
#> 40:   male age_group4 income   978510   60002672   60003344
#> 41:      m age_group4 income   978510   60002672   60003344
#> 42:   male age_group5 income   393134   21541805   21489029
#> 43:      m age_group5 income   393134   21541805   21489029
#> 44:   male age_group6 income    33693    1972355    1991406
#> 45:      m age_group6 income    33693    1972355    1991406
#>        sex        age  vname      uws         ws        pws

tab$numtab("savings")

#>        sex        age   vname     uws        ws         pws
#>  1:  Total      Total savings 2273532 135234314 135233049.2
#>  2:  Total age_group1 savings  982386  58688160  58677429.4
#>  3:  Total       ag1a savings  982386  58688160  58677429.4
#>  4:  Total age_group2 savings  552336  33103736  33099559.2
#>  5:  Total       ag2a savings  552336  33103736  33099559.2
#>  6:  Total age_group3 savings  437101  25369617  25371975.1
#>  7:  Total age_group4 savings  214661  13144692  13146424.6
#>  8:  Total age_group5 savings   80451   4575088   4575142.5
#>  9:  Total age_group6 savings    6597    353021    353068.8
#> 10: female      Total savings 1113716  65866414  65870380.9
#> 11:      f      Total savings 1113716  65866414  65870380.9
#> 12: female age_group1 savings  464726  27620562  27611854.5
#> 13:      f age_group1 savings  464726  27620562  27611854.5
#> 14: female       ag1a savings  464726  27620562  27611854.5
#> 15:      f       ag1a savings  464726  27620562  27611854.5
#> 16: female age_group2 savings  271413  15831252  15836820.7
#> 17:      f age_group2 savings  271413  15831252  15836820.7
#> 18: female       ag2a savings  271413  15831252  15836820.7
#> 19:      f       ag2a savings  271413  15831252  15836820.7
#> 20: female age_group3 savings  222131  13110880  13108140.5
#> 21:      f age_group3 savings  222131  13110880  13108140.5
#> 22: female age_group4 savings  115241   6965582   6964091.8
#> 23:      f age_group4 savings  115241   6965582   6964091.8
#> 24: female age_group5 savings   37218   2197290   2194020.4
#> 25:      f age_group5 savings   37218   2197290   2194020.4
#> 26: female age_group6 savings    2987    140848    142912.7
#> 27:      f age_group6 savings    2987    140848    142912.7
#> 28:   male      Total savings 1159816  69367900  69366815.5
#> 29:      m      Total savings 1159816  69367900  69366815.5
#> 30:   male age_group1 savings  517660  31067598  31067667.6
#> 31:      m age_group1 savings  517660  31067598  31067667.6
#> 32:   male       ag1a savings  517660  31067598  31067667.6
#> 33:      m       ag1a savings  517660  31067598  31067667.6
#> 34:   male age_group2 savings  280923  17272484  17270902.2
#> 35:      m age_group2 savings  280923  17272484  17270902.2
#> 36:   male       ag2a savings  280923  17272484  17270902.2
#> 37:      m       ag2a savings  280923  17272484  17270902.2
#> 38:   male age_group3 savings  214970  12258737  12256543.3
#> 39:      m age_group3 savings  214970  12258737  12256543.3
#> 40:   male age_group4 savings   99420   6179110   6179610.2
#> 41:      m age_group4 savings   99420   6179110   6179610.2
#> 42:   male age_group5 savings   43233   2377798   2375541.6
#> 43:      m age_group5 savings   43233   2377798   2375541.6
#> 44:   male age_group6 savings    3610    212173    213121.3
#> 45:      m age_group6 savings    3610    212173    213121.3
#>        sex        age   vname     uws        ws         pws

# results can be resetted, too
tab$reset_cntvars(v = "cnt_males")

# we can then set other parameters and perturb again
tab$params_cnts_set(val = p_cnts1, v = "cnt_males")

#> --> setting perturbation parameters for variable 'cnt_males'

tab$perturb(v = "cnt_males")

#> Count variable 'cnt_males' was perturbed.

# write to a file "outtab.csv" (.csv is automatically added to the path)
if (FALSE) {
tab$freqtab(v = c("total", "cnt_males"), path = "outtab.csv")
}

# show results containing weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))

#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 273086 4581 273145.6258
#>  2:  Total age_group1     total 1969 118141 1968 118080.9995
#>  3:  Total       ag1a     total 1969 118141 1968 118080.9995
#>  4:  Total age_group2     total 1143  68719 1143  68719.0000
#>  5:  Total       ag2a     total 1143  68719 1143  68719.0000
#>  6:  Total age_group3     total  864  50212  862  50095.7685
#>  7:  Total age_group4     total  423  25768  424  25828.9173
#>  8:  Total age_group5     total  168   9547  168   9547.0000
#>  9:  Total age_group6     total   13    699   13    699.0000
#> 10: female      Total     total 2284 135726 2284 135726.0000
#> 11:      f      Total     total 2284 135726 2284 135726.0000
#> 12: female age_group1     total  954  57024  955  57083.7736
#> 13:      f age_group1     total  954  57024  955  57083.7736
#> 14: female       ag1a     total  954  57024  955  57083.7736
#> 15:      f       ag1a     total  954  57024  955  57083.7736
#> 16: female age_group2     total  572  33785  574  33903.1294
#> 17:      f age_group2     total  572  33785  574  33903.1294
#> 18: female       ag2a     total  572  33785  574  33903.1294
#> 19:      f       ag2a     total  572  33785  574  33903.1294
#> 20: female age_group3     total  440  25885  440  25885.0000
#> 21:      f age_group3     total  440  25885  440  25885.0000
#> 22: female age_group4     total  228  13755  228  13755.0000
#> 23:      f age_group4     total  228  13755  228  13755.0000
#> 24: female age_group5     total   84   4978   85   5037.2619
#> 25:      f age_group5     total   84   4978   85   5037.2619
#> 26: female age_group6     total    6    299    8    398.6667
#> 27:      f age_group6     total    6    299    8    398.6667
#> 28:   male      Total     total 2296 137360 2295 137300.1742
#> 29:      m      Total     total 2296 137360 2295 137300.1742
#> 30:   male age_group1     total 1015  61117 1015  61117.0000
#> 31:      m age_group1     total 1015  61117 1015  61117.0000
#> 32:   male       ag1a     total 1015  61117 1015  61117.0000
#> 33:      m       ag1a     total 1015  61117 1015  61117.0000
#> 34:   male age_group2     total  571  34934  571  34934.0000
#> 35:      m age_group2     total  571  34934  571  34934.0000
#> 36:   male       ag2a     total  571  34934  571  34934.0000
#> 37:      m       ag2a     total  571  34934  571  34934.0000
#> 38:   male age_group3     total  424  24327  424  24327.0000
#> 39:      m age_group3     total  424  24327  424  24327.0000
#> 40:   male age_group4     total  195  12013  194  11951.3949
#> 41:      m age_group4     total  195  12013  194  11951.3949
#> 42:   male age_group5     total   84   4569   85   4623.3929
#> 43:      m age_group5     total   84   4569   85   4623.3929
#> 44:   male age_group6     total    7    400    7    400.0000
#> 45:      m age_group6     total    7    400    7    400.0000
#> 46:  Total      Total cnt_males 2296 137360 2295 137300.1742
#> 47:  Total age_group1 cnt_males 1015  61117 1015  61117.0000
#> 48:  Total       ag1a cnt_males 1015  61117 1015  61117.0000
#> 49:  Total age_group2 cnt_males  571  34934  571  34934.0000
#> 50:  Total       ag2a cnt_males  571  34934  571  34934.0000
#> 51:  Total age_group3 cnt_males  424  24327  424  24327.0000
#> 52:  Total age_group4 cnt_males  195  12013  194  11951.3949
#> 53:  Total age_group5 cnt_males   84   4569   85   4623.3929
#> 54:  Total age_group6 cnt_males    7    400    7    400.0000
#> 55: female      Total cnt_males    0      0    0      0.0000
#> 56:      f      Total cnt_males    0      0    0      0.0000
#> 57: female age_group1 cnt_males    0      0    0      0.0000
#> 58:      f age_group1 cnt_males    0      0    0      0.0000
#> 59: female       ag1a cnt_males    0      0    0      0.0000
#> 60:      f       ag1a cnt_males    0      0    0      0.0000
#> 61: female age_group2 cnt_males    0      0    0      0.0000
#> 62:      f age_group2 cnt_males    0      0    0      0.0000
#> 63: female       ag2a cnt_males    0      0    0      0.0000
#> 64:      f       ag2a cnt_males    0      0    0      0.0000
#> 65: female age_group3 cnt_males    0      0    0      0.0000
#> 66:      f age_group3 cnt_males    0      0    0      0.0000
#> 67: female age_group4 cnt_males    0      0    0      0.0000
#> 68:      f age_group4 cnt_males    0      0    0      0.0000
#> 69: female age_group5 cnt_males    0      0    0      0.0000
#> 70:      f age_group5 cnt_males    0      0    0      0.0000
#> 71: female age_group6 cnt_males    0      0    0      0.0000
#> 72:      f age_group6 cnt_males    0      0    0      0.0000
#> 73:   male      Total cnt_males 2296 137360 2295 137300.1742
#> 74:      m      Total cnt_males 2296 137360 2295 137300.1742
#> 75:   male age_group1 cnt_males 1015  61117 1015  61117.0000
#> 76:      m age_group1 cnt_males 1015  61117 1015  61117.0000
#> 77:   male       ag1a cnt_males 1015  61117 1015  61117.0000
#> 78:      m       ag1a cnt_males 1015  61117 1015  61117.0000
#> 79:   male age_group2 cnt_males  571  34934  571  34934.0000
#> 80:      m age_group2 cnt_males  571  34934  571  34934.0000
#> 81:   male       ag2a cnt_males  571  34934  571  34934.0000
#> 82:      m       ag2a cnt_males  571  34934  571  34934.0000
#> 83:   male age_group3 cnt_males  424  24327  424  24327.0000
#> 84:      m age_group3 cnt_males  424  24327  424  24327.0000
#> 85:   male age_group4 cnt_males  195  12013  194  11951.3949
#> 86:      m age_group4 cnt_males  195  12013  194  11951.3949
#> 87:   male age_group5 cnt_males   84   4569   85   4623.3929
#> 88:      m age_group5 cnt_males   84   4569   85   4623.3929
#> 89:   male age_group6 cnt_males    7    400    7    400.0000
#> 90:      m age_group6 cnt_males    7    400    7    400.0000
#>        sex        age     vname  uwc     wc puwc         pwc

# utility measures for a count variable
tab$measures_cnts(v = "total", exclude_zeros = TRUE)

#> $overview
#>    noise cnt        pct
#> 1:    -2   6 0.13333333
#> 2:    -1  10 0.22222222
#> 3:     0  22 0.48888889
#> 4:     1   6 0.13333333
#> 5:     2   1 0.02222222
#> 
#> $measures
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 0.000 0.000 0.000
#>  6:   Mean 0.667 0.017 0.031
#>  7: Median 1.000 0.000 0.007
#>  8:    Q60 1.000 0.001 0.013
#>  9:    Q70 1.000 0.002 0.023
#> 10:    Q80 1.000 0.003 0.042
#> 11:    Q90 2.000 0.012 0.054
#> 12:    Q95 2.000 0.012 0.054
#> 13:    Q99 2.000 0.333 0.379
#> 14:    Max 2.000 0.333 0.379
#> 
#> $cumdistr_d1
#>    cat cnt       pct
#> 1:   0  22 0.4888889
#> 2:   1  38 0.8444444
#> 3:   2  45 1.0000000
#> 
#> $cumdistr_d2
#>            cat cnt       pct
#> 1:    [0,0.02]  43 0.9555556
#> 2: (0.02,0.05]  43 0.9555556
#> 3:  (0.05,0.1]  43 0.9555556
#> 4:   (0.1,0.2]  43 0.9555556
#> 5:   (0.2,0.3]  43 0.9555556
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $cumdistr_d3
#>            cat cnt       pct
#> 1:    [0,0.02]  31 0.6888889
#> 2: (0.02,0.05]  39 0.8666667
#> 3:  (0.05,0.1]  43 0.9555556
#> 4:   (0.1,0.2]  43 0.9555556
#> 5:   (0.2,0.3]  43 0.9555556
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $false_zero
#> [1] 0
#> 
#> $false_nonzero
#> [1] 0
#> 
#> $exclude_zeros
#> [1] TRUE
#> 

# modifications for perturbed count variables
tab$mod_cnts()

#>        sex        age row_nr pert      ckey  countvar
#>   1: Total      Total     16    1 0.7087783     total
#>   2: Total age_group1     14   -1 0.1912354     total
#>   3: Total       ag1a     14   -1 0.1912354     total
#>   4: Total age_group2     15    0 0.6058147     total
#>   5: Total       ag2a     15    0 0.6058147     total
#>  ---                                                 
#> 131:     m age_group4     14   -1 0.2376177 cnt_males
#> 132:  male age_group5     16    1 0.7124368 cnt_males
#> 133:     m age_group5     16    1 0.7124368 cnt_males
#> 134:  male age_group6     15    0 0.5618537 cnt_males
#> 135:     m age_group6     15    0 0.5618537 cnt_males

# display a summary about utility measures
tab$summary()

#> ┌──────────────────────────────────────────────┐
#> │Utility measures for perturbed count variables│
#> └──────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#>          countvar Min Q10 Q20 Q30 Q40   Mean Median Q60 Q70 Q80 Q90 Q95 Q99 Max
#> 1:          total  -2  -1   0   0   0  0.311      0 0.0   1   1   2 2.0   2   2
#> 2: cnt_highincome  -4  -2  -1  -1   0  0.067      0 0.4   1   2   2 2.0   3   3
#> 3:      cnt_males  -1  -1   0   0   0 -0.067      0 0.0   0   0   0 0.8   1   1
#> 
#> ── Distance-based measures ─────────────────────────────────────────────────────
#> ✔ Variable: 'total'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 0.000 0.000 0.000
#>  6:   Mean 0.667 0.017 0.031
#>  7: Median 1.000 0.000 0.007
#>  8:    Q60 1.000 0.001 0.013
#>  9:    Q70 1.000 0.002 0.023
#> 10:    Q80 1.000 0.003 0.042
#> 11:    Q90 2.000 0.012 0.054
#> 12:    Q95 2.000 0.012 0.054
#> 13:    Q99 2.000 0.333 0.379
#> 14:    Max 2.000 0.333 0.379
#> 
#> ✔ Variable: 'cnt_males'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 0.000 0.000 0.000
#>  6:   Mean 0.333 0.002 0.011
#>  7: Median 0.000 0.000 0.000
#>  8:    Q60 0.000 0.000 0.000
#>  9:    Q70 1.000 0.000 0.010
#> 10:    Q80 1.000 0.004 0.031
#> 11:    Q90 1.000 0.008 0.043
#> 12:    Q95 1.000 0.012 0.054
#> 13:    Q99 1.000 0.012 0.054
#> 14:    Max 1.000 0.012 0.054
#> 
#> ✔ Variable: 'cnt_highincome'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 1.000 0.009 0.053
#>  4:    Q30 1.000 0.011 0.062
#>  5:    Q40 1.000 0.015 0.066
#>  6:   Mean 1.425 0.044 0.109
#>  7: Median 1.000 0.020 0.078
#>  8:    Q60 2.000 0.024 0.100
#>  9:    Q70 2.000 0.026 0.131
#> 10:    Q80 2.000 0.035 0.136
#> 11:    Q90 3.000 0.143 0.196
#> 12:    Q95 3.000 0.200 0.370
#> 13:    Q99 3.610 0.252 0.498
#> 14:    Max 4.000 0.286 0.579
#> 
#> ┌──────────────────────────────────────────────────┐
#> │Utility measures for perturbed numerical variables│
#> └──────────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────

#> Warning: no non-missing arguments to min; returning Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to min; returning Inf

#> Warning: no non-missing arguments to max; returning -Inf

#>      vname        Min        Q10        Q20        Q30        Q40      Mean
#> 1:  expend        Inf         NA         NA         NA         NA       NaN
#> 2:  income -118733.65 -52776.173 -38402.426 -18758.878 -14592.170 -2811.088
#> 3: savings  -10730.63  -8707.499  -3269.637  -2243.875  -1581.797 -1257.744
#> 4:   mixed        Inf         NA         NA         NA         NA       NaN
#>       Median      Q60      Q70      Q80       Q90       Q95       Q99       Max
#> 1:        NA       NA       NA       NA        NA        NA        NA      -Inf
#> 2:  3430.831 9408.921 25884.60 34327.54 46028.783 74474.602 74474.602 74474.602
#> 3: -1264.839   60.498   414.08  1799.03  3966.919  5568.729  5568.729  5568.729
#> 4:        NA       NA       NA       NA        NA        NA        NA      -Inf