Set parameters required to perturb numeric variables using a simple approach

ck_simpleparams() allows to define parameters for a simple perturbation approach based on a single magnitude parameter (m). The values of epsilon are used to "weight" parameter m in case type == "top_contr" is set in ck_params_nums().

ck_simpleparams(p, epsilon = 1)

Arguments

p
a percentage value used as magnitude for perturbation

epsilon
a numeric vector in descending order with all values >= 0 and <= 1 with the first element forced to equal 1. The length of this vector must correspond with the number top_k specified in ck_params_nums() when creating parameters for type == "top_contr" which is checked at runtime. This setting allows to use different flex-functions for the largest top_k contributors.

Value

an object suitable as input for ck_params_nums().

Details

details about the flex function can be found in Deliverable D4.2, Part I in SGA "Open Source tools for perturbative confidentiality methods"

See also

Examples

x <- ck_create_testdata()

# create some 0/1 variables that should be perturbed later
x[, cnt_females := ifelse(sex == "male", 0, 1)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              69          25.00000           0
#>    2:   2530      28       1              78          25.00000           1
#>    3:   6920     550       1              59          25.00000           0
#>    4:   7960     870       1              78          25.00000           0
#>    5:   9030      20       2              37          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              23          16.66667           1
#> 4577:   1420     987    1000              67          16.66667           0
#> 4578:   8900     684    1000              72          16.66667           0
#> 4579:   3880     294    1000              89          16.66667           1
#> 4580:   4830     911    1000              49          16.66667           0

x[, cnt_males := ifelse(sex == "male", 1, 0)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              69          25.00000           0
#>    2:   2530      28       1              78          25.00000           1
#>    3:   6920     550       1              59          25.00000           0
#>    4:   7960     870       1              78          25.00000           0
#>    5:   9030      20       2              37          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              23          16.66667           1
#> 4577:   1420     987    1000              67          16.66667           0
#> 4578:   8900     684    1000              72          16.66667           0
#> 4579:   3880     294    1000              89          16.66667           1
#> 4580:   4830     911    1000              49          16.66667           0
#>       cnt_males
#>    1:         1
#>    2:         0
#>    3:         1
#>    4:         1
#>    5:         1
#>   ---          
#> 4576:         0
#> 4577:         1
#> 4578:         1
#> 4579:         0
#> 4580:         1

x[, cnt_highincome := ifelse(income >= 9000, 1, 0)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              69          25.00000           0
#>    2:   2530      28       1              78          25.00000           1
#>    3:   6920     550       1              59          25.00000           0
#>    4:   7960     870       1              78          25.00000           0
#>    5:   9030      20       2              37          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              23          16.66667           1
#> 4577:   1420     987    1000              67          16.66667           0
#> 4578:   8900     684    1000              72          16.66667           0
#> 4579:   3880     294    1000              89          16.66667           1
#> 4580:   4830     911    1000              49          16.66667           0
#>       cnt_males cnt_highincome
#>    1:         1              0
#>    2:         0              0
#>    3:         1              0
#>    4:         1              0
#>    5:         1              1
#>   ---                         
#> 4576:         0              0
#> 4577:         1              0
#> 4578:         1              0
#> 4579:         0              0
#> 4580:         1              0

# a variable with positive and negative contributions
x[, mixed := sample(-10:10, nrow(x), replace = TRUE)]

#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              69          25.00000           0
#>    2:   2530      28       1              78          25.00000           1
#>    3:   6920     550       1              59          25.00000           0
#>    4:   7960     870       1              78          25.00000           0
#>    5:   9030      20       2              37          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              23          16.66667           1
#> 4577:   1420     987    1000              67          16.66667           0
#> 4578:   8900     684    1000              72          16.66667           0
#> 4579:   3880     294    1000              89          16.66667           1
#> 4580:   4830     911    1000              49          16.66667           0
#>       cnt_males cnt_highincome mixed
#>    1:         1              0     8
#>    2:         0              0     1
#>    3:         1              0    10
#>    4:         1              0     9
#>    5:         1              1     0
#>   ---                               
#> 4576:         0              0     4
#> 4577:         1              0    -8
#> 4578:         1              0    -1
#> 4579:         0              0    -1
#> 4580:         1              0    -3

# create record keys
x$rkey <- ck_generate_rkeys(dat = x)

# define required inputs

# hierarchy with some bogus codes
d_sex <- hier_create(root = "Total", nodes = c("male", "female"))
d_sex <- hier_add(d_sex, root = "female", "f")
d_sex <- hier_add(d_sex, root = "male", "m")

d_age <- hier_create(root = "Total", nodes = paste0("age_group", 1:6))
d_age <- hier_add(d_age, root = "age_group1", "ag1a")
d_age <- hier_add(d_age, root = "age_group2", "ag2a")

# define the cell key object
countvars <- c("cnt_females", "cnt_males", "cnt_highincome")
numvars <- c("expend", "income", "savings", "mixed")
tab <- ck_setup(
  x = x,
  rkey = "rkey",
  dims = list(sex = d_sex, age = d_age),
  w = "sampling_weight",
  countvars = countvars,
  numvars = numvars)

# show some information about this table instance
tab$print() # identical with print(tab)

#> ── Table Information ───────────────────────────────────────────────────────────
#> ✔ 45 cells in 2 dimensions ('sex', 'age')
#> ✔ weights: yes
#> ── Tabulated / Perturbed countvars ─────────────────────────────────────────────
#> ☐ 'total'
#> ☐ 'cnt_females'
#> ☐ 'cnt_males'
#> ☐ 'cnt_highincome'
#> ── Tabulated / Perturbed numvars ───────────────────────────────────────────────
#> ☐ 'expend'
#> ☐ 'income'
#> ☐ 'savings'
#> ☐ 'mixed'

# information about the hierarchies
tab$hierarchy_info()

#> $sex
#>      code level is_leaf parent
#> 1:  Total     1   FALSE   <NA>
#> 2:   male     2   FALSE  Total
#> 3: female     2   FALSE  Total
#> 4:      f     3    TRUE female
#> 5:      m     3    TRUE   male
#> 
#> $age
#>          code level is_leaf     parent
#> 1:      Total     1   FALSE       <NA>
#> 2: age_group1     2   FALSE      Total
#> 3: age_group2     2   FALSE      Total
#> 4: age_group3     2    TRUE      Total
#> 5: age_group4     2    TRUE      Total
#> 6: age_group5     2    TRUE      Total
#> 7: age_group6     2    TRUE      Total
#> 8:       ag1a     3    TRUE age_group1
#> 9:       ag2a     3    TRUE age_group2
#> 

# which variables have been defined?
tab$allvars()

#> $cntvars
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"
#> 
#> $numvars
#> [1] "expend"  "income"  "savings" "mixed"  
#> 

# count variables
tab$cntvars()

#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"

# continuous variables
tab$numvars()

#> [1] "expend"  "income"  "savings" "mixed"  

# create perturbation parameters for "total" variable and
# write to yaml-file

# create a ptable using functionality from the ptable-pkg
f_yaml <- tempfile(fileext = ".yaml")
p_cnts1 <- ck_params_cnts(
  ptab = ptable::pt_ex_cnts(),
  path = f_yaml)

#> yaml configuration '/tmp/RtmpYzze5a/file486624034a21.yaml' successfully written.

# read parameters from yaml-file and set them for variable `"total"`
p_cnts1 <- ck_read_yaml(path = f_yaml)

tab$params_cnts_set(val = p_cnts1, v = "total")

#> --> setting perturbation parameters for variable 'total'

# create alternative perturbation parameters by specifying parameters
para2 <- ptable::pt_create_pParams(
  D = 8, V = 3, js = 2, table = "cnts")

#> Ignoring argument `step` for freqency tables

p_cnts2 <- ck_params_cnts(ptab = para2)

# use these ptable it for the remaining variables
tab$params_cnts_set(val = p_cnts2, v = countvars)

#> --> setting perturbation parameters for variable 'cnt_females'

#> --> setting perturbation parameters for variable 'cnt_males'

#> --> setting perturbation parameters for variable 'cnt_highincome'

# perturb a variable
tab$perturb(v = "total")

#> Count variable 'total' was perturbed.

# multiple variables can be perturbed as well
tab$perturb(v = c("cnt_males", "cnt_highincome"))

#> Count variable 'cnt_males' was perturbed.

#> Count variable 'cnt_highincome' was perturbed.

# return weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))

#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 275400 4580 275400.0000
#>  2:  Total age_group1     total 1969 118766 1969 118766.0000
#>  3:  Total       ag1a     total 1969 118766 1969 118766.0000
#>  4:  Total age_group2     total 1143  69184 1144  69244.5284
#>  5:  Total       ag2a     total 1143  69184 1144  69244.5284
#>  6:  Total age_group3     total  864  51018  866  51136.0972
#>  7:  Total age_group4     total  423  25479  421  25358.5319
#>  8:  Total age_group5     total  168  10173  167  10112.4464
#>  9:  Total age_group6     total   13    780   11    660.0000
#> 10: female      Total     total 2284 137217 2285 137277.0775
#> 11:      f      Total     total 2284 137217 2285 137277.0775
#> 12: female age_group1     total  954  57639  954  57639.0000
#> 13:      f age_group1     total  954  57639  954  57639.0000
#> 14: female       ag1a     total  954  57639  954  57639.0000
#> 15:      f       ag1a     total  954  57639  954  57639.0000
#> 16: female age_group2     total  572  34521  573  34581.3514
#> 17:      f age_group2     total  572  34521  573  34581.3514
#> 18: female       ag2a     total  572  34521  573  34581.3514
#> 19:      f       ag2a     total  572  34521  573  34581.3514
#> 20: female age_group3     total  440  25910  441  25968.8864
#> 21:      f age_group3     total  440  25910  441  25968.8864
#> 22: female age_group4     total  228  13704  229  13764.1053
#> 23:      f age_group4     total  228  13704  229  13764.1053
#> 24: female age_group5     total   84   5035   85   5094.9405
#> 25:      f age_group5     total   84   5035   85   5094.9405
#> 26: female age_group6     total    6    408    7    476.0000
#> 27:      f age_group6     total    6    408    7    476.0000
#> 28:   male      Total     total 2296 138183 2297 138243.1842
#> 29:      m      Total     total 2296 138183 2297 138243.1842
#> 30:   male age_group1     total 1015  61127 1013  61006.5527
#> 31:      m age_group1     total 1015  61127 1013  61006.5527
#> 32:   male       ag1a     total 1015  61127 1013  61006.5527
#> 33:      m       ag1a     total 1015  61127 1013  61006.5527
#> 34:   male age_group2     total  571  34663  572  34723.7058
#> 35:      m age_group2     total  571  34663  572  34723.7058
#> 36:   male       ag2a     total  571  34663  572  34723.7058
#> 37:      m       ag2a     total  571  34663  572  34723.7058
#> 38:   male age_group3     total  424  25108  423  25048.7830
#> 39:      m age_group3     total  424  25108  423  25048.7830
#> 40:   male age_group4     total  195  11775  194  11714.6154
#> 41:      m age_group4     total  195  11775  194  11714.6154
#> 42:   male age_group5     total   84   5138   83   5076.8333
#> 43:      m age_group5     total   84   5138   83   5076.8333
#> 44:   male age_group6     total    7    372    6    318.8571
#> 45:      m age_group6     total    7    372    6    318.8571
#> 46:  Total      Total cnt_males 2296 138183 2297 138243.1842
#> 47:  Total age_group1 cnt_males 1015  61127 1012  60946.3291
#> 48:  Total       ag1a cnt_males 1015  61127 1012  60946.3291
#> 49:  Total age_group2 cnt_males  571  34663  573  34784.4116
#> 50:  Total       ag2a cnt_males  571  34663  573  34784.4116
#> 51:  Total age_group3 cnt_males  424  25108  422  24989.5660
#> 52:  Total age_group4 cnt_males  195  11775  193  11654.2308
#> 53:  Total age_group5 cnt_males   84   5138   83   5076.8333
#> 54:  Total age_group6 cnt_males    7    372    6    318.8571
#> 55: female      Total cnt_males    0      0    0      0.0000
#> 56:      f      Total cnt_males    0      0    0      0.0000
#> 57: female age_group1 cnt_males    0      0    0      0.0000
#> 58:      f age_group1 cnt_males    0      0    0      0.0000
#> 59: female       ag1a cnt_males    0      0    0      0.0000
#> 60:      f       ag1a cnt_males    0      0    0      0.0000
#> 61: female age_group2 cnt_males    0      0    0      0.0000
#> 62:      f age_group2 cnt_males    0      0    0      0.0000
#> 63: female       ag2a cnt_males    0      0    0      0.0000
#> 64:      f       ag2a cnt_males    0      0    0      0.0000
#> 65: female age_group3 cnt_males    0      0    0      0.0000
#> 66:      f age_group3 cnt_males    0      0    0      0.0000
#> 67: female age_group4 cnt_males    0      0    0      0.0000
#> 68:      f age_group4 cnt_males    0      0    0      0.0000
#> 69: female age_group5 cnt_males    0      0    0      0.0000
#> 70:      f age_group5 cnt_males    0      0    0      0.0000
#> 71: female age_group6 cnt_males    0      0    0      0.0000
#> 72:      f age_group6 cnt_males    0      0    0      0.0000
#> 73:   male      Total cnt_males 2296 138183 2297 138243.1842
#> 74:      m      Total cnt_males 2296 138183 2297 138243.1842
#> 75:   male age_group1 cnt_males 1015  61127 1012  60946.3291
#> 76:      m age_group1 cnt_males 1015  61127 1012  60946.3291
#> 77:   male       ag1a cnt_males 1015  61127 1012  60946.3291
#> 78:      m       ag1a cnt_males 1015  61127 1012  60946.3291
#> 79:   male age_group2 cnt_males  571  34663  573  34784.4116
#> 80:      m age_group2 cnt_males  571  34663  573  34784.4116
#> 81:   male       ag2a cnt_males  571  34663  573  34784.4116
#> 82:      m       ag2a cnt_males  571  34663  573  34784.4116
#> 83:   male age_group3 cnt_males  424  25108  422  24989.5660
#> 84:      m age_group3 cnt_males  424  25108  422  24989.5660
#> 85:   male age_group4 cnt_males  195  11775  193  11654.2308
#> 86:      m age_group4 cnt_males  195  11775  193  11654.2308
#> 87:   male age_group5 cnt_males   84   5138   83   5076.8333
#> 88:      m age_group5 cnt_males   84   5138   83   5076.8333
#> 89:   male age_group6 cnt_males    7    372    6    318.8571
#> 90:      m age_group6 cnt_males    7    372    6    318.8571
#>        sex        age     vname  uwc     wc puwc         pwc

# numerical variables (positive variables using flex-function)
# we also write the config to a yaml file
f_yaml <- tempfile(fileext = ".yaml")

# create a ptable using functionality from the ptable-pkg
# a single ptable for all cells
ptab1 <- ptable::pt_ex_nums(parity = TRUE, separation = FALSE)

# a single ptab for all cells except for very small ones
ptab2 <- ptable::pt_ex_nums(parity = TRUE, separation = TRUE)

# different ptables for cells with even/odd number of contributors
# and very small cells
ptab3 <- ptable::pt_ex_nums(parity = FALSE, separation = TRUE)

p_nums1 <- ck_params_nums(
  ptab = ptab1,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.30, 0.03),
    epsilon = c(1, 0.5, 0.2),
    q = 3),
  mu_c = 2,
  same_key = FALSE,
  use_zero_rkeys = FALSE,
  path = f_yaml)

#> yaml configuration '/tmp/RtmpYzze5a/file4866420fa302.yaml' successfully written.

# we read the parameters from the yaml-file
p_nums1 <- ck_read_yaml(path = f_yaml)

# for variables with positive and negative values
p_nums2 <- ck_params_nums(
  ptab = ptab2,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.15, 0.02),
    epsilon = c(1, 0.4, 0.15),
    q = 3),
  mu_c = 2,
  same_key = FALSE)

# simple perturbation parameters (not using the flex-function approach)
p_nums3 <- ck_params_nums(
  ptab = ptab3,
  type = "mean",
  mult_params = ck_simpleparams(p = 0.25),
  mu_c = 2,
  same_key = FALSE)

# use `p_nums1` for all variables
tab$params_nums_set(p_nums1, c("savings", "income", "expend"))

#> --> setting perturbation parameters for variable 'savings'

#> --> setting perturbation parameters for variable 'income'

#> --> setting perturbation parameters for variable 'expend'

# use different parameters for variable `mixed`
tab$params_nums_set(p_nums2, v = "mixed")

#> --> setting perturbation parameters for variable 'mixed'

# identify sensitive cells to which extra protection (`mu_c`) is added.
tab$supp_p(v = "income", p = 85)

#> p%-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)

tab$supp_pq(v = "income", p = 85, q = 90)

#> pq-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)

tab$supp_nk(v = "income", n = 2, k = 90)

#> nk-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)

tab$supp_freq(v = "income", n = 14, weighted = FALSE)

#> freq-rule: 5 new sensitive cells (incl. duplicates) found (total: 5)

tab$supp_val(v = "income", n = 10000, weighted = TRUE)

#> val-rule: 0 new sensitive cells (incl. duplicates) found (total: 5)

tab$supp_cells(
  v = "income",
  inp = data.frame(
    sex = c("female", "female"),
    "age" = c("age_group1", "age_group3")
  )
)

#> cell-rule: 2 new sensitive cells (incl. duplicates) found (total: 7)

# perturb variables
tab$perturb(v = c("income", "savings"))

#> Numeric variable 'income' was perturbed.

#> Numeric variable 'savings' was perturbed.

# extract results
tab$numtab("income", mean_before_sum = TRUE)

#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1386853880 1386861305
#>  2:  Total age_group1 income  9810547  596321132  596330739
#>  3:  Total       ag1a income  9810547  596321132  596330739
#>  4:  Total age_group2 income  5692119  341969861  341993495
#>  5:  Total       ag2a income  5692119  341969861  341993495
#>  6:  Total age_group3 income  4406946  263540591  263509065
#>  7:  Total age_group4 income  2133543  128933854  128895052
#>  8:  Total age_group5 income   848151   52424134   52419604
#>  9:  Total age_group6 income    61672    3664308    3745007
#> 10: female      Total income 11690929  706125041  706195978
#> 11:      f      Total income 11690929  706125041  706195978
#> 12: female age_group1 income  4933383  299401420  299309598
#> 13:      f age_group1 income  4933383  299401420  299309598
#> 14: female       ag1a income  4933383  299401420  299309598
#> 15:      f       ag1a income  4933383  299401420  299309598
#> 16: female age_group2 income  2880740  172749535  172769750
#> 17:      f age_group2 income  2880740  172749535  172769750
#> 18: female       ag2a income  2880740  172749535  172769750
#> 19:      f       ag2a income  2880740  172749535  172769750
#> 20: female age_group3 income  2238777  134269729  134293591
#> 21:      f age_group3 income  2238777  134269729  134293591
#> 22: female age_group4 income  1155033   69818738   69840816
#> 23:      f age_group4 income  1155033   69818738   69840816
#> 24: female age_group5 income   455017   27892210   27805985
#> 25:      f age_group5 income   455017   27892210   27805985
#> 26: female age_group6 income    27979    1993409    1999926
#> 27:      f age_group6 income    27979    1993409    1999926
#> 28:   male      Total income 11262049  680728839  680575368
#> 29:      m      Total income 11262049  680728839  680575368
#> 30:   male age_group1 income  4877164  296919712  296910135
#> 31:      m age_group1 income  4877164  296919712  296910135
#> 32:   male       ag1a income  4877164  296919712  296910135
#> 33:      m       ag1a income  4877164  296919712  296910135
#> 34:   male age_group2 income  2811379  169220326  169149315
#> 35:      m age_group2 income  2811379  169220326  169149315
#> 36:   male       ag2a income  2811379  169220326  169149315
#> 37:      m       ag2a income  2811379  169220326  169149315
#> 38:   male age_group3 income  2168169  129270862  129406962
#> 39:      m age_group3 income  2168169  129270862  129406962
#> 40:   male age_group4 income   978510   59115116   59168988
#> 41:      m age_group4 income   978510   59115116   59168988
#> 42:   male age_group5 income   393134   24531924   24532799
#> 43:      m age_group5 income   393134   24531924   24532799
#> 44:   male age_group6 income    33693    1670899    1770108
#> 45:      m age_group6 income    33693    1670899    1770108
#>        sex        age  vname      uws         ws        pws

tab$numtab("income", mean_before_sum = FALSE)

#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1386853880 1386857593
#>  2:  Total age_group1 income  9810547  596321132  596325936
#>  3:  Total       ag1a income  9810547  596321132  596325936
#>  4:  Total age_group2 income  5692119  341969861  341981678
#>  5:  Total       ag2a income  5692119  341969861  341981678
#>  6:  Total age_group3 income  4406946  263540591  263524828
#>  7:  Total age_group4 income  2133543  128933854  128914452
#>  8:  Total age_group5 income   848151   52424134   52421869
#>  9:  Total age_group6 income    61672    3664308    3704438
#> 10: female      Total income 11690929  706125041  706160509
#> 11:      f      Total income 11690929  706125041  706160509
#> 12: female age_group1 income  4933383  299401420  299355506
#> 13:      f age_group1 income  4933383  299401420  299355506
#> 14: female       ag1a income  4933383  299401420  299355506
#> 15:      f       ag1a income  4933383  299401420  299355506
#> 16: female age_group2 income  2880740  172749535  172759642
#> 17:      f age_group2 income  2880740  172749535  172759642
#> 18: female       ag2a income  2880740  172749535  172759642
#> 19:      f       ag2a income  2880740  172749535  172759642
#> 20: female age_group3 income  2238777  134269729  134281660
#> 21:      f age_group3 income  2238777  134269729  134281660
#> 22: female age_group4 income  1155033   69818738   69829776
#> 23:      f age_group4 income  1155033   69818738   69829776
#> 24: female age_group5 income   455017   27892210   27849064
#> 25:      f age_group5 income   455017   27892210   27849064
#> 26: female age_group6 income    27979    1993409    1996665
#> 27:      f age_group6 income    27979    1993409    1996665
#> 28:   male      Total income 11262049  680728839  680652099
#> 29:      m      Total income 11262049  680728839  680652099
#> 30:   male age_group1 income  4877164  296919712  296914923
#> 31:      m age_group1 income  4877164  296919712  296914923
#> 32:   male       ag1a income  4877164  296919712  296914923
#> 33:      m       ag1a income  4877164  296919712  296914923
#> 34:   male age_group2 income  2811379  169220326  169184817
#> 35:      m age_group2 income  2811379  169220326  169184817
#> 36:   male       ag2a income  2811379  169220326  169184817
#> 37:      m       ag2a income  2811379  169220326  169184817
#> 38:   male age_group3 income  2168169  129270862  129338894
#> 39:      m age_group3 income  2168169  129270862  129338894
#> 40:   male age_group4 income   978510   59115116   59142046
#> 41:      m age_group4 income   978510   59115116   59142046
#> 42:   male age_group5 income   393134   24531924   24532362
#> 43:      m age_group5 income   393134   24531924   24532362
#> 44:   male age_group6 income    33693    1670899    1719788
#> 45:      m age_group6 income    33693    1670899    1719788
#>        sex        age  vname      uws         ws        pws

tab$numtab("savings")

#>        sex        age   vname     uws        ws         pws
#>  1:  Total      Total savings 2273532 136816190 136826351.7
#>  2:  Total age_group1 savings  982386  59288100  59284132.3
#>  3:  Total       ag1a savings  982386  59288100  59284132.3
#>  4:  Total age_group2 savings  552336  33518908  33520613.4
#>  5:  Total       ag2a savings  552336  33518908  33520613.4
#>  6:  Total age_group3 savings  437101  25798713  25801102.7
#>  7:  Total age_group4 savings  214661  12924977  12922294.8
#>  8:  Total age_group5 savings   80451   4903855   4903811.1
#>  9:  Total age_group6 savings    6597    381637    381472.2
#> 10: female      Total savings 1113716  67027458  67018609.1
#> 11:      f      Total savings 1113716  67027458  67018609.1
#> 12: female age_group1 savings  464726  28194967  28195251.3
#> 13:      f age_group1 savings  464726  28194967  28195251.3
#> 14: female       ag1a savings  464726  28194967  28195251.3
#> 15:      f       ag1a savings  464726  28194967  28195251.3
#> 16: female age_group2 savings  271413  16309077  16309782.8
#> 17:      f age_group2 savings  271413  16309077  16309782.8
#> 18: female       ag2a savings  271413  16309077  16309782.8
#> 19:      f       ag2a savings  271413  16309077  16309782.8
#> 20: female age_group3 savings  222131  13055109  13055872.2
#> 21:      f age_group3 savings  222131  13055109  13055872.2
#> 22: female age_group4 savings  115241   7006463   7008045.8
#> 23:      f age_group4 savings  115241   7006463   7008045.8
#> 24: female age_group5 savings   37218   2266714   2265180.8
#> 25:      f age_group5 savings   37218   2266714   2265180.8
#> 26: female age_group6 savings    2987    195128    194041.8
#> 27:      f age_group6 savings    2987    195128    194041.8
#> 28:   male      Total savings 1159816  69788732  69789747.2
#> 29:      m      Total savings 1159816  69788732  69789747.2
#> 30:   male age_group1 savings  517660  31093133  31093829.4
#> 31:      m age_group1 savings  517660  31093133  31093829.4
#> 32:   male       ag1a savings  517660  31093133  31093829.4
#> 33:      m       ag1a savings  517660  31093133  31093829.4
#> 34:   male age_group2 savings  280923  17209831  17204394.0
#> 35:      m age_group2 savings  280923  17209831  17204394.0
#> 36:   male       ag2a savings  280923  17209831  17204394.0
#> 37:      m       ag2a savings  280923  17209831  17204394.0
#> 38:   male age_group3 savings  214970  12743604  12745987.7
#> 39:      m age_group3 savings  214970  12743604  12745987.7
#> 40:   male age_group4 savings   99420   5918514   5922127.8
#> 41:      m age_group4 savings   99420   5918514   5922127.8
#> 42:   male age_group5 savings   43233   2637141   2636940.2
#> 43:      m age_group5 savings   43233   2637141   2636940.2
#> 44:   male age_group6 savings    3610    186509    188509.6
#> 45:      m age_group6 savings    3610    186509    188509.6
#>        sex        age   vname     uws        ws         pws

# results can be resetted, too
tab$reset_cntvars(v = "cnt_males")

# we can then set other parameters and perturb again
tab$params_cnts_set(val = p_cnts1, v = "cnt_males")

#> --> setting perturbation parameters for variable 'cnt_males'

tab$perturb(v = "cnt_males")

#> Count variable 'cnt_males' was perturbed.

# write to a file "outtab.csv" (.csv is automatically added to the path)
if (FALSE) {
tab$freqtab(v = c("total", "cnt_males"), path = "outtab.csv")
}

# show results containing weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))

#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 275400 4580 275400.0000
#>  2:  Total age_group1     total 1969 118766 1969 118766.0000
#>  3:  Total       ag1a     total 1969 118766 1969 118766.0000
#>  4:  Total age_group2     total 1143  69184 1144  69244.5284
#>  5:  Total       ag2a     total 1143  69184 1144  69244.5284
#>  6:  Total age_group3     total  864  51018  866  51136.0972
#>  7:  Total age_group4     total  423  25479  421  25358.5319
#>  8:  Total age_group5     total  168  10173  167  10112.4464
#>  9:  Total age_group6     total   13    780   11    660.0000
#> 10: female      Total     total 2284 137217 2285 137277.0775
#> 11:      f      Total     total 2284 137217 2285 137277.0775
#> 12: female age_group1     total  954  57639  954  57639.0000
#> 13:      f age_group1     total  954  57639  954  57639.0000
#> 14: female       ag1a     total  954  57639  954  57639.0000
#> 15:      f       ag1a     total  954  57639  954  57639.0000
#> 16: female age_group2     total  572  34521  573  34581.3514
#> 17:      f age_group2     total  572  34521  573  34581.3514
#> 18: female       ag2a     total  572  34521  573  34581.3514
#> 19:      f       ag2a     total  572  34521  573  34581.3514
#> 20: female age_group3     total  440  25910  441  25968.8864
#> 21:      f age_group3     total  440  25910  441  25968.8864
#> 22: female age_group4     total  228  13704  229  13764.1053
#> 23:      f age_group4     total  228  13704  229  13764.1053
#> 24: female age_group5     total   84   5035   85   5094.9405
#> 25:      f age_group5     total   84   5035   85   5094.9405
#> 26: female age_group6     total    6    408    7    476.0000
#> 27:      f age_group6     total    6    408    7    476.0000
#> 28:   male      Total     total 2296 138183 2297 138243.1842
#> 29:      m      Total     total 2296 138183 2297 138243.1842
#> 30:   male age_group1     total 1015  61127 1013  61006.5527
#> 31:      m age_group1     total 1015  61127 1013  61006.5527
#> 32:   male       ag1a     total 1015  61127 1013  61006.5527
#> 33:      m       ag1a     total 1015  61127 1013  61006.5527
#> 34:   male age_group2     total  571  34663  572  34723.7058
#> 35:      m age_group2     total  571  34663  572  34723.7058
#> 36:   male       ag2a     total  571  34663  572  34723.7058
#> 37:      m       ag2a     total  571  34663  572  34723.7058
#> 38:   male age_group3     total  424  25108  423  25048.7830
#> 39:      m age_group3     total  424  25108  423  25048.7830
#> 40:   male age_group4     total  195  11775  194  11714.6154
#> 41:      m age_group4     total  195  11775  194  11714.6154
#> 42:   male age_group5     total   84   5138   83   5076.8333
#> 43:      m age_group5     total   84   5138   83   5076.8333
#> 44:   male age_group6     total    7    372    6    318.8571
#> 45:      m age_group6     total    7    372    6    318.8571
#> 46:  Total      Total cnt_males 2296 138183 2297 138243.1842
#> 47:  Total age_group1 cnt_males 1015  61127 1013  61006.5527
#> 48:  Total       ag1a cnt_males 1015  61127 1013  61006.5527
#> 49:  Total age_group2 cnt_males  571  34663  572  34723.7058
#> 50:  Total       ag2a cnt_males  571  34663  572  34723.7058
#> 51:  Total age_group3 cnt_males  424  25108  423  25048.7830
#> 52:  Total age_group4 cnt_males  195  11775  194  11714.6154
#> 53:  Total age_group5 cnt_males   84   5138   83   5076.8333
#> 54:  Total age_group6 cnt_males    7    372    6    318.8571
#> 55: female      Total cnt_males    0      0    0      0.0000
#> 56:      f      Total cnt_males    0      0    0      0.0000
#> 57: female age_group1 cnt_males    0      0    0      0.0000
#> 58:      f age_group1 cnt_males    0      0    0      0.0000
#> 59: female       ag1a cnt_males    0      0    0      0.0000
#> 60:      f       ag1a cnt_males    0      0    0      0.0000
#> 61: female age_group2 cnt_males    0      0    0      0.0000
#> 62:      f age_group2 cnt_males    0      0    0      0.0000
#> 63: female       ag2a cnt_males    0      0    0      0.0000
#> 64:      f       ag2a cnt_males    0      0    0      0.0000
#> 65: female age_group3 cnt_males    0      0    0      0.0000
#> 66:      f age_group3 cnt_males    0      0    0      0.0000
#> 67: female age_group4 cnt_males    0      0    0      0.0000
#> 68:      f age_group4 cnt_males    0      0    0      0.0000
#> 69: female age_group5 cnt_males    0      0    0      0.0000
#> 70:      f age_group5 cnt_males    0      0    0      0.0000
#> 71: female age_group6 cnt_males    0      0    0      0.0000
#> 72:      f age_group6 cnt_males    0      0    0      0.0000
#> 73:   male      Total cnt_males 2296 138183 2297 138243.1842
#> 74:      m      Total cnt_males 2296 138183 2297 138243.1842
#> 75:   male age_group1 cnt_males 1015  61127 1013  61006.5527
#> 76:      m age_group1 cnt_males 1015  61127 1013  61006.5527
#> 77:   male       ag1a cnt_males 1015  61127 1013  61006.5527
#> 78:      m       ag1a cnt_males 1015  61127 1013  61006.5527
#> 79:   male age_group2 cnt_males  571  34663  572  34723.7058
#> 80:      m age_group2 cnt_males  571  34663  572  34723.7058
#> 81:   male       ag2a cnt_males  571  34663  572  34723.7058
#> 82:      m       ag2a cnt_males  571  34663  572  34723.7058
#> 83:   male age_group3 cnt_males  424  25108  423  25048.7830
#> 84:      m age_group3 cnt_males  424  25108  423  25048.7830
#> 85:   male age_group4 cnt_males  195  11775  194  11714.6154
#> 86:      m age_group4 cnt_males  195  11775  194  11714.6154
#> 87:   male age_group5 cnt_males   84   5138   83   5076.8333
#> 88:      m age_group5 cnt_males   84   5138   83   5076.8333
#> 89:   male age_group6 cnt_males    7    372    6    318.8571
#> 90:      m age_group6 cnt_males    7    372    6    318.8571
#>        sex        age     vname  uwc     wc puwc         pwc

# utility measures for a count variable
tab$measures_cnts(v = "total", exclude_zeros = TRUE)

#> $overview
#>    noise cnt        pct
#> 1:    -2   1 0.02222222
#> 2:    -1  22 0.48888889
#> 3:     0   7 0.15555556
#> 4:     1   9 0.20000000
#> 5:     2   6 0.13333333
#> 
#> $measures
#>       what d1    d2    d3
#>  1:    Min  0 0.000 0.000
#>  2:    Q10  0 0.000 0.000
#>  3:    Q20  1 0.000 0.010
#>  4:    Q30  1 0.002 0.021
#>  5:    Q40  1 0.002 0.021
#>  6:   Mean  1 0.020 0.045
#>  7: Median  1 0.002 0.024
#>  8:    Q60  1 0.002 0.031
#>  9:    Q70  1 0.004 0.034
#> 10:    Q80  1 0.007 0.050
#> 11:    Q90  2 0.090 0.140
#> 12:    Q95  2 0.152 0.196
#> 13:    Q99  2 0.167 0.248
#> 14:    Max  2 0.167 0.289
#> 
#> $cumdistr_d1
#>    cat cnt       pct
#> 1:   0   7 0.1555556
#> 2:   1  38 0.8444444
#> 3:   2  45 1.0000000
#> 
#> $cumdistr_d2
#>            cat cnt       pct
#> 1:    [0,0.02]  40 0.8888889
#> 2: (0.02,0.05]  40 0.8888889
#> 3:  (0.05,0.1]  40 0.8888889
#> 4:   (0.1,0.2]  45 1.0000000
#> 5:   (0.2,0.3]  45 1.0000000
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $cumdistr_d3
#>            cat cnt       pct
#> 1:    [0,0.02]  13 0.2888889
#> 2: (0.02,0.05]  36 0.8000000
#> 3:  (0.05,0.1]  40 0.8888889
#> 4:   (0.1,0.2]  44 0.9777778
#> 5:   (0.2,0.3]  45 1.0000000
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $false_zero
#> [1] 0
#> 
#> $false_nonzero
#> [1] 0
#> 
#> $exclude_zeros
#> [1] TRUE
#> 

# modifications for perturbed count variables
tab$mod_cnts()

#>        sex        age row_nr pert      ckey  countvar
#>   1: Total      Total     15    0 0.5531533     total
#>   2: Total age_group1     15    0 0.5645369     total
#>   3: Total       ag1a     15    0 0.5645369     total
#>   4: Total age_group2     16    1 0.7713398     total
#>   5: Total       ag2a     16    1 0.7713398     total
#>  ---                                                 
#> 131:     m age_group4     14   -1 0.1854716 cnt_males
#> 132:  male age_group5     14   -1 0.2518601 cnt_males
#> 133:     m age_group5     14   -1 0.2518601 cnt_males
#> 134:  male age_group6     14   -1 0.2937306 cnt_males
#> 135:     m age_group6     14   -1 0.2937306 cnt_males

# display a summary about utility measures
tab$summary()

#> ┌──────────────────────────────────────────────┐
#> │Utility measures for perturbed count variables│
#> └──────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#>          countvar Min Q10 Q20 Q30  Q40   Mean Median Q60 Q70 Q80 Q90 Q95  Q99
#> 1:          total  -2  -2  -1  -1  0.0  0.067      1   1   1 1.0   1   1 1.56
#> 2: cnt_highincome  -4  -2  -1  -1  0.0 -0.156      0   0   0 1.0   2   2 2.00
#> 3:      cnt_males  -2  -2  -1  -1 -0.4 -0.333      0   0   0 0.2   1   1 1.00
#>    Max
#> 1:   2
#> 2:   2
#> 3:   1
#> 
#> ── Distance-based measures ─────────────────────────────────────────────────────
#> ✔ Variable: 'total'
#> 
#>       what d1    d2    d3
#>  1:    Min  0 0.000 0.000
#>  2:    Q10  0 0.000 0.000
#>  3:    Q20  1 0.000 0.010
#>  4:    Q30  1 0.002 0.021
#>  5:    Q40  1 0.002 0.021
#>  6:   Mean  1 0.020 0.045
#>  7: Median  1 0.002 0.024
#>  8:    Q60  1 0.002 0.031
#>  9:    Q70  1 0.004 0.034
#> 10:    Q80  1 0.007 0.050
#> 11:    Q90  2 0.090 0.140
#> 12:    Q95  2 0.152 0.196
#> 13:    Q99  2 0.167 0.248
#> 14:    Max  2 0.167 0.289
#> 
#> ✔ Variable: 'cnt_males'
#> 
#>       what    d1    d2    d3
#>  1:    Min 1.000 0.000 0.010
#>  2:    Q10 1.000 0.001 0.017
#>  3:    Q20 1.000 0.002 0.021
#>  4:    Q30 1.000 0.002 0.021
#>  5:    Q40 1.000 0.002 0.024
#>  6:   Mean 1.222 0.019 0.047
#>  7: Median 1.000 0.002 0.031
#>  8:    Q60 1.000 0.002 0.031
#>  9:    Q70 1.000 0.005 0.036
#> 10:    Q80 1.800 0.011 0.051
#> 11:    Q90 2.000 0.064 0.111
#> 12:    Q95 2.000 0.143 0.196
#> 13:    Q99 2.000 0.143 0.196
#> 14:    Max 2.000 0.143 0.196
#> 
#> ✔ Variable: 'cnt_highincome'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 1.000 0.008 0.047
#>  6:   Mean 1.025 0.027 0.077
#>  7: Median 1.000 0.010 0.049
#>  8:    Q60 1.000 0.014 0.067
#>  9:    Q70 2.000 0.021 0.087
#> 10:    Q80 2.000 0.034 0.124
#> 11:    Q90 2.000 0.067 0.160
#> 12:    Q95 2.000 0.105 0.224
#> 13:    Q99 3.220 0.215 0.441
#> 14:    Max 4.000 0.286 0.579
#> 
#> ┌──────────────────────────────────────────────────┐
#> │Utility measures for perturbed numerical variables│
#> └──────────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────

#> Warning: no non-missing arguments to min; returning Inf

#> Warning: no non-missing arguments to max; returning -Inf

#> Warning: no non-missing arguments to min; returning Inf

#> Warning: no non-missing arguments to max; returning -Inf

#>      vname        Min        Q10        Q20        Q30       Q40      Mean
#> 1:  expend        Inf         NA         NA         NA        NA       NaN
#> 2:  income -76739.822 -45914.448 -35509.110 -13568.473 -3274.443 -2057.259
#> 3: savings  -8848.935  -5437.005  -1763.009   -200.836   284.285  -233.023
#> 4:   mixed        Inf         NA         NA         NA        NA       NaN
#>      Median       Q60       Q70       Q80       Q90       Q95       Q99
#> 1:       NA        NA        NA        NA        NA        NA        NA
#> 2: 3255.700 10107.347 11038.175 14930.449 38264.844 48889.389 68032.091
#> 3:  696.371   705.804   964.808  1705.439  2383.666  3368.943  7280.615
#> 4:       NA        NA        NA        NA        NA        NA        NA
#>         Max
#> 1:     -Inf
#> 2: 68032.09
#> 3: 10161.72
#> 4:     -Inf