Set parameters required to perturb numeric variables using a simple approach

ck_simpleparams() allows to define parameters for a simple perturbation approach based on a single magnitude parameter (m). The values of epsilon are used to "weight" parameter m in case type == "top_contr" is set in ck_params_nums().

ck_simpleparams(p, epsilon = 1)

Arguments

p

a percentage value used as magnitude for perturbation

epsilon

a numeric vector in descending order with all values >= 0 and <= 1 with the first element forced to equal 1. The length of this vector must correspond with the number top_k specified in ck_params_nums() when creating parameters for type == "top_contr" which is checked at runtime. This setting allows to use different flex-functions for the largest top_k contributors.

Value

an object suitable as input for ck_params_nums().

Details

details about the flex function can be found in Deliverable D4.2, Part I in SGA "Open Source tools for perturbative confidentiality methods"

See also

Examples

x <- ck_create_testdata()

# create some 0/1 variables that should be perturbed later
x[, cnt_females := ifelse(sex == "male", 0, 1)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              63          25.00000           0
#>    2:   2530      28       1              94          25.00000           1
#>    3:   6920     550       1              30          25.00000           0
#>    4:   7960     870       1              52          25.00000           0
#>    5:   9030      20       2              85          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              32          16.66667           1
#> 4577:   1420     987    1000              74          16.66667           0
#> 4578:   8900     684    1000              86          16.66667           0
#> 4579:   3880     294    1000              74          16.66667           1
#> 4580:   4830     911    1000              31          16.66667           0
x[, cnt_males := ifelse(sex == "male", 1, 0)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              63          25.00000           0
#>    2:   2530      28       1              94          25.00000           1
#>    3:   6920     550       1              30          25.00000           0
#>    4:   7960     870       1              52          25.00000           0
#>    5:   9030      20       2              85          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              32          16.66667           1
#> 4577:   1420     987    1000              74          16.66667           0
#> 4578:   8900     684    1000              86          16.66667           0
#> 4579:   3880     294    1000              74          16.66667           1
#> 4580:   4830     911    1000              31          16.66667           0
#>       cnt_males
#>    1:         1
#>    2:         0
#>    3:         1
#>    4:         1
#>    5:         1
#>   ---          
#> 4576:         0
#> 4577:         1
#> 4578:         1
#> 4579:         0
#> 4580:         1
x[, cnt_highincome := ifelse(income >= 9000, 1, 0)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              63          25.00000           0
#>    2:   2530      28       1              94          25.00000           1
#>    3:   6920     550       1              30          25.00000           0
#>    4:   7960     870       1              52          25.00000           0
#>    5:   9030      20       2              85          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              32          16.66667           1
#> 4577:   1420     987    1000              74          16.66667           0
#> 4578:   8900     684    1000              86          16.66667           0
#> 4579:   3880     294    1000              74          16.66667           1
#> 4580:   4830     911    1000              31          16.66667           0
#>       cnt_males cnt_highincome
#>    1:         1              0
#>    2:         0              0
#>    3:         1              0
#>    4:         1              0
#>    5:         1              1
#>   ---                         
#> 4576:         0              0
#> 4577:         1              0
#> 4578:         1              0
#> 4579:         0              0
#> 4580:         1              0
# a variable with positive and negative contributions
x[, mixed := sample(-10:10, nrow(x), replace = TRUE)]
#>       urbrur roof walls water electcon relat    sex        age hhcivil expend
#>    1:      2    4     3     3        1     1   male age_group3       2   9093
#>    2:      2    4     3     3        1     2 female age_group3       2   2734
#>    3:      2    4     3     3        1     3   male age_group1       1   2652
#>    4:      2    4     3     3        1     3   male age_group1       1   1807
#>    5:      2    4     2     3        1     1   male age_group4       2    671
#>   ---                                                                        
#> 4576:      2    4     3     4        1     2 female age_group3       2   3696
#> 4577:      2    4     3     4        1     3   male age_group1       1    282
#> 4578:      2    4     3     4        1     3   male age_group1       1    840
#> 4579:      2    4     3     4        1     3 female age_group1       1   6258
#> 4580:      2    4     3     4        1     3   male age_group1       1   7019
#>       income savings ori_hid sampling_weight household_weights cnt_females
#>    1:   5780      12       1              63          25.00000           0
#>    2:   2530      28       1              94          25.00000           1
#>    3:   6920     550       1              30          25.00000           0
#>    4:   7960     870       1              52          25.00000           0
#>    5:   9030      20       2              85          16.66667           0
#>   ---                                                                     
#> 4576:   7900     278    1000              32          16.66667           1
#> 4577:   1420     987    1000              74          16.66667           0
#> 4578:   8900     684    1000              86          16.66667           0
#> 4579:   3880     294    1000              74          16.66667           1
#> 4580:   4830     911    1000              31          16.66667           0
#>       cnt_males cnt_highincome mixed
#>    1:         1              0    -1
#>    2:         0              0     2
#>    3:         1              0    -4
#>    4:         1              0   -10
#>    5:         1              1    -1
#>   ---                               
#> 4576:         0              0     2
#> 4577:         1              0    -1
#> 4578:         1              0     8
#> 4579:         0              0    -5
#> 4580:         1              0     0

# create record keys
x$rkey <- ck_generate_rkeys(dat = x)

# define required inputs

# hierarchy with some bogus codes
d_sex <- hier_create(root = "Total", nodes = c("male", "female"))
d_sex <- hier_add(d_sex, root = "female", "f")
d_sex <- hier_add(d_sex, root = "male", "m")

d_age <- hier_create(root = "Total", nodes = paste0("age_group", 1:6))
d_age <- hier_add(d_age, root = "age_group1", "ag1a")
d_age <- hier_add(d_age, root = "age_group2", "ag2a")

# define the cell key object
countvars <- c("cnt_females", "cnt_males", "cnt_highincome")
numvars <- c("expend", "income", "savings", "mixed")
tab <- ck_setup(
  x = x,
  rkey = "rkey",
  dims = list(sex = d_sex, age = d_age),
  w = "sampling_weight",
  countvars = countvars,
  numvars = numvars)

# show some information about this table instance
tab$print() # identical with print(tab)
#> ── Table Information ───────────────────────────────────────────────────────────
#> ✔ 45 cells in 2 dimensions ('sex', 'age')
#> ✔ weights: yes
#> ── Tabulated / Perturbed countvars ─────────────────────────────────────────────
#> ☐ 'total'
#> ☐ 'cnt_females'
#> ☐ 'cnt_males'
#> ☐ 'cnt_highincome'
#> ── Tabulated / Perturbed numvars ───────────────────────────────────────────────
#> ☐ 'expend'
#> ☐ 'income'
#> ☐ 'savings'
#> ☐ 'mixed'

# information about the hierarchies
tab$hierarchy_info()
#> $sex
#>      code level is_leaf parent
#> 1:  Total     1   FALSE  Total
#> 2:   male     2   FALSE  Total
#> 3:      m     3    TRUE   male
#> 4: female     2   FALSE  Total
#> 5:      f     3    TRUE female
#> 
#> $age
#>          code level is_leaf     parent
#> 1:      Total     1   FALSE      Total
#> 2: age_group1     2   FALSE      Total
#> 3:       ag1a     3    TRUE age_group1
#> 4: age_group2     2   FALSE      Total
#> 5:       ag2a     3    TRUE age_group2
#> 6: age_group3     2    TRUE      Total
#> 7: age_group4     2    TRUE      Total
#> 8: age_group5     2    TRUE      Total
#> 9: age_group6     2    TRUE      Total
#> 

# which variables have been defined?
tab$allvars()
#> $cntvars
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"
#> 
#> $numvars
#> [1] "expend"  "income"  "savings" "mixed"  
#> 

# count variables
tab$cntvars()
#> [1] "total"          "cnt_females"    "cnt_males"      "cnt_highincome"

# continuous variables
tab$numvars()
#> [1] "expend"  "income"  "savings" "mixed"  

# create perturbation parameters for "total" variable and
# write to yaml-file

# create a ptable using functionality from the ptable-pkg
f_yaml <- tempfile(fileext = ".yaml")
p_cnts1 <- ck_params_cnts(
  ptab = ptable::pt_ex_cnts(),
  path = f_yaml)
#> yaml configuration '/tmp/RtmpRLPm23/file1e1b31879971.yaml' successfully written.

# read parameters from yaml-file and set them for variable `"total"`
p_cnts1 <- ck_read_yaml(path = f_yaml)

tab$params_cnts_set(val = p_cnts1, v = "total")
#> --> setting perturbation parameters for variable 'total'

# create alternative perturbation parameters by specifying parameters
para2 <- ptable::create_num_ptable(
  D = 8, V = 3, js = 2, create = FALSE)
#> Error in ptable::create_num_ptable(D = 8, V = 3, js = 2, create = FALSE): unused argument (js = 2)

p_cnts2 <- ck_params_cnts(ptab = para2)
#> Error in ck_params_cnts(ptab = para2): object 'para2' not found

# use these ptable it for the remaining variables
tab$params_cnts_set(val = p_cnts2, v = countvars)
#> Error in tab$params_cnts_set(val = p_cnts2, v = countvars): object 'p_cnts2' not found

# perturb a variable
tab$perturb(v = "total")
#> Count variable 'total' was perturbed.

# multiple variables can be perturbed as well
tab$perturb(v = c("cnt_males", "cnt_highincome"))
#> Error: No perturbation parameters found for variable 'cnt_males' - Please use method `$params_cnts_set().

# return weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))
#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 274202 4578 274082.2611
#>  2:  Total age_group1     total 1969 117405 1969 117405.0000
#>  3:  Total       ag1a     total 1969 117405 1969 117405.0000
#>  4:  Total age_group2     total 1143  68808 1145  68928.3990
#>  5:  Total       ag2a     total 1143  68808 1145  68928.3990
#>  6:  Total age_group3     total  864  51712  866  51831.7037
#>  7:  Total age_group4     total  423  25055  423  25055.0000
#>  8:  Total age_group5     total  168  10246  170  10367.9762
#>  9:  Total age_group6     total   13    976   12    900.9231
#> 10:   male      Total     total 2296 135996 2296 135996.0000
#> 11:      m      Total     total 2296 135996 2296 135996.0000
#> 12:   male age_group1     total 1015  59962 1014  59902.9241
#> 13:      m age_group1     total 1015  59962 1014  59902.9241
#> 14:   male       ag1a     total 1015  59962 1014  59902.9241
#> 15:      m       ag1a     total 1015  59962 1014  59902.9241
#> 16:   male age_group2     total  571  34272  570  34211.9790
#> 17:      m age_group2     total  571  34272  570  34211.9790
#> 18:   male       ag2a     total  571  34272  570  34211.9790
#> 19:      m       ag2a     total  571  34272  570  34211.9790
#> 20:   male age_group3     total  424  24859  424  24859.0000
#> 21:      m age_group3     total  424  24859  424  24859.0000
#> 22:   male age_group4     total  195  11476  196  11534.8513
#> 23:      m age_group4     total  195  11476  196  11534.8513
#> 24:   male age_group5     total   84   4937   84   4937.0000
#> 25:      m age_group5     total   84   4937   84   4937.0000
#> 26:   male age_group6     total    7    490    6    420.0000
#> 27:      m age_group6     total    7    490    6    420.0000
#> 28: female      Total     total 2284 138206 2284 138206.0000
#> 29:      f      Total     total 2284 138206 2284 138206.0000
#> 30: female age_group1     total  954  57443  953  57382.7872
#> 31:      f age_group1     total  954  57443  953  57382.7872
#> 32: female       ag1a     total  954  57443  953  57382.7872
#> 33:      f       ag1a     total  954  57443  953  57382.7872
#> 34: female age_group2     total  572  34536  572  34536.0000
#> 35:      f age_group2     total  572  34536  572  34536.0000
#> 36: female       ag2a     total  572  34536  572  34536.0000
#> 37:      f       ag2a     total  572  34536  572  34536.0000
#> 38: female age_group3     total  440  26853  439  26791.9705
#> 39:      f age_group3     total  440  26853  439  26791.9705
#> 40: female age_group4     total  228  13579  229  13638.5570
#> 41:      f age_group4     total  228  13579  229  13638.5570
#> 42: female age_group5     total   84   5309   84   5309.0000
#> 43:      f age_group5     total   84   5309   84   5309.0000
#> 44: female age_group6     total    6    486    8    648.0000
#> 45:      f age_group6     total    6    486    8    648.0000
#> 46:  Total      Total cnt_males 2296 135996   NA          NA
#> 47:  Total age_group1 cnt_males 1015  59962   NA          NA
#> 48:  Total       ag1a cnt_males 1015  59962   NA          NA
#> 49:  Total age_group2 cnt_males  571  34272   NA          NA
#> 50:  Total       ag2a cnt_males  571  34272   NA          NA
#> 51:  Total age_group3 cnt_males  424  24859   NA          NA
#> 52:  Total age_group4 cnt_males  195  11476   NA          NA
#> 53:  Total age_group5 cnt_males   84   4937   NA          NA
#> 54:  Total age_group6 cnt_males    7    490   NA          NA
#> 55:   male      Total cnt_males 2296 135996   NA          NA
#> 56:      m      Total cnt_males 2296 135996   NA          NA
#> 57:   male age_group1 cnt_males 1015  59962   NA          NA
#> 58:      m age_group1 cnt_males 1015  59962   NA          NA
#> 59:   male       ag1a cnt_males 1015  59962   NA          NA
#> 60:      m       ag1a cnt_males 1015  59962   NA          NA
#> 61:   male age_group2 cnt_males  571  34272   NA          NA
#> 62:      m age_group2 cnt_males  571  34272   NA          NA
#> 63:   male       ag2a cnt_males  571  34272   NA          NA
#> 64:      m       ag2a cnt_males  571  34272   NA          NA
#> 65:   male age_group3 cnt_males  424  24859   NA          NA
#> 66:      m age_group3 cnt_males  424  24859   NA          NA
#> 67:   male age_group4 cnt_males  195  11476   NA          NA
#> 68:      m age_group4 cnt_males  195  11476   NA          NA
#> 69:   male age_group5 cnt_males   84   4937   NA          NA
#> 70:      m age_group5 cnt_males   84   4937   NA          NA
#> 71:   male age_group6 cnt_males    7    490   NA          NA
#> 72:      m age_group6 cnt_males    7    490   NA          NA
#> 73: female      Total cnt_males    0      0   NA          NA
#> 74:      f      Total cnt_males    0      0   NA          NA
#> 75: female age_group1 cnt_males    0      0   NA          NA
#> 76:      f age_group1 cnt_males    0      0   NA          NA
#> 77: female       ag1a cnt_males    0      0   NA          NA
#> 78:      f       ag1a cnt_males    0      0   NA          NA
#> 79: female age_group2 cnt_males    0      0   NA          NA
#> 80:      f age_group2 cnt_males    0      0   NA          NA
#> 81: female       ag2a cnt_males    0      0   NA          NA
#> 82:      f       ag2a cnt_males    0      0   NA          NA
#> 83: female age_group3 cnt_males    0      0   NA          NA
#> 84:      f age_group3 cnt_males    0      0   NA          NA
#> 85: female age_group4 cnt_males    0      0   NA          NA
#> 86:      f age_group4 cnt_males    0      0   NA          NA
#> 87: female age_group5 cnt_males    0      0   NA          NA
#> 88:      f age_group5 cnt_males    0      0   NA          NA
#> 89: female age_group6 cnt_males    0      0   NA          NA
#> 90:      f age_group6 cnt_males    0      0   NA          NA
#>        sex        age     vname  uwc     wc puwc         pwc

# numerical variables (positive variables using flex-function)
# we also write the config to a yaml file
f_yaml <- tempfile(fileext = ".yaml")

# create a ptable using functionality from the ptable-pkg
# a single ptable for all cells
ptab1 <- ptable::pt_ex_nums(parity = TRUE, separation = FALSE)

# a single ptab for all cells except for very small ones
ptab2 <- ptable::pt_ex_nums(parity = TRUE, separation = TRUE)

# different ptables for cells with even/odd number of contributors
# and very small cells
ptab3 <- ptable::pt_ex_nums(parity = FALSE, separation = TRUE)

p_nums1 <- ck_params_nums(
  ptab = ptab1,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.30, 0.03),
    epsilon = c(1, 0.5, 0.2),
    q = 3),
  mu_c = 2,
  same_key = FALSE,
  use_zero_rkeys = FALSE,
  path = f_yaml)
#> yaml configuration '/tmp/RtmpRLPm23/file1e1b5e718f85.yaml' successfully written.

# we read the parameters from the yaml-file
p_nums1 <- ck_read_yaml(path = f_yaml)

# for variables with positive and negative values
p_nums2 <- ck_params_nums(
  ptab = ptab2,
  type = "top_contr",
  top_k = 3,
  mult_params = ck_flexparams(
    fp = 1000,
    p = c(0.15, 0.02),
    epsilon = c(1, 0.4, 0.15),
    q = 3),
  mu_c = 2,
  same_key = FALSE)

# simple perturbation parameters (not using the flex-function approach)
p_nums3 <- ck_params_nums(
  ptab = ptab3,
  type = "mean",
  mult_params = ck_simpleparams(p = 0.25),
  mu_c = 2,
  same_key = FALSE)

# use `p_nums1` for all variables
tab$params_nums_set(p_nums1, c("savings", "income", "expend"))
#> --> setting perturbation parameters for variable 'savings'
#> --> setting perturbation parameters for variable 'income'
#> --> setting perturbation parameters for variable 'expend'

# use different parameters for variable `mixed`
tab$params_nums_set(p_nums2, v = "mixed")
#> --> setting perturbation parameters for variable 'mixed'

# identify sensitive cells to which extra protection (`mu_c`) is added.
tab$supp_p(v = "income", p = 85)
#> computing contributing indices | rawdata <--> table; this might take a while
#> p%-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_pq(v = "income", p = 85, q = 90)
#> computing contributing indices | rawdata <--> table; this might take a while
#> pq-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_nk(v = "income", n = 2, k = 90)
#> computing contributing indices | rawdata <--> table; this might take a while
#> nk-rule: 0 new sensitive cells (incl. duplicates) found (total: 0)
tab$supp_freq(v = "income", n = 14, weighted = FALSE)
#> freq-rule: 5 new sensitive cells (incl. duplicates) found (total: 5)
tab$supp_val(v = "income", n = 10000, weighted = TRUE)
#> val-rule: 0 new sensitive cells (incl. duplicates) found (total: 5)
tab$supp_cells(
  v = "income",
  inp = data.frame(
    sex = c("female", "female"),
    "age" = c("age_group1", "age_group3")
  )
)
#> cell-rule: 2 new sensitive cells (incl. duplicates) found (total: 7)

# perturb variables
tab$perturb(v = c("income", "savings"))
#> Numeric variable 'income' was perturbed.
#> Numeric variable 'savings' was perturbed.

# extract results
tab$numtab("income", mean_before_sum = TRUE)
#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1370252349 1370250250
#>  2:  Total age_group1 income  9810547  582605606  582586907
#>  3:  Total       ag1a income  9810547  582605606  582586907
#>  4:  Total age_group2 income  5692119  339674911  339658002
#>  5:  Total       ag2a income  5692119  339674911  339658002
#>  6:  Total age_group3 income  4406946  264273559  264308773
#>  7:  Total age_group4 income  2133543  126965556  126913901
#>  8:  Total age_group5 income   848151   51812094   51835726
#>  9:  Total age_group6 income    61672    4920623    5018586
#> 10:   male      Total income 11262049  664222206  664241089
#> 11:      m      Total income 11262049  664222206  664241089
#> 12:   male age_group1 income  4877164  287017023  287023024
#> 13:      m age_group1 income  4877164  287017023  287023024
#> 14:   male       ag1a income  4877164  287017023  287023024
#> 15:      m       ag1a income  4877164  287017023  287023024
#> 16:   male age_group2 income  2811379  167435375  167525706
#> 17:      m age_group2 income  2811379  167435375  167525706
#> 18:   male       ag2a income  2811379  167435375  167525706
#> 19:      m       ag2a income  2811379  167435375  167525706
#> 20:   male age_group3 income  2168169  126596217  126635072
#> 21:      m age_group3 income  2168169  126596217  126635072
#> 22:   male age_group4 income   978510   57979387   57998030
#> 23:      m age_group4 income   978510   57979387   57998030
#> 24:   male age_group5 income   393134   22567895   22718203
#> 25:      m age_group5 income   393134   22567895   22718203
#> 26:   male age_group6 income    33693    2626309    2596630
#> 27:      m age_group6 income    33693    2626309    2596630
#> 28: female      Total income 11690929  706030143  705972819
#> 29:      f      Total income 11690929  706030143  705972819
#> 30: female age_group1 income  4933383  295588583  295650292
#> 31:      f age_group1 income  4933383  295588583  295650292
#> 32: female       ag1a income  4933383  295588583  295650292
#> 33:      f       ag1a income  4933383  295588583  295650292
#> 34: female age_group2 income  2880740  172239536  172291161
#> 35:      f age_group2 income  2880740  172239536  172291161
#> 36: female       ag2a income  2880740  172239536  172291161
#> 37:      f       ag2a income  2880740  172239536  172291161
#> 38: female age_group3 income  2238777  137677342  137637157
#> 39:      f age_group3 income  2238777  137677342  137637157
#> 40: female age_group4 income  1155033   68986169   68898376
#> 41:      f age_group4 income  1155033   68986169   68898376
#> 42: female age_group5 income   455017   29244199   29272125
#> 43:      f age_group5 income   455017   29244199   29272125
#> 44: female age_group6 income    27979    2294314    2160888
#> 45:      f age_group6 income    27979    2294314    2160888
#>        sex        age  vname      uws         ws        pws
tab$numtab("income", mean_before_sum = FALSE)
#>        sex        age  vname      uws         ws        pws
#>  1:  Total      Total income 22952978 1370252349 1370251300
#>  2:  Total age_group1 income  9810547  582605606  582596257
#>  3:  Total       ag1a income  9810547  582605606  582596257
#>  4:  Total age_group2 income  5692119  339674911  339666456
#>  5:  Total       ag2a income  5692119  339674911  339666456
#>  6:  Total age_group3 income  4406946  264273559  264291165
#>  7:  Total age_group4 income  2133543  126965556  126939726
#>  8:  Total age_group5 income   848151   51812094   51823909
#>  9:  Total age_group6 income    61672    4920623    4969363
#> 10:   male      Total income 11262049  664222206  664231647
#> 11:      m      Total income 11262049  664222206  664231647
#> 12:   male age_group1 income  4877164  287017023  287020024
#> 13:      m age_group1 income  4877164  287017023  287020024
#> 14:   male       ag1a income  4877164  287017023  287020024
#> 15:      m       ag1a income  4877164  287017023  287020024
#> 16:   male age_group2 income  2811379  167435375  167480534
#> 17:      m age_group2 income  2811379  167435375  167480534
#> 18:   male       ag2a income  2811379  167435375  167480534
#> 19:      m       ag2a income  2811379  167435375  167480534
#> 20:   male age_group3 income  2168169  126596217  126615643
#> 21:      m age_group3 income  2168169  126596217  126615643
#> 22:   male age_group4 income   978510   57979387   57988708
#> 23:      m age_group4 income   978510   57979387   57988708
#> 24:   male age_group5 income   393134   22567895   22642924
#> 25:      m age_group5 income   393134   22567895   22642924
#> 26:   male age_group6 income    33693    2626309    2611427
#> 27:      m age_group6 income    33693    2626309    2611427
#> 28: female      Total income 11690929  706030143  706001480
#> 29:      f      Total income 11690929  706030143  706001480
#> 30: female age_group1 income  4933383  295588583  295619436
#> 31:      f age_group1 income  4933383  295588583  295619436
#> 32: female       ag1a income  4933383  295588583  295619436
#> 33:      f       ag1a income  4933383  295588583  295619436
#> 34: female age_group2 income  2880740  172239536  172265346
#> 35:      f age_group2 income  2880740  172239536  172265346
#> 36: female       ag2a income  2880740  172239536  172265346
#> 37:      f       ag2a income  2880740  172239536  172265346
#> 38: female age_group3 income  2238777  137677342  137657248
#> 39:      f age_group3 income  2238777  137677342  137657248
#> 40: female age_group4 income  1155033   68986169   68942258
#> 41:      f age_group4 income  1155033   68986169   68942258
#> 42: female age_group5 income   455017   29244199   29258159
#> 43:      f age_group5 income   455017   29244199   29258159
#> 44: female age_group6 income    27979    2294314    2226602
#> 45:      f age_group6 income    27979    2294314    2226602
#>        sex        age  vname      uws         ws        pws
tab$numtab("savings")
#>        sex        age   vname     uws        ws         pws
#>  1:  Total      Total savings 2273532 135441366 135435804.2
#>  2:  Total age_group1 savings  982386  58351435  58354909.6
#>  3:  Total       ag1a savings  982386  58351435  58354909.6
#>  4:  Total age_group2 savings  552336  33382327  33381985.8
#>  5:  Total       ag2a savings  552336  33382327  33381985.8
#>  6:  Total age_group3 savings  437101  25917456  25911850.6
#>  7:  Total age_group4 savings  214661  12574911  12567690.9
#>  8:  Total age_group5 savings   80451   4732098   4732919.0
#>  9:  Total age_group6 savings    6597    483139    484694.7
#> 10:   male      Total savings 1159816  68113388  68114708.3
#> 11:      m      Total savings 1159816  68113388  68114708.3
#> 12:   male age_group1 savings  517660  30418547  30419687.5
#> 13:      m age_group1 savings  517660  30418547  30419687.5
#> 14:   male       ag1a savings  517660  30418547  30419687.5
#> 15:      m       ag1a savings  517660  30418547  30419687.5
#> 16:   male age_group2 savings  280923  16777919  16782617.6
#> 17:      m age_group2 savings  280923  16777919  16782617.6
#> 18:   male       ag2a savings  280923  16777919  16782617.6
#> 19:      m       ag2a savings  280923  16777919  16782617.6
#> 20:   male age_group3 savings  214970  12446508  12447552.9
#> 21:      m age_group3 savings  214970  12446508  12447552.9
#> 22:   male age_group4 savings   99420   5733640   5729266.9
#> 23:      m age_group4 savings   99420   5733640   5729266.9
#> 24:   male age_group5 savings   43233   2480407   2483708.6
#> 25:      m age_group5 savings   43233   2480407   2483708.6
#> 26:   male age_group6 savings    3610    256367    251603.9
#> 27:      m age_group6 savings    3610    256367    251603.9
#> 28: female      Total savings 1113716  67327978  67319660.6
#> 29:      f      Total savings 1113716  67327978  67319660.6
#> 30: female age_group1 savings  464726  27932888  27934082.4
#> 31:      f age_group1 savings  464726  27932888  27934082.4
#> 32: female       ag1a savings  464726  27932888  27934082.4
#> 33:      f       ag1a savings  464726  27932888  27934082.4
#> 34: female age_group2 savings  271413  16604408  16608395.7
#> 35:      f age_group2 savings  271413  16604408  16608395.7
#> 36: female       ag2a savings  271413  16604408  16608395.7
#> 37:      f       ag2a savings  271413  16604408  16608395.7
#> 38: female age_group3 savings  222131  13470948  13472532.6
#> 39:      f age_group3 savings  222131  13470948  13472532.6
#> 40: female age_group4 savings  115241   6841271   6836131.6
#> 41:      f age_group4 savings  115241   6841271   6836131.6
#> 42: female age_group5 savings   37218   2251691   2252612.2
#> 43:      f age_group5 savings   37218   2251691   2252612.2
#> 44: female age_group6 savings    2987    226772    225266.6
#> 45:      f age_group6 savings    2987    226772    225266.6
#>        sex        age   vname     uws        ws         pws

# results can be resetted, too
tab$reset_cntvars(v = "cnt_males")
#> Error: Invalid variables specified in `v`: Possible choices are:
#>  'total'

# we can then set other parameters and perturb again
tab$params_cnts_set(val = p_cnts1, v = "cnt_males")
#> --> setting perturbation parameters for variable 'cnt_males'
tab$perturb(v = "cnt_males")
#> Count variable 'cnt_males' was perturbed.

# write to a file "outtab.csv" (.csv is automatically added to the path)
if (FALSE) {
tab$freqtab(v = c("total", "cnt_males"), path = "outtab.csv")
}

# show results containing weighted and unweighted results
tab$freqtab(v = c("total", "cnt_males"))
#>        sex        age     vname  uwc     wc puwc         pwc
#>  1:  Total      Total     total 4580 274202 4578 274082.2611
#>  2:  Total age_group1     total 1969 117405 1969 117405.0000
#>  3:  Total       ag1a     total 1969 117405 1969 117405.0000
#>  4:  Total age_group2     total 1143  68808 1145  68928.3990
#>  5:  Total       ag2a     total 1143  68808 1145  68928.3990
#>  6:  Total age_group3     total  864  51712  866  51831.7037
#>  7:  Total age_group4     total  423  25055  423  25055.0000
#>  8:  Total age_group5     total  168  10246  170  10367.9762
#>  9:  Total age_group6     total   13    976   12    900.9231
#> 10:   male      Total     total 2296 135996 2296 135996.0000
#> 11:      m      Total     total 2296 135996 2296 135996.0000
#> 12:   male age_group1     total 1015  59962 1014  59902.9241
#> 13:      m age_group1     total 1015  59962 1014  59902.9241
#> 14:   male       ag1a     total 1015  59962 1014  59902.9241
#> 15:      m       ag1a     total 1015  59962 1014  59902.9241
#> 16:   male age_group2     total  571  34272  570  34211.9790
#> 17:      m age_group2     total  571  34272  570  34211.9790
#> 18:   male       ag2a     total  571  34272  570  34211.9790
#> 19:      m       ag2a     total  571  34272  570  34211.9790
#> 20:   male age_group3     total  424  24859  424  24859.0000
#> 21:      m age_group3     total  424  24859  424  24859.0000
#> 22:   male age_group4     total  195  11476  196  11534.8513
#> 23:      m age_group4     total  195  11476  196  11534.8513
#> 24:   male age_group5     total   84   4937   84   4937.0000
#> 25:      m age_group5     total   84   4937   84   4937.0000
#> 26:   male age_group6     total    7    490    6    420.0000
#> 27:      m age_group6     total    7    490    6    420.0000
#> 28: female      Total     total 2284 138206 2284 138206.0000
#> 29:      f      Total     total 2284 138206 2284 138206.0000
#> 30: female age_group1     total  954  57443  953  57382.7872
#> 31:      f age_group1     total  954  57443  953  57382.7872
#> 32: female       ag1a     total  954  57443  953  57382.7872
#> 33:      f       ag1a     total  954  57443  953  57382.7872
#> 34: female age_group2     total  572  34536  572  34536.0000
#> 35:      f age_group2     total  572  34536  572  34536.0000
#> 36: female       ag2a     total  572  34536  572  34536.0000
#> 37:      f       ag2a     total  572  34536  572  34536.0000
#> 38: female age_group3     total  440  26853  439  26791.9705
#> 39:      f age_group3     total  440  26853  439  26791.9705
#> 40: female age_group4     total  228  13579  229  13638.5570
#> 41:      f age_group4     total  228  13579  229  13638.5570
#> 42: female age_group5     total   84   5309   84   5309.0000
#> 43:      f age_group5     total   84   5309   84   5309.0000
#> 44: female age_group6     total    6    486    8    648.0000
#> 45:      f age_group6     total    6    486    8    648.0000
#> 46:  Total      Total cnt_males 2296 135996 2296 135996.0000
#> 47:  Total age_group1 cnt_males 1015  59962 1014  59902.9241
#> 48:  Total       ag1a cnt_males 1015  59962 1014  59902.9241
#> 49:  Total age_group2 cnt_males  571  34272  570  34211.9790
#> 50:  Total       ag2a cnt_males  571  34272  570  34211.9790
#> 51:  Total age_group3 cnt_males  424  24859  424  24859.0000
#> 52:  Total age_group4 cnt_males  195  11476  196  11534.8513
#> 53:  Total age_group5 cnt_males   84   4937   84   4937.0000
#> 54:  Total age_group6 cnt_males    7    490    6    420.0000
#> 55:   male      Total cnt_males 2296 135996 2296 135996.0000
#> 56:      m      Total cnt_males 2296 135996 2296 135996.0000
#> 57:   male age_group1 cnt_males 1015  59962 1014  59902.9241
#> 58:      m age_group1 cnt_males 1015  59962 1014  59902.9241
#> 59:   male       ag1a cnt_males 1015  59962 1014  59902.9241
#> 60:      m       ag1a cnt_males 1015  59962 1014  59902.9241
#> 61:   male age_group2 cnt_males  571  34272  570  34211.9790
#> 62:      m age_group2 cnt_males  571  34272  570  34211.9790
#> 63:   male       ag2a cnt_males  571  34272  570  34211.9790
#> 64:      m       ag2a cnt_males  571  34272  570  34211.9790
#> 65:   male age_group3 cnt_males  424  24859  424  24859.0000
#> 66:      m age_group3 cnt_males  424  24859  424  24859.0000
#> 67:   male age_group4 cnt_males  195  11476  196  11534.8513
#> 68:      m age_group4 cnt_males  195  11476  196  11534.8513
#> 69:   male age_group5 cnt_males   84   4937   84   4937.0000
#> 70:      m age_group5 cnt_males   84   4937   84   4937.0000
#> 71:   male age_group6 cnt_males    7    490    6    420.0000
#> 72:      m age_group6 cnt_males    7    490    6    420.0000
#> 73: female      Total cnt_males    0      0    0      0.0000
#> 74:      f      Total cnt_males    0      0    0      0.0000
#> 75: female age_group1 cnt_males    0      0    0      0.0000
#> 76:      f age_group1 cnt_males    0      0    0      0.0000
#> 77: female       ag1a cnt_males    0      0    0      0.0000
#> 78:      f       ag1a cnt_males    0      0    0      0.0000
#> 79: female age_group2 cnt_males    0      0    0      0.0000
#> 80:      f age_group2 cnt_males    0      0    0      0.0000
#> 81: female       ag2a cnt_males    0      0    0      0.0000
#> 82:      f       ag2a cnt_males    0      0    0      0.0000
#> 83: female age_group3 cnt_males    0      0    0      0.0000
#> 84:      f age_group3 cnt_males    0      0    0      0.0000
#> 85: female age_group4 cnt_males    0      0    0      0.0000
#> 86:      f age_group4 cnt_males    0      0    0      0.0000
#> 87: female age_group5 cnt_males    0      0    0      0.0000
#> 88:      f age_group5 cnt_males    0      0    0      0.0000
#> 89: female age_group6 cnt_males    0      0    0      0.0000
#> 90:      f age_group6 cnt_males    0      0    0      0.0000
#>        sex        age     vname  uwc     wc puwc         pwc

# utility measures for a count variable
tab$measures_cnts(v = "total", exclude_zeros = TRUE)
#> $overview
#>    noise cnt        pct
#> 1:    -2   6 0.13333333
#> 2:    -1   4 0.08888889
#> 3:     0  17 0.37777778
#> 4:     1  17 0.37777778
#> 5:     2   1 0.02222222
#> 
#> $measures
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 1.000 0.001 0.015
#>  6:   Mean 0.778 0.024 0.042
#>  7: Median 1.000 0.001 0.016
#>  8:    Q60 1.000 0.002 0.021
#>  9:    Q70 1.000 0.002 0.024
#> 10:    Q80 1.000 0.004 0.033
#> 11:    Q90 2.000 0.051 0.116
#> 12:    Q95 2.000 0.143 0.196
#> 13:    Q99 2.000 0.333 0.379
#> 14:    Max 2.000 0.333 0.379
#> 
#> $cumdistr_d1
#>    cat cnt       pct
#> 1:   0  17 0.3777778
#> 2:   1  38 0.8444444
#> 3:   2  45 1.0000000
#> 
#> $cumdistr_d2
#>            cat cnt       pct
#> 1:    [0,0.02]  40 0.8888889
#> 2: (0.02,0.05]  40 0.8888889
#> 3:  (0.05,0.1]  41 0.9111111
#> 4:   (0.1,0.2]  43 0.9555556
#> 5:   (0.2,0.3]  43 0.9555556
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $cumdistr_d3
#>            cat cnt       pct
#> 1:    [0,0.02]  26 0.5777778
#> 2: (0.02,0.05]  39 0.8666667
#> 3:  (0.05,0.1]  40 0.8888889
#> 4:   (0.1,0.2]  43 0.9555556
#> 5:   (0.2,0.3]  43 0.9555556
#> 6:   (0.3,0.4]  45 1.0000000
#> 7:   (0.4,0.5]  45 1.0000000
#> 8:   (0.5,Inf]  45 1.0000000
#> 
#> $false_zero
#> [1] 0
#> 
#> $false_nonzero
#> [1] 0
#> 
#> $exclude_zeros
#> [1] TRUE
#> 

# modifications for perturbed count variables
tab$mod_cnts()
#>        sex        age row_nr pert       ckey  countvar
#>  1:  Total      Total     13   -2 0.01455182     total
#>  2:  Total age_group1     15    0 0.43362424     total
#>  3:  Total       ag1a     15    0 0.43362424     total
#>  4:  Total age_group2     17    2 0.93463657     total
#>  5:  Total       ag2a     17    2 0.93463657     total
#>  6:  Total age_group3     17    2 0.93004553     total
#>  7:  Total age_group4     15    0 0.51303567     total
#>  8:  Total age_group5     17    2 0.95390799     total
#>  9:  Total age_group6     14   -1 0.24930182     total
#> 10:   male      Total     15    0 0.59053939     total
#> 11:      m      Total     15    0 0.59053939     total
#> 12:   male age_group1     14   -1 0.14725874     total
#> 13:      m age_group1     14   -1 0.14725874     total
#> 14:   male       ag1a     14   -1 0.14725874     total
#> 15:      m       ag1a     14   -1 0.14725874     total
#> 16:   male age_group2     14   -1 0.28576806     total
#> 17:      m age_group2     14   -1 0.28576806     total
#> 18:   male       ag2a     14   -1 0.28576806     total
#> 19:      m       ag2a     14   -1 0.28576806     total
#> 20:   male age_group3     15    0 0.64794971     total
#> 21:      m age_group3     15    0 0.64794971     total
#> 22:   male age_group4     16    1 0.80699331     total
#> 23:      m age_group4     16    1 0.80699331     total
#> 24:   male age_group5     15    0 0.39665878     total
#> 25:      m age_group5     15    0 0.39665878     total
#> 26:   male age_group6     14   -1 0.30591079     total
#> 27:      m age_group6     14   -1 0.30591079     total
#> 28: female      Total     15    0 0.42401243     total
#> 29:      f      Total     15    0 0.42401243     total
#> 30: female age_group1     14   -1 0.28636550     total
#> 31:      f age_group1     14   -1 0.28636550     total
#> 32: female       ag1a     14   -1 0.28636550     total
#> 33:      f       ag1a     14   -1 0.28636550     total
#> 34: female age_group2     15    0 0.64886851     total
#> 35:      f age_group2     15    0 0.64886851     total
#> 36: female       ag2a     15    0 0.64886851     total
#> 37:      f       ag2a     15    0 0.64886851     total
#> 38: female age_group3     14   -1 0.28209582     total
#> 39:      f age_group3     14   -1 0.28209582     total
#> 40: female age_group4     16    1 0.70604236     total
#> 41:      f age_group4     16    1 0.70604236     total
#> 42: female age_group5     15    0 0.55724921     total
#> 43:      f age_group5     15    0 0.55724921     total
#> 44: female age_group6     17    2 0.94339103     total
#> 45:      f age_group6     17    2 0.94339103     total
#> 46:  Total      Total     15    0 0.59053939 cnt_males
#> 47:  Total age_group1     14   -1 0.14725874 cnt_males
#> 48:  Total       ag1a     14   -1 0.14725874 cnt_males
#> 49:  Total age_group2     14   -1 0.28576806 cnt_males
#> 50:  Total       ag2a     14   -1 0.28576806 cnt_males
#> 51:  Total age_group3     15    0 0.64794971 cnt_males
#> 52:  Total age_group4     16    1 0.80699331 cnt_males
#> 53:  Total age_group5     15    0 0.39665878 cnt_males
#> 54:  Total age_group6     14   -1 0.30591079 cnt_males
#> 55:   male      Total     15    0 0.59053939 cnt_males
#> 56:      m      Total     15    0 0.59053939 cnt_males
#> 57:   male age_group1     14   -1 0.14725874 cnt_males
#> 58:      m age_group1     14   -1 0.14725874 cnt_males
#> 59:   male       ag1a     14   -1 0.14725874 cnt_males
#> 60:      m       ag1a     14   -1 0.14725874 cnt_males
#> 61:   male age_group2     14   -1 0.28576806 cnt_males
#> 62:      m age_group2     14   -1 0.28576806 cnt_males
#> 63:   male       ag2a     14   -1 0.28576806 cnt_males
#> 64:      m       ag2a     14   -1 0.28576806 cnt_males
#> 65:   male age_group3     15    0 0.64794971 cnt_males
#> 66:      m age_group3     15    0 0.64794971 cnt_males
#> 67:   male age_group4     16    1 0.80699331 cnt_males
#> 68:      m age_group4     16    1 0.80699331 cnt_males
#> 69:   male age_group5     15    0 0.39665878 cnt_males
#> 70:      m age_group5     15    0 0.39665878 cnt_males
#> 71:   male age_group6     14   -1 0.30591079 cnt_males
#> 72:      m age_group6     14   -1 0.30591079 cnt_males
#> 73: female      Total     -1    0 0.00000000 cnt_males
#> 74:      f      Total     -1    0 0.00000000 cnt_males
#> 75: female age_group1     -1    0 0.00000000 cnt_males
#> 76:      f age_group1     -1    0 0.00000000 cnt_males
#> 77: female       ag1a     -1    0 0.00000000 cnt_males
#> 78:      f       ag1a     -1    0 0.00000000 cnt_males
#> 79: female age_group2     -1    0 0.00000000 cnt_males
#> 80:      f age_group2     -1    0 0.00000000 cnt_males
#> 81: female       ag2a     -1    0 0.00000000 cnt_males
#> 82:      f       ag2a     -1    0 0.00000000 cnt_males
#> 83: female age_group3     -1    0 0.00000000 cnt_males
#> 84:      f age_group3     -1    0 0.00000000 cnt_males
#> 85: female age_group4     -1    0 0.00000000 cnt_males
#> 86:      f age_group4     -1    0 0.00000000 cnt_males
#> 87: female age_group5     -1    0 0.00000000 cnt_males
#> 88:      f age_group5     -1    0 0.00000000 cnt_males
#> 89: female age_group6     -1    0 0.00000000 cnt_males
#> 90:      f age_group6     -1    0 0.00000000 cnt_males
#>        sex        age row_nr pert       ckey  countvar

# display a summary about utility measures
tab$summary()
#> ┌──────────────────────────────────────────────┐
#> │Utility measures for perturbed count variables│
#> └──────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#>     countvar Min Q10 Q20 Q30  Q40   Mean Median Q60 Q70 Q80 Q90 Q95 Q99 Max
#> 1:     total  -2  -1  -1  -1 -0.4 -0.067      0   0   0   1   2 2.0   2   2
#> 2: cnt_males  -1  -1  -1  -1  0.0 -0.267      0   0   0   0   0 0.8   1   1
#> 
#> ── Distance-based measures ─────────────────────────────────────────────────────
#> ✔ Variable: 'total'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 1.000 0.001 0.015
#>  6:   Mean 0.778 0.024 0.042
#>  7: Median 1.000 0.001 0.016
#>  8:    Q60 1.000 0.002 0.021
#>  9:    Q70 1.000 0.002 0.024
#> 10:    Q80 1.000 0.004 0.033
#> 11:    Q90 2.000 0.051 0.116
#> 12:    Q95 2.000 0.143 0.196
#> 13:    Q99 2.000 0.333 0.379
#> 14:    Max 2.000 0.333 0.379
#> 
#> ✔ Variable: 'cnt_males'
#> 
#>       what    d1    d2    d3
#>  1:    Min 0.000 0.000 0.000
#>  2:    Q10 0.000 0.000 0.000
#>  3:    Q20 0.000 0.000 0.000
#>  4:    Q30 0.000 0.000 0.000
#>  5:    Q40 1.000 0.001 0.016
#>  6:   Mean 0.667 0.017 0.034
#>  7: Median 1.000 0.001 0.016
#>  8:    Q60 1.000 0.002 0.021
#>  9:    Q70 1.000 0.002 0.021
#> 10:    Q80 1.000 0.004 0.033
#> 11:    Q90 1.000 0.060 0.100
#> 12:    Q95 1.000 0.143 0.196
#> 13:    Q99 1.000 0.143 0.196
#> 14:    Max 1.000 0.143 0.196
#> 
#> ┌──────────────────────────────────────────────────┐
#> │Utility measures for perturbed numerical variables│
#> └──────────────────────────────────────────────────┘
#> ── Distribution statistics of perturbations ────────────────────────────────────
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to min; returning Inf
#> Warning: no non-missing arguments to max; returning -Inf
#>      vname        Min        Q10        Q20       Q30      Q40     Mean
#> 1:  expend        Inf         NA         NA        NA       NA      NaN
#> 2:  income -67712.152 -28662.554 -15924.216 -8454.695 3000.563 7528.859
#> 3: savings  -8317.433  -5392.838  -4451.097  -341.178  995.375   55.327
#> 4:   mixed        Inf         NA         NA        NA       NA      NaN
#>      Median      Q60      Q70       Q80       Q90       Q95       Q99       Max
#> 1:       NA       NA       NA        NA        NA        NA        NA      -Inf
#> 2: 9441.489 15418.25 25810.49 30852.937 45159.198 48024.076 75029.156 75029.156
#> 3: 1140.524  1194.44  1578.82  3474.636  3987.668  4698.634  4698.634  4698.634
#> 4:       NA       NA       NA        NA        NA        NA        NA      -Inf